Terrestrial Mammal Bioegeography Data with Phylogeny

Data on terrestrial mammals biogeographic distributions across the world's ecoregions joined to a pf object (for {phyf})

Usage

mammal_biogeo

Format

mammal_biogeo

A 'pf' data frame (subclasses tibble) with 10,648 rows and 844 columns:

label

Node labels including species name for the tip labels

phlo

The phylogenetic flow column which stores the phylogenetic information

IUCN_binomial

Species name used by the IUCN, matches with IUCN range polygons

body_mass_median

Mammal species' body mass

litter_clutch_size

Mammal species' average clutch size (# of offspring in a litter)

activity)

Mammal species' primary time of activity

hab_breadth

...

volant

...

diet_5cat

...

range_size_km2

Mammal species' range size in kilometers squared

threat

Mammal species' IUCN threat category

ecoregion:eco_id

Each of the 832 columns starting with "ecoregion:" represents the proportion of the mammal species' range that fall in the ecoregion with id equal to "eco_id".

Source

https://ecoregions.appspot.com, https://www.iucnredlist.org/resources/spatial-data-download

References

None yet.