Return a pfc with ancestral features

Return a pfc with ancestral features

Usage

pf_anc(x, replace = 0)

Arguments

x

A pfc object

replace

Value to replace edge feature with no ancestral value (such as when the ancestor is the root)

Value

A new pfc with the same structure as x, but with the features of each edge's ancestors instead

Examples

pf_anc(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] --0.0000->Node16--1.4850->Node24--0.0464->t21                                                                                                                                                                           
#>   [2] -0.00000->Node16-1.48504->Node24-0.04642->Node43-0.02804->Node55-0.00734->t75                                                                                                                                           
#>   [3] -0.00000->Node16-1.48504->Node24-0.04642->Node43-0.02804->Node55-0.00734->t12                                                                                                                                           
#>   [4] -0.00000->Node16-1.48504->Node24-0.04642->Node43-0.02804->Node79-0.01923->Node98-0.00607->t4                                                                                                                            
#>   [5] -0.00000->Node16-1.48504->Node24-0.04642->Node43-0.02804->Node79-0.01923->Node98-0.00607->t50                                                                                                                           
#>   [6] --0.0000->Node16--1.4850->Node24--0.0464->Node43--0.0280->Node79--0.0192->t47                                                                                                                                           
#>   [7] --0.0000->Node16--1.4850->Node90--0.0965->t42                                                                                                                                                                           
#>   [8] --0.0000->Node16--1.4850->Node90--0.0965->t3                                                                                                                                                                            
#>   [9] --0.000-->Node2--0.915-->Node9--0.486-->t34                                                                                                                                                                             
#>  [10] --0.000-->Node2--0.915-->Node9--0.486-->Node38--0.151-->t10                                                                                                                                                             
#>  [11] -0.00000->Node2-0.91474->Node9-0.48577->Node38-0.15082->Node45-0.00856->t2                                                                                                                                              
#>  [12] -0.00000->Node2-0.91474->Node9-0.48577->Node38-0.15082->Node45-0.00856->t44                                                                                                                                             
#>  [13] --0.0000->Node2--0.9147->Node3--0.1332->Node11--0.3876->Node17--0.0499->Node19--0.0229->t72                                                                                                                             
#>  [14] --0.0000->Node2--0.9147->Node3--0.1332->Node11--0.3876->Node17--0.0499->Node19--0.0229->t15                                                                                                                             
#>  [15] --0.0000->Node2--0.9147->Node3--0.1332->Node11--0.3876->Node17--0.0499->t57                                                                                                                                             
#>  [16] --0.000-->Node2--0.915-->Node3--0.133-->Node11--0.388-->t22                                                                                                                                                             
#>  [17] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node5--0.1355->Node18--0.1902->Node29--0.0355->t97                                                                                                               
#>  [18] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node5--0.1355->Node18--0.1902->Node29--0.0355->t79                                                                                                               
#>  [19] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node5--0.1355->Node18--0.1902->Node33--0.0402->t20                                                                                                               
#>  [20] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node5--0.1355->Node18--0.1902->Node33--0.0402->t80                                                                                                               
#>  [21] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node20-0.18934->Node77-0.06504->Node96-0.00571->t59                                                                                
#>  [22] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node20-0.18934->Node77-0.06504->Node96-0.00571->t31                                                                                
#>  [23] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node20-0.18934->Node77-0.06504->Node82-0.00082->t62                                                                                
#>  [24] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node20-0.18934->Node77-0.06504->Node82-0.00082->t8                                                                                 
#>  [25] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node20-0.18934->t81                                                                                                                
#>  [26] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node35-0.07317->Node62-0.02385->Node85-0.00948->t25                                                                
#>  [27] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node35-0.07317->Node62-0.02385->Node85-0.00948->t7                                                                 
#>  [28] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node35-0.07317->Node62-0.02385->t40                                                                                
#>  [29] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node35-0.07317->Node40-0.00685->Node49-0.00835->Node54-0.00260->t55                                                
#>  [30] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node35-0.07317->Node40-0.00685->Node49-0.00835->Node54-0.00260->Node57-0.00173->t87                                
#>  [31] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node35-0.07317->Node40-0.00685->Node49-0.00835->Node54-0.00260->Node57-0.00173->t39                                
#>  [32] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node35-0.07317->Node40-0.00685->Node49-0.00835->t91                                                                
#>  [33] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node35-0.07317->Node40-0.00685->t37                                                                                
#>  [34] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node21-0.03943->Node92-0.06802->t88                                                                                
#>  [35] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node21-0.03943->Node92-0.06802->t69                                                                                
#>  [36] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node21-0.03943->t28                                                                                                
#>  [37] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node6--0.1425->Node14--0.1302->Node22--0.0718->Node34--0.0217->t73                                                                                               
#>  [38] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node6--0.1425->Node14--0.1302->Node22--0.0718->Node34--0.0217->Node37--0.0023->t61                                                                               
#>  [39] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node6--0.1425->Node14--0.1302->Node22--0.0718->Node34--0.0217->Node37--0.0023->Node88--0.0317->t63                                                               
#>  [40] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node6--0.1425->Node14--0.1302->Node22--0.0718->Node34--0.0217->Node37--0.0023->Node88--0.0317->t19                                                               
#>  [41] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node6--0.1425->Node14--0.1302->Node22--0.0718->t94                                                                                                               
#>  [42] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node30-0.01598->Node36-0.00627->Node47-0.01290->Node74-0.01545->t90                                               
#>  [43] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node30-0.01598->Node36-0.00627->Node47-0.01290->Node74-0.01545->t54                                               
#>  [44] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node30-0.01598->Node36-0.00627->Node47-0.01290->Node75-0.01617->t32                                               
#>  [45] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node30-0.01598->Node36-0.00627->Node47-0.01290->Node75-0.01617->Node83-0.00144->t14                               
#>  [46] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node30-0.01598->Node36-0.00627->Node47-0.01290->Node75-0.01617->Node83-0.00144->Node94-0.00452->t11               
#>  [47] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node30-0.01598->Node36-0.00627->Node47-0.01290->Node75-0.01617->Node83-0.00144->Node94-0.00452->t46               
#>  [48] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node30-0.01598->Node36-0.00627->Node53-0.01643->t41                                                               
#>  [49] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node30-0.01598->Node36-0.00627->Node53-0.01643->Node89-0.01631->t45                                               
#>  [50] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node30-0.01598->Node36-0.00627->Node53-0.01643->Node89-0.01631->t65                                               
#>  [51] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node6--0.1425->Node14--0.1302->Node23--0.0728->Node30--0.0160->Node61--0.0285->t56                                                                               
#>  [52] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node6--0.1425->Node14--0.1302->Node23--0.0728->Node30--0.0160->Node61--0.0285->t43                                                                               
#>  [53] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node6--0.1425->Node14--0.1302->Node23--0.0728->Node84--0.0534->t5                                                                                                
#>  [54] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node84-0.05343->Node93-0.00335->t96                                                                               
#>  [55] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node84-0.05343->Node93-0.00335->Node99-0.00163->t23                                                               
#>  [56] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node84-0.05343->Node93-0.00335->Node99-0.00163->t24                                                               
#>  [57] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node39-0.00719->t60                                                                                
#>  [58] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node39-0.00719->Node56-0.01455->t6                                                                 
#>  [59] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node39-0.00719->Node56-0.01455->Node76-0.01107->t82                                                
#>  [60] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node39-0.00719->Node56-0.01455->Node76-0.01107->t36                                                
#>  [61] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node44-0.01419->Node50-0.00412->Node63-0.00817->t95                                                
#>  [62] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node10-0.084505>Node32-0.132279>Node44-0.014195>Node50-0.004120>Node63-0.008166>Node65-0.000465>Node68-0.001593>Node72-0.002490>t99
#>  [63] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node10-0.084505>Node32-0.132279>Node44-0.014195>Node50-0.004120>Node63-0.008166>Node65-0.000465>Node68-0.001593>Node72-0.002490>t86
#>  [64] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node10-0.084505>Node32-0.132279>Node44-0.014195>Node50-0.004120>Node63-0.008166>Node65-0.000465>Node68-0.001593>t53                
#>  [65] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node10-0.084505>Node32-0.132279>Node44-0.014195>Node50-0.004120>Node63-0.008166>Node65-0.000465>t26                                
#>  [66] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node44-0.01419->Node50-0.00412->t70                                                                
#>  [67] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node44-0.01419->Node51-0.00534->Node78-0.01362->Node97-0.00614->t74                                
#>  [68] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node44-0.01419->Node51-0.00534->Node78-0.01362->Node97-0.00614->t83                                
#>  [69] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node44-0.01419->Node51-0.00534->Node78-0.01362->t92                                                
#>  [70] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node10-0.084505>Node32-0.132279>Node44-0.014195>Node51-0.005336>Node52-0.000153>Node59-0.004827>t52                                
#>  [71] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node10-0.084505>Node32-0.132279>Node44-0.014195>Node51-0.005336>Node52-0.000153>Node59-0.004827>t49                                
#>  [72] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node10-0.084505>Node32-0.132279>Node44-0.014195>Node51-0.005336>Node52-0.000153>t27                                                
#>  [73] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node25-0.12014->Node58-0.03636->t58                                                                                
#>  [74] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node25-0.12014->Node58-0.03636->t38                                                                                
#>  [75] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node25-0.12014->t76                                                                                                
#>  [76] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node42-0.10506->Node60-0.01241->t85                                                                
#>  [77] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node42-0.10506->Node60-0.01241->t29                                                                
#>  [78] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node42-0.10506->Node64-0.01342->Node67-0.00122->t1                                                 
#>  [79] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node42-0.10506->Node64-0.01342->Node67-0.00122->Node71-0.00192->Node91-0.00672->t89                
#>  [80] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node42-0.10506->Node64-0.01342->Node67-0.00122->Node71-0.00192->Node91-0.00672->t35                
#>  [81] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node42-0.10506->Node64-0.01342->Node67-0.00122->Node71-0.00192->t77                                
#>  [82] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node42-0.10506->Node64-0.01342->t48                                                                
#>  [83] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node26-0.08285->Node41-0.02005->Node46-0.00398->t18                                                
#>  [84] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node26-0.08285->Node41-0.02005->Node46-0.00398->Node48-0.00247->t33                                
#>  [85] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node26-0.08285->Node41-0.02005->Node46-0.00398->Node48-0.00247->Node80-0.01628->t30                
#>  [86] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node26-0.08285->Node41-0.02005->Node46-0.00398->Node48-0.00247->Node80-0.01628->t67                
#>  [87] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node26-0.08285->Node41-0.02005->t13                                                                
#>  [88] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node26-0.08285->Node87-0.04482->t68                                                                
#>  [89] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node26-0.08285->Node87-0.04482->Node95-0.00272->t16                                                
#>  [90] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node26-0.08285->Node87-0.04482->Node95-0.00272->t71                                                
#>  [91] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node27-0.08976->t78                                                                                                
#>  [92] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node12-0.120110>Node27-0.089763>Node28-0.000151>Node31-0.005094>Node70-0.030675>t66                                                
#>  [93] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node12-0.120110>Node27-0.089763>Node28-0.000151>Node31-0.005094>Node70-0.030675>t98                                                
#>  [94] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node12-0.120110>Node27-0.089763>Node28-0.000151>Node31-0.005094>t9                                                                 
#>  [95] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node12-0.120110>Node27-0.089763>Node28-0.000151>Node66-0.034443>t100                                                               
#>  [96] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node12-0.120110>Node27-0.089763>Node28-0.000151>Node66-0.034443>Node69-0.000928>Node73-0.003053>t51                                
#>  [97] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node12-0.120110>Node27-0.089763>Node28-0.000151>Node66-0.034443>Node69-0.000928>Node73-0.003053>Node81-0.002120>Node86-0.001941>t64
#>  [98] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node12-0.120110>Node27-0.089763>Node28-0.000151>Node66-0.034443>Node69-0.000928>Node73-0.003053>Node81-0.002120>Node86-0.001941>t93
#>  [99] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node12-0.120110>Node27-0.089763>Node28-0.000151>Node66-0.034443>Node69-0.000928>Node73-0.003053>Node81-0.002120>t17                
#> [100] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node12-0.120110>Node27-0.089763>Node28-0.000151>Node66-0.034443>Node69-0.000928>t84                                                
#> [101] ----0---->Node2                                                                                                                                                                                                         
#> [102] --0.000-->Node2--0.915-->Node3                                                                                                                                                                                          
#> [103] --0.000-->Node2--0.915-->Node3--0.133-->Node4                                                                                                                                                                           
#> [104] --0.000-->Node2--0.915-->Node3--0.133-->Node4--0.133-->Node5                                                                                                                                                            
#> [105] --0.000-->Node2--0.915-->Node3--0.133-->Node4--0.133-->Node6                                                                                                                                                            
#> [106] --0.000-->Node2--0.915-->Node3--0.133-->Node4--0.133-->Node5--0.136-->Node7                                                                                                                                             
#> [107] --0.000-->Node2--0.915-->Node3--0.133-->Node4--0.133-->Node6--0.142-->Node8                                                                                                                                             
#> [108] --0.000-->Node2--0.915-->Node9                                                                                                                                                                                          
#> [109] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10                                                                                                                             
#> [110] --0.000-->Node2--0.915-->Node3--0.133-->Node11                                                                                                                                                                          
#> [111] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12                                                                                                                             
#> [112] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13                                                                                                             
#> [113] --0.000-->Node2--0.915-->Node3--0.133-->Node4--0.133-->Node6--0.142-->Node14                                                                                                                                            
#> [114] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15                                                                                                                             
#> [115] ----0---->Node16                                                                                                                                                                                                        
#> [116] --0.000-->Node2--0.915-->Node3--0.133-->Node11--0.388-->Node17                                                                                                                                                          
#> [117] --0.000-->Node2--0.915-->Node3--0.133-->Node4--0.133-->Node5--0.136-->Node18                                                                                                                                            
#> [118] --0.0000->Node2--0.9147->Node3--0.1332->Node11--0.3876->Node17--0.0499->Node19                                                                                                                                          
#> [119] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node20                                                                                                                             
#> [120] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node21                                                                                                             
#> [121] --0.000-->Node2--0.915-->Node3--0.133-->Node4--0.133-->Node6--0.142-->Node14--0.130-->Node22                                                                                                                            
#> [122] --0.000-->Node2--0.915-->Node3--0.133-->Node4--0.133-->Node6--0.142-->Node14--0.130-->Node23                                                                                                                            
#> [123] ---0.00-->Node16---1.49-->Node24                                                                                                                                                                                        
#> [124] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node25                                                                                                             
#> [125] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node26                                                                                             
#> [126] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node27                                                                                                             
#> [127] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node27-0.08976->Node28                                                                                             
#> [128] --0.000-->Node2--0.915-->Node3--0.133-->Node4--0.133-->Node5--0.136-->Node18--0.190-->Node29                                                                                                                            
#> [129] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node6--0.1425->Node14--0.1302->Node23--0.0728->Node30                                                                                                            
#> [130] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node12-0.120110>Node27-0.089763>Node28-0.000151>Node31                                                                             
#> [131] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32                                                                                                             
#> [132] --0.000-->Node2--0.915-->Node3--0.133-->Node4--0.133-->Node5--0.136-->Node18--0.190-->Node33                                                                                                                            
#> [133] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node6--0.1425->Node14--0.1302->Node22--0.0718->Node34                                                                                                            
#> [134] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node35                                                                                                             
#> [135] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node6--0.1425->Node14--0.1302->Node23--0.0728->Node30--0.0160->Node36                                                                                            
#> [136] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node6--0.1425->Node14--0.1302->Node22--0.0718->Node34--0.0217->Node37                                                                                            
#> [137] --0.000-->Node2--0.915-->Node9--0.486-->Node38                                                                                                                                                                          
#> [138] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node39                                                                                             
#> [139] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node35-0.07317->Node40                                                                                             
#> [140] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node26-0.08285->Node41                                                                             
#> [141] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node42                                                                                             
#> [142] --0.0000->Node16--1.4850->Node24--0.0464->Node43                                                                                                                                                                        
#> [143] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node44                                                                                             
#> [144] --0.000-->Node2--0.915-->Node9--0.486-->Node38--0.151-->Node45                                                                                                                                                          
#> [145] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node26-0.08285->Node41-0.02005->Node46                                                             
#> [146] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node30-0.01598->Node36-0.00627->Node47                                                                            
#> [147] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node26-0.08285->Node41-0.02005->Node46-0.00398->Node48                                             
#> [148] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node35-0.07317->Node40-0.00685->Node49                                                                             
#> [149] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node44-0.01419->Node50                                                                             
#> [150] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node44-0.01419->Node51                                                                             
#> [151] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node44-0.01419->Node51-0.00534->Node52                                                             
#> [152] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node30-0.01598->Node36-0.00627->Node53                                                                            
#> [153] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node35-0.07317->Node40-0.00685->Node49-0.00835->Node54                                                             
#> [154] --0.0000->Node16--1.4850->Node24--0.0464->Node43--0.0280->Node55                                                                                                                                                        
#> [155] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node39-0.00719->Node56                                                                             
#> [156] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node35-0.07317->Node40-0.00685->Node49-0.00835->Node54-0.00260->Node57                                             
#> [157] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node25-0.12014->Node58                                                                                             
#> [158] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node10-0.084505>Node32-0.132279>Node44-0.014195>Node51-0.005336>Node52-0.000153>Node59                                             
#> [159] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node42-0.10506->Node60                                                                             
#> [160] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node6--0.1425->Node14--0.1302->Node23--0.0728->Node30--0.0160->Node61                                                                                            
#> [161] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node35-0.07317->Node62                                                                                             
#> [162] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node44-0.01419->Node50-0.00412->Node63                                                             
#> [163] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node42-0.10506->Node64                                                                             
#> [164] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node44-0.01419->Node50-0.00412->Node63-0.00817->Node65                                             
#> [165] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node12-0.120110>Node27-0.089763>Node28-0.000151>Node66                                                                             
#> [166] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node42-0.10506->Node64-0.01342->Node67                                                             
#> [167] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node10-0.084505>Node32-0.132279>Node44-0.014195>Node50-0.004120>Node63-0.008166>Node65-0.000465>Node68                             
#> [168] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node12-0.120110>Node27-0.089763>Node28-0.000151>Node66-0.034443>Node69                                                             
#> [169] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node12-0.120110>Node27-0.089763>Node28-0.000151>Node31-0.005094>Node70                                                             
#> [170] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node42-0.10506->Node64-0.01342->Node67-0.00122->Node71                                             
#> [171] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node10-0.084505>Node32-0.132279>Node44-0.014195>Node50-0.004120>Node63-0.008166>Node65-0.000465>Node68-0.001593>Node72             
#> [172] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node12-0.120110>Node27-0.089763>Node28-0.000151>Node66-0.034443>Node69-0.000928>Node73                                             
#> [173] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node30-0.01598->Node36-0.00627->Node47-0.01290->Node74                                                            
#> [174] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node30-0.01598->Node36-0.00627->Node47-0.01290->Node75                                                            
#> [175] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node39-0.00719->Node56-0.01455->Node76                                                             
#> [176] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node20-0.18934->Node77                                                                                                             
#> [177] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node44-0.01419->Node51-0.00534->Node78                                                             
#> [178] --0.0000->Node16--1.4850->Node24--0.0464->Node43--0.0280->Node79                                                                                                                                                        
#> [179] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node26-0.08285->Node41-0.02005->Node46-0.00398->Node48-0.00247->Node80                             
#> [180] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node12-0.120110>Node27-0.089763>Node28-0.000151>Node66-0.034443>Node69-0.000928>Node73-0.003053>Node81                             
#> [181] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node20-0.18934->Node77-0.06504->Node82                                                                                             
#> [182] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node30-0.01598->Node36-0.00627->Node47-0.01290->Node75-0.01617->Node83                                            
#> [183] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node6--0.1425->Node14--0.1302->Node23--0.0728->Node84                                                                                                            
#> [184] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node35-0.07317->Node62-0.02385->Node85                                                                             
#> [185] -0.000000>Node2-0.914745>Node3-0.133227>Node4-0.132980>Node6-0.142496>Node8-0.005075>Node12-0.120110>Node27-0.089763>Node28-0.000151>Node66-0.034443>Node69-0.000928>Node73-0.003053>Node81-0.002120>Node86             
#> [186] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node26-0.08285->Node87                                                                             
#> [187] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node6--0.1425->Node14--0.1302->Node22--0.0718->Node34--0.0217->Node37--0.0023->Node88                                                                            
#> [188] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node30-0.01598->Node36-0.00627->Node53-0.01643->Node89                                                            
#> [189] ---0.00-->Node16---1.49-->Node90                                                                                                                                                                                        
#> [190] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node42-0.10506->Node64-0.01342->Node67-0.00122->Node71-0.00192->Node91                             
#> [191] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node15-0.15051->Node21-0.03943->Node92                                                                                             
#> [192] --0.0000->Node2--0.9147->Node3--0.1332->Node4--0.1330->Node6--0.1425->Node14--0.1302->Node23--0.0728->Node84--0.0534->Node93                                                                                            
#> [193] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node30-0.01598->Node36-0.00627->Node47-0.01290->Node75-0.01617->Node83-0.00144->Node94                            
#> [194] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node12-0.12011->Node13-0.00473->Node26-0.08285->Node87-0.04482->Node95                                                             
#> [195] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node5-0.13553->Node7-0.00746->Node20-0.18934->Node77-0.06504->Node96                                                                                             
#> [196] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node8-0.00507->Node10-0.08450->Node32-0.13228->Node44-0.01419->Node51-0.00534->Node78-0.01362->Node97                                             
#> [197] --0.0000->Node16--1.4850->Node24--0.0464->Node43--0.0280->Node79--0.0192->Node98                                                                                                                                        
#> [198] -0.00000->Node2-0.91474->Node3-0.13323->Node4-0.13298->Node6-0.14250->Node14-0.13016->Node23-0.07280->Node84-0.05343->Node93-0.00335->Node99