Return a pfc with descendent's features

Note that in the context of a phylogenetic flow, each element of the flow only has a single descendent, which is the next edge on the sequence between the root and the terminal node of the flow in question.

Usage

pf_desc(x, replace = 0)

Arguments

x

A pfc object

replace

Value to replace edge feature with no descendent values (such as when the edges are the tip edges).

Value

A new pfc with the same structure as x, but with the features of each edge's descendant instead

Examples

pf_desc(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] -0.843201>Node2-0.000156>Node10-0.112268>Node11-0.037133>Node33-0.000000>t42                                                                                                                                             
#>   [2] -0.843201>Node2-0.000156>Node10-0.112268>Node11-0.032710>Node33-0.004423>Node83-0.000000>t85                                                                                                                             
#>   [3] -0.843201>Node2-0.000156>Node10-0.112268>Node11-0.032710>Node33-0.004423>Node83-0.000000>t23                                                                                                                             
#>   [4] -0.843201>Node2-0.000156>Node10-0.116107>Node11-0.033019>Node35-0.000275>Node98-0.000000>t90                                                                                                                             
#>   [5] -0.843201>Node2-0.000156>Node10-0.116107>Node11-0.033019>Node35-0.000275>Node98-0.000000>t87                                                                                                                             
#>   [6] -0.843201>Node2-0.000156>Node10-0.116107>Node11-0.013935>Node35-0.013517>Node45-0.005843>Node74-0.000000>t61                                                                                                             
#>   [7] -0.843201>Node2-0.000156>Node10-0.116107>Node11-0.013935>Node35-0.013517>Node45-0.005843>Node74-0.000000>t14                                                                                                             
#>   [8] -0.843201>Node2-0.000156>Node10-0.116107>Node11-0.013935>Node35-0.013634>Node45-0.005096>Node75-0.000630>Node97-0.000000>t83                                                                                             
#>   [9] -0.843201>Node2-0.000156>Node10-0.116107>Node11-0.013935>Node35-0.013634>Node45-0.005096>Node75-0.000630>Node97-0.000000>t84                                                                                             
#>  [10] -0.843201>Node2-0.000156>Node10-0.116107>Node11-0.013935>Node35-0.013634>Node45-0.005726>Node75-0.000000>t86                                                                                                             
#>  [11] -0.84320->Node2-0.03723->Node10-0.08867->Node15-0.00119->Node39-0.01588->Node41-0.00659->Node71-0.00000->t31                                                                                                             
#>  [12] -0.84320->Node2-0.03723->Node10-0.08867->Node15-0.00119->Node39-0.01588->Node41-0.00659->Node71-0.00000->t59                                                                                                             
#>  [13] -0.84320->Node2-0.03723->Node10-0.08867->Node15-0.00119->Node39-0.02247->Node41-0.00000->t58                                                                                                                             
#>  [14] --0.8432->Node2--0.0372->Node10--0.0887->Node15--0.0193->Node39--0.0044->Node84--0.0000->t99                                                                                                                             
#>  [15] -0.84320->Node2-0.03723->Node10-0.08867->Node15-0.01926->Node39-0.00092->Node84-0.00348->Node86-0.00000->t73                                                                                                             
#>  [16] -0.84320->Node2-0.03723->Node10-0.08867->Node15-0.01926->Node39-0.00092->Node84-0.00348->Node86-0.00000->t68                                                                                                             
#>  [17] -0.84320->Node2-0.03723->Node10-0.04519->Node15-0.05850->Node22-0.00398->Node66-0.00467->Node82-0.00000->t92                                                                                                             
#>  [18] -0.84320->Node2-0.03723->Node10-0.04519->Node15-0.05850->Node22-0.00398->Node66-0.00467->Node82-0.00000->t47                                                                                                             
#>  [19] -0.84320->Node2-0.03723->Node10-0.04519->Node15-0.05850->Node22-0.00865->Node66-0.00000->t75                                                                                                                             
#>  [20] --0.8432->Node2--0.0372->Node10--0.0452->Node15--0.0223->Node22--0.0264->Node30--0.0185->Node46--0.0000->t64                                                                                                             
#>  [21] --0.8432->Node2--0.0372->Node10--0.0452->Node15--0.0223->Node22--0.0264->Node30--0.0185->Node46--0.0000->t5                                                                                                              
#>  [22] --0.8432->Node2--0.0372->Node10--0.0452->Node15--0.0223->Node22--0.0337->Node30--0.0112->Node58--0.0000->t1                                                                                                              
#>  [23] -0.84320->Node2-0.03723->Node10-0.04519->Node15-0.02226->Node22-0.03368->Node30-0.00496->Node58-0.00625->Node73-0.00000->t74                                                                                             
#>  [24] -0.843201>Node2-0.037226>Node10-0.045186>Node15-0.022259>Node22-0.033677>Node30-0.004963>Node58-0.005598>Node73-0.000649>Node96-0.000000>t16                                                                             
#>  [25] -0.843201>Node2-0.037226>Node10-0.045186>Node15-0.022259>Node22-0.033677>Node30-0.004963>Node58-0.005598>Node73-0.000649>Node96-0.000000>t52                                                                             
#>  [26] -0.02832->Node2-0.74086->Node3-0.05481->Node6-0.15414->Node9-0.00947->Node53-0.00516->Node78-0.00000->t4                                                                                                                 
#>  [27] -0.02832->Node2-0.74086->Node3-0.05481->Node6-0.15414->Node9-0.00947->Node53-0.00295->Node78-0.00221->Node89-0.00000->t60                                                                                                
#>  [28] -0.02832->Node2-0.74086->Node3-0.05481->Node6-0.15414->Node9-0.00947->Node53-0.00295->Node78-0.00221->Node89-0.00000->t88                                                                                                
#>  [29] -0.028315>Node2-0.740860>Node3-0.054810>Node6-0.154142>Node9-0.009212>Node53-0.004677>Node77-0.000539>Node95-0.000204>Node99-0.000000>t24                                                                                
#>  [30] -0.028315>Node2-0.740860>Node3-0.054810>Node6-0.154142>Node9-0.009212>Node53-0.004677>Node77-0.000539>Node95-0.000204>Node99-0.000000>t18                                                                                
#>  [31] -0.028315>Node2-0.740860>Node3-0.054810>Node6-0.154142>Node9-0.009212>Node53-0.004677>Node77-0.000743>Node95-0.000000>t98                                                                                                
#>  [32] -0.028315>Node2-0.740860>Node3-0.054810>Node6-0.154142>Node9-0.009212>Node53-0.000261>Node77-0.005159>Node79-0.000000>t17                                                                                                
#>  [33] -0.028315>Node2-0.740860>Node3-0.054810>Node6-0.154142>Node9-0.009212>Node53-0.000261>Node77-0.005159>Node79-0.000000>t50                                                                                                
#>  [34] -0.02832->Node2-0.74086->Node3-0.05481->Node6-0.14473->Node9-0.02273->Node38-0.00132->Node92-0.00000->t37                                                                                                                
#>  [35] -0.02832->Node2-0.74086->Node3-0.05481->Node6-0.14473->Node9-0.02273->Node38-0.00132->Node92-0.00000->t72                                                                                                                
#>  [36] -0.02832->Node2-0.74086->Node3-0.05481->Node6-0.14473->Node9-0.01506->Node38-0.00899->Node63-0.00000->t45                                                                                                                
#>  [37] -0.02832->Node2-0.74086->Node3-0.05481->Node6-0.14473->Node9-0.01506->Node38-0.00899->Node63-0.00000->t38                                                                                                                
#>  [38] -0.02832->Node2-0.74086->Node3-0.13675->Node6-0.05008->Node16-0.02455->Node34-0.00294->Node57-0.00925->Node62-0.00000->t70                                                                                               
#>  [39] -0.02832->Node2-0.74086->Node3-0.13675->Node6-0.05008->Node16-0.02455->Node34-0.00294->Node57-0.00771->Node62-0.00154->Node91-0.00000->t100                                                                              
#>  [40] -0.02832->Node2-0.74086->Node3-0.13675->Node6-0.05008->Node16-0.02455->Node34-0.00294->Node57-0.00771->Node62-0.00154->Node91-0.00000->t95                                                                               
#>  [41] --0.0283->Node2--0.7409->Node3--0.1368->Node6--0.0501->Node16--0.0246->Node34--0.0122->Node57--0.0000->t43                                                                                                               
#>  [42] --0.0283->Node2--0.7409->Node3--0.1368->Node6--0.0501->Node16--0.0367->Node34--0.0000->t2                                                                                                                                
#>  [43] --0.0283->Node2--0.7409->Node3--0.1368->Node6--0.0656->Node16--0.0212->Node42--0.0000->t76                                                                                                                               
#>  [44] -0.028315>Node2-0.740860>Node3-0.136755>Node6-0.065600>Node16-0.000594>Node42-0.020634>Node44-0.000000>t27                                                                                                               
#>  [45] -0.028315>Node2-0.740860>Node3-0.136755>Node6-0.065600>Node16-0.000594>Node42-0.020634>Node44-0.000000>t63                                                                                                               
#>  [46] --0.0283->Node2--0.5476->Node3--0.4168->Node4--0.0000->t28                                                                                                                                                               
#>  [47] --0.0283->Node2--0.5476->Node3--0.1093->Node4--0.1729->Node5--0.1174->Node13--0.0172->Node47--0.0000->t81                                                                                                                
#>  [48] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.11744->Node13-0.01164->Node47-0.00553->Node76-0.00000->t20                                                                                                
#>  [49] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.11744->Node13-0.01164->Node47-0.00553->Node76-0.00000->t41                                                                                                
#>  [50] --0.0283->Node2--0.5476->Node3--0.1093->Node4--0.1729->Node5--0.0139->Node13--0.0501->Node14--0.0707->Node21--0.0000->t29                                                                                                
#>  [51] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.00247->Node28-0.00797->Node29-0.04159->Node32-0.00000->t78                                                
#>  [52] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.00247->Node28-0.00797->Node29-0.04159->Node32-0.00000->t22                                                
#>  [53] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.00247->Node28-0.03263->Node29-0.00980->Node48-0.00713->Node70-0.00000->t8                                 
#>  [54] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.00247->Node28-0.03263->Node29-0.00980->Node48-0.00713->Node70-0.00000->t26                                
#>  [55] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.00247->Node28-0.03263->Node29-0.00173->Node48-0.01521->Node52-0.00000->t49                                
#>  [56] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.00247->Node28-0.03263->Node29-0.00173->Node48-0.00135->Node52-0.00751->Node56-0.00635->Node72-0.00000->t15
#>  [57] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.00247->Node28-0.03263->Node29-0.00173->Node48-0.00135->Node52-0.00751->Node56-0.00635->Node72-0.00000->t19
#>  [58] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.00247->Node28-0.03263->Node29-0.00173->Node48-0.00135->Node52-0.01385->Node56-0.00000->t82                
#>  [59] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.04889->Node28-0.00314->Node88-0.00000->t3                                                                 
#>  [60] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.04889->Node28-0.00314->Node88-0.00000->t32                                                                
#>  [61] --0.0283->Node2--0.5476->Node3--0.1093->Node4--0.1729->Node5--0.0139->Node13--0.1048->Node14--0.0159->Node51--0.0000->t12                                                                                                
#>  [62] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.10482->Node14-0.00717->Node51-0.00472->Node65-0.00401->Node85-0.00000->t91                                                                
#>  [63] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.10482->Node14-0.00717->Node51-0.00472->Node65-0.00401->Node85-0.00000->t33                                                                
#>  [64] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.10482->Node14-0.00717->Node51-0.00675->Node65-0.00198->Node90-0.00000->t56                                                                
#>  [65] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.10482->Node14-0.00717->Node51-0.00675->Node65-0.00198->Node90-0.00000->t7                                                                 
#>  [66] --0.0283->Node2--0.5476->Node3--0.1093->Node4--0.1342->Node5--0.1733->Node7--0.0000->t55                                                                                                                                 
#>  [67] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.00205->Node17-0.05818->Node18-0.01645->Node49-0.00000->t96                                                                  
#>  [68] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.00205->Node17-0.05818->Node18-0.00705->Node49-0.00183->Node61-0.00757->Node68-0.00000->t51                                  
#>  [69] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.00205->Node17-0.05818->Node18-0.00705->Node49-0.00183->Node61-0.00757->Node68-0.00000->t69                                  
#>  [70] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.00205->Node17-0.05818->Node18-0.00705->Node49-0.00940->Node61-0.00000->t77                                                  
#>  [71] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.00205->Node17-0.05154->Node18-0.01815->Node40-0.00495->Node80-0.00000->t35                                                  
#>  [72] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.00205->Node17-0.05154->Node18-0.01815->Node40-0.00495->Node80-0.00000->t6                                                   
#>  [73] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.00205->Node17-0.05154->Node18-0.02310->Node40-0.00000->t9                                                                   
#>  [74] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.03286->Node17-0.01196->Node31-0.03187->Node36-0.00000->t21                                                                  
#>  [75] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.03286->Node17-0.01196->Node31-0.02856->Node36-0.00330->Node87-0.00000->t40                                                  
#>  [76] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.03286->Node17-0.01196->Node31-0.02856->Node36-0.00330->Node87-0.00000->t94                                                  
#>  [77] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.03286->Node17-0.03912->Node31-0.00471->Node81-0.00000->t46                                                                  
#>  [78] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.03286->Node17-0.03912->Node31-0.00471->Node81-0.00000->t79                                                                  
#>  [79] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01433->Node19-0.05050->Node25-0.00805->Node67-0.00000->t71                                                                  
#>  [80] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01433->Node19-0.05050->Node25-0.00805->Node67-0.00000->t57                                                                  
#>  [81] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01433->Node19-0.00364->Node25-0.02862->Node27-0.00509->Node37-0.02120->Node43-0.00000->t34                                  
#>  [82] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01433->Node19-0.00364->Node25-0.02862->Node27-0.00509->Node37-0.02015->Node43-0.00106->Node93-0.00000->t36                  
#>  [83] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01433->Node19-0.00364->Node25-0.02862->Node27-0.00509->Node37-0.02015->Node43-0.00106->Node93-0.00000->t67                  
#>  [84] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01433->Node19-0.00364->Node25-0.02862->Node27-0.02630->Node37-0.00000->t25                                                  
#>  [85] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01433->Node19-0.00364->Node25-0.05492->Node27-0.00000->t13                                                                  
#>  [86] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01468->Node19-0.05819->Node26-0.00000->t39                                                                                  
#>  [87] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01468->Node19-0.04801->Node26-0.01018->Node60-0.00000->t66                                                                  
#>  [88] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01468->Node19-0.04801->Node26-0.01018->Node60-0.00000->t30                                                                  
#>  [89] --0.0684->Node12--0.0116->Node20--0.0594->Node24--0.0000->t10                                                                                                                                                            
#>  [90] -0.068364>Node12-0.011564>Node20-0.044907>Node24-0.003853>Node55-0.003217>Node59-0.006567>Node69-0.000833>Node94-0.000000>t48                                                                                            
#>  [91] -0.068364>Node12-0.011564>Node20-0.044907>Node24-0.003853>Node55-0.003217>Node59-0.006567>Node69-0.000833>Node94-0.000000>t65                                                                                            
#>  [92] -0.06836->Node12-0.01156->Node20-0.04491->Node24-0.00385->Node55-0.00322->Node59-0.00740->Node69-0.00000->t97                                                                                                            
#>  [93] -0.06836->Node12-0.01156->Node20-0.04491->Node24-0.00385->Node55-0.00183->Node59-0.00878->Node64-0.00000->t44                                                                                                            
#>  [94] -0.06836->Node12-0.01156->Node20-0.04491->Node24-0.00385->Node55-0.00183->Node59-0.00878->Node64-0.00000->t54                                                                                                            
#>  [95] --0.0684->Node12--0.0116->Node20--0.0449->Node24--0.0145->Node55--0.0000->t89                                                                                                                                            
#>  [96] --0.0684->Node12--0.0565->Node20--0.0145->Node54--0.0000->t11                                                                                                                                                            
#>  [97] --0.0684->Node12--0.0565->Node20--0.0145->Node54--0.0000->t93                                                                                                                                                            
#>  [98] --0.0722->Node12--0.0507->Node23--0.0164->Node50--0.0000->t80                                                                                                                                                            
#>  [99] --0.0722->Node12--0.0507->Node23--0.0164->Node50--0.0000->t53                                                                                                                                                            
#> [100] --0.0722->Node12--0.0671->Node23--0.0000->t62                                                                                                                                                                            
#> [101] ----0---->Node2                                                                                                                                                                                                          
#> [102] --0.0283->Node2--0.0000->Node3                                                                                                                                                                                           
#> [103] --0.0283->Node2--0.5476->Node3--0.0000->Node4                                                                                                                                                                            
#> [104] --0.0283->Node2--0.5476->Node3--0.1093->Node4--0.0000->Node5                                                                                                                                                             
#> [105] --0.0283->Node2--0.7409->Node3--0.0000->Node6                                                                                                                                                                            
#> [106] --0.0283->Node2--0.5476->Node3--0.1093->Node4--0.1342->Node5--0.0000->Node7                                                                                                                                              
#> [107] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.00000->Node8                                                                                                                               
#> [108] --0.0283->Node2--0.7409->Node3--0.0548->Node6--0.0000->Node9                                                                                                                                                             
#> [109] --0.843-->Node2--0.000-->Node10                                                                                                                                                                                          
#> [110] -0.843201>Node2-0.000156>Node10-0.000000>Node11                                                                                                                                                                          
#> [111] ----0---->Node12                                                                                                                                                                                                         
#> [112] --0.0283->Node2--0.5476->Node3--0.1093->Node4--0.1729->Node5--0.0000->Node13                                                                                                                                             
#> [113] --0.0283->Node2--0.5476->Node3--0.1093->Node4--0.1729->Node5--0.0139->Node13--0.0000->Node14                                                                                                                             
#> [114] --0.8432->Node2--0.0372->Node10--0.0000->Node15                                                                                                                                                                          
#> [115] --0.0283->Node2--0.7409->Node3--0.1368->Node6--0.0000->Node16                                                                                                                                                            
#> [116] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.00000->Node17                                                                                                               
#> [117] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.00205->Node17-0.00000->Node18                                                                                               
#> [118] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.00000->Node19                                                                                                               
#> [119] --0.0684->Node12--0.0000->Node20                                                                                                                                                                                         
#> [120] --0.0283->Node2--0.5476->Node3--0.1093->Node4--0.1729->Node5--0.0139->Node13--0.0501->Node14--0.0000->Node21                                                                                                             
#> [121] --0.8432->Node2--0.0372->Node10--0.0452->Node15--0.0000->Node22                                                                                                                                                          
#> [122] --0.0722->Node12--0.0000->Node23                                                                                                                                                                                         
#> [123] --0.0684->Node12--0.0116->Node20--0.0000->Node24                                                                                                                                                                         
#> [124] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01433->Node19-0.00000->Node25                                                                                               
#> [125] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01468->Node19-0.00000->Node26                                                                                               
#> [126] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01433->Node19-0.00364->Node25-0.00000->Node27                                                                               
#> [127] --0.0283->Node2--0.5476->Node3--0.1093->Node4--0.1729->Node5--0.0139->Node13--0.0501->Node14--0.0186->Node21--0.0000->Node28                                                                                             
#> [128] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.00247->Node28-0.00000->Node29                                                                             
#> [129] --0.8432->Node2--0.0372->Node10--0.0452->Node15--0.0223->Node22--0.0000->Node30                                                                                                                                          
#> [130] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.03286->Node17-0.00000->Node31                                                                                               
#> [131] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.00247->Node28-0.00797->Node29-0.00000->Node32                                                             
#> [132] -0.843201>Node2-0.000156>Node10-0.112268>Node11-0.000000>Node33                                                                                                                                                          
#> [133] --0.0283->Node2--0.7409->Node3--0.1368->Node6--0.0501->Node16--0.0000->Node34                                                                                                                                            
#> [134] -0.843201>Node2-0.000156>Node10-0.116107>Node11-0.000000>Node35                                                                                                                                                          
#> [135] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.03286->Node17-0.01196->Node31-0.00000->Node36                                                                               
#> [136] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01433->Node19-0.00364->Node25-0.02862->Node27-0.00000->Node37                                                               
#> [137] --0.0283->Node2--0.7409->Node3--0.0548->Node6--0.1447->Node9--0.0000->Node38                                                                                                                                             
#> [138] --0.8432->Node2--0.0372->Node10--0.0887->Node15--0.0000->Node39                                                                                                                                                          
#> [139] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.00205->Node17-0.05154->Node18-0.00000->Node40                                                                               
#> [140] -0.84320->Node2-0.03723->Node10-0.08867->Node15-0.00119->Node39-0.00000->Node41                                                                                                                                          
#> [141] --0.0283->Node2--0.7409->Node3--0.1368->Node6--0.0656->Node16--0.0000->Node42                                                                                                                                            
#> [142] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01433->Node19-0.00364->Node25-0.02862->Node27-0.00509->Node37-0.00000->Node43                                               
#> [143] -0.028315>Node2-0.740860>Node3-0.136755>Node6-0.065600>Node16-0.000594>Node42-0.000000>Node44                                                                                                                            
#> [144] -0.843201>Node2-0.000156>Node10-0.116107>Node11-0.013935>Node35-0.000000>Node45                                                                                                                                          
#> [145] --0.8432->Node2--0.0372->Node10--0.0452->Node15--0.0223->Node22--0.0264->Node30--0.0000->Node46                                                                                                                          
#> [146] --0.0283->Node2--0.5476->Node3--0.1093->Node4--0.1729->Node5--0.1174->Node13--0.0000->Node47                                                                                                                             
#> [147] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.00247->Node28-0.03263->Node29-0.00000->Node48                                                             
#> [148] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.00205->Node17-0.05818->Node18-0.00000->Node49                                                                               
#> [149] --0.0722->Node12--0.0507->Node23--0.0000->Node50                                                                                                                                                                         
#> [150] --0.0283->Node2--0.5476->Node3--0.1093->Node4--0.1729->Node5--0.0139->Node13--0.1048->Node14--0.0000->Node51                                                                                                             
#> [151] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.00247->Node28-0.03263->Node29-0.00173->Node48-0.00000->Node52                                             
#> [152] --0.0283->Node2--0.7409->Node3--0.0548->Node6--0.1541->Node9--0.0000->Node53                                                                                                                                             
#> [153] --0.0684->Node12--0.0565->Node20--0.0000->Node54                                                                                                                                                                         
#> [154] --0.0684->Node12--0.0116->Node20--0.0449->Node24--0.0000->Node55                                                                                                                                                         
#> [155] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.00247->Node28-0.03263->Node29-0.00173->Node48-0.00135->Node52-0.00000->Node56                             
#> [156] --0.0283->Node2--0.7409->Node3--0.1368->Node6--0.0501->Node16--0.0246->Node34--0.0000->Node57                                                                                                                            
#> [157] --0.8432->Node2--0.0372->Node10--0.0452->Node15--0.0223->Node22--0.0337->Node30--0.0000->Node58                                                                                                                          
#> [158] -0.06836->Node12-0.01156->Node20-0.04491->Node24-0.00385->Node55-0.00000->Node59                                                                                                                                         
#> [159] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01468->Node19-0.04801->Node26-0.00000->Node60                                                                               
#> [160] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.00205->Node17-0.05818->Node18-0.00705->Node49-0.00000->Node61                                                               
#> [161] -0.02832->Node2-0.74086->Node3-0.13675->Node6-0.05008->Node16-0.02455->Node34-0.00294->Node57-0.00000->Node62                                                                                                            
#> [162] --0.0283->Node2--0.7409->Node3--0.0548->Node6--0.1447->Node9--0.0151->Node38--0.0000->Node63                                                                                                                             
#> [163] -0.06836->Node12-0.01156->Node20-0.04491->Node24-0.00385->Node55-0.00183->Node59-0.00000->Node64                                                                                                                         
#> [164] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.10482->Node14-0.00717->Node51-0.00000->Node65                                                                                             
#> [165] --0.8432->Node2--0.0372->Node10--0.0452->Node15--0.0585->Node22--0.0000->Node66                                                                                                                                          
#> [166] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01433->Node19-0.05050->Node25-0.00000->Node67                                                                               
#> [167] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.00205->Node17-0.05818->Node18-0.00705->Node49-0.00183->Node61-0.00000->Node68                                               
#> [168] -0.06836->Node12-0.01156->Node20-0.04491->Node24-0.00385->Node55-0.00322->Node59-0.00000->Node69                                                                                                                         
#> [169] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.00247->Node28-0.03263->Node29-0.00980->Node48-0.00000->Node70                                             
#> [170] -0.84320->Node2-0.03723->Node10-0.08867->Node15-0.00119->Node39-0.01588->Node41-0.00000->Node71                                                                                                                          
#> [171] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.05006->Node14-0.01864->Node21-0.00247->Node28-0.03263->Node29-0.00173->Node48-0.00135->Node52-0.00751->Node56-0.00000->Node72             
#> [172] -0.84320->Node2-0.03723->Node10-0.04519->Node15-0.02226->Node22-0.03368->Node30-0.00496->Node58-0.00000->Node73                                                                                                          
#> [173] -0.843201>Node2-0.000156>Node10-0.116107>Node11-0.013935>Node35-0.013517>Node45-0.000000>Node74                                                                                                                          
#> [174] -0.843201>Node2-0.000156>Node10-0.116107>Node11-0.013935>Node35-0.013634>Node45-0.000000>Node75                                                                                                                          
#> [175] --0.0283->Node2--0.5476->Node3--0.1093->Node4--0.1729->Node5--0.1174->Node13--0.0116->Node47--0.0000->Node76                                                                                                             
#> [176] -0.02832->Node2-0.74086->Node3-0.05481->Node6-0.15414->Node9-0.00921->Node53-0.00000->Node77                                                                                                                             
#> [177] -0.02832->Node2-0.74086->Node3-0.05481->Node6-0.15414->Node9-0.00947->Node53-0.00000->Node78                                                                                                                             
#> [178] -0.028315>Node2-0.740860>Node3-0.054810>Node6-0.154142>Node9-0.009212>Node53-0.000261>Node77-0.000000>Node79                                                                                                             
#> [179] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.00205->Node17-0.05154->Node18-0.01815->Node40-0.00000->Node80                                                               
#> [180] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.03286->Node17-0.03912->Node31-0.00000->Node81                                                                               
#> [181] -0.84320->Node2-0.03723->Node10-0.04519->Node15-0.05850->Node22-0.00398->Node66-0.00000->Node82                                                                                                                          
#> [182] -0.843201>Node2-0.000156>Node10-0.112268>Node11-0.032710>Node33-0.000000>Node83                                                                                                                                          
#> [183] --0.8432->Node2--0.0372->Node10--0.0887->Node15--0.0193->Node39--0.0000->Node84                                                                                                                                          
#> [184] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.10482->Node14-0.00717->Node51-0.00472->Node65-0.00000->Node85                                                                             
#> [185] -0.84320->Node2-0.03723->Node10-0.08867->Node15-0.01926->Node39-0.00092->Node84-0.00000->Node86                                                                                                                          
#> [186] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09246->Node8-0.03286->Node17-0.01196->Node31-0.02856->Node36-0.00000->Node87                                                               
#> [187] --0.0283->Node2--0.5476->Node3--0.1093->Node4--0.1729->Node5--0.0139->Node13--0.0501->Node14--0.0186->Node21--0.0489->Node28--0.0000->Node88                                                                             
#> [188] -0.02832->Node2-0.74086->Node3-0.05481->Node6-0.15414->Node9-0.00947->Node53-0.00295->Node78-0.00000->Node89                                                                                                             
#> [189] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.17288->Node5-0.01389->Node13-0.10482->Node14-0.00717->Node51-0.00675->Node65-0.00000->Node90                                                                             
#> [190] -0.02832->Node2-0.74086->Node3-0.13675->Node6-0.05008->Node16-0.02455->Node34-0.00294->Node57-0.00771->Node62-0.00000->Node91                                                                                            
#> [191] --0.0283->Node2--0.7409->Node3--0.0548->Node6--0.1447->Node9--0.0227->Node38--0.0000->Node92                                                                                                                             
#> [192] -0.02832->Node2-0.54760->Node3-0.10934->Node4-0.13418->Node5-0.00416->Node7-0.09627->Node8-0.01433->Node19-0.00364->Node25-0.02862->Node27-0.00509->Node37-0.02015->Node43-0.00000->Node93                               
#> [193] -0.06836->Node12-0.01156->Node20-0.04491->Node24-0.00385->Node55-0.00322->Node59-0.00657->Node69-0.00000->Node94                                                                                                         
#> [194] -0.02832->Node2-0.74086->Node3-0.05481->Node6-0.15414->Node9-0.00921->Node53-0.00468->Node77-0.00000->Node95                                                                                                             
#> [195] -0.84320->Node2-0.03723->Node10-0.04519->Node15-0.02226->Node22-0.03368->Node30-0.00496->Node58-0.00560->Node73-0.00000->Node96                                                                                          
#> [196] -0.843201>Node2-0.000156>Node10-0.116107>Node11-0.013935>Node35-0.013634>Node45-0.005096>Node75-0.000000>Node97                                                                                                          
#> [197] -0.843201>Node2-0.000156>Node10-0.116107>Node11-0.033019>Node35-0.000000>Node98                                                                                                                                          
#> [198] -0.028315>Node2-0.740860>Node3-0.054810>Node6-0.154142>Node9-0.009212>Node53-0.004677>Node77-0.000539>Node95-0.000000>Node99