Apply a function along edges within a pfc object

Features of the returned pfc will be the output of the function applied to the original features. Therefore the functionm provided must return a vector the same length of the input.

Usage

pf_edge_apply(x, fun, ...)

Arguments

x

A pfc object

fun

A function to apply to the edge vectors, which will be passed as the first argument. Must return a vector the same length and type as the input.

...

Any additional arguments to pass to fun

Value

A pfc object with the same structure as x but with features replaced with the output of fun

Examples

## average features of descendents of each edge as a pfc:
pf_edge_apply(pf_desc(pf_indexes(rpfc(100))), function(x) rep(mean(x), length(x)))
#> <pfc<e:198>[198]>
#>   [1] ---3.03-->Node9---5.50-->Node20                                                                                                                                                                             
#>   [2] ---3.03-->Node9---5.50-->Node20                                                                                                                                                                             
#>   [3] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--11.50-->Node17--13.50-->Node38                                                                                                                             
#>   [4] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--11.50-->Node17--13.50-->Node38                                                                                                                             
#>   [5] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--11.50-->Node17--15.50-->Node45                                                                                                                             
#>   [6] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--11.50-->Node17--15.50-->Node45                                                                                                                             
#>   [7] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--19.50-->Node19--21.50-->Node68--23.50-->Node98                                                                                             
#>   [8] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--19.50-->Node19--21.50-->Node68--23.50-->Node98                                                                                             
#>   [9] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--19.50-->Node19--21.50-->Node68--25.50-->Node90                                                                                             
#>  [10] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--19.50-->Node19--21.50-->Node68--25.50-->Node90                                                                                             
#>  [11] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--19.50-->Node19--27.17-->Node63                                                                                                             
#>  [12] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--19.50-->Node19--27.17-->Node63--29.75-->Node64--31.50-->Node75                                                                             
#>  [13] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--19.50-->Node19--27.17-->Node63--29.75-->Node64--31.50-->Node75                                                                             
#>  [14] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--19.50-->Node19--27.17-->Node63--29.75-->Node64                                                                                             
#>  [15] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--37.30-->Node42--39.50-->Node72                                                                             
#>  [16] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--37.30-->Node42--39.50-->Node72                                                                             
#>  [17] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--37.30-->Node42--41.17-->Node51                                                                             
#>  [18] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--37.30-->Node42--41.17-->Node51--43.75-->Node61--45.50-->Node85                                             
#>  [19] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--37.30-->Node42--41.17-->Node51--43.75-->Node61--45.50-->Node85                                             
#>  [20] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--37.30-->Node42--41.17-->Node51--43.75-->Node61                                                             
#>  [21] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--51.75-->Node31--53.62-->Node35--55.75-->Node55--57.50-->Node96             
#>  [22] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--51.75-->Node31--53.62-->Node35--55.75-->Node55--57.50-->Node96             
#>  [23] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--51.75-->Node31--53.62-->Node35--55.75-->Node55                             
#>  [24] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--51.75-->Node31--53.62-->Node35--59.50-->Node50                             
#>  [25] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--51.75-->Node31--53.62-->Node35--59.50-->Node50                             
#>  [26] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--51.75-->Node31--61.50-->Node73                                             
#>  [27] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--51.75-->Node31--61.50-->Node73                                             
#>  [28] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--63.93-->Node49--65.58-->Node54--67.50-->Node59--69.50-->Node65             
#>  [29] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--63.93-->Node49--65.58-->Node54--67.50-->Node59--69.50-->Node65             
#>  [30] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--63.93-->Node49--65.58-->Node54--67.50-->Node59--71.50-->Node62             
#>  [31] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--63.93-->Node49--65.58-->Node54--67.50-->Node59--71.50-->Node62             
#>  [32] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--63.93-->Node49--65.58-->Node54--73.25-->Node76                             
#>  [33] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--63.93-->Node49--65.58-->Node54--73.25-->Node76--75.50-->Node95             
#>  [34] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--63.93-->Node49--65.58-->Node54--73.25-->Node76--75.50-->Node95             
#>  [35] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--63.93-->Node49                                                             
#>  [36] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24                                                                                             
#>  [37] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--79.31-->Node39--81.25-->Node52                                                                             
#>  [38] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--79.31-->Node39--81.25-->Node52--83.50-->Node60                                                             
#>  [39] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--79.31-->Node39--81.25-->Node52--83.50-->Node60                                                             
#>  [40] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--79.31-->Node39--85.50-->Node53--87.75-->Node77--89.50-->Node80                                             
#>  [41] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--79.31-->Node39--85.50-->Node53--87.75-->Node77--89.50-->Node80                                             
#>  [42] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--79.31-->Node39--85.50-->Node53--87.75-->Node77                                                             
#>  [43] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--79.31-->Node39--85.50-->Node53--91.25-->Node89                                                             
#>  [44] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--79.31-->Node39--85.50-->Node53--91.25-->Node89--93.50-->Node97                                             
#>  [45] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--79.31-->Node39--85.50-->Node53--91.25-->Node89--93.50-->Node97                                             
#>  [46] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--95.50-->Node71                                                                                             
#>  [47] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--95.50-->Node71                                                                                             
#>  [48] ---3.03-->Node9---7.99-->Node11                                                                                                                                                                             
#>  [49] ---97.1-->Node2---99.2-->Node25                                                                                                                                                                             
#>  [50] ---97.1-->Node2---99.2-->Node25--101.2-->Node56                                                                                                                                                             
#>  [51] ---97.1-->Node2---99.2-->Node25--101.2-->Node56--103.5-->Node78                                                                                                                                             
#>  [52] ---97.1-->Node2---99.2-->Node25--101.2-->Node56--103.5-->Node78                                                                                                                                             
#>  [53] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--109.9-->Node12--111.2-->Node37                                                                                                                               
#>  [54] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--109.9-->Node12--111.2-->Node37--113.8-->Node58--115.5-->Node74                                                                                               
#>  [55] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--109.9-->Node12--111.2-->Node37--113.8-->Node58--115.5-->Node74                                                                                               
#>  [56] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--109.9-->Node12--111.2-->Node37--113.8-->Node58                                                                                                               
#>  [57] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--109.9-->Node12                                                                                                                                               
#>  [58] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--121.4-->Node23--123.2-->Node44                                                                                                
#>  [59] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--121.4-->Node23--123.2-->Node44--125.5-->Node70                                                                                
#>  [60] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--121.4-->Node23--123.2-->Node44--125.5-->Node70                                                                                
#>  [61] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--121.4-->Node23--127.5-->Node57--129.5-->Node66                                                                                
#>  [62] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--121.4-->Node23--127.5-->Node57--129.5-->Node66                                                                                
#>  [63] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--121.4-->Node23--127.5-->Node57--131.5-->Node84                                                                                
#>  [64] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--121.4-->Node23--127.5-->Node57--131.5-->Node84                                                                                
#>  [65] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--133.5-->Node34--135.5-->Node67                                                                                                
#>  [66] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--133.5-->Node34--135.5-->Node67                                                                                                
#>  [67] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--133.5-->Node34--137.5-->Node88                                                                                                
#>  [68] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--133.5-->Node34--137.5-->Node88                                                                                                
#>  [69] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--143.5-->Node93                                                                                                
#>  [70] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--143.5-->Node93                                                                                                
#>  [71] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--147.9-->Node46--149.2-->Node81                                                                
#>  [72] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--147.9-->Node46--149.2-->Node81--151.2-->Node87                                                
#>  [73] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--147.9-->Node46--149.2-->Node81--151.2-->Node87--153.5-->Node92                                
#>  [74] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--147.9-->Node46--149.2-->Node81--151.2-->Node87--153.5-->Node92                                
#>  [75] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--147.9-->Node46                                                                                
#>  [76] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--157.8-->Node48--159.5-->Node79                                                
#>  [77] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--157.8-->Node48--159.5-->Node79                                                
#>  [78] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--157.8-->Node48                                                                
#>  [79] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--161.9-->Node33--163.4-->Node41--165.8-->Node47--167.5-->Node99                
#>  [80] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--161.9-->Node33--163.4-->Node41--165.8-->Node47--167.5-->Node99                
#>  [81] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--161.9-->Node33--163.4-->Node41--165.8-->Node47                                
#>  [82] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--161.9-->Node33--163.4-->Node41--169.2-->Node82                                
#>  [83] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--161.9-->Node33--163.4-->Node41--169.2-->Node82--171.2-->Node86                
#>  [84] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--161.9-->Node33--163.4-->Node41--169.2-->Node82--171.2-->Node86--173.5-->Node91
#>  [85] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--161.9-->Node33--163.4-->Node41--169.2-->Node82--171.2-->Node86--173.5-->Node91
#>  [86] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--161.9-->Node33                                                                
#>  [87] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--175.5-->Node18                                                                                                                
#>  [88] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--175.5-->Node18                                                                                                                
#>  [89] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--179.5-->Node32--181.5-->Node83                                                                                                                               
#>  [90] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--179.5-->Node32--181.5-->Node83                                                                                                                               
#>  [91] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--179.5-->Node32--183.5-->Node94                                                                                                                               
#>  [92] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--179.5-->Node32--183.5-->Node94                                                                                                                               
#>  [93] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--185.4-->Node6--187.8-->Node43--189.5-->Node69                                                                                                                
#>  [94] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--185.4-->Node6--187.8-->Node43--189.5-->Node69                                                                                                                
#>  [95] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--185.4-->Node6--187.8-->Node43                                                                                                                                
#>  [96] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--185.4-->Node6--191.4-->Node7--193.5-->Node30                                                                                                                 
#>  [97] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--185.4-->Node6--191.4-->Node7--193.5-->Node30                                                                                                                 
#>  [98] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--185.4-->Node6--191.4-->Node7--195.8-->Node27--197.5-->Node40                                                                                                 
#>  [99] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--185.4-->Node6--191.4-->Node7--195.8-->Node27--197.5-->Node40                                                                                                 
#> [100] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--185.4-->Node6--191.4-->Node7--195.8-->Node27                                                                                                                 
#> [101] ->                                                                                                                                                                                                          
#> [102] ---97.1-->Node2                                                                                                                                                                                             
#> [103] ---97.1-->Node2--105.8-->Node3                                                                                                                                                                              
#> [104] ---97.1-->Node2--105.8-->Node3                                                                                                                                                                              
#> [105] ---97.1-->Node2--105.8-->Node3--177.3-->Node4                                                                                                                                                               
#> [106] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--185.4-->Node6                                                                                                                                                
#> [107] ---97.1-->Node2--105.8-->Node3--107.1-->Node5                                                                                                                                                               
#> [108] ->                                                                                                                                                                                                          
#> [109] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8                                                                                                                                                
#> [110] ---3.03-->Node9                                                                                                                                                                                             
#> [111] ---97.1-->Node2--105.8-->Node3--107.1-->Node5                                                                                                                                                               
#> [112] ---3.03-->Node9---7.99-->Node11                                                                                                                                                                             
#> [113] ---3.03-->Node9---7.99-->Node11---9.08-->Node13                                                                                                                                                             
#> [114] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8                                                                                                                                                
#> [115] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15                                                                                                                                
#> [116] ---3.03-->Node9---7.99-->Node11---9.08-->Node13                                                                                                                                                             
#> [117] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15                                                                                                                                
#> [118] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14                                                                                                                                             
#> [119] ---3.03-->Node9                                                                                                                                                                                             
#> [120] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14                                                                                                                                             
#> [121] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21                                                                                                                             
#> [122] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10                                                                                                                                
#> [123] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22                                                                                                             
#> [124] ---97.1-->Node2                                                                                                                                                                                             
#> [125] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16                                                                                                                
#> [126] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--185.4-->Node6--191.4-->Node7                                                                                                                                 
#> [127] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24                                                                                             
#> [128] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26                                                                                                
#> [129] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--185.4-->Node6--191.4-->Node7                                                                                                                                 
#> [130] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28                                                                             
#> [131] ---97.1-->Node2--105.8-->Node3--177.3-->Node4                                                                                                                                                               
#> [132] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29                                                                                
#> [133] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10                                                                                                                                
#> [134] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--51.75-->Node31                                                             
#> [135] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21                                                                                                                             
#> [136] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--109.9-->Node12                                                                                                                                               
#> [137] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--11.50-->Node17                                                                                                                                             
#> [138] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36                                                                                                             
#> [139] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--185.4-->Node6--191.4-->Node7--195.8-->Node27                                                                                                                 
#> [140] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--161.9-->Node33                                                                
#> [141] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22                                                                                                             
#> [142] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--185.4-->Node6                                                                                                                                                
#> [143] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--121.4-->Node23                                                                                                                
#> [144] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--11.50-->Node17                                                                                                                                             
#> [145] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26                                                                                                
#> [146] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--161.9-->Node33--163.4-->Node41                                                
#> [147] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29                                                                                
#> [148] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28                                                                             
#> [149] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--51.75-->Node31--53.62-->Node35                                             
#> [150] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--37.30-->Node42                                                                                             
#> [151] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--79.31-->Node39                                                                                             
#> [152] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--79.31-->Node39                                                                                             
#> [153] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--63.93-->Node49                                                             
#> [154] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--51.75-->Node31--53.62-->Node35                                             
#> [155] ---97.1-->Node2---99.2-->Node25                                                                                                                                                                             
#> [156] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--121.4-->Node23                                                                                                                
#> [157] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--109.9-->Node12--111.2-->Node37                                                                                                                               
#> [158] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--63.93-->Node49--65.58-->Node54                                             
#> [159] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--79.31-->Node39--81.25-->Node52                                                                             
#> [160] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--37.30-->Node42--41.17-->Node51                                                                             
#> [161] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--63.93-->Node49--65.58-->Node54--67.50-->Node59                             
#> [162] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--19.50-->Node19                                                                                                                             
#> [163] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--19.50-->Node19--27.17-->Node63                                                                                                             
#> [164] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--63.93-->Node49--65.58-->Node54--67.50-->Node59                             
#> [165] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--121.4-->Node23--127.5-->Node57                                                                                                
#> [166] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--133.5-->Node34                                                                                                                
#> [167] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--19.50-->Node19                                                                                                                             
#> [168] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--185.4-->Node6--187.8-->Node43                                                                                                                                
#> [169] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--121.4-->Node23--123.2-->Node44                                                                                                
#> [170] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36                                                                                                             
#> [171] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--37.30-->Node42                                                                                             
#> [172] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--51.75-->Node31                                                             
#> [173] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--109.9-->Node12--111.2-->Node37--113.8-->Node58                                                                                                               
#> [174] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--19.50-->Node19--27.17-->Node63--29.75-->Node64                                                                                             
#> [175] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--63.93-->Node49--65.58-->Node54                                             
#> [176] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--79.31-->Node39--85.50-->Node53                                                                             
#> [177] ---97.1-->Node2---99.2-->Node25--101.2-->Node56                                                                                                                                                             
#> [178] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--157.8-->Node48                                                                
#> [179] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--79.31-->Node39--85.50-->Node53--87.75-->Node77                                                             
#> [180] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--147.9-->Node46                                                                                
#> [181] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--161.9-->Node33--163.4-->Node41                                                
#> [182] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--179.5-->Node32                                                                                                                                               
#> [183] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--121.4-->Node23--127.5-->Node57                                                                                                
#> [184] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--37.30-->Node42--41.17-->Node51--43.75-->Node61                                                             
#> [185] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--161.9-->Node33--163.4-->Node41--169.2-->Node82                                
#> [186] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--147.9-->Node46--149.2-->Node81                                                                
#> [187] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--119.7-->Node10--133.5-->Node34                                                                                                                
#> [188] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--79.31-->Node39--85.50-->Node53                                                                             
#> [189] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--19.50-->Node19--21.50-->Node68                                                                                                             
#> [190] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--161.9-->Node33--163.4-->Node41--169.2-->Node82--171.2-->Node86                
#> [191] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--147.9-->Node46--149.2-->Node81--151.2-->Node87                                                
#> [192] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16                                                                                                                
#> [193] ---97.1-->Node2--105.8-->Node3--177.3-->Node4--179.5-->Node32                                                                                                                                               
#> [194] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--63.93-->Node49--65.58-->Node54--73.25-->Node76                             
#> [195] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--35.26-->Node22--47.97-->Node24--49.46-->Node28--51.75-->Node31--53.62-->Node35--55.75-->Node55                             
#> [196] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--33.67-->Node21--77.85-->Node36--79.31-->Node39--85.50-->Node53--91.25-->Node89                                                             
#> [197] ---3.03-->Node9---7.99-->Node11---9.08-->Node13--17.19-->Node14--19.50-->Node19--21.50-->Node68                                                                                                             
#> [198] ---97.1-->Node2--105.8-->Node3--107.1-->Node5--117.3-->Node8--139.9-->Node15--141.1-->Node16--145.3-->Node26--155.2-->Node29--161.9-->Node33--163.4-->Node41--165.8-->Node47