Segments the edges of a phylogeny by splitting them at particular positions

Segments the edges of a phylogeny by splitting them at particular positions

Usage

pf_edge_segmentize(x, edges, positions)

Arguments

x

A pfc object

edges

A vector of edge names

positions

A vector of position to cut the edges (must be same length as edges)

Value

A pfc object with edges segmented

Examples

pf_edge_segmentize(rpfc(100), "t1", 0.01)
#> Joining with `by = join_by(edge)`
#> <pfc<e:199>[100]>
#> First  10 phylogenetic flows: 
#> [1] ◎──0.08574──→ Node2_1 ──0.83897──→ Node3_1 ──0.16378──→ Node4_1 ──0.29874──→ Node12_1 ──0.01317──→ Node14_1 ──0.01970──→ Node15_1 ──0.06368──→ Node33_1 ──0.04118──→ Node69_1 ──0.00054──→ Node73_1 ──0.00534──→ Node84_1 ──0.01000──→ t1_1 ──-0.00556──→ t1_2 
#> 
#> [2] ◎──0.0857──→ Node2_1 ──0.8390──→ Node3_1 ──0.2625──→ Node5_1 ──0.0577──→ Node6_1 ──0.2792──→ Node65_1 ──0.0072──→ Node87_1 ──0.0040──→ t10_1 
#> 
#> [3] ◎──0.08574──→ Node2_1 ──0.83897──→ Node3_1 ──0.16378──→ Node4_1 ──0.29874──→ Node12_1 ──0.01317──→ Node14_1 ──0.01970──→ Node15_1 ──0.06368──→ Node33_1 ──0.04118──→ Node69_1 ──0.00054──→ Node73_1 ──0.00534──→ Node84_1 ──0.00444──→ t100_1 
#> 
#> [4] ◎──0.08574──→ Node2_1 ──0.83897──→ Node3_1 ──0.16378──→ Node4_1 ──0.25744──→ Node8_1 ──0.00201──→ Node9_1 ──0.16541──→ Node51_1 ──0.00495──→ Node57_1 ──0.01651──→ Node99_1 ──0.00047──→ t11_1 
#> 
#> [5] ◎──0.086──→ Node2_1 ──0.839──→ Node3_1 ──0.262──→ Node5_1 ──0.058──→ Node6_1 ──0.176──→ Node16_1 ──0.051──→ Node29_1 ──0.047──→ Node59_1 ──0.012──→ Node88_1 ──0.004──→ t12_1 
#> 
#> [6] ◎──0.086──→ Node2_1 ──0.839──→ Node3_1 ──0.262──→ Node5_1 ──0.096──→ Node7_1 ──0.159──→ Node21_1 ──0.027──→ Node28_1 ──0.048──→ Node56_1 ──0.018──→ t13_1 
#> 
#> [7] ◎──0.0857──→ Node2_1 ──1.2709──→ Node10_1 ──0.0097──→ Node11_1 ──0.0640──→ Node17_1 ──0.0696──→ Node42_1 ──0.0253──→ Node71_1 ──0.0101──→ t14_1 
#> 
#> [8] ◎──0.086──→ Node2_1 ──0.839──→ Node3_1 ──0.164──→ Node4_1 ──0.299──→ Node12_1 ──0.013──→ Node14_1 ──0.020──→ Node15_1 ──0.012──→ Node18_1 ──0.024──→ Node24_1 ──0.079──→ t15_1 
#> 
#> [9] ◎──0.08574──→ Node2_1 ──0.83897──→ Node3_1 ──0.16378──→ Node4_1 ──0.25744──→ Node8_1 ──0.05101──→ Node13_1 ──0.03859──→ Node19_1 ──0.01799──→ Node22_1 ──0.00043──→ Node23_1 ──0.03550──→ Node37_1 ──0.03147──→ Node60_1 ──0.01435──→ t16_1 
#> 
#> [10] ◎──0.086──→ Node2_1 ──0.839──→ Node3_1 ──0.164──→ Node4_1 ──0.299──→ Node12_1 ──0.148──→ t17_1