Extract the features of the end edge of each phylogenetic flow

Extract the features of the end edge of each phylogenetic flow

Usage

pf_end_features(x)

Arguments

x

A pfc object

Value

A numeric vector of the features of the end edge of each phylogenetic flow

Examples

pf_end_features(rpfc(100))
#>   [1] 1.594371e-02 1.171957e-02 5.100813e-03 5.100813e-03 1.104208e-02
#>   [6] 8.398816e-03 1.713179e-03 1.713179e-03 4.967871e-03 1.292900e-02
#>  [11] 1.052601e-04 1.052601e-04 9.275165e-02 4.459258e-02 4.459258e-02
#>  [16] 6.924401e-03 6.924401e-03 8.923905e-03 3.453284e-02 1.636072e-02
#>  [21] 1.699601e-03 1.699601e-03 1.714855e-03 2.696078e-02 9.560706e-04
#>  [26] 9.560706e-04 3.351056e-03 3.351056e-03 6.169451e-02 8.060212e-03
#>  [31] 7.848166e-03 7.848166e-03 3.345792e-04 3.345792e-04 1.044071e-01
#>  [36] 2.194803e-02 2.194803e-02 1.909374e-02 1.909374e-02 9.155195e-02
#>  [41] 2.643510e-03 6.787395e-04 6.787395e-04 1.458288e-02 3.540768e-02
#>  [46] 4.337132e-03 2.645176e-03 2.645176e-03 3.809307e-03 5.369838e-03
#>  [51] 5.369838e-03 4.644746e-02 7.975834e-02 1.075666e-03 1.075666e-03
#>  [56] 1.459680e-03 2.736706e-01 1.660649e-03 1.660649e-03 9.390856e-02
#>  [61] 4.420718e-02 4.420718e-02 2.375275e-02 2.375275e-02 1.206622e-03
#>  [66] 1.206622e-03 4.487980e-03 2.662581e-03 2.662581e-03 2.242544e-02
#>  [71] 4.080181e-02 4.080181e-02 3.172385e-02 1.924065e-03 1.924065e-03
#>  [76] 9.525583e-03 6.243480e-03 6.243480e-03 1.498474e-04 1.498474e-04
#>  [81] 2.552721e-02 1.647378e-02 1.647378e-02 2.872056e-03 2.764826e-03
#>  [86] 2.764826e-03 1.592385e-03 1.592385e-03 4.715290e-03 3.253199e-03
#>  [91] 3.253199e-03 2.917546e-02 1.989703e-03 1.989703e-03 5.825419e-02
#>  [96] 4.813664e-03 4.248646e-03 4.248646e-03 7.539312e-03 7.539312e-03
#> [101] 2.411364e+00 2.883734e+00 5.007433e-01 3.669725e-02 6.262831e-03
#> [106] 9.137292e-02 3.466394e-02 1.188823e-01 6.477159e-01 4.243757e-02
#> [111] 8.736671e-02 2.366343e-01 3.478929e-02 1.773681e-01 3.967488e-02
#> [116] 3.538612e-02 1.671727e-03 5.573441e-02 8.731936e-02 3.483221e-01
#> [121] 2.918840e-02 1.351886e-01 1.352423e-02 3.427479e-02 1.685451e-02
#> [126] 9.335172e-03 3.331088e-02 4.815907e-02 1.372163e-02 6.091418e-03
#> [131] 1.911776e-02 5.477848e-02 2.129302e-01 3.648334e-02 1.103978e-02
#> [136] 8.043908e-03 1.046224e-02 9.924256e-02 4.894867e-03 1.561597e-02
#> [141] 1.309816e-02 3.667977e-01 2.654585e-02 3.720833e-02 1.221340e-01
#> [146] 5.407261e-02 7.287663e-03 7.245821e-02 2.721855e-02 9.053428e-03
#> [151] 1.817213e-02 8.641778e-03 2.516552e-01 2.232350e-03 1.165648e-02
#> [156] 1.991789e-03 2.669276e-03 2.500635e-02 2.007717e-01 1.395622e-01
#> [161] 2.643263e-03 1.308592e-02 2.120457e-04 3.226538e-02 1.999505e-03
#> [166] 1.572048e-02 3.282104e-03 5.031496e-03 6.618754e-03 3.430945e-03
#> [171] 3.499103e-02 2.446017e-02 2.216981e-03 6.064202e-03 5.650180e-04
#> [176] 5.278253e-04 1.632206e-02 1.462091e-03 1.444195e-02 1.072304e-04
#> [181] 1.825399e-03 1.164131e-03 1.193937e-02 3.208062e-02 2.979978e-02
#> [186] 1.464586e-02 3.254692e-03 1.525409e-05 2.720100e-01 1.572162e-02
#> [191] 3.982478e-01 5.498338e-03 3.840144e-04 1.871704e-02 1.964771e-03
#> [196] 2.081155e-02 3.847551e-02 1.282374e-02