Calculate the sum of features for each flow

Cumulative sum the phylogenetic flow features for each flow in a pfc, returning a pfc.

Usage

pf_flow_cumsum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A pfc object with cumulative sums of features in place of the original features

Examples

pf_flow_cumsum(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.015-->Node14--2.119-->Node39--2.134-->Node60--2.138-->Node73--2.146-->t29                                                                                                
#>   [2] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.015-->Node14--2.119-->Node39--2.134-->Node60--2.138-->Node73--2.146-->t78                                                                                                
#>   [3] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.015-->Node14--2.119-->Node39--2.134-->Node60--2.146-->t55                                                                                                                
#>   [4] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.015-->Node14--2.119-->Node39--2.146-->t80                                                                                                                                
#>   [5] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.015-->Node14--2.135-->Node63--2.136-->Node65--2.146-->t99                                                                                                                
#>   [6] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.015-->Node14--2.135-->Node63--2.136-->Node65--2.146-->t38                                                                                                                
#>   [7] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.015-->Node14--2.135-->Node63--2.146-->t91                                                                                                                                
#>   [8] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.146-->t86                                                                                                                                                
#>   [9] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.132-->Node55--2.140-->Node81--2.146-->t97                                                                
#>  [10] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.132-->Node55--2.140-->Node81--2.146-->t70                                                                
#>  [11] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.132-->Node55--2.133-->Node59--2.146-->t82                                                                
#>  [12] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.132-->Node55--2.133-->Node59--2.135-->Node64--2.146-->t90                                                
#>  [13] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.132-->Node55--2.133-->Node59--2.135-->Node64--2.137-->Node66--2.137-->Node68--2.146-->t65                
#>  [14] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.132-->Node55--2.133-->Node59--2.135-->Node64--2.137-->Node66--2.137-->Node68--2.144-->Node95--2.146-->t36
#>  [15] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.132-->Node55--2.133-->Node59--2.135-->Node64--2.137-->Node66--2.137-->Node68--2.144-->Node95--2.146-->t77
#>  [16] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.132-->Node55--2.133-->Node59--2.135-->Node64--2.137-->Node66--2.146-->t85                                
#>  [17] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.127-->Node49--2.137-->Node69--2.146-->t19                                                                
#>  [18] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.127-->Node49--2.137-->Node69--2.146-->t59                                                                
#>  [19] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.127-->Node49--2.138-->Node71--2.146-->t27                                                                
#>  [20] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.127-->Node49--2.138-->Node71--2.146-->t42                                                                
#>  [21] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.120-->Node43--2.146-->t48                                                                                                
#>  [22] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.120-->Node43--2.146-->t45                                                                                                
#>  [23] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48--2.134-->Node61--2.137-->Node70--2.138-->Node75--2.140-->Node79--2.146-->t76                                                
#>  [24] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48--2.134-->Node61--2.137-->Node70--2.138-->Node75--2.140-->Node79--2.146-->t92                                                
#>  [25] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48--2.134-->Node61--2.137-->Node70--2.138-->Node75--2.146-->t2                                                                 
#>  [26] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48--2.134-->Node61--2.137-->Node70--2.146-->t62                                                                                
#>  [27] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48--2.134-->Node61--2.141-->Node84--2.146-->t12                                                                                
#>  [28] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48--2.134-->Node61--2.141-->Node84--2.144-->Node97--2.146-->t16                                                                
#>  [29] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48--2.134-->Node61--2.141-->Node84--2.144-->Node97--2.146-->t5                                                                 
#>  [30] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48--2.132-->Node56--2.146-->t79                                                                                                
#>  [31] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48--2.132-->Node56--2.146-->t34                                                                                                
#>  [32] --0.896-->Node2--1.111-->Node3--1.930-->Node6--2.120-->Node42--2.146-->t1                                                                                                                                                                 
#>  [33] --0.896-->Node2--1.111-->Node3--1.930-->Node6--2.120-->Node42--2.129-->Node52--2.132-->Node57--2.146-->t10                                                                                                                                
#>  [34] --0.896-->Node2--1.111-->Node3--1.930-->Node6--2.120-->Node42--2.129-->Node52--2.132-->Node57--2.146-->t93                                                                                                                                
#>  [35] --0.896-->Node2--1.111-->Node3--1.930-->Node6--2.120-->Node42--2.129-->Node52--2.140-->Node80--2.143-->Node91--2.146-->t35                                                                                                                
#>  [36] --0.896-->Node2--1.111-->Node3--1.930-->Node6--2.120-->Node42--2.129-->Node52--2.140-->Node80--2.143-->Node91--2.146-->t24                                                                                                                
#>  [37] --0.896-->Node2--1.111-->Node3--1.930-->Node6--2.120-->Node42--2.129-->Node52--2.140-->Node80--2.146-->t95                                                                                                                                
#>  [38] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.146-->t31                                                                                                                                                                                
#>  [39] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21--2.102-->Node25--2.114-->Node33--2.118-->Node38--2.130-->Node53--2.142-->Node86--2.146-->t72                                                                                
#>  [40] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21--2.102-->Node25--2.114-->Node33--2.118-->Node38--2.130-->Node53--2.142-->Node86--2.146-->t32                                                                                
#>  [41] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21--2.102-->Node25--2.114-->Node33--2.118-->Node38--2.130-->Node53--2.146-->t33                                                                                                
#>  [42] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21--2.102-->Node25--2.114-->Node33--2.118-->Node38--2.146-->t15                                                                                                                
#>  [43] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21--2.102-->Node25--2.114-->Node33--2.118-->Node37--2.129-->Node51--2.146-->t54                                                                                                
#>  [44] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21--2.102-->Node25--2.114-->Node33--2.118-->Node37--2.129-->Node51--2.146-->t75                                                                                                
#>  [45] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21--2.102-->Node25--2.114-->Node33--2.118-->Node37--2.146-->t18                                                                                                                
#>  [46] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21--2.102-->Node25--2.146-->t43                                                                                                                                                
#>  [47] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21--2.146-->t66                                                                                                                                                                
#>  [48] --0.896-->Node2--2.092-->Node23--2.105-->Node27--2.119-->Node40--2.144-->Node92--2.146-->t98                                                                                                                                              
#>  [49] --0.896-->Node2--2.092-->Node23--2.105-->Node27--2.119-->Node40--2.144-->Node92--2.146-->t3                                                                                                                                               
#>  [50] --0.896-->Node2--2.092-->Node23--2.105-->Node27--2.119-->Node40--2.146-->t13                                                                                                                                                              
#>  [51] --0.896-->Node2--2.092-->Node23--2.105-->Node27--2.146-->t74                                                                                                                                                                              
#>  [52] --0.896-->Node2--2.092-->Node23--2.105-->Node28--2.146-->t37                                                                                                                                                                              
#>  [53] --0.896-->Node2--2.092-->Node23--2.105-->Node28--2.135-->Node62--2.143-->Node90--2.146-->t57                                                                                                                                              
#>  [54] --0.896-->Node2--2.092-->Node23--2.105-->Node28--2.135-->Node62--2.143-->Node90--2.146-->t87                                                                                                                                              
#>  [55] --0.896-->Node2--2.092-->Node23--2.105-->Node28--2.135-->Node62--2.146-->t6                                                                                                                                                               
#>  [56] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.14-->Node93---2.15-->t4                                                                                                                                                                
#>  [57] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.14-->Node93---2.15-->t28                                                                                                                                                               
#>  [58] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node35---2.13-->Node50---2.14-->Node76---2.15-->t96                                                                                                               
#>  [59] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node35---2.13-->Node50---2.14-->Node76---2.14-->Node78---2.15-->t68                                                                                               
#>  [60] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node35---2.13-->Node50---2.14-->Node76---2.14-->Node78---2.15-->t94                                                                                               
#>  [61] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node35---2.13-->Node50---2.14-->Node83---2.15-->t14                                                                                                               
#>  [62] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node35---2.13-->Node50---2.14-->Node83---2.15-->t23                                                                                                               
#>  [63] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node35---2.14-->Node74---2.15-->t25                                                                                                                               
#>  [64] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node35---2.14-->Node74---2.15-->t53                                                                                                                               
#>  [65] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node34---2.15-->t73                                                                                                                                               
#>  [66] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node34---2.14-->Node96---2.15-->t71                                                                                                                               
#>  [67] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node34---2.14-->Node96---2.15-->t11                                                                                                                               
#>  [68] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---2.04-->Node16---2.15-->t30                                                                                                                                                                
#>  [69] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---2.04-->Node16---2.09-->Node22---2.10-->Node24---2.15-->t64                                                                                                                                
#>  [70] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---2.04-->Node16---2.09-->Node22---2.10-->Node24---2.13-->Node54---2.15-->t51                                                                                                                
#>  [71] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---2.04-->Node16---2.09-->Node22---2.10-->Node24---2.13-->Node54---2.14-->Node87---2.15-->t83                                                                                                
#>  [72] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---2.04-->Node16---2.09-->Node22---2.10-->Node24---2.13-->Node54---2.14-->Node87---2.15-->t88                                                                                                
#>  [73] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---2.04-->Node16---2.09-->Node22---2.13-->Node58---2.15-->t81                                                                                                                                
#>  [74] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---2.04-->Node16---2.09-->Node22---2.13-->Node58---2.15-->t26                                                                                                                                
#>  [75] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---1.99-->Node11---2.05-->Node18---2.14-->Node85---2.15-->t56                                                                                                                                
#>  [76] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---1.99-->Node11---2.05-->Node18---2.14-->Node85---2.15-->t52                                                                                                                                
#>  [77] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---1.99-->Node11---2.05-->Node18---2.11-->Node32---2.15-->t89                                                                                                                                
#>  [78] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---1.99-->Node11---2.05-->Node18---2.11-->Node32---2.12-->Node46---2.15-->t49                                                                                                                
#>  [79] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---1.99-->Node11---2.05-->Node18---2.11-->Node32---2.12-->Node46---2.14-->Node88---2.15-->t41                                                                                                
#>  [80] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---1.99-->Node11---2.05-->Node18---2.11-->Node32---2.12-->Node46---2.14-->Node88---2.15-->t61                                                                                                
#>  [81] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---1.99-->Node11---2.12-->Node41---2.15-->t84                                                                                                                                                
#>  [82] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---1.99-->Node11---2.12-->Node41---2.15-->t69                                                                                                                                                
#>  [83] ---1.77-->Node4---1.98-->Node9---2.03-->Node15---2.14-->Node67---2.15-->t8                                                                                                                                                                
#>  [84] ---1.77-->Node4---1.98-->Node9---2.03-->Node15---2.14-->Node67---2.15-->t58                                                                                                                                                               
#>  [85] ---1.77-->Node4---1.98-->Node9---2.03-->Node15---2.08-->Node19---2.15-->t17                                                                                                                                                               
#>  [86] ---1.77-->Node4---1.98-->Node9---2.03-->Node15---2.08-->Node19---2.14-->Node77---2.14-->Node89---2.15-->t7                                                                                                                                
#>  [87] ---1.77-->Node4---1.98-->Node9---2.03-->Node15---2.08-->Node19---2.14-->Node77---2.14-->Node89---2.15-->t46                                                                                                                               
#>  [88] ---1.77-->Node4---1.98-->Node9---2.03-->Node15---2.08-->Node19---2.14-->Node77---2.14-->Node82---2.15-->t40                                                                                                                               
#>  [89] ---1.77-->Node4---1.98-->Node9---2.03-->Node15---2.08-->Node19---2.14-->Node77---2.14-->Node82---2.15-->t67                                                                                                                               
#>  [90] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.11-->Node29---2.11-->Node31---2.14-->Node94---2.15-->t60                                                                                                               
#>  [91] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.11-->Node29---2.11-->Node31---2.14-->Node94---2.15-->t100                                                                                                              
#>  [92] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.11-->Node29---2.11-->Node31---2.15-->t44                                                                                                                               
#>  [93] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.11-->Node29---2.14-->Node72---2.15-->t20                                                                                                                               
#>  [94] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.11-->Node29---2.14-->Node72---2.15-->Node99---2.15-->t50                                                                                                               
#>  [95] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.11-->Node29---2.14-->Node72---2.15-->Node99---2.15-->t63                                                                                                               
#>  [96] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.12-->Node44---2.15-->t47                                                                                                                                               
#>  [97] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.12-->Node44---2.12-->Node47---2.15-->t22                                                                                                                               
#>  [98] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.12-->Node44---2.12-->Node47---2.15-->Node98---2.15-->t9                                                                                                                
#>  [99] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.12-->Node44---2.12-->Node47---2.15-->Node98---2.15-->t39                                                                                                               
#> [100] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.15-->t21                                                                                                                                                                               
#> [101] --0.896-->Node2                                                                                                                                                                                                                           
#> [102] --0.896-->Node2--1.111-->Node3                                                                                                                                                                                                            
#> [103] ---1.77-->Node4                                                                                                                                                                                                                           
#> [104] ---1.77-->Node4---1.84-->Node5                                                                                                                                                                                                            
#> [105] --0.896-->Node2--1.111-->Node3--1.930-->Node6                                                                                                                                                                                             
#> [106] --0.896-->Node2--1.111-->Node3--1.944-->Node7                                                                                                                                                                                             
#> [107] ---1.77-->Node4---1.84-->Node5---1.98-->Node8                                                                                                                                                                                             
#> [108] ---1.77-->Node4---1.98-->Node9                                                                                                                                                                                                            
#> [109] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10                                                                                                                                                                             
#> [110] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---1.99-->Node11                                                                                                                                                                             
#> [111] ---1.77-->Node4---1.98-->Node9---1.99-->Node12                                                                                                                                                                                            
#> [112] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13                                                                                                                                                                            
#> [113] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.015-->Node14                                                                                                                                                             
#> [114] ---1.77-->Node4---1.98-->Node9---2.03-->Node15                                                                                                                                                                                            
#> [115] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---2.04-->Node16                                                                                                                                                                             
#> [116] ---1.77-->Node4---1.84-->Node5---2.04-->Node17                                                                                                                                                                                            
#> [117] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---1.99-->Node11---2.05-->Node18                                                                                                                                                             
#> [118] ---1.77-->Node4---1.98-->Node9---2.03-->Node15---2.08-->Node19                                                                                                                                                                            
#> [119] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20                                                                                                                                                             
#> [120] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21                                                                                                                                                                             
#> [121] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---2.04-->Node16---2.09-->Node22                                                                                                                                                             
#> [122] --0.896-->Node2--2.092-->Node23                                                                                                                                                                                                           
#> [123] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---2.04-->Node16---2.09-->Node22---2.10-->Node24                                                                                                                                             
#> [124] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21--2.102-->Node25                                                                                                                                                             
#> [125] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26                                                                                                                                                                            
#> [126] --0.896-->Node2--2.092-->Node23--2.105-->Node27                                                                                                                                                                                           
#> [127] --0.896-->Node2--2.092-->Node23--2.105-->Node28                                                                                                                                                                                           
#> [128] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.11-->Node29                                                                                                                                                            
#> [129] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30                                                                                                                                             
#> [130] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.11-->Node29---2.11-->Node31                                                                                                                                            
#> [131] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---1.99-->Node11---2.05-->Node18---2.11-->Node32                                                                                                                                             
#> [132] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21--2.102-->Node25--2.114-->Node33                                                                                                                                             
#> [133] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node34                                                                                                                                                            
#> [134] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node35                                                                                                                                                            
#> [135] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36                                                                                                                             
#> [136] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21--2.102-->Node25--2.114-->Node33--2.118-->Node37                                                                                                                             
#> [137] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21--2.102-->Node25--2.114-->Node33--2.118-->Node38                                                                                                                             
#> [138] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.015-->Node14--2.119-->Node39                                                                                                                                             
#> [139] --0.896-->Node2--2.092-->Node23--2.105-->Node27--2.119-->Node40                                                                                                                                                                           
#> [140] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---1.99-->Node11---2.12-->Node41                                                                                                                                                             
#> [141] --0.896-->Node2--1.111-->Node3--1.930-->Node6--2.120-->Node42                                                                                                                                                                             
#> [142] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.120-->Node43                                                                                                             
#> [143] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.12-->Node44                                                                                                                                                            
#> [144] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45                                                                                                             
#> [145] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---1.99-->Node11---2.05-->Node18---2.11-->Node32---2.12-->Node46                                                                                                                             
#> [146] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.12-->Node44---2.12-->Node47                                                                                                                                            
#> [147] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48                                                                                                                             
#> [148] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.127-->Node49                                                                                             
#> [149] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node35---2.13-->Node50                                                                                                                                            
#> [150] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21--2.102-->Node25--2.114-->Node33--2.118-->Node37--2.129-->Node51                                                                                                             
#> [151] --0.896-->Node2--1.111-->Node3--1.930-->Node6--2.120-->Node42--2.129-->Node52                                                                                                                                                             
#> [152] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21--2.102-->Node25--2.114-->Node33--2.118-->Node38--2.130-->Node53                                                                                                             
#> [153] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---2.04-->Node16---2.09-->Node22---2.10-->Node24---2.13-->Node54                                                                                                                             
#> [154] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.132-->Node55                                                                                             
#> [155] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48--2.132-->Node56                                                                                                             
#> [156] --0.896-->Node2--1.111-->Node3--1.930-->Node6--2.120-->Node42--2.129-->Node52--2.132-->Node57                                                                                                                                             
#> [157] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---2.04-->Node16---2.09-->Node22---2.13-->Node58                                                                                                                                             
#> [158] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.132-->Node55--2.133-->Node59                                                                             
#> [159] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.015-->Node14--2.119-->Node39--2.134-->Node60                                                                                                                             
#> [160] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48--2.134-->Node61                                                                                                             
#> [161] --0.896-->Node2--2.092-->Node23--2.105-->Node28--2.135-->Node62                                                                                                                                                                           
#> [162] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.015-->Node14--2.135-->Node63                                                                                                                                             
#> [163] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.132-->Node55--2.133-->Node59--2.135-->Node64                                                             
#> [164] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.015-->Node14--2.135-->Node63--2.136-->Node65                                                                                                                             
#> [165] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.132-->Node55--2.133-->Node59--2.135-->Node64--2.137-->Node66                                             
#> [166] ---1.77-->Node4---1.98-->Node9---2.03-->Node15---2.14-->Node67                                                                                                                                                                            
#> [167] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.132-->Node55--2.133-->Node59--2.135-->Node64--2.137-->Node66--2.137-->Node68                             
#> [168] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.127-->Node49--2.137-->Node69                                                                             
#> [169] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48--2.134-->Node61--2.137-->Node70                                                                                             
#> [170] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.127-->Node49--2.138-->Node71                                                                             
#> [171] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.11-->Node29---2.14-->Node72                                                                                                                                            
#> [172] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.015-->Node14--2.119-->Node39--2.134-->Node60--2.138-->Node73                                                                                                             
#> [173] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node35---2.14-->Node74                                                                                                                                            
#> [174] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48--2.134-->Node61--2.137-->Node70--2.138-->Node75                                                                             
#> [175] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node35---2.13-->Node50---2.14-->Node76                                                                                                                            
#> [176] ---1.77-->Node4---1.98-->Node9---2.03-->Node15---2.08-->Node19---2.14-->Node77                                                                                                                                                            
#> [177] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node35---2.13-->Node50---2.14-->Node76---2.14-->Node78                                                                                                            
#> [178] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48--2.134-->Node61--2.137-->Node70--2.138-->Node75--2.140-->Node79                                                             
#> [179] --0.896-->Node2--1.111-->Node3--1.930-->Node6--2.120-->Node42--2.129-->Node52--2.140-->Node80                                                                                                                                             
#> [180] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.132-->Node55--2.140-->Node81                                                                             
#> [181] ---1.77-->Node4---1.98-->Node9---2.03-->Node15---2.08-->Node19---2.14-->Node77---2.14-->Node82                                                                                                                                            
#> [182] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node35---2.13-->Node50---2.14-->Node83                                                                                                                            
#> [183] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48--2.134-->Node61--2.141-->Node84                                                                                             
#> [184] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---1.99-->Node11---2.05-->Node18---2.14-->Node85                                                                                                                                             
#> [185] --0.896-->Node2--1.111-->Node3--1.944-->Node7--2.090-->Node21--2.102-->Node25--2.114-->Node33--2.118-->Node38--2.130-->Node53--2.142-->Node86                                                                                             
#> [186] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---2.04-->Node16---2.09-->Node22---2.10-->Node24---2.13-->Node54---2.14-->Node87                                                                                                             
#> [187] ---1.77-->Node4---1.84-->Node5---1.98-->Node8---1.99-->Node11---2.05-->Node18---2.11-->Node32---2.12-->Node46---2.14-->Node88                                                                                                             
#> [188] ---1.77-->Node4---1.98-->Node9---2.03-->Node15---2.08-->Node19---2.14-->Node77---2.14-->Node89                                                                                                                                            
#> [189] --0.896-->Node2--2.092-->Node23--2.105-->Node28--2.135-->Node62--2.143-->Node90                                                                                                                                                           
#> [190] --0.896-->Node2--1.111-->Node3--1.930-->Node6--2.120-->Node42--2.129-->Node52--2.140-->Node80--2.143-->Node91                                                                                                                             
#> [191] --0.896-->Node2--2.092-->Node23--2.105-->Node27--2.119-->Node40--2.144-->Node92                                                                                                                                                           
#> [192] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.14-->Node93                                                                                                                                                                            
#> [193] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.11-->Node29---2.11-->Node31---2.14-->Node94                                                                                                                            
#> [194] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.117-->Node36--2.123-->Node45--2.132-->Node55--2.133-->Node59--2.135-->Node64--2.137-->Node66--2.137-->Node68--2.144-->Node95             
#> [195] ---1.77-->Node4---1.84-->Node5---2.04-->Node17---2.10-->Node26---2.12-->Node34---2.14-->Node96                                                                                                                                            
#> [196] --0.896-->Node2--1.111-->Node3--1.930-->Node6--1.983-->Node10--2.085-->Node20--2.106-->Node30--2.125-->Node48--2.134-->Node61--2.141-->Node84--2.144-->Node97                                                                             
#> [197] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.12-->Node44---2.12-->Node47---2.15-->Node98                                                                                                                            
#> [198] ---1.77-->Node4---1.98-->Node9---1.99-->Node12---2.00-->Node13---2.11-->Node29---2.14-->Node72---2.15-->Node99