Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>       t49       t82       t59       t39       t90       t40        t6       t67 
#> 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 
#>       t51       t13       t48       t55       t27       t31       t41       t64 
#> 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 
#>       t58       t80       t93       t23        t1       t65       t97       t37 
#> 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 
#>       t74       t30        t4       t12       t71       t26       t14       t16 
#> 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 
#>       t95       t85       t50       t94       t73       t70       t56       t84 
#> 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 
#>       t47       t79       t38       t22       t89       t75       t29       t66 
#> 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 
#>       t76       t42        t9       t20       t19       t72       t32       t57 
#> 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 
#>       t33       t24       t11      t100       t69       t53       t45       t17 
#> 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 
#>       t86       t54       t15        t7       t88       t77       t98       t25 
#> 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 
#>       t99       t46       t52       t44       t96       t92       t28       t18 
#> 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 
#>       t60       t10       t83       t91       t68       t81       t78       t87 
#> 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 
#>       t61        t2       t34        t8       t62       t35       t36        t5 
#> 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 1.3503926 
#>        t3       t63       t43       t21     Node2     Node3     Node4     Node5 
#> 1.3503926 1.3503926 1.3503926 1.3503926 0.2676562 0.7181764 0.8001699 0.8131283 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 0.9987979 1.0436570 1.1243271 1.1657375 1.1709619 1.1986325 1.1999622 1.2049010 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 1.2083296 1.2227293 1.2329336 1.2413811 1.2443703 1.2444792 1.2617729 1.2624966 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 1.2648964 1.2673047 1.2717047 1.2746571 1.2777328 1.2805655 1.2842530 1.2877238 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 1.2879598 1.2968562 1.2994254 1.3013465 1.3024256 1.3070888 1.3078270 1.3105309 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 1.3109013 1.3112466 1.3120241 1.3131541 1.3135113 1.3138610 1.3242216 1.3245273 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 1.3256201 1.3256463 1.3262933 1.3269410 1.3272996 1.3292018 1.3298437 1.3300597 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 1.3316937 1.3336226 1.3336290 1.3341535 1.3343974 1.3344035 1.3348408 1.3356333 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 1.3358111 1.3366541 1.3369647 1.3373426 1.3403095 1.3422417 1.3424414 1.3429073 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 1.3433617 1.3434177 1.3435654 1.3436116 1.3438490 1.3438912 1.3442382 1.3446009 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 1.3446615 1.3448269 1.3449478 1.3450816 1.3450816 1.3451589 1.3455578 1.3455821 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 1.3469377 1.3474506 1.3477959 1.3479838 1.3480172 1.3488656 1.3488869 1.3489483 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 1.3492911 1.3495604 1.3496392 1.3500030 1.3500531 1.3501845