Replace feature with edge index in a pfc

This is similar to base R's col() function, it returns a pfc with features that are just the index of the edge. Useful for complex indexing procedures

Usage

pf_indexes(x)

Arguments

x

A pfc object

Value

A new pfc with the same structure as x, but with the features of each edge equal to the edge index (the column number in the sparse matrix representation)

Examples

pf_indexes(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] ----2---->Node3----4---->Node5----6---->Node49----8---->t86                                                                                                                                                                
#>   [2] ----2---->Node3----4---->Node5----6---->Node49----7---->t7                                                                                                                                                                 
#>   [3] ----2---->Node3----4---->Node5----5---->Node10----10--->t38                                                                                                                                                                
#>   [4] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----14--->Node43----16--->t48                                                                                                                
#>   [5] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----14--->Node43----15--->t23                                                                                                                
#>   [6] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----20--->t35                                                                                                
#>   [7] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40----24--->Node44----26--->Node77----28--->t91                                
#>   [8] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40----24--->Node44----26--->Node77----27--->t60                                
#>   [9] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40----24--->Node44----25--->Node96----30--->t87                                
#>  [10] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40----24--->Node44----25--->Node96----29--->t21                                
#>  [11] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40----23--->Node48----32--->t30                                                
#>  [12] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40----23--->Node48----31--->Node62----34--->t84                                
#>  [13] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40----23--->Node48----31--->Node62----33--->Node74----36--->Node93----38--->t68
#>  [14] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40----23--->Node48----31--->Node62----33--->Node74----36--->Node93----37--->t75
#>  [15] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40----23--->Node48----31--->Node62----33--->Node74----35--->t49                
#>  [16] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----21--->Node83----40--->t57                                                                
#>  [17] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----21--->Node83----39--->t44                                                                
#>  [18] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39----42--->Node59----44--->t64                                                                                
#>  [19] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39----42--->Node59----43--->t66                                                                                
#>  [20] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39----41--->Node52----46--->Node68----48--->t29                                                                
#>  [21] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39----41--->Node52----46--->Node68----47--->Node72----50--->t59                                                
#>  [22] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39----41--->Node52----46--->Node68----47--->Node72----49--->Node80----52--->Node85----54--->t89                
#>  [23] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39----41--->Node52----46--->Node68----47--->Node72----49--->Node80----52--->Node85----53--->t3                 
#>  [24] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39----41--->Node52----46--->Node68----47--->Node72----49--->Node80----51--->t15                                
#>  [25] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39----41--->Node52----45--->Node79----56--->t32                                                                
#>  [26] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39----41--->Node52----45--->Node79----55--->t28                                                                
#>  [27] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22----58--->t61                                                                                                                                
#>  [28] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22----57--->Node25----60--->Node45----62--->Node75----64--->t76                                                                                
#>  [29] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22----57--->Node25----60--->Node45----62--->Node75----63--->t54                                                                                
#>  [30] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22----57--->Node25----60--->Node45----61--->t63                                                                                                
#>  [31] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22----57--->Node25----59--->Node29----66--->Node89----68--->t5                                                                                 
#>  [32] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22----57--->Node25----59--->Node29----66--->Node89----67--->t62                                                                                
#>  [33] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22----57--->Node25----59--->Node29----65--->Node87----70--->Node92----72--->t4                                                                 
#>  [34] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22----57--->Node25----59--->Node29----65--->Node87----70--->Node92----71--->t24                                                                
#>  [35] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22----57--->Node25----59--->Node29----65--->Node87----69--->t53                                                                                
#>  [36] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----80--->Node26----82--->Node66----84--->Node73----86--->t55                                                                                 
#>  [37] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----80--->Node26----82--->Node66----84--->Node73----85--->Node94----88--->t2                                                                  
#>  [38] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----80--->Node26----82--->Node66----84--->Node73----85--->Node94----87--->t73                                                                 
#>  [39] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----80--->Node26----82--->Node66----83--->Node86----90--->t6                                                                                  
#>  [40] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----80--->Node26----82--->Node66----83--->Node86----89--->t20                                                                                 
#>  [41] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----80--->Node26----81--->Node88----92--->t9                                                                                                  
#>  [42] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----80--->Node26----81--->Node88----91--->t93                                                                                                 
#>  [43] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----79--->Node50----94--->Node54----96--->t40                                                                                                 
#>  [44] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----79--->Node50----94--->Node54----95--->t56                                                                                                 
#>  [45] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----79--->Node50----93--->t51                                                                                                                 
#>  [46] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----77--->Node57----98--->t47                                                                                                                                 
#>  [47] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----77--->Node57----97--->t31                                                                                                                                 
#>  [48] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---102--->t12                                                                                                                                 
#>  [49] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---106--->Node32---108--->Node61---110--->t100                                                                
#>  [50] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---106--->Node32---108--->Node61---109--->t80                                                                 
#>  [51] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---106--->Node32---107--->Node33---112--->Node51---114--->Node65---116--->t82                                 
#>  [52] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---106--->Node32---107--->Node33---112--->Node51---114--->Node65---115--->t42                                 
#>  [53] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---106--->Node32---107--->Node33---112--->Node51---113--->t85                                                 
#>  [54] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---106--->Node32---107--->Node33---111--->t37                                                                 
#>  [55] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---105--->Node37---118--->Node78---120--->t67                                                                 
#>  [56] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---105--->Node37---118--->Node78---119--->Node99---122--->t58                                                 
#>  [57] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---105--->Node37---118--->Node78---119--->Node99---121--->t10                                                 
#>  [58] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---105--->Node37---117--->Node46---124--->Node91---126--->t71                                                 
#>  [59] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---105--->Node37---117--->Node46---124--->Node91---125--->t69                                                 
#>  [60] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---105--->Node37---117--->Node46---123--->t25                                                                 
#>  [61] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---103--->t22                                                                                                                 
#>  [62] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11----99--->Node69---128--->t1                                                                                                                                  
#>  [63] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11----99--->Node69---127--->t27                                                                                                                                 
#>  [64] ----2---->Node3----3---->Node7----73--->Node8---130--->Node53---132--->t17                                                                                                                                                 
#>  [65] ----2---->Node3----3---->Node7----73--->Node8---130--->Node53---131--->t50                                                                                                                                                 
#>  [66] ----2---->Node3----3---->Node7----73--->Node8---129--->Node35---134--->t16                                                                                                                                                 
#>  [67] ----2---->Node3----3---->Node7----73--->Node8---129--->Node35---133--->Node38---136--->Node67---138--->t39                                                                                                                 
#>  [68] ----2---->Node3----3---->Node7----73--->Node8---129--->Node35---133--->Node38---136--->Node67---137--->t90                                                                                                                 
#>  [69] ----2---->Node3----3---->Node7----73--->Node8---129--->Node35---133--->Node38---135--->t8                                                                                                                                  
#>  [70] ----1---->Node2---140--->Node6---142--->Node17---144--->Node24---146--->t98                                                                                                                                                
#>  [71] ----1---->Node2---140--->Node6---142--->Node17---144--->Node24---145--->Node27---148--->Node41---150--->t43                                                                                                                
#>  [72] ----1---->Node2---140--->Node6---142--->Node17---144--->Node24---145--->Node27---148--->Node41---149--->t88                                                                                                                
#>  [73] ----1---->Node2---140--->Node6---142--->Node17---144--->Node24---145--->Node27---147--->t11                                                                                                                                
#>  [74] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---152--->Node56---154--->t77                                                                                                                                
#>  [75] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---152--->Node56---153--->Node70---156--->t95                                                                                                                
#>  [76] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---152--->Node56---153--->Node70---155--->Node71---158--->t45                                                                                                
#>  [77] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---152--->Node56---153--->Node70---155--->Node71---157--->t14                                                                                                
#>  [78] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---151--->Node47---160--->Node55---162--->Node60---164--->t52                                                                                                
#>  [79] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---151--->Node47---160--->Node55---162--->Node60---163--->t83                                                                                                
#>  [80] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---151--->Node47---160--->Node55---161--->t72                                                                                                                
#>  [81] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---151--->Node47---159--->Node58---166--->Node81---168--->t36                                                                                                
#>  [82] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---151--->Node47---159--->Node58---166--->Node81---167--->Node98---170--->t13                                                                                
#>  [83] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---151--->Node47---159--->Node58---166--->Node81---167--->Node98---169--->t70                                                                                
#>  [84] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---151--->Node47---159--->Node58---165--->Node84---172--->t78                                                                                                
#>  [85] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---151--->Node47---159--->Node58---165--->Node84---171--->t34                                                                                                
#>  [86] ----1---->Node2---140--->Node6---141--->Node76---174--->t41                                                                                                                                                                
#>  [87] ----1---->Node2---140--->Node6---141--->Node76---173--->Node82---176--->t97                                                                                                                                                
#>  [88] ----1---->Node2---140--->Node6---141--->Node76---173--->Node82---175--->t74                                                                                                                                                
#>  [89] ----1---->Node2---139--->Node4---178--->Node95---180--->t96                                                                                                                                                                
#>  [90] ----1---->Node2---139--->Node4---178--->Node95---179--->t99                                                                                                                                                                
#>  [91] ----1---->Node2---139--->Node4---177--->Node15---182--->Node20---184--->t65                                                                                                                                                
#>  [92] ----1---->Node2---139--->Node4---177--->Node15---182--->Node20---183--->Node21---186--->t33                                                                                                                                
#>  [93] ----1---->Node2---139--->Node4---177--->Node15---182--->Node20---183--->Node21---185--->t26                                                                                                                                
#>  [94] ----1---->Node2---139--->Node4---177--->Node15---181--->Node23---188--->Node63---190--->Node64---192--->t81                                                                                                                
#>  [95] ----1---->Node2---139--->Node4---177--->Node15---181--->Node23---188--->Node63---190--->Node64---191--->t94                                                                                                                
#>  [96] ----1---->Node2---139--->Node4---177--->Node15---181--->Node23---188--->Node63---189--->t79                                                                                                                                
#>  [97] ----1---->Node2---139--->Node4---177--->Node15---181--->Node23---187--->Node42---194--->Node90---196--->Node97---198--->t19                                                                                                
#>  [98] ----1---->Node2---139--->Node4---177--->Node15---181--->Node23---187--->Node42---194--->Node90---196--->Node97---197--->t46                                                                                                
#>  [99] ----1---->Node2---139--->Node4---177--->Node15---181--->Node23---187--->Node42---194--->Node90---195--->t18                                                                                                                
#> [100] ----1---->Node2---139--->Node4---177--->Node15---181--->Node23---187--->Node42---193--->t92                                                                                                                                
#> [101] ----1---->Node2                                                                                                                                                                                                            
#> [102] ----2---->Node3                                                                                                                                                                                                            
#> [103] ----1---->Node2---139--->Node4                                                                                                                                                                                             
#> [104] ----2---->Node3----4---->Node5                                                                                                                                                                                             
#> [105] ----1---->Node2---140--->Node6                                                                                                                                                                                             
#> [106] ----2---->Node3----3---->Node7                                                                                                                                                                                             
#> [107] ----2---->Node3----3---->Node7----73--->Node8                                                                                                                                                                              
#> [108] ----2---->Node3----3---->Node7----74--->Node9                                                                                                                                                                              
#> [109] ----2---->Node3----4---->Node5----5---->Node10                                                                                                                                                                             
#> [110] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11                                                                                                                                                              
#> [111] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12                                                                                                                                                             
#> [112] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13                                                                                                                                                              
#> [113] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14                                                                                                                                             
#> [114] ----1---->Node2---139--->Node4---177--->Node15                                                                                                                                                                             
#> [115] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16                                                                                                                                              
#> [116] ----1---->Node2---140--->Node6---142--->Node17                                                                                                                                                                             
#> [117] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18                                                                                                                                              
#> [118] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19                                                                                                                             
#> [119] ----1---->Node2---139--->Node4---177--->Node15---182--->Node20                                                                                                                                                             
#> [120] ----1---->Node2---139--->Node4---177--->Node15---182--->Node20---183--->Node21                                                                                                                                             
#> [121] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22                                                                                                                                             
#> [122] ----1---->Node2---139--->Node4---177--->Node15---181--->Node23                                                                                                                                                             
#> [123] ----1---->Node2---140--->Node6---142--->Node17---144--->Node24                                                                                                                                                             
#> [124] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22----57--->Node25                                                                                                                             
#> [125] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----80--->Node26                                                                                                                              
#> [126] ----1---->Node2---140--->Node6---142--->Node17---144--->Node24---145--->Node27                                                                                                                                             
#> [127] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28                                                                                                                              
#> [128] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22----57--->Node25----59--->Node29                                                                                                             
#> [129] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30                                                                                                             
#> [130] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31                                                                                                              
#> [131] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---106--->Node32                                                                                              
#> [132] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---106--->Node32---107--->Node33                                                                              
#> [133] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34                                                                                             
#> [134] ----2---->Node3----3---->Node7----73--->Node8---129--->Node35                                                                                                                                                              
#> [135] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36                                                                                                                                                             
#> [136] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---105--->Node37                                                                                              
#> [137] ----2---->Node3----3---->Node7----73--->Node8---129--->Node35---133--->Node38                                                                                                                                              
#> [138] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39                                                                                                             
#> [139] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40                                                                             
#> [140] ----1---->Node2---140--->Node6---142--->Node17---144--->Node24---145--->Node27---148--->Node41                                                                                                                             
#> [141] ----1---->Node2---139--->Node4---177--->Node15---181--->Node23---187--->Node42                                                                                                                                             
#> [142] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----14--->Node43                                                                                                                             
#> [143] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40----24--->Node44                                                             
#> [144] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22----57--->Node25----60--->Node45                                                                                                             
#> [145] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---105--->Node37---117--->Node46                                                                              
#> [146] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---151--->Node47                                                                                                                                             
#> [147] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40----23--->Node48                                                             
#> [148] ----2---->Node3----4---->Node5----6---->Node49                                                                                                                                                                             
#> [149] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----79--->Node50                                                                                                                              
#> [150] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---106--->Node32---107--->Node33---112--->Node51                                                              
#> [151] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39----41--->Node52                                                                                             
#> [152] ----2---->Node3----3---->Node7----73--->Node8---130--->Node53                                                                                                                                                              
#> [153] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----79--->Node50----94--->Node54                                                                                                              
#> [154] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---151--->Node47---160--->Node55                                                                                                                             
#> [155] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---152--->Node56                                                                                                                                             
#> [156] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----77--->Node57                                                                                                                                              
#> [157] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---151--->Node47---159--->Node58                                                                                                                             
#> [158] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39----42--->Node59                                                                                             
#> [159] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---151--->Node47---160--->Node55---162--->Node60                                                                                                             
#> [160] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---106--->Node32---108--->Node61                                                                              
#> [161] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40----23--->Node48----31--->Node62                                             
#> [162] ----1---->Node2---139--->Node4---177--->Node15---181--->Node23---188--->Node63                                                                                                                                             
#> [163] ----1---->Node2---139--->Node4---177--->Node15---181--->Node23---188--->Node63---190--->Node64                                                                                                                             
#> [164] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---106--->Node32---107--->Node33---112--->Node51---114--->Node65                                              
#> [165] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----80--->Node26----82--->Node66                                                                                                              
#> [166] ----2---->Node3----3---->Node7----73--->Node8---129--->Node35---133--->Node38---136--->Node67                                                                                                                              
#> [167] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39----41--->Node52----46--->Node68                                                                             
#> [168] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11----99--->Node69                                                                                                                                              
#> [169] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---152--->Node56---153--->Node70                                                                                                                             
#> [170] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---152--->Node56---153--->Node70---155--->Node71                                                                                                             
#> [171] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39----41--->Node52----46--->Node68----47--->Node72                                                             
#> [172] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----80--->Node26----82--->Node66----84--->Node73                                                                                              
#> [173] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40----23--->Node48----31--->Node62----33--->Node74                             
#> [174] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22----57--->Node25----60--->Node45----62--->Node75                                                                                             
#> [175] ----1---->Node2---140--->Node6---141--->Node76                                                                                                                                                                             
#> [176] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40----24--->Node44----26--->Node77                                             
#> [177] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---105--->Node37---118--->Node78                                                                              
#> [178] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39----41--->Node52----45--->Node79                                                                             
#> [179] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39----41--->Node52----46--->Node68----47--->Node72----49--->Node80                                             
#> [180] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---151--->Node47---159--->Node58---166--->Node81                                                                                                             
#> [181] ----1---->Node2---140--->Node6---141--->Node76---173--->Node82                                                                                                                                                             
#> [182] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----21--->Node83                                                                             
#> [183] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---151--->Node47---159--->Node58---165--->Node84                                                                                                             
#> [184] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----17--->Node39----41--->Node52----46--->Node68----47--->Node72----49--->Node80----52--->Node85                             
#> [185] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----80--->Node26----82--->Node66----83--->Node86                                                                                              
#> [186] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22----57--->Node25----59--->Node29----65--->Node87                                                                                             
#> [187] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----80--->Node26----81--->Node88                                                                                                              
#> [188] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22----57--->Node25----59--->Node29----66--->Node89                                                                                             
#> [189] ----1---->Node2---139--->Node4---177--->Node15---181--->Node23---187--->Node42---194--->Node90                                                                                                                             
#> [190] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---105--->Node37---117--->Node46---124--->Node91                                                              
#> [191] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----11--->Node22----57--->Node25----59--->Node29----65--->Node87----70--->Node92                                                                             
#> [192] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40----23--->Node48----31--->Node62----33--->Node74----36--->Node93             
#> [193] ----2---->Node3----3---->Node7----74--->Node9----76--->Node13----78--->Node16----80--->Node26----82--->Node66----84--->Node73----85--->Node94                                                                              
#> [194] ----1---->Node2---139--->Node4---178--->Node95                                                                                                                                                                             
#> [195] ----2---->Node3----4---->Node5----5---->Node10----9---->Node12----12--->Node14----13--->Node19----18--->Node30----19--->Node34----22--->Node40----24--->Node44----25--->Node96                                             
#> [196] ----1---->Node2---139--->Node4---177--->Node15---181--->Node23---187--->Node42---194--->Node90---196--->Node97                                                                                                             
#> [197] ----1---->Node2---140--->Node6---142--->Node17---143--->Node36---151--->Node47---159--->Node58---166--->Node81---167--->Node98                                                                                             
#> [198] ----2---->Node3----3---->Node7----74--->Node9----75--->Node11---100--->Node18---101--->Node28---104--->Node31---105--->Node37---118--->Node78---119--->Node99