Calculate a kronecker product when a pfc is the multiplicand or the multiplier.

Calculate a kronecker product when a pfc is the multiplicand or the multiplier.

Usage

pf_kronecker(x, y, ...)

Arguments

x

The kronecker multiplicand.

y

The kronecker multiplier.

...

Other arguments passed to or from other methods.

Value

A pfc object

Examples

pf_kronecker(rpfc(20), rpfc(20))
#> <pfc<e:1444>[1444]>
#> First  10 phylogenetic flows: 
#> [1] ◎──2.6e+00──→ Node3:Node6 ──2.5e-02──→ Node3:Node9 ──9.4e-02──→ Node3:Node13 ──2.2e-05──→ Node3:Node14 ──8.8e-02──→ Node3:t7 ──5.4e-01──→ Node6:Node6 ──5.2e-03──→ Node6:Node9 ──1.9e-02──→ Node6:Node13 ──4.5e-06──→ Node6:Node14 ──1.8e-02──→ Node6:t7 ──7.8e-02──→ Node7:Node6 ──7.5e-04──→ Node7:Node9 ──2.8e-03──→ Node7:Node13 ──6.5e-07──→ Node7:Node14 ──2.6e-03──→ Node7:t7 ──2.5e-01──→ t11:Node6 ──2.4e-03──→ t11:Node9 ──8.9e-03──→ t11:Node13 ──2.1e-06──→ t11:Node14 ──8.3e-03──→ t11:t7 
#> 
#> [2] ◎──2.6e+00──→ Node3:Node6 ──2.5e-02──→ Node3:Node9 ──9.4e-02──→ Node3:Node13 ──2.2e-05──→ Node3:Node14 ──8.8e-02──→ Node3:t20 ──5.4e-01──→ Node6:Node6 ──5.2e-03──→ Node6:Node9 ──1.9e-02──→ Node6:Node13 ──4.5e-06──→ Node6:Node14 ──1.8e-02──→ Node6:t20 ──7.8e-02──→ Node7:Node6 ──7.5e-04──→ Node7:Node9 ──2.8e-03──→ Node7:Node13 ──6.5e-07──→ Node7:Node14 ──2.6e-03──→ Node7:t20 ──2.5e-01──→ t11:Node6 ──2.4e-03──→ t11:Node9 ──8.9e-03──→ t11:Node13 ──2.1e-06──→ t11:Node14 ──8.3e-03──→ t11:t20 
#> 
#> [3] ◎──2.64285──→ Node3:Node6 ──0.02549──→ Node3:Node9 ──0.09415──→ Node3:Node13 ──0.08783──→ Node3:t11 ──0.54061──→ Node6:Node6 ──0.00521──→ Node6:Node9 ──0.01926──→ Node6:Node13 ──0.01797──→ Node6:t11 ──0.07804──→ Node7:Node6 ──0.00075──→ Node7:Node9 ──0.00278──→ Node7:Node13 ──0.00259──→ Node7:t11 ──0.24895──→ t11:Node6 ──0.00240──→ t11:Node9 ──0.00887──→ t11:Node13 ──0.00827──→ t11:t11 
#> 
#> [4] ◎──2.64285──→ Node3:Node6 ──0.02549──→ Node3:Node9 ──0.13134──→ Node3:Node16 ──0.05064──→ Node3:t2 ──0.54061──→ Node6:Node6 ──0.00521──→ Node6:Node9 ──0.02687──→ Node6:Node16 ──0.01036──→ Node6:t2 ──0.07804──→ Node7:Node6 ──0.00075──→ Node7:Node9 ──0.00388──→ Node7:Node16 ──0.00150──→ Node7:t2 ──0.24895──→ t11:Node6 ──0.00240──→ t11:Node9 ──0.01237──→ t11:Node16 ──0.00477──→ t11:t2 
#> 
#> [5] ◎──2.64285──→ Node3:Node6 ──0.02549──→ Node3:Node9 ──0.13134──→ Node3:Node16 ──0.05064──→ Node3:t10 ──0.54061──→ Node6:Node6 ──0.00521──→ Node6:Node9 ──0.02687──→ Node6:Node16 ──0.01036──→ Node6:t10 ──0.07804──→ Node7:Node6 ──0.00075──→ Node7:Node9 ──0.00388──→ Node7:Node16 ──0.00150──→ Node7:t10 ──0.24895──→ t11:Node6 ──0.00240──→ t11:Node9 ──0.01237──→ t11:Node16 ──0.00477──→ t11:t10 
#> 
#> [6] ◎──2.6428──→ Node3:Node6 ──0.1392──→ Node3:Node15 ──0.0683──→ Node3:t5 ──0.5406──→ Node6:Node6 ──0.0285──→ Node6:Node15 ──0.0140──→ Node6:t5 ──0.0780──→ Node7:Node6 ──0.0041──→ Node7:Node15 ──0.0020──→ Node7:t5 ──0.2489──→ t11:Node6 ──0.0131──→ t11:Node15 ──0.0064──→ t11:t5 
#> 
#> [7] ◎──2.6428──→ Node3:Node6 ──0.1392──→ Node3:Node15 ──0.0683──→ Node3:t6 ──0.5406──→ Node6:Node6 ──0.0285──→ Node6:Node15 ──0.0140──→ Node6:t6 ──0.0780──→ Node7:Node6 ──0.0041──→ Node7:Node15 ──0.0020──→ Node7:t6 ──0.2489──→ t11:Node6 ──0.0131──→ t11:Node15 ──0.0064──→ t11:t6 
#> 
#> [8] ◎──1.4520──→ Node3:Node2 ──1.2367──→ Node3:Node11 ──0.1616──→ Node3:t17 ──0.2970──→ Node6:Node2 ──0.2530──→ Node6:Node11 ──0.0331──→ Node6:t17 ──0.0429──→ Node7:Node2 ──0.0365──→ Node7:Node11 ──0.0048──→ Node7:t17 ──0.1368──→ t11:Node2 ──0.1165──→ t11:Node11 ──0.0152──→ t11:t17 
#> 
#> [9] ◎──1.4520──→ Node3:Node2 ──1.2367──→ Node3:Node11 ──0.1616──→ Node3:t9 ──0.2970──→ Node6:Node2 ──0.2530──→ Node6:Node11 ──0.0331──→ Node6:t9 ──0.0429──→ Node7:Node2 ──0.0365──→ Node7:Node11 ──0.0048──→ Node7:t9 ──0.1368──→ t11:Node2 ──0.1165──→ t11:Node11 ──0.0152──→ t11:t9 
#> 
#> [10] ◎──1.4520──→ Node3:Node2 ──0.3414──→ Node3:Node3 ──0.4087──→ Node3:Node4 ──0.5314──→ Node3:Node12 ──0.0750──→ Node3:Node17 ──0.0418──→ Node3:t4 ──0.2970──→ Node6:Node2 ──0.0698──→ Node6:Node3 ──0.0836──→ Node6:Node4 ──0.1087──→ Node6:Node12 ──0.0154──→ Node6:Node17 ──0.0086──→ Node6:t4 ──0.0429──→ Node7:Node2 ──0.0101──→ Node7:Node3 ──0.0121──→ Node7:Node4 ──0.0157──→ Node7:Node12 ──0.0022──→ Node7:Node17 ──0.0012──→ Node7:t4 ──0.1368──→ t11:Node2 ──0.0322──→ t11:Node3 ──0.0385──→ t11:Node4 ──0.0501──→ t11:Node12 ──0.0071──→ t11:Node17 ──0.0039──→ t11:t4