Return a vector of labels for a pfc object

Return a vector of labels for a pfc object

Usage

pf_labels(x)

Arguments

x

A pfc object

Value

A character vector of labels

Examples

pf_labels(rpfc(100))
#>   [1] "t46"    "t15"    "t28"    "t20"    "t96"    "t55"    "t14"    "t83"   
#>   [9] "t97"    "t18"    "t67"    "t86"    "t34"    "t5"     "t29"    "t21"   
#>  [17] "t26"    "t30"    "t1"     "t7"     "t100"   "t56"    "t77"    "t32"   
#>  [25] "t99"    "t72"    "t54"    "t17"    "t38"    "t95"    "t80"    "t22"   
#>  [33] "t6"     "t90"    "t31"    "t73"    "t85"    "t75"    "t70"    "t19"   
#>  [41] "t12"    "t16"    "t91"    "t23"    "t51"    "t63"    "t47"    "t66"   
#>  [49] "t40"    "t74"    "t87"    "t49"    "t42"    "t36"    "t61"    "t98"   
#>  [57] "t82"    "t78"    "t89"    "t81"    "t43"    "t79"    "t44"    "t9"    
#>  [65] "t76"    "t68"    "t35"    "t58"    "t11"    "t92"    "t69"    "t4"    
#>  [73] "t88"    "t41"    "t59"    "t48"    "t27"    "t65"    "t62"    "t50"   
#>  [81] "t71"    "t52"    "t84"    "t64"    "t60"    "t57"    "t10"    "t13"   
#>  [89] "t37"    "t24"    "t8"     "t33"    "t2"     "t45"    "t94"    "t53"   
#>  [97] "t3"     "t93"    "t39"    "t25"    "Node2"  "Node3"  "Node4"  "Node5" 
#> [105] "Node6"  "Node7"  "Node8"  "Node9"  "Node10" "Node11" "Node12" "Node13"
#> [113] "Node14" "Node15" "Node16" "Node17" "Node18" "Node19" "Node20" "Node21"
#> [121] "Node22" "Node23" "Node24" "Node25" "Node26" "Node27" "Node28" "Node29"
#> [129] "Node30" "Node31" "Node32" "Node33" "Node34" "Node35" "Node36" "Node37"
#> [137] "Node38" "Node39" "Node40" "Node41" "Node42" "Node43" "Node44" "Node45"
#> [145] "Node46" "Node47" "Node48" "Node49" "Node50" "Node51" "Node52" "Node53"
#> [153] "Node54" "Node55" "Node56" "Node57" "Node58" "Node59" "Node60" "Node61"
#> [161] "Node62" "Node63" "Node64" "Node65" "Node66" "Node67" "Node68" "Node69"
#> [169] "Node70" "Node71" "Node72" "Node73" "Node74" "Node75" "Node76" "Node77"
#> [177] "Node78" "Node79" "Node80" "Node81" "Node82" "Node83" "Node84" "Node85"
#> [185] "Node86" "Node87" "Node88" "Node89" "Node90" "Node91" "Node92" "Node93"
#> [193] "Node94" "Node95" "Node96" "Node97" "Node98" "Node99"