Extract pfp object from pfc, the paths of each flow from root to terminal node.

Extract pfp object from pfc, the paths of each flow from root to terminal node.

Usage

pf_path(x)

Arguments

x

A pfc object

Value

A pfp object

Examples

pf_path(rpfc(100))
#> <pfp[198]>
#>   [1] 2->4->6->8                                                               
#>   [2] 2->4->6->7->10->12                                                       
#>   [3] 2->4->6->7->10->11                                                       
#>   [4] 2->4->6->7->9                                                            
#>   [5] 2->4->5                                                                  
#>   [6] 2->3->14->16->18                                                         
#>   [7] 2->3->14->16->17->20                                                     
#>   [8] 2->3->14->16->17->19->22                                                 
#>   [9] 2->3->14->16->17->19->21->24                                             
#>  [10] 2->3->14->16->17->19->21->23                                             
#>  [11] 2->3->14->15->26->28->30                                                 
#>  [12] 2->3->14->15->26->28->29                                                 
#>  [13] 2->3->14->15->26->27                                                     
#>  [14] 2->3->14->15->25                                                         
#>  [15] 2->3->13->32->34->36->38                                                 
#>  [16] 2->3->13->32->34->36->37                                                 
#>  [17] 2->3->13->32->34->35                                                     
#>  [18] 2->3->13->32->33->40                                                     
#>  [19] 2->3->13->32->33->39                                                     
#>  [20] 2->3->13->31->42                                                         
#>  [21] 2->3->13->31->41                                                         
#>  [22] 1->44->46->48->50                                                        
#>  [23] 1->44->46->48->49->52                                                    
#>  [24] 1->44->46->48->49->51                                                    
#>  [25] 1->44->46->47->54->56                                                    
#>  [26] 1->44->46->47->54->55->58                                                
#>  [27] 1->44->46->47->54->55->57                                                
#>  [28] 1->44->46->47->53->60->62                                                
#>  [29] 1->44->46->47->53->60->61->64->66->68                                    
#>  [30] 1->44->46->47->53->60->61->64->66->67                                    
#>  [31] 1->44->46->47->53->60->61->64->65->70->72                                
#>  [32] 1->44->46->47->53->60->61->64->65->70->71->74                            
#>  [33] 1->44->46->47->53->60->61->64->65->70->71->73                            
#>  [34] 1->44->46->47->53->60->61->64->65->69->76                                
#>  [35] 1->44->46->47->53->60->61->64->65->69->75                                
#>  [36] 1->44->46->47->53->60->61->63                                            
#>  [37] 1->44->46->47->53->59->78                                                
#>  [38] 1->44->46->47->53->59->77->80->82->84->86                                
#>  [39] 1->44->46->47->53->59->77->80->82->84->85->88                            
#>  [40] 1->44->46->47->53->59->77->80->82->84->85->87                            
#>  [41] 1->44->46->47->53->59->77->80->82->83                                    
#>  [42] 1->44->46->47->53->59->77->80->81->90                                    
#>  [43] 1->44->46->47->53->59->77->80->81->89                                    
#>  [44] 1->44->46->47->53->59->77->79->92->94                                    
#>  [45] 1->44->46->47->53->59->77->79->92->93->96                                
#>  [46] 1->44->46->47->53->59->77->79->92->93->95->98                            
#>  [47] 1->44->46->47->53->59->77->79->92->93->95->97                            
#>  [48] 1->44->46->47->53->59->77->79->91->100->102->104->106->108->110          
#>  [49] 1->44->46->47->53->59->77->79->91->100->102->104->106->108->109          
#>  [50] 1->44->46->47->53->59->77->79->91->100->102->104->106->107               
#>  [51] 1->44->46->47->53->59->77->79->91->100->102->104->105->112->114->116->118
#>  [52] 1->44->46->47->53->59->77->79->91->100->102->104->105->112->114->116->117
#>  [53] 1->44->46->47->53->59->77->79->91->100->102->104->105->112->114->115     
#>  [54] 1->44->46->47->53->59->77->79->91->100->102->104->105->112->113          
#>  [55] 1->44->46->47->53->59->77->79->91->100->102->104->105->111->120          
#>  [56] 1->44->46->47->53->59->77->79->91->100->102->104->105->111->119          
#>  [57] 1->44->46->47->53->59->77->79->91->100->102->103->122->124               
#>  [58] 1->44->46->47->53->59->77->79->91->100->102->103->122->123               
#>  [59] 1->44->46->47->53->59->77->79->91->100->102->103->121                    
#>  [60] 1->44->46->47->53->59->77->79->91->100->101->126->128->130->132->134->136
#>  [61] 1->44->46->47->53->59->77->79->91->100->101->126->128->130->132->134->135
#>  [62] 1->44->46->47->53->59->77->79->91->100->101->126->128->130->132->133     
#>  [63] 1->44->46->47->53->59->77->79->91->100->101->126->128->130->131          
#>  [64] 1->44->46->47->53->59->77->79->91->100->101->126->128->129               
#>  [65] 1->44->46->47->53->59->77->79->91->100->101->126->127                    
#>  [66] 1->44->46->47->53->59->77->79->91->100->101->125->138                    
#>  [67] 1->44->46->47->53->59->77->79->91->100->101->125->137->140               
#>  [68] 1->44->46->47->53->59->77->79->91->100->101->125->137->139               
#>  [69] 1->44->46->47->53->59->77->79->91->99->142                               
#>  [70] 1->44->46->47->53->59->77->79->91->99->141->144                          
#>  [71] 1->44->46->47->53->59->77->79->91->99->141->143->146                     
#>  [72] 1->44->46->47->53->59->77->79->91->99->141->143->145                     
#>  [73] 1->44->45->148->150                                                      
#>  [74] 1->44->45->148->149->152->154                                            
#>  [75] 1->44->45->148->149->152->153                                            
#>  [76] 1->44->45->148->149->151                                                 
#>  [77] 1->44->45->147->156->158->160                                            
#>  [78] 1->44->45->147->156->158->159                                            
#>  [79] 1->44->45->147->156->157->162->164                                       
#>  [80] 1->44->45->147->156->157->162->163                                       
#>  [81] 1->44->45->147->156->157->161->166                                       
#>  [82] 1->44->45->147->156->157->161->165                                       
#>  [83] 1->44->45->147->155->168->170->172                                       
#>  [84] 1->44->45->147->155->168->170->171                                       
#>  [85] 1->44->45->147->155->168->169->174->176                                  
#>  [86] 1->44->45->147->155->168->169->174->175                                  
#>  [87] 1->44->45->147->155->168->169->173->178->180                             
#>  [88] 1->44->45->147->155->168->169->173->178->179                             
#>  [89] 1->44->45->147->155->168->169->173->177->182->184->186                   
#>  [90] 1->44->45->147->155->168->169->173->177->182->184->185                   
#>  [91] 1->44->45->147->155->168->169->173->177->182->183                        
#>  [92] 1->44->45->147->155->168->169->173->177->181->188                        
#>  [93] 1->44->45->147->155->168->169->173->177->181->187                        
#>  [94] 1->44->45->147->155->167                                                 
#>  [95] 1->43->190->192->194                                                     
#>  [96] 1->43->190->192->193                                                     
#>  [97] 1->43->190->191->196                                                     
#>  [98] 1->43->190->191->195                                                     
#>  [99] 1->43->189->198                                                          
#> [100] 1->43->189->197                                                          
#> [101] 2                                                                        
#> [102] 1                                                                        
#> [103] 1->44                                                                    
#> [104] 1->44->45                                                                
#> [105] 2->3                                                                     
#> [106] 1->44->46                                                                
#> [107] 1->43                                                                    
#> [108] 1->44->45->147                                                           
#> [109] 2->4                                                                     
#> [110] 1->44->46->47                                                            
#> [111] 2->3->13                                                                 
#> [112] 1->44->45->147->156                                                      
#> [113] 1->44->45->147->155                                                      
#> [114] 1->44->46->47->53                                                        
#> [115] 1->44->46->47->53->59                                                    
#> [116] 1->44->46->48                                                            
#> [117] 1->44->46->47->53->59->77                                                
#> [118] 1->44->46->47->53->59->77->79                                            
#> [119] 2->3->14                                                                 
#> [120] 1->44->45->147->155->168                                                 
#> [121] 1->43->190                                                               
#> [122] 2->4->6                                                                  
#> [123] 1->44->46->47->53->59->77->79->91                                        
#> [124] 2->3->13->32                                                             
#> [125] 1->44->46->47->53->59->77->80                                            
#> [126] 1->44->46->47->53->59->77->79->92                                        
#> [127] 1->44->46->48->49                                                        
#> [128] 1->44->46->47->53->59->77->79->91->100                                   
#> [129] 2->3->14->16                                                             
#> [130] 1->44->46->47->53->60                                                    
#> [131] 1->44->45->147->155->168->169                                            
#> [132] 2->3->13->32->34                                                         
#> [133] 1->44->46->47->53->60->61                                                
#> [134] 1->44->46->47->53->59->77->79->91->99                                    
#> [135] 1->44->46->47->53->59->77->79->91->100->102                              
#> [136] 1->44->46->47->53->59->77->80->82                                        
#> [137] 1->44->45->147->155->168->169->173                                       
#> [138] 1->44->46->47->53->59->77->79->92->93                                    
#> [139] 2->4->6->7                                                               
#> [140] 1->44->46->47->53->59->77->79->91->100->101                              
#> [141] 1->44->45->147->155->168->169->173->177                                  
#> [142] 2->3->14->15                                                             
#> [143] 1->44->45->148                                                           
#> [144] 1->44->45->147->155->168->169->174                                       
#> [145] 1->44->45->147->156->157                                                 
#> [146] 1->44->46->47->53->59->77->79->91->100->102->104                         
#> [147] 1->44->46->47->53->59->77->79->91->100->101->125                         
#> [148] 1->44->45->147->156->157->162                                            
#> [149] 1->44->46->47->53->59->77->79->91->100->101->126                         
#> [150] 1->44->46->47->53->59->77->79->91->99->141                               
#> [151] 1->44->45->148->149                                                      
#> [152] 2->3->14->16->17                                                         
#> [153] 1->44->46->47->53->59->77->79->91->100->102->104->105                    
#> [154] 1->44->46->47->53->60->61->64                                            
#> [155] 2->3->14->16->17->19                                                     
#> [156] 1->44->46->47->53->60->61->64->66                                        
#> [157] 2->3->14->15->26                                                         
#> [158] 2->3->14->15->26->28                                                     
#> [159] 1->44->45->147->155->168->169->173->178                                  
#> [160] 1->44->46->47->53->60->61->64->65                                        
#> [161] 2->3->13->32->34->36                                                     
#> [162] 1->44->46->47->53->59->77->80->82->84                                    
#> [163] 1->44->46->47->53->59->77->80->82->84->85                                
#> [164] 1->44->46->47->53->59->77->79->91->100->102->104->106                    
#> [165] 1->44->45->147->155->168->169->173->177->181                             
#> [166] 1->44->46->47->53->60->61->64->65->70                                    
#> [167] 1->44->46->47->53->59->77->79->91->100->101->125->137                    
#> [168] 1->43->190->191                                                          
#> [169] 1->44->45->147->156->157->161                                            
#> [170] 1->44->46->47->54                                                        
#> [171] 1->44->46->47->53->59->77->79->91->100->102->103                         
#> [172] 1->44->46->47->53->60->61->64->65->70->71                                
#> [173] 1->44->46->47->53->59->77->79->91->100->102->103->122                    
#> [174] 1->44->46->47->53->59->77->79->91->100->102->104->105->111               
#> [175] 1->44->46->47->53->59->77->79->91->100->102->104->106->108               
#> [176] 1->44->45->147->156->158                                                 
#> [177] 1->44->45->147->155->168->169->173->177->182                             
#> [178] 1->44->46->47->53->59->77->79->91->100->102->104->105->112               
#> [179] 2->3->13->31                                                             
#> [180] 1->44->46->47->53->59->77->79->91->100->101->126->128                    
#> [181] 1->44->46->47->53->59->77->79->91->99->141->143                          
#> [182] 1->44->46->47->53->59->77->79->91->100->101->126->128->130               
#> [183] 1->44->46->47->53->59->77->79->91->100->101->126->128->130->132          
#> [184] 2->4->6->7->10                                                           
#> [185] 1->44->45->147->155->168->170                                            
#> [186] 2->3->13->32->33                                                         
#> [187] 1->43->189                                                               
#> [188] 1->44->46->47->53->60->61->64->65->69                                    
#> [189] 1->44->46->47->53->59->77->79->92->93->95                                
#> [190] 1->43->190->192                                                          
#> [191] 1->44->45->147->155->168->169->173->177->182->184                        
#> [192] 1->44->46->47->53->59->77->79->91->100->102->104->105->112->114          
#> [193] 1->44->46->47->53->59->77->79->91->100->101->126->128->130->132->134     
#> [194] 1->44->46->47->53->59->77->80->81                                        
#> [195] 1->44->45->148->149->152                                                 
#> [196] 1->44->46->47->53->59->77->79->91->100->102->104->105->112->114->116     
#> [197] 1->44->46->47->54->55                                                    
#> [198] 2->3->14->16->17->19->21