Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──0.1355──→ Node3:Node2 ──0.2890──→ Node3:Node7 ──0.0210──→ Node3:Node9 ──0.0257──→ Node3:Node11 ──0.0862──→ Node3:t5 ──0.2015──→ Node17:Node2 ──0.4297──→ Node17:Node7 ──0.0312──→ Node17:Node9 ──0.0382──→ Node17:Node11 ──0.1282──→ Node17:t5 ──0.0089──→ t17:Node2 ──0.0190──→ t17:Node7 ──0.0014──→ t17:Node9 ──0.0017──→ t17:Node11 ──0.0057──→ t17:t5 
#> 
#> [2] ◎──0.1355──→ Node3:Node2 ──0.2890──→ Node3:Node7 ──0.0210──→ Node3:Node9 ──0.0257──→ Node3:Node11 ──0.0862──→ Node3:t3 ──0.2015──→ Node17:Node2 ──0.4297──→ Node17:Node7 ──0.0312──→ Node17:Node9 ──0.0382──→ Node17:Node11 ──0.1282──→ Node17:t3 ──0.0089──→ t5:Node2 ──0.0190──→ t5:Node7 ──0.0014──→ t5:Node9 ──0.0017──→ t5:Node11 ──0.0057──→ t5:t3 
#> 
#> [3] ◎──0.1355──→ Node3:Node2 ──0.2890──→ Node3:Node7 ──0.0210──→ Node3:Node9 ──0.1119──→ Node3:t1 ──0.1470──→ Node6:Node2 ──0.3136──→ Node6:Node7 ──0.0228──→ Node6:Node9 ──0.1214──→ Node6:t1 ──0.0634──→ t3:Node2 ──0.1351──→ t3:Node7 ──0.0098──→ t3:Node9 ──0.0523──→ t3:t1 
#> 
#> [4] ◎──0.1355──→ Node3:Node2 ──0.2890──→ Node3:Node7 ──0.1329──→ Node3:t12 ──0.1470──→ Node6:Node2 ──0.3136──→ Node6:Node7 ──0.1442──→ Node6:t12 ──0.0076──→ Node7:Node2 ──0.0162──→ Node7:Node7 ──0.0074──→ Node7:t12 ──0.0116──→ Node8:Node2 ──0.0247──→ Node8:Node7 ──0.0114──→ Node8:t12 ──0.0289──→ Node13:Node2 ──0.0617──→ Node13:Node7 ──0.0284──→ Node13:t12 ──0.0029──→ Node16:Node2 ──0.0061──→ Node16:Node7 ──0.0028──→ Node16:t12 ──0.0124──→ t14:Node2 ──0.0265──→ t14:Node7 ──0.0122──→ t14:t12 
#> 
#> [5] ◎──0.13552──→ Node3:Node2 ──0.17511──→ Node3:Node3 ──0.07411──→ Node3:Node4 ──0.01764──→ Node3:Node5 ──0.15507──→ Node3:t2 ──0.14704──→ Node6:Node2 ──0.18999──→ Node6:Node3 ──0.08041──→ Node6:Node4 ──0.01914──→ Node6:Node5 ──0.16824──→ Node6:t2 ──0.00758──→ Node7:Node2 ──0.00979──→ Node7:Node3 ──0.00414──→ Node7:Node4 ──0.00099──→ Node7:Node5 ──0.00867──→ Node7:t2 ──0.01158──→ Node8:Node2 ──0.01496──→ Node8:Node3 ──0.00633──→ Node8:Node4 ──0.00151──→ Node8:Node5 ──0.01325──→ Node8:t2 ──0.02892──→ Node13:Node2 ──0.03737──→ Node13:Node3 ──0.01582──→ Node13:Node4 ──0.00376──→ Node13:Node5 ──0.03309──→ Node13:t2 ──0.00286──→ Node16:Node2 ──0.00370──→ Node16:Node3 ──0.00157──→ Node16:Node4 ──0.00037──→ Node16:Node5 ──0.00328──→ Node16:t2 ──0.00873──→ Node19:Node2 ──0.01128──→ Node19:Node3 ──0.00477──→ Node19:Node4 ──0.00114──→ Node19:Node5 ──0.00999──→ Node19:t2 ──0.00369──→ t20:Node2 ──0.00477──→ t20:Node3 ──0.00202──→ t20:Node4 ──0.00048──→ t20:Node5 ──0.00423──→ t20:t2 
#> 
#> [6] ◎──0.13552──→ Node3:Node2 ──0.17511──→ Node3:Node3 ──0.07411──→ Node3:Node4 ──0.01764──→ Node3:Node5 ──0.03972──→ Node3:Node8 ──0.11534──→ Node3:t13 ──0.14704──→ Node6:Node2 ──0.18999──→ Node6:Node3 ──0.08041──→ Node6:Node4 ──0.01914──→ Node6:Node5 ──0.04310──→ Node6:Node8 ──0.12514──→ Node6:t13 ──0.00758──→ Node7:Node2 ──0.00979──→ Node7:Node3 ──0.00414──→ Node7:Node4 ──0.00099──→ Node7:Node5 ──0.00222──→ Node7:Node8 ──0.00645──→ Node7:t13 ──0.01158──→ Node8:Node2 ──0.01496──→ Node8:Node3 ──0.00633──→ Node8:Node4 ──0.00151──→ Node8:Node5 ──0.00339──→ Node8:Node8 ──0.00985──→ Node8:t13 ──0.02892──→ Node13:Node2 ──0.03737──→ Node13:Node3 ──0.01582──→ Node13:Node4 ──0.00376──→ Node13:Node5 ──0.00848──→ Node13:Node8 ──0.02462──→ Node13:t13 ──0.00286──→ Node16:Node2 ──0.00370──→ Node16:Node3 ──0.00157──→ Node16:Node4 ──0.00037──→ Node16:Node5 ──0.00084──→ Node16:Node8 ──0.00244──→ Node16:t13 ──0.00873──→ Node19:Node2 ──0.01128──→ Node19:Node3 ──0.00477──→ Node19:Node4 ──0.00114──→ Node19:Node5 ──0.00256──→ Node19:Node8 ──0.00743──→ Node19:t13 ──0.00369──→ t11:Node2 ──0.00477──→ t11:Node3 ──0.00202──→ t11:Node4 ──0.00048──→ t11:Node5 ──0.00108──→ t11:Node8 ──0.00314──→ t11:t13 
#> 
#> [7] ◎──0.13552──→ Node3:Node2 ──0.17511──→ Node3:Node3 ──0.07411──→ Node3:Node4 ──0.01764──→ Node3:Node5 ──0.03972──→ Node3:Node8 ──0.11534──→ Node3:t4 ──0.14704──→ Node6:Node2 ──0.18999──→ Node6:Node3 ──0.08041──→ Node6:Node4 ──0.01914──→ Node6:Node5 ──0.04310──→ Node6:Node8 ──0.12514──→ Node6:t4 ──0.00758──→ Node7:Node2 ──0.00979──→ Node7:Node3 ──0.00414──→ Node7:Node4 ──0.00099──→ Node7:Node5 ──0.00222──→ Node7:Node8 ──0.00645──→ Node7:t4 ──0.01158──→ Node8:Node2 ──0.01496──→ Node8:Node3 ──0.00633──→ Node8:Node4 ──0.00151──→ Node8:Node5 ──0.00339──→ Node8:Node8 ──0.00985──→ Node8:t4 ──0.02892──→ Node13:Node2 ──0.03737──→ Node13:Node3 ──0.01582──→ Node13:Node4 ──0.00376──→ Node13:Node5 ──0.00848──→ Node13:Node8 ──0.02462──→ Node13:t4 ──0.01528──→ t6:Node2 ──0.01975──→ t6:Node3 ──0.00836──→ t6:Node4 ──0.00199──→ t6:Node5 ──0.00448──→ t6:Node8 ──0.01301──→ t6:t4 
#> 
#> [8] ◎──0.1355──→ Node3:Node2 ──0.1751──→ Node3:Node3 ──0.0741──→ Node3:Node4 ──0.1727──→ Node3:t6 ──0.1470──→ Node6:Node2 ──0.1900──→ Node6:Node3 ──0.0804──→ Node6:Node4 ──0.1874──→ Node6:t6 ──0.0076──→ Node7:Node2 ──0.0098──→ Node7:Node3 ──0.0041──→ Node7:Node4 ──0.0097──→ Node7:t6 ──0.0116──→ Node8:Node2 ──0.0150──→ Node8:Node3 ──0.0063──→ Node8:Node4 ──0.0148──→ Node8:t6 ──0.0403──→ Node18:Node2 ──0.0521──→ Node18:Node3 ──0.0221──→ Node18:Node4 ──0.0514──→ Node18:t6 ──0.0039──→ t16:Node2 ──0.0050──→ t16:Node3 ──0.0021──→ t16:Node4 ──0.0049──→ t16:t6 
#> 
#> [9] ◎──0.13552──→ Node3:Node2 ──0.17511──→ Node3:Node3 ──0.15754──→ Node3:Node10 ──0.08423──→ Node3:Node18 ──0.00505──→ Node3:t9 ──0.14704──→ Node6:Node2 ──0.18999──→ Node6:Node3 ──0.17092──→ Node6:Node10 ──0.09139──→ Node6:Node18 ──0.00547──→ Node6:t9 ──0.00758──→ Node7:Node2 ──0.00979──→ Node7:Node3 ──0.00881──→ Node7:Node10 ──0.00471──→ Node7:Node18 ──0.00028──→ Node7:t9 ──0.01158──→ Node8:Node2 ──0.01496──→ Node8:Node3 ──0.01346──→ Node8:Node10 ──0.00720──→ Node8:Node18 ──0.00043──→ Node8:t9 ──0.04033──→ Node18:Node2 ──0.05211──→ Node18:Node3 ──0.04688──→ Node18:Node10 ──0.02507──→ Node18:Node18 ──0.00150──→ Node18:t9 ──0.00388──→ t8:Node2 ──0.00501──→ t8:Node3 ──0.00451──→ t8:Node10 ──0.00241──→ t8:Node18 ──0.00014──→ t8:t9 
#> 
#> [10] ◎──0.13552──→ Node3:Node2 ──0.17511──→ Node3:Node3 ──0.15754──→ Node3:Node10 ──0.08423──→ Node3:Node18 ──0.00505──→ Node3:t10 ──0.14704──→ Node6:Node2 ──0.18999──→ Node6:Node3 ──0.17092──→ Node6:Node10 ──0.09139──→ Node6:Node18 ──0.00547──→ Node6:t10 ──0.00758──→ Node7:Node2 ──0.00979──→ Node7:Node3 ──0.00881──→ Node7:Node10 ──0.00471──→ Node7:Node18 ──0.00028──→ Node7:t10 ──0.03093──→ Node11:Node2 ──0.03997──→ Node11:Node3 ──0.03596──→ Node11:Node10 ──0.01923──→ Node11:Node18 ──0.00115──→ Node11:t10 ──0.01206──→ Node14:Node2 ──0.01558──→ Node14:Node3 ──0.01402──→ Node14:Node10 ──0.00749──→ Node14:Node18 ──0.00045──→ Node14:t10 ──0.01279──→ t2:Node2 ──0.01653──→ t2:Node3 ──0.01487──→ t2:Node10 ──0.00795──→ t2:Node18 ──0.00048──→ t2:t10