Scale the phylogenetic flow features to a constant sum.

Can be used to standardise branch lengths to reasonable values

Usage

pf_scale_flow_sum(x, scale_to = 1)

Arguments

x

a pfc object

scale_to

The value to scale the sums to. Default is 1.

Value

a scaled pfc object

Examples

pf_scale_flow_sum(rpfc(100))
#> <pfc<e:198>[198]>
#> First  10 phylogenetic flows: 
#> [1] ◎──0.527──→ Node3 ──0.424──→ Node26 ──0.037──→ Node59 ──0.013──→ t82 
#> 
#> [2] ◎──0.527──→ Node3 ──0.424──→ Node26 ──0.037──→ Node59 ──0.013──→ t43 
#> 
#> [3] ◎── 0.53──→ Node3 ── 0.42──→ Node26 ── 0.05──→ t18 
#> 
#> [4] ◎──0.527──→ Node3 ──0.176──→ Node5 ──0.107──→ Node6 ──0.157──→ Node34 ──0.034──→ t34 
#> 
#> [5] ◎──0.527──→ Node3 ──0.176──→ Node5 ──0.107──→ Node6 ──0.157──→ Node34 ──0.034──→ t52 
#> 
#> [6] ◎──5.3e-01──→ Node3 ──1.8e-01──→ Node5 ──1.1e-01──→ Node6 ──2.2e-02──→ Node7 ──3.5e-02──→ Node11 ──1.2e-01──→ Node50 ──5.6e-03──→ Node58 ──8.1e-03──→ Node77 ──3.5e-03──→ Node92 ──4.9e-05──→ Node93 ──1.3e-03──→ t76 
#> 
#> [7] ◎──5.3e-01──→ Node3 ──1.8e-01──→ Node5 ──1.1e-01──→ Node6 ──2.2e-02──→ Node7 ──3.5e-02──→ Node11 ──1.2e-01──→ Node50 ──5.6e-03──→ Node58 ──8.1e-03──→ Node77 ──3.5e-03──→ Node92 ──4.9e-05──→ Node93 ──1.3e-03──→ t57 
#> 
#> [8] ◎──0.5265──→ Node3 ──0.1759──→ Node5 ──0.1069──→ Node6 ──0.0216──→ Node7 ──0.0350──→ Node11 ──0.1156──→ Node50 ──0.0056──→ Node58 ──0.0081──→ Node77 ──0.0035──→ Node92 ──0.0014──→ t81 
#> 
#> [9] ◎──0.5265──→ Node3 ──0.1759──→ Node5 ──0.1069──→ Node6 ──0.0216──→ Node7 ──0.0350──→ Node11 ──0.1156──→ Node50 ──0.0056──→ Node58 ──0.0081──→ Node77 ──0.0015──→ Node82 ──0.0034──→ t19 
#> 
#> [10] ◎──0.5265──→ Node3 ──0.1759──→ Node5 ──0.1069──→ Node6 ──0.0216──→ Node7 ──0.0350──→ Node11 ──0.1156──→ Node50 ──0.0056──→ Node58 ──0.0081──→ Node77 ──0.0015──→ Node82 ──0.0034──→ t22