Calculate a 'second order' pfc, which represents summed branch length from each node to the terminal node along each flow. When used in modelling traits, a second order pfc implies that rates of evolution of the trait are themselves evolving according to Brownian motion.

Calculate a 'second order' pfc, which represents summed branch length from each node to the terminal node along each flow. When used in modelling traits, a second order pfc implies that rates of evolution of the trait are themselves evolving according to Brownian motion.

Usage

pf_second_order(x, ...)

Arguments

x

A pfc object

...

Additional arguments passed to or other methods.

Value

A pfc object

Examples

pf_second_order(rpfc(100))
#> <pfc<e:198>[198]>
#> First  10 phylogenetic flows: 
#> [1] ◎──2.029──→ Node2 ──1.165──→ Node3 ──0.912──→ Node5 ──0.301──→ Node7 ──0.253──→ Node11 ──0.167──→ Node12 ──0.166──→ Node45 ──0.025──→ t63 
#> 
#> [2] ◎──2.029──→ Node2 ──1.165──→ Node3 ──0.912──→ Node5 ──0.301──→ Node7 ──0.253──→ Node11 ──0.167──→ Node12 ──0.166──→ Node45 ──0.025──→ t62 
#> 
#> [3] ◎──2.029──→ Node2 ──1.165──→ Node3 ──0.912──→ Node5 ──0.301──→ Node7 ──0.253──→ Node11 ──0.167──→ Node12 ──0.166──→ Node19 ──0.092──→ Node38 ──0.033──→ Node41 ──0.032──→ t54 
#> 
#> [4] ◎──2.029──→ Node2 ──1.165──→ Node3 ──0.912──→ Node5 ──0.301──→ Node7 ──0.253──→ Node11 ──0.167──→ Node12 ──0.166──→ Node19 ──0.092──→ Node38 ──0.033──→ Node41 ──0.032──→ t40 
#> 
#> [5] ◎──2.029──→ Node2 ──1.165──→ Node3 ──0.912──→ Node5 ──0.301──→ Node7 ──0.253──→ Node11 ──0.167──→ Node12 ──0.166──→ Node19 ──0.092──→ Node38 ──0.033──→ Node50 ──0.020──→ t4 
#> 
#> [6] ◎──2.0287──→ Node2 ──1.1652──→ Node3 ──0.9125──→ Node5 ──0.3009──→ Node7 ──0.2535──→ Node11 ──0.1669──→ Node12 ──0.1662──→ Node19 ──0.0924──→ Node38 ──0.0332──→ Node50 ──0.0196──→ Node63 ──0.0111──→ Node65 ──0.0105──→ Node74 ──0.0079──→ t90 
#> 
#> [7] ◎──2.0287──→ Node2 ──1.1652──→ Node3 ──0.9125──→ Node5 ──0.3009──→ Node7 ──0.2535──→ Node11 ──0.1669──→ Node12 ──0.1662──→ Node19 ──0.0924──→ Node38 ──0.0332──→ Node50 ──0.0196──→ Node63 ──0.0111──→ Node65 ──0.0105──→ Node74 ──0.0079──→ Node87 ──0.0030──→ t21 
#> 
#> [8] ◎──2.0287──→ Node2 ──1.1652──→ Node3 ──0.9125──→ Node5 ──0.3009──→ Node7 ──0.2535──→ Node11 ──0.1669──→ Node12 ──0.1662──→ Node19 ──0.0924──→ Node38 ──0.0332──→ Node50 ──0.0196──→ Node63 ──0.0111──→ Node65 ──0.0105──→ Node74 ──0.0079──→ Node87 ──0.0030──→ t80 
#> 
#> [9] ◎──2.0287──→ Node2 ──1.1652──→ Node3 ──0.9125──→ Node5 ──0.3009──→ Node7 ──0.2535──→ Node11 ──0.1669──→ Node12 ──0.1662──→ Node19 ──0.0924──→ Node38 ──0.0332──→ Node50 ──0.0196──→ Node63 ──0.0111──→ Node65 ──0.0105──→ Node77 ──0.0066──→ t12 
#> 
#> [10] ◎──2.0287──→ Node2 ──1.1652──→ Node3 ──0.9125──→ Node5 ──0.3009──→ Node7 ──0.2535──→ Node11 ──0.1669──→ Node12 ──0.1662──→ Node19 ──0.0924──→ Node38 ──0.0332──→ Node50 ──0.0196──→ Node63 ──0.0111──→ Node65 ──0.0105──→ Node77 ──0.0066──→ Node88 ──0.0030──→ t72