Standardise the phylogenetic flow features to an implied typical variance of 1.

Can be used to make random effects based on pfcs comparable.

Usage

pf_standardise(x)

Arguments

x

a pfc object

Value

a scaled pfc object

Examples

pf_standardise(rpfc(100))
#> <pfc<e:198>[198]>
#> First  10 phylogenetic flows: 
#> [1] ◎──0.9169──→ Node8 ──0.0771──→ Node25 ──0.0023──→ Node27 ──0.0054──→ Node37 ──0.0067──→ Node56 ──0.0036──→ Node76 ──0.0024──→ t86 
#> 
#> [2] ◎──0.9169──→ Node8 ──0.0771──→ Node25 ──0.0023──→ Node27 ──0.0054──→ Node37 ──0.0067──→ Node56 ──0.0036──→ Node76 ──0.0024──→ t82 
#> 
#> [3] ◎──0.9169──→ Node8 ──0.0771──→ Node25 ──0.0023──→ Node27 ──0.0054──→ Node37 ──0.0067──→ Node56 ──0.0060──→ t84 
#> 
#> [4] ◎──0.9169──→ Node8 ──0.0771──→ Node25 ──0.0023──→ Node27 ──0.0054──→ Node37 ──0.0077──→ Node61 ──0.0049──→ t91 
#> 
#> [5] ◎──9.2e-01──→ Node8 ──7.7e-02──→ Node25 ──2.3e-03──→ Node27 ──5.4e-03──→ Node37 ──7.7e-03──→ Node61 ──4.9e-03──→ Node97 ──6.1e-05──→ t62 
#> 
#> [6] ◎──9.2e-01──→ Node8 ──7.7e-02──→ Node25 ──2.3e-03──→ Node27 ──5.4e-03──→ Node37 ──7.7e-03──→ Node61 ──4.9e-03──→ Node97 ──6.1e-05──→ t39 
#> 
#> [7] ◎──0.9169──→ Node8 ──0.0771──→ Node25 ──0.0023──→ Node27 ──0.0060──→ Node38 ──0.0120──→ t28 
#> 
#> [8] ◎──0.9169──→ Node8 ──0.0771──→ Node25 ──0.0023──→ Node27 ──0.0060──→ Node38 ──0.0120──→ t52 
#> 
#> [9] ◎──0.917──→ Node8 ──0.077──→ Node25 ──0.009──→ Node41 ──0.011──→ t16 
#> 
#> [10] ◎──0.9169──→ Node8 ──0.0771──→ Node25 ──0.0090──→ Node41 ──0.0066──→ Node64 ──0.0047──→ t77