Primate Diet Diversity and Threat Status Data with Phylogeny

Data on primates diets and their threat status joined to a pf object (for {phyf})

Usage

primate_diet

Format

primate_diet

A 'pf' data frame (subclasses tibble) with 504 rows and 55 columns:

label

Node labels including species name for the tip labels

phlo

The phylogenetic flow column which stores the phylogenetic information

Threat status

IUCN threat category

Threat status source

Threat status data source – see primate_refs dataset

Body mass (g)

Primate species' mean body mass in grams

Body mass source)

Body mass data source – see primate_refs dataset

Range size (km2)

Primate species' range size in kilometers

Range size source)

Range size data source – see primate_refs dataset

Diet disparity (PSV)

Primate species' disparity of diet items as measured by Phylogenetic Species Variability (PSV) metric.

Diet breadth

Primate species' breadth of diet items as measured by the number of different diet items (richness).

Diet diversity (DDI)

Primate species' diversity of diet items as measured by the DDI metric.

Trophic guild

Primate species' trophic guild. Possible values: "Omnivore", "Frugivore", "Gummivore", "Insectivore", or "Folivore-frugivore"

diet_item:item (40 columns)

Each of the 40 columns starting with "diet_item:" represents a different type of item in primates' diets. These are binary integer columns with a 1 if the species feeds on that diet item or 0 if it does not.

Source

https://zslpublications.onlinelibrary.wiley.com/doi/full/10.1111/acv.12823

References

Machado, F. F., Jardim, L., Dinnage, R., Brito, D., & Cardillo, M. (2022). Diet disparity and diversity predict extinction risk in primates. Animal Conservation.