AVONET Bird Trait Data with Phylogeny

The AVONET Bird Trait Database joined to a pf object (for {phyf})

Usage

avonet

Format

who

A 'pf' data frame (subclasses tibble) with 13,338 rows and 38 columns:

node_name

Node labels including species name for the tip labels

phyf

The phylogenetic flow column which stores the phylogenetic information

Species3, Family3, Order3

Taxonomic names -- Names of the species, family and order, respectively

Total.individuals

Number of individuals used to measure the data

Beak.Length_Culmen:Species.Status

Various traits of the bird species, see Source section to get more detailed information

...

Source

https://figshare.com/articles/dataset/AVONET_morphological_ecological_and_geographical_data_for_all_birds_Tobias_et_al_2021_Ecology_Letters_/16586228