Return a pfc with ancestral features

Return a pfc with ancestral features

Usage

pf_anc(x, replace = 0)

Arguments

x

A pfc object

replace

Value to replace edge feature with no ancestral value (such as when the ancestor is the root)

Value

A new pfc with the same structure as x, but with the features of each edge's ancestors instead

Examples

pf_anc(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] --0.0000->Node2--0.0126->Node8--1.2077->Node13--0.0914->t22                                                                                                                                                               
#>   [2] --0.0000->Node2--0.0126->Node8--1.2077->Node13--0.0914->Node37--0.0540->t4                                                                                                                                                
#>   [3] --0.0000->Node2--0.0126->Node8--1.2077->Node13--0.0914->Node37--0.0540->Node48--0.0123->t88                                                                                                                               
#>   [4] --0.0000->Node2--0.0126->Node8--1.2077->Node13--0.0914->Node37--0.0540->Node48--0.0123->t98                                                                                                                               
#>   [5] --0.0000->Node2--0.0126->Node8--1.2077->Node9--0.0172->Node18--0.0927->Node50--0.0505->t37                                                                                                                                
#>   [6] --0.0000->Node2--0.0126->Node8--1.2077->Node9--0.0172->Node18--0.0927->Node50--0.0505->t84                                                                                                                                
#>   [7] --0.0000->Node2--0.0126->Node8--1.2077->Node9--0.0172->Node18--0.0927->Node61--0.0579->t100                                                                                                                               
#>   [8] --0.0000->Node2--0.0126->Node8--1.2077->Node9--0.0172->Node18--0.0927->Node61--0.0579->t41                                                                                                                                
#>   [9] --0.0000->Node2--0.0126->Node8--1.2077->Node9--0.0172->Node15--0.0809->Node63--0.0701->t91                                                                                                                                
#>  [10] --0.0000->Node2--0.0126->Node8--1.2077->Node9--0.0172->Node15--0.0809->Node63--0.0701->t25                                                                                                                                
#>  [11] -0.00000->Node2-0.01257->Node8-1.20767->Node9-0.01716->Node15-0.08087->Node16-0.00778->Node24-0.01258->t60                                                                                                                
#>  [12] -0.00000->Node2-0.01257->Node8-1.20767->Node9-0.01716->Node15-0.08087->Node16-0.00778->Node24-0.01258->t44                                                                                                                
#>  [13] -0.00000->Node2-0.01257->Node8-1.20767->Node9-0.01716->Node15-0.08087->Node16-0.00778->Node32-0.03266->Node42-0.01540->t17                                                                                                
#>  [14] -0.00000->Node2-0.01257->Node8-1.20767->Node9-0.01716->Node15-0.08087->Node16-0.00778->Node32-0.03266->Node42-0.01540->t58                                                                                                
#>  [15] -0.00000->Node2-0.01257->Node8-1.20767->Node9-0.01716->Node15-0.08087->Node16-0.00778->Node32-0.03266->t11                                                                                                                
#>  [16] --0.0000->Node2--0.0126->t65                                                                                                                                                                                              
#>  [17] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node21--0.2381->Node28--0.0171->Node90--0.0500->t89                                                                                                 
#>  [18] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node21--0.2381->Node28--0.0171->Node90--0.0500->t90                                                                                                 
#>  [19] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node21--0.2381->Node28--0.0171->Node47--0.0284->Node70--0.0163->t12                                                                                 
#>  [20] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node21--0.2381->Node28--0.0171->Node47--0.0284->Node70--0.0163->t87                                                                                 
#>  [21] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node6-0.10891->Node21-0.23809->Node28-0.01710->Node47-0.02844->Node73-0.01675->Node75-0.00201->t92                                                                 
#>  [22] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node6-0.10891->Node21-0.23809->Node28-0.01710->Node47-0.02844->Node73-0.01675->Node75-0.00201->t49                                                                 
#>  [23] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node21--0.2381->Node28--0.0171->Node47--0.0284->Node73--0.0167->t75                                                                                 
#>  [24] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node21--0.2381->Node36--0.0313->t95                                                                                                                 
#>  [25] -0.000000>Node3-0.688143>Node4-0.090989>Node5-0.206203>Node6-0.108908>Node21-0.238094>Node36-0.031319>Node49-0.015897>Node85-0.019488>Node92-0.000898>Node98-0.001744>t57                                                 
#>  [26] -0.000000>Node3-0.688143>Node4-0.090989>Node5-0.206203>Node6-0.108908>Node21-0.238094>Node36-0.031319>Node49-0.015897>Node85-0.019488>Node92-0.000898>Node98-0.001744>t73                                                 
#>  [27] -0.000000>Node3-0.688143>Node4-0.090989>Node5-0.206203>Node6-0.108908>Node21-0.238094>Node36-0.031319>Node49-0.015897>Node85-0.019488>Node92-0.000898>t2                                                                  
#>  [28] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node21--0.2381->Node36--0.0313->Node49--0.0159->Node85--0.0195->t30                                                                                 
#>  [29] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node21--0.2381->Node36--0.0313->Node49--0.0159->t76                                                                                                 
#>  [30] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node20--0.2370->Node38--0.0346->t1                                                                                                                  
#>  [31] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node20--0.2370->Node38--0.0346->t80                                                                                                                 
#>  [32] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node6--0.109-->Node20--0.237-->Node31--0.025-->t59                                                                                                                 
#>  [33] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node6-0.10891->Node20-0.23700->Node31-0.02499->Node40-0.01700->Node56-0.01376->Node62-0.00132->t18                                                                 
#>  [34] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node6-0.10891->Node20-0.23700->Node31-0.02499->Node40-0.01700->Node56-0.01376->Node62-0.00132->t81                                                                 
#>  [35] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node20--0.2370->Node31--0.0250->Node40--0.0170->Node56--0.0138->t56                                                                                 
#>  [36] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node6--0.109-->Node20--0.237-->Node31--0.025-->Node40--0.017-->t83                                                                                                 
#>  [37] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node10--0.277-->Node12--0.014-->Node45--0.101-->t26                                                                                                                
#>  [38] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node10--0.277-->Node12--0.014-->Node45--0.101-->t32                                                                                                                
#>  [39] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node58-0.03754->Node66-0.00253->t53                                                
#>  [40] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node58-0.03754->Node66-0.00253->t5                                                 
#>  [41] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node58-0.03754->Node99-0.01426->t9                                                 
#>  [42] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node58-0.03754->Node99-0.01426->t16                                                
#>  [43] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node43-0.02431->Node71-0.01998->t19                                                
#>  [44] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node43-0.02431->Node71-0.01998->Node83-0.00458->t23                                
#>  [45] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node43-0.02431->Node71-0.01998->Node83-0.00458->t42                                
#>  [46] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node43-0.02431->Node44-0.00077->Node80-0.02287->t39                                
#>  [47] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node43-0.02431->Node44-0.00077->Node80-0.02287->t86                                
#>  [48] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node43-0.02431->Node44-0.00077->t35                                                
#>  [49] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node55-0.05122->t6                                                                                 
#>  [50] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node55-0.05122->t66                                                                                
#>  [51] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node10--0.2771->Node12--0.0140->Node22--0.0569->Node34--0.0271->Node68--0.0328->t34                                                                                
#>  [52] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node34-0.02705->Node68-0.03284->Node82-0.00534->t31                                                                
#>  [53] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node34-0.02705->Node68-0.03284->Node82-0.00534->t68                                                                
#>  [54] -0.000000>Node3-0.688143>Node4-0.090989>Node5-0.206203>Node10-0.277055>Node12-0.014000>Node22-0.056940>Node34-0.027050>Node81-0.037922>Node88-0.000857>Node96-0.002411>t71                                                
#>  [55] -0.000000>Node3-0.688143>Node4-0.090989>Node5-0.206203>Node10-0.277055>Node12-0.014000>Node22-0.056940>Node34-0.027050>Node81-0.037922>Node88-0.000857>Node96-0.002411>t14                                                
#>  [56] -0.000000>Node3-0.688143>Node4-0.090989>Node5-0.206203>Node10-0.277055>Node12-0.014000>Node22-0.056940>Node34-0.027050>Node81-0.037922>Node88-0.000857>t69                                                                
#>  [57] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node10--0.2771->Node12--0.0140->Node22--0.0569->Node34--0.0271->Node81--0.0379->t29                                                                                
#>  [58] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->t8                                                                                                                 
#>  [59] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node84-0.03783->t99                                                                                
#>  [60] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node84-0.03783->t63                                                                                
#>  [61] -0.000000>Node3-0.688143>Node4-0.090989>Node5-0.206203>Node10-0.277055>Node17-0.064795>Node19-0.003214>Node35-0.030715>Node41-0.012682>Node53-0.010374>Node67-0.006888>Node77-0.005864>Node78-0.000678>Node93-0.002348>t47
#>  [62] -0.000000>Node3-0.688143>Node4-0.090989>Node5-0.206203>Node10-0.277055>Node17-0.064795>Node19-0.003214>Node35-0.030715>Node41-0.012682>Node53-0.010374>Node67-0.006888>Node77-0.005864>Node78-0.000678>Node93-0.002348>t51
#>  [63] -0.000000>Node3-0.688143>Node4-0.090989>Node5-0.206203>Node10-0.277055>Node17-0.064795>Node19-0.003214>Node35-0.030715>Node41-0.012682>Node53-0.010374>Node67-0.006888>Node77-0.005864>Node78-0.000678>Node89-0.001688>t28
#>  [64] -0.000000>Node3-0.688143>Node4-0.090989>Node5-0.206203>Node10-0.277055>Node17-0.064795>Node19-0.003214>Node35-0.030715>Node41-0.012682>Node53-0.010374>Node67-0.006888>Node77-0.005864>Node78-0.000678>Node89-0.001688>t20
#>  [65] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41-0.01268->Node53-0.01037->Node67-0.00689->Node77-0.00586->t72                                
#>  [66] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41-0.01268->Node53-0.01037->Node67-0.00689->Node95-0.00938->t3                                 
#>  [67] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41-0.01268->Node53-0.01037->Node67-0.00689->Node95-0.00938->t96                                
#>  [68] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41-0.01268->Node53-0.01037->Node74-0.01049->t38                                                
#>  [69] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41-0.01268->Node53-0.01037->Node74-0.01049->Node79-0.00304->t10                                
#>  [70] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41-0.01268->Node53-0.01037->Node74-0.01049->Node79-0.00304->t27                                
#>  [71] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41-0.01268->Node69-0.02009->t77                                                                
#>  [72] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41-0.01268->Node69-0.02009->t94                                                                
#>  [73] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node46-0.03203->t15                                                                                
#>  [74] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node46-0.03203->Node59-0.00982->Node76-0.00932->t7                                                 
#>  [75] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node46-0.03203->Node59-0.00982->Node76-0.00932->Node86-0.00227->t45                                
#>  [76] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node46-0.03203->Node59-0.00982->Node76-0.00932->Node86-0.00227->t85                                
#>  [77] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node46-0.03203->Node59-0.00982->Node65-0.00161->Node87-0.01004->t46                                
#>  [78] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node46-0.03203->Node59-0.00982->Node65-0.00161->Node87-0.01004->t97                                
#>  [79] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node46-0.03203->Node59-0.00982->Node65-0.00161->t50                                                
#>  [80] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node64-0.04300->Node94-0.01166->t55                                                                
#>  [81] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node64-0.04300->Node94-0.01166->t40                                                                
#>  [82] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node64-0.04300->t33                                                                                
#>  [83] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node33-0.01785->Node39-0.00943->Node52-0.01271->Node57-0.00548->t62                                                
#>  [84] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node33-0.01785->Node39-0.00943->Node52-0.01271->Node57-0.00548->t74                                                
#>  [85] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node33-0.01785->Node39-0.00943->Node52-0.01271->t61                                                                
#>  [86] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node33-0.01785->Node39-0.00943->t52                                                                                
#>  [87] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node10--0.2771->Node17--0.0648->Node25--0.0148->Node33--0.0179->Node54--0.0260->t78                                                                                
#>  [88] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node10--0.2771->Node17--0.0648->Node25--0.0148->Node33--0.0179->Node54--0.0260->t21                                                                                
#>  [89] --0.000-->Node3--0.688-->Node4--0.091-->Node7--0.399-->t48                                                                                                                                                                
#>  [90] -0.00000->Node3-0.68814->Node4-0.09099->Node7-0.39867->Node14-0.13980->Node91-0.08183->Node97-0.00222->t43                                                                                                                
#>  [91] -0.00000->Node3-0.68814->Node4-0.09099->Node7-0.39867->Node14-0.13980->Node91-0.08183->Node97-0.00222->t13                                                                                                                
#>  [92] --0.0000->Node3--0.6881->Node4--0.0910->Node7--0.3987->Node14--0.1398->Node91--0.0818->t79                                                                                                                                
#>  [93] --0.000-->Node3--0.688-->Node4--0.091-->Node7--0.399-->Node14--0.140-->Node72--0.077-->t36                                                                                                                                
#>  [94] --0.000-->Node3--0.688-->Node4--0.091-->Node7--0.399-->Node14--0.140-->Node72--0.077-->t54                                                                                                                                
#>  [95] -0.00000->Node3-0.68814->Node11-0.58437->Node26-0.07089->Node29-0.00642->t67                                                                                                                                              
#>  [96] -0.00000->Node3-0.68814->Node11-0.58437->Node26-0.07089->Node29-0.00642->Node60-0.03776->t64                                                                                                                              
#>  [97] -0.00000->Node3-0.68814->Node11-0.58437->Node26-0.07089->Node29-0.00642->Node60-0.03776->t24                                                                                                                              
#>  [98] --0.0000->Node3--0.6881->Node11--0.5844->Node26--0.0709->Node51--0.0383->t82                                                                                                                                              
#>  [99] --0.0000->Node3--0.6881->Node11--0.5844->Node26--0.0709->Node51--0.0383->t70                                                                                                                                              
#> [100] --0.000-->Node3--0.688-->Node11--0.584-->t93                                                                                                                                                                              
#> [101] ----0---->Node2                                                                                                                                                                                                           
#> [102] ----0---->Node3                                                                                                                                                                                                           
#> [103] --0.000-->Node3--0.688-->Node4                                                                                                                                                                                            
#> [104] --0.000-->Node3--0.688-->Node4--0.091-->Node5                                                                                                                                                                             
#> [105] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node6                                                                                                                                                              
#> [106] --0.000-->Node3--0.688-->Node4--0.091-->Node7                                                                                                                                                                             
#> [107] --0.0000->Node2--0.0126->Node8                                                                                                                                                                                            
#> [108] --0.0000->Node2--0.0126->Node8--1.2077->Node9                                                                                                                                                                             
#> [109] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node10                                                                                                                                                             
#> [110] --0.000-->Node3--0.688-->Node11                                                                                                                                                                                           
#> [111] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node10--0.277-->Node12                                                                                                                                             
#> [112] --0.0000->Node2--0.0126->Node8--1.2077->Node13                                                                                                                                                                            
#> [113] --0.000-->Node3--0.688-->Node4--0.091-->Node7--0.399-->Node14                                                                                                                                                             
#> [114] --0.0000->Node2--0.0126->Node8--1.2077->Node9--0.0172->Node15                                                                                                                                                             
#> [115] --0.0000->Node2--0.0126->Node8--1.2077->Node9--0.0172->Node15--0.0809->Node16                                                                                                                                             
#> [116] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node10--0.277-->Node17                                                                                                                                             
#> [117] --0.0000->Node2--0.0126->Node8--1.2077->Node9--0.0172->Node18                                                                                                                                                             
#> [118] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node10--0.2771->Node17--0.0648->Node19                                                                                                                             
#> [119] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node6--0.109-->Node20                                                                                                                                              
#> [120] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node6--0.109-->Node21                                                                                                                                              
#> [121] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node10--0.277-->Node12--0.014-->Node22                                                                                                                             
#> [122] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node10--0.2771->Node12--0.0140->Node22--0.0569->Node23                                                                                                             
#> [123] -0.00000->Node2-0.01257->Node8-1.20767->Node9-0.01716->Node15-0.08087->Node16-0.00778->Node24                                                                                                                             
#> [124] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node10--0.2771->Node17--0.0648->Node25                                                                                                                             
#> [125] --0.000-->Node3--0.688-->Node11--0.584-->Node26                                                                                                                                                                           
#> [126] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node10--0.2771->Node17--0.0648->Node25--0.0148->Node27                                                                                                             
#> [127] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node6--0.109-->Node21--0.238-->Node28                                                                                                                              
#> [128] --0.0000->Node3--0.6881->Node11--0.5844->Node26--0.0709->Node29                                                                                                                                                           
#> [129] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30                                                                                             
#> [130] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node6--0.109-->Node20--0.237-->Node31                                                                                                                              
#> [131] -0.00000->Node2-0.01257->Node8-1.20767->Node9-0.01716->Node15-0.08087->Node16-0.00778->Node32                                                                                                                             
#> [132] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node10--0.2771->Node17--0.0648->Node25--0.0148->Node33                                                                                                             
#> [133] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node10--0.2771->Node12--0.0140->Node22--0.0569->Node34                                                                                                             
#> [134] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35                                                                                                             
#> [135] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node6--0.109-->Node21--0.238-->Node36                                                                                                                              
#> [136] --0.0000->Node2--0.0126->Node8--1.2077->Node13--0.0914->Node37                                                                                                                                                            
#> [137] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node6--0.109-->Node20--0.237-->Node38                                                                                                                              
#> [138] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node10--0.2771->Node17--0.0648->Node25--0.0148->Node33--0.0179->Node39                                                                                             
#> [139] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node6--0.109-->Node20--0.237-->Node31--0.025-->Node40                                                                                                              
#> [140] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41                                                                                             
#> [141] -0.00000->Node2-0.01257->Node8-1.20767->Node9-0.01716->Node15-0.08087->Node16-0.00778->Node32-0.03266->Node42                                                                                                             
#> [142] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node43                                                                             
#> [143] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node43-0.02431->Node44                                                             
#> [144] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node10--0.277-->Node12--0.014-->Node45                                                                                                                             
#> [145] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node46                                                                                             
#> [146] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node21--0.2381->Node28--0.0171->Node47                                                                                                              
#> [147] --0.0000->Node2--0.0126->Node8--1.2077->Node13--0.0914->Node37--0.0540->Node48                                                                                                                                            
#> [148] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node21--0.2381->Node36--0.0313->Node49                                                                                                              
#> [149] --0.0000->Node2--0.0126->Node8--1.2077->Node9--0.0172->Node18--0.0927->Node50                                                                                                                                             
#> [150] --0.0000->Node3--0.6881->Node11--0.5844->Node26--0.0709->Node51                                                                                                                                                           
#> [151] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node33-0.01785->Node39-0.00943->Node52                                                                             
#> [152] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41-0.01268->Node53                                                                             
#> [153] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node10--0.2771->Node17--0.0648->Node25--0.0148->Node33--0.0179->Node54                                                                                             
#> [154] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node55                                                                                             
#> [155] --0.000-->Node3--0.688-->Node4--0.091-->Node5--0.206-->Node6--0.109-->Node20--0.237-->Node31--0.025-->Node40--0.017-->Node56                                                                                              
#> [156] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node33-0.01785->Node39-0.00943->Node52-0.01271->Node57                                                             
#> [157] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node58                                                                             
#> [158] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node46-0.03203->Node59                                                                             
#> [159] -0.00000->Node3-0.68814->Node11-0.58437->Node26-0.07089->Node29-0.00642->Node60                                                                                                                                           
#> [160] --0.0000->Node2--0.0126->Node8--1.2077->Node9--0.0172->Node18--0.0927->Node61                                                                                                                                             
#> [161] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node20--0.2370->Node31--0.0250->Node40--0.0170->Node56--0.0138->Node62                                                                              
#> [162] --0.0000->Node2--0.0126->Node8--1.2077->Node9--0.0172->Node15--0.0809->Node63                                                                                                                                             
#> [163] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node64                                                                                             
#> [164] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node46-0.03203->Node59-0.00982->Node65                                                             
#> [165] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node58-0.03754->Node66                                                             
#> [166] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41-0.01268->Node53-0.01037->Node67                                                             
#> [167] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node10--0.2771->Node12--0.0140->Node22--0.0569->Node34--0.0271->Node68                                                                                             
#> [168] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41-0.01268->Node69                                                                             
#> [169] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node21--0.2381->Node28--0.0171->Node47--0.0284->Node70                                                                                              
#> [170] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node43-0.02431->Node71                                                             
#> [171] --0.000-->Node3--0.688-->Node4--0.091-->Node7--0.399-->Node14--0.140-->Node72                                                                                                                                             
#> [172] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node21--0.2381->Node28--0.0171->Node47--0.0284->Node73                                                                                              
#> [173] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41-0.01268->Node53-0.01037->Node74                                                             
#> [174] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node21--0.2381->Node28--0.0171->Node47--0.0284->Node73--0.0167->Node75                                                                              
#> [175] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node46-0.03203->Node59-0.00982->Node76                                                             
#> [176] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41-0.01268->Node53-0.01037->Node67-0.00689->Node77                                             
#> [177] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41-0.01268->Node53-0.01037->Node67-0.00689->Node77-0.00586->Node78                             
#> [178] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41-0.01268->Node53-0.01037->Node74-0.01049->Node79                                             
#> [179] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node43-0.02431->Node44-0.00077->Node80                                             
#> [180] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node10--0.2771->Node12--0.0140->Node22--0.0569->Node34--0.0271->Node81                                                                                             
#> [181] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node10--0.2771->Node12--0.0140->Node22--0.0569->Node34--0.0271->Node68--0.0328->Node82                                                                             
#> [182] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node43-0.02431->Node71-0.01998->Node83                                             
#> [183] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node84                                                                                             
#> [184] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node21--0.2381->Node36--0.0313->Node49--0.0159->Node85                                                                                              
#> [185] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node46-0.03203->Node59-0.00982->Node76-0.00932->Node86                                             
#> [186] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node46-0.03203->Node59-0.00982->Node65-0.00161->Node87                                             
#> [187] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node10--0.2771->Node12--0.0140->Node22--0.0569->Node34--0.0271->Node81--0.0379->Node88                                                                             
#> [188] -0.000000>Node3-0.688143>Node4-0.090989>Node5-0.206203>Node10-0.277055>Node17-0.064795>Node19-0.003214>Node35-0.030715>Node41-0.012682>Node53-0.010374>Node67-0.006888>Node77-0.005864>Node78-0.000678>Node89             
#> [189] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node21--0.2381->Node28--0.0171->Node90                                                                                                              
#> [190] --0.000-->Node3--0.688-->Node4--0.091-->Node7--0.399-->Node14--0.140-->Node91                                                                                                                                             
#> [191] --0.0000->Node3--0.6881->Node4--0.0910->Node5--0.2062->Node6--0.1089->Node21--0.2381->Node36--0.0313->Node49--0.0159->Node85--0.0195->Node92                                                                              
#> [192] -0.000000>Node3-0.688143>Node4-0.090989>Node5-0.206203>Node10-0.277055>Node17-0.064795>Node19-0.003214>Node35-0.030715>Node41-0.012682>Node53-0.010374>Node67-0.006888>Node77-0.005864>Node78-0.000678>Node93             
#> [193] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node25-0.01476->Node27-0.00371->Node64-0.04300->Node94                                                                             
#> [194] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node17-0.06479->Node19-0.00321->Node35-0.03071->Node41-0.01268->Node53-0.01037->Node67-0.00689->Node95                                             
#> [195] -0.000000>Node3-0.688143>Node4-0.090989>Node5-0.206203>Node10-0.277055>Node12-0.014000>Node22-0.056940>Node34-0.027050>Node81-0.037922>Node88-0.000857>Node96                                                             
#> [196] --0.0000->Node3--0.6881->Node4--0.0910->Node7--0.3987->Node14--0.1398->Node91--0.0818->Node97                                                                                                                             
#> [197] -0.000000>Node3-0.688143>Node4-0.090989>Node5-0.206203>Node6-0.108908>Node21-0.238094>Node36-0.031319>Node49-0.015897>Node85-0.019488>Node92-0.000898>Node98                                                              
#> [198] -0.00000->Node3-0.68814->Node4-0.09099->Node5-0.20620->Node10-0.27705->Node12-0.01400->Node22-0.05694->Node23-0.00152->Node30-0.01505->Node58-0.03754->Node99