Return a pfc with ancestral features

Return a pfc with ancestral features

Usage

pf_anc(x, replace = 0)

Arguments

x

A pfc object

replace

Value to replace edge feature with no ancestral value (such as when the ancestor is the root)

Value

A new pfc with the same structure as x, but with the features of each edge's ancestors instead

Examples

pf_anc(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] ---0.00-->Node44---1.28-->t23                                                                                                                                                                            
#>   [2] ---0.00-->Node44---1.28-->t54                                                                                                                                                                            
#>   [3] --0.000-->Node2--0.816-->Node4--0.101-->Node60--0.371-->t13                                                                                                                                              
#>   [4] --0.0000->Node2--0.8159->Node4--0.1015->Node60--0.3713->Node95--0.0123->t15                                                                                                                              
#>   [5] --0.0000->Node2--0.8159->Node4--0.1015->Node60--0.3713->Node95--0.0123->t28                                                                                                                              
#>   [6] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node17--0.0121->Node20--0.0122->Node77--0.0690->t98                                                                                              
#>   [7] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node17--0.0121->Node20--0.0122->Node77--0.0690->t44                                                                                              
#>   [8] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node17--0.0121->Node20--0.0122->t58                                                                                                              
#>   [9] -0.00000->Node2-0.81589->Node4-0.10146->Node13-0.28495->Node17-0.01207->Node26-0.03732->Node31-0.00978->t49                                                                                              
#>  [10] -0.00000->Node2-0.81589->Node4-0.10146->Node13-0.28495->Node17-0.01207->Node26-0.03732->Node31-0.00978->Node67-0.02987->t40                                                                              
#>  [11] -0.00000->Node2-0.81589->Node4-0.10146->Node13-0.28495->Node17-0.01207->Node26-0.03732->Node31-0.00978->Node67-0.02987->t36                                                                              
#>  [12] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node17--0.0121->Node26--0.0373->Node40--0.0215->Node73--0.0206->t57                                                                              
#>  [13] -0.00000->Node2-0.81589->Node4-0.10146->Node13-0.28495->Node17-0.01207->Node26-0.03732->Node40-0.02150->Node73-0.02055->Node79-0.00278->t9                                                               
#>  [14] -0.00000->Node2-0.81589->Node4-0.10146->Node13-0.28495->Node17-0.01207->Node26-0.03732->Node40-0.02150->Node73-0.02055->Node79-0.00278->t68                                                              
#>  [15] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node17--0.0121->Node26--0.0373->Node40--0.0215->t27                                                                                              
#>  [16] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node16--0.0102->Node29--0.0447->Node76--0.0378->t43                                                                                              
#>  [17] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node16--0.0102->Node29--0.0447->Node76--0.0378->t67                                                                                              
#>  [18] -0.00000->Node2-0.81589->Node4-0.10146->Node13-0.28495->Node16-0.01021->Node29-0.04471->Node32-0.00493->t65                                                                                              
#>  [19] -0.00000->Node2-0.81589->Node4-0.10146->Node13-0.28495->Node16-0.01021->Node29-0.04471->Node32-0.00493->t55                                                                                              
#>  [20] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node16--0.0102->Node24--0.0340->Node59--0.0411->t6                                                                                               
#>  [21] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node16--0.0102->Node24--0.0340->Node59--0.0411->t81                                                                                              
#>  [22] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node16--0.0102->Node24--0.0340->Node39--0.0258->Node70--0.0198->t83                                                                              
#>  [23] -0.00000->Node2-0.81589->Node4-0.10146->Node13-0.28495->Node16-0.01021->Node24-0.03398->Node39-0.02575->Node70-0.01976->Node98-0.00937->t70                                                              
#>  [24] -0.00000->Node2-0.81589->Node4-0.10146->Node13-0.28495->Node16-0.01021->Node24-0.03398->Node39-0.02575->Node70-0.01976->Node98-0.00937->t12                                                              
#>  [25] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node16--0.0102->Node24--0.0340->Node39--0.0258->Node92--0.0283->t79                                                                              
#>  [26] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node16--0.0102->Node24--0.0340->Node39--0.0258->Node92--0.0283->t3                                                                               
#>  [27] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node7--0.0570->Node15--0.1189->t87                                                                                                                
#>  [28] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node7--0.0570->Node15--0.1189->Node23--0.0360->t26                                                                                                
#>  [29] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node7--0.0570->Node15--0.1189->Node23--0.0360->t97                                                                                                
#>  [30] -0.000000>Node2-0.815890>Node3-0.048889>Node6-0.168846>Node7-0.057002>Node37-0.179426>Node43-0.007792>Node48-0.005094>Node49-0.000906>t75                                                                
#>  [31] -0.000000>Node2-0.815890>Node3-0.048889>Node6-0.168846>Node7-0.057002>Node37-0.179426>Node43-0.007792>Node48-0.005094>Node49-0.000906>Node51-0.001229>Node65-0.005590>t92                                
#>  [32] -0.000000>Node2-0.815890>Node3-0.048889>Node6-0.168846>Node7-0.057002>Node37-0.179426>Node43-0.007792>Node48-0.005094>Node49-0.000906>Node51-0.001229>Node65-0.005590>Node89-0.008842>t8                 
#>  [33] -0.000000>Node2-0.815890>Node3-0.048889>Node6-0.168846>Node7-0.057002>Node37-0.179426>Node43-0.007792>Node48-0.005094>Node49-0.000906>Node51-0.001229>Node65-0.005590>Node89-0.008842>t34                
#>  [34] -0.000000>Node2-0.815890>Node3-0.048889>Node6-0.168846>Node7-0.057002>Node37-0.179426>Node43-0.007792>Node48-0.005094>Node49-0.000906>Node51-0.001229>Node62-0.004694>t31                                
#>  [35] -0.000000>Node2-0.815890>Node3-0.048889>Node6-0.168846>Node7-0.057002>Node37-0.179426>Node43-0.007792>Node48-0.005094>Node49-0.000906>Node51-0.001229>Node62-0.004694>t69                                
#>  [36] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node7-0.05700->Node37-0.17943->Node43-0.00779->Node48-0.00509->t56                                                                                
#>  [37] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node7-0.05700->Node37-0.17943->Node43-0.00779->Node46-0.00448->Node58-0.00503->Node94-0.01344->t38                                                
#>  [38] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node7-0.05700->Node37-0.17943->Node43-0.00779->Node46-0.00448->Node58-0.00503->Node94-0.01344->t74                                                
#>  [39] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node7-0.05700->Node37-0.17943->Node43-0.00779->Node46-0.00448->Node58-0.00503->t5                                                                 
#>  [40] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node7-0.05700->Node37-0.17943->Node43-0.00779->Node46-0.00448->t1                                                                                 
#>  [41] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node7--0.0570->Node37--0.1794->Node50--0.0144->Node81--0.0128->t35                                                                                
#>  [42] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node7--0.0570->Node37--0.1794->Node50--0.0144->Node81--0.0128->t42                                                                                
#>  [43] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node7-0.05700->Node37-0.17943->Node50-0.01439->Node66-0.00651->Node75-0.00348->Node85-0.00422->t64                                                
#>  [44] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node7-0.05700->Node37-0.17943->Node50-0.01439->Node66-0.00651->Node75-0.00348->Node85-0.00422->t48                                                
#>  [45] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node7-0.05700->Node37-0.17943->Node50-0.01439->Node66-0.00651->Node75-0.00348->t61                                                                
#>  [46] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node7-0.05700->Node37-0.17943->Node50-0.01439->Node66-0.00651->t100                                                                               
#>  [47] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11--0.1505->Node14--0.0201->Node21--0.0334->t80                                                                                               
#>  [48] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node11-0.15049->Node14-0.02013->Node21-0.03342->Node22-0.00641->Node53-0.04236->Node82-0.01102->t33                                               
#>  [49] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node11-0.15049->Node14-0.02013->Node21-0.03342->Node22-0.00641->Node53-0.04236->Node82-0.01102->t59                                               
#>  [50] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node11-0.15049->Node14-0.02013->Node21-0.03342->Node22-0.00641->Node53-0.04236->Node74-0.00798->t20                                               
#>  [51] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node11-0.15049->Node14-0.02013->Node21-0.03342->Node22-0.00641->Node53-0.04236->Node74-0.00798->t21                                               
#>  [52] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node11-0.15049->Node14-0.02013->Node21-0.03342->Node22-0.00641->Node99-0.05744->t52                                                               
#>  [53] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node11-0.15049->Node14-0.02013->Node21-0.03342->Node22-0.00641->Node99-0.05744->t60                                                               
#>  [54] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11--0.1505->Node14--0.0201->Node35--0.0645->Node41--0.0056->t91                                                                               
#>  [55] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11--0.1505->Node14--0.0201->Node35--0.0645->Node41--0.0056->Node91--0.0256->t7                                                                
#>  [56] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11--0.1505->Node14--0.0201->Node35--0.0645->Node41--0.0056->Node91--0.0256->t47                                                               
#>  [57] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11--0.1505->Node14--0.0201->Node35--0.0645->Node86--0.0300->t46                                                                               
#>  [58] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11--0.1505->Node14--0.0201->Node35--0.0645->Node86--0.0300->t93                                                                               
#>  [59] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11--0.1505->Node52--0.1022->t45                                                                                                               
#>  [60] -0.000000>Node2-0.815890>Node3-0.048889>Node6-0.168846>Node11-0.150489>Node52-0.102228>Node57-0.000785>t50                                                                                               
#>  [61] -0.000000>Node2-0.815890>Node3-0.048889>Node6-0.168846>Node11-0.150489>Node52-0.102228>Node57-0.000785>Node88-0.012286>t22                                                                               
#>  [62] -0.000000>Node2-0.815890>Node3-0.048889>Node6-0.168846>Node11-0.150489>Node52-0.102228>Node57-0.000785>Node88-0.012286>t10                                                                               
#>  [63] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node10--0.0838->Node33--0.0824->t14                                                                                                
#>  [64] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node10--0.0838->Node33--0.0824->t88                                                                                                
#>  [65] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node69-0.07450->Node84-0.00598->t24                                                                
#>  [66] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node69-0.07450->Node84-0.00598->t11                                                                
#>  [67] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node10--0.0838->Node18--0.0347->Node69--0.0745->t62                                                                                
#>  [68] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node10--0.0838->Node18--0.0347->Node25--0.0300->Node63--0.0431->t66                                                                
#>  [69] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node10--0.0838->Node18--0.0347->Node25--0.0300->Node63--0.0431->t17                                                                
#>  [70] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node30-0.00520->Node36-0.01117->t90                                
#>  [71] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node30-0.00520->Node36-0.01117->Node61-0.01978->t96                
#>  [72] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node30-0.00520->Node36-0.01117->Node61-0.01978->t53                
#>  [73] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node30-0.00520->Node68-0.03253->t86                                
#>  [74] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node30-0.00520->Node68-0.03253->Node90-0.00857->t71                
#>  [75] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node30-0.00520->Node68-0.03253->Node90-0.00857->t77                
#>  [76] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node34-0.01158->t30                                                
#>  [77] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node34-0.01158->Node42-0.01072->t37                                
#>  [78] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node34-0.01158->Node42-0.01072->Node54-0.01094->Node87-0.01228->t94
#>  [79] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node34-0.01158->Node42-0.01072->Node54-0.01094->Node87-0.01228->t78
#>  [80] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node34-0.01158->Node42-0.01072->Node54-0.01094->t76                
#>  [81] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node9--0.0232->Node12--0.0720->t39                                                                                                 
#>  [82] -0.000000>Node2-0.815890>Node3-0.048889>Node5-0.065601>Node8-0.168452>Node9-0.023246>Node12-0.072011>Node78-0.102112>Node93-0.004513>Node96-0.000331>t82                                                 
#>  [83] -0.000000>Node2-0.815890>Node3-0.048889>Node5-0.065601>Node8-0.168452>Node9-0.023246>Node12-0.072011>Node78-0.102112>Node93-0.004513>Node96-0.000331>t2                                                  
#>  [84] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node9-0.02325->Node12-0.07201->Node78-0.10211->Node93-0.00451->t25                                                                 
#>  [85] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node9--0.0232->Node12--0.0720->Node78--0.1021->t18                                                                                 
#>  [86] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node9--0.0232->t29                                                                                                                 
#>  [87] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node47--0.0606->Node72--0.0107->t16                                                                                               
#>  [88] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node47--0.0606->Node72--0.0107->t85                                                                                               
#>  [89] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node19-0.29183->Node47-0.06058->Node71-0.00947->Node97-0.00895->t19                                                                               
#>  [90] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node19-0.29183->Node47-0.06058->Node71-0.00947->Node97-0.00895->t95                                                                               
#>  [91] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node19-0.29183->Node47-0.06058->Node71-0.00947->Node83-0.00525->t84                                                                               
#>  [92] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node19-0.29183->Node47-0.06058->Node71-0.00947->Node83-0.00525->t41                                                                               
#>  [93] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node27--0.0304->Node38--0.0185->Node45--0.0104->t73                                                                               
#>  [94] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node27--0.0304->Node38--0.0185->Node45--0.0104->t89                                                                               
#>  [95] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node27--0.0304->Node38--0.0185->Node55--0.0159->Node80--0.0103->t51                                                               
#>  [96] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node27--0.0304->Node38--0.0185->Node55--0.0159->Node80--0.0103->t99                                                               
#>  [97] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node27--0.0304->Node38--0.0185->Node55--0.0159->t72                                                                               
#>  [98] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node27--0.0304->Node56--0.0343->Node64--0.0037->t63                                                                               
#>  [99] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node27--0.0304->Node56--0.0343->Node64--0.0037->t32                                                                               
#> [100] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node27--0.0304->Node56--0.0343->t4                                                                                                
#> [101] ----0---->Node2                                                                                                                                                                                          
#> [102] --0.000-->Node2--0.816-->Node3                                                                                                                                                                           
#> [103] --0.000-->Node2--0.816-->Node4                                                                                                                                                                           
#> [104] --0.0000->Node2--0.8159->Node3--0.0489->Node5                                                                                                                                                            
#> [105] --0.0000->Node2--0.8159->Node3--0.0489->Node6                                                                                                                                                            
#> [106] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node7                                                                                                                                             
#> [107] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8                                                                                                                                             
#> [108] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node9                                                                                                                              
#> [109] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node10                                                                                                                             
#> [110] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11                                                                                                                                            
#> [111] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node9--0.0232->Node12                                                                                                              
#> [112] --0.000-->Node2--0.816-->Node4--0.101-->Node13                                                                                                                                                           
#> [113] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11--0.1505->Node14                                                                                                                            
#> [114] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node7--0.0570->Node15                                                                                                                             
#> [115] --0.000-->Node2--0.816-->Node4--0.101-->Node13--0.285-->Node16                                                                                                                                           
#> [116] --0.000-->Node2--0.816-->Node4--0.101-->Node13--0.285-->Node17                                                                                                                                           
#> [117] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node10--0.0838->Node18                                                                                                             
#> [118] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19                                                                                                                                            
#> [119] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node17--0.0121->Node20                                                                                                                           
#> [120] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11--0.1505->Node14--0.0201->Node21                                                                                                            
#> [121] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11--0.1505->Node14--0.0201->Node21--0.0334->Node22                                                                                            
#> [122] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node7--0.0570->Node15--0.1189->Node23                                                                                                             
#> [123] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node16--0.0102->Node24                                                                                                                           
#> [124] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node10--0.0838->Node18--0.0347->Node25                                                                                             
#> [125] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node17--0.0121->Node26                                                                                                                           
#> [126] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node27                                                                                                                            
#> [127] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node10--0.0838->Node18--0.0347->Node25--0.0300->Node28                                                                             
#> [128] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node16--0.0102->Node29                                                                                                                           
#> [129] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node30                                                             
#> [130] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node17--0.0121->Node26--0.0373->Node31                                                                                                           
#> [131] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node16--0.0102->Node29--0.0447->Node32                                                                                                           
#> [132] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node10--0.0838->Node33                                                                                                             
#> [133] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node34                                                             
#> [134] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11--0.1505->Node14--0.0201->Node35                                                                                                            
#> [135] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node30-0.00520->Node36                                             
#> [136] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node7--0.0570->Node37                                                                                                                             
#> [137] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node27--0.0304->Node38                                                                                                            
#> [138] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node16--0.0102->Node24--0.0340->Node39                                                                                                           
#> [139] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node17--0.0121->Node26--0.0373->Node40                                                                                                           
#> [140] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11--0.1505->Node14--0.0201->Node35--0.0645->Node41                                                                                            
#> [141] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node34-0.01158->Node42                                             
#> [142] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node7--0.0570->Node37--0.1794->Node43                                                                                                             
#> [143] ----0---->Node44                                                                                                                                                                                         
#> [144] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node27--0.0304->Node38--0.0185->Node45                                                                                            
#> [145] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node7-0.05700->Node37-0.17943->Node43-0.00779->Node46                                                                                             
#> [146] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node47                                                                                                                            
#> [147] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node7-0.05700->Node37-0.17943->Node43-0.00779->Node48                                                                                             
#> [148] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node7-0.05700->Node37-0.17943->Node43-0.00779->Node48-0.00509->Node49                                                                             
#> [149] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node7--0.0570->Node37--0.1794->Node50                                                                                                             
#> [150] -0.000000>Node2-0.815890>Node3-0.048889>Node6-0.168846>Node7-0.057002>Node37-0.179426>Node43-0.007792>Node48-0.005094>Node49-0.000906>Node51                                                             
#> [151] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11--0.1505->Node52                                                                                                                            
#> [152] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node11-0.15049->Node14-0.02013->Node21-0.03342->Node22-0.00641->Node53                                                                            
#> [153] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node34-0.01158->Node42-0.01072->Node54                             
#> [154] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node27--0.0304->Node38--0.0185->Node55                                                                                            
#> [155] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node27--0.0304->Node56                                                                                                            
#> [156] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11--0.1505->Node52--0.1022->Node57                                                                                                            
#> [157] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node7-0.05700->Node37-0.17943->Node43-0.00779->Node46-0.00448->Node58                                                                             
#> [158] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node16--0.0102->Node24--0.0340->Node59                                                                                                           
#> [159] --0.000-->Node2--0.816-->Node4--0.101-->Node60                                                                                                                                                           
#> [160] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node30-0.00520->Node36-0.01117->Node61                             
#> [161] -0.000000>Node2-0.815890>Node3-0.048889>Node6-0.168846>Node7-0.057002>Node37-0.179426>Node43-0.007792>Node48-0.005094>Node49-0.000906>Node51-0.001229>Node62                                             
#> [162] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node10--0.0838->Node18--0.0347->Node25--0.0300->Node63                                                                             
#> [163] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node27--0.0304->Node56--0.0343->Node64                                                                                            
#> [164] -0.000000>Node2-0.815890>Node3-0.048889>Node6-0.168846>Node7-0.057002>Node37-0.179426>Node43-0.007792>Node48-0.005094>Node49-0.000906>Node51-0.001229>Node65                                             
#> [165] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node7--0.0570->Node37--0.1794->Node50--0.0144->Node66                                                                                             
#> [166] -0.00000->Node2-0.81589->Node4-0.10146->Node13-0.28495->Node17-0.01207->Node26-0.03732->Node31-0.00978->Node67                                                                                           
#> [167] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node30-0.00520->Node68                                             
#> [168] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node10--0.0838->Node18--0.0347->Node69                                                                                             
#> [169] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node16--0.0102->Node24--0.0340->Node39--0.0258->Node70                                                                                           
#> [170] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node47--0.0606->Node71                                                                                                            
#> [171] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node47--0.0606->Node72                                                                                                            
#> [172] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node17--0.0121->Node26--0.0373->Node40--0.0215->Node73                                                                                           
#> [173] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node11-0.15049->Node14-0.02013->Node21-0.03342->Node22-0.00641->Node53-0.04236->Node74                                                            
#> [174] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node7-0.05700->Node37-0.17943->Node50-0.01439->Node66-0.00651->Node75                                                                             
#> [175] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node16--0.0102->Node29--0.0447->Node76                                                                                                           
#> [176] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node17--0.0121->Node20--0.0122->Node77                                                                                                           
#> [177] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node9--0.0232->Node12--0.0720->Node78                                                                                              
#> [178] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node17--0.0121->Node26--0.0373->Node40--0.0215->Node73--0.0206->Node79                                                                           
#> [179] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node19--0.2918->Node27--0.0304->Node38--0.0185->Node55--0.0159->Node80                                                                            
#> [180] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node7--0.0570->Node37--0.1794->Node50--0.0144->Node81                                                                                             
#> [181] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node11-0.15049->Node14-0.02013->Node21-0.03342->Node22-0.00641->Node53-0.04236->Node82                                                            
#> [182] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node19-0.29183->Node47-0.06058->Node71-0.00947->Node83                                                                                            
#> [183] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node10--0.0838->Node18--0.0347->Node69--0.0745->Node84                                                                             
#> [184] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node7-0.05700->Node37-0.17943->Node50-0.01439->Node66-0.00651->Node75-0.00348->Node85                                                             
#> [185] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11--0.1505->Node14--0.0201->Node35--0.0645->Node86                                                                                            
#> [186] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node34-0.01158->Node42-0.01072->Node54-0.01094->Node87             
#> [187] -0.000000>Node2-0.815890>Node3-0.048889>Node6-0.168846>Node11-0.150489>Node52-0.102228>Node57-0.000785>Node88                                                                                            
#> [188] -0.000000>Node2-0.815890>Node3-0.048889>Node6-0.168846>Node7-0.057002>Node37-0.179426>Node43-0.007792>Node48-0.005094>Node49-0.000906>Node51-0.001229>Node65-0.005590>Node89                             
#> [189] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node10-0.08377->Node18-0.03475->Node25-0.03003->Node28-0.00624->Node30-0.00520->Node68-0.03253->Node90                             
#> [190] --0.0000->Node2--0.8159->Node3--0.0489->Node6--0.1688->Node11--0.1505->Node14--0.0201->Node35--0.0645->Node41--0.0056->Node91                                                                            
#> [191] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node16--0.0102->Node24--0.0340->Node39--0.0258->Node92                                                                                           
#> [192] --0.0000->Node2--0.8159->Node3--0.0489->Node5--0.0656->Node8--0.1685->Node9--0.0232->Node12--0.0720->Node78--0.1021->Node93                                                                              
#> [193] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node7-0.05700->Node37-0.17943->Node43-0.00779->Node46-0.00448->Node58-0.00503->Node94                                                             
#> [194] --0.000-->Node2--0.816-->Node4--0.101-->Node60--0.371-->Node95                                                                                                                                           
#> [195] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node8-0.16845->Node9-0.02325->Node12-0.07201->Node78-0.10211->Node93-0.00451->Node96                                                              
#> [196] -0.00000->Node2-0.81589->Node3-0.04889->Node5-0.06560->Node19-0.29183->Node47-0.06058->Node71-0.00947->Node97                                                                                            
#> [197] --0.0000->Node2--0.8159->Node4--0.1015->Node13--0.2849->Node16--0.0102->Node24--0.0340->Node39--0.0258->Node70--0.0198->Node98                                                                           
#> [198] -0.00000->Node2-0.81589->Node3-0.04889->Node6-0.16885->Node11-0.15049->Node14-0.02013->Node21-0.03342->Node22-0.00641->Node99