Return a pfc with ancestral features

Return a pfc with ancestral features

Usage

pf_anc(x, replace = 0)

Arguments

x

A pfc object

replace

Value to replace edge feature with no ancestral value (such as when the ancestor is the root)

Value

A new pfc with the same structure as x, but with the features of each edge's ancestors instead

Examples

pf_anc(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node10--0.0355->Node58--0.1087->Node92--0.0139->t10                                                                                                                
#>   [2] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node10--0.0355->Node58--0.1087->Node92--0.0139->t73                                                                                                                
#>   [3] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node58-0.10868->Node71-0.00701->t89                                                                                                                
#>   [4] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node58-0.10868->Node71-0.00701->t81                                                                                                                
#>   [5] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26-0.04460->Node36-0.02342->Node59-0.02236->Node89-0.01367->t23                                                
#>   [6] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26-0.04460->Node36-0.02342->Node59-0.02236->Node89-0.01367->t38                                                
#>   [7] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26-0.04460->Node36-0.02342->Node59-0.02236->t45                                                                
#>   [8] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26-0.04460->Node36-0.02342->Node41-0.00876->t94                                                                
#>   [9] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26-0.04460->Node36-0.02342->Node41-0.00876->t17                                                                
#>  [10] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26-0.04460->Node40-0.02897->Node75-0.02485->Node87-0.00458->t87                                                
#>  [11] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26-0.04460->Node40-0.02897->Node75-0.02485->Node87-0.00458->t9                                                 
#>  [12] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26-0.04460->Node40-0.02897->Node75-0.02485->t7                                                                 
#>  [13] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26-0.04460->Node40-0.02897->t29                                                                                
#>  [14] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28-0.03127->Node76-0.04946->t63                                                                
#>  [15] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28-0.03127->Node76-0.04946->t56                                                                
#>  [16] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28-0.03127->Node30-0.00609->Node42-0.02201->Node55-0.01167->t37                                
#>  [17] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28-0.03127->Node30-0.00609->Node42-0.02201->Node55-0.01167->t60                                
#>  [18] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28-0.03127->Node30-0.00609->Node42-0.02201->Node50-0.00897->Node53-0.00125->t52                
#>  [19] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28-0.03127->Node30-0.00609->Node42-0.02201->Node50-0.00897->Node53-0.00125->Node62-0.00565->t47
#>  [20] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28-0.03127->Node30-0.00609->Node42-0.02201->Node50-0.00897->Node53-0.00125->Node62-0.00565->t88
#>  [21] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28-0.03127->Node30-0.00609->Node42-0.02201->Node50-0.00897->t78                                
#>  [22] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28-0.03127->Node30-0.00609->t54                                                                
#>  [23] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node18-0.00768->Node24-0.01669->t32                                                                
#>  [24] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node18-0.00768->Node24-0.01669->t22                                                                
#>  [25] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node18-0.00768->Node63-0.06764->Node67-0.00214->t19                                                
#>  [26] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node18-0.00768->Node63-0.06764->Node67-0.00214->Node96-0.00988->t99                                
#>  [27] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node18-0.00768->Node63-0.06764->Node67-0.00214->Node96-0.00988->t14                                
#>  [28] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node18-0.00768->Node63-0.06764->t50                                                                
#>  [29] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node23-0.01108->Node66-0.05712->Node80-0.00544->Node84-0.00179->t53                                                
#>  [30] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node23-0.01108->Node66-0.05712->Node80-0.00544->Node84-0.00179->t95                                                
#>  [31] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node23-0.01108->Node66-0.05712->Node80-0.00544->t42                                                                
#>  [32] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node23-0.01108->Node66-0.05712->t59                                                                                
#>  [33] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node23-0.01108->Node78-0.06177->t67                                                                                
#>  [34] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node23-0.01108->Node78-0.06177->t75                                                                                
#>  [35] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38-0.02186->Node54-0.01533->t62                                                                
#>  [36] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38-0.02186->Node54-0.01533->t31                                                                
#>  [37] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38-0.02186->Node60-0.01989->Node68-0.00312->Node81-0.00588->Node94-0.00328->t8                 
#>  [38] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38-0.02186->Node60-0.01989->Node68-0.00312->Node81-0.00588->Node94-0.00328->t40                
#>  [39] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38-0.02186->Node60-0.01989->Node68-0.00312->Node81-0.00588->t83                                
#>  [40] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38-0.02186->Node60-0.01989->Node68-0.00312->Node70-0.00146->t97                                
#>  [41] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38-0.02186->Node60-0.01989->Node68-0.00312->Node70-0.00146->t57                                
#>  [42] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38-0.02186->Node60-0.01989->Node93-0.01222->t5                                                 
#>  [43] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38-0.02186->Node60-0.01989->Node93-0.01222->t55                                                
#>  [44] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node34-0.01091->t71                                                                                
#>  [45] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node34-0.01091->Node51-0.02499->Node79-0.01371->t100                                               
#>  [46] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node34-0.01091->Node51-0.02499->Node79-0.01371->Node86-0.00292->t39                                
#>  [47] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node34-0.01091->Node51-0.02499->Node79-0.01371->Node86-0.00292->t69                                
#>  [48] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node34-0.01091->Node51-0.02499->Node97-0.01889->t18                                                
#>  [49] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node34-0.01091->Node51-0.02499->Node97-0.01889->t2                                                 
#>  [50] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node14--0.0579->Node25--0.0401->t27                                                                                                                                
#>  [51] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node14--0.0579->Node25--0.0401->Node32--0.0155->t61                                                                                                                
#>  [52] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node14-0.05786->Node25-0.04008->Node32-0.01552->Node57-0.03069->Node64-0.00352->Node72-0.00363->t26                                                                
#>  [53] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node14-0.05786->Node25-0.04008->Node32-0.01552->Node57-0.03069->Node64-0.00352->Node72-0.00363->Node82-0.00397->t33                                                
#>  [54] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node14-0.05786->Node25-0.04008->Node32-0.01552->Node57-0.03069->Node64-0.00352->Node72-0.00363->Node82-0.00397->t24                                                
#>  [55] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node14-0.05786->Node25-0.04008->Node32-0.01552->Node57-0.03069->Node64-0.00352->t64                                                                                
#>  [56] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node14--0.0579->Node25--0.0401->Node32--0.0155->Node57--0.0307->t86                                                                                                
#>  [57] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node14--0.0579->Node45--0.0757->t65                                                                                                                                
#>  [58] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node14-0.05786->Node45-0.07575->Node65-0.01420->Node85-0.00882->t25                                                                                                
#>  [59] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node14-0.05786->Node45-0.07575->Node65-0.01420->Node85-0.00882->t12                                                                                                
#>  [60] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node14--0.0579->Node45--0.0757->Node65--0.0142->t1                                                                                                                 
#>  [61] --0.000-->Node2--1.410-->Node4--0.401-->Node83--0.485-->t16                                                                                                                                                               
#>  [62] --0.000-->Node2--1.410-->Node4--0.401-->Node83--0.485-->t48                                                                                                                                                               
#>  [63] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node20--0.1811->Node37--0.0422->Node49--0.0144->t15                                                                                                                
#>  [64] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node20--0.1811->Node37--0.0422->Node49--0.0144->t70                                                                                                                
#>  [65] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node20--0.1811->Node37--0.0422->t41                                                                                                                                
#>  [66] --0.000-->Node2--1.410-->Node3--0.340-->Node5--0.291-->Node20--0.181-->t98                                                                                                                                                
#>  [67] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->t66                                                                                                                                                 
#>  [68] -0.00000->Node2-1.41026->Node3-0.33980->Node5-0.29132->Node7-0.08243->Node8-0.01012->Node17-0.08008->Node33-0.04033->Node35-0.00566->t49                                                                                  
#>  [69] -0.00000->Node2-1.41026->Node3-0.33980->Node5-0.29132->Node7-0.08243->Node8-0.01012->Node17-0.08008->Node33-0.04033->Node35-0.00566->t84                                                                                  
#>  [70] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node17--0.0801->Node33--0.0403->t72                                                                                                  
#>  [71] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node17--0.0801->Node21--0.0136->Node39--0.0396->Node56--0.0161->t20                                                                  
#>  [72] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node17--0.0801->Node21--0.0136->Node39--0.0396->Node56--0.0161->t96                                                                  
#>  [73] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node17--0.0801->Node21--0.0136->Node39--0.0396->t44                                                                                  
#>  [74] -0.00000->Node2-1.41026->Node3-0.33980->Node5-0.29132->Node7-0.08243->Node8-0.01012->Node17-0.08008->Node21-0.01362->Node46-0.04822->Node47-0.00223->t92                                                                  
#>  [75] -0.00000->Node2-1.41026->Node3-0.33980->Node5-0.29132->Node7-0.08243->Node8-0.01012->Node17-0.08008->Node21-0.01362->Node46-0.04822->Node47-0.00223->Node69-0.01288->Node88-0.00658->t30                                  
#>  [76] -0.00000->Node2-1.41026->Node3-0.33980->Node5-0.29132->Node7-0.08243->Node8-0.01012->Node17-0.08008->Node21-0.01362->Node46-0.04822->Node47-0.00223->Node69-0.01288->Node88-0.00658->Node91-0.00109->t79                  
#>  [77] -0.00000->Node2-1.41026->Node3-0.33980->Node5-0.29132->Node7-0.08243->Node8-0.01012->Node17-0.08008->Node21-0.01362->Node46-0.04822->Node47-0.00223->Node69-0.01288->Node88-0.00658->Node91-0.00109->t13                  
#>  [78] -0.00000->Node2-1.41026->Node3-0.33980->Node5-0.29132->Node7-0.08243->Node8-0.01012->Node17-0.08008->Node21-0.01362->Node46-0.04822->Node47-0.00223->Node69-0.01288->t3                                                   
#>  [79] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node17--0.0801->Node21--0.0136->Node46--0.0482->t90                                                                                  
#>  [80] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node27--0.1064->Node31--0.0132->Node95--0.0460->t76                                                                                  
#>  [81] -0.00000->Node2-1.41026->Node3-0.33980->Node5-0.29132->Node7-0.08243->Node8-0.01012->Node27-0.10637->Node31-0.01318->Node95-0.04601->Node99-0.00092->t36                                                                  
#>  [82] -0.00000->Node2-1.41026->Node3-0.33980->Node5-0.29132->Node7-0.08243->Node8-0.01012->Node27-0.10637->Node31-0.01318->Node95-0.04601->Node99-0.00092->t74                                                                  
#>  [83] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node27--0.1064->Node31--0.0132->Node77--0.0404->t35                                                                                  
#>  [84] -0.00000->Node2-1.41026->Node3-0.33980->Node5-0.29132->Node7-0.08243->Node8-0.01012->Node27-0.10637->Node31-0.01318->Node77-0.04042->Node98-0.00648->t80                                                                  
#>  [85] -0.00000->Node2-1.41026->Node3-0.33980->Node5-0.29132->Node7-0.08243->Node8-0.01012->Node27-0.10637->Node31-0.01318->Node77-0.04042->Node98-0.00648->t46                                                                  
#>  [86] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node27--0.1064->Node73--0.0517->t28                                                                                                  
#>  [87] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node27--0.1064->Node73--0.0517->t68                                                                                                  
#>  [88] -0.00000->Node2-1.41026->Node3-0.33980->Node12-0.43909->Node43-0.08345->Node48-0.00609->t6                                                                                                                                
#>  [89] -0.00000->Node2-1.41026->Node3-0.33980->Node12-0.43909->Node43-0.08345->Node48-0.00609->Node52-0.00285->Node74-0.01049->t21                                                                                               
#>  [90] -0.00000->Node2-1.41026->Node3-0.33980->Node12-0.43909->Node43-0.08345->Node48-0.00609->Node52-0.00285->Node74-0.01049->Node90-0.00649->t11                                                                               
#>  [91] -0.00000->Node2-1.41026->Node3-0.33980->Node12-0.43909->Node43-0.08345->Node48-0.00609->Node52-0.00285->Node74-0.01049->Node90-0.00649->t82                                                                               
#>  [92] -0.00000->Node2-1.41026->Node3-0.33980->Node12-0.43909->Node43-0.08345->Node48-0.00609->Node52-0.00285->Node61-0.00518->t77                                                                                               
#>  [93] -0.00000->Node2-1.41026->Node3-0.33980->Node12-0.43909->Node43-0.08345->Node48-0.00609->Node52-0.00285->Node61-0.00518->t91                                                                                               
#>  [94] -0.00e+00>Node2-1.41e+00>Node3-3.40e-01>Node12-4.39e-01>Node43-8.35e-02>Node44-5.04e-05>t4                                                                                                                                
#>  [95] -0.00e+00>Node2-1.41e+00>Node3-3.40e-01>Node12-4.39e-01>Node43-8.35e-02>Node44-5.04e-05>t51                                                                                                                               
#>  [96] --0.000-->Node2--1.410-->Node3--0.340-->Node12--0.439-->t43                                                                                                                                                               
#>  [97] ---0.00-->Node6---2.12-->t58                                                                                                                                                                                              
#>  [98] --0.0000->Node6--2.1189->Node15--0.0853->t93                                                                                                                                                                              
#>  [99] --0.0000->Node6--2.1189->Node15--0.0853->Node22--0.0261->t85                                                                                                                                                              
#> [100] --0.0000->Node6--2.1189->Node15--0.0853->Node22--0.0261->t34                                                                                                                                                              
#> [101] ----0---->Node2                                                                                                                                                                                                           
#> [102] ---0.00-->Node2---1.41-->Node3                                                                                                                                                                                            
#> [103] ---0.00-->Node2---1.41-->Node4                                                                                                                                                                                            
#> [104] ---0.00-->Node2---1.41-->Node3---0.34-->Node5                                                                                                                                                                             
#> [105] ----0---->Node6                                                                                                                                                                                                           
#> [106] --0.000-->Node2--1.410-->Node3--0.340-->Node5--0.291-->Node7                                                                                                                                                              
#> [107] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8                                                                                                                                               
#> [108] --0.000-->Node2--1.410-->Node4--0.401-->Node9                                                                                                                                                                             
#> [109] --0.000-->Node2--1.410-->Node4--0.401-->Node9--0.330-->Node10                                                                                                                                                             
#> [110] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node10--0.0355->Node11                                                                                                                                             
#> [111] ---0.00-->Node2---1.41-->Node3---0.34-->Node12                                                                                                                                                                            
#> [112] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13                                                                                                                             
#> [113] --0.000-->Node2--1.410-->Node4--0.401-->Node9--0.330-->Node14                                                                                                                                                             
#> [114] ---0.00-->Node6---2.12-->Node15                                                                                                                                                                                           
#> [115] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16                                                                                                             
#> [116] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node17                                                                                                                               
#> [117] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node18                                                                                             
#> [118] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19                                                                                                                             
#> [119] --0.000-->Node2--1.410-->Node3--0.340-->Node5--0.291-->Node20                                                                                                                                                             
#> [120] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node17--0.0801->Node21                                                                                                               
#> [121] --0.0000->Node6--2.1189->Node15--0.0853->Node22                                                                                                                                                                           
#> [122] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node23                                                                                                             
#> [123] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node18-0.00768->Node24                                                                             
#> [124] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node14--0.0579->Node25                                                                                                                                             
#> [125] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26                                                                                                             
#> [126] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node27                                                                                                                               
#> [127] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28                                                                                             
#> [128] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29                                                                                                             
#> [129] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28-0.03127->Node30                                                                             
#> [130] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node27--0.1064->Node31                                                                                                               
#> [131] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node14--0.0579->Node25--0.0401->Node32                                                                                                                             
#> [132] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node17--0.0801->Node33                                                                                                               
#> [133] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node34                                                                                             
#> [134] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node17--0.0801->Node33--0.0403->Node35                                                                                               
#> [135] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26-0.04460->Node36                                                                                             
#> [136] --0.000-->Node2--1.410-->Node3--0.340-->Node5--0.291-->Node20--0.181-->Node37                                                                                                                                             
#> [137] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38                                                                                             
#> [138] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node17--0.0801->Node21--0.0136->Node39                                                                                               
#> [139] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26-0.04460->Node40                                                                                             
#> [140] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26-0.04460->Node36-0.02342->Node41                                                                             
#> [141] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28-0.03127->Node30-0.00609->Node42                                                             
#> [142] --0.000-->Node2--1.410-->Node3--0.340-->Node12--0.439-->Node43                                                                                                                                                            
#> [143] --0.0000->Node2--1.4103->Node3--0.3398->Node12--0.4391->Node43--0.0835->Node44                                                                                                                                            
#> [144] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node14--0.0579->Node45                                                                                                                                             
#> [145] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node17--0.0801->Node21--0.0136->Node46                                                                                               
#> [146] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node17--0.0801->Node21--0.0136->Node46--0.0482->Node47                                                                               
#> [147] --0.0000->Node2--1.4103->Node3--0.3398->Node12--0.4391->Node43--0.0835->Node48                                                                                                                                            
#> [148] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node20--0.1811->Node37--0.0422->Node49                                                                                                                             
#> [149] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28-0.03127->Node30-0.00609->Node42-0.02201->Node50                                             
#> [150] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node34-0.01091->Node51                                                                             
#> [151] -0.00000->Node2-1.41026->Node3-0.33980->Node12-0.43909->Node43-0.08345->Node48-0.00609->Node52                                                                                                                            
#> [152] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28-0.03127->Node30-0.00609->Node42-0.02201->Node50-0.00897->Node53                             
#> [153] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38-0.02186->Node54                                                                             
#> [154] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28-0.03127->Node30-0.00609->Node42-0.02201->Node55                                             
#> [155] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node17--0.0801->Node21--0.0136->Node39--0.0396->Node56                                                                               
#> [156] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node14--0.0579->Node25--0.0401->Node32--0.0155->Node57                                                                                                             
#> [157] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node10--0.0355->Node58                                                                                                                                             
#> [158] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26-0.04460->Node36-0.02342->Node59                                                                             
#> [159] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38-0.02186->Node60                                                                             
#> [160] -0.00000->Node2-1.41026->Node3-0.33980->Node12-0.43909->Node43-0.08345->Node48-0.00609->Node52-0.00285->Node61                                                                                                            
#> [161] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28-0.03127->Node30-0.00609->Node42-0.02201->Node50-0.00897->Node53-0.00125->Node62             
#> [162] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node18-0.00768->Node63                                                                             
#> [163] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node14--0.0579->Node25--0.0401->Node32--0.0155->Node57--0.0307->Node64                                                                                             
#> [164] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node14--0.0579->Node45--0.0757->Node65                                                                                                                             
#> [165] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node23-0.01108->Node66                                                                                             
#> [166] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node18-0.00768->Node63-0.06764->Node67                                                             
#> [167] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38-0.02186->Node60-0.01989->Node68                                                             
#> [168] -0.00000->Node2-1.41026->Node3-0.33980->Node5-0.29132->Node7-0.08243->Node8-0.01012->Node17-0.08008->Node21-0.01362->Node46-0.04822->Node47-0.00223->Node69                                                               
#> [169] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38-0.02186->Node60-0.01989->Node68-0.00312->Node70                                             
#> [170] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node10--0.0355->Node58--0.1087->Node71                                                                                                                             
#> [171] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node14-0.05786->Node25-0.04008->Node32-0.01552->Node57-0.03069->Node64-0.00352->Node72                                                                             
#> [172] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node27--0.1064->Node73                                                                                                               
#> [173] -0.00000->Node2-1.41026->Node3-0.33980->Node12-0.43909->Node43-0.08345->Node48-0.00609->Node52-0.00285->Node74                                                                                                            
#> [174] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26-0.04460->Node40-0.02897->Node75                                                                             
#> [175] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node28-0.03127->Node76                                                                             
#> [176] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node27--0.1064->Node31--0.0132->Node77                                                                                               
#> [177] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node23-0.01108->Node78                                                                                             
#> [178] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node34-0.01091->Node51-0.02499->Node79                                                             
#> [179] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node23-0.01108->Node66-0.05712->Node80                                                                             
#> [180] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38-0.02186->Node60-0.01989->Node68-0.00312->Node81                                             
#> [181] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node14-0.05786->Node25-0.04008->Node32-0.01552->Node57-0.03069->Node64-0.00352->Node72-0.00363->Node82                                                             
#> [182] --0.000-->Node2--1.410-->Node4--0.401-->Node83                                                                                                                                                                            
#> [183] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node23-0.01108->Node66-0.05712->Node80-0.00544->Node84                                                             
#> [184] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node14--0.0579->Node45--0.0757->Node65--0.0142->Node85                                                                                                             
#> [185] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node34-0.01091->Node51-0.02499->Node79-0.01371->Node86                                             
#> [186] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26-0.04460->Node40-0.02897->Node75-0.02485->Node87                                                             
#> [187] -0.00000->Node2-1.41026->Node3-0.33980->Node5-0.29132->Node7-0.08243->Node8-0.01012->Node17-0.08008->Node21-0.01362->Node46-0.04822->Node47-0.00223->Node69-0.01288->Node88                                               
#> [188] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node26-0.04460->Node36-0.02342->Node59-0.02236->Node89                                                             
#> [189] -0.00000->Node2-1.41026->Node3-0.33980->Node12-0.43909->Node43-0.08345->Node48-0.00609->Node52-0.00285->Node74-0.01049->Node90                                                                                            
#> [190] -0.00000->Node2-1.41026->Node3-0.33980->Node5-0.29132->Node7-0.08243->Node8-0.01012->Node17-0.08008->Node21-0.01362->Node46-0.04822->Node47-0.00223->Node69-0.01288->Node88-0.00658->Node91                               
#> [191] --0.0000->Node2--1.4103->Node4--0.4007->Node9--0.3297->Node10--0.0355->Node58--0.1087->Node92                                                                                                                             
#> [192] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38-0.02186->Node60-0.01989->Node93                                                             
#> [193] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node38-0.02186->Node60-0.01989->Node68-0.00312->Node81-0.00588->Node94                             
#> [194] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node27--0.1064->Node31--0.0132->Node95                                                                                               
#> [195] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node13-0.01446->Node16-0.01779->Node18-0.00768->Node63-0.06764->Node67-0.00214->Node96                                             
#> [196] -0.00000->Node2-1.41026->Node4-0.40069->Node9-0.32968->Node10-0.03550->Node11-0.00386->Node19-0.04110->Node29-0.02376->Node34-0.01091->Node51-0.02499->Node97                                                             
#> [197] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node27--0.1064->Node31--0.0132->Node77--0.0404->Node98                                                                               
#> [198] --0.0000->Node2--1.4103->Node3--0.3398->Node5--0.2913->Node7--0.0824->Node8--0.0101->Node27--0.1064->Node31--0.0132->Node95--0.0460->Node99