Return a pfc with ancestral features

Return a pfc with ancestral features

Usage

pf_anc(x, replace = 0)

Arguments

x

A pfc object

replace

Value to replace edge feature with no ancestral value (such as when the ancestor is the root)

Value

A new pfc with the same structure as x, but with the features of each edge's ancestor instead

Examples

pf_anc(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] -0.00000->Node2-0.37424->Node13-1.42115->Node29-0.10919->Node31-0.00168->Node67-0.03379->t22                                                                                                           
#>   [2] -0.00000->Node2-0.37424->Node13-1.42115->Node29-0.10919->Node31-0.00168->Node67-0.03379->t99                                                                                                           
#>   [3] -0.00000->Node2-0.37424->Node13-1.42115->Node29-0.10919->Node31-0.00168->Node80-0.03726->t30                                                                                                           
#>   [4] -0.00000->Node2-0.37424->Node13-1.42115->Node29-0.10919->Node31-0.00168->Node80-0.03726->Node83-0.00131->Node94-0.00285->t18                                                                           
#>   [5] -0.00000->Node2-0.37424->Node13-1.42115->Node29-0.10919->Node31-0.00168->Node80-0.03726->Node83-0.00131->Node94-0.00285->t20                                                                           
#>   [6] -0.00000->Node2-0.37424->Node13-1.42115->Node29-0.10919->Node31-0.00168->Node80-0.03726->Node83-0.00131->t36                                                                                           
#>   [7] --0.000-->Node2--0.374-->Node13--1.421-->Node29--0.109-->t10                                                                                                                                           
#>   [8] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node27--0.0151->t80                                                                                                                           
#>   [9] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node27--0.0151->Node41--0.0258->t72                                                                                                           
#>  [10] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node27--0.0151->Node41--0.0258->t73                                                                                                           
#>  [11] -0.00000->Node2-0.37424->Node13-1.42115->Node21-0.08735->Node23-0.01095->Node47-0.03611->Node75-0.01204->Node91-0.00562->t23                                                                           
#>  [12] -0.00000->Node2-0.37424->Node13-1.42115->Node21-0.08735->Node23-0.01095->Node47-0.03611->Node75-0.01204->Node91-0.00562->t93                                                                           
#>  [13] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node23--0.0109->Node47--0.0361->Node75--0.0120->t54                                                                                           
#>  [14] -0.00000->Node2-0.37424->Node13-1.42115->Node21-0.08735->Node23-0.01095->Node47-0.03611->Node76-0.01220->Node98-0.00653->t55                                                                           
#>  [15] -0.000000>Node2-0.374243>Node13-1.421149>Node21-0.087352>Node23-0.010948>Node47-0.036106>Node76-0.012205>Node98-0.006526>Node99-0.000139>t21                                                           
#>  [16] -0.000000>Node2-0.374243>Node13-1.421149>Node21-0.087352>Node23-0.010948>Node47-0.036106>Node76-0.012205>Node98-0.006526>Node99-0.000139>t43                                                           
#>  [17] -0.000000>Node2-0.374243>Node13-1.421149>Node21-0.087352>Node23-0.010948>Node47-0.036106>Node76-0.012205>Node77-0.000183>t14                                                                           
#>  [18] -0.000000>Node2-0.374243>Node13-1.421149>Node21-0.087352>Node23-0.010948>Node47-0.036106>Node76-0.012205>Node77-0.000183>t42                                                                           
#>  [19] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node23--0.0109->Node36--0.0219->t27                                                                                                           
#>  [20] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node23--0.0109->Node36--0.0219->Node57--0.0222->t37                                                                                           
#>  [21] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node23--0.0109->Node36--0.0219->Node57--0.0222->t45                                                                                           
#>  [22] --0.000-->Node2--0.374-->Node3--0.924-->Node46--0.631-->t88                                                                                                                                            
#>  [23] -0.00000->Node2-0.37424->Node3-0.92390->Node46-0.63121->Node64-0.01005->Node70-0.00116->t85                                                                                                            
#>  [24] -0.00000->Node2-0.37424->Node3-0.92390->Node46-0.63121->Node64-0.01005->Node70-0.00116->t94                                                                                                            
#>  [25] --0.0000->Node2--0.3742->Node3--0.9239->Node46--0.6312->Node64--0.0101->t29                                                                                                                            
#>  [26] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node10-0.16541->Node20-0.11699->Node42-0.04311->Node48-0.00526->Node78-0.01212->t19                                                             
#>  [27] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node10-0.16541->Node20-0.11699->Node42-0.04311->Node48-0.00526->Node78-0.01212->Node87-0.00356->t56                                             
#>  [28] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node10-0.16541->Node20-0.11699->Node42-0.04311->Node48-0.00526->Node78-0.01212->Node87-0.00356->Node90-0.00145->t49                             
#>  [29] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node10-0.16541->Node20-0.11699->Node42-0.04311->Node48-0.00526->Node78-0.01212->Node87-0.00356->Node90-0.00145->t13                             
#>  [30] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node10-0.16541->Node20-0.11699->Node42-0.04311->Node48-0.00526->t12                                                                             
#>  [31] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node20--0.1170->Node42--0.0431->Node65--0.0146->t9                                                                              
#>  [32] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node10-0.16541->Node20-0.11699->Node42-0.04311->Node65-0.01464->Node73-0.00165->t8                                                              
#>  [33] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node10-0.16541->Node20-0.11699->Node42-0.04311->Node65-0.01464->Node73-0.00165->t66                                                             
#>  [34] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node20--0.1170->Node45--0.0462->Node60--0.0106->t4                                                                              
#>  [35] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node10-0.16541->Node20-0.11699->Node45-0.04619->Node60-0.01064->Node72-0.00253->t28                                                             
#>  [36] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node10-0.16541->Node20-0.11699->Node45-0.04619->Node60-0.01064->Node72-0.00253->t17                                                             
#>  [37] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node20--0.1170->Node45--0.0462->t95                                                                                             
#>  [38] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node15--0.0764->Node22--0.0461->Node33--0.0193->Node40--0.0135->t47                                                             
#>  [39] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node15--0.0764->Node22--0.0461->Node33--0.0193->Node40--0.0135->Node82--0.0243->t74                                             
#>  [40] -0.000000>Node2-0.374243>Node3-0.923899>Node4-0.301381>Node10-0.165408>Node15-0.076447>Node22-0.046050>Node33-0.019304>Node40-0.013476>Node82-0.024350>Node85-0.000496>t2                              
#>  [41] -0.000000>Node2-0.374243>Node3-0.923899>Node4-0.301381>Node10-0.165408>Node15-0.076447>Node22-0.046050>Node33-0.019304>Node40-0.013476>Node82-0.024350>Node85-0.000496>t75                             
#>  [42] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node15--0.0764->Node22--0.0461->Node33--0.0193->Node61--0.0323->t53                                                             
#>  [43] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node15--0.0764->Node22--0.0461->Node33--0.0193->Node61--0.0323->t11                                                             
#>  [44] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node15--0.0764->Node22--0.0461->Node79--0.0561->t76                                                                             
#>  [45] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node15--0.0764->Node22--0.0461->Node79--0.0561->t7                                                                              
#>  [46] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node15--0.0764->t100                                                                                                            
#>  [47] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node24-0.11998->Node28-0.00525->Node35-0.01593->Node84-0.02944->t61                
#>  [48] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node24-0.11998->Node28-0.00525->Node35-0.01593->Node84-0.02944->t3                 
#>  [49] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node24-0.11998->Node28-0.00525->Node35-0.01593->Node68-0.02474->t87                
#>  [50] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node24-0.11998->Node28-0.00525->Node35-0.01593->Node68-0.02474->t77                
#>  [51] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node24-0.11998->Node28-0.00525->Node51-0.03555->t70                                
#>  [52] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node24-0.11998->Node28-0.00525->Node51-0.03555->t34                                
#>  [53] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node24-0.11998->Node26-0.00317->t32                                                
#>  [54] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node24-0.11998->Node26-0.00317->t90                                                
#>  [55] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node7--0.0929->Node11--0.0468->Node14--0.0607->t40                                                                
#>  [56] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node14-0.06073->Node44-0.09297->Node49-0.00334->Node62-0.00792->Node96-0.00924->t82
#>  [57] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node14-0.06073->Node44-0.09297->Node49-0.00334->Node62-0.00792->Node96-0.00924->t1 
#>  [58] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node14-0.06073->Node44-0.09297->Node49-0.00334->Node62-0.00792->Node81-0.00481->t84
#>  [59] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node14-0.06073->Node44-0.09297->Node49-0.00334->Node62-0.00792->Node81-0.00481->t68
#>  [60] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node14-0.06073->Node44-0.09297->Node49-0.00334->Node66-0.00843->t97                
#>  [61] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node14-0.06073->Node44-0.09297->Node49-0.00334->Node66-0.00843->Node97-0.00876->t92
#>  [62] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node14-0.06073->Node44-0.09297->Node49-0.00334->Node66-0.00843->Node97-0.00876->t58
#>  [63] -0.000000>Node2-0.374243>Node3-0.923899>Node4-0.301381>Node5-0.019065>Node6-0.015908>Node7-0.092944>Node11-0.046802>Node14-0.060735>Node44-0.092974>Node92-0.019562>Node95-0.000321>t89                
#>  [64] -0.000000>Node2-0.374243>Node3-0.923899>Node4-0.301381>Node5-0.019065>Node6-0.015908>Node7-0.092944>Node11-0.046802>Node14-0.060735>Node44-0.092974>Node92-0.019562>Node95-0.000321>t59                
#>  [65] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node7--0.0929->Node11--0.0468->Node14--0.0607->Node44--0.0930->Node92--0.0196->t86                                
#>  [66] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node7--0.0929->Node18--0.1264->t31                                                                                
#>  [67] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node7--0.0929->Node18--0.1264->t48                                                                                
#>  [68] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node50--0.3003->t35                                                                                               
#>  [69] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node50--0.3003->t25                                                                                               
#>  [70] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node9--0.1261->Node19--0.1305->Node30--0.0304->t44                                                                               
#>  [71] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node9-0.12612->Node19-0.13051->Node30-0.03038->Node32-0.00106->Node56-0.03107->t81                                               
#>  [72] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node9-0.12612->Node19-0.13051->Node30-0.03038->Node32-0.00106->Node56-0.03107->t91                                               
#>  [73] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node9-0.12612->Node19-0.13051->Node30-0.03038->Node32-0.00106->Node59-0.03180->Node74-0.00303->t51                               
#>  [74] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node9-0.12612->Node19-0.13051->Node30-0.03038->Node32-0.00106->Node59-0.03180->Node74-0.00303->t38                               
#>  [75] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node9-0.12612->Node19-0.13051->Node30-0.03038->Node32-0.00106->Node59-0.03180->t52                                               
#>  [76] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node9--0.1261->Node19--0.1305->t63                                                                                               
#>  [77] -0.000000>Node2-0.374243>Node3-0.923899>Node4-0.301381>Node5-0.019065>Node9-0.126122>Node52-0.191405>Node53-0.000706>Node58-0.001428>Node89-0.009056>t69                                               
#>  [78] -0.000000>Node2-0.374243>Node3-0.923899>Node4-0.301381>Node5-0.019065>Node9-0.126122>Node52-0.191405>Node53-0.000706>Node58-0.001428>Node89-0.009056>t96                                               
#>  [79] -0.000000>Node2-0.374243>Node3-0.923899>Node4-0.301381>Node5-0.019065>Node9-0.126122>Node52-0.191405>Node53-0.000706>Node58-0.001428>t67                                                               
#>  [80] -0.000000>Node2-0.374243>Node3-0.923899>Node4-0.301381>Node5-0.019065>Node9-0.126122>Node52-0.191405>Node53-0.000706>Node54-0.000856>t98                                                               
#>  [81] -0.000000>Node2-0.374243>Node3-0.923899>Node4-0.301381>Node5-0.019065>Node9-0.126122>Node52-0.191405>Node53-0.000706>Node54-0.000856>t71                                                               
#>  [82] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node9-0.12612->Node52-0.19141->Node86-0.00976->t78                                                                               
#>  [83] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node9-0.12612->Node52-0.19141->Node86-0.00976->t83                                                                               
#>  [84] --0.0000->Node8--1.7295->Node12--0.0587->Node25--0.1076->t62                                                                                                                                           
#>  [85] --0.0000->Node8--1.7295->Node12--0.0587->Node25--0.1076->t65                                                                                                                                           
#>  [86] -0.00e+00>Node8-1.73e+00>Node12-5.87e-02>Node16-5.70e-02>Node17-7.37e-05>Node63-9.41e-02>t64                                                                                                           
#>  [87] -0.00e+00>Node8-1.73e+00>Node12-5.87e-02>Node16-5.70e-02>Node17-7.37e-05>Node63-9.41e-02>t41                                                                                                           
#>  [88] -0.00e+00>Node8-1.73e+00>Node12-5.87e-02>Node16-5.70e-02>Node17-7.37e-05>Node38-7.13e-02>t60                                                                                                           
#>  [89] -0.00e+00>Node8-1.73e+00>Node12-5.87e-02>Node16-5.70e-02>Node17-7.37e-05>Node38-7.13e-02>Node69-2.39e-02>Node93-7.12e-03>t6                                                                            
#>  [90] -0.00e+00>Node8-1.73e+00>Node12-5.87e-02>Node16-5.70e-02>Node17-7.37e-05>Node38-7.13e-02>Node69-2.39e-02>Node93-7.12e-03>t79                                                                           
#>  [91] -0.00e+00>Node8-1.73e+00>Node12-5.87e-02>Node16-5.70e-02>Node17-7.37e-05>Node38-7.13e-02>Node69-2.39e-02>t50                                                                                           
#>  [92] -0.00000->Node8-1.72948->Node12-0.05872->Node16-0.05699->Node34-0.06425->Node37-0.00668->t26                                                                                                           
#>  [93] -0.00000->Node8-1.72948->Node12-0.05872->Node16-0.05699->Node34-0.06425->Node37-0.00668->Node39-0.00338->t39                                                                                           
#>  [94] -0.00000->Node8-1.72948->Node12-0.05872->Node16-0.05699->Node34-0.06425->Node37-0.00668->Node39-0.00338->Node43-0.00802->t24                                                                           
#>  [95] -0.00000->Node8-1.72948->Node12-0.05872->Node16-0.05699->Node34-0.06425->Node37-0.00668->Node39-0.00338->Node43-0.00802->t57                                                                           
#>  [96] --0.0000->Node8--1.7295->Node12--0.0587->Node16--0.0570->Node34--0.0643->Node88--0.0371->t33                                                                                                           
#>  [97] --0.0000->Node8--1.7295->Node12--0.0587->Node16--0.0570->Node34--0.0643->Node88--0.0371->t46                                                                                                           
#>  [98] --0.000-->Node8--1.729-->Node55--0.208-->t5                                                                                                                                                            
#>  [99] -0.00000->Node8-1.72948->Node55-0.20820->Node71-0.00332->t16                                                                                                                                           
#> [100] -0.00000->Node8-1.72948->Node55-0.20820->Node71-0.00332->t15                                                                                                                                           
#> [101] ----0---->Node2                                                                                                                                                                                        
#> [102] --0.000-->Node2--0.374-->Node3                                                                                                                                                                         
#> [103] --0.000-->Node2--0.374-->Node3--0.924-->Node4                                                                                                                                                          
#> [104] --0.000-->Node2--0.374-->Node3--0.924-->Node4--0.301-->Node5                                                                                                                                           
#> [105] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6                                                                                                                            
#> [106] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node7                                                                                                             
#> [107] ----0---->Node8                                                                                                                                                                                        
#> [108] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node9                                                                                                                            
#> [109] --0.000-->Node2--0.374-->Node3--0.924-->Node4--0.301-->Node10                                                                                                                                          
#> [110] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node7--0.0929->Node11                                                                                             
#> [111] ---0.00-->Node8---1.73-->Node12                                                                                                                                                                        
#> [112] --0.000-->Node2--0.374-->Node13                                                                                                                                                                        
#> [113] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node7--0.0929->Node11--0.0468->Node14                                                                             
#> [114] --0.000-->Node2--0.374-->Node3--0.924-->Node4--0.301-->Node10--0.165-->Node15                                                                                                                          
#> [115] --0.0000->Node8--1.7295->Node12--0.0587->Node16                                                                                                                                                        
#> [116] --0.0000->Node8--1.7295->Node12--0.0587->Node16--0.0570->Node17                                                                                                                                        
#> [117] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node7--0.0929->Node18                                                                                             
#> [118] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node9--0.1261->Node19                                                                                                            
#> [119] --0.000-->Node2--0.374-->Node3--0.924-->Node4--0.301-->Node10--0.165-->Node20                                                                                                                          
#> [120] --0.000-->Node2--0.374-->Node13--1.421-->Node21                                                                                                                                                        
#> [121] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node15--0.0764->Node22                                                                                                          
#> [122] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node23                                                                                                                                        
#> [123] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node7--0.0929->Node11--0.0468->Node24                                                                             
#> [124] --0.0000->Node8--1.7295->Node12--0.0587->Node25                                                                                                                                                        
#> [125] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node7--0.0929->Node11--0.0468->Node24--0.1200->Node26                                                             
#> [126] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node27                                                                                                                                        
#> [127] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node7--0.0929->Node11--0.0468->Node24--0.1200->Node28                                                             
#> [128] --0.000-->Node2--0.374-->Node13--1.421-->Node29                                                                                                                                                        
#> [129] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node9--0.1261->Node19--0.1305->Node30                                                                                            
#> [130] --0.000-->Node2--0.374-->Node13--1.421-->Node29--0.109-->Node31                                                                                                                                        
#> [131] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node9--0.1261->Node19--0.1305->Node30--0.0304->Node32                                                                            
#> [132] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node15--0.0764->Node22--0.0461->Node33                                                                                          
#> [133] --0.0000->Node8--1.7295->Node12--0.0587->Node16--0.0570->Node34                                                                                                                                        
#> [134] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node24-0.11998->Node28-0.00525->Node35                                             
#> [135] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node23--0.0109->Node36                                                                                                                        
#> [136] --0.0000->Node8--1.7295->Node12--0.0587->Node16--0.0570->Node34--0.0643->Node37                                                                                                                        
#> [137] -0.00e+00>Node8-1.73e+00>Node12-5.87e-02>Node16-5.70e-02>Node17-7.37e-05>Node38                                                                                                                        
#> [138] -0.00000->Node8-1.72948->Node12-0.05872->Node16-0.05699->Node34-0.06425->Node37-0.00668->Node39                                                                                                        
#> [139] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node15--0.0764->Node22--0.0461->Node33--0.0193->Node40                                                                          
#> [140] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node27--0.0151->Node41                                                                                                                        
#> [141] --0.000-->Node2--0.374-->Node3--0.924-->Node4--0.301-->Node10--0.165-->Node20--0.117-->Node42                                                                                                          
#> [142] -0.00000->Node8-1.72948->Node12-0.05872->Node16-0.05699->Node34-0.06425->Node37-0.00668->Node39-0.00338->Node43                                                                                        
#> [143] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node7--0.0929->Node11--0.0468->Node14--0.0607->Node44                                                             
#> [144] --0.000-->Node2--0.374-->Node3--0.924-->Node4--0.301-->Node10--0.165-->Node20--0.117-->Node45                                                                                                          
#> [145] --0.000-->Node2--0.374-->Node3--0.924-->Node46                                                                                                                                                         
#> [146] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node23--0.0109->Node47                                                                                                                        
#> [147] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node20--0.1170->Node42--0.0431->Node48                                                                                          
#> [148] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node7--0.0929->Node11--0.0468->Node14--0.0607->Node44--0.0930->Node49                                             
#> [149] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node50                                                                                                            
#> [150] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node24-0.11998->Node28-0.00525->Node51                                             
#> [151] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node9--0.1261->Node52                                                                                                            
#> [152] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node9--0.1261->Node52--0.1914->Node53                                                                                            
#> [153] -0.000000>Node2-0.374243>Node3-0.923899>Node4-0.301381>Node5-0.019065>Node9-0.126122>Node52-0.191405>Node53-0.000706>Node54                                                                            
#> [154] ---0.00-->Node8---1.73-->Node55                                                                                                                                                                        
#> [155] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node9-0.12612->Node19-0.13051->Node30-0.03038->Node32-0.00106->Node56                                                            
#> [156] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node23--0.0109->Node36--0.0219->Node57                                                                                                        
#> [157] -0.000000>Node2-0.374243>Node3-0.923899>Node4-0.301381>Node5-0.019065>Node9-0.126122>Node52-0.191405>Node53-0.000706>Node58                                                                            
#> [158] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node9-0.12612->Node19-0.13051->Node30-0.03038->Node32-0.00106->Node59                                                            
#> [159] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node20--0.1170->Node45--0.0462->Node60                                                                                          
#> [160] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node15--0.0764->Node22--0.0461->Node33--0.0193->Node61                                                                          
#> [161] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node14-0.06073->Node44-0.09297->Node49-0.00334->Node62                             
#> [162] -0.00e+00>Node8-1.73e+00>Node12-5.87e-02>Node16-5.70e-02>Node17-7.37e-05>Node63                                                                                                                        
#> [163] --0.000-->Node2--0.374-->Node3--0.924-->Node46--0.631-->Node64                                                                                                                                         
#> [164] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node20--0.1170->Node42--0.0431->Node65                                                                                          
#> [165] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node14-0.06073->Node44-0.09297->Node49-0.00334->Node66                             
#> [166] -0.00000->Node2-0.37424->Node13-1.42115->Node29-0.10919->Node31-0.00168->Node67                                                                                                                        
#> [167] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node24-0.11998->Node28-0.00525->Node35-0.01593->Node68                             
#> [168] -0.00e+00>Node8-1.73e+00>Node12-5.87e-02>Node16-5.70e-02>Node17-7.37e-05>Node38-7.13e-02>Node69                                                                                                        
#> [169] --0.0000->Node2--0.3742->Node3--0.9239->Node46--0.6312->Node64--0.0101->Node70                                                                                                                         
#> [170] --0.000-->Node8--1.729-->Node55--0.208-->Node71                                                                                                                                                        
#> [171] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node20--0.1170->Node45--0.0462->Node60--0.0106->Node72                                                                          
#> [172] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node20--0.1170->Node42--0.0431->Node65--0.0146->Node73                                                                          
#> [173] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node9-0.12612->Node19-0.13051->Node30-0.03038->Node32-0.00106->Node59-0.03180->Node74                                            
#> [174] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node23--0.0109->Node47--0.0361->Node75                                                                                                        
#> [175] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node23--0.0109->Node47--0.0361->Node76                                                                                                        
#> [176] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node23--0.0109->Node47--0.0361->Node76--0.0122->Node77                                                                                        
#> [177] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node10-0.16541->Node20-0.11699->Node42-0.04311->Node48-0.00526->Node78                                                                          
#> [178] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node15--0.0764->Node22--0.0461->Node79                                                                                          
#> [179] -0.00000->Node2-0.37424->Node13-1.42115->Node29-0.10919->Node31-0.00168->Node80                                                                                                                        
#> [180] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node14-0.06073->Node44-0.09297->Node49-0.00334->Node62-0.00792->Node81             
#> [181] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node15--0.0764->Node22--0.0461->Node33--0.0193->Node40--0.0135->Node82                                                          
#> [182] -0.00000->Node2-0.37424->Node13-1.42115->Node29-0.10919->Node31-0.00168->Node80-0.03726->Node83                                                                                                        
#> [183] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node24-0.11998->Node28-0.00525->Node35-0.01593->Node84                             
#> [184] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node10--0.1654->Node15--0.0764->Node22--0.0461->Node33--0.0193->Node40--0.0135->Node82--0.0243->Node85                                          
#> [185] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node9--0.1261->Node52--0.1914->Node86                                                                                            
#> [186] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node10-0.16541->Node20-0.11699->Node42-0.04311->Node48-0.00526->Node78-0.01212->Node87                                                          
#> [187] --0.0000->Node8--1.7295->Node12--0.0587->Node16--0.0570->Node34--0.0643->Node88                                                                                                                        
#> [188] -0.000000>Node2-0.374243>Node3-0.923899>Node4-0.301381>Node5-0.019065>Node9-0.126122>Node52-0.191405>Node53-0.000706>Node58-0.001428>Node89                                                            
#> [189] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node10-0.16541->Node20-0.11699->Node42-0.04311->Node48-0.00526->Node78-0.01212->Node87-0.00356->Node90                                          
#> [190] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node23--0.0109->Node47--0.0361->Node75--0.0120->Node91                                                                                        
#> [191] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node7--0.0929->Node11--0.0468->Node14--0.0607->Node44--0.0930->Node92                                             
#> [192] -0.00e+00>Node8-1.73e+00>Node12-5.87e-02>Node16-5.70e-02>Node17-7.37e-05>Node38-7.13e-02>Node69-2.39e-02>Node93                                                                                        
#> [193] -0.00000->Node2-0.37424->Node13-1.42115->Node29-0.10919->Node31-0.00168->Node80-0.03726->Node83-0.00131->Node94                                                                                        
#> [194] --0.0000->Node2--0.3742->Node3--0.9239->Node4--0.3014->Node5--0.0191->Node6--0.0159->Node7--0.0929->Node11--0.0468->Node14--0.0607->Node44--0.0930->Node92--0.0196->Node95                             
#> [195] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node14-0.06073->Node44-0.09297->Node49-0.00334->Node62-0.00792->Node96             
#> [196] -0.00000->Node2-0.37424->Node3-0.92390->Node4-0.30138->Node5-0.01906->Node6-0.01591->Node7-0.09294->Node11-0.04680->Node14-0.06073->Node44-0.09297->Node49-0.00334->Node66-0.00843->Node97             
#> [197] --0.0000->Node2--0.3742->Node13--1.4211->Node21--0.0874->Node23--0.0109->Node47--0.0361->Node76--0.0122->Node98                                                                                        
#> [198] -0.00000->Node2-0.37424->Node13-1.42115->Node21-0.08735->Node23-0.01095->Node47-0.03611->Node76-0.01220->Node98-0.00653->Node99