Return a pfc with ancestral features

Return a pfc with ancestral features

Usage

pf_anc(x, replace = 0)

Arguments

x

A pfc object

replace

Value to replace edge feature with no ancestral value (such as when the ancestor is the root)

Value

A new pfc with the same structure as x, but with the features of each edge's ancestors instead

Examples

pf_anc(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node28-0.04340->Node32-0.01212->Node35-0.00847->t69                                                                             
#>   [2] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node28-0.04340->Node32-0.01212->Node35-0.00847->Node80-0.03371->t72                                                             
#>   [3] -0.000000>Node2-0.299340>Node6-1.092459>Node13-0.130436>Node16-0.025368>Node28-0.043401>Node32-0.012123>Node35-0.008467>Node80-0.033713>Node85-0.001197>Node86-0.000329>t78                             
#>   [4] -0.000000>Node2-0.299340>Node6-1.092459>Node13-0.130436>Node16-0.025368>Node28-0.043401>Node32-0.012123>Node35-0.008467>Node80-0.033713>Node85-0.001197>Node86-0.000329>t18                             
#>   [5] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node28-0.04340->Node32-0.01212->Node35-0.00847->Node80-0.03371->Node85-0.00120->t6                                              
#>   [6] --0.0000->Node2--0.2993->Node6--1.0925->Node13--0.1304->Node16--0.0254->Node28--0.0434->Node32--0.0121->t35                                                                                             
#>   [7] --0.0000->Node2--0.2993->Node6--1.0925->Node13--0.1304->Node16--0.0254->Node28--0.0434->t73                                                                                                             
#>   [8] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node44-0.05416->Node54-0.01175->t28                                                             
#>   [9] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node44-0.05416->Node54-0.01175->Node96-0.01375->t71                                             
#>  [10] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node44-0.05416->Node54-0.01175->Node96-0.01375->t31                                             
#>  [11] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node44-0.05416->Node52-0.01074->Node77-0.01139->t46                                             
#>  [12] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node44-0.05416->Node52-0.01074->Node77-0.01139->t32                                             
#>  [13] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node44-0.05416->Node52-0.01074->Node68-0.00926->Node91-0.00459->t88                             
#>  [14] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node44-0.05416->Node52-0.01074->Node68-0.00926->Node91-0.00459->t37                             
#>  [15] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node44-0.05416->Node52-0.01074->Node68-0.00926->t34                                             
#>  [16] -0.00e+00>Node2-2.99e-01>Node6-1.09e+00>Node13-1.30e-01>Node16-2.54e-02>Node22-1.89e-02>Node23-2.02e-03>Node24-1.03e-02>Node27-1.11e-02>Node93-5.82e-02>Node94-1.22e-05>t14                             
#>  [17] -0.00e+00>Node2-2.99e-01>Node6-1.09e+00>Node13-1.30e-01>Node16-2.54e-02>Node22-1.89e-02>Node23-2.02e-03>Node24-1.03e-02>Node27-1.11e-02>Node93-5.82e-02>Node94-1.22e-05>t39                             
#>  [18] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node24-0.01028->Node27-0.01108->Node93-0.05820->t22                                             
#>  [19] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node24-0.01028->Node27-0.01108->t43                                                             
#>  [20] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node24-0.01028->t11                                                                             
#>  [21] --0.0000->Node2--0.2993->Node6--1.0925->Node13--0.1304->Node16--0.0254->Node22--0.0189->Node95--0.0816->t84                                                                                             
#>  [22] --0.0000->Node2--0.2993->Node6--1.0925->Node13--0.1304->Node16--0.0254->Node22--0.0189->Node95--0.0816->t26                                                                                             
#>  [23] --0.0000->Node2--0.2993->Node6--1.0925->Node13--0.1304->Node20--0.0390->Node92--0.0868->t7                                                                                                              
#>  [24] --0.0000->Node2--0.2993->Node6--1.0925->Node13--0.1304->Node20--0.0390->Node92--0.0868->t33                                                                                                             
#>  [25] --0.000-->Node2--0.299-->Node6--1.092-->Node13--0.130-->Node20--0.039-->t17                                                                                                                             
#>  [26] --0.000-->Node2--0.299-->Node6--1.092-->t92                                                                                                                                                             
#>  [27] --0.000-->Node2--0.299-->Node3--0.734-->t48                                                                                                                                                             
#>  [28] -0.00000->Node2-0.29934->Node3-0.73447->Node45-0.59138->Node56-0.00995->Node74-0.00871->t64                                                                                                             
#>  [29] -0.00000->Node2-0.29934->Node3-0.73447->Node45-0.59138->Node56-0.00995->Node74-0.00871->t59                                                                                                             
#>  [30] -0.00000->Node2-0.29934->Node3-0.73447->Node45-0.59138->Node56-0.00995->t82                                                                                                                             
#>  [31] --0.000-->Node2--0.299-->Node3--0.734-->Node45--0.591-->t55                                                                                                                                             
#>  [32] --0.0000->Node4--1.0360->Node5--0.2861->Node31--0.2798->Node90--0.0453->t53                                                                                                                             
#>  [33] --0.0000->Node4--1.0360->Node5--0.2861->Node31--0.2798->Node90--0.0453->t79                                                                                                                             
#>  [34] --0.0000->Node4--1.0360->Node5--0.2861->Node31--0.2798->Node36--0.0121->t36                                                                                                                             
#>  [35] --0.0000->Node4--1.0360->Node5--0.2861->Node31--0.2798->Node36--0.0121->t87                                                                                                                             
#>  [36] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node15--0.1380->Node18--0.0148->Node25--0.0285->Node33--0.0229->Node41--0.0122->t21                                                              
#>  [37] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node15-0.13805->Node18-0.01477->Node25-0.02854->Node33-0.02295->Node41-0.01216->Node57-0.01500->Node71-0.00719->t98                              
#>  [38] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node15-0.13805->Node18-0.01477->Node25-0.02854->Node33-0.02295->Node41-0.01216->Node57-0.01500->Node71-0.00719->t29                              
#>  [39] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node15--0.1380->Node18--0.0148->Node25--0.0285->Node33--0.0229->Node41--0.0122->Node57--0.0150->t24                                              
#>  [40] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node15-0.13805->Node18-0.01477->Node25-0.02854->Node33-0.02295->Node59-0.02927->Node88-0.00905->t63                                              
#>  [41] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node15-0.13805->Node18-0.01477->Node25-0.02854->Node33-0.02295->Node59-0.02927->Node88-0.00905->t40                                              
#>  [42] -0.000000>Node4-1.035983>Node5-0.286098>Node7-0.082261>Node15-0.138046>Node18-0.014765>Node25-0.028541>Node33-0.022947>Node59-0.029269>Node60-0.000491>t56                                              
#>  [43] -0.000000>Node4-1.035983>Node5-0.286098>Node7-0.082261>Node15-0.138046>Node18-0.014765>Node25-0.028541>Node33-0.022947>Node59-0.029269>Node60-0.000491>Node63-0.002355>t70                              
#>  [44] -0.000000>Node4-1.035983>Node5-0.286098>Node7-0.082261>Node15-0.138046>Node18-0.014765>Node25-0.028541>Node33-0.022947>Node59-0.029269>Node60-0.000491>Node63-0.002355>Node72-0.002517>t77              
#>  [45] -0.000000>Node4-1.035983>Node5-0.286098>Node7-0.082261>Node15-0.138046>Node18-0.014765>Node25-0.028541>Node33-0.022947>Node59-0.029269>Node60-0.000491>Node63-0.002355>Node72-0.002517>t86              
#>  [46] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node15-0.13805->Node18-0.01477->Node25-0.02854->Node29-0.00762->t5                                                                               
#>  [47] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node15-0.13805->Node18-0.01477->Node25-0.02854->Node29-0.00762->Node50-0.03791->t89                                                              
#>  [48] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node15-0.13805->Node18-0.01477->Node25-0.02854->Node29-0.00762->Node50-0.03791->t95                                                              
#>  [49] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node15-0.13805->Node18-0.01477->Node21-0.00902->Node40-0.05426->t9                                                                               
#>  [50] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node15-0.13805->Node18-0.01477->Node21-0.00902->Node40-0.05426->t44                                                                              
#>  [51] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node15-0.13805->Node18-0.01477->Node21-0.00902->t91                                                                                              
#>  [52] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node15--0.1380->t67                                                                                                                              
#>  [53] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node43--0.2045->Node79--0.0236->t27                                                                                               
#>  [54] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node8-0.01281->Node43-0.20453->Node79-0.02360->Node87-0.00161->t65                                                                               
#>  [55] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node8-0.01281->Node43-0.20453->Node79-0.02360->Node87-0.00161->t23                                                                               
#>  [56] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node43--0.2045->t96                                                                                                               
#>  [57] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node11--0.0116->Node34--0.1034->t100                                                               
#>  [58] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node11--0.0116->Node34--0.1034->t47                                                                
#>  [59] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node11--0.0116->Node17--0.0439->Node38--0.0680->t83                                                
#>  [60] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node8-0.01281->Node9-0.03124->Node10-0.04561->Node11-0.01155->Node17-0.04390->Node38-0.06796->Node46-0.00825->Node64-0.01551->t51                
#>  [61] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node8-0.01281->Node9-0.03124->Node10-0.04561->Node11-0.01155->Node17-0.04390->Node38-0.06796->Node46-0.00825->Node64-0.01551->t80                
#>  [62] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node8-0.01281->Node9-0.03124->Node10-0.04561->Node11-0.01155->Node17-0.04390->Node38-0.06796->Node46-0.00825->t99                                
#>  [63] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node11--0.0116->Node17--0.0439->Node58--0.0870->t76                                                
#>  [64] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node8-0.01281->Node9-0.03124->Node10-0.04561->Node11-0.01155->Node17-0.04390->Node58-0.08703->Node73-0.00724->Node82-0.00239->t15                
#>  [65] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node8-0.01281->Node9-0.03124->Node10-0.04561->Node11-0.01155->Node17-0.04390->Node58-0.08703->Node73-0.00724->Node82-0.00239->t66                
#>  [66] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node8-0.01281->Node9-0.03124->Node10-0.04561->Node11-0.01155->Node17-0.04390->Node58-0.08703->Node73-0.00724->Node98-0.00460->t75                
#>  [67] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node8-0.01281->Node9-0.03124->Node10-0.04561->Node11-0.01155->Node17-0.04390->Node58-0.08703->Node73-0.00724->Node98-0.00460->t42                
#>  [68] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node14--0.0350->Node47--0.0978->Node83--0.0193->t97                                                
#>  [69] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node8-0.01281->Node9-0.03124->Node10-0.04561->Node14-0.03504->Node47-0.09776->Node83-0.01934->Node97-0.00214->t85                                
#>  [70] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node8-0.01281->Node9-0.03124->Node10-0.04561->Node14-0.03504->Node47-0.09776->Node83-0.01934->Node97-0.00214->t49                                
#>  [71] -0.000000>Node4-1.035983>Node5-0.286098>Node7-0.082261>Node8-0.012805>Node9-0.031235>Node10-0.045610>Node14-0.035040>Node47-0.097763>Node48-0.000772>Node55-0.007372>t68                                
#>  [72] -0.000000>Node4-1.035983>Node5-0.286098>Node7-0.082261>Node8-0.012805>Node9-0.031235>Node10-0.045610>Node14-0.035040>Node47-0.097763>Node48-0.000772>Node55-0.007372>t13                                
#>  [73] -0.000000>Node4-1.035983>Node5-0.286098>Node7-0.082261>Node8-0.012805>Node9-0.031235>Node10-0.045610>Node14-0.035040>Node47-0.097763>Node48-0.000772>Node69-0.015123>Node75-0.001624>Node89-0.002795>t61
#>  [74] -0.000000>Node4-1.035983>Node5-0.286098>Node7-0.082261>Node8-0.012805>Node9-0.031235>Node10-0.045610>Node14-0.035040>Node47-0.097763>Node48-0.000772>Node69-0.015123>Node75-0.001624>Node89-0.002795>t60
#>  [75] -0.000000>Node4-1.035983>Node5-0.286098>Node7-0.082261>Node8-0.012805>Node9-0.031235>Node10-0.045610>Node14-0.035040>Node47-0.097763>Node48-0.000772>Node69-0.015123>Node75-0.001624>t90                
#>  [76] -0.000000>Node4-1.035983>Node5-0.286098>Node7-0.082261>Node8-0.012805>Node9-0.031235>Node10-0.045610>Node14-0.035040>Node47-0.097763>Node48-0.000772>Node69-0.015123>t2                                 
#>  [77] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node14--0.0350->t8                                                                                 
#>  [78] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->t62                                                                                                                
#>  [79] -0.00000->Node4-1.03598->Node12-0.48367->Node26-0.06685->Node39-0.03272->Node51-0.01250->Node76-0.01284->Node99-0.00393->t54                                                                            
#>  [80] -0.00000->Node4-1.03598->Node12-0.48367->Node26-0.06685->Node39-0.03272->Node51-0.01250->Node76-0.01284->Node99-0.00393->t16                                                                            
#>  [81] --0.0000->Node4--1.0360->Node12--0.4837->Node26--0.0668->Node39--0.0327->Node51--0.0125->Node76--0.0128->t3                                                                                             
#>  [82] --0.0000->Node4--1.0360->Node12--0.4837->Node26--0.0668->Node39--0.0327->Node51--0.0125->Node84--0.0146->t93                                                                                            
#>  [83] --0.0000->Node4--1.0360->Node12--0.4837->Node26--0.0668->Node39--0.0327->Node51--0.0125->Node84--0.0146->t81                                                                                            
#>  [84] --0.0000->Node4--1.0360->Node12--0.4837->Node26--0.0668->Node39--0.0327->t19                                                                                                                            
#>  [85] -0.00000->Node4-1.03598->Node12-0.48367->Node26-0.06685->Node30-0.00798->Node65-0.04676->t25                                                                                                            
#>  [86] -0.00000->Node4-1.03598->Node12-0.48367->Node26-0.06685->Node30-0.00798->Node65-0.04676->t74                                                                                                            
#>  [87] -0.00000->Node4-1.03598->Node12-0.48367->Node26-0.06685->Node30-0.00798->Node37-0.02230->t94                                                                                                            
#>  [88] -0.00000->Node4-1.03598->Node12-0.48367->Node26-0.06685->Node30-0.00798->Node37-0.02230->Node49-0.01116->Node70-0.01484->t57                                                                            
#>  [89] -0.00000->Node4-1.03598->Node12-0.48367->Node26-0.06685->Node30-0.00798->Node37-0.02230->Node49-0.01116->Node70-0.01484->Node78-0.00232->t12                                                            
#>  [90] -0.00000->Node4-1.03598->Node12-0.48367->Node26-0.06685->Node30-0.00798->Node37-0.02230->Node49-0.01116->Node70-0.01484->Node78-0.00232->t41                                                            
#>  [91] -0.00000->Node4-1.03598->Node12-0.48367->Node26-0.06685->Node30-0.00798->Node37-0.02230->Node49-0.01116->Node61-0.01234->t1                                                                             
#>  [92] -0.000000>Node4-1.035983>Node12-0.483671>Node26-0.066849>Node30-0.007977>Node37-0.022295>Node49-0.011156>Node61-0.012335>Node62-0.000359>Node67-0.001140>t38                                            
#>  [93] -0.000000>Node4-1.035983>Node12-0.483671>Node26-0.066849>Node30-0.007977>Node37-0.022295>Node49-0.011156>Node61-0.012335>Node62-0.000359>Node67-0.001140>t20                                            
#>  [94] -0.000000>Node4-1.035983>Node12-0.483671>Node26-0.066849>Node30-0.007977>Node37-0.022295>Node49-0.011156>Node61-0.012335>Node62-0.000359>Node81-0.005468>t30                                            
#>  [95] -0.000000>Node4-1.035983>Node12-0.483671>Node26-0.066849>Node30-0.007977>Node37-0.022295>Node49-0.011156>Node61-0.012335>Node62-0.000359>Node81-0.005468>t4                                             
#>  [96] --0.0000->Node4--1.0360->Node12--0.4837->Node19--0.0395->Node42--0.0624->t10                                                                                                                            
#>  [97] -0.00000->Node4-1.03598->Node12-0.48367->Node19-0.03952->Node42-0.06241->Node53-0.01201->Node66-0.00794->t58                                                                                            
#>  [98] -0.00000->Node4-1.03598->Node12-0.48367->Node19-0.03952->Node42-0.06241->Node53-0.01201->Node66-0.00794->t50                                                                                            
#>  [99] --0.0000->Node4--1.0360->Node12--0.4837->Node19--0.0395->Node42--0.0624->Node53--0.0120->t52                                                                                                            
#> [100] --0.0000->Node4--1.0360->Node12--0.4837->Node19--0.0395->t45                                                                                                                                            
#> [101] ----0---->Node2                                                                                                                                                                                         
#> [102] --0.000-->Node2--0.299-->Node3                                                                                                                                                                          
#> [103] ----0---->Node4                                                                                                                                                                                         
#> [104] ---0.00-->Node4---1.04-->Node5                                                                                                                                                                          
#> [105] --0.000-->Node2--0.299-->Node6                                                                                                                                                                          
#> [106] --0.000-->Node4--1.036-->Node5--0.286-->Node7                                                                                                                                                           
#> [107] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8                                                                                                                                            
#> [108] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9                                                                                                                             
#> [109] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10                                                                                                             
#> [110] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node11                                                                                             
#> [111] ---0.00-->Node4---1.04-->Node12                                                                                                                                                                         
#> [112] --0.000-->Node2--0.299-->Node6--1.092-->Node13                                                                                                                                                          
#> [113] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node14                                                                                             
#> [114] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node15                                                                                                                                           
#> [115] --0.000-->Node2--0.299-->Node6--1.092-->Node13--0.130-->Node16                                                                                                                                          
#> [116] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node11--0.0116->Node17                                                                             
#> [117] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node15--0.1380->Node18                                                                                                                           
#> [118] --0.000-->Node4--1.036-->Node12--0.484-->Node19                                                                                                                                                         
#> [119] --0.000-->Node2--0.299-->Node6--1.092-->Node13--0.130-->Node20                                                                                                                                          
#> [120] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node15--0.1380->Node18--0.0148->Node21                                                                                                           
#> [121] --0.0000->Node2--0.2993->Node6--1.0925->Node13--0.1304->Node16--0.0254->Node22                                                                                                                          
#> [122] --0.0000->Node2--0.2993->Node6--1.0925->Node13--0.1304->Node16--0.0254->Node22--0.0189->Node23                                                                                                          
#> [123] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node24                                                                                          
#> [124] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node15--0.1380->Node18--0.0148->Node25                                                                                                           
#> [125] --0.000-->Node4--1.036-->Node12--0.484-->Node26                                                                                                                                                         
#> [126] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node24-0.01028->Node27                                                                          
#> [127] --0.0000->Node2--0.2993->Node6--1.0925->Node13--0.1304->Node16--0.0254->Node28                                                                                                                          
#> [128] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node15--0.1380->Node18--0.0148->Node25--0.0285->Node29                                                                                           
#> [129] --0.0000->Node4--1.0360->Node12--0.4837->Node26--0.0668->Node30                                                                                                                                         
#> [130] --0.000-->Node4--1.036-->Node5--0.286-->Node31                                                                                                                                                          
#> [131] --0.0000->Node2--0.2993->Node6--1.0925->Node13--0.1304->Node16--0.0254->Node28--0.0434->Node32                                                                                                          
#> [132] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node15--0.1380->Node18--0.0148->Node25--0.0285->Node33                                                                                           
#> [133] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node11--0.0116->Node34                                                                             
#> [134] --0.0000->Node2--0.2993->Node6--1.0925->Node13--0.1304->Node16--0.0254->Node28--0.0434->Node32--0.0121->Node35                                                                                          
#> [135] --0.000-->Node4--1.036-->Node5--0.286-->Node31--0.280-->Node36                                                                                                                                          
#> [136] -0.00000->Node4-1.03598->Node12-0.48367->Node26-0.06685->Node30-0.00798->Node37                                                                                                                         
#> [137] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node11--0.0116->Node17--0.0439->Node38                                                             
#> [138] --0.0000->Node4--1.0360->Node12--0.4837->Node26--0.0668->Node39                                                                                                                                         
#> [139] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node15-0.13805->Node18-0.01477->Node21-0.00902->Node40                                                                                           
#> [140] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node15--0.1380->Node18--0.0148->Node25--0.0285->Node33--0.0229->Node41                                                                           
#> [141] --0.0000->Node4--1.0360->Node12--0.4837->Node19--0.0395->Node42                                                                                                                                         
#> [142] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node43                                                                                                                            
#> [143] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node44                                                                                          
#> [144] --0.000-->Node2--0.299-->Node3--0.734-->Node45                                                                                                                                                          
#> [145] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node11--0.0116->Node17--0.0439->Node38--0.0680->Node46                                             
#> [146] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node14--0.0350->Node47                                                                             
#> [147] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node14--0.0350->Node47--0.0978->Node48                                                             
#> [148] -0.00000->Node4-1.03598->Node12-0.48367->Node26-0.06685->Node30-0.00798->Node37-0.02230->Node49                                                                                                         
#> [149] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node15-0.13805->Node18-0.01477->Node25-0.02854->Node29-0.00762->Node50                                                                           
#> [150] --0.0000->Node4--1.0360->Node12--0.4837->Node26--0.0668->Node39--0.0327->Node51                                                                                                                         
#> [151] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node44-0.05416->Node52                                                                          
#> [152] --0.0000->Node4--1.0360->Node12--0.4837->Node19--0.0395->Node42--0.0624->Node53                                                                                                                         
#> [153] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node44-0.05416->Node54                                                                          
#> [154] -0.000000>Node4-1.035983>Node5-0.286098>Node7-0.082261>Node8-0.012805>Node9-0.031235>Node10-0.045610>Node14-0.035040>Node47-0.097763>Node48-0.000772>Node55                                             
#> [155] --0.000-->Node2--0.299-->Node3--0.734-->Node45--0.591-->Node56                                                                                                                                          
#> [156] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node15--0.1380->Node18--0.0148->Node25--0.0285->Node33--0.0229->Node41--0.0122->Node57                                                           
#> [157] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node11--0.0116->Node17--0.0439->Node58                                                             
#> [158] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node15--0.1380->Node18--0.0148->Node25--0.0285->Node33--0.0229->Node59                                                                           
#> [159] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node15--0.1380->Node18--0.0148->Node25--0.0285->Node33--0.0229->Node59--0.0293->Node60                                                           
#> [160] -0.00000->Node4-1.03598->Node12-0.48367->Node26-0.06685->Node30-0.00798->Node37-0.02230->Node49-0.01116->Node61                                                                                         
#> [161] -0.00000->Node4-1.03598->Node12-0.48367->Node26-0.06685->Node30-0.00798->Node37-0.02230->Node49-0.01116->Node61-0.01234->Node62                                                                         
#> [162] -0.000000>Node4-1.035983>Node5-0.286098>Node7-0.082261>Node15-0.138046>Node18-0.014765>Node25-0.028541>Node33-0.022947>Node59-0.029269>Node60-0.000491>Node63                                           
#> [163] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node8-0.01281->Node9-0.03124->Node10-0.04561->Node11-0.01155->Node17-0.04390->Node38-0.06796->Node46-0.00825->Node64                             
#> [164] -0.00000->Node4-1.03598->Node12-0.48367->Node26-0.06685->Node30-0.00798->Node65                                                                                                                         
#> [165] --0.0000->Node4--1.0360->Node12--0.4837->Node19--0.0395->Node42--0.0624->Node53--0.0120->Node66                                                                                                         
#> [166] -0.000000>Node4-1.035983>Node12-0.483671>Node26-0.066849>Node30-0.007977>Node37-0.022295>Node49-0.011156>Node61-0.012335>Node62-0.000359>Node67                                                         
#> [167] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node44-0.05416->Node52-0.01074->Node68                                                          
#> [168] -0.000000>Node4-1.035983>Node5-0.286098>Node7-0.082261>Node8-0.012805>Node9-0.031235>Node10-0.045610>Node14-0.035040>Node47-0.097763>Node48-0.000772>Node69                                             
#> [169] -0.00000->Node4-1.03598->Node12-0.48367->Node26-0.06685->Node30-0.00798->Node37-0.02230->Node49-0.01116->Node70                                                                                         
#> [170] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node15--0.1380->Node18--0.0148->Node25--0.0285->Node33--0.0229->Node41--0.0122->Node57--0.0150->Node71                                           
#> [171] -0.000000>Node4-1.035983>Node5-0.286098>Node7-0.082261>Node15-0.138046>Node18-0.014765>Node25-0.028541>Node33-0.022947>Node59-0.029269>Node60-0.000491>Node63-0.002355>Node72                           
#> [172] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node11--0.0116->Node17--0.0439->Node58--0.0870->Node73                                             
#> [173] -0.00000->Node2-0.29934->Node3-0.73447->Node45-0.59138->Node56-0.00995->Node74                                                                                                                          
#> [174] -0.000000>Node4-1.035983>Node5-0.286098>Node7-0.082261>Node8-0.012805>Node9-0.031235>Node10-0.045610>Node14-0.035040>Node47-0.097763>Node48-0.000772>Node69-0.015123>Node75                             
#> [175] --0.0000->Node4--1.0360->Node12--0.4837->Node26--0.0668->Node39--0.0327->Node51--0.0125->Node76                                                                                                         
#> [176] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node44-0.05416->Node52-0.01074->Node77                                                          
#> [177] -0.00000->Node4-1.03598->Node12-0.48367->Node26-0.06685->Node30-0.00798->Node37-0.02230->Node49-0.01116->Node70-0.01484->Node78                                                                         
#> [178] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node43--0.2045->Node79                                                                                                            
#> [179] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node28-0.04340->Node32-0.01212->Node35-0.00847->Node80                                                                          
#> [180] -0.000000>Node4-1.035983>Node12-0.483671>Node26-0.066849>Node30-0.007977>Node37-0.022295>Node49-0.011156>Node61-0.012335>Node62-0.000359>Node81                                                         
#> [181] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node8-0.01281->Node9-0.03124->Node10-0.04561->Node11-0.01155->Node17-0.04390->Node58-0.08703->Node73-0.00724->Node82                             
#> [182] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node14--0.0350->Node47--0.0978->Node83                                                             
#> [183] --0.0000->Node4--1.0360->Node12--0.4837->Node26--0.0668->Node39--0.0327->Node51--0.0125->Node84                                                                                                         
#> [184] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node28-0.04340->Node32-0.01212->Node35-0.00847->Node80-0.03371->Node85                                                          
#> [185] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node28-0.04340->Node32-0.01212->Node35-0.00847->Node80-0.03371->Node85-0.00120->Node86                                          
#> [186] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node43--0.2045->Node79--0.0236->Node87                                                                                            
#> [187] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node15--0.1380->Node18--0.0148->Node25--0.0285->Node33--0.0229->Node59--0.0293->Node88                                                           
#> [188] -0.000000>Node4-1.035983>Node5-0.286098>Node7-0.082261>Node8-0.012805>Node9-0.031235>Node10-0.045610>Node14-0.035040>Node47-0.097763>Node48-0.000772>Node69-0.015123>Node75-0.001624>Node89             
#> [189] --0.000-->Node4--1.036-->Node5--0.286-->Node31--0.280-->Node90                                                                                                                                          
#> [190] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node44-0.05416->Node52-0.01074->Node68-0.00926->Node91                                          
#> [191] --0.000-->Node2--0.299-->Node6--1.092-->Node13--0.130-->Node20--0.039-->Node92                                                                                                                          
#> [192] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node24-0.01028->Node27-0.01108->Node93                                                          
#> [193] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node24-0.01028->Node27-0.01108->Node93-0.05820->Node94                                          
#> [194] --0.0000->Node2--0.2993->Node6--1.0925->Node13--0.1304->Node16--0.0254->Node22--0.0189->Node95                                                                                                          
#> [195] -0.00000->Node2-0.29934->Node6-1.09246->Node13-0.13044->Node16-0.02537->Node22-0.01886->Node23-0.00202->Node44-0.05416->Node54-0.01175->Node96                                                          
#> [196] --0.0000->Node4--1.0360->Node5--0.2861->Node7--0.0823->Node8--0.0128->Node9--0.0312->Node10--0.0456->Node14--0.0350->Node47--0.0978->Node83--0.0193->Node97                                             
#> [197] -0.00000->Node4-1.03598->Node5-0.28610->Node7-0.08226->Node8-0.01281->Node9-0.03124->Node10-0.04561->Node11-0.01155->Node17-0.04390->Node58-0.08703->Node73-0.00724->Node98                             
#> [198] --0.0000->Node4--1.0360->Node12--0.4837->Node26--0.0668->Node39--0.0327->Node51--0.0125->Node76--0.0128->Node99