Return a pfc with ancestral features

Return a pfc with ancestral features

Usage

pf_anc(x, replace = 0)

Arguments

x

A pfc object

replace

Value to replace edge feature with no ancestral value (such as when the ancestor is the root)

Value

A new pfc with the same structure as x, but with the features of each edge's ancestors instead

Examples

pf_anc(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node12--0.0695->Node18--0.0530->t63                                                                                                                
#>   [2] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node12--0.0695->Node18--0.0530->Node43--0.0785->t34                                                                                                
#>   [3] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node12--0.0695->Node18--0.0530->Node43--0.0785->t31                                                                                                
#>   [4] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node12--0.0695->Node95--0.1599->t8                                                                                                                 
#>   [5] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node12--0.0695->Node95--0.1599->t46                                                                                                                
#>   [6] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node11--0.0554->Node19--0.0789->Node37--0.0578->t25                                                                                                
#>   [7] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node11--0.0554->Node19--0.0789->Node37--0.0578->t62                                                                                                
#>   [8] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>t65                                                                                
#>   [9] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node90-0.051361>Node91-0.000307>t37                                
#>  [10] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node90-0.051361>Node91-0.000307>t11                                
#>  [11] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node90-0.051361>t71                                                
#>  [12] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40-0.020440>Node67-0.021053>Node78-0.005301>t89                
#>  [13] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40-0.020440>Node67-0.021053>Node78-0.005301>t22                
#>  [14] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40-0.020440>Node67-0.021053>t97                                
#>  [15] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40-0.020440>Node63-0.019601>Node76-0.006286>Node79-0.001044>t32
#>  [16] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40-0.020440>Node63-0.019601>Node76-0.006286>Node79-0.001044>t66
#>  [17] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40-0.020440>Node63-0.019601>Node76-0.006286>t20                
#>  [18] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40-0.020440>Node63-0.019601>Node84-0.009307>Node96-0.003708>t79
#>  [19] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40-0.020440>Node63-0.019601>Node84-0.009307>Node96-0.003708>t92
#>  [20] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40-0.020440>Node63-0.019601>Node84-0.009307>t45                
#>  [21] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>t24                                                                                                
#>  [22] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node11--0.0554->Node92--0.1732->t84                                                                                                                
#>  [23] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node11--0.0554->Node92--0.1732->t43                                                                                                                
#>  [24] --0.000-->Node2--0.157-->Node4--0.300-->Node8--0.208-->Node17--0.147-->t33                                                                                                                                               
#>  [25] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node17--0.1471->Node59--0.0935->t69                                                                                                                               
#>  [26] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node17--0.1471->Node59--0.0935->t47                                                                                                                               
#>  [27] --0.000-->Node2--0.157-->Node4--0.300-->Node38--0.428-->Node94--0.037-->t86                                                                                                                                              
#>  [28] --0.000-->Node2--0.157-->Node4--0.300-->Node38--0.428-->Node94--0.037-->t9                                                                                                                                               
#>  [29] --0.000-->Node2--0.157-->Node4--0.300-->Node38--0.428-->t98                                                                                                                                                              
#>  [30] --0.000-->Node2--0.157-->Node3--0.137-->Node5--0.279-->Node23--0.271-->t40                                                                                                                                               
#>  [31] --0.0000->Node2--0.1571->Node3--0.1371->Node5--0.2792->Node23--0.2715->Node28--0.0183->Node44--0.0347->Node48--0.0019->t15                                                                                               
#>  [32] -0.00000->Node2-0.15709->Node3-0.13711->Node5-0.27918->Node23-0.27147->Node28-0.01832->Node44-0.03466->Node48-0.00190->Node77-0.01657->Node82-0.00197->t3                                                                
#>  [33] -0.00000->Node2-0.15709->Node3-0.13711->Node5-0.27918->Node23-0.27147->Node28-0.01832->Node44-0.03466->Node48-0.00190->Node77-0.01657->Node82-0.00197->Node89-0.00259->t38                                               
#>  [34] -0.00000->Node2-0.15709->Node3-0.13711->Node5-0.27918->Node23-0.27147->Node28-0.01832->Node44-0.03466->Node48-0.00190->Node77-0.01657->Node82-0.00197->Node89-0.00259->t18                                               
#>  [35] --0.0000->Node2--0.1571->Node3--0.1371->Node5--0.2792->Node23--0.2715->Node28--0.0183->Node44--0.0347->Node48--0.0019->Node77--0.0166->t96                                                                               
#>  [36] -0.00000->Node2-0.15709->Node3-0.13711->Node5-0.27918->Node23-0.27147->Node28-0.01832->Node44-0.03466->Node52-0.00537->t4                                                                                                
#>  [37] -0.00000->Node2-0.15709->Node3-0.13711->Node5-0.27918->Node23-0.27147->Node28-0.01832->Node44-0.03466->Node52-0.00537->t91                                                                                               
#>  [38] --0.0000->Node2--0.1571->Node3--0.1371->Node5--0.2792->Node23--0.2715->Node28--0.0183->Node29--0.0020->Node36--0.0147->t12                                                                                               
#>  [39] --0.0000->Node2--0.1571->Node3--0.1371->Node5--0.2792->Node23--0.2715->Node28--0.0183->Node29--0.0020->Node36--0.0147->t85                                                                                               
#>  [40] --0.0000->Node2--0.1571->Node3--0.1371->Node5--0.2792->Node23--0.2715->Node28--0.0183->Node29--0.0020->Node66--0.0459->t26                                                                                               
#>  [41] --0.0000->Node2--0.1571->Node3--0.1371->Node5--0.2792->Node23--0.2715->Node28--0.0183->Node29--0.0020->Node66--0.0459->t95                                                                                               
#>  [42] --0.000-->Node2--0.157-->Node3--0.137-->Node5--0.279-->Node31--0.294-->t10                                                                                                                                               
#>  [43] --0.000-->Node2--0.157-->Node3--0.137-->Node5--0.279-->Node31--0.294-->t72                                                                                                                                               
#>  [44] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node13--0.1525->Node41--0.1239->t100                                                                                                               
#>  [45] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node13--0.1525->Node41--0.1239->t57                                                                                                                
#>  [46] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node13--0.1525->Node16--0.0392->Node24--0.0404->t49                                                                                                
#>  [47] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node13--0.1525->Node16--0.0392->Node24--0.0404->Node34--0.0307->t19                                                                                
#>  [48] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node13--0.1525->Node16--0.0392->Node24--0.0404->Node34--0.0307->t29                                                                                
#>  [49] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node13--0.1525->Node16--0.0392->Node39--0.0836->Node70--0.0247->t30                                                                                
#>  [50] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node13-0.15255->Node16-0.03918->Node39-0.08359->Node70-0.02474->Node87-0.00588->t28                                                                
#>  [51] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node13-0.15255->Node16-0.03918->Node39-0.08359->Node70-0.02474->Node87-0.00588->t5                                                                 
#>  [52] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node13--0.1525->Node16--0.0392->Node39--0.0836->t74                                                                                                
#>  [53] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node14--0.0472->t50                                                                                                                
#>  [54] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node45-0.00622->Node49-0.00257->Node68-0.01086->Node93-0.01079->t48                                
#>  [55] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node45-0.00622->Node49-0.00257->Node68-0.01086->Node93-0.01079->t75                                
#>  [56] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node45-0.00622->Node49-0.00257->Node68-0.01086->Node73-0.00395->t64                                
#>  [57] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node45-0.00622->Node49-0.00257->Node68-0.01086->Node73-0.00395->t67                                
#>  [58] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node45-0.00622->Node49-0.00257->t80                                                                
#>  [59] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node45-0.00622->Node72-0.01726->Node80-0.00202->t13                                                
#>  [60] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node45-0.00622->Node72-0.01726->Node80-0.00202->t44                                                
#>  [61] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node45-0.00622->Node72-0.01726->t61                                                                
#>  [62] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node46-0.00653->Node50-0.00369->Node61-0.00652->t60                                                
#>  [63] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node46-0.00653->Node50-0.00369->Node61-0.00652->t73                                                
#>  [64] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node46-0.00653->Node50-0.00369->Node56-0.00420->Node88-0.01408->t93                                
#>  [65] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node46-0.00653->Node50-0.00369->Node56-0.00420->Node88-0.01408->t90                                
#>  [66] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node46-0.00653->Node50-0.00369->Node56-0.00420->Node97-0.01701->t27                                
#>  [67] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node46-0.00653->Node50-0.00369->Node56-0.00420->Node97-0.01701->t56                                
#>  [68] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node46-0.00653->Node85-0.02114->t81                                                                
#>  [69] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node46-0.00653->Node85-0.02114->t21                                                                
#>  [70] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node15--0.0598->Node51--0.1222->Node74--0.0123->t41                                                                                
#>  [71] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node15--0.0598->Node51--0.1222->Node74--0.0123->t77                                                                                
#>  [72] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node15-0.05975->Node51-0.12217->Node60-0.00502->Node98-0.01513->t35                                                                
#>  [73] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node15-0.05975->Node51-0.12217->Node60-0.00502->Node98-0.01513->t2                                                                 
#>  [74] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node15-0.05975->Node51-0.12217->Node60-0.00502->t6                                                                                 
#>  [75] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node15--0.0598->Node21--0.0477->Node26--0.0266->t99                                                                                
#>  [76] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node15--0.0598->Node21--0.0477->Node26--0.0266->Node53--0.0483->t23                                                                
#>  [77] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node15-0.05975->Node21-0.04765->Node26-0.02659->Node53-0.04826->Node54-0.00116->t51                                                
#>  [78] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node15-0.05975->Node21-0.04765->Node26-0.02659->Node53-0.04826->Node54-0.00116->Node69-0.00922->t1                                 
#>  [79] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node15-0.05975->Node21-0.04765->Node26-0.02659->Node53-0.04826->Node54-0.00116->Node69-0.00922->t94                                
#>  [80] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node15--0.0598->Node21--0.0477->Node55--0.0761->t54                                                                                
#>  [81] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node15--0.0598->Node21--0.0477->Node55--0.0761->t68                                                                                
#>  [82] -0.000000>Node2-0.157092>Node3-0.137111>Node6-0.280654>Node27-0.287398>Node58-0.044190>Node83-0.012239>Node86-0.000802>t17                                                                                               
#>  [83] -0.000000>Node2-0.157092>Node3-0.137111>Node6-0.280654>Node27-0.287398>Node58-0.044190>Node83-0.012239>Node86-0.000802>t52                                                                                               
#>  [84] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node27--0.2874->Node58--0.0442->Node83--0.0122->t36                                                                                                               
#>  [85] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node27--0.2874->Node58--0.0442->Node62--0.0029->t7                                                                                                                
#>  [86] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node27--0.2874->Node58--0.0442->Node62--0.0029->t88                                                                                                               
#>  [87] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node27-0.28740->Node30-0.00314->Node47-0.03432->Node75-0.01616->t83                                                                                               
#>  [88] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node27-0.28740->Node30-0.00314->Node47-0.03432->Node75-0.01616->t59                                                                                               
#>  [89] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node27-0.28740->Node30-0.00314->Node47-0.03432->Node71-0.01445->t16                                                                                               
#>  [90] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node27-0.28740->Node30-0.00314->Node47-0.03432->Node71-0.01445->t58                                                                                               
#>  [91] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node27-0.28740->Node30-0.00314->Node99-0.05807->t14                                                                                                               
#>  [92] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node27-0.28740->Node30-0.00314->Node99-0.05807->t70                                                                                                               
#>  [93] --0.0000->Node22--0.8403->Node32--0.0272->t42                                                                                                                                                                            
#>  [94] --0.0000->Node22--0.8403->Node32--0.0272->Node35--0.0119->t78                                                                                                                                                            
#>  [95] -0.00000->Node22-0.84034->Node32-0.02717->Node35-0.01188->Node65-0.03142->Node81-0.00683->t55                                                                                                                            
#>  [96] -0.00000->Node22-0.84034->Node32-0.02717->Node35-0.01188->Node65-0.03142->Node81-0.00683->t39                                                                                                                            
#>  [97] --0.0000->Node22--0.8403->Node32--0.0272->Node35--0.0119->Node65--0.0314->t76                                                                                                                                            
#>  [98] --0.000-->Node22--0.840-->Node57--0.066-->t53                                                                                                                                                                            
#>  [99] -0.00000->Node22-0.84034->Node57-0.06600->Node64-0.00442->t87                                                                                                                                                            
#> [100] -0.00000->Node22-0.84034->Node57-0.06600->Node64-0.00442->t82                                                                                                                                                            
#> [101] ----0---->Node2                                                                                                                                                                                                          
#> [102] --0.000-->Node2--0.157-->Node3                                                                                                                                                                                           
#> [103] --0.000-->Node2--0.157-->Node4                                                                                                                                                                                           
#> [104] --0.000-->Node2--0.157-->Node3--0.137-->Node5                                                                                                                                                                            
#> [105] --0.000-->Node2--0.157-->Node3--0.137-->Node6                                                                                                                                                                            
#> [106] --0.000-->Node2--0.157-->Node3--0.137-->Node6--0.281-->Node7                                                                                                                                                             
#> [107] --0.000-->Node2--0.157-->Node4--0.300-->Node8                                                                                                                                                                            
#> [108] --0.000-->Node2--0.157-->Node4--0.300-->Node8--0.208-->Node9                                                                                                                                                             
#> [109] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10                                                                                                                                             
#> [110] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node11                                                                                                                                             
#> [111] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node12                                                                                                                                             
#> [112] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node13                                                                                                                                             
#> [113] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node14                                                                                                                             
#> [114] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node15                                                                                                                             
#> [115] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node13--0.1525->Node16                                                                                                                             
#> [116] --0.000-->Node2--0.157-->Node4--0.300-->Node8--0.208-->Node17                                                                                                                                                            
#> [117] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node12--0.0695->Node18                                                                                                                             
#> [118] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node11--0.0554->Node19                                                                                                                             
#> [119] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node11--0.0554->Node19--0.0789->Node20                                                                                                             
#> [120] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node15--0.0598->Node21                                                                                                             
#> [121] ----0---->Node22                                                                                                                                                                                                         
#> [122] --0.000-->Node2--0.157-->Node3--0.137-->Node5--0.279-->Node23                                                                                                                                                            
#> [123] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node13--0.1525->Node16--0.0392->Node24                                                                                                             
#> [124] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25                                                                                             
#> [125] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node15--0.0598->Node21--0.0477->Node26                                                                                             
#> [126] --0.000-->Node2--0.157-->Node3--0.137-->Node6--0.281-->Node27                                                                                                                                                            
#> [127] --0.000-->Node2--0.157-->Node3--0.137-->Node5--0.279-->Node23--0.271-->Node28                                                                                                                                            
#> [128] --0.0000->Node2--0.1571->Node3--0.1371->Node5--0.2792->Node23--0.2715->Node28--0.0183->Node29                                                                                                                            
#> [129] --0.000-->Node2--0.157-->Node3--0.137-->Node6--0.281-->Node27--0.287-->Node30                                                                                                                                            
#> [130] --0.000-->Node2--0.157-->Node3--0.137-->Node5--0.279-->Node31                                                                                                                                                            
#> [131] ---0.00-->Node22---0.84-->Node32                                                                                                                                                                                         
#> [132] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33                                                                             
#> [133] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node13--0.1525->Node16--0.0392->Node24--0.0404->Node34                                                                                             
#> [134] --0.0000->Node22--0.8403->Node32--0.0272->Node35                                                                                                                                                                         
#> [135] --0.0000->Node2--0.1571->Node3--0.1371->Node5--0.2792->Node23--0.2715->Node28--0.0183->Node29--0.0020->Node36                                                                                                            
#> [136] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node11--0.0554->Node19--0.0789->Node37                                                                                                             
#> [137] --0.000-->Node2--0.157-->Node4--0.300-->Node38                                                                                                                                                                           
#> [138] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node13--0.1525->Node16--0.0392->Node39                                                                                                             
#> [139] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40                                                             
#> [140] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node13--0.1525->Node41                                                                                                                             
#> [141] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node14--0.0472->Node42                                                                                                             
#> [142] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node12--0.0695->Node18--0.0530->Node43                                                                                                             
#> [143] --0.0000->Node2--0.1571->Node3--0.1371->Node5--0.2792->Node23--0.2715->Node28--0.0183->Node44                                                                                                                            
#> [144] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node14--0.0472->Node42--0.1234->Node45                                                                                             
#> [145] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node14--0.0472->Node42--0.1234->Node46                                                                                             
#> [146] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node27-0.28740->Node30-0.00314->Node47                                                                                                                            
#> [147] --0.0000->Node2--0.1571->Node3--0.1371->Node5--0.2792->Node23--0.2715->Node28--0.0183->Node44--0.0347->Node48                                                                                                            
#> [148] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node45-0.00622->Node49                                                                             
#> [149] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node46-0.00653->Node50                                                                             
#> [150] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node15--0.0598->Node51                                                                                                             
#> [151] --0.0000->Node2--0.1571->Node3--0.1371->Node5--0.2792->Node23--0.2715->Node28--0.0183->Node44--0.0347->Node52                                                                                                            
#> [152] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node15--0.0598->Node21--0.0477->Node26--0.0266->Node53                                                                             
#> [153] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node15--0.0598->Node21--0.0477->Node26--0.0266->Node53--0.0483->Node54                                                             
#> [154] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node15--0.0598->Node21--0.0477->Node55                                                                                             
#> [155] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node46-0.00653->Node50-0.00369->Node56                                                             
#> [156] ---0.00-->Node22---0.84-->Node57                                                                                                                                                                                         
#> [157] --0.000-->Node2--0.157-->Node3--0.137-->Node6--0.281-->Node27--0.287-->Node58                                                                                                                                            
#> [158] --0.000-->Node2--0.157-->Node4--0.300-->Node8--0.208-->Node17--0.147-->Node59                                                                                                                                            
#> [159] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node15--0.0598->Node51--0.1222->Node60                                                                                             
#> [160] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node46-0.00653->Node50-0.00369->Node61                                                             
#> [161] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node27--0.2874->Node58--0.0442->Node62                                                                                                                            
#> [162] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40-0.020440>Node63                                             
#> [163] --0.000-->Node22--0.840-->Node57--0.066-->Node64                                                                                                                                                                         
#> [164] --0.0000->Node22--0.8403->Node32--0.0272->Node35--0.0119->Node65                                                                                                                                                         
#> [165] --0.0000->Node2--0.1571->Node3--0.1371->Node5--0.2792->Node23--0.2715->Node28--0.0183->Node29--0.0020->Node66                                                                                                            
#> [166] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40-0.020440>Node67                                             
#> [167] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node45-0.00622->Node49-0.00257->Node68                                                             
#> [168] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node15-0.05975->Node21-0.04765->Node26-0.02659->Node53-0.04826->Node54-0.00116->Node69                                             
#> [169] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node13--0.1525->Node16--0.0392->Node39--0.0836->Node70                                                                                             
#> [170] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node27-0.28740->Node30-0.00314->Node47-0.03432->Node71                                                                                                            
#> [171] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node45-0.00622->Node72                                                                             
#> [172] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node45-0.00622->Node49-0.00257->Node68-0.01086->Node73                                             
#> [173] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node10--0.1071->Node15--0.0598->Node51--0.1222->Node74                                                                                             
#> [174] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node27-0.28740->Node30-0.00314->Node47-0.03432->Node75                                                                                                            
#> [175] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40-0.020440>Node63-0.019601>Node76                             
#> [176] --0.0000->Node2--0.1571->Node3--0.1371->Node5--0.2792->Node23--0.2715->Node28--0.0183->Node44--0.0347->Node48--0.0019->Node77                                                                                            
#> [177] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40-0.020440>Node67-0.021053>Node78                             
#> [178] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40-0.020440>Node63-0.019601>Node76-0.006286>Node79             
#> [179] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node45-0.00622->Node72-0.01726->Node80                                                             
#> [180] --0.0000->Node22--0.8403->Node32--0.0272->Node35--0.0119->Node65--0.0314->Node81                                                                                                                                         
#> [181] --0.0000->Node2--0.1571->Node3--0.1371->Node5--0.2792->Node23--0.2715->Node28--0.0183->Node44--0.0347->Node48--0.0019->Node77--0.0166->Node82                                                                            
#> [182] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node27--0.2874->Node58--0.0442->Node83                                                                                                                            
#> [183] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40-0.020440>Node63-0.019601>Node84                             
#> [184] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node46-0.00653->Node85                                                                             
#> [185] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node27--0.2874->Node58--0.0442->Node83--0.0122->Node86                                                                                                            
#> [186] --0.0000->Node2--0.1571->Node3--0.1371->Node6--0.2807->Node7--0.0393->Node13--0.1525->Node16--0.0392->Node39--0.0836->Node70--0.0247->Node87                                                                             
#> [187] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node46-0.00653->Node50-0.00369->Node56-0.00420->Node88                                             
#> [188] -0.00000->Node2-0.15709->Node3-0.13711->Node5-0.27918->Node23-0.27147->Node28-0.01832->Node44-0.03466->Node48-0.00190->Node77-0.01657->Node82-0.00197->Node89                                                            
#> [189] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node90                                                             
#> [190] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node90-0.051361>Node91                                             
#> [191] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node11--0.0554->Node92                                                                                                                             
#> [192] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node45-0.00622->Node49-0.00257->Node68-0.01086->Node93                                             
#> [193] --0.000-->Node2--0.157-->Node4--0.300-->Node38--0.428-->Node94                                                                                                                                                           
#> [194] --0.0000->Node2--0.1571->Node4--0.3004->Node8--0.2085->Node9--0.0274->Node12--0.0695->Node95                                                                                                                             
#> [195] -0.000000>Node2-0.157092>Node4-0.300362>Node8-0.208456>Node9-0.027449>Node11-0.055419>Node19-0.078895>Node20-0.000755>Node25-0.026073>Node33-0.015285>Node40-0.020440>Node63-0.019601>Node84-0.009307>Node96             
#> [196] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node14-0.04716->Node42-0.12339->Node46-0.00653->Node50-0.00369->Node56-0.00420->Node97                                             
#> [197] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node7-0.03925->Node10-0.10714->Node15-0.05975->Node51-0.12217->Node60-0.00502->Node98                                                                             
#> [198] -0.00000->Node2-0.15709->Node3-0.13711->Node6-0.28065->Node27-0.28740->Node30-0.00314->Node99