Return a pfc with ancestral features

Return a pfc with ancestral features

Usage

pf_anc(x, replace = 0)

Arguments

x

A pfc object

replace

Value to replace edge feature with no ancestral value (such as when the ancestor is the root)

Value

A new pfc with the same structure as x, but with the features of each edge's ancestors instead

Examples

pf_anc(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] --0.000-->Node2--0.965-->Node7--0.798-->Node10--0.115-->Node22--0.129-->t13                                                                                                                                                                               
#>   [2] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node22-0.12904->Node45-0.04267->Node46-0.00142->Node49-0.00151->t18                                                                                                                               
#>   [3] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node22-0.12904->Node45-0.04267->Node46-0.00142->Node49-0.00151->Node59-0.00711->t39                                                                                                               
#>   [4] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node22-0.12904->Node45-0.04267->Node46-0.00142->Node49-0.00151->Node59-0.00711->Node75-0.00813->t64                                                                                               
#>   [5] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node22-0.12904->Node45-0.04267->Node46-0.00142->Node49-0.00151->Node59-0.00711->Node75-0.00813->t30                                                                                               
#>   [6] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node22-0.12904->Node45-0.04267->Node46-0.00142->Node78-0.01722->t80                                                                                                                               
#>   [7] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node22-0.12904->Node45-0.04267->Node46-0.00142->Node78-0.01722->Node97-0.00555->t71                                                                                                               
#>   [8] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node22-0.12904->Node45-0.04267->Node46-0.00142->Node78-0.01722->Node97-0.00555->t62                                                                                                               
#>   [9] -0.000000>Node2-0.965129>Node7-0.798408>Node10-0.115394>Node22-0.129040>Node45-0.042667>Node90-0.022584>Node91-0.000505>t33                                                                                                                               
#>  [10] -0.000000>Node2-0.965129>Node7-0.798408>Node10-0.115394>Node22-0.129040>Node45-0.042667>Node90-0.022584>Node91-0.000505>t93                                                                                                                               
#>  [11] --0.0000->Node2--0.9651->Node7--0.7984->Node10--0.1154->Node22--0.1290->Node45--0.0427->Node90--0.0226->t25                                                                                                                                               
#>  [12] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node16-0.08370->Node23-0.04739->Node25-0.00808->Node67-0.04722->Node81-0.00447->t46                                                                                                               
#>  [13] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node16-0.08370->Node23-0.04739->Node25-0.00808->Node67-0.04722->Node81-0.00447->t41                                                                                                               
#>  [14] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node16-0.08370->Node23-0.04739->Node25-0.00808->Node67-0.04722->t24                                                                                                                               
#>  [15] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node16-0.08370->Node23-0.04739->Node25-0.00808->Node88-0.05477->t58                                                                                                                               
#>  [16] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node16-0.08370->Node23-0.04739->Node25-0.00808->Node88-0.05477->t38                                                                                                                               
#>  [17] --0.0000->Node2--0.9651->Node7--0.7984->Node10--0.1154->Node16--0.0837->Node23--0.0474->Node51--0.0445->t59                                                                                                                                               
#>  [18] --0.0000->Node2--0.9651->Node7--0.7984->Node10--0.1154->Node16--0.0837->Node23--0.0474->Node51--0.0445->t56                                                                                                                                               
#>  [19] --0.0000->Node2--0.9651->Node7--0.7984->Node10--0.1154->Node16--0.0837->t63                                                                                                                                                                               
#>  [20] --0.000-->Node2--0.965-->Node7--0.798-->Node84--0.308-->t96                                                                                                                                                                                               
#>  [21] --0.000-->Node2--0.965-->Node7--0.798-->Node84--0.308-->t99                                                                                                                                                                                               
#>  [22] --0.000-->Node2--0.965-->Node3--0.449-->Node6--0.328-->t23                                                                                                                                                                                                
#>  [23] --0.000-->Node2--0.965-->Node3--0.449-->Node6--0.328-->Node35--0.297-->t34                                                                                                                                                                                
#>  [24] --0.000-->Node2--0.965-->Node3--0.449-->Node6--0.328-->Node35--0.297-->t79                                                                                                                                                                                
#>  [25] --0.000-->Node2--0.965-->Node3--0.449-->Node4--0.135-->Node29--0.484-->t17                                                                                                                                                                                
#>  [26] --0.0000->Node2--0.9651->Node3--0.4487->Node4--0.1349->Node29--0.4843->Node99--0.0422->t43                                                                                                                                                                
#>  [27] --0.0000->Node2--0.9651->Node3--0.4487->Node4--0.1349->Node29--0.4843->Node99--0.0422->t28                                                                                                                                                                
#>  [28] --0.0000->Node2--0.9651->Node3--0.4487->Node4--0.1349->Node11--0.3690->Node14--0.0330->Node42--0.0974->t7                                                                                                                                                 
#>  [29] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node14-0.03298->Node42-0.09739->Node54-0.00825->Node64-0.00629->t37                                                                                                                
#>  [30] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node14-0.03298->Node42-0.09739->Node54-0.00825->Node64-0.00629->Node66-0.00192->Node79-0.00485->t82                                                                                
#>  [31] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node14-0.03298->Node42-0.09739->Node54-0.00825->Node64-0.00629->Node66-0.00192->Node79-0.00485->t15                                                                                
#>  [32] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node14-0.03298->Node42-0.09739->Node54-0.00825->Node64-0.00629->Node66-0.00192->t89                                                                                                
#>  [33] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node14-0.03298->Node42-0.09739->Node54-0.00825->t81                                                                                                                                
#>  [34] --0.0000->Node2--0.9651->Node3--0.4487->Node4--0.1349->Node11--0.3690->Node14--0.0330->Node33--0.0861->t11                                                                                                                                                
#>  [35] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node14-0.03298->Node33-0.08610->Node40-0.00746->Node89-0.02882->Node95-0.00159->t100                                                                                               
#>  [36] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node14-0.03298->Node33-0.08610->Node40-0.00746->Node89-0.02882->Node95-0.00159->t69                                                                                                
#>  [37] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node14-0.03298->Node33-0.08610->Node40-0.00746->Node89-0.02882->t48                                                                                                                
#>  [38] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node14-0.032978>Node33-0.086095>Node40-0.007456>Node94-0.030007>Node98-0.000565>t57                                                                                                
#>  [39] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node14-0.032978>Node33-0.086095>Node40-0.007456>Node94-0.030007>Node98-0.000565>t3                                                                                                 
#>  [40] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node14-0.03298->Node33-0.08610->Node40-0.00746->Node94-0.03001->t1                                                                                                                 
#>  [41] --0.0000->Node2--0.9651->Node3--0.4487->Node4--0.1349->Node11--0.3690->Node13--0.0218->Node26--0.0868->t14                                                                                                                                                
#>  [42] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node61-0.006161>Node65-0.002155>Node69-0.003173>Node71-0.000348>t16
#>  [43] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node61-0.006161>Node65-0.002155>Node69-0.003173>Node71-0.000348>t92
#>  [44] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node61-0.006161>Node65-0.002155>Node69-0.003173>t19                
#>  [45] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node61-0.006161>Node65-0.002155>Node77-0.005833>t22                
#>  [46] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node61-0.006161>Node65-0.002155>Node77-0.005833>t47                
#>  [47] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node61-0.006161>Node62-0.001331>t52                                
#>  [48] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node61-0.006161>Node62-0.001331>Node70-0.004058>t29                
#>  [49] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node61-0.006161>Node62-0.001331>Node70-0.004058>t88                
#>  [50] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node55-0.002076>t51                                                
#>  [51] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node55-0.002076>Node83-0.013895>Node85-0.001544>t45                
#>  [52] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node55-0.002076>Node83-0.013895>Node85-0.001544>t67                
#>  [53] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node55-0.002076>Node83-0.013895>t73                                
#>  [54] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node13-0.02182->Node26-0.08681->Node32-0.00905->Node36-0.00326->Node48-0.01435->Node50-0.00146->Node68-0.01096->t72                                                                
#>  [55] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node13-0.02182->Node26-0.08681->Node32-0.00905->Node36-0.00326->Node48-0.01435->Node50-0.00146->Node68-0.01096->Node72-0.00194->t83                                                
#>  [56] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node13-0.02182->Node26-0.08681->Node32-0.00905->Node36-0.00326->Node48-0.01435->Node50-0.00146->Node68-0.01096->Node72-0.00194->t27                                                
#>  [57] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node13-0.02182->Node26-0.08681->Node32-0.00905->Node36-0.00326->Node48-0.01435->t60                                                                                                
#>  [58] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node76-0.030136>Node80-0.000897>Node93-0.004303>t8                                                                 
#>  [59] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node76-0.030136>Node80-0.000897>Node93-0.004303>t20                                                                
#>  [60] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node76-0.030136>Node80-0.000897>Node96-0.005087>t90                                                                
#>  [61] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node76-0.030136>Node80-0.000897>Node96-0.005087>t5                                                                 
#>  [62] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node13-0.02182->Node26-0.08681->Node32-0.00905->Node36-0.00326->Node76-0.03014->t2                                                                                                 
#>  [63] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node13-0.02182->Node26-0.08681->Node32-0.00905->t61                                                                                                                                
#>  [64] --0.0000->Node2--0.9651->Node3--0.4487->Node4--0.1349->Node11--0.3690->Node13--0.0218->Node47--0.1134->t35                                                                                                                                                
#>  [65] --0.0000->Node2--0.9651->Node3--0.4487->Node4--0.1349->Node11--0.3690->Node13--0.0218->Node47--0.1134->t86                                                                                                                                                
#>  [66] --0.0000->Node5--1.5856->Node12--0.3336->Node20--0.0844->Node82--0.0665->t4                                                                                                                                                                               
#>  [67] --0.0000->Node5--1.5856->Node12--0.3336->Node20--0.0844->Node82--0.0665->t68                                                                                                                                                                              
#>  [68] --0.0000->Node5--1.5856->Node12--0.3336->Node20--0.0844->Node73--0.0647->t10                                                                                                                                                                              
#>  [69] -0.00000->Node5-1.58563->Node12-0.33360->Node20-0.08436->Node73-0.06468->Node86-0.00424->t49                                                                                                                                                              
#>  [70] -0.00000->Node5-1.58563->Node12-0.33360->Node20-0.08436->Node73-0.06468->Node86-0.00424->t54                                                                                                                                                              
#>  [71] --0.000-->Node5--1.586-->Node12--0.334-->Node57--0.140-->t78                                                                                                                                                                                              
#>  [72] --0.000-->Node5--1.586-->Node12--0.334-->Node57--0.140-->t74                                                                                                                                                                                              
#>  [73] --0.0000->Node5--1.5856->Node8--0.1780->Node17--0.2186->Node18--0.0175->Node21--0.0071->Node37--0.0330->Node74--0.0287->t12                                                                                                                               
#>  [74] --0.0000->Node5--1.5856->Node8--0.1780->Node17--0.2186->Node18--0.0175->Node21--0.0071->Node37--0.0330->Node74--0.0287->t75                                                                                                                               
#>  [75] --0.0000->Node5--1.5856->Node8--0.1780->Node17--0.2186->Node18--0.0175->Node21--0.0071->Node37--0.0330->Node43--0.0101->t70                                                                                                                               
#>  [76] --0.0000->Node5--1.5856->Node8--0.1780->Node17--0.2186->Node18--0.0175->Node21--0.0071->Node37--0.0330->Node43--0.0101->t85                                                                                                                               
#>  [77] -0.00000->Node5-1.58563->Node8-0.17803->Node17-0.21863->Node18-0.01745->Node21-0.00710->Node24-0.00939->t77                                                                                                                                               
#>  [78] -0.00000->Node5-1.58563->Node8-0.17803->Node17-0.21863->Node18-0.01745->Node21-0.00710->Node24-0.00939->Node28-0.01405->t55                                                                                                                               
#>  [79] -0.00000->Node5-1.58563->Node8-0.17803->Node17-0.21863->Node18-0.01745->Node21-0.00710->Node24-0.00939->Node28-0.01405->t84                                                                                                                               
#>  [80] --0.0000->Node5--1.5856->Node8--0.1780->Node17--0.2186->Node18--0.0175->t44                                                                                                                                                                               
#>  [81] --0.0000->Node5--1.5856->Node8--0.1780->Node17--0.2186->Node41--0.0649->t32                                                                                                                                                                               
#>  [82] --0.0000->Node5--1.5856->Node8--0.1780->Node17--0.2186->Node41--0.0649->t26                                                                                                                                                                               
#>  [83] --0.0000->Node5--1.5856->Node8--0.1780->Node9--0.0037->Node15--0.1851->Node27--0.0758->Node38--0.0121->t9                                                                                                                                                 
#>  [84] --0.0000->Node5--1.5856->Node8--0.1780->Node9--0.0037->Node15--0.1851->Node27--0.0758->Node38--0.0121->Node58--0.0197->t76                                                                                                                                
#>  [85] --0.0000->Node5--1.5856->Node8--0.1780->Node9--0.0037->Node15--0.1851->Node27--0.0758->Node38--0.0121->Node58--0.0197->t95                                                                                                                                
#>  [86] -0.00000->Node5-1.58563->Node8-0.17803->Node9-0.00370->Node15-0.18512->Node27-0.07583->Node31-0.00696->t94                                                                                                                                                
#>  [87] -0.00000->Node5-1.58563->Node8-0.17803->Node9-0.00370->Node15-0.18512->Node27-0.07583->Node31-0.00696->Node87-0.03760->t50                                                                                                                                
#>  [88] -0.000000>Node5-1.585628>Node8-0.178030>Node9-0.003702>Node15-0.185116>Node27-0.075828>Node31-0.006962>Node87-0.037597>Node92-0.000957>t42                                                                                                                
#>  [89] -0.000000>Node5-1.585628>Node8-0.178030>Node9-0.003702>Node15-0.185116>Node27-0.075828>Node31-0.006962>Node87-0.037597>Node92-0.000957>t6                                                                                                                 
#>  [90] --0.0000->Node5--1.5856->Node8--0.1780->Node9--0.0037->Node15--0.1851->Node53--0.1036->t36                                                                                                                                                                
#>  [91] --0.0000->Node5--1.5856->Node8--0.1780->Node9--0.0037->Node15--0.1851->Node53--0.1036->t66                                                                                                                                                                
#>  [92] -0.00000->Node5-1.58563->Node8-0.17803->Node9-0.00370->Node19-0.23435->Node30-0.03170->Node39-0.00866->Node44-0.00795->Node60-0.01070->t97                                                                                                                
#>  [93] -0.00000->Node5-1.58563->Node8-0.17803->Node9-0.00370->Node19-0.23435->Node30-0.03170->Node39-0.00866->Node44-0.00795->Node60-0.01070->Node63-0.00184->t91                                                                                                
#>  [94] -0.00000->Node5-1.58563->Node8-0.17803->Node9-0.00370->Node19-0.23435->Node30-0.03170->Node39-0.00866->Node44-0.00795->Node60-0.01070->Node63-0.00184->t31                                                                                                
#>  [95] -0.00000->Node5-1.58563->Node8-0.17803->Node9-0.00370->Node19-0.23435->Node30-0.03170->Node39-0.00866->Node44-0.00795->Node56-0.00888->t40                                                                                                                
#>  [96] -0.00000->Node5-1.58563->Node8-0.17803->Node9-0.00370->Node19-0.23435->Node30-0.03170->Node39-0.00866->Node44-0.00795->Node56-0.00888->t65                                                                                                                
#>  [97] -0.00000->Node5-1.58563->Node8-0.17803->Node9-0.00370->Node19-0.23435->Node30-0.03170->Node39-0.00866->t53                                                                                                                                                
#>  [98] -0.00000->Node5-1.58563->Node8-0.17803->Node9-0.00370->Node19-0.23435->Node30-0.03170->Node34-0.00379->t98                                                                                                                                                
#>  [99] -0.00000->Node5-1.58563->Node8-0.17803->Node9-0.00370->Node19-0.23435->Node30-0.03170->Node34-0.00379->t21                                                                                                                                                
#> [100] --0.0000->Node5--1.5856->Node8--0.1780->Node9--0.0037->Node19--0.2344->t87                                                                                                                                                                                
#> [101] ----0---->Node2                                                                                                                                                                                                                                           
#> [102] --0.000-->Node2--0.965-->Node3                                                                                                                                                                                                                            
#> [103] --0.000-->Node2--0.965-->Node3--0.449-->Node4                                                                                                                                                                                                             
#> [104] ----0---->Node5                                                                                                                                                                                                                                           
#> [105] --0.000-->Node2--0.965-->Node3--0.449-->Node6                                                                                                                                                                                                             
#> [106] --0.000-->Node2--0.965-->Node7                                                                                                                                                                                                                            
#> [107] ---0.00-->Node5---1.59-->Node8                                                                                                                                                                                                                            
#> [108] --0.000-->Node5--1.586-->Node8--0.178-->Node9                                                                                                                                                                                                             
#> [109] --0.000-->Node2--0.965-->Node7--0.798-->Node10                                                                                                                                                                                                            
#> [110] --0.000-->Node2--0.965-->Node3--0.449-->Node4--0.135-->Node11                                                                                                                                                                                             
#> [111] ---0.00-->Node5---1.59-->Node12                                                                                                                                                                                                                           
#> [112] --0.000-->Node2--0.965-->Node3--0.449-->Node4--0.135-->Node11--0.369-->Node13                                                                                                                                                                             
#> [113] --0.000-->Node2--0.965-->Node3--0.449-->Node4--0.135-->Node11--0.369-->Node14                                                                                                                                                                             
#> [114] --0.0000->Node5--1.5856->Node8--0.1780->Node9--0.0037->Node15                                                                                                                                                                                             
#> [115] --0.000-->Node2--0.965-->Node7--0.798-->Node10--0.115-->Node16                                                                                                                                                                                            
#> [116] --0.000-->Node5--1.586-->Node8--0.178-->Node17                                                                                                                                                                                                            
#> [117] --0.000-->Node5--1.586-->Node8--0.178-->Node17--0.219-->Node18                                                                                                                                                                                            
#> [118] --0.0000->Node5--1.5856->Node8--0.1780->Node9--0.0037->Node19                                                                                                                                                                                             
#> [119] --0.000-->Node5--1.586-->Node12--0.334-->Node20                                                                                                                                                                                                           
#> [120] --0.0000->Node5--1.5856->Node8--0.1780->Node17--0.2186->Node18--0.0175->Node21                                                                                                                                                                            
#> [121] --0.000-->Node2--0.965-->Node7--0.798-->Node10--0.115-->Node22                                                                                                                                                                                            
#> [122] --0.0000->Node2--0.9651->Node7--0.7984->Node10--0.1154->Node16--0.0837->Node23                                                                                                                                                                            
#> [123] --0.0000->Node5--1.5856->Node8--0.1780->Node17--0.2186->Node18--0.0175->Node21--0.0071->Node24                                                                                                                                                            
#> [124] --0.0000->Node2--0.9651->Node7--0.7984->Node10--0.1154->Node16--0.0837->Node23--0.0474->Node25                                                                                                                                                            
#> [125] --0.0000->Node2--0.9651->Node3--0.4487->Node4--0.1349->Node11--0.3690->Node13--0.0218->Node26                                                                                                                                                             
#> [126] --0.0000->Node5--1.5856->Node8--0.1780->Node9--0.0037->Node15--0.1851->Node27                                                                                                                                                                             
#> [127] -0.00000->Node5-1.58563->Node8-0.17803->Node17-0.21863->Node18-0.01745->Node21-0.00710->Node24-0.00939->Node28                                                                                                                                            
#> [128] --0.000-->Node2--0.965-->Node3--0.449-->Node4--0.135-->Node29                                                                                                                                                                                             
#> [129] --0.0000->Node5--1.5856->Node8--0.1780->Node9--0.0037->Node19--0.2344->Node30                                                                                                                                                                             
#> [130] --0.0000->Node5--1.5856->Node8--0.1780->Node9--0.0037->Node15--0.1851->Node27--0.0758->Node31                                                                                                                                                             
#> [131] --0.0000->Node2--0.9651->Node3--0.4487->Node4--0.1349->Node11--0.3690->Node13--0.0218->Node26--0.0868->Node32                                                                                                                                             
#> [132] --0.000-->Node2--0.965-->Node3--0.449-->Node4--0.135-->Node11--0.369-->Node14--0.033-->Node33                                                                                                                                                             
#> [133] --0.0000->Node5--1.5856->Node8--0.1780->Node9--0.0037->Node19--0.2344->Node30--0.0317->Node34                                                                                                                                                             
#> [134] --0.000-->Node2--0.965-->Node3--0.449-->Node6--0.328-->Node35                                                                                                                                                                                             
#> [135] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node13-0.02182->Node26-0.08681->Node32-0.00905->Node36                                                                                                                             
#> [136] --0.0000->Node5--1.5856->Node8--0.1780->Node17--0.2186->Node18--0.0175->Node21--0.0071->Node37                                                                                                                                                            
#> [137] --0.0000->Node5--1.5856->Node8--0.1780->Node9--0.0037->Node15--0.1851->Node27--0.0758->Node38                                                                                                                                                             
#> [138] --0.0000->Node5--1.5856->Node8--0.1780->Node9--0.0037->Node19--0.2344->Node30--0.0317->Node39                                                                                                                                                             
#> [139] --0.0000->Node2--0.9651->Node3--0.4487->Node4--0.1349->Node11--0.3690->Node14--0.0330->Node33--0.0861->Node40                                                                                                                                             
#> [140] --0.000-->Node5--1.586-->Node8--0.178-->Node17--0.219-->Node41                                                                                                                                                                                            
#> [141] --0.000-->Node2--0.965-->Node3--0.449-->Node4--0.135-->Node11--0.369-->Node14--0.033-->Node42                                                                                                                                                             
#> [142] --0.0000->Node5--1.5856->Node8--0.1780->Node17--0.2186->Node18--0.0175->Node21--0.0071->Node37--0.0330->Node43                                                                                                                                            
#> [143] -0.00000->Node5-1.58563->Node8-0.17803->Node9-0.00370->Node19-0.23435->Node30-0.03170->Node39-0.00866->Node44                                                                                                                                             
#> [144] --0.000-->Node2--0.965-->Node7--0.798-->Node10--0.115-->Node22--0.129-->Node45                                                                                                                                                                            
#> [145] --0.0000->Node2--0.9651->Node7--0.7984->Node10--0.1154->Node22--0.1290->Node45--0.0427->Node46                                                                                                                                                            
#> [146] --0.0000->Node2--0.9651->Node3--0.4487->Node4--0.1349->Node11--0.3690->Node13--0.0218->Node47                                                                                                                                                             
#> [147] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node13-0.02182->Node26-0.08681->Node32-0.00905->Node36-0.00326->Node48                                                                                                             
#> [148] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node22-0.12904->Node45-0.04267->Node46-0.00142->Node49                                                                                                                                            
#> [149] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node13-0.02182->Node26-0.08681->Node32-0.00905->Node36-0.00326->Node48-0.01435->Node50                                                                                             
#> [150] --0.0000->Node2--0.9651->Node7--0.7984->Node10--0.1154->Node16--0.0837->Node23--0.0474->Node51                                                                                                                                                            
#> [151] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node13-0.02182->Node26-0.08681->Node32-0.00905->Node36-0.00326->Node48-0.01435->Node50-0.00146->Node52                                                                             
#> [152] --0.0000->Node5--1.5856->Node8--0.1780->Node9--0.0037->Node15--0.1851->Node53                                                                                                                                                                             
#> [153] --0.0000->Node2--0.9651->Node3--0.4487->Node4--0.1349->Node11--0.3690->Node14--0.0330->Node42--0.0974->Node54                                                                                                                                             
#> [154] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node55                                                             
#> [155] -0.00000->Node5-1.58563->Node8-0.17803->Node9-0.00370->Node19-0.23435->Node30-0.03170->Node39-0.00866->Node44-0.00795->Node56                                                                                                                             
#> [156] --0.000-->Node5--1.586-->Node12--0.334-->Node57                                                                                                                                                                                                           
#> [157] --0.0000->Node5--1.5856->Node8--0.1780->Node9--0.0037->Node15--0.1851->Node27--0.0758->Node38--0.0121->Node58                                                                                                                                             
#> [158] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node22-0.12904->Node45-0.04267->Node46-0.00142->Node49-0.00151->Node59                                                                                                                            
#> [159] -0.00000->Node5-1.58563->Node8-0.17803->Node9-0.00370->Node19-0.23435->Node30-0.03170->Node39-0.00866->Node44-0.00795->Node60                                                                                                                             
#> [160] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node61                                                             
#> [161] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node61-0.006161>Node62                                             
#> [162] -0.00000->Node5-1.58563->Node8-0.17803->Node9-0.00370->Node19-0.23435->Node30-0.03170->Node39-0.00866->Node44-0.00795->Node60-0.01070->Node63                                                                                                             
#> [163] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node14-0.03298->Node42-0.09739->Node54-0.00825->Node64                                                                                                                             
#> [164] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node61-0.006161>Node65                                             
#> [165] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node14-0.03298->Node42-0.09739->Node54-0.00825->Node64-0.00629->Node66                                                                                                             
#> [166] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node16-0.08370->Node23-0.04739->Node25-0.00808->Node67                                                                                                                                            
#> [167] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node13-0.02182->Node26-0.08681->Node32-0.00905->Node36-0.00326->Node48-0.01435->Node50-0.00146->Node68                                                                             
#> [168] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node61-0.006161>Node65-0.002155>Node69                             
#> [169] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node61-0.006161>Node62-0.001331>Node70                             
#> [170] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node61-0.006161>Node65-0.002155>Node69-0.003173>Node71             
#> [171] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node13-0.02182->Node26-0.08681->Node32-0.00905->Node36-0.00326->Node48-0.01435->Node50-0.00146->Node68-0.01096->Node72                                                             
#> [172] --0.0000->Node5--1.5856->Node12--0.3336->Node20--0.0844->Node73                                                                                                                                                                                           
#> [173] --0.0000->Node5--1.5856->Node8--0.1780->Node17--0.2186->Node18--0.0175->Node21--0.0071->Node37--0.0330->Node74                                                                                                                                            
#> [174] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node22-0.12904->Node45-0.04267->Node46-0.00142->Node49-0.00151->Node59-0.00711->Node75                                                                                                            
#> [175] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node13-0.02182->Node26-0.08681->Node32-0.00905->Node36-0.00326->Node76                                                                                                             
#> [176] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node61-0.006161>Node65-0.002155>Node77                             
#> [177] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node22-0.12904->Node45-0.04267->Node46-0.00142->Node78                                                                                                                                            
#> [178] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node14-0.03298->Node42-0.09739->Node54-0.00825->Node64-0.00629->Node66-0.00192->Node79                                                                                             
#> [179] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node13-0.02182->Node26-0.08681->Node32-0.00905->Node36-0.00326->Node76-0.03014->Node80                                                                                             
#> [180] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node16-0.08370->Node23-0.04739->Node25-0.00808->Node67-0.04722->Node81                                                                                                                            
#> [181] --0.0000->Node5--1.5856->Node12--0.3336->Node20--0.0844->Node82                                                                                                                                                                                           
#> [182] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node55-0.002076>Node83                                             
#> [183] --0.000-->Node2--0.965-->Node7--0.798-->Node84                                                                                                                                                                                                            
#> [184] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node48-0.014348>Node50-0.001459>Node52-0.000251>Node55-0.002076>Node83-0.013895>Node85                             
#> [185] --0.0000->Node5--1.5856->Node12--0.3336->Node20--0.0844->Node73--0.0647->Node86                                                                                                                                                                           
#> [186] -0.00000->Node5-1.58563->Node8-0.17803->Node9-0.00370->Node15-0.18512->Node27-0.07583->Node31-0.00696->Node87                                                                                                                                             
#> [187] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node16-0.08370->Node23-0.04739->Node25-0.00808->Node88                                                                                                                                            
#> [188] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node14-0.03298->Node33-0.08610->Node40-0.00746->Node89                                                                                                                             
#> [189] --0.0000->Node2--0.9651->Node7--0.7984->Node10--0.1154->Node22--0.1290->Node45--0.0427->Node90                                                                                                                                                            
#> [190] --0.0000->Node2--0.9651->Node7--0.7984->Node10--0.1154->Node22--0.1290->Node45--0.0427->Node90--0.0226->Node91                                                                                                                                            
#> [191] -0.00000->Node5-1.58563->Node8-0.17803->Node9-0.00370->Node15-0.18512->Node27-0.07583->Node31-0.00696->Node87-0.03760->Node92                                                                                                                             
#> [192] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node76-0.030136>Node80-0.000897>Node93                                                                             
#> [193] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node14-0.03298->Node33-0.08610->Node40-0.00746->Node94                                                                                                                             
#> [194] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node14-0.03298->Node33-0.08610->Node40-0.00746->Node89-0.02882->Node95                                                                                                             
#> [195] -0.000000>Node2-0.965129>Node3-0.448686>Node4-0.134891>Node11-0.369041>Node13-0.021815>Node26-0.086805>Node32-0.009051>Node36-0.003265>Node76-0.030136>Node80-0.000897>Node96                                                                             
#> [196] -0.00000->Node2-0.96513->Node7-0.79841->Node10-0.11539->Node22-0.12904->Node45-0.04267->Node46-0.00142->Node78-0.01722->Node97                                                                                                                            
#> [197] -0.00000->Node2-0.96513->Node3-0.44869->Node4-0.13489->Node11-0.36904->Node14-0.03298->Node33-0.08610->Node40-0.00746->Node94-0.03001->Node98                                                                                                             
#> [198] --0.000-->Node2--0.965-->Node3--0.449-->Node4--0.135-->Node29--0.484-->Node99