Return a pfc with ancestral features

Return a pfc with ancestral features

Usage

pf_anc(x, replace = 0)

Arguments

x

A pfc object

replace

Value to replace edge feature with no ancestral value (such as when the ancestor is the root)

Value

A new pfc with the same structure as x, but with the features of each edge's ancestors instead

Examples

pf_anc(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] --0.000-->Node11--2.089-->Node16--0.098-->t16                                                                                                                                                             
#>   [2] -0.000000>Node11-2.089342>Node16-0.097978>Node58-0.116893>Node90-0.014077>Node97-0.000433>t43                                                                                                             
#>   [3] -0.000000>Node11-2.089342>Node16-0.097978>Node58-0.116893>Node90-0.014077>Node97-0.000433>t28                                                                                                             
#>   [4] --0.0000->Node11--2.0893->Node16--0.0980->Node58--0.1169->Node90--0.0141->t66                                                                                                                             
#>   [5] --0.000-->Node11--2.089-->Node16--0.098-->Node58--0.117-->t93                                                                                                                                             
#>   [6] --0.0000->Node11--2.0893->Node15--0.0711->Node42--0.1263->t95                                                                                                                                             
#>   [7] -0.00000->Node11-2.08934->Node15-0.07109->Node42-0.12632->Node51-0.01312->Node62-0.00599->Node77-0.00857->t50                                                                                             
#>   [8] -0.00000->Node11-2.08934->Node15-0.07109->Node42-0.12632->Node51-0.01312->Node62-0.00599->Node77-0.00857->t45                                                                                             
#>   [9] -0.00000->Node11-2.08934->Node15-0.07109->Node42-0.12632->Node51-0.01312->Node62-0.00599->t7                                                                                                              
#>  [10] --0.0000->Node11--2.0893->Node15--0.0711->Node42--0.1263->Node51--0.0131->t73                                                                                                                             
#>  [11] --0.0000->Node11--2.0893->Node15--0.0711->Node18--0.0416->t22                                                                                                                                             
#>  [12] --0.0000->Node11--2.0893->Node15--0.0711->Node18--0.0416->Node54--0.0989->t97                                                                                                                             
#>  [13] --0.0000->Node11--2.0893->Node15--0.0711->Node18--0.0416->Node54--0.0989->t63                                                                                                                             
#>  [14] --0.000-->Node2--1.547-->Node4--0.223-->Node8--0.225-->t72                                                                                                                                                
#>  [15] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node71-0.11901->Node81-0.00377->t62                                                                                
#>  [16] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node71-0.11901->Node81-0.00377->t80                                                                                
#>  [17] --0.000-->Node2--1.547-->Node4--0.223-->Node8--0.225-->Node12--0.146-->Node17--0.052-->Node71--0.119-->t42                                                                                                
#>  [18] --0.0000->Node2--1.5469->Node4--0.2232->Node8--0.2251->Node12--0.1455->Node17--0.0520->Node20--0.0193->t18                                                                                                
#>  [19] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node52-0.03637->Node74-0.01392->Node91-0.00434->t92                                
#>  [20] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node52-0.03637->Node74-0.01392->Node91-0.00434->t100                               
#>  [21] --0.0000->Node2--1.5469->Node4--0.2232->Node8--0.2251->Node12--0.1455->Node17--0.0520->Node20--0.0193->Node29--0.0516->Node52--0.0364->Node74--0.0139->t24                                                
#>  [22] --0.0000->Node2--1.5469->Node4--0.2232->Node8--0.2251->Node12--0.1455->Node17--0.0520->Node20--0.0193->Node29--0.0516->Node52--0.0364->t84                                                                
#>  [23] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node43-0.01984->t67                                                
#>  [24] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node43-0.01984->Node68-0.02280->t90                                
#>  [25] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node43-0.01984->Node68-0.02280->t89                                
#>  [26] -0.000000>Node2-1.546919>Node4-0.223244>Node8-0.225077>Node12-0.145506>Node17-0.052020>Node20-0.019290>Node29-0.051617>Node33-0.003645>Node47-0.025441>Node60-0.012490>Node61-0.000069>t76                
#>  [27] -0.000000>Node2-1.546919>Node4-0.223244>Node8-0.225077>Node12-0.145506>Node17-0.052020>Node20-0.019290>Node29-0.051617>Node33-0.003645>Node47-0.025441>Node60-0.012490>Node61-0.000069>t60                
#>  [28] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node47-0.02544->Node60-0.01249->Node93-0.01308->t12                
#>  [29] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node47-0.02544->Node60-0.01249->Node93-0.01308->t32                
#>  [30] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node47-0.02544->Node49-0.00104->Node64-0.01411->Node69-0.00236->t2 
#>  [31] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node47-0.02544->Node49-0.00104->Node64-0.01411->Node69-0.00236->t53
#>  [32] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node47-0.02544->Node49-0.00104->Node64-0.01411->t11                
#>  [33] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node47-0.02544->Node49-0.00104->t47                                
#>  [34] --0.0000->Node2--1.5469->Node4--0.2232->Node8--0.2251->Node12--0.1455->Node28--0.1201->Node50--0.0335->Node98--0.0246->t39                                                                                
#>  [35] --0.0000->Node2--1.5469->Node4--0.2232->Node8--0.2251->Node12--0.1455->Node28--0.1201->Node50--0.0335->Node98--0.0246->t23                                                                                
#>  [36] --0.0000->Node2--1.5469->Node4--0.2232->Node8--0.2251->Node12--0.1455->Node28--0.1201->Node50--0.0335->Node67--0.0156->t77                                                                                
#>  [37] --0.0000->Node2--1.5469->Node4--0.2232->Node8--0.2251->Node12--0.1455->Node28--0.1201->Node50--0.0335->Node67--0.0156->t40                                                                                
#>  [38] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node28-0.12014->Node34-0.00865->t33                                                                                                
#>  [39] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node28-0.12014->Node34-0.00865->Node59-0.03552->Node72-0.00685->t58                                                                
#>  [40] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node28-0.12014->Node34-0.00865->Node59-0.03552->Node72-0.00685->t41                                                                
#>  [41] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node28-0.12014->Node34-0.00865->Node59-0.03552->Node84-0.01160->t94                                                                
#>  [42] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node28-0.12014->Node34-0.00865->Node59-0.03552->Node84-0.01160->t13                                                                
#>  [43] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7--0.1156->t75                                                                                                                                 
#>  [44] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node13-0.05888->Node14-0.00109->Node32-0.12482->Node96-0.05125->t6                                                  
#>  [45] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node13-0.05888->Node14-0.00109->Node32-0.12482->Node96-0.05125->t20                                                 
#>  [46] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node13-0.05888->Node14-0.00109->Node32-0.12482->t9                                                                  
#>  [47] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node13-0.05888->Node14-0.00109->Node35-0.12841->Node89-0.04738->t19                                                 
#>  [48] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node13-0.05888->Node14-0.00109->Node35-0.12841->Node89-0.04738->t52                                                 
#>  [49] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node13-0.05888->Node14-0.00109->Node35-0.12841->t54                                                                 
#>  [50] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7--0.1156->Node10--0.1019->Node13--0.0589->Node22--0.0857->t74                                                                                 
#>  [51] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7--0.1156->Node10--0.1019->Node13--0.0589->Node22--0.0857->t38                                                                                 
#>  [52] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7--0.1156->Node10--0.1019->Node24--0.1540->Node39--0.0469->t55                                                                                 
#>  [53] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node39-0.04687->Node48-0.00988->Node99-0.02582->t85                                                 
#>  [54] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node39-0.04687->Node48-0.00988->Node99-0.02582->t96                                                 
#>  [55] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node39-0.04687->Node48-0.00988->Node63-0.01409->t68                                                 
#>  [56] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node39-0.04687->Node48-0.00988->Node63-0.01409->t69                                                 
#>  [57] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7--0.1156->Node10--0.1019->Node24--0.1540->Node27--0.0213->Node38--0.0215->t79                                                                 
#>  [58] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node27-0.02129->Node38-0.02145->Node40-0.00589->t83                                                 
#>  [59] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node27-0.02129->Node38-0.02145->Node40-0.00589->Node44-0.00378->t64                                 
#>  [60] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node27-0.02129->Node38-0.02145->Node40-0.00589->Node44-0.00378->Node83-0.02712->t57                 
#>  [61] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node27-0.02129->Node38-0.02145->Node40-0.00589->Node44-0.00378->Node83-0.02712->t65                 
#>  [62] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node27-0.02129->Node31-0.00902->Node66-0.04300->t26                                                 
#>  [63] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node27-0.02129->Node31-0.00902->Node66-0.04300->t56                                                 
#>  [64] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node27-0.02129->Node31-0.00902->Node56-0.03642->t37                                                 
#>  [65] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node27-0.02129->Node31-0.00902->Node56-0.03642->Node95-0.01512->t34                                 
#>  [66] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node27-0.02129->Node31-0.00902->Node56-0.03642->Node95-0.01512->t14                                 
#>  [67] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node26--0.3906->t4                                                                                                                                 
#>  [68] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node26--0.3906->Node76--0.0587->t21                                                                                                                
#>  [69] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node26--0.3906->Node76--0.0587->t29                                                                                                                
#>  [70] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node9--0.0538->Node19--0.1742->Node65--0.1032->t44                                                                                                 
#>  [71] -0.00000->Node2-1.54692->Node3-0.02492->Node6-0.40557->Node9-0.05383->Node19-0.17422->Node65-0.10322->Node88-0.00925->t27                                                                                 
#>  [72] -0.00000->Node2-1.54692->Node3-0.02492->Node6-0.40557->Node9-0.05383->Node19-0.17422->Node65-0.10322->Node88-0.00925->t49                                                                                 
#>  [73] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node9--0.0538->Node19--0.1742->Node85--0.1115->t61                                                                                                 
#>  [74] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node9--0.0538->Node19--0.1742->Node85--0.1115->t10                                                                                                 
#>  [75] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node9--0.0538->Node37--0.2420->Node92--0.0451->t70                                                                                                 
#>  [76] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node9--0.0538->Node37--0.2420->Node92--0.0451->t8                                                                                                  
#>  [77] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node9--0.0538->Node37--0.2420->t51                                                                                                                 
#>  [78] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node23--0.0130->Node70--0.0793->t31                                                                                                
#>  [79] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node23--0.0130->Node70--0.0793->t81                                                                                                
#>  [80] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node23--0.0130->Node25--0.0242->Node36--0.0165->Node75--0.0423->t71                                                                
#>  [81] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node23--0.0130->Node25--0.0242->Node36--0.0165->Node75--0.0423->t17                                                                
#>  [82] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node23--0.0130->Node25--0.0242->Node36--0.0165->t98                                                                                
#>  [83] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node23--0.0130->Node25--0.0242->Node80--0.0598->t78                                                                                
#>  [84] -0.000000>Node2-1.546919>Node3-0.024918>Node6-0.405568>Node21-0.240641>Node23-0.013001>Node25-0.024241>Node80-0.059848>Node82-0.000911>t3                                                                 
#>  [85] -0.000000>Node2-1.546919>Node3-0.024918>Node6-0.405568>Node21-0.240641>Node23-0.013001>Node25-0.024241>Node80-0.059848>Node82-0.000911>t5                                                                 
#>  [86] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node30--0.0481->Node87--0.0517->t59                                                                                                
#>  [87] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node30--0.0481->Node87--0.0517->t30                                                                                                
#>  [88] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node30--0.0481->Node41--0.0197->Node55--0.0170->Node86--0.0148->t99                                                                
#>  [89] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node30--0.0481->Node41--0.0197->Node55--0.0170->Node86--0.0148->t36                                                                
#>  [90] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node30--0.0481->Node41--0.0197->Node55--0.0170->t82                                                                                
#>  [91] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node30--0.0481->Node41--0.0197->Node78--0.0292->t1                                                                                 
#>  [92] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node30--0.0481->Node41--0.0197->Node78--0.0292->t46                                                                                
#>  [93] -0.000000>Node2-1.546919>Node3-0.024918>Node45-0.717277>Node46-0.000104>t48                                                                                                                               
#>  [94] -0.000000>Node2-1.546919>Node3-0.024918>Node45-0.717277>Node46-0.000104>t88                                                                                                                               
#>  [95] --0.0000->Node2--1.5469->Node3--0.0249->Node45--0.7173->Node53--0.0115->t15                                                                                                                               
#>  [96] -0.00000->Node2-1.54692->Node3-0.02492->Node45-0.71728->Node53-0.01154->Node57-0.00262->Node73-0.01042->t91                                                                                               
#>  [97] -0.00000->Node2-1.54692->Node3-0.02492->Node45-0.71728->Node53-0.01154->Node57-0.00262->Node73-0.01042->t87                                                                                               
#>  [98] -0.00000->Node2-1.54692->Node3-0.02492->Node45-0.71728->Node53-0.01154->Node57-0.00262->Node79-0.01184->Node94-0.00323->t35                                                                               
#>  [99] -0.00000->Node2-1.54692->Node3-0.02492->Node45-0.71728->Node53-0.01154->Node57-0.00262->Node79-0.01184->Node94-0.00323->t25                                                                               
#> [100] -0.00000->Node2-1.54692->Node3-0.02492->Node45-0.71728->Node53-0.01154->Node57-0.00262->Node79-0.01184->t86                                                                                               
#> [101] ----0---->Node2                                                                                                                                                                                           
#> [102] ---0.00-->Node2---1.55-->Node3                                                                                                                                                                            
#> [103] ---0.00-->Node2---1.55-->Node4                                                                                                                                                                            
#> [104] --0.000-->Node2--1.547-->Node4--0.223-->Node5                                                                                                                                                             
#> [105] --0.0000->Node2--1.5469->Node3--0.0249->Node6                                                                                                                                                             
#> [106] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7                                                                                                                                              
#> [107] --0.000-->Node2--1.547-->Node4--0.223-->Node8                                                                                                                                                             
#> [108] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node9                                                                                                                                              
#> [109] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7--0.1156->Node10                                                                                                                              
#> [110] ----0---->Node11                                                                                                                                                                                          
#> [111] --0.000-->Node2--1.547-->Node4--0.223-->Node8--0.225-->Node12                                                                                                                                             
#> [112] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7--0.1156->Node10--0.1019->Node13                                                                                                              
#> [113] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7--0.1156->Node10--0.1019->Node13--0.0589->Node14                                                                                              
#> [114] ---0.00-->Node11---2.09-->Node15                                                                                                                                                                          
#> [115] ---0.00-->Node11---2.09-->Node16                                                                                                                                                                          
#> [116] --0.000-->Node2--1.547-->Node4--0.223-->Node8--0.225-->Node12--0.146-->Node17                                                                                                                             
#> [117] --0.0000->Node11--2.0893->Node15--0.0711->Node18                                                                                                                                                          
#> [118] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node9--0.0538->Node19                                                                                                                              
#> [119] --0.000-->Node2--1.547-->Node4--0.223-->Node8--0.225-->Node12--0.146-->Node17--0.052-->Node20                                                                                                             
#> [120] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21                                                                                                                                             
#> [121] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7--0.1156->Node10--0.1019->Node13--0.0589->Node22                                                                                              
#> [122] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node23                                                                                                                             
#> [123] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7--0.1156->Node10--0.1019->Node24                                                                                                              
#> [124] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node23--0.0130->Node25                                                                                                             
#> [125] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node26                                                                                                                                             
#> [126] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7--0.1156->Node10--0.1019->Node24--0.1540->Node27                                                                                              
#> [127] --0.000-->Node2--1.547-->Node4--0.223-->Node8--0.225-->Node12--0.146-->Node28                                                                                                                             
#> [128] --0.0000->Node2--1.5469->Node4--0.2232->Node8--0.2251->Node12--0.1455->Node17--0.0520->Node20--0.0193->Node29                                                                                             
#> [129] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node30                                                                                                                             
#> [130] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7--0.1156->Node10--0.1019->Node24--0.1540->Node27--0.0213->Node31                                                                              
#> [131] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node13-0.05888->Node14-0.00109->Node32                                                                              
#> [132] --0.0000->Node2--1.5469->Node4--0.2232->Node8--0.2251->Node12--0.1455->Node17--0.0520->Node20--0.0193->Node29--0.0516->Node33                                                                             
#> [133] --0.000-->Node2--1.547-->Node4--0.223-->Node8--0.225-->Node12--0.146-->Node28--0.120-->Node34                                                                                                             
#> [134] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node13-0.05888->Node14-0.00109->Node35                                                                              
#> [135] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node23--0.0130->Node25--0.0242->Node36                                                                                             
#> [136] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node9--0.0538->Node37                                                                                                                              
#> [137] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7--0.1156->Node10--0.1019->Node24--0.1540->Node27--0.0213->Node38                                                                              
#> [138] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7--0.1156->Node10--0.1019->Node24--0.1540->Node39                                                                                              
#> [139] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7--0.1156->Node10--0.1019->Node24--0.1540->Node27--0.0213->Node38--0.0215->Node40                                                              
#> [140] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node30--0.0481->Node41                                                                                                             
#> [141] --0.0000->Node11--2.0893->Node15--0.0711->Node42                                                                                                                                                          
#> [142] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node43                                                             
#> [143] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node27-0.02129->Node38-0.02145->Node40-0.00589->Node44                                              
#> [144] --0.0000->Node2--1.5469->Node3--0.0249->Node45                                                                                                                                                            
#> [145] --0.0000->Node2--1.5469->Node3--0.0249->Node45--0.7173->Node46                                                                                                                                            
#> [146] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node47                                                             
#> [147] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node7--0.1156->Node10--0.1019->Node24--0.1540->Node39--0.0469->Node48                                                                              
#> [148] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node47-0.02544->Node49                                             
#> [149] --0.000-->Node2--1.547-->Node4--0.223-->Node8--0.225-->Node12--0.146-->Node28--0.120-->Node50                                                                                                             
#> [150] --0.0000->Node11--2.0893->Node15--0.0711->Node42--0.1263->Node51                                                                                                                                          
#> [151] --0.0000->Node2--1.5469->Node4--0.2232->Node8--0.2251->Node12--0.1455->Node17--0.0520->Node20--0.0193->Node29--0.0516->Node52                                                                             
#> [152] --0.0000->Node2--1.5469->Node3--0.0249->Node45--0.7173->Node53                                                                                                                                            
#> [153] --0.0000->Node11--2.0893->Node15--0.0711->Node18--0.0416->Node54                                                                                                                                          
#> [154] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node30--0.0481->Node41--0.0197->Node55                                                                                             
#> [155] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node27-0.02129->Node31-0.00902->Node56                                                              
#> [156] --0.0000->Node2--1.5469->Node3--0.0249->Node45--0.7173->Node53--0.0115->Node57                                                                                                                            
#> [157] --0.000-->Node11--2.089-->Node16--0.098-->Node58                                                                                                                                                          
#> [158] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node28-0.12014->Node34-0.00865->Node59                                                                                             
#> [159] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node47-0.02544->Node60                                             
#> [160] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node47-0.02544->Node60-0.01249->Node61                             
#> [161] --0.0000->Node11--2.0893->Node15--0.0711->Node42--0.1263->Node51--0.0131->Node62                                                                                                                          
#> [162] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node39-0.04687->Node48-0.00988->Node63                                                              
#> [163] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node47-0.02544->Node49-0.00104->Node64                             
#> [164] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node9--0.0538->Node19--0.1742->Node65                                                                                                              
#> [165] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node27-0.02129->Node31-0.00902->Node66                                                              
#> [166] --0.0000->Node2--1.5469->Node4--0.2232->Node8--0.2251->Node12--0.1455->Node28--0.1201->Node50--0.0335->Node67                                                                                             
#> [167] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node43-0.01984->Node68                                             
#> [168] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node47-0.02544->Node49-0.00104->Node64-0.01411->Node69             
#> [169] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node23--0.0130->Node70                                                                                                             
#> [170] --0.000-->Node2--1.547-->Node4--0.223-->Node8--0.225-->Node12--0.146-->Node17--0.052-->Node71                                                                                                             
#> [171] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node28-0.12014->Node34-0.00865->Node59-0.03552->Node72                                                                             
#> [172] -0.00000->Node2-1.54692->Node3-0.02492->Node45-0.71728->Node53-0.01154->Node57-0.00262->Node73                                                                                                            
#> [173] --0.0000->Node2--1.5469->Node4--0.2232->Node8--0.2251->Node12--0.1455->Node17--0.0520->Node20--0.0193->Node29--0.0516->Node52--0.0364->Node74                                                             
#> [174] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node23--0.0130->Node25--0.0242->Node36--0.0165->Node75                                                                             
#> [175] --0.0000->Node2--1.5469->Node4--0.2232->Node5--0.0948->Node26--0.3906->Node76                                                                                                                             
#> [176] -0.00000->Node11-2.08934->Node15-0.07109->Node42-0.12632->Node51-0.01312->Node62-0.00599->Node77                                                                                                          
#> [177] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node30--0.0481->Node41--0.0197->Node78                                                                                             
#> [178] -0.00000->Node2-1.54692->Node3-0.02492->Node45-0.71728->Node53-0.01154->Node57-0.00262->Node79                                                                                                            
#> [179] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node23--0.0130->Node25--0.0242->Node80                                                                                             
#> [180] --0.000-->Node2--1.547-->Node4--0.223-->Node8--0.225-->Node12--0.146-->Node17--0.052-->Node71--0.119-->Node81                                                                                             
#> [181] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node23--0.0130->Node25--0.0242->Node80--0.0598->Node82                                                                             
#> [182] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node27-0.02129->Node38-0.02145->Node40-0.00589->Node44-0.00378->Node83                              
#> [183] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node28-0.12014->Node34-0.00865->Node59-0.03552->Node84                                                                             
#> [184] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node9--0.0538->Node19--0.1742->Node85                                                                                                              
#> [185] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node30--0.0481->Node41--0.0197->Node55--0.0170->Node86                                                                             
#> [186] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node21--0.2406->Node30--0.0481->Node87                                                                                                             
#> [187] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node9--0.0538->Node19--0.1742->Node65--0.1032->Node88                                                                                              
#> [188] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node13-0.05888->Node14-0.00109->Node35-0.12841->Node89                                                              
#> [189] --0.000-->Node11--2.089-->Node16--0.098-->Node58--0.117-->Node90                                                                                                                                          
#> [190] --0.0000->Node2--1.5469->Node4--0.2232->Node8--0.2251->Node12--0.1455->Node17--0.0520->Node20--0.0193->Node29--0.0516->Node52--0.0364->Node74--0.0139->Node91                                             
#> [191] --0.0000->Node2--1.5469->Node3--0.0249->Node6--0.4056->Node9--0.0538->Node37--0.2420->Node92                                                                                                              
#> [192] -0.00000->Node2-1.54692->Node4-0.22324->Node8-0.22508->Node12-0.14551->Node17-0.05202->Node20-0.01929->Node29-0.05162->Node33-0.00364->Node47-0.02544->Node60-0.01249->Node93                             
#> [193] -0.00000->Node2-1.54692->Node3-0.02492->Node45-0.71728->Node53-0.01154->Node57-0.00262->Node79-0.01184->Node94                                                                                            
#> [194] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node27-0.02129->Node31-0.00902->Node56-0.03642->Node95                                              
#> [195] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node13-0.05888->Node14-0.00109->Node32-0.12482->Node96                                                              
#> [196] --0.0000->Node11--2.0893->Node16--0.0980->Node58--0.1169->Node90--0.0141->Node97                                                                                                                          
#> [197] --0.0000->Node2--1.5469->Node4--0.2232->Node8--0.2251->Node12--0.1455->Node28--0.1201->Node50--0.0335->Node98                                                                                             
#> [198] -0.00000->Node2-1.54692->Node4-0.22324->Node5-0.09481->Node7-0.11560->Node10-0.10189->Node24-0.15405->Node39-0.04687->Node48-0.00988->Node99