Return a pfc with descendent's features

Note that in the context of a phylogenetic flow, each element of the flow only has a single descendent, which is the next edge on the sequence between the root and the terminal node of the flow in question.

Usage

pf_desc(x, replace = 0)

Arguments

x

A pfc object

replace

Value to replace edge feature with no descendent values (such as when the edges are the tip edges).

Value

A new pfc with the same structure as x, but with the features of each edge's descendant instead

Examples

pf_desc(rpfc(100))
#> <pfc<e:198>[198]>
#> First  10 phylogenetic flows: 
#> [1] ◎──0.2637──→ Node4 ──0.0246──→ Node7 ──0.0327──→ Node8 ──0.0046──→ Node10 ──0.1193──→ Node11 ──0.0069──→ Node80 ──0.0000──→ t8 
#> 
#> [2] ◎──0.2637──→ Node4 ──0.0246──→ Node7 ──0.0327──→ Node8 ──0.0046──→ Node10 ──0.1193──→ Node11 ──0.0035──→ Node80 ──0.0033──→ Node89 ──0.0000──→ t38 
#> 
#> [3] ◎──0.2637──→ Node4 ──0.0246──→ Node7 ──0.0327──→ Node8 ──0.0046──→ Node10 ──0.1193──→ Node11 ──0.0035──→ Node80 ──0.0033──→ Node89 ──0.0000──→ t19 
#> 
#> [4] ◎──0.2637──→ Node4 ──0.0246──→ Node7 ──0.0327──→ Node8 ──0.0046──→ Node10 ──0.1257──→ Node11 ──0.0004──→ Node96 ──0.0000──→ t98 
#> 
#> [5] ◎──0.2637──→ Node4 ──0.0246──→ Node7 ──0.0327──→ Node8 ──0.0046──→ Node10 ──0.1257──→ Node11 ──0.0004──→ Node96 ──0.0000──→ t61 
#> 
#> [6] ◎──0.264──→ Node4 ──0.025──→ Node7 ──0.033──→ Node8 ──0.040──→ Node10 ──0.091──→ Node20 ──0.000──→ t84 
#> 
#> [7] ◎──0.264──→ Node4 ──0.025──→ Node7 ──0.033──→ Node8 ──0.040──→ Node10 ──0.066──→ Node20 ──0.013──→ Node46 ──0.012──→ Node67 ──0.000──→ t29 
#> 
#> [8] ◎──0.2637──→ Node4 ──0.0246──→ Node7 ──0.0327──→ Node8 ──0.0401──→ Node10 ──0.0662──→ Node20 ──0.0129──→ Node46 ──0.0076──→ Node67 ──0.0039──→ Node88 ──0.0000──→ t18 
#> 
#> [9] ◎──0.2637──→ Node4 ──0.0246──→ Node7 ──0.0327──→ Node8 ──0.0401──→ Node10 ──0.0662──→ Node20 ──0.0129──→ Node46 ──0.0076──→ Node67 ──0.0039──→ Node88 ──0.0000──→ t43 
#> 
#> [10] ◎──0.2637──→ Node4 ──0.0246──→ Node7 ──0.0327──→ Node8 ──0.0401──→ Node10 ──0.0662──→ Node20 ──0.0167──→ Node46 ──0.0028──→ Node76 ──0.0050──→ Node84 ──0.0000──→ t14