Return a pfc with descendent's features

Note that in the context of a phylogenetic flow, each element of the flow only has a single descendent, which is the next edge on the sequence between the root and the terminal node of the flow in question.

Usage

pf_desc(x, replace = 0)

Arguments

x

A pfc object

replace

Value to replace edge feature with no descendent values (such as when the edges are the tip edges).

Value

A new pfc with the same structure as x, but with the features of each edge's descendant instead

Examples

pf_desc(rpfc(100))
#> <pfc<e:198>[198]>
#> First  10 phylogenetic flows: 
#> [1] ◎──0.266──→ Node2 ──0.147──→ Node3 ──0.137──→ Node6 ──0.067──→ Node17 ──0.032──→ Node28 ──0.000──→ t11 
#> 
#> [2] ◎──0.2659──→ Node2 ──0.1473──→ Node3 ──0.1373──→ Node6 ──0.0675──→ Node17 ──0.0156──→ Node28 ──0.0093──→ Node42 ──0.0075──→ Node63 ──0.0000──→ t59 
#> 
#> [3] ◎──0.2659──→ Node2 ──0.1473──→ Node3 ──0.1373──→ Node6 ──0.0675──→ Node17 ──0.0156──→ Node28 ──0.0093──→ Node42 ──0.0075──→ Node63 ──0.0000──→ t68 
#> 
#> [4] ◎──0.266──→ Node2 ──0.147──→ Node3 ──0.137──→ Node6 ──0.067──→ Node17 ──0.016──→ Node28 ──0.017──→ Node42 ──0.000──→ t91 
#> 
#> [5] ◎── 0.27──→ Node2 ── 0.15──→ Node3 ── 0.14──→ Node6 ── 0.10──→ Node17 ── 0.00──→ t54 
#> 
#> [6] ◎──0.2659──→ Node2 ──0.1473──→ Node3 ──0.1779──→ Node6 ──0.0524──→ Node22 ──0.0060──→ Node68 ──0.0009──→ Node97 ──0.0000──→ t61 
#> 
#> [7] ◎──0.2659──→ Node2 ──0.1473──→ Node3 ──0.1779──→ Node6 ──0.0524──→ Node22 ──0.0060──→ Node68 ──0.0009──→ Node97 ──0.0000──→ t43 
#> 
#> [8] ◎──0.2659──→ Node2 ──0.1473──→ Node3 ──0.1779──→ Node6 ──0.0524──→ Node22 ──0.0014──→ Node68 ──0.0055──→ Node73 ──0.0000──→ t7 
#> 
#> [9] ◎──0.2659──→ Node2 ──0.1473──→ Node3 ──0.1779──→ Node6 ──0.0524──→ Node22 ──0.0014──→ Node68 ──0.0055──→ Node73 ──0.0000──→ t22 
#> 
#> [10] ◎──0.2659──→ Node2 ──0.1473──→ Node3 ──0.1779──→ Node6 ──0.0581──→ Node22 ──0.0012──→ Node95 ──0.0000──→ t8