Segments the edges of a phylogeny by splitting them at particular positions

Segments the edges of a phylogeny by splitting them at particular positions

Usage

pf_edge_segmentize(x, edges, positions)

Arguments

x

A pfc object

edges

A vector of edge names

positions

A vector of position to cut the edges (must be same length as edges)

Value

A pfc object with edges segmented

Examples

pf_edge_segmentize(rpfc(100), "t1", 0.01)
#> Joining with `by = join_by(edge)`
#> <pfc<e:199>[100]>
#> First  10 phylogenetic flows: 
#> [1] ◎──0.6895──→ Node2_1 ──0.2325──→ Node3_1 ──0.3628──→ Node4_1 ──0.4784──→ Node10_1 ──0.0713──→ Node16_1 ──0.1010──→ Node86_1 ──0.0017──→ Node93_1 ──0.0100──→ t1_1 ──-0.0086──→ t1_2 
#> 
#> [2] ◎──0.6895──→ Node2_1 ──0.2325──→ Node3_1 ──0.7278──→ Node7_1 ──0.1877──→ Node17_1 ──0.0443──→ Node23_1 ──0.0250──→ Node34_1 ──0.0052──→ Node38_1 ──0.0065──→ Node43_1 ──0.0060──→ Node51_1 ──0.0089──→ Node74_1 ──0.0053──→ t10_1 
#> 
#> [3] ◎──0.6895──→ Node2_1 ──0.2325──→ Node3_1 ──0.7278──→ Node7_1 ──0.1877──→ Node17_1 ──0.0443──→ Node23_1 ──0.0250──→ Node34_1 ──0.0052──→ Node38_1 ──0.0065──→ Node43_1 ──0.0031──→ Node48_1 ──0.0024──→ Node50_1 ──0.0022──→ Node53_1 ──0.0126──→ t100_1 
#> 
#> [4] ◎──0.68952──→ Node2_1 ──0.74841──→ Node5_1 ──0.17160──→ Node6_1 ──0.13866──→ Node9_1 ──0.05886──→ Node13_1 ──0.01450──→ Node15_1 ──0.07214──→ Node26_1 ──0.00388──→ Node27_1 ──0.03493──→ Node73_1 ──0.00553──→ Node97_1 ──0.00056──→ t11_1 
#> 
#> [5] ◎──0.690──→ Node2_1 ──0.233──→ Node3_1 ──0.728──→ Node7_1 ──0.216──→ Node22_1 ──0.023──→ Node25_1 ──0.049──→ t12_1 
#> 
#> [6] ◎──0.690──→ Node2_1 ──0.748──→ Node5_1 ──0.420──→ Node19_1 ──0.081──→ t13_1 
#> 
#> [7] ◎──0.6895──→ Node2_1 ──0.2325──→ Node3_1 ──0.3628──→ Node4_1 ──0.4784──→ Node10_1 ──0.0713──→ Node16_1 ──0.0221──→ Node18_1 ──0.0083──→ Node21_1 ──0.0736──→ t14_1 
#> 
#> [8] ◎──0.690──→ Node2_1 ──0.748──→ Node5_1 ──0.172──→ Node6_1 ──0.191──→ Node11_1 ──0.110──→ Node36_1 ──0.027──→ t15_1 
#> 
#> [9] ◎──0.68952──→ Node2_1 ──0.23251──→ Node3_1 ──0.36282──→ Node4_1 ──0.47843──→ Node10_1 ──0.09699──→ Node20_1 ──0.02491──→ Node24_1 ──0.04919──→ Node80_1 ──0.00086──→ Node84_1 ──0.00337──→ t16_1 
#> 
#> [10] ◎──0.6895──→ Node2_1 ──0.2325──→ Node3_1 ──0.3628──→ Node4_1 ──0.4784──→ Node10_1 ──0.0713──→ Node16_1 ──0.0221──→ Node18_1 ──0.0083──→ Node21_1 ──0.0347──→ Node28_1 ──0.0012──→ Node30_1 ──0.0211──→ Node49_1 ──0.0166──→ t17_1