Extract the features of the end edge of each phylogenetic flow

Extract the features of the end edge of each phylogenetic flow

Usage

pf_end_features(x)

Arguments

x

A pfc object

Value

A numeric vector of the features of the end edge of each phylogenetic flow

Examples

pf_end_features(rpfc(100))
#>   [1] 0.0037160063 0.0037160063 0.0093413746 0.0093413746 0.0052309027
#>   [6] 0.0052309027 0.0171349976 0.0171349976 0.0265325064 0.0088526197
#>  [11] 0.0031325645 0.0024003794 0.0024003794 0.0138674849 0.0060392892
#>  [16] 0.0060392892 0.0018785134 0.0018785134 0.0070871457 0.0083298354
#>  [21] 0.0083298354 0.0420217811 0.0083008410 0.0029304522 0.0029304522
#>  [26] 0.0089068022 0.0089068022 0.0095812723 0.0153766538 0.0141175015
#>  [31] 0.0009379487 0.0009379487 0.0089908880 0.0078239559 0.0078239559
#>  [36] 0.0279293293 0.0065369265 0.0065369265 0.0166905810 0.0164313962
#>  [41] 0.0164313962 0.0205190428 0.0075759215 0.0075759215 0.0088708995
#>  [46] 0.0238666271 0.0238666271 0.0183343890 0.0183343890 0.0195826700
#>  [51] 0.0195826700 0.0343649924 0.0997541096 0.0252377459 0.0035698784
#>  [56] 0.0035698784 0.0149548629 0.1146689014 0.0343203834 0.0343203834
#>  [61] 0.0079977579 0.0079977579 0.0211121609 0.0160740357 0.0160740357
#>  [66] 0.0166457811 0.0166457811 0.0239235173 0.0049147380 0.0016031662
#>  [71] 0.0016031662 0.0155968163 0.0177363497 0.0115698631 0.0115698631
#>  [76] 0.0278504026 0.0077675764 0.0077675764 0.0093075271 0.0093075271
#>  [81] 0.0010716929 0.0010716929 0.0215030509 0.0215030509 0.2642021494
#>  [86] 0.0592248053 0.0072740658 0.0072740658 0.0036275150 0.0036275150
#>  [91] 0.0081173740 0.0020462273 0.0020462273 0.0165475115 0.0086813840
#>  [96] 0.0086813840 0.0001680238 0.0001680238 0.0033934699 0.0033934699
#> [101] 1.5344827566 0.0682667072 0.2737915132 2.0263782760 0.0072969602
#> [106] 0.2198246406 0.0730148813 0.0645150265 0.0391365091 0.7203490111
#> [111] 0.0378860913 0.1128079281 0.2938075291 0.0250951755 0.1959058241
#> [116] 0.0149147918 0.2749958815 0.0599856952 0.5185825988 0.0062974172
#> [121] 0.1091508371 0.0429114061 0.8036998034 0.3086288797 0.1198111801
#> [126] 0.0046212148 0.0187757939 0.1733185596 0.0301332202 0.3992066136
#> [131] 0.0454579983 0.0029097286 0.0240244322 0.0441939372 0.1235438187
#> [136] 0.0007959302 0.0388456396 0.0745163637 0.0405979110 0.0089162733
#> [141] 0.0069357376 0.0055514215 0.0023834325 0.0863206734 0.0122638576
#> [146] 0.0147823224 0.0474914603 0.1336578716 0.0093975088 0.0100702928
#> [151] 0.0002022720 0.0112387483 0.0072777362 0.0047124767 0.0192612328
#> [156] 0.0050381252 0.0489246119 0.0805086894 0.0102828829 0.0160570841
#> [161] 0.0817678417 0.0012868052 0.0408503181 0.0061664866 0.0045085763
#> [166] 0.0297551402 0.0074743805 0.3618903367 0.0006744701 0.0116481433
#> [171] 0.0050148652 0.0078661275 0.0336919457 0.0057890076 0.0064047002
#> [176] 0.0717806238 0.0011669321 0.0092166031 0.0012949780 0.0473295246
#> [181] 0.0537104293 0.0101536544 0.0486785139 0.0601472433 0.0106820783
#> [186] 0.0353805085 0.0108945592 0.0113849845 0.0202499508 0.0057200552
#> [191] 0.0053703888 0.0007321851 0.0060711467 0.0052086323 0.0033115718
#> [196] 0.0157102146 0.0131795528 0.0210919645