Calculate the sum of features for each flow

Cumulative sum the phylogenetic flow features for each flow in a pfc, returning a pfc.

Usage

pf_flow_cumsum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A pfc object with cumulative sums of features in place of the original features

Examples

pf_flow_cumsum(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31---2.84-->Node54---2.85-->t55                                                                                
#>   [2] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31---2.84-->Node54---2.85-->t68                                                                                
#>   [3] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31---2.81-->Node32---2.83-->Node52---2.85-->t38                                                                
#>   [4] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31---2.81-->Node32---2.83-->Node52---2.85-->t14                                                                
#>   [5] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31---2.81-->Node32---2.83-->Node48---2.83-->Node50---2.85-->t21                                                
#>   [6] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31---2.81-->Node32---2.83-->Node48---2.83-->Node50---2.85-->t15                                                
#>   [7] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31---2.81-->Node32---2.83-->Node48---2.84-->Node58---2.85-->t93                                                
#>   [8] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31---2.81-->Node32---2.83-->Node48---2.84-->Node58---2.84-->Node75---2.85-->t84                                
#>   [9] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31---2.81-->Node32---2.83-->Node48---2.84-->Node58---2.84-->Node75---2.85-->t70                                
#>  [10] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.83-->Node39---2.85-->t32                                                                                                
#>  [11] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.83-->Node39---2.83-->Node46---2.84-->Node61---2.85-->Node88---2.85-->Node91---2.85-->t74                                
#>  [12] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.83-->Node39---2.83-->Node46---2.84-->Node61---2.85-->Node88---2.85-->Node91---2.85-->t95                                
#>  [13] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.83-->Node39---2.83-->Node46---2.84-->Node61---2.85-->Node88---2.85-->t47                                                
#>  [14] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.83-->Node39---2.83-->Node46---2.84-->Node61---2.85-->t96                                                                
#>  [15] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.83-->Node39---2.83-->Node46---2.85-->Node90---2.85-->t52                                                                
#>  [16] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.83-->Node39---2.83-->Node46---2.85-->Node90---2.85-->t8                                                                 
#>  [17] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.83-->Node40---2.85-->t83                                                                                                                
#>  [18] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.83-->Node40---2.84-->Node66---2.85-->t31                                                                                                
#>  [19] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.83-->Node40---2.84-->Node66---2.85-->Node80---2.85-->t69                                                                                
#>  [20] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.83-->Node40---2.84-->Node66---2.85-->Node80---2.85-->t77                                                                                
#>  [21] ---1.76-->Node2---2.21-->Node4---2.85-->t46                                                                                                                                                
#>  [22] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.76-->Node18---2.85-->t17                                                                                                 
#>  [23] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.76-->Node18---2.80-->Node28---2.85-->Node84---2.85-->t90                                                                 
#>  [24] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.76-->Node18---2.80-->Node28---2.85-->Node84---2.85-->t88                                                                 
#>  [25] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.76-->Node18---2.80-->Node28---2.85-->t4                                                                                  
#>  [26] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.84-->Node53---2.85-->Node83---2.85-->t2                                                                  
#>  [27] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.84-->Node53---2.85-->Node83---2.85-->t98                                                                 
#>  [28] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.84-->Node53---2.84-->Node68---2.85-->t6                                                                  
#>  [29] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.84-->Node53---2.84-->Node68---2.85-->t89                                                                 
#>  [30] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.83-->Node51---2.85-->Node93---2.85-->t43                                                                 
#>  [31] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.83-->Node51---2.85-->Node93---2.85-->t29                                                                 
#>  [32] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.83-->Node51---2.84-->Node65---2.85-->t41                                                                 
#>  [33] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.83-->Node51---2.84-->Node65---2.84-->Node74---2.85-->t59                                                 
#>  [34] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.83-->Node51---2.84-->Node65---2.84-->Node74---2.84-->Node76---2.85-->t82                                 
#>  [35] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.83-->Node51---2.84-->Node65---2.84-->Node74---2.84-->Node76---2.85-->t34                                 
#>  [36] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.77-->Node20---2.80-->Node30---2.84-->Node70---2.85-->t100                                                                 
#>  [37] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.77-->Node20---2.80-->Node30---2.84-->Node70---2.85-->t33                                                                  
#>  [38] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.77-->Node20---2.80-->Node30---2.85-->t75                                                                                  
#>  [39] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.77-->Node20---2.84-->Node67---2.85-->t49                                                                                  
#>  [40] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.77-->Node20---2.84-->Node67---2.84-->Node69---2.85-->t50                                                                  
#>  [41] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.77-->Node20---2.84-->Node67---2.84-->Node69---2.85-->t65                                                                  
#>  [42] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.67-->Node11---2.85-->t54                                                                                                  
#>  [43] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.67-->Node11---2.75-->Node16---2.82-->Node36---2.83-->Node42---2.85-->t72                                                  
#>  [44] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.67-->Node11---2.75-->Node16---2.82-->Node36---2.83-->Node42---2.85-->t44                                                  
#>  [45] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.67-->Node11---2.75-->Node16---2.82-->Node36---2.85-->t94                                                                  
#>  [46] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.67-->Node11---2.75-->Node16---2.78-->Node22---2.85-->Node82---2.85-->t99                                                  
#>  [47] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.67-->Node11---2.75-->Node16---2.78-->Node22---2.85-->Node82---2.85-->t63                                                  
#>  [48] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.67-->Node11---2.75-->Node16---2.78-->Node22---2.84-->Node71---2.85-->t28                                                  
#>  [49] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.67-->Node11---2.75-->Node16---2.78-->Node22---2.84-->Node71---2.85-->t61                                                  
#>  [50] ---1.76-->Node2---2.03-->Node3---2.85-->t67                                                                                                                                                
#>  [51] ---2.44-->Node5---2.58-->Node7---2.69-->Node13---2.85-->t48                                                                                                                                
#>  [52] ---2.44-->Node5---2.58-->Node7---2.69-->Node13---2.80-->Node29---2.84-->Node59---2.85-->t9                                                                                                 
#>  [53] ---2.44-->Node5---2.58-->Node7---2.69-->Node13---2.80-->Node29---2.84-->Node59---2.85-->t23                                                                                                
#>  [54] ---2.44-->Node5---2.58-->Node7---2.69-->Node13---2.80-->Node29---2.83-->Node38---2.84-->Node60---2.85-->Node86---2.85-->t40                                                                
#>  [55] ---2.44-->Node5---2.58-->Node7---2.69-->Node13---2.80-->Node29---2.83-->Node38---2.84-->Node60---2.85-->Node86---2.85-->t81                                                                
#>  [56] ---2.44-->Node5---2.58-->Node7---2.69-->Node13---2.80-->Node29---2.83-->Node38---2.84-->Node60---2.85-->t3                                                                                 
#>  [57] ---2.44-->Node5---2.58-->Node7---2.69-->Node13---2.80-->Node29---2.83-->Node38---2.85-->t16                                                                                                
#>  [58] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.83-->Node45---2.85-->t80                                                                                                
#>  [59] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.83-->Node45---2.84-->Node62---2.85-->t13                                                                                
#>  [60] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.83-->Node45---2.84-->Node62---2.85-->t36                                                                                
#>  [61] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.81-->Node33---2.83-->Node43---2.83-->Node47---2.85-->t11                                                                
#>  [62] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.81-->Node33---2.83-->Node43---2.83-->Node47---2.84-->Node57---2.85-->t85                                                
#>  [63] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.81-->Node33---2.83-->Node43---2.83-->Node47---2.84-->Node57---2.85-->Node79---2.85-->t5                                 
#>  [64] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.81-->Node33---2.83-->Node43---2.83-->Node47---2.84-->Node57---2.85-->Node79---2.85-->t71                                
#>  [65] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.81-->Node33---2.83-->Node43---2.85-->t76                                                                                
#>  [66] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.81-->Node33---2.84-->Node73---2.85-->t97                                                                                
#>  [67] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.81-->Node33---2.84-->Node73---2.85-->Node77---2.85-->t24                                                                
#>  [68] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.81-->Node33---2.84-->Node73---2.85-->Node77---2.85-->t86                                                                
#>  [69] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.84-->Node72---2.85-->t7                                                                                                 
#>  [70] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.84-->Node72---2.85-->t27                                                                                                
#>  [71] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17---2.85-->t35                                                                                                
#>  [72] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17---2.81-->Node34---2.85-->t57                                                                                
#>  [73] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17---2.81-->Node34---2.82-->Node37---2.85-->t25                                                                
#>  [74] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17---2.81-->Node34---2.82-->Node37---2.84-->Node55---2.85-->Node81---2.85-->t51                                
#>  [75] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17---2.81-->Node34---2.82-->Node37---2.84-->Node55---2.85-->Node81---2.85-->Node89---2.85-->Node95---2.85-->t39
#>  [76] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17---2.81-->Node34---2.82-->Node37---2.84-->Node55---2.85-->Node81---2.85-->Node89---2.85-->Node95---2.85-->t58
#>  [77] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17---2.81-->Node34---2.82-->Node37---2.84-->Node55---2.85-->Node81---2.85-->Node89---2.85-->Node97---2.85-->t18
#>  [78] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17---2.81-->Node34---2.82-->Node37---2.84-->Node55---2.85-->Node81---2.85-->Node89---2.85-->Node97---2.85-->t92
#>  [79] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17---2.81-->Node34---2.82-->Node37---2.84-->Node55---2.85-->t30                                                
#>  [80] ---2.44-->Node5---2.59-->Node8---2.80-->Node27---2.85-->t20                                                                                                                                
#>  [81] ---2.44-->Node5---2.59-->Node8---2.80-->Node27---2.85-->t12                                                                                                                                
#>  [82] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.85-->Node98---2.85-->Node99---2.85-->t79                                                                                                
#>  [83] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.85-->Node98---2.85-->Node99---2.85-->t19                                                                                                
#>  [84] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.85-->Node98---2.85-->t60                                                                                                                
#>  [85] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.79-->Node24---2.84-->Node64---2.85-->Node85---2.85-->t1                                                                 
#>  [86] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.79-->Node24---2.84-->Node64---2.85-->Node85---2.85-->Node94---2.85-->t42                                                
#>  [87] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.79-->Node24---2.84-->Node64---2.85-->Node85---2.85-->Node94---2.85-->t66                                                
#>  [88] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.79-->Node24---2.84-->Node64---2.85-->t87                                                                                
#>  [89] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.79-->Node24---2.85-->Node87---2.85-->t22                                                                                
#>  [90] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.79-->Node24---2.85-->Node87---2.85-->t91                                                                                
#>  [91] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.82-->Node35---2.83-->Node41---2.84-->Node56---2.84-->Node63---2.85-->t10                                
#>  [92] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.82-->Node35---2.83-->Node41---2.84-->Node56---2.84-->Node63---2.85-->t26                                
#>  [93] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.82-->Node35---2.83-->Node41---2.84-->Node56---2.85-->Node96---2.85-->t62                                
#>  [94] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.82-->Node35---2.83-->Node41---2.84-->Node56---2.85-->Node96---2.85-->t64                                
#>  [95] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.82-->Node35---2.83-->Node41---2.85-->Node78---2.85-->t45                                                
#>  [96] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.82-->Node35---2.83-->Node41---2.85-->Node78---2.85-->t73                                                
#>  [97] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.82-->Node35---2.85-->Node92---2.85-->t78                                                                
#>  [98] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.82-->Node35---2.85-->Node92---2.85-->t37                                                                
#>  [99] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.83-->Node49---2.85-->t53                                                                                
#> [100] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.83-->Node49---2.85-->t56                                                                                
#> [101] ---1.76-->Node2                                                                                                                                                                            
#> [102] ---1.76-->Node2---2.03-->Node3                                                                                                                                                             
#> [103] ---1.76-->Node2---2.21-->Node4                                                                                                                                                             
#> [104] ---2.44-->Node5                                                                                                                                                                            
#> [105] ---1.76-->Node2---2.03-->Node3---2.52-->Node6                                                                                                                                              
#> [106] ---2.44-->Node5---2.58-->Node7                                                                                                                                                             
#> [107] ---2.44-->Node5---2.59-->Node8                                                                                                                                                             
#> [108] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9                                                                                                                               
#> [109] ---2.44-->Node5---2.58-->Node7---2.63-->Node10                                                                                                                                             
#> [110] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.67-->Node11                                                                                                               
#> [111] ---2.44-->Node5---2.59-->Node8---2.68-->Node12                                                                                                                                             
#> [112] ---2.44-->Node5---2.58-->Node7---2.69-->Node13                                                                                                                                             
#> [113] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14                                                                                                                             
#> [114] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15                                                                                                                              
#> [115] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.67-->Node11---2.75-->Node16                                                                                               
#> [116] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17                                                                                                             
#> [117] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.76-->Node18                                                                                                              
#> [118] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19                                                                                                                             
#> [119] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.77-->Node20                                                                                                               
#> [120] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21                                                                                                             
#> [121] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.67-->Node11---2.75-->Node16---2.78-->Node22                                                                               
#> [122] ---1.76-->Node2---2.21-->Node4---2.79-->Node23                                                                                                                                             
#> [123] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.79-->Node24                                                                                                             
#> [124] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25                                                                                                                             
#> [125] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26                                                                                                                             
#> [126] ---2.44-->Node5---2.59-->Node8---2.80-->Node27                                                                                                                                             
#> [127] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.76-->Node18---2.80-->Node28                                                                                              
#> [128] ---2.44-->Node5---2.58-->Node7---2.69-->Node13---2.80-->Node29                                                                                                                             
#> [129] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.77-->Node20---2.80-->Node30                                                                                               
#> [130] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31                                                                                                             
#> [131] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31---2.81-->Node32                                                                                             
#> [132] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.81-->Node33                                                                                                             
#> [133] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17---2.81-->Node34                                                                                             
#> [134] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.82-->Node35                                                                                             
#> [135] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.67-->Node11---2.75-->Node16---2.82-->Node36                                                                               
#> [136] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17---2.81-->Node34---2.82-->Node37                                                                             
#> [137] ---2.44-->Node5---2.58-->Node7---2.69-->Node13---2.80-->Node29---2.83-->Node38                                                                                                             
#> [138] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.83-->Node39                                                                                                             
#> [139] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.83-->Node40                                                                                                                             
#> [140] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.82-->Node35---2.83-->Node41                                                                             
#> [141] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.67-->Node11---2.75-->Node16---2.82-->Node36---2.83-->Node42                                                               
#> [142] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.81-->Node33---2.83-->Node43                                                                                             
#> [143] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44                                                                                                              
#> [144] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.83-->Node45                                                                                                             
#> [145] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.83-->Node39---2.83-->Node46                                                                                             
#> [146] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.81-->Node33---2.83-->Node43---2.83-->Node47                                                                             
#> [147] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31---2.81-->Node32---2.83-->Node48                                                                             
#> [148] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.83-->Node49                                                                                             
#> [149] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31---2.81-->Node32---2.83-->Node48---2.83-->Node50                                                             
#> [150] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.83-->Node51                                                                                              
#> [151] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31---2.81-->Node32---2.83-->Node52                                                                             
#> [152] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.84-->Node53                                                                                              
#> [153] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31---2.84-->Node54                                                                                             
#> [154] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17---2.81-->Node34---2.82-->Node37---2.84-->Node55                                                             
#> [155] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.82-->Node35---2.83-->Node41---2.84-->Node56                                                             
#> [156] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.81-->Node33---2.83-->Node43---2.83-->Node47---2.84-->Node57                                                             
#> [157] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31---2.81-->Node32---2.83-->Node48---2.84-->Node58                                                             
#> [158] ---2.44-->Node5---2.58-->Node7---2.69-->Node13---2.80-->Node29---2.84-->Node59                                                                                                             
#> [159] ---2.44-->Node5---2.58-->Node7---2.69-->Node13---2.80-->Node29---2.83-->Node38---2.84-->Node60                                                                                             
#> [160] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.83-->Node39---2.83-->Node46---2.84-->Node61                                                                             
#> [161] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.83-->Node45---2.84-->Node62                                                                                             
#> [162] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.82-->Node35---2.83-->Node41---2.84-->Node56---2.84-->Node63                                             
#> [163] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.79-->Node24---2.84-->Node64                                                                                             
#> [164] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.83-->Node51---2.84-->Node65                                                                              
#> [165] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.83-->Node40---2.84-->Node66                                                                                                             
#> [166] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.77-->Node20---2.84-->Node67                                                                                               
#> [167] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.84-->Node53---2.84-->Node68                                                                              
#> [168] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.77-->Node20---2.84-->Node67---2.84-->Node69                                                                               
#> [169] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.77-->Node20---2.80-->Node30---2.84-->Node70                                                                               
#> [170] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.67-->Node11---2.75-->Node16---2.78-->Node22---2.84-->Node71                                                               
#> [171] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.84-->Node72                                                                                                             
#> [172] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.81-->Node33---2.84-->Node73                                                                                             
#> [173] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.83-->Node51---2.84-->Node65---2.84-->Node74                                                              
#> [174] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.81-->Node31---2.81-->Node32---2.83-->Node48---2.84-->Node58---2.84-->Node75                                             
#> [175] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.83-->Node51---2.84-->Node65---2.84-->Node74---2.84-->Node76                                              
#> [176] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.81-->Node33---2.84-->Node73---2.85-->Node77                                                                             
#> [177] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.82-->Node35---2.83-->Node41---2.85-->Node78                                                             
#> [178] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.79-->Node25---2.81-->Node33---2.83-->Node43---2.83-->Node47---2.84-->Node57---2.85-->Node79                                             
#> [179] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.83-->Node40---2.84-->Node66---2.85-->Node80                                                                                             
#> [180] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17---2.81-->Node34---2.82-->Node37---2.84-->Node55---2.85-->Node81                                             
#> [181] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.60-->Node9---2.67-->Node11---2.75-->Node16---2.78-->Node22---2.85-->Node82                                                               
#> [182] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.84-->Node53---2.85-->Node83                                                                              
#> [183] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.76-->Node18---2.80-->Node28---2.85-->Node84                                                                              
#> [184] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.79-->Node24---2.84-->Node64---2.85-->Node85                                                                             
#> [185] ---2.44-->Node5---2.58-->Node7---2.69-->Node13---2.80-->Node29---2.83-->Node38---2.84-->Node60---2.85-->Node86                                                                             
#> [186] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.79-->Node24---2.85-->Node87                                                                                             
#> [187] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.83-->Node39---2.83-->Node46---2.84-->Node61---2.85-->Node88                                                             
#> [188] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17---2.81-->Node34---2.82-->Node37---2.84-->Node55---2.85-->Node81---2.85-->Node89                             
#> [189] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.83-->Node39---2.83-->Node46---2.85-->Node90                                                                             
#> [190] ---1.76-->Node2---2.21-->Node4---2.79-->Node23---2.80-->Node26---2.83-->Node39---2.83-->Node46---2.84-->Node61---2.85-->Node88---2.85-->Node91                                             
#> [191] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.82-->Node35---2.85-->Node92                                                                             
#> [192] ---1.76-->Node2---2.03-->Node3---2.52-->Node6---2.74-->Node15---2.83-->Node44---2.83-->Node51---2.85-->Node93                                                                              
#> [193] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.79-->Node24---2.84-->Node64---2.85-->Node85---2.85-->Node94                                                             
#> [194] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17---2.81-->Node34---2.82-->Node37---2.84-->Node55---2.85-->Node81---2.85-->Node89---2.85-->Node95             
#> [195] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.76-->Node19---2.77-->Node21---2.82-->Node35---2.83-->Node41---2.84-->Node56---2.85-->Node96                                             
#> [196] ---2.44-->Node5---2.58-->Node7---2.63-->Node10---2.70-->Node14---2.76-->Node17---2.81-->Node34---2.82-->Node37---2.84-->Node55---2.85-->Node81---2.85-->Node89---2.85-->Node97             
#> [197] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.85-->Node98                                                                                                                             
#> [198] ---2.44-->Node5---2.59-->Node8---2.68-->Node12---2.85-->Node98---2.85-->Node99