Calculate the sum of features for each flow

Cumulative sum the phylogenetic flow features for each flow in a pfc, returning a pfc.

Usage

pf_flow_cumsum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A pfc object with cumulative sums of features in place of the original features

Examples

pf_flow_cumsum(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] ---2.88-->Node2---3.78-->Node24---3.81-->Node40---3.83-->Node65---3.84-->t8                                                                                                                                                                 
#>   [2] ---2.88-->Node2---3.78-->Node24---3.81-->Node40---3.83-->Node65---3.84-->t33                                                                                                                                                                
#>   [3] ---2.88-->Node2---3.78-->Node24---3.81-->Node40---3.84-->t85                                                                                                                                                                                
#>   [4] ---2.88-->Node2---3.78-->Node24---3.83-->Node52---3.84-->t49                                                                                                                                                                                
#>   [5] ---2.88-->Node2---3.78-->Node24---3.83-->Node52---3.84-->t2                                                                                                                                                                                 
#>   [6] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12---3.80-->Node33---3.84-->t59                                                                                                                                                  
#>   [7] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12---3.80-->Node33---3.83-->Node68---3.84-->t51                                                                                                                                  
#>   [8] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12---3.80-->Node33---3.83-->Node68---3.83-->Node69---3.83-->Node77---3.84-->t98                                                                                                  
#>   [9] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12---3.80-->Node33---3.83-->Node68---3.83-->Node69---3.83-->Node77---3.84-->t89                                                                                                  
#>  [10] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12---3.80-->Node33---3.83-->Node68---3.83-->Node69---3.84-->t96                                                                                                                  
#>  [11] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12---3.82-->Node43---3.83-->Node49---3.83-->Node64---3.84-->t76                                                                                                                  
#>  [12] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12---3.82-->Node43---3.83-->Node49---3.83-->Node64---3.84-->t73                                                                                                                  
#>  [13] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12---3.82-->Node43---3.83-->Node49---3.84-->t9                                                                                                                                   
#>  [14] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12---3.82-->Node43---3.84-->t69                                                                                                                                                  
#>  [15] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.62-->Node10---3.72-->Node15---3.81-->Node39---3.83-->Node47---3.84-->t1                                                                      
#>  [16] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.62-->Node10---3.72-->Node15---3.81-->Node39---3.83-->Node47---3.84-->t37                                                                     
#>  [17] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.62-->Node10---3.72-->Node15---3.81-->Node39---3.84-->Node98---3.84-->t71                                                                     
#>  [18] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.62-->Node10---3.72-->Node15---3.81-->Node39---3.84-->Node98---3.84-->t41                                                                     
#>  [19] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.62-->Node10---3.72-->Node15---3.84-->t12                                                                                                     
#>  [20] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.62-->Node10---3.84-->t16                                                                                                                     
#>  [21] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.80-->Node31---3.83-->Node78---3.84-->t45                                                                                     
#>  [22] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.80-->Node31---3.83-->Node78---3.84-->Node92---3.84-->t61                                                                     
#>  [23] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.80-->Node31---3.83-->Node78---3.84-->Node92---3.84-->t27                                                                     
#>  [24] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.80-->Node31---3.83-->Node79---3.84-->Node87---3.84-->t34                                                                     
#>  [25] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.80-->Node31---3.83-->Node79---3.84-->Node87---3.84-->t63                                                                     
#>  [26] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.80-->Node31---3.83-->Node79---3.84-->Node93---3.84-->t50                                                                     
#>  [27] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.80-->Node31---3.83-->Node79---3.84-->Node93---3.84-->t57                                                                     
#>  [28] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.81-->Node38---3.84-->Node94---3.84-->t38                                                                                     
#>  [29] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.81-->Node38---3.84-->Node94---3.84-->t78                                                                                     
#>  [30] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.81-->Node38---3.83-->Node70---3.84-->t53                                                                                     
#>  [31] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.81-->Node38---3.83-->Node70---3.84-->Node86---3.84-->t35                                                                     
#>  [32] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.81-->Node38---3.83-->Node70---3.84-->Node86---3.84-->t13                                                                     
#>  [33] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.76-->Node19---3.83-->Node59---3.84-->t17                                                                                                                    
#>  [34] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.76-->Node19---3.83-->Node59---3.84-->Node97---3.84-->t15                                                                                                    
#>  [35] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.76-->Node19---3.83-->Node59---3.84-->Node97---3.84-->t40                                                                                                    
#>  [36] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.76-->Node19---3.84-->t87                                                                                                                                    
#>  [37] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.84-->t66                                                                                                                                                                   
#>  [38] ---2.88-->Node2---3.12-->Node3---3.58-->Node8---3.84-->t5                                                                                                                                                                                   
#>  [39] ---2.88-->Node2---3.12-->Node3---3.58-->Node8---3.80-->Node34---3.83-->Node62---3.84-->t7                                                                                                                                                   
#>  [40] ---2.88-->Node2---3.12-->Node3---3.58-->Node8---3.80-->Node34---3.83-->Node62---3.84-->t22                                                                                                                                                  
#>  [41] ---2.88-->Node2---3.12-->Node3---3.58-->Node8---3.80-->Node34---3.82-->Node45---3.83-->Node51---3.83-->Node57---3.84-->t46                                                                                                                  
#>  [42] ---2.88-->Node2---3.12-->Node3---3.58-->Node8---3.80-->Node34---3.82-->Node45---3.83-->Node51---3.83-->Node57---3.83-->Node84---3.84-->t95                                                                                                  
#>  [43] ---2.88-->Node2---3.12-->Node3---3.58-->Node8---3.80-->Node34---3.82-->Node45---3.83-->Node51---3.83-->Node57---3.83-->Node84---3.84-->t44                                                                                                  
#>  [44] ---2.88-->Node2---3.12-->Node3---3.58-->Node8---3.80-->Node34---3.82-->Node45---3.83-->Node51---3.83-->Node81---3.84-->t55                                                                                                                  
#>  [45] ---2.88-->Node2---3.12-->Node3---3.58-->Node8---3.80-->Node34---3.82-->Node45---3.83-->Node51---3.83-->Node81---3.84-->t18                                                                                                                  
#>  [46] ---2.88-->Node2---3.12-->Node3---3.58-->Node8---3.80-->Node34---3.82-->Node45---3.84-->t36                                                                                                                                                  
#>  [47] ---3.51-->Node5---3.70-->Node13---3.83-->Node83---3.84-->t68                                                                                                                                                                                
#>  [48] ---3.51-->Node5---3.70-->Node13---3.83-->Node83---3.84-->t94                                                                                                                                                                                
#>  [49] ---3.51-->Node5---3.70-->Node13---3.73-->Node16---3.84-->t58                                                                                                                                                                                
#>  [50] ---3.51-->Node5---3.70-->Node13---3.73-->Node16---3.76-->Node20---3.82-->Node44---3.84-->t82                                                                                                                                                
#>  [51] ---3.51-->Node5---3.70-->Node13---3.73-->Node16---3.76-->Node20---3.82-->Node44---3.83-->Node50---3.84-->t26                                                                                                                                
#>  [52] ---3.51-->Node5---3.70-->Node13---3.73-->Node16---3.76-->Node20---3.82-->Node44---3.83-->Node50---3.84-->t29                                                                                                                                
#>  [53] ---3.51-->Node5---3.70-->Node13---3.73-->Node16---3.76-->Node20---3.84-->t24                                                                                                                                                                
#>  [54] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.83-->Node74---3.84-->t62                                                                                                                                                
#>  [55] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.83-->Node74---3.84-->t10                                                                                                                                                
#>  [56] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.78-->Node27---3.78-->Node28---3.84-->t54                                                                                                                                
#>  [57] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.78-->Node27---3.78-->Node28---3.81-->Node36---3.84-->t84                                                                                                                
#>  [58] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.78-->Node27---3.78-->Node28---3.81-->Node36---3.84-->Node90---3.84-->t47                                                                                                
#>  [59] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.78-->Node27---3.78-->Node28---3.81-->Node36---3.84-->Node90---3.84-->t79                                                                                                
#>  [60] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.78-->Node27---3.80-->Node32---3.83-->Node80---3.84-->t74                                                                                                                
#>  [61] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.78-->Node27---3.80-->Node32---3.83-->Node80---3.84-->t6                                                                                                                 
#>  [62] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.78-->Node27---3.80-->Node32---3.84-->Node88---3.84-->t64                                                                                                                
#>  [63] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.78-->Node27---3.80-->Node32---3.84-->Node88---3.84-->t93                                                                                                                
#>  [64] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.84-->t81                                                                                                                                                                
#>  [65] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node25---3.83-->Node55---3.83-->Node66---3.84-->t90                                                                                                
#>  [66] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node25---3.83-->Node55---3.83-->Node66---3.84-->t25                                                                                                
#>  [67] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node25---3.83-->Node55---3.84-->t70                                                                                                                
#>  [68] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node25---3.84-->t23                                                                                                                                
#>  [69] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.84-->Node99---3.84-->t48                                                                                                                
#>  [70] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.84-->Node99---3.84-->t32                                                                                                                
#>  [71] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.82-->Node46---3.83-->Node60---3.84-->t14                                                                                
#>  [72] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.82-->Node46---3.83-->Node60---3.83-->Node82---3.84-->t91                                                                
#>  [73] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.82-->Node46---3.83-->Node60---3.83-->Node82---3.84-->t72                                                                
#>  [74] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.82-->Node46---3.84-->Node89---3.84-->t77                                                                                
#>  [75] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.82-->Node46---3.84-->Node89---3.84-->t99                                                                                
#>  [76] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41---3.82-->Node42---3.84-->t21                                                                
#>  [77] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41---3.82-->Node42---3.83-->Node48---3.84-->t30                                                
#>  [78] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41---3.82-->Node42---3.83-->Node48---3.83-->Node53---3.84-->t3                                 
#>  [79] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41---3.82-->Node42---3.83-->Node48---3.83-->Node53---3.83-->Node61---3.84-->t28                
#>  [80] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41---3.82-->Node42---3.83-->Node48---3.83-->Node53---3.83-->Node61---3.84-->Node96---3.84-->t56
#>  [81] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41---3.82-->Node42---3.83-->Node48---3.83-->Node53---3.83-->Node61---3.84-->Node96---3.84-->t80
#>  [82] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41---3.83-->Node67---3.84-->t52                                                                
#>  [83] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41---3.83-->Node67---3.83-->Node76---3.84-->t39                                                
#>  [84] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41---3.83-->Node67---3.83-->Node76---3.84-->t4                                                 
#>  [85] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.81-->Node37---3.83-->Node54---3.83-->Node56---3.84-->t97                                                
#>  [86] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.81-->Node37---3.83-->Node54---3.83-->Node56---3.83-->Node63---3.84-->t67                                
#>  [87] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.81-->Node37---3.83-->Node54---3.83-->Node56---3.83-->Node63---3.83-->Node75---3.84-->Node85---3.84-->t86
#>  [88] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.81-->Node37---3.83-->Node54---3.83-->Node56---3.83-->Node63---3.83-->Node75---3.84-->Node85---3.84-->t43
#>  [89] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.81-->Node37---3.83-->Node54---3.83-->Node56---3.83-->Node63---3.83-->Node75---3.84-->t100               
#>  [90] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.81-->Node37---3.83-->Node54---3.83-->Node73---3.84-->t11                                                
#>  [91] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.81-->Node37---3.83-->Node54---3.83-->Node73---3.84-->t42                                                
#>  [92] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.81-->Node37---3.84-->t19                                                                                
#>  [93] ---3.51-->Node5---3.66-->Node11---3.79-->Node29---3.83-->Node58---3.83-->Node72---3.84-->t60                                                                                                                                                
#>  [94] ---3.51-->Node5---3.66-->Node11---3.79-->Node29---3.83-->Node58---3.83-->Node72---3.84-->Node91---3.84-->t65                                                                                                                                
#>  [95] ---3.51-->Node5---3.66-->Node11---3.79-->Node29---3.83-->Node58---3.83-->Node72---3.84-->Node91---3.84-->t92                                                                                                                                
#>  [96] ---3.51-->Node5---3.66-->Node11---3.79-->Node29---3.83-->Node58---3.84-->Node95---3.84-->t31                                                                                                                                                
#>  [97] ---3.51-->Node5---3.66-->Node11---3.79-->Node29---3.83-->Node58---3.84-->Node95---3.84-->t88                                                                                                                                                
#>  [98] ---3.51-->Node5---3.66-->Node11---3.79-->Node29---3.80-->Node30---3.83-->Node71---3.84-->t83                                                                                                                                                
#>  [99] ---3.51-->Node5---3.66-->Node11---3.79-->Node29---3.80-->Node30---3.83-->Node71---3.84-->t20                                                                                                                                                
#> [100] ---3.51-->Node5---3.66-->Node11---3.79-->Node29---3.80-->Node30---3.84-->t75                                                                                                                                                                
#> [101] ---2.88-->Node2                                                                                                                                                                                                                             
#> [102] ---2.88-->Node2---3.12-->Node3                                                                                                                                                                                                              
#> [103] ---2.88-->Node2---3.12-->Node3---3.18-->Node4                                                                                                                                                                                               
#> [104] ---3.51-->Node5                                                                                                                                                                                                                             
#> [105] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6                                                                                                                                                                                
#> [106] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7                                                                                                                                                                 
#> [107] ---2.88-->Node2---3.12-->Node3---3.58-->Node8                                                                                                                                                                                               
#> [108] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9                                                                                                                                                  
#> [109] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.62-->Node10                                                                                                                                  
#> [110] ---3.51-->Node5---3.66-->Node11                                                                                                                                                                                                             
#> [111] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12                                                                                                                                                                               
#> [112] ---3.51-->Node5---3.70-->Node13                                                                                                                                                                                                             
#> [113] ---3.51-->Node5---3.66-->Node11---3.71-->Node14                                                                                                                                                                                             
#> [114] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.62-->Node10---3.72-->Node15                                                                                                                  
#> [115] ---3.51-->Node5---3.70-->Node13---3.73-->Node16                                                                                                                                                                                             
#> [116] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17                                                                                                                                                                             
#> [117] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18                                                                                                                                                                             
#> [118] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.76-->Node19                                                                                                                                                 
#> [119] ---3.51-->Node5---3.70-->Node13---3.73-->Node16---3.76-->Node20                                                                                                                                                                             
#> [120] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21                                                                                                                                                             
#> [121] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22                                                                                                                                             
#> [122] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23                                                                                                                                  
#> [123] ---2.88-->Node2---3.78-->Node24                                                                                                                                                                                                             
#> [124] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node25                                                                                                                                             
#> [125] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26                                                                                                                             
#> [126] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.78-->Node27                                                                                                                                                             
#> [127] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.78-->Node27---3.78-->Node28                                                                                                                                             
#> [128] ---3.51-->Node5---3.66-->Node11---3.79-->Node29                                                                                                                                                                                             
#> [129] ---3.51-->Node5---3.66-->Node11---3.79-->Node29---3.80-->Node30                                                                                                                                                                             
#> [130] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.80-->Node31                                                                                                                  
#> [131] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.78-->Node27---3.80-->Node32                                                                                                                                             
#> [132] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12---3.80-->Node33                                                                                                                                                               
#> [133] ---2.88-->Node2---3.12-->Node3---3.58-->Node8---3.80-->Node34                                                                                                                                                                               
#> [134] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35                                                                                                             
#> [135] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.78-->Node27---3.78-->Node28---3.81-->Node36                                                                                                                             
#> [136] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.81-->Node37                                                                                             
#> [137] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.81-->Node38                                                                                                                  
#> [138] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.62-->Node10---3.72-->Node15---3.81-->Node39                                                                                                  
#> [139] ---2.88-->Node2---3.78-->Node24---3.81-->Node40                                                                                                                                                                                             
#> [140] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41                                                                                             
#> [141] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41---3.82-->Node42                                                                             
#> [142] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12---3.82-->Node43                                                                                                                                                               
#> [143] ---3.51-->Node5---3.70-->Node13---3.73-->Node16---3.76-->Node20---3.82-->Node44                                                                                                                                                             
#> [144] ---2.88-->Node2---3.12-->Node3---3.58-->Node8---3.80-->Node34---3.82-->Node45                                                                                                                                                               
#> [145] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.82-->Node46                                                                                                             
#> [146] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.62-->Node10---3.72-->Node15---3.81-->Node39---3.83-->Node47                                                                                  
#> [147] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41---3.82-->Node42---3.83-->Node48                                                             
#> [148] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12---3.82-->Node43---3.83-->Node49                                                                                                                                               
#> [149] ---3.51-->Node5---3.70-->Node13---3.73-->Node16---3.76-->Node20---3.82-->Node44---3.83-->Node50                                                                                                                                             
#> [150] ---2.88-->Node2---3.12-->Node3---3.58-->Node8---3.80-->Node34---3.82-->Node45---3.83-->Node51                                                                                                                                               
#> [151] ---2.88-->Node2---3.78-->Node24---3.83-->Node52                                                                                                                                                                                             
#> [152] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41---3.82-->Node42---3.83-->Node48---3.83-->Node53                                             
#> [153] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.81-->Node37---3.83-->Node54                                                                             
#> [154] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node25---3.83-->Node55                                                                                                                             
#> [155] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.81-->Node37---3.83-->Node54---3.83-->Node56                                                             
#> [156] ---2.88-->Node2---3.12-->Node3---3.58-->Node8---3.80-->Node34---3.82-->Node45---3.83-->Node51---3.83-->Node57                                                                                                                               
#> [157] ---3.51-->Node5---3.66-->Node11---3.79-->Node29---3.83-->Node58                                                                                                                                                                             
#> [158] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.76-->Node19---3.83-->Node59                                                                                                                                 
#> [159] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.82-->Node46---3.83-->Node60                                                                                             
#> [160] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41---3.82-->Node42---3.83-->Node48---3.83-->Node53---3.83-->Node61                             
#> [161] ---2.88-->Node2---3.12-->Node3---3.58-->Node8---3.80-->Node34---3.83-->Node62                                                                                                                                                               
#> [162] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.81-->Node37---3.83-->Node54---3.83-->Node56---3.83-->Node63                                             
#> [163] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12---3.82-->Node43---3.83-->Node49---3.83-->Node64                                                                                                                               
#> [164] ---2.88-->Node2---3.78-->Node24---3.81-->Node40---3.83-->Node65                                                                                                                                                                             
#> [165] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node25---3.83-->Node55---3.83-->Node66                                                                                                             
#> [166] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41---3.83-->Node67                                                                             
#> [167] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12---3.80-->Node33---3.83-->Node68                                                                                                                                               
#> [168] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12---3.80-->Node33---3.83-->Node68---3.83-->Node69                                                                                                                               
#> [169] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.81-->Node38---3.83-->Node70                                                                                                  
#> [170] ---3.51-->Node5---3.66-->Node11---3.79-->Node29---3.80-->Node30---3.83-->Node71                                                                                                                                                             
#> [171] ---3.51-->Node5---3.66-->Node11---3.79-->Node29---3.83-->Node58---3.83-->Node72                                                                                                                                                             
#> [172] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.81-->Node37---3.83-->Node54---3.83-->Node73                                                             
#> [173] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.83-->Node74                                                                                                                                                             
#> [174] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.81-->Node37---3.83-->Node54---3.83-->Node56---3.83-->Node63---3.83-->Node75                             
#> [175] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41---3.83-->Node67---3.83-->Node76                                                             
#> [176] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.69-->Node12---3.80-->Node33---3.83-->Node68---3.83-->Node69---3.83-->Node77                                                                                                               
#> [177] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.80-->Node31---3.83-->Node78                                                                                                  
#> [178] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.80-->Node31---3.83-->Node79                                                                                                  
#> [179] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.78-->Node27---3.80-->Node32---3.83-->Node80                                                                                                                             
#> [180] ---2.88-->Node2---3.12-->Node3---3.58-->Node8---3.80-->Node34---3.82-->Node45---3.83-->Node51---3.83-->Node81                                                                                                                               
#> [181] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.82-->Node46---3.83-->Node60---3.83-->Node82                                                                             
#> [182] ---3.51-->Node5---3.70-->Node13---3.83-->Node83                                                                                                                                                                                             
#> [183] ---2.88-->Node2---3.12-->Node3---3.58-->Node8---3.80-->Node34---3.82-->Node45---3.83-->Node51---3.83-->Node57---3.83-->Node84                                                                                                               
#> [184] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.81-->Node37---3.83-->Node54---3.83-->Node56---3.83-->Node63---3.83-->Node75---3.84-->Node85             
#> [185] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.81-->Node38---3.83-->Node70---3.84-->Node86                                                                                  
#> [186] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.80-->Node31---3.83-->Node79---3.84-->Node87                                                                                  
#> [187] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.78-->Node27---3.80-->Node32---3.84-->Node88                                                                                                                             
#> [188] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.82-->Node46---3.84-->Node89                                                                                             
#> [189] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node18---3.78-->Node27---3.78-->Node28---3.81-->Node36---3.84-->Node90                                                                                                             
#> [190] ---3.51-->Node5---3.66-->Node11---3.79-->Node29---3.83-->Node58---3.83-->Node72---3.84-->Node91                                                                                                                                             
#> [191] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.80-->Node31---3.83-->Node78---3.84-->Node92                                                                                  
#> [192] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.80-->Node31---3.83-->Node79---3.84-->Node93                                                                                  
#> [193] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.78-->Node23---3.81-->Node38---3.84-->Node94                                                                                                  
#> [194] ---3.51-->Node5---3.66-->Node11---3.79-->Node29---3.83-->Node58---3.84-->Node95                                                                                                                                                             
#> [195] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.78-->Node26---3.81-->Node35---3.82-->Node41---3.82-->Node42---3.83-->Node48---3.83-->Node53---3.83-->Node61---3.84-->Node96             
#> [196] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.76-->Node19---3.83-->Node59---3.84-->Node97                                                                                                                 
#> [197] ---2.88-->Node2---3.12-->Node3---3.18-->Node4---3.53-->Node6---3.58-->Node7---3.61-->Node9---3.62-->Node10---3.72-->Node15---3.81-->Node39---3.84-->Node98                                                                                  
#> [198] ---3.51-->Node5---3.66-->Node11---3.71-->Node14---3.74-->Node17---3.77-->Node21---3.78-->Node22---3.84-->Node99