Calculate the sum of features for each flow

Cumulative sum the phylogenetic flow features for each flow in a pfc, returning a pfc.

Usage

pf_flow_cumsum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A pfc object with cumulative sums of features in place of the original features

Examples

pf_flow_cumsum(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] ---1.27-->Node6---1.51-->Node11---1.66-->t47                                                                                                                                                                                               
#>   [2] ---1.27-->Node6---1.51-->Node11---1.63-->Node45---1.66-->t46                                                                                                                                                                               
#>   [3] ---1.27-->Node6---1.51-->Node11---1.63-->Node45---1.66-->t59                                                                                                                                                                               
#>   [4] ---1.27-->Node6---1.59-->Node21---1.65-->Node63---1.66-->Node78---1.66-->Node83---1.66-->t89                                                                                                                                               
#>   [5] ---1.27-->Node6---1.59-->Node21---1.65-->Node63---1.66-->Node78---1.66-->Node83---1.66-->t58                                                                                                                                               
#>   [6] ---1.27-->Node6---1.59-->Node21---1.65-->Node63---1.66-->Node78---1.66-->Node87---1.66-->t16                                                                                                                                               
#>   [7] ---1.27-->Node6---1.59-->Node21---1.65-->Node63---1.66-->Node78---1.66-->Node87---1.66-->t38                                                                                                                                               
#>   [8] ---1.27-->Node6---1.59-->Node21---1.65-->Node63---1.66-->Node81---1.66-->t62                                                                                                                                                               
#>   [9] ---1.27-->Node6---1.59-->Node21---1.65-->Node63---1.66-->Node81---1.66-->t19                                                                                                                                                               
#>  [10] ---1.27-->Node6---1.59-->Node21---1.59-->Node22---1.65-->Node77---1.66-->t57                                                                                                                                                               
#>  [11] ---1.27-->Node6---1.59-->Node21---1.59-->Node22---1.65-->Node77---1.66-->t55                                                                                                                                                               
#>  [12] ---1.27-->Node6---1.59-->Node21---1.59-->Node22---1.66-->t1                                                                                                                                                                                
#>  [13] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17--1.619-->Node33--1.660-->t60                                                                                                                                
#>  [14] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17--1.619-->Node33--1.640-->Node52--1.660-->t33                                                                                                                
#>  [15] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17--1.619-->Node33--1.640-->Node52--1.653-->Node70--1.658-->Node90--1.660-->t18                                                                                
#>  [16] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17--1.619-->Node33--1.640-->Node52--1.653-->Node70--1.658-->Node90--1.660-->t85                                                                                
#>  [17] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17--1.619-->Node33--1.640-->Node52--1.653-->Node70--1.657-->Node85--1.660-->t80                                                                                
#>  [18] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17--1.619-->Node33--1.640-->Node52--1.653-->Node70--1.657-->Node85--1.660-->t51                                                                                
#>  [19] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17--1.633-->Node44--1.660-->t40                                                                                                                                
#>  [20] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17--1.633-->Node44--1.655-->Node76--1.660-->t100                                                                                                               
#>  [21] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17--1.633-->Node44--1.655-->Node76--1.660-->t30                                                                                                                
#>  [22] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node43--1.660-->t4                                                 
#>  [23] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node43--1.648-->Node60--1.654-->Node74--1.660-->t31                
#>  [24] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node43--1.648-->Node60--1.654-->Node74--1.656-->Node80--1.660-->t78
#>  [25] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node43--1.648-->Node60--1.654-->Node74--1.656-->Node80--1.660-->t26
#>  [26] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node43--1.648-->Node60--1.660-->t2                                 
#>  [27] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node42--1.640-->Node53--1.660-->t83                                
#>  [28] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node42--1.640-->Node53--1.653-->Node68--1.656-->Node84--1.660-->t84
#>  [29] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node42--1.640-->Node53--1.653-->Node68--1.656-->Node84--1.660-->t82
#>  [30] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node42--1.640-->Node53--1.653-->Node68--1.660-->t66                
#>  [31] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node42--1.660-->t9                                                 
#>  [32] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.617-->Node32--1.660-->t53                                                                
#>  [33] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.617-->Node32--1.631-->Node39--1.659-->Node94--1.660-->t63                                
#>  [34] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.617-->Node32--1.631-->Node39--1.659-->Node94--1.660-->t75                                
#>  [35] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.617-->Node32--1.631-->Node39--1.659-->Node93--1.660-->t3                                 
#>  [36] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.617-->Node32--1.631-->Node39--1.659-->Node93--1.660-->t29                                
#>  [37] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.612-->Node28--1.615-->Node30--1.652-->Node66--1.660-->t15                                                
#>  [38] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.612-->Node28--1.615-->Node30--1.652-->Node66--1.660-->t98                                                
#>  [39] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.612-->Node28--1.615-->Node30--1.660-->t68                                                                
#>  [40] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.612-->Node28--1.643-->Node56--1.660-->Node98--1.660-->t88                                                
#>  [41] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.612-->Node28--1.643-->Node56--1.660-->Node98--1.660-->t74                                                
#>  [42] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.612-->Node28--1.643-->Node56--1.660-->t92                                                                
#>  [43] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.632-->Node41--1.638-->Node48--1.653-->Node71--1.660-->t35                                                                
#>  [44] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.632-->Node41--1.638-->Node48--1.653-->Node71--1.660-->t64                                                                
#>  [45] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.632-->Node41--1.638-->Node48--1.660-->t93                                                                                
#>  [46] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.632-->Node41--1.660-->t90                                                                                                
#>  [47] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.614-->Node29--1.627-->Node37--1.660-->t61                                                                                                
#>  [48] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.614-->Node29--1.627-->Node37--1.639-->Node51--1.660-->t77                                                                                
#>  [49] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.614-->Node29--1.627-->Node37--1.639-->Node51--1.660-->t99                                                                                
#>  [50] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.614-->Node29--1.660-->t41                                                                                                                
#>  [51] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.654-->Node73--1.660-->t10                                                                                                                                
#>  [52] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.654-->Node73--1.660-->t72                                                                                                                                
#>  [53] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.585-->Node20--1.595-->Node25--1.638-->Node49--1.660-->t94                                                                                                                
#>  [54] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.585-->Node20--1.595-->Node25--1.638-->Node49--1.660-->t27                                                                                                                
#>  [55] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.585-->Node20--1.595-->Node25--1.643-->Node55--1.652-->Node64--1.659-->Node96--1.660-->t12                                                                                
#>  [56] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.585-->Node20--1.595-->Node25--1.643-->Node55--1.652-->Node64--1.659-->Node96--1.660-->t73                                                                                
#>  [57] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.585-->Node20--1.595-->Node25--1.643-->Node55--1.652-->Node64--1.660-->t37                                                                                                
#>  [58] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.585-->Node20--1.595-->Node25--1.643-->Node55--1.660-->t44                                                                                                                
#>  [59] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.585-->Node20--1.647-->Node59--1.660-->t49                                                                                                                                
#>  [60] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.585-->Node20--1.647-->Node59--1.660-->t65                                                                                                                                
#>  [61] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.537-->Node14--1.660-->t52                                                                                                                                                
#>  [62] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.537-->Node14--1.584-->Node19--1.649-->Node61--1.660-->t28                                                                                                                
#>  [63] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.537-->Node14--1.584-->Node19--1.649-->Node61--1.660-->t70                                                                                                                
#>  [64] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.537-->Node14--1.584-->Node19--1.620-->Node34--1.641-->Node54--1.660-->t20                                                                                                
#>  [65] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.537-->Node14--1.584-->Node19--1.620-->Node34--1.641-->Node54--1.653-->Node69--1.660-->t87                                                                                
#>  [66] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.537-->Node14--1.584-->Node19--1.620-->Node34--1.641-->Node54--1.653-->Node69--1.658-->Node89--1.660-->t95                                                                
#>  [67] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.537-->Node14--1.584-->Node19--1.620-->Node34--1.641-->Node54--1.653-->Node69--1.658-->Node89--1.660-->t36                                                                
#>  [68] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.537-->Node14--1.584-->Node19--1.620-->Node34--1.660-->t97                                                                                                                
#>  [69] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.636-->Node47--1.660-->t54                                                                                                                    
#>  [70] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.636-->Node47--1.638-->Node50--1.651-->Node62--1.660-->t71                                                                                    
#>  [71] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.636-->Node47--1.638-->Node50--1.651-->Node62--1.658-->Node91--1.660-->t11                                                                    
#>  [72] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.636-->Node47--1.638-->Node50--1.651-->Node62--1.658-->Node91--1.660-->t13                                                                    
#>  [73] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.636-->Node47--1.638-->Node50--1.653-->Node67--1.660-->t76                                                                                    
#>  [74] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.636-->Node47--1.638-->Node50--1.653-->Node67--1.660-->t86                                                                                    
#>  [75] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.653-->Node72--1.660-->t79                                                                                                                    
#>  [76] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.653-->Node72--1.656-->Node82--1.660-->Node97--1.660-->t25                                                                                    
#>  [77] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.653-->Node72--1.656-->Node82--1.660-->Node97--1.660-->t67                                                                                    
#>  [78] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.653-->Node72--1.656-->Node82--1.657-->Node86--1.660-->t69                                                                                    
#>  [79] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.653-->Node72--1.656-->Node82--1.657-->Node86--1.660-->t56                                                                                    
#>  [80] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.660-->t7                                                                                                                                                    
#>  [81] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.660-->t22                                                                                                                                                                  
#>  [82] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.652-->Node65--1.660-->t5                                                                                                                                                                  
#>  [83] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.652-->Node65--1.660-->t8                                                                                                                                                                  
#>  [84] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.647-->Node58--1.659-->Node95--1.660-->t39                                                                                                                                                
#>  [85] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.647-->Node58--1.659-->Node95--1.660-->t34                                                                                                                                                
#>  [86] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.647-->Node58--1.660-->t91                                                                                                                                                                
#>  [87] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.626-->Node36--1.660-->t23                                                                                                                                                                
#>  [88] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.626-->Node36--1.629-->Node38--1.636-->Node46--1.660-->Node99--1.660-->t45                                                                                                                
#>  [89] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.626-->Node36--1.629-->Node38--1.636-->Node46--1.660-->Node99--1.660-->t42                                                                                                                
#>  [90] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.626-->Node36--1.629-->Node38--1.636-->Node46--1.644-->Node57--1.660-->t81                                                                                                                
#>  [91] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.626-->Node36--1.629-->Node38--1.636-->Node46--1.644-->Node57--1.660-->t24                                                                                                                
#>  [92] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.626-->Node36--1.629-->Node38--1.656-->Node79--1.660-->t32                                                                                                                                
#>  [93] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.626-->Node36--1.629-->Node38--1.656-->Node79--1.660-->t17                                                                                                                                
#>  [94] --0.452-->Node2--1.436-->Node9--1.546-->Node16--1.603-->Node27--1.655-->Node75--1.660-->t43                                                                                                                                                
#>  [95] --0.452-->Node2--1.436-->Node9--1.546-->Node16--1.603-->Node27--1.655-->Node75--1.660-->t6                                                                                                                                                 
#>  [96] --0.452-->Node2--1.436-->Node9--1.546-->Node16--1.603-->Node27--1.660-->t21                                                                                                                                                                
#>  [97] --0.452-->Node2--1.436-->Node9--1.546-->Node16--1.617-->Node31--1.660-->t96                                                                                                                                                                
#>  [98] --0.452-->Node2--1.436-->Node9--1.546-->Node16--1.617-->Node31--1.658-->Node88--1.658-->Node92--1.660-->t50                                                                                                                                
#>  [99] --0.452-->Node2--1.436-->Node9--1.546-->Node16--1.617-->Node31--1.658-->Node88--1.658-->Node92--1.660-->t48                                                                                                                                
#> [100] --0.452-->Node2--1.436-->Node9--1.546-->Node16--1.617-->Node31--1.658-->Node88--1.660-->t14                                                                                                                                                
#> [101] --0.452-->Node2                                                                                                                                                                                                                            
#> [102] --0.452-->Node2--0.514-->Node3                                                                                                                                                                                                             
#> [103] --0.452-->Node2--0.514-->Node3--0.907-->Node4                                                                                                                                                                                              
#> [104] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5                                                                                                                                                                               
#> [105] ---1.27-->Node6                                                                                                                                                                                                                            
#> [106] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7                                                                                                                                                                
#> [107] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8                                                                                                                                                 
#> [108] --0.452-->Node2--1.436-->Node9                                                                                                                                                                                                             
#> [109] --0.452-->Node2--0.514-->Node3--1.465-->Node10                                                                                                                                                                                             
#> [110] ---1.27-->Node6---1.51-->Node11                                                                                                                                                                                                            
#> [111] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12                                                                                                                                                                             
#> [112] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13                                                                                                                                                                             
#> [113] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.537-->Node14                                                                                                                                                             
#> [114] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15                                                                                                                                                             
#> [115] --0.452-->Node2--1.436-->Node9--1.546-->Node16                                                                                                                                                                                             
#> [116] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17                                                                                                                                                             
#> [117] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18                                                                                                                                             
#> [118] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.537-->Node14--1.584-->Node19                                                                                                                                             
#> [119] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.585-->Node20                                                                                                                                                             
#> [120] ---1.27-->Node6---1.59-->Node21                                                                                                                                                                                                            
#> [121] ---1.27-->Node6---1.59-->Node21---1.59-->Node22                                                                                                                                                                                            
#> [122] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23                                                                                                                             
#> [123] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24                                                                                                             
#> [124] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.585-->Node20--1.595-->Node25                                                                                                                                             
#> [125] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26                                                                                             
#> [126] --0.452-->Node2--1.436-->Node9--1.546-->Node16--1.603-->Node27                                                                                                                                                                             
#> [127] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.612-->Node28                                                                                             
#> [128] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.614-->Node29                                                                                                                             
#> [129] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.612-->Node28--1.615-->Node30                                                                             
#> [130] --0.452-->Node2--1.436-->Node9--1.546-->Node16--1.617-->Node31                                                                                                                                                                             
#> [131] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.617-->Node32                                                                             
#> [132] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17--1.619-->Node33                                                                                                                                             
#> [133] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.537-->Node14--1.584-->Node19--1.620-->Node34                                                                                                                             
#> [134] --0.452-->Node2--1.436-->Node9--1.623-->Node35                                                                                                                                                                                             
#> [135] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.626-->Node36                                                                                                                                                                             
#> [136] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.614-->Node29--1.627-->Node37                                                                                                             
#> [137] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.626-->Node36--1.629-->Node38                                                                                                                                                             
#> [138] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.617-->Node32--1.631-->Node39                                                             
#> [139] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40                                                                             
#> [140] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.632-->Node41                                                                                                             
#> [141] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node42                                                             
#> [142] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node43                                                             
#> [143] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17--1.633-->Node44                                                                                                                                             
#> [144] ---1.27-->Node6---1.51-->Node11---1.63-->Node45                                                                                                                                                                                            
#> [145] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.626-->Node36--1.629-->Node38--1.636-->Node46                                                                                                                                             
#> [146] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.636-->Node47                                                                                                                                 
#> [147] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.632-->Node41--1.638-->Node48                                                                                             
#> [148] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.585-->Node20--1.595-->Node25--1.638-->Node49                                                                                                                             
#> [149] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.636-->Node47--1.638-->Node50                                                                                                                 
#> [150] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.614-->Node29--1.627-->Node37--1.639-->Node51                                                                                             
#> [151] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17--1.619-->Node33--1.640-->Node52                                                                                                                             
#> [152] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node42--1.640-->Node53                                             
#> [153] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.537-->Node14--1.584-->Node19--1.620-->Node34--1.641-->Node54                                                                                                             
#> [154] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.585-->Node20--1.595-->Node25--1.643-->Node55                                                                                                                             
#> [155] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.612-->Node28--1.643-->Node56                                                                             
#> [156] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.626-->Node36--1.629-->Node38--1.636-->Node46--1.644-->Node57                                                                                                                             
#> [157] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.647-->Node58                                                                                                                                                                             
#> [158] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.585-->Node20--1.647-->Node59                                                                                                                                             
#> [159] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node43--1.648-->Node60                                             
#> [160] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.537-->Node14--1.584-->Node19--1.649-->Node61                                                                                                                             
#> [161] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.636-->Node47--1.638-->Node50--1.651-->Node62                                                                                                 
#> [162] ---1.27-->Node6---1.59-->Node21---1.65-->Node63                                                                                                                                                                                            
#> [163] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.585-->Node20--1.595-->Node25--1.643-->Node55--1.652-->Node64                                                                                                             
#> [164] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.652-->Node65                                                                                                                                                                              
#> [165] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.612-->Node28--1.615-->Node30--1.652-->Node66                                                             
#> [166] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.636-->Node47--1.638-->Node50--1.653-->Node67                                                                                                 
#> [167] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node42--1.640-->Node53--1.653-->Node68                             
#> [168] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.537-->Node14--1.584-->Node19--1.620-->Node34--1.641-->Node54--1.653-->Node69                                                                                             
#> [169] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17--1.619-->Node33--1.640-->Node52--1.653-->Node70                                                                                                             
#> [170] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.632-->Node41--1.638-->Node48--1.653-->Node71                                                                             
#> [171] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.653-->Node72                                                                                                                                 
#> [172] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.654-->Node73                                                                                                                                             
#> [173] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node43--1.648-->Node60--1.654-->Node74                             
#> [174] --0.452-->Node2--1.436-->Node9--1.546-->Node16--1.603-->Node27--1.655-->Node75                                                                                                                                                             
#> [175] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17--1.633-->Node44--1.655-->Node76                                                                                                                             
#> [176] ---1.27-->Node6---1.59-->Node21---1.59-->Node22---1.65-->Node77                                                                                                                                                                            
#> [177] ---1.27-->Node6---1.59-->Node21---1.65-->Node63---1.66-->Node78                                                                                                                                                                            
#> [178] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.626-->Node36--1.629-->Node38--1.656-->Node79                                                                                                                                             
#> [179] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node43--1.648-->Node60--1.654-->Node74--1.656-->Node80             
#> [180] ---1.27-->Node6---1.59-->Node21---1.65-->Node63---1.66-->Node81                                                                                                                                                                            
#> [181] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.653-->Node72--1.656-->Node82                                                                                                                 
#> [182] ---1.27-->Node6---1.59-->Node21---1.65-->Node63---1.66-->Node78---1.66-->Node83                                                                                                                                                            
#> [183] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.631-->Node40--1.633-->Node42--1.640-->Node53--1.653-->Node68--1.656-->Node84             
#> [184] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17--1.619-->Node33--1.640-->Node52--1.653-->Node70--1.657-->Node85                                                                                             
#> [185] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.653-->Node72--1.656-->Node82--1.657-->Node86                                                                                                 
#> [186] ---1.27-->Node6---1.59-->Node21---1.65-->Node63---1.66-->Node78---1.66-->Node87                                                                                                                                                            
#> [187] --0.452-->Node2--1.436-->Node9--1.546-->Node16--1.617-->Node31--1.658-->Node88                                                                                                                                                             
#> [188] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.537-->Node14--1.584-->Node19--1.620-->Node34--1.641-->Node54--1.653-->Node69--1.658-->Node89                                                                             
#> [189] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.578-->Node17--1.619-->Node33--1.640-->Node52--1.653-->Node70--1.658-->Node90                                                                                             
#> [190] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.636-->Node47--1.638-->Node50--1.651-->Node62--1.658-->Node91                                                                                 
#> [191] --0.452-->Node2--1.436-->Node9--1.546-->Node16--1.617-->Node31--1.658-->Node88--1.658-->Node92                                                                                                                                             
#> [192] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.617-->Node32--1.631-->Node39--1.659-->Node93                                             
#> [193] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.602-->Node26--1.617-->Node32--1.631-->Node39--1.659-->Node94                                             
#> [194] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.647-->Node58--1.659-->Node95                                                                                                                                                             
#> [195] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.519-->Node13--1.585-->Node20--1.595-->Node25--1.643-->Node55--1.652-->Node64--1.659-->Node96                                                                                             
#> [196] --0.452-->Node2--0.514-->Node3--0.907-->Node4--1.052-->Node5--1.328-->Node7--1.347-->Node8--1.653-->Node72--1.656-->Node82--1.660-->Node97                                                                                                 
#> [197] --0.452-->Node2--0.514-->Node3--1.465-->Node10--1.507-->Node12--1.540-->Node15--1.578-->Node18--1.592-->Node23--1.593-->Node24--1.612-->Node28--1.643-->Node56--1.660-->Node98                                                             
#> [198] --0.452-->Node2--1.436-->Node9--1.623-->Node35--1.626-->Node36--1.629-->Node38--1.636-->Node46--1.660-->Node99