Calculate the sum of features for each flow

Cumulative sum the phylogenetic flow features for each flow in a pfc, returning a pfc.

Usage

pf_flow_cumsum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A pfc object with cumulative sums of features in place of the original features

Examples

pf_flow_cumsum(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] --0.867-->Node5--1.078-->Node11--1.149-->Node28--1.184-->Node58--1.197-->t89                                                                                                                                            
#>   [2] --0.867-->Node5--1.078-->Node11--1.149-->Node28--1.184-->Node58--1.193-->Node79--1.197-->t2                                                                                                                             
#>   [3] --0.867-->Node5--1.078-->Node11--1.149-->Node28--1.184-->Node58--1.193-->Node79--1.196-->Node95--1.197-->t27                                                                                                            
#>   [4] --0.867-->Node5--1.078-->Node11--1.149-->Node28--1.184-->Node58--1.193-->Node79--1.196-->Node95--1.197-->t12                                                                                                            
#>   [5] --0.867-->Node5--1.078-->Node11--1.149-->Node28--1.184-->Node57--1.185-->Node59--1.197-->t63                                                                                                                            
#>   [6] --0.867-->Node5--1.078-->Node11--1.149-->Node28--1.184-->Node57--1.185-->Node59--1.197-->t79                                                                                                                            
#>   [7] --0.867-->Node5--1.078-->Node11--1.149-->Node28--1.184-->Node57--1.197-->t92                                                                                                                                            
#>   [8] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.175-->Node44--1.181-->Node52--1.197-->t15                                                                                                                            
#>   [9] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.175-->Node44--1.181-->Node52--1.181-->Node53--1.183-->Node54--1.197-->t87                                                                                            
#>  [10] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.175-->Node44--1.181-->Node52--1.181-->Node53--1.183-->Node54--1.184-->Node55--1.197-->t20                                                                            
#>  [11] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.175-->Node44--1.181-->Node52--1.181-->Node53--1.183-->Node54--1.184-->Node55--1.197-->t78                                                                            
#>  [12] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.175-->Node44--1.181-->Node52--1.181-->Node53--1.189-->Node71--1.197-->t96                                                                                            
#>  [13] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.175-->Node44--1.181-->Node52--1.181-->Node53--1.189-->Node71--1.196-->Node96--1.197-->t29                                                                            
#>  [14] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.175-->Node44--1.181-->Node52--1.181-->Node53--1.189-->Node71--1.196-->Node96--1.197-->t13                                                                            
#>  [15] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.175-->Node44--1.197-->t67                                                                                                                                            
#>  [16] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.161-->Node34--1.186-->Node61--1.195-->Node93--1.197-->t28                                                                                                            
#>  [17] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.161-->Node34--1.186-->Node61--1.195-->Node93--1.197-->t74                                                                                                            
#>  [18] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.161-->Node34--1.186-->Node61--1.186-->Node63--1.187-->Node65--1.189-->Node69--1.197-->t11                                                                            
#>  [19] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.161-->Node34--1.186-->Node61--1.186-->Node63--1.187-->Node65--1.189-->Node69--1.197-->t9                                                                             
#>  [20] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.161-->Node34--1.186-->Node61--1.186-->Node63--1.187-->Node65--1.197-->t21                                                                                            
#>  [21] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.161-->Node34--1.186-->Node61--1.186-->Node63--1.197-->t30                                                                                                            
#>  [22] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.161-->Node34--1.192-->Node76--1.197-->t17                                                                                                                            
#>  [23] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.161-->Node34--1.192-->Node76--1.197-->t61                                                                                                                            
#>  [24] --0.867-->Node5--1.195-->Node88--1.197-->t44                                                                                                                                                                            
#>  [25] --0.867-->Node5--1.195-->Node88--1.197-->t58                                                                                                                                                                            
#>  [26] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.084-->Node12--1.171-->Node42--1.194-->Node86--1.197-->t83                                                                                              
#>  [27] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.084-->Node12--1.171-->Node42--1.194-->Node86--1.197-->t18                                                                                              
#>  [28] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.084-->Node12--1.171-->Node42--1.197-->t85                                                                                                              
#>  [29] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.084-->Node12--1.149-->Node29--1.194-->Node85--1.197-->t25                                                                                              
#>  [30] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.084-->Node12--1.149-->Node29--1.194-->Node85--1.197-->t90                                                                                              
#>  [31] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.084-->Node12--1.149-->Node29--1.197-->t68                                                                                                              
#>  [32] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.168-->Node39--1.197-->t57                                                                                                                              
#>  [33] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.168-->Node39--1.171-->Node41--1.186-->Node64--1.197-->t76                                                                                              
#>  [34] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.168-->Node39--1.171-->Node41--1.186-->Node64--1.197-->t5                                                                                               
#>  [35] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.168-->Node39--1.171-->Node41--1.191-->Node74--1.197-->t4                                                                                               
#>  [36] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.168-->Node39--1.171-->Node41--1.191-->Node74--1.197-->t33                                                                                              
#>  [37] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.141-->Node27--1.196-->Node97--1.197-->t24                                                                                              
#>  [38] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.141-->Node27--1.196-->Node97--1.197-->t53                                                                                              
#>  [39] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.141-->Node27--1.180-->Node51--1.197-->t88                                                                                              
#>  [40] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.141-->Node27--1.180-->Node51--1.197-->t19                                                                                              
#>  [41] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.155-->Node32--1.195-->Node87--1.195-->Node94--1.197-->Node98--1.197-->t1                                                               
#>  [42] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.155-->Node32--1.195-->Node87--1.195-->Node94--1.197-->Node98--1.197-->t60                                                              
#>  [43] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.155-->Node32--1.195-->Node87--1.195-->Node94--1.197-->t84                                                                              
#>  [44] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.155-->Node32--1.195-->Node87--1.197-->t99                                                                                              
#>  [45] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.155-->Node32--1.165-->Node36--1.197-->t51                                                                                              
#>  [46] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.155-->Node32--1.165-->Node36--1.197-->t81                                                                                              
#>  [47] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.175-->Node45--1.186-->Node62--1.197-->t69                                                                                                              
#>  [48] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.175-->Node45--1.186-->Node62--1.193-->Node78--1.197-->t59                                                                                              
#>  [49] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.175-->Node45--1.186-->Node62--1.193-->Node78--1.197-->t55                                                                                              
#>  [50] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.175-->Node45--1.176-->Node46--1.197-->t49                                                                                                              
#>  [51] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.175-->Node45--1.176-->Node46--1.197-->t98                                                                                                              
#>  [52] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37--1.197-->t97                                                                                                
#>  [53] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37--1.167-->Node38--1.185-->Node60--1.187-->Node66--1.197-->t45                                                
#>  [54] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37--1.167-->Node38--1.185-->Node60--1.187-->Node66--1.192-->Node75--1.193-->Node80--1.195-->Node92--1.197-->t70
#>  [55] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37--1.167-->Node38--1.185-->Node60--1.187-->Node66--1.192-->Node75--1.193-->Node80--1.195-->Node92--1.197-->t75
#>  [56] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37--1.167-->Node38--1.185-->Node60--1.187-->Node66--1.192-->Node75--1.193-->Node80--1.197-->t6                 
#>  [57] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37--1.167-->Node38--1.185-->Node60--1.187-->Node66--1.192-->Node75--1.197-->t46                                
#>  [58] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37--1.167-->Node38--1.185-->Node60--1.197-->t62                                                                
#>  [59] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37--1.167-->Node38--1.177-->Node49--1.197-->t95                                                                
#>  [60] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37--1.167-->Node38--1.177-->Node49--1.197-->t35                                                                
#>  [61] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.197-->t52                                                                                                                
#>  [62] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.193-->Node82--1.197-->t40                                                                                                                
#>  [63] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.193-->Node82--1.197-->t32                                                                                                                
#>  [64] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.140-->Node26--1.153-->Node30--1.197-->t65                                                                                                               
#>  [65] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.140-->Node26--1.153-->Node30--1.162-->Node35--1.197-->t54                                                                                               
#>  [66] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.140-->Node26--1.153-->Node30--1.162-->Node35--1.197-->t39                                                                                               
#>  [67] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.140-->Node26--1.197-->t47                                                                                                                               
#>  [68] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.114-->Node18--1.135-->Node25--1.177-->Node48--1.197-->t36                                                                                               
#>  [69] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.114-->Node18--1.135-->Node25--1.177-->Node48--1.197-->t14                                                                                               
#>  [70] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.114-->Node18--1.135-->Node25--1.189-->Node70--1.189-->Node72--1.197-->t94                                                                               
#>  [71] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.114-->Node18--1.135-->Node25--1.189-->Node70--1.189-->Node72--1.197-->t77                                                                               
#>  [72] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.114-->Node18--1.135-->Node25--1.189-->Node70--1.197-->t80                                                                                               
#>  [73] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.114-->Node18--1.195-->Node91--1.197-->t22                                                                                                               
#>  [74] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.114-->Node18--1.195-->Node91--1.197-->t3                                                                                                                
#>  [75] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.124-->Node21--1.169-->Node40--1.192-->Node77--1.197-->t31                                                               
#>  [76] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.124-->Node21--1.169-->Node40--1.192-->Node77--1.197-->t72                                                               
#>  [77] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.124-->Node21--1.169-->Node40--1.176-->Node47--1.197-->t48                                                               
#>  [78] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.124-->Node21--1.169-->Node40--1.176-->Node47--1.197-->t91                                                               
#>  [79] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.124-->Node21--1.197-->t71                                                                                               
#>  [80] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.129-->Node22--1.174-->Node43--1.193-->Node81--1.197-->t82                                                               
#>  [81] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.129-->Node22--1.174-->Node43--1.193-->Node81--1.194-->Node84--1.195-->Node89--1.197-->t86                               
#>  [82] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.129-->Node22--1.174-->Node43--1.193-->Node81--1.194-->Node84--1.195-->Node89--1.197-->t38                               
#>  [83] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.129-->Node22--1.174-->Node43--1.193-->Node81--1.194-->Node84--1.197-->t26                                               
#>  [84] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.129-->Node22--1.174-->Node43--1.197-->t93                                                                               
#>  [85] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.129-->Node22--1.197-->t23                                                                                               
#>  [86] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.129-->Node23--1.179-->Node50--1.193-->Node83--1.197-->t34                                                                               
#>  [87] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.129-->Node23--1.179-->Node50--1.193-->Node83--1.197-->t16                                                                               
#>  [88] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.129-->Node23--1.179-->Node50--1.197-->t8                                                                                                
#>  [89] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.129-->Node23--1.197-->t42                                                                                                               
#>  [90] --0.517-->Node2--1.095-->Node14--1.098-->Node15--1.184-->Node56--1.197-->t66                                                                                                                                            
#>  [91] --0.517-->Node2--1.095-->Node14--1.098-->Node15--1.184-->Node56--1.197-->t41                                                                                                                                            
#>  [92] --0.517-->Node2--1.095-->Node14--1.098-->Node15--1.154-->Node31--1.195-->Node90--1.197-->t10                                                                                                                            
#>  [93] --0.517-->Node2--1.095-->Node14--1.098-->Node15--1.154-->Node31--1.195-->Node90--1.197-->Node99--1.197-->t43                                                                                                            
#>  [94] --0.517-->Node2--1.095-->Node14--1.098-->Node15--1.154-->Node31--1.195-->Node90--1.197-->Node99--1.197-->t73                                                                                                            
#>  [95] --0.517-->Node2--1.095-->Node14--1.098-->Node15--1.154-->Node31--1.197-->t50                                                                                                                                            
#>  [96] --0.517-->Node2--1.095-->Node14--1.119-->Node20--1.188-->Node68--1.197-->t64                                                                                                                                            
#>  [97] --0.517-->Node2--1.095-->Node14--1.119-->Node20--1.188-->Node68--1.197-->t56                                                                                                                                            
#>  [98] --0.517-->Node2--1.095-->Node14--1.119-->Node20--1.187-->Node67--1.197-->t100                                                                                                                                           
#>  [99] --0.517-->Node2--1.095-->Node14--1.119-->Node20--1.187-->Node67--1.191-->Node73--1.197-->t7                                                                                                                             
#> [100] --0.517-->Node2--1.095-->Node14--1.119-->Node20--1.187-->Node67--1.191-->Node73--1.197-->t37                                                                                                                            
#> [101] --0.517-->Node2                                                                                                                                                                                                         
#> [102] --0.517-->Node2--0.631-->Node3                                                                                                                                                                                          
#> [103] --0.517-->Node2--0.631-->Node3--0.841-->Node4                                                                                                                                                                           
#> [104] --0.867-->Node5                                                                                                                                                                                                         
#> [105] --0.517-->Node2--0.631-->Node3--0.947-->Node6                                                                                                                                                                           
#> [106] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7                                                                                                                                                            
#> [107] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8                                                                                                                                                            
#> [108] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9                                                                                                                                             
#> [109] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10                                                                                                                                                           
#> [110] --0.867-->Node5--1.078-->Node11                                                                                                                                                                                         
#> [111] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.084-->Node12                                                                                                                                           
#> [112] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13                                                                                                                                            
#> [113] --0.517-->Node2--1.095-->Node14                                                                                                                                                                                         
#> [114] --0.517-->Node2--1.095-->Node14--1.098-->Node15                                                                                                                                                                         
#> [115] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16                                                                                                                             
#> [116] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17                                                                                                                                                           
#> [117] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.114-->Node18                                                                                                                                            
#> [118] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19                                                                                                                            
#> [119] --0.517-->Node2--1.095-->Node14--1.119-->Node20                                                                                                                                                                         
#> [120] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.124-->Node21                                                                                                            
#> [121] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.129-->Node22                                                                                                            
#> [122] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.129-->Node23                                                                                                                            
#> [123] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24                                                                                                                                           
#> [124] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.114-->Node18--1.135-->Node25                                                                                                                            
#> [125] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.140-->Node26                                                                                                                                            
#> [126] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.141-->Node27                                                                                                                           
#> [127] --0.867-->Node5--1.078-->Node11--1.149-->Node28                                                                                                                                                                         
#> [128] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.084-->Node12--1.149-->Node29                                                                                                                           
#> [129] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.140-->Node26--1.153-->Node30                                                                                                                            
#> [130] --0.517-->Node2--1.095-->Node14--1.098-->Node15--1.154-->Node31                                                                                                                                                         
#> [131] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.155-->Node32                                                                                                                           
#> [132] --0.867-->Node5--1.078-->Node11--1.161-->Node33                                                                                                                                                                         
#> [133] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.161-->Node34                                                                                                                                                         
#> [134] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.140-->Node26--1.153-->Node30--1.162-->Node35                                                                                                            
#> [135] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.155-->Node32--1.165-->Node36                                                                                                           
#> [136] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37                                                                                                             
#> [137] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37--1.167-->Node38                                                                                             
#> [138] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.168-->Node39                                                                                                                                           
#> [139] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.124-->Node21--1.169-->Node40                                                                                            
#> [140] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.168-->Node39--1.171-->Node41                                                                                                                           
#> [141] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.084-->Node12--1.171-->Node42                                                                                                                           
#> [142] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.129-->Node22--1.174-->Node43                                                                                            
#> [143] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.175-->Node44                                                                                                                                                         
#> [144] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.175-->Node45                                                                                                                                           
#> [145] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.175-->Node45--1.176-->Node46                                                                                                                           
#> [146] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.124-->Node21--1.169-->Node40--1.176-->Node47                                                                            
#> [147] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.114-->Node18--1.135-->Node25--1.177-->Node48                                                                                                            
#> [148] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37--1.167-->Node38--1.177-->Node49                                                                             
#> [149] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.129-->Node23--1.179-->Node50                                                                                                            
#> [150] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.141-->Node27--1.180-->Node51                                                                                                           
#> [151] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.175-->Node44--1.181-->Node52                                                                                                                                         
#> [152] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.175-->Node44--1.181-->Node52--1.181-->Node53                                                                                                                         
#> [153] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.175-->Node44--1.181-->Node52--1.181-->Node53--1.183-->Node54                                                                                                         
#> [154] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.175-->Node44--1.181-->Node52--1.181-->Node53--1.183-->Node54--1.184-->Node55                                                                                         
#> [155] --0.517-->Node2--1.095-->Node14--1.098-->Node15--1.184-->Node56                                                                                                                                                         
#> [156] --0.867-->Node5--1.078-->Node11--1.149-->Node28--1.184-->Node57                                                                                                                                                         
#> [157] --0.867-->Node5--1.078-->Node11--1.149-->Node28--1.184-->Node58                                                                                                                                                         
#> [158] --0.867-->Node5--1.078-->Node11--1.149-->Node28--1.184-->Node57--1.185-->Node59                                                                                                                                         
#> [159] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37--1.167-->Node38--1.185-->Node60                                                                             
#> [160] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.161-->Node34--1.186-->Node61                                                                                                                                         
#> [161] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.175-->Node45--1.186-->Node62                                                                                                                           
#> [162] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.161-->Node34--1.186-->Node61--1.186-->Node63                                                                                                                         
#> [163] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.168-->Node39--1.171-->Node41--1.186-->Node64                                                                                                           
#> [164] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.161-->Node34--1.186-->Node61--1.186-->Node63--1.187-->Node65                                                                                                         
#> [165] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37--1.167-->Node38--1.185-->Node60--1.187-->Node66                                                             
#> [166] --0.517-->Node2--1.095-->Node14--1.119-->Node20--1.187-->Node67                                                                                                                                                         
#> [167] --0.517-->Node2--1.095-->Node14--1.119-->Node20--1.188-->Node68                                                                                                                                                         
#> [168] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.161-->Node34--1.186-->Node61--1.186-->Node63--1.187-->Node65--1.189-->Node69                                                                                         
#> [169] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.114-->Node18--1.135-->Node25--1.189-->Node70                                                                                                            
#> [170] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.175-->Node44--1.181-->Node52--1.181-->Node53--1.189-->Node71                                                                                                         
#> [171] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.114-->Node18--1.135-->Node25--1.189-->Node70--1.189-->Node72                                                                                            
#> [172] --0.517-->Node2--1.095-->Node14--1.119-->Node20--1.187-->Node67--1.191-->Node73                                                                                                                                         
#> [173] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.168-->Node39--1.171-->Node41--1.191-->Node74                                                                                                           
#> [174] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37--1.167-->Node38--1.185-->Node60--1.187-->Node66--1.192-->Node75                                             
#> [175] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.161-->Node34--1.192-->Node76                                                                                                                                         
#> [176] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.124-->Node21--1.169-->Node40--1.192-->Node77                                                                            
#> [177] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.175-->Node45--1.186-->Node62--1.193-->Node78                                                                                                           
#> [178] --0.867-->Node5--1.078-->Node11--1.149-->Node28--1.184-->Node58--1.193-->Node79                                                                                                                                         
#> [179] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37--1.167-->Node38--1.185-->Node60--1.187-->Node66--1.192-->Node75--1.193-->Node80                             
#> [180] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.129-->Node22--1.174-->Node43--1.193-->Node81                                                                            
#> [181] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.193-->Node82                                                                                                                             
#> [182] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.129-->Node23--1.179-->Node50--1.193-->Node83                                                                                            
#> [183] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.129-->Node22--1.174-->Node43--1.193-->Node81--1.194-->Node84                                                            
#> [184] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.084-->Node12--1.149-->Node29--1.194-->Node85                                                                                                           
#> [185] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.060-->Node10--1.084-->Node12--1.171-->Node42--1.194-->Node86                                                                                                           
#> [186] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.155-->Node32--1.195-->Node87                                                                                                           
#> [187] --0.867-->Node5--1.195-->Node88                                                                                                                                                                                         
#> [188] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.088-->Node13--1.117-->Node19--1.129-->Node22--1.174-->Node43--1.193-->Node81--1.194-->Node84--1.195-->Node89                                            
#> [189] --0.517-->Node2--1.095-->Node14--1.098-->Node15--1.154-->Node31--1.195-->Node90                                                                                                                                         
#> [190] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.013-->Node7--1.114-->Node18--1.195-->Node91                                                                                                                            
#> [191] --0.517-->Node2--0.631-->Node3--0.947-->Node6--1.051-->Node8--1.057-->Node9--1.099-->Node16--1.166-->Node37--1.167-->Node38--1.185-->Node60--1.187-->Node66--1.192-->Node75--1.193-->Node80--1.195-->Node92             
#> [192] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.161-->Node34--1.186-->Node61--1.195-->Node93                                                                                                                         
#> [193] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.155-->Node32--1.195-->Node87--1.195-->Node94                                                                                           
#> [194] --0.867-->Node5--1.078-->Node11--1.149-->Node28--1.184-->Node58--1.193-->Node79--1.196-->Node95                                                                                                                         
#> [195] --0.867-->Node5--1.078-->Node11--1.161-->Node33--1.175-->Node44--1.181-->Node52--1.181-->Node53--1.189-->Node71--1.196-->Node96                                                                                         
#> [196] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.141-->Node27--1.196-->Node97                                                                                                           
#> [197] --0.517-->Node2--0.631-->Node3--0.841-->Node4--1.107-->Node17--1.130-->Node24--1.155-->Node32--1.195-->Node87--1.195-->Node94--1.197-->Node98                                                                           
#> [198] --0.517-->Node2--1.095-->Node14--1.098-->Node15--1.154-->Node31--1.195-->Node90--1.197-->Node99