Calculate the sum of features for each flow

Cumulative sum the phylogenetic flow features for each flow in a pfc, returning a pfc.

Usage

pf_flow_cumsum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A pfc object with cumulative sums of features in place of the original features

Examples

pf_flow_cumsum(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] ---1.83-->Node6---2.00-->Node10---2.12-->t43                                                                                                                                                                            
#>   [2] ---1.83-->Node6---2.00-->Node10---2.12-->t70                                                                                                                                                                            
#>   [3] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.12-->t39                                                                                                                                             
#>   [4] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31---2.09-->Node45---2.10-->Node54---2.12-->Node97---2.12-->t97                                                             
#>   [5] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31---2.09-->Node45---2.10-->Node54---2.12-->Node97---2.12-->t2                                                              
#>   [6] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31---2.09-->Node45---2.10-->Node54---2.12-->Node99---2.12-->t10                                                             
#>   [7] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31---2.09-->Node45---2.10-->Node54---2.12-->Node99---2.12-->t42                                                             
#>   [8] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31---2.09-->Node45---2.12-->t26                                                                                             
#>   [9] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31---2.10-->Node57---2.12-->t93                                                                                             
#>  [10] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31---2.10-->Node57---2.10-->Node59---2.12-->t88                                                                             
#>  [11] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31---2.10-->Node57---2.10-->Node59---2.11-->Node75---2.12-->t21                                                             
#>  [12] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31---2.10-->Node57---2.10-->Node59---2.11-->Node75---2.12-->t80                                                             
#>  [13] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.12-->Node93---2.12-->t76                                                                                                             
#>  [14] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.12-->Node93---2.12-->t82                                                                                                             
#>  [15] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25---2.10-->Node56---2.11-->Node70---2.12-->t16                                                                                                             
#>  [16] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25---2.10-->Node56---2.11-->Node70---2.12-->t62                                                                                                             
#>  [17] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25---2.10-->Node56---2.12-->t58                                                                                                                             
#>  [18] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25---2.10-->Node63---2.11-->Node79---2.12-->Node94---2.12-->t48                                                                                             
#>  [19] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25---2.10-->Node63---2.11-->Node79---2.12-->Node94---2.12-->t71                                                                                             
#>  [20] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25---2.10-->Node63---2.11-->Node79---2.12-->Node90---2.12-->Node98---2.12-->t12                                                                             
#>  [21] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25---2.10-->Node63---2.11-->Node79---2.12-->Node90---2.12-->Node98---2.12-->t49                                                                             
#>  [22] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25---2.10-->Node63---2.11-->Node79---2.12-->Node90---2.12-->t6                                                                                              
#>  [23] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25---2.10-->Node63---2.12-->t51                                                                                                                             
#>  [24] ---1.83-->Node6---1.86-->Node8---2.05-->Node21---2.10-->Node60---2.11-->Node84---2.12-->t17                                                                                                                             
#>  [25] ---1.83-->Node6---1.86-->Node8---2.05-->Node21---2.10-->Node60---2.11-->Node84---2.12-->Node96---2.12-->t65                                                                                                             
#>  [26] ---1.83-->Node6---1.86-->Node8---2.05-->Node21---2.10-->Node60---2.11-->Node84---2.12-->Node96---2.12-->t59                                                                                                             
#>  [27] ---1.83-->Node6---1.86-->Node8---2.05-->Node21---2.10-->Node60---2.12-->t45                                                                                                                                             
#>  [28] ---1.83-->Node6---1.86-->Node8---2.05-->Node21---2.08-->Node36---2.11-->Node66---2.12-->t11                                                                                                                             
#>  [29] ---1.83-->Node6---1.86-->Node8---2.05-->Node21---2.08-->Node36---2.11-->Node66---2.11-->Node67---2.12-->t20                                                                                                             
#>  [30] ---1.83-->Node6---1.86-->Node8---2.05-->Node21---2.08-->Node36---2.11-->Node66---2.11-->Node67---2.12-->t4                                                                                                              
#>  [31] ---1.83-->Node6---1.86-->Node8---2.05-->Node21---2.08-->Node36---2.12-->t73                                                                                                                                             
#>  [32] --0.988-->Node2--2.087-->Node40--2.095-->Node48--2.119-->t79                                                                                                                                                            
#>  [33] --0.988-->Node2--2.087-->Node40--2.095-->Node48--2.119-->t55                                                                                                                                                            
#>  [34] --0.988-->Node2--2.087-->Node40--2.114-->Node85--2.119-->t44                                                                                                                                                            
#>  [35] --0.988-->Node2--2.087-->Node40--2.114-->Node85--2.119-->t46                                                                                                                                                            
#>  [36] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.088-->Node42--2.119-->t5                                                                                                                               
#>  [37] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.088-->Node42--2.112-->Node81--2.119-->t52                                                                                                              
#>  [38] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.088-->Node42--2.112-->Node81--2.119-->t56                                                                                                              
#>  [39] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.080-->Node32--2.106-->Node72--2.119-->t99                                                                                                              
#>  [40] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.080-->Node32--2.106-->Node72--2.119-->t69                                                                                                              
#>  [41] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.080-->Node32--2.083-->Node38--2.088-->Node41--2.105-->Node65--2.119-->t34                                                                              
#>  [42] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.080-->Node32--2.083-->Node38--2.088-->Node41--2.105-->Node65--2.119-->t47                                                                              
#>  [43] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.080-->Node32--2.083-->Node38--2.088-->Node41--2.119-->t64                                                                                              
#>  [44] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.080-->Node32--2.083-->Node38--2.116-->Node91--2.119-->t25                                                                                              
#>  [45] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.080-->Node32--2.083-->Node38--2.116-->Node91--2.119-->t40                                                                                              
#>  [46] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.024-->Node14--2.089-->Node43--2.119-->t14                                                                                                                              
#>  [47] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.024-->Node14--2.089-->Node43--2.106-->Node71--2.119-->t98                                                                                                              
#>  [48] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.024-->Node14--2.089-->Node43--2.106-->Node71--2.109-->Node76--2.119-->t61                                                                                              
#>  [49] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.024-->Node14--2.089-->Node43--2.106-->Node71--2.109-->Node76--2.119-->t37                                                                                              
#>  [50] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.024-->Node14--2.107-->Node73--2.119-->t28                                                                                                                              
#>  [51] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.024-->Node14--2.107-->Node73--2.113-->Node82--2.119-->t53                                                                                                              
#>  [52] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.024-->Node14--2.107-->Node73--2.113-->Node82--2.119-->t74                                                                                                              
#>  [53] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.075-->Node29--2.114-->Node86--2.119-->t75                                                                                               
#>  [54] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.075-->Node29--2.114-->Node86--2.119-->t57                                                                                               
#>  [55] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.075-->Node29--2.107-->Node74--2.119-->t8                                                                                                
#>  [56] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.075-->Node29--2.107-->Node74--2.119-->t18                                                                                               
#>  [57] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.066-->Node26--2.080-->Node33--2.097-->Node50--2.119-->t63                                                                               
#>  [58] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.066-->Node26--2.080-->Node33--2.097-->Node50--2.105-->Node69--2.119-->t92                                                               
#>  [59] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.066-->Node26--2.080-->Node33--2.097-->Node50--2.105-->Node69--2.119-->t66                                                               
#>  [60] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.066-->Node26--2.080-->Node33--2.119-->t38                                                                                               
#>  [61] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.066-->Node26--2.074-->Node28--2.114-->Node83--2.119-->t33                                                                               
#>  [62] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.066-->Node26--2.074-->Node28--2.114-->Node83--2.119-->t81                                                                               
#>  [63] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.066-->Node26--2.074-->Node28--2.119-->t3                                                                                                
#>  [64] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.061-->Node24--2.119-->t29                                                                                                
#>  [65] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.061-->Node24--2.086-->Node39--2.102-->Node58--2.119-->t85                                                                
#>  [66] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.061-->Node24--2.086-->Node39--2.102-->Node58--2.119-->t23                                                                
#>  [67] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.061-->Node24--2.086-->Node39--2.092-->Node46--2.119-->t32                                                                
#>  [68] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.061-->Node24--2.086-->Node39--2.092-->Node46--2.103-->Node62--2.109-->Node78--2.119-->t86                                
#>  [69] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.061-->Node24--2.086-->Node39--2.092-->Node46--2.103-->Node62--2.109-->Node78--2.119-->t24                                
#>  [70] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.061-->Node24--2.086-->Node39--2.092-->Node46--2.103-->Node62--2.119-->t72                                                
#>  [71] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19--2.119-->t7                                                                 
#>  [72] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19--2.095-->Node47--2.096-->Node49--2.101-->Node53--2.119-->t87                
#>  [73] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19--2.095-->Node47--2.096-->Node49--2.101-->Node53--2.119-->t67                
#>  [74] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19--2.095-->Node47--2.096-->Node49--2.115-->Node88--2.119-->t54                
#>  [75] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19--2.095-->Node47--2.096-->Node49--2.115-->Node88--2.118-->Node95--2.119-->t90
#>  [76] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19--2.095-->Node47--2.096-->Node49--2.115-->Node88--2.118-->Node95--2.119-->t41
#>  [77] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19--2.095-->Node47--2.103-->Node61--2.119-->t94                                
#>  [78] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19--2.095-->Node47--2.103-->Node61--2.111-->Node80--2.119-->t19                
#>  [79] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19--2.095-->Node47--2.103-->Node61--2.111-->Node80--2.119-->t35                
#>  [80] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.119-->t78                                                                                
#>  [81] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.081-->Node34--2.105-->Node68--2.119-->t22                                                
#>  [82] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.081-->Node34--2.105-->Node68--2.119-->t9                                                 
#>  [83] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.081-->Node34--2.119-->t36                                                                
#>  [84] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44--2.101-->Node52--2.102-->Node55--2.119-->t96                                
#>  [85] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44--2.101-->Node52--2.102-->Node55--2.114-->Node87--2.119-->t60                
#>  [86] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44--2.101-->Node52--2.102-->Node55--2.114-->Node87--2.119-->t68                
#>  [87] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44--2.101-->Node52--2.109-->Node77--2.119-->t27                                
#>  [88] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44--2.101-->Node52--2.109-->Node77--2.119-->t83                                
#>  [89] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44--2.105-->Node64--2.119-->t95                                                
#>  [90] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44--2.105-->Node64--2.115-->Node89--2.119-->t89                                
#>  [91] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44--2.105-->Node64--2.115-->Node89--2.116-->Node92--2.119-->t50                
#>  [92] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44--2.105-->Node64--2.115-->Node89--2.116-->Node92--2.119-->t84                
#>  [93] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.028-->Node15--2.053-->Node20--2.083-->Node35--2.119-->t13                                                                                
#>  [94] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.028-->Node15--2.053-->Node20--2.083-->Node35--2.119-->t100                                                                               
#>  [95] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.028-->Node15--2.053-->Node20--2.119-->t1                                                                                                 
#>  [96] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.028-->Node15--2.083-->Node37--2.119-->t15                                                                                                
#>  [97] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.028-->Node15--2.083-->Node37--2.119-->t77                                                                                                
#>  [98] --0.988-->Node2--1.199-->Node3--1.479-->Node4--2.037-->Node16--2.099-->Node51--2.119-->t91                                                                                                                              
#>  [99] --0.988-->Node2--1.199-->Node3--1.479-->Node4--2.037-->Node16--2.099-->Node51--2.119-->t31                                                                                                                              
#> [100] --0.988-->Node2--1.199-->Node3--1.479-->Node4--2.037-->Node16--2.119-->t30                                                                                                                                              
#> [101] --0.988-->Node2                                                                                                                                                                                                         
#> [102] --0.988-->Node2--1.199-->Node3                                                                                                                                                                                          
#> [103] --0.988-->Node2--1.199-->Node3--1.479-->Node4                                                                                                                                                                           
#> [104] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5                                                                                                                                                            
#> [105] ---1.83-->Node6                                                                                                                                                                                                         
#> [106] --0.988-->Node2--1.199-->Node3--1.837-->Node7                                                                                                                                                                           
#> [107] ---1.83-->Node6---1.86-->Node8                                                                                                                                                                                          
#> [108] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9                                                                                                                                             
#> [109] ---1.83-->Node6---2.00-->Node10                                                                                                                                                                                         
#> [110] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11                                                                                                                             
#> [111] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12                                                                                                             
#> [112] ---1.83-->Node6---1.86-->Node8---2.02-->Node13                                                                                                                                                                          
#> [113] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.024-->Node14                                                                                                                                                           
#> [114] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.028-->Node15                                                                                                                             
#> [115] --0.988-->Node2--1.199-->Node3--1.479-->Node4--2.037-->Node16                                                                                                                                                           
#> [116] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17                                                                                             
#> [117] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18                                                                                             
#> [118] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19                                                                             
#> [119] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.028-->Node15--2.053-->Node20                                                                                                             
#> [120] ---1.83-->Node6---1.86-->Node8---2.05-->Node21                                                                                                                                                                          
#> [121] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22                                                                                                                                                          
#> [122] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23                                                                                                                                            
#> [123] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.061-->Node24                                                                                                             
#> [124] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25                                                                                                                                                          
#> [125] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.066-->Node26                                                                                                                            
#> [126] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27                                                                                                                                                           
#> [127] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.066-->Node26--2.074-->Node28                                                                                                            
#> [128] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.075-->Node29                                                                                                                            
#> [129] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30                                                                                                                                          
#> [130] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31                                                                                                                          
#> [131] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.080-->Node32                                                                                                                                           
#> [132] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.066-->Node26--2.080-->Node33                                                                                                            
#> [133] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.081-->Node34                                                                             
#> [134] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.028-->Node15--2.053-->Node20--2.083-->Node35                                                                                             
#> [135] ---1.83-->Node6---1.86-->Node8---2.05-->Node21---2.08-->Node36                                                                                                                                                          
#> [136] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.028-->Node15--2.083-->Node37                                                                                                             
#> [137] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.080-->Node32--2.083-->Node38                                                                                                                           
#> [138] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.061-->Node24--2.086-->Node39                                                                                             
#> [139] --0.988-->Node2--2.087-->Node40                                                                                                                                                                                         
#> [140] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.080-->Node32--2.083-->Node38--2.088-->Node41                                                                                                           
#> [141] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.088-->Node42                                                                                                                                           
#> [142] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.024-->Node14--2.089-->Node43                                                                                                                                           
#> [143] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44                                                                             
#> [144] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31---2.09-->Node45                                                                                                          
#> [145] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.061-->Node24--2.086-->Node39--2.092-->Node46                                                                             
#> [146] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19--2.095-->Node47                                                             
#> [147] --0.988-->Node2--2.087-->Node40--2.095-->Node48                                                                                                                                                                         
#> [148] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19--2.095-->Node47--2.096-->Node49                                             
#> [149] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.066-->Node26--2.080-->Node33--2.097-->Node50                                                                                            
#> [150] --0.988-->Node2--1.199-->Node3--1.479-->Node4--2.037-->Node16--2.099-->Node51                                                                                                                                           
#> [151] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44--2.101-->Node52                                                             
#> [152] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19--2.095-->Node47--2.096-->Node49--2.101-->Node53                             
#> [153] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31---2.09-->Node45---2.10-->Node54                                                                                          
#> [154] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44--2.101-->Node52--2.102-->Node55                                             
#> [155] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25---2.10-->Node56                                                                                                                                          
#> [156] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31---2.10-->Node57                                                                                                          
#> [157] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.061-->Node24--2.086-->Node39--2.102-->Node58                                                                             
#> [158] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31---2.10-->Node57---2.10-->Node59                                                                                          
#> [159] ---1.83-->Node6---1.86-->Node8---2.05-->Node21---2.10-->Node60                                                                                                                                                          
#> [160] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19--2.095-->Node47--2.103-->Node61                                             
#> [161] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.061-->Node24--2.086-->Node39--2.092-->Node46--2.103-->Node62                                                             
#> [162] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25---2.10-->Node63                                                                                                                                          
#> [163] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44--2.105-->Node64                                                             
#> [164] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.080-->Node32--2.083-->Node38--2.088-->Node41--2.105-->Node65                                                                                           
#> [165] ---1.83-->Node6---1.86-->Node8---2.05-->Node21---2.08-->Node36---2.11-->Node66                                                                                                                                          
#> [166] ---1.83-->Node6---1.86-->Node8---2.05-->Node21---2.08-->Node36---2.11-->Node66---2.11-->Node67                                                                                                                          
#> [167] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.081-->Node34--2.105-->Node68                                                             
#> [168] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.066-->Node26--2.080-->Node33--2.097-->Node50--2.105-->Node69                                                                            
#> [169] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25---2.10-->Node56---2.11-->Node70                                                                                                                          
#> [170] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.024-->Node14--2.089-->Node43--2.106-->Node71                                                                                                                           
#> [171] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.080-->Node32--2.106-->Node72                                                                                                                           
#> [172] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.024-->Node14--2.107-->Node73                                                                                                                                           
#> [173] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.075-->Node29--2.107-->Node74                                                                                                            
#> [174] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31---2.10-->Node57---2.10-->Node59---2.11-->Node75                                                                          
#> [175] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.024-->Node14--2.089-->Node43--2.106-->Node71--2.109-->Node76                                                                                                           
#> [176] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44--2.101-->Node52--2.109-->Node77                                             
#> [177] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.061-->Node24--2.086-->Node39--2.092-->Node46--2.103-->Node62--2.109-->Node78                                             
#> [178] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25---2.10-->Node63---2.11-->Node79                                                                                                                          
#> [179] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19--2.095-->Node47--2.103-->Node61--2.111-->Node80                             
#> [180] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.088-->Node42--2.112-->Node81                                                                                                                           
#> [181] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.024-->Node14--2.107-->Node73--2.113-->Node82                                                                                                                           
#> [182] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.066-->Node26--2.074-->Node28--2.114-->Node83                                                                                            
#> [183] ---1.83-->Node6---1.86-->Node8---2.05-->Node21---2.10-->Node60---2.11-->Node84                                                                                                                                          
#> [184] --0.988-->Node2--2.087-->Node40--2.114-->Node85                                                                                                                                                                         
#> [185] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--2.057-->Node23--2.075-->Node29--2.114-->Node86                                                                                                            
#> [186] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44--2.101-->Node52--2.102-->Node55--2.114-->Node87                             
#> [187] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19--2.095-->Node47--2.096-->Node49--2.115-->Node88                             
#> [188] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44--2.105-->Node64--2.115-->Node89                                             
#> [189] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25---2.10-->Node63---2.11-->Node79---2.12-->Node90                                                                                                          
#> [190] --0.988-->Node2--1.199-->Node3--1.837-->Node7--2.073-->Node27--2.080-->Node32--2.083-->Node38--2.116-->Node91                                                                                                           
#> [191] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.048-->Node18--2.089-->Node44--2.105-->Node64--2.115-->Node89--2.116-->Node92                             
#> [192] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.12-->Node93                                                                                                                          
#> [193] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25---2.10-->Node63---2.11-->Node79---2.12-->Node94                                                                                                          
#> [194] --0.988-->Node2--1.199-->Node3--1.479-->Node4--1.539-->Node5--1.910-->Node9--2.004-->Node11--2.012-->Node12--2.042-->Node17--2.050-->Node19--2.095-->Node47--2.096-->Node49--2.115-->Node88--2.118-->Node95             
#> [195] ---1.83-->Node6---1.86-->Node8---2.05-->Node21---2.10-->Node60---2.11-->Node84---2.12-->Node96                                                                                                                          
#> [196] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31---2.09-->Node45---2.10-->Node54---2.12-->Node97                                                                          
#> [197] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node25---2.10-->Node63---2.11-->Node79---2.12-->Node90---2.12-->Node98                                                                                          
#> [198] ---1.83-->Node6---1.86-->Node8---2.02-->Node13---2.06-->Node22---2.07-->Node30---2.08-->Node31---2.09-->Node45---2.10-->Node54---2.12-->Node99