Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>        t93       t100        t38        t90        t99        t62        t28 
#> 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 
#>        t39        t77        t85        t52        t25        t51        t78 
#> 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 
#>        t60        t23        t67        t81        t27        t12        t21 
#> 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 
#>        t50        t64        t55        t74        t79        t65        t32 
#> 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 
#>        t82        t53        t18        t94        t95        t13        t86 
#> 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 
#>        t88        t87        t56        t35        t59        t63        t92 
#> 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 
#>        t84        t40         t7        t98        t48        t57        t14 
#> 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 
#>        t33         t8        t73        t58        t26        t97         t1 
#> 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 
#>        t54        t66        t96        t72         t9        t44        t20 
#> 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 
#>         t3        t36        t46        t47        t43        t16        t41 
#> 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 
#>        t11        t80         t2        t69        t34         t6        t75 
#> 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 
#>        t71        t76         t4        t45        t37        t68        t19 
#> 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 
#>        t89        t22        t49        t83        t61        t24         t5 
#> 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 
#>        t15        t29        t31        t17        t42        t10        t70 
#> 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 0.41444942 
#>        t30        t91      Node2      Node3      Node4      Node5      Node6 
#> 0.41444942 0.41444942 0.02414808 0.04174451 0.10956553 0.14653817 0.20843183 
#>      Node7      Node8      Node9     Node10     Node11     Node12     Node13 
#> 0.26088243 0.26997668 0.27107325 0.28647356 0.30191830 0.32848312 0.32977463 
#>     Node14     Node15     Node16     Node17     Node18     Node19     Node20 
#> 0.33082755 0.34689791 0.34813871 0.34834946 0.35078810 0.35161870 0.35675706 
#>     Node21     Node22     Node23     Node24     Node25     Node26     Node27 
#> 0.35808598 0.35940980 0.36073938 0.36291780 0.36573869 0.36983887 0.37066690 
#>     Node28     Node29     Node30     Node31     Node32     Node33     Node34 
#> 0.37345576 0.37587226 0.37895844 0.37980598 0.37996984 0.38061849 0.38216799 
#>     Node35     Node36     Node37     Node38     Node39     Node40     Node41 
#> 0.38235124 0.38552559 0.38574841 0.38642936 0.38881951 0.38908726 0.38927092 
#>     Node42     Node43     Node44     Node45     Node46     Node47     Node48 
#> 0.39078098 0.39156050 0.39178145 0.39241045 0.39328639 0.39375061 0.39482525 
#>     Node49     Node50     Node51     Node52     Node53     Node54     Node55 
#> 0.39541491 0.39601896 0.39646463 0.39654440 0.39724496 0.39806793 0.39816057 
#>     Node56     Node57     Node58     Node59     Node60     Node61     Node62 
#> 0.39954685 0.39964820 0.39994639 0.40109236 0.40174994 0.40195786 0.40234611 
#>     Node63     Node64     Node65     Node66     Node67     Node68     Node69 
#> 0.40236962 0.40271640 0.40324743 0.40345377 0.40396573 0.40420043 0.40425342 
#>     Node70     Node71     Node72     Node73     Node74     Node75     Node76 
#> 0.40427077 0.40462842 0.40611021 0.40646584 0.40694365 0.40735480 0.40773313 
#>     Node77     Node78     Node79     Node80     Node81     Node82     Node83 
#> 0.40778253 0.40781371 0.40800889 0.40801496 0.40815484 0.40817300 0.40825568 
#>     Node84     Node85     Node86     Node87     Node88     Node89     Node90 
#> 0.40836712 0.40855269 0.40910870 0.40933178 0.40958869 0.40969562 0.41040664 
#>     Node91     Node92     Node93     Node94     Node95     Node96     Node97 
#> 0.41116010 0.41122110 0.41272843 0.41273213 0.41344761 0.41345253 0.41345538 
#>     Node98     Node99 
#> 0.41409845 0.41438822