Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>       t42       t78       t96       t98       t48       t58       t57       t66 
#> 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 
#>       t29        t2       t32       t15       t22       t39       t20       t59 
#> 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 
#>       t46       t24       t55       t63       t18       t26       t12       t51 
#> 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 
#>       t52       t56        t5       t61       t28       t14        t3       t86 
#> 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 
#>       t13       t50       t69       t25        t8       t72        t7       t65 
#> 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 
#>       t62       t91       t77       t11       t64       t99       t49       t73 
#> 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 
#>       t45       t23       t93       t44       t84       t33       t47       t31 
#> 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 
#>       t27       t30       t34        t6       t37       t89       t67       t40 
#> 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 
#>       t74       t87       t71       t41        t1       t90       t60       t10 
#> 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 
#>       t97       t76       t83       t35       t88       t21       t36       t16 
#> 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 
#>       t94       t81       t79       t70       t43       t82       t68       t92 
#> 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 
#>       t38       t95       t85        t9       t53      t100       t54       t17 
#> 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 1.8406703 
#>       t80        t4       t75       t19     Node2     Node3     Node4     Node5 
#> 1.8406703 1.8406703 1.8406703 1.8406703 0.9022823 1.4188167 1.4829451 1.5501046 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 1.5983893 1.6137092 1.6173353 1.6347310 1.6754899 1.7397053 1.7462682 1.7473049 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 1.7486060 1.7565677 1.7577573 1.7664401 1.7713194 1.7716780 1.7717424 1.7728654 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 1.7764778 1.7805664 1.7851631 1.7871687 1.7872259 1.7887701 1.7917949 1.7987027 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 1.8024330 1.8043162 1.8055343 1.8060710 1.8069358 1.8081683 1.8092817 1.8117783 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 1.8118683 1.8128493 1.8132419 1.8137839 1.8147819 1.8165242 1.8172651 1.8179218 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 1.8198882 1.8202174 1.8207408 1.8211054 1.8215369 1.8223957 1.8227360 1.8232589 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 1.8233759 1.8247431 1.8247884 1.8256665 1.8272026 1.8275932 1.8289494 1.8295422 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 1.8295862 1.8300064 1.8310222 1.8310664 1.8312685 1.8313054 1.8314125 1.8319560 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 1.8324631 1.8327905 1.8328544 1.8328975 1.8332626 1.8336327 1.8338063 1.8339271 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 1.8345663 1.8347137 1.8349524 1.8349867 1.8352497 1.8354804 1.8358698 1.8359216 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 1.8363048 1.8363205 1.8370942 1.8376515 1.8377201 1.8382970 1.8387955 1.8389977 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 1.8392014 1.8396304 1.8399551 1.8401218 1.8404283 1.8405964