Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>      t89      t21      t26      t60      t19      t59      t49      t28 
#> 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 
#>      t94      t91      t65      t42      t90      t22       t9      t20 
#> 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 
#>      t66      t35      t85      t17      t53      t29      t45      t24 
#> 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 
#>      t13      t33      t44      t61      t98      t18      t97      t93 
#> 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 
#>       t3      t99      t68      t47      t25       t7      t37      t57 
#> 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 
#>     t100      t84      t62      t80      t51      t69      t81      t95 
#> 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 
#>      t83      t78      t73      t46      t86      t87      t92      t30 
#> 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 
#>      t54      t34      t50      t96      t12      t52      t76      t82 
#> 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 
#>      t67      t40      t38      t41      t10      t43       t4      t23 
#> 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 
#>      t39      t48      t16      t32      t58      t11      t64       t2 
#> 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 
#>      t56       t1      t70      t71      t63      t27      t72       t6 
#> 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 
#>       t8      t79      t74      t15      t55      t36       t5      t14 
#> 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 4.805109 
#>      t88      t75      t31      t77    Node2    Node3    Node4    Node5 
#> 4.805109 4.805109 4.805109 4.805109 3.867568 4.402151 4.406596 4.527735 
#>    Node6    Node7    Node8    Node9   Node10   Node11   Node12   Node13 
#> 4.576420 4.594477 4.601185 4.601375 4.645999 4.694449 4.701565 4.710570 
#>   Node14   Node15   Node16   Node17   Node18   Node19   Node20   Node21 
#> 4.714232 4.717924 4.724326 4.725798 4.728070 4.735964 4.738751 4.758051 
#>   Node22   Node23   Node24   Node25   Node26   Node27   Node28   Node29 
#> 4.758195 4.761458 4.761638 4.762052 4.762696 4.764266 4.765472 4.766761 
#>   Node30   Node31   Node32   Node33   Node34   Node35   Node36   Node37 
#> 4.767507 4.768728 4.769374 4.772347 4.773719 4.775160 4.775184 4.775255 
#>   Node38   Node39   Node40   Node41   Node42   Node43   Node44   Node45 
#> 4.775713 4.776100 4.777254 4.777866 4.778130 4.778425 4.779070 4.779789 
#>   Node46   Node47   Node48   Node49   Node50   Node51   Node52   Node53 
#> 4.780859 4.782732 4.783023 4.783641 4.783754 4.785244 4.786440 4.786897 
#>   Node54   Node55   Node56   Node57   Node58   Node59   Node60   Node61 
#> 4.787268 4.788774 4.789264 4.789968 4.790180 4.790859 4.790916 4.791260 
#>   Node62   Node63   Node64   Node65   Node66   Node67   Node68   Node69 
#> 4.791601 4.791767 4.791796 4.792066 4.793073 4.793390 4.795748 4.796418 
#>   Node70   Node71   Node72   Node73   Node74   Node75   Node76   Node77 
#> 4.796723 4.796799 4.798043 4.798375 4.798482 4.798876 4.799667 4.799762 
#>   Node78   Node79   Node80   Node81   Node82   Node83   Node84   Node85 
#> 4.799966 4.800276 4.800337 4.800724 4.801542 4.801611 4.802104 4.802128 
#>   Node86   Node87   Node88   Node89   Node90   Node91   Node92   Node93 
#> 4.802159 4.802287 4.802741 4.802775 4.802955 4.803291 4.803596 4.803999 
#>   Node94   Node95   Node96   Node97   Node98   Node99 
#> 4.804054 4.804165 4.804440 4.804739 4.804765 4.804846