Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>       t74       t25       t27        t7       t83       t60       t43       t31 
#> 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 
#>       t38       t90       t48       t69       t76       t26       t95        t8 
#> 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 
#>       t13       t67       t82       t17       t24       t11       t61       t40 
#> 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 
#>       t63       t58       t45       t66       t15       t92       t65       t98 
#> 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 
#>       t44        t9       t89       t54       t96       t94       t73       t52 
#> 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 
#>       t86       t34        t5       t72       t41       t50       t70       t71 
#> 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 
#>       t99       t37       t19       t64       t14        t6       t87       t59 
#> 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 
#>       t78      t100       t81       t77       t68       t47       t20       t84 
#> 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 
#>       t33       t88       t39       t56       t93       t91       t23       t10 
#> 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 
#>       t55       t35       t75       t79       t53        t1       t30       t49 
#> 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 
#>       t21       t18        t2       t32       t22       t42       t85        t3 
#> 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 
#>       t28       t36       t29       t12       t80       t16       t97       t57 
#> 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 1.2211920 
#>       t62       t46        t4       t51     Node2     Node3     Node4     Node5 
#> 1.2211920 1.2211920 1.2211920 1.2211920 0.4963606 0.6906593 0.8387379 0.8670138 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 0.8765231 0.9082578 0.9117826 0.9415964 1.0238248 1.0365642 1.0371970 1.0692767 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 1.0829575 1.0831482 1.0906649 1.1062730 1.1185494 1.1195065 1.1237173 1.1253424 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 1.1382909 1.1388901 1.1447119 1.1532288 1.1540430 1.1591482 1.1649270 1.1734762 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 1.1738896 1.1743351 1.1750623 1.1799034 1.1817173 1.1834363 1.1843200 1.1848761 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 1.1868023 1.1896115 1.1911062 1.1914989 1.1951397 1.1955332 1.1956451 1.1960296 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 1.1964142 1.1978610 1.1991605 1.2000789 1.2009067 1.2015318 1.2020618 1.2027188 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 1.2037092 1.2047573 1.2051645 1.2053240 1.2056296 1.2057919 1.2066326 1.2068455 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 1.2074910 1.2094595 1.2096490 1.2103776 1.2113036 1.2115543 1.2117374 1.2117777 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 1.2119141 1.2120147 1.2121539 1.2133523 1.2139373 1.2143939 1.2146763 1.2151033 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 1.2152115 1.2154661 1.2155448 1.2157593 1.2159119 1.2161482 1.2161926 1.2168264 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 1.2169444 1.2172273 1.2174042 1.2178531 1.2179199 1.2182590 1.2183319 1.2183995 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 1.2190515 1.2194346 1.2200160 1.2202848 1.2206655 1.2210218