Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>      t67      t93      t81      t92      t72      t73      t23      t83 
#> 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 
#>      t40      t84      t63      t15      t98      t97     t100      t22 
#> 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 
#>      t14      t85       t1      t27      t54      t69      t77      t75 
#> 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 
#>      t82      t60      t51      t47      t31      t99      t76      t34 
#> 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 
#>       t2      t35      t16      t70      t46      t58       t3      t37 
#> 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 
#>      t91      t74      t36      t71      t78      t62      t21      t42 
#> 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 
#>      t64      t11      t20      t18      t56      t52       t9      t96 
#> 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 
#>      t13      t66      t33      t28      t38      t43      t24      t79 
#> 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 
#>      t45      t32       t4      t55      t25       t7      t94      t48 
#> 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 
#>      t50      t59      t61      t39       t5       t6      t26      t90 
#> 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 
#>      t49      t53      t68      t89      t57      t88      t10      t41 
#> 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 
#>      t30      t86       t8      t17      t12      t87      t29      t44 
#> 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 1.576534 
#>      t95      t19      t65      t80    Node2    Node3    Node4    Node5 
#> 1.576534 1.576534 1.576534 1.576534 1.008974 1.212507 1.276281 1.277970 
#>    Node6    Node7    Node8    Node9   Node10   Node11   Node12   Node13 
#> 1.355304 1.376651 1.380992 1.387896 1.399504 1.430262 1.438981 1.452485 
#>   Node14   Node15   Node16   Node17   Node18   Node19   Node20   Node21 
#> 1.453255 1.477558 1.486610 1.489831 1.490848 1.490976 1.500334 1.504592 
#>   Node22   Node23   Node24   Node25   Node26   Node27   Node28   Node29 
#> 1.505968 1.508699 1.513668 1.517314 1.517394 1.518501 1.525898 1.526039 
#>   Node30   Node31   Node32   Node33   Node34   Node35   Node36   Node37 
#> 1.528349 1.530278 1.536275 1.537922 1.538692 1.539159 1.543480 1.545151 
#>   Node38   Node39   Node40   Node41   Node42   Node43   Node44   Node45 
#> 1.545396 1.545423 1.549210 1.550052 1.551560 1.551892 1.552226 1.553348 
#>   Node46   Node47   Node48   Node49   Node50   Node51   Node52   Node53 
#> 1.553703 1.555139 1.556349 1.558312 1.558668 1.558984 1.559031 1.559492 
#>   Node54   Node55   Node56   Node57   Node58   Node59   Node60   Node61 
#> 1.560157 1.560465 1.561129 1.561392 1.561952 1.562005 1.562256 1.563418 
#>   Node62   Node63   Node64   Node65   Node66   Node67   Node68   Node69 
#> 1.563440 1.563633 1.563947 1.564180 1.566007 1.567398 1.567548 1.567662 
#>   Node70   Node71   Node72   Node73   Node74   Node75   Node76   Node77 
#> 1.569419 1.569589 1.569757 1.569992 1.570241 1.570278 1.571104 1.571406 
#>   Node78   Node79   Node80   Node81   Node82   Node83   Node84   Node85 
#> 1.571685 1.571817 1.571985 1.572235 1.572279 1.572990 1.573030 1.573342 
#>   Node86   Node87   Node88   Node89   Node90   Node91   Node92   Node93 
#> 1.573541 1.573759 1.574324 1.574426 1.574599 1.574790 1.574811 1.574923 
#>   Node94   Node95   Node96   Node97   Node98   Node99 
#> 1.575046 1.575667 1.575770 1.575859 1.576105 1.576431