Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>       t86       t51       t83       t34       t62        t7       t71       t81 
#> 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 
#>       t53       t46       t25       t18       t27       t45       t41       t30 
#> 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 
#>        t8       t19       t80       t90       t28       t22       t94       t87 
#> 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 
#>       t56        t1       t85       t38       t33       t14       t11       t82 
#> 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 
#>      t100       t13       t75       t77       t23       t84       t20       t96 
#> 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 
#>       t99       t55       t16       t54        t9        t2       t50       t91 
#> 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 
#>       t74       t48       t67       t79       t61       t17       t52       t15 
#> 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 
#>        t3       t29       t76       t70       t63       t66       t32       t69 
#> 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 
#>       t47       t10       t24       t78       t65       t97        t6       t92 
#> 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 
#>       t57       t98       t44       t39       t36       t12       t64       t68 
#> 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 
#>       t42       t72        t5       t49       t73       t35       t60        t4 
#> 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 
#>       t37       t95       t89       t59       t43       t93       t21       t31 
#> 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 0.7633703 
#>       t26       t88       t40       t58     Node2     Node3     Node4     Node5 
#> 0.7633703 0.7633703 0.7633703 0.7633703 0.3813484 0.4226268 0.4395117 0.4497572 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 0.4611966 0.5306646 0.5333481 0.5574029 0.5595044 0.5676977 0.5996343 0.6229671 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 0.6458261 0.6666543 0.6680226 0.6774194 0.6809601 0.6809683 0.6822972 0.6870394 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 0.6973371 0.7008832 0.7065339 0.7102814 0.7147041 0.7218757 0.7226631 0.7236812 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 0.7238389 0.7239785 0.7245785 0.7257338 0.7264275 0.7287104 0.7301439 0.7311172 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 0.7314232 0.7324910 0.7331944 0.7338501 0.7344193 0.7350349 0.7358229 0.7388545 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 0.7408509 0.7412296 0.7426432 0.7427823 0.7430810 0.7432578 0.7435781 0.7444427 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 0.7458849 0.7460253 0.7460812 0.7469676 0.7470588 0.7486989 0.7499037 0.7507957 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 0.7507967 0.7509440 0.7510298 0.7519625 0.7521520 0.7522353 0.7528295 0.7531153 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 0.7533298 0.7536612 0.7538325 0.7542759 0.7544698 0.7547651 0.7548710 0.7549545 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 0.7558644 0.7563403 0.7565275 0.7565407 0.7568199 0.7570697 0.7571590 0.7577504 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 0.7600414 0.7600959 0.7601165 0.7604003 0.7609348 0.7611634 0.7614228 0.7618144 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 0.7621754 0.7622614 0.7623318 0.7624271 0.7626435 0.7633501