Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>       t81       t18       t32       t86       t35        t1       t95       t10 
#> 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 
#>       t84        t6       t74        t4       t87       t19       t60       t30 
#> 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 
#>       t39       t33       t65       t68       t27       t55       t21       t23 
#> 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 
#>       t77        t9       t64       t52       t78       t75       t20       t72 
#> 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 
#>       t11       t28       t66       t85       t37       t49       t91       t34 
#> 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 
#>       t57       t70       t16       t96       t31       t15       t48       t67 
#> 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 
#>        t7        t3       t24       t13        t8       t71       t89       t94 
#> 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 
#>       t83       t17       t29       t92       t44      t100       t51       t90 
#> 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 
#>       t46       t45       t47       t25       t82       t54       t43       t50 
#> 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 
#>       t26       t14       t56       t88       t63       t59       t12       t76 
#> 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 
#>       t98       t73       t22       t38       t58       t53        t5       t62 
#> 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 
#>       t42       t41       t61        t2       t97       t93       t36       t40 
#> 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 1.3757434 
#>       t69       t80       t99       t79     Node2     Node3     Node4     Node5 
#> 1.3757434 1.3757434 1.3757434 1.3757434 0.7524042 0.7965625 1.0329595 1.0346239 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 1.1312553 1.1319446 1.1651497 1.1653070 1.1901236 1.2099974 1.2597216 1.2734582 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 1.2768634 1.2959801 1.2978257 1.2986414 1.2986909 1.3017396 1.3021627 1.3032384 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 1.3090550 1.3106898 1.3113597 1.3169126 1.3211884 1.3258367 1.3290320 1.3320669 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 1.3353139 1.3356971 1.3394245 1.3404635 1.3407297 1.3448081 1.3475837 1.3511403 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 1.3511929 1.3519182 1.3539594 1.3545559 1.3545630 1.3546773 1.3569956 1.3572872 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 1.3581259 1.3585306 1.3589508 1.3592232 1.3594367 1.3608015 1.3610678 1.3617482 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 1.3618029 1.3619623 1.3621686 1.3622784 1.3623460 1.3630865 1.3636635 1.3637689 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 1.3642174 1.3642752 1.3650032 1.3656226 1.3656487 1.3666730 1.3670350 1.3675208 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 1.3684103 1.3687061 1.3689070 1.3689734 1.3690566 1.3694004 1.3700668 1.3704057 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 1.3705328 1.3707609 1.3710267 1.3722057 1.3722285 1.3723026 1.3723574 1.3725487 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 1.3728403 1.3730500 1.3730903 1.3735033 1.3736947 1.3739327 1.3742241 1.3742310 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 1.3743361 1.3747177 1.3749408 1.3749951 1.3751976 1.3752042