Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>       t90      t100       t89       t87       t63       t37        t4       t70 
#> 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 
#>       t82       t45        t2       t65       t71       t78       t49       t20 
#> 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 
#>       t91       t93       t43       t84       t80       t40       t23       t10 
#> 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 
#>       t62       t46       t21       t54       t13       t76       t18       t51 
#> 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 
#>       t55       t99       t29       t50       t72       t41       t24       t33 
#> 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 
#>       t38       t27       t56       t59       t53       t15        t8       t48 
#> 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 
#>       t31       t22       t32       t98       t58       t69       t34        t5 
#> 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 
#>       t17       t95        t7       t19       t96       t60       t94       t86 
#> 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 
#>       t85       t28       t83       t68       t57        t1       t12       t73 
#> 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 
#>       t39       t14       t66       t44       t36       t52       t16       t92 
#> 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 
#>       t79        t3       t47        t6       t64       t67       t30       t74 
#> 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 
#>        t9       t75       t77       t81       t35       t11       t61       t97 
#> 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 1.7388568 
#>       t26       t25       t42       t88     Node2     Node3     Node4     Node5 
#> 1.7388568 1.7388568 1.7388568 1.7388568 0.4455933 0.5912382 0.7518130 1.3023039 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 1.4193556 1.4426912 1.4596119 1.4791262 1.5303089 1.5426251 1.5587453 1.5805193 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 1.5840907 1.6157252 1.6202245 1.6420425 1.6451837 1.6464500 1.6482976 1.6530983 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 1.6576721 1.6637524 1.6671277 1.6748080 1.6824031 1.6827529 1.6891971 1.6895564 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 1.6956440 1.7013141 1.7022652 1.7022981 1.7068811 1.7073049 1.7073472 1.7080262 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 1.7089463 1.7092908 1.7100021 1.7100726 1.7105804 1.7110750 1.7110967 1.7119105 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 1.7123229 1.7131830 1.7145995 1.7153076 1.7169602 1.7170902 1.7171584 1.7183150 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 1.7183364 1.7193981 1.7195545 1.7198259 1.7202614 1.7203725 1.7216473 1.7218731 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 1.7221308 1.7229201 1.7230489 1.7246279 1.7260581 1.7278221 1.7279817 1.7294498 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 1.7297589 1.7299080 1.7302004 1.7308922 1.7310538 1.7311323 1.7328365 1.7331474 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 1.7333587 1.7333895 1.7335500 1.7336098 1.7336397 1.7337497 1.7344262 1.7344546 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 1.7345114 1.7345495 1.7348296 1.7355918 1.7357101 1.7358044 1.7360076 1.7369533 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 1.7373276 1.7377228 1.7378313 1.7379173 1.7383197 1.7386879