Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>      t27      t47      t36      t12      t69      t92      t64      t95 
#> 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 
#>      t67      t11      t58      t66      t62      t39      t44      t46 
#> 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 
#>      t89      t99       t5      t96      t74       t3      t17      t13 
#> 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 
#>      t16      t81      t32      t60      t40      t38      t61      t98 
#> 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 
#>      t34      t35      t97      t53      t72      t54      t42       t9 
#> 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 
#>      t82      t18       t4      t19      t21      t30      t59      t52 
#> 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 
#>      t45      t93       t6      t51      t55       t7       t8       t2 
#> 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 
#>      t63      t41      t78      t86      t65      t15      t50      t48 
#> 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 
#>      t24      t73      t56      t14      t80      t26      t84      t37 
#> 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 
#>      t28      t77     t100      t79      t43      t33      t10      t94 
#> 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 
#>      t31      t85      t83      t76      t49      t25      t91      t20 
#> 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 
#>      t75      t23      t70      t57       t1      t88      t68      t22 
#> 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 1.965542 
#>      t90      t71      t87      t29    Node2    Node3    Node4    Node5 
#> 1.965542 1.965542 1.965542 1.965542 1.139925 1.200918 1.248770 1.386074 
#>    Node6    Node7    Node8    Node9   Node10   Node11   Node12   Node13 
#> 1.587011 1.659704 1.699243 1.709651 1.751907 1.776871 1.838431 1.850114 
#>   Node14   Node15   Node16   Node17   Node18   Node19   Node20   Node21 
#> 1.860024 1.860588 1.862325 1.867393 1.872144 1.877131 1.881239 1.883504 
#>   Node22   Node23   Node24   Node25   Node26   Node27   Node28   Node29 
#> 1.884617 1.887568 1.892224 1.894488 1.894537 1.895237 1.904899 1.905617 
#>   Node30   Node31   Node32   Node33   Node34   Node35   Node36   Node37 
#> 1.915054 1.917019 1.921450 1.922619 1.923309 1.923843 1.926021 1.927107 
#>   Node38   Node39   Node40   Node41   Node42   Node43   Node44   Node45 
#> 1.931075 1.931762 1.936721 1.937052 1.937873 1.939847 1.939951 1.940075 
#>   Node46   Node47   Node48   Node49   Node50   Node51   Node52   Node53 
#> 1.941257 1.943129 1.943399 1.944384 1.946750 1.947427 1.947804 1.948431 
#>   Node54   Node55   Node56   Node57   Node58   Node59   Node60   Node61 
#> 1.948955 1.949766 1.950170 1.950899 1.950946 1.951290 1.951683 1.952715 
#>   Node62   Node63   Node64   Node65   Node66   Node67   Node68   Node69 
#> 1.953412 1.954085 1.954328 1.954717 1.954738 1.956031 1.956407 1.956577 
#>   Node70   Node71   Node72   Node73   Node74   Node75   Node76   Node77 
#> 1.956598 1.957236 1.958174 1.959038 1.959863 1.960113 1.960372 1.960402 
#>   Node78   Node79   Node80   Node81   Node82   Node83   Node84   Node85 
#> 1.960651 1.960989 1.961036 1.961092 1.961512 1.961759 1.961864 1.962503 
#>   Node86   Node87   Node88   Node89   Node90   Node91   Node92   Node93 
#> 1.963152 1.963637 1.963920 1.963979 1.964089 1.964119 1.964171 1.964257 
#>   Node94   Node95   Node96   Node97   Node98   Node99 
#> 1.964381 1.964620 1.964648 1.964683 1.964955 1.965534