Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>       t28        t8       t38       t14       t79       t92       t20       t86 
#> 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 
#>       t13       t37       t75        t7       t42       t18       t68       t53 
#> 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 
#>        t5       t34      t100        t4       t74       t44       t63       t55 
#> 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 
#>       t85       t46       t58        t2       t19       t76       t64       t97 
#> 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 
#>       t91       t60        t1       t25       t22       t17       t48       t89 
#> 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 
#>       t16       t35       t61       t87       t32       t50       t65       t39 
#> 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 
#>       t67       t15       t73       t99       t47       t95       t80       t49 
#> 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 
#>       t57       t98       t33       t82       t69       t11       t21        t3 
#> 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 
#>       t56       t36        t6       t88       t52       t72       t62       t77 
#> 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 
#>       t90       t93       t31        t9       t45       t26       t66       t94 
#> 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 
#>       t24       t30       t43       t96       t27       t81       t51       t40 
#> 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 
#>       t12       t84       t41       t59       t70       t54       t23       t83 
#> 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 0.8826033 
#>       t10       t29       t71       t78     Node2     Node3     Node4     Node5 
#> 0.8826033 0.8826033 0.8826033 0.8826033 0.2090232 0.4847516 0.4869042 0.5506035 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 0.5614172 0.6572996 0.6707814 0.6981923 0.7191277 0.7389719 0.7434947 0.7601158 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 0.7658831 0.7723628 0.7844808 0.8065775 0.8081587 0.8178929 0.8199745 0.8209283 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 0.8218872 0.8220756 0.8225242 0.8232129 0.8295314 0.8305448 0.8355596 0.8365899 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 0.8374110 0.8378156 0.8384796 0.8391506 0.8412750 0.8414476 0.8417138 0.8420782 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 0.8425374 0.8464021 0.8511298 0.8523469 0.8527831 0.8546652 0.8595448 0.8598012 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 0.8607703 0.8623181 0.8626902 0.8636224 0.8650624 0.8650696 0.8652164 0.8652551 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 0.8661684 0.8662816 0.8662853 0.8675868 0.8676839 0.8684540 0.8687829 0.8695064 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 0.8699096 0.8704110 0.8711814 0.8713546 0.8716859 0.8719893 0.8721091 0.8723515 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 0.8731446 0.8732989 0.8738763 0.8739274 0.8746334 0.8751194 0.8754236 0.8754251 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 0.8759716 0.8759906 0.8760356 0.8762059 0.8763890 0.8764559 0.8765206 0.8775365 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 0.8775755 0.8777377 0.8778863 0.8779653 0.8779757 0.8781231 0.8807547 0.8810319 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 0.8810424 0.8814564 0.8815315 0.8820824 0.8821230 0.8824783