Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>       t32       t57       t90       t91       t88       t68       t63        t5 
#> 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 
#>       t21       t69       t20       t60       t26        t3       t50       t76 
#> 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 
#>       t66       t23       t45       t83       t72       t33       t74       t31 
#> 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 
#>        t7       t51        t1       t98       t48       t92       t14       t89 
#> 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 
#>       t40       t86       t17       t16       t58       t79       t11       t18 
#> 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 
#>       t39       t52       t96       t70       t67       t46       t94       t87 
#> 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 
#>       t77       t53       t27       t54       t19        t4       t55       t99 
#> 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 
#>       t97       t84       t75        t2       t24       t37       t41       t62 
#> 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 
#>       t42       t49       t59       t47       t95       t65       t15       t80 
#> 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 
#>       t61        t8       t56       t73       t35       t28       t85        t9 
#> 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 
#>       t44       t43       t78       t71       t34       t12      t100       t10 
#> 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 
#>       t82        t6       t64       t22       t13       t30       t25       t93 
#> 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 0.8988556 
#>       t81       t38       t36       t29     Node2     Node3     Node4     Node5 
#> 0.8988556 0.8988556 0.8988556 0.8988556 0.5363140 0.5895693 0.6420756 0.7294814 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 0.7362627 0.7367440 0.7533085 0.7608951 0.7739242 0.7920646 0.7962307 0.7964372 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 0.7993200 0.8050122 0.8057266 0.8093857 0.8163887 0.8232687 0.8247618 0.8250937 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 0.8281490 0.8304630 0.8402150 0.8403820 0.8406586 0.8412735 0.8510377 0.8536674 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 0.8549197 0.8557965 0.8572689 0.8591252 0.8660228 0.8673786 0.8684966 0.8712242 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 0.8723587 0.8729496 0.8735621 0.8737798 0.8738835 0.8739783 0.8751780 0.8759271 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 0.8766448 0.8779169 0.8788896 0.8798288 0.8800202 0.8814499 0.8825869 0.8828088 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 0.8830073 0.8832007 0.8834277 0.8835343 0.8839911 0.8843593 0.8843974 0.8858905 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 0.8860616 0.8865211 0.8867125 0.8874563 0.8882127 0.8883610 0.8883749 0.8887885 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 0.8897670 0.8898109 0.8900546 0.8902539 0.8911100 0.8911308 0.8912433 0.8915836 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 0.8920364 0.8922657 0.8925313 0.8929353 0.8931882 0.8940549 0.8942714 0.8953267 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 0.8953678 0.8954902 0.8956939 0.8957832 0.8962827 0.8966739 0.8974274 0.8975003 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 0.8976426 0.8977597 0.8979761 0.8979990 0.8980614 0.8981046