Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>      t77      t50       t2      t92      t99      t33       t7      t59 
#> 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 
#>      t44      t53      t97       t1      t37      t31      t90       t6 
#> 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 
#>      t86      t56      t30      t26      t23      t27      t73      t63 
#> 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 
#>      t83       t9      t58      t93      t70      t51      t94      t85 
#> 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 
#>      t19      t32      t64      t17      t55      t12      t95      t72 
#> 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 
#>      t46      t60      t38      t42      t24      t20      t43      t39 
#> 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 
#>      t11      t66      t82      t40      t13      t22       t3      t80 
#> 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 
#>      t14      t47      t18      t98      t25      t16      t57      t28 
#> 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 
#>      t10      t62      t15       t5      t96      t74      t84      t54 
#> 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 
#>      t89     t100      t79      t76      t48      t71      t21      t78 
#> 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 
#>      t35      t34      t88      t87      t61      t91      t65       t4 
#> 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 
#>      t68       t8      t81      t49      t45      t36      t67      t29 
#> 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 6.510823 
#>      t75      t69      t52      t41    Node2    Node3    Node4    Node5 
#> 6.510823 6.510823 6.510823 6.510823 6.032393 6.082167 6.109995 6.185835 
#>    Node6    Node7    Node8    Node9   Node10   Node11   Node12   Node13 
#> 6.211474 6.241265 6.273857 6.275611 6.314349 6.321589 6.331742 6.336281 
#>   Node14   Node15   Node16   Node17   Node18   Node19   Node20   Node21 
#> 6.354382 6.363219 6.368117 6.407202 6.418016 6.423168 6.424495 6.437476 
#>   Node22   Node23   Node24   Node25   Node26   Node27   Node28   Node29 
#> 6.444748 6.446096 6.447201 6.449180 6.450918 6.456536 6.459189 6.461788 
#>   Node30   Node31   Node32   Node33   Node34   Node35   Node36   Node37 
#> 6.464717 6.473704 6.473853 6.476665 6.479218 6.479358 6.479963 6.482532 
#>   Node38   Node39   Node40   Node41   Node42   Node43   Node44   Node45 
#> 6.484255 6.487333 6.487596 6.488403 6.488937 6.488993 6.491349 6.491767 
#>   Node46   Node47   Node48   Node49   Node50   Node51   Node52   Node53 
#> 6.491973 6.493036 6.493112 6.494097 6.494196 6.494461 6.494987 6.495140 
#>   Node54   Node55   Node56   Node57   Node58   Node59   Node60   Node61 
#> 6.495148 6.495204 6.495953 6.498561 6.498674 6.498868 6.498923 6.499683 
#>   Node62   Node63   Node64   Node65   Node66   Node67   Node68   Node69 
#> 6.499780 6.499879 6.500518 6.502632 6.502921 6.503042 6.503394 6.503401 
#>   Node70   Node71   Node72   Node73   Node74   Node75   Node76   Node77 
#> 6.503518 6.503993 6.504668 6.504792 6.505068 6.505322 6.505358 6.505443 
#>   Node78   Node79   Node80   Node81   Node82   Node83   Node84   Node85 
#> 6.505613 6.505986 6.506363 6.506407 6.507978 6.507978 6.508233 6.508577 
#>   Node86   Node87   Node88   Node89   Node90   Node91   Node92   Node93 
#> 6.508899 6.509099 6.509162 6.509272 6.509301 6.509335 6.509568 6.509795 
#>   Node94   Node95   Node96   Node97   Node98   Node99 
#> 6.510054 6.510303 6.510476 6.510605 6.510612 6.510661