Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>         t2        t99        t25        t91        t98        t86        t56 
#> 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 
#>        t63        t77        t89        t85        t95        t49        t32 
#> 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 
#>        t31        t41        t20        t23        t45        t97        t51 
#> 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 
#>        t90        t17        t15        t67        t40        t62        t34 
#> 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 
#>        t43        t83        t76        t84         t3        t75        t10 
#> 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 
#>        t50        t14        t58        t72        t42        t37        t82 
#> 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 
#>        t47        t94        t19        t88        t93        t55        t66 
#> 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 
#>         t8        t12        t79        t26        t16        t28        t68 
#> 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 
#>        t29        t35        t21        t46        t24        t80        t54 
#> 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 
#>        t44        t65         t7        t71         t6        t53        t13 
#> 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 
#>        t27        t33        t70        t69        t48        t11        t87 
#> 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 
#>         t9        t81         t5        t22        t78        t61         t4 
#> 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 
#>        t96        t39        t18       t100        t57        t38        t36 
#> 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 
#>        t64         t1        t73        t60        t92        t59        t52 
#> 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 0.73555423 
#>        t74        t30      Node2      Node3      Node4      Node5      Node6 
#> 0.73555423 0.73555423 0.02663617 0.15289501 0.27861068 0.36683732 0.42229949 
#>      Node7      Node8      Node9     Node10     Node11     Node12     Node13 
#> 0.45750708 0.57307078 0.57448736 0.59123916 0.59151752 0.59982118 0.60109774 
#>     Node14     Node15     Node16     Node17     Node18     Node19     Node20 
#> 0.63160268 0.63166326 0.63493569 0.65172479 0.65387328 0.65603161 0.66019364 
#>     Node21     Node22     Node23     Node24     Node25     Node26     Node27 
#> 0.66115550 0.66269416 0.66638798 0.66774194 0.67400762 0.67644709 0.67900350 
#>     Node28     Node29     Node30     Node31     Node32     Node33     Node34 
#> 0.67908543 0.67931163 0.67976125 0.68565388 0.68855158 0.69165649 0.69200055 
#>     Node35     Node36     Node37     Node38     Node39     Node40     Node41 
#> 0.69204140 0.69978712 0.70187758 0.70585444 0.70652443 0.70702031 0.70839702 
#>     Node42     Node43     Node44     Node45     Node46     Node47     Node48 
#> 0.70863062 0.70883978 0.71098130 0.71171053 0.71195702 0.71271569 0.71308624 
#>     Node49     Node50     Node51     Node52     Node53     Node54     Node55 
#> 0.71679698 0.71751774 0.71753223 0.71874324 0.71891563 0.72036725 0.72124517 
#>     Node56     Node57     Node58     Node59     Node60     Node61     Node62 
#> 0.72219740 0.72223241 0.72249805 0.72280390 0.72292970 0.72342004 0.72431488 
#>     Node63     Node64     Node65     Node66     Node67     Node68     Node69 
#> 0.72495795 0.72499739 0.72521230 0.72593622 0.72683046 0.72748041 0.72754071 
#>     Node70     Node71     Node72     Node73     Node74     Node75     Node76 
#> 0.72760259 0.72815070 0.72860198 0.72860966 0.72864667 0.72917920 0.72980928 
#>     Node77     Node78     Node79     Node80     Node81     Node82     Node83 
#> 0.72983289 0.72985678 0.73029592 0.73055263 0.73085446 0.73165028 0.73198713 
#>     Node84     Node85     Node86     Node87     Node88     Node89     Node90 
#> 0.73214152 0.73247665 0.73249082 0.73271206 0.73292067 0.73318779 0.73369205 
#>     Node91     Node92     Node93     Node94     Node95     Node96     Node97 
#> 0.73392026 0.73443835 0.73467841 0.73479614 0.73491333 0.73495301 0.73500494 
#>     Node98     Node99 
#> 0.73525928 0.73550093