Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>        t98        t83        t22        t13         t8        t26        t99 
#> 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 
#>        t47        t45        t81        t89        t24        t95        t17 
#> 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 
#>        t11        t90        t71        t55        t20        t57        t62 
#> 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 
#>        t59        t74        t39        t30        t54         t7        t87 
#> 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 
#>        t97        t94         t9        t48        t61        t75        t68 
#> 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 
#>        t84        t46        t76        t79        t92        t34        t51 
#> 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 
#>        t53         t1        t44        t25        t67        t60        t73 
#> 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 
#>        t29        t52        t15        t91        t49        t88        t35 
#> 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 
#>        t78        t18        t63        t42         t2        t77        t70 
#> 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 
#>        t14        t56        t38        t58        t93        t50        t85 
#> 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 
#>        t86        t16        t65        t64        t19        t80        t66 
#> 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 
#>        t10        t27         t4        t32        t31         t3        t36 
#> 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 
#>        t21        t43        t82        t40        t72        t28        t96 
#> 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 
#>         t5        t37        t23        t41        t33        t12         t6 
#> 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 0.60742575 
#>        t69       t100      Node2      Node3      Node4      Node5      Node6 
#> 0.60742575 0.60742575 0.04074126 0.11554729 0.17810725 0.23453825 0.38893831 
#>      Node7      Node8      Node9     Node10     Node11     Node12     Node13 
#> 0.43478995 0.43882552 0.47485510 0.47773824 0.48412763 0.48771298 0.49773489 
#>     Node14     Node15     Node16     Node17     Node18     Node19     Node20 
#> 0.50652640 0.50724241 0.50998340 0.52954858 0.53578396 0.54158534 0.54251627 
#>     Node21     Node22     Node23     Node24     Node25     Node26     Node27 
#> 0.54592971 0.54596855 0.55044382 0.55052027 0.55317732 0.55537073 0.55675174 
#>     Node28     Node29     Node30     Node31     Node32     Node33     Node34 
#> 0.56077889 0.56396707 0.56449237 0.56796119 0.56829329 0.57150634 0.57216968 
#>     Node35     Node36     Node37     Node38     Node39     Node40     Node41 
#> 0.57350552 0.57480776 0.57492034 0.57522433 0.57562227 0.57600378 0.57637776 
#>     Node42     Node43     Node44     Node45     Node46     Node47     Node48 
#> 0.57658154 0.57764459 0.57871117 0.58149207 0.58193151 0.58302950 0.58416170 
#>     Node49     Node50     Node51     Node52     Node53     Node54     Node55 
#> 0.58446637 0.58503210 0.58688222 0.58768911 0.58815711 0.58861936 0.58911343 
#>     Node56     Node57     Node58     Node59     Node60     Node61     Node62 
#> 0.58950311 0.58957750 0.58996420 0.59043060 0.59092288 0.59097921 0.59123914 
#>     Node63     Node64     Node65     Node66     Node67     Node68     Node69 
#> 0.59171943 0.59245317 0.59282377 0.59297320 0.59341494 0.59429065 0.59510512 
#>     Node70     Node71     Node72     Node73     Node74     Node75     Node76 
#> 0.59686839 0.59805918 0.59807521 0.59860206 0.59883390 0.59884537 0.59891434 
#>     Node77     Node78     Node79     Node80     Node81     Node82     Node83 
#> 0.59912169 0.59913422 0.59933634 0.60022413 0.60118606 0.60138496 0.60194821 
#>     Node84     Node85     Node86     Node87     Node88     Node89     Node90 
#> 0.60286292 0.60310408 0.60389205 0.60394683 0.60406089 0.60411614 0.60415407 
#>     Node91     Node92     Node93     Node94     Node95     Node96     Node97 
#> 0.60456882 0.60457135 0.60475370 0.60486281 0.60533629 0.60576728 0.60628758 
#>     Node98     Node99 
#> 0.60684910 0.60741869