Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>       t62       t19       t43       t84       t60       t96       t16        t6 
#> 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 
#>       t63       t24       t26       t28       t58       t61        t3       t57 
#> 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 
#>       t52       t75       t41       t25       t95        t1       t29       t48 
#> 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 
#>       t44       t27       t36       t70       t51       t18       t94        t7 
#> 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 
#>       t56       t59       t14       t47       t85       t82       t76       t17 
#> 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 
#>       t65       t67       t81       t89       t83       t30       t71       t21 
#> 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 
#>       t64       t87       t31       t40       t99       t98       t78        t4 
#> 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 
#>       t69       t32       t97       t33       t49       t79       t86       t35 
#> 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 
#>       t46       t23        t5       t15       t68       t90       t54       t77 
#> 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 
#>       t12       t92       t20       t42       t39        t9       t72       t37 
#> 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 
#>      t100       t53        t2       t73       t91       t50       t22       t66 
#> 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 
#>       t55       t11       t93       t80       t34       t88       t45        t8 
#> 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 1.3011576 
#>       t38       t10       t13       t74     Node2     Node3     Node4     Node5 
#> 1.3011576 1.3011576 1.3011576 1.3011576 0.5526488 0.5938915 0.6209863 0.9230010 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 0.9929099 1.0016346 1.0238837 1.0505996 1.0527721 1.1568753 1.1741446 1.1747692 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 1.1790004 1.1825998 1.1928977 1.2035422 1.2121183 1.2138767 1.2145868 1.2160156 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 1.2203395 1.2280117 1.2345444 1.2417017 1.2458036 1.2506215 1.2519362 1.2541130 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 1.2548542 1.2572880 1.2582807 1.2587117 1.2595634 1.2617675 1.2620165 1.2663429 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 1.2701731 1.2706252 1.2713716 1.2720973 1.2744830 1.2745628 1.2749438 1.2762159 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 1.2766346 1.2781180 1.2785893 1.2790481 1.2793844 1.2823685 1.2828051 1.2834436 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 1.2836972 1.2837127 1.2853053 1.2861104 1.2868866 1.2871982 1.2884978 1.2889213 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 1.2894031 1.2894259 1.2894268 1.2907426 1.2915507 1.2925174 1.2928181 1.2930434 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 1.2930666 1.2930776 1.2932975 1.2941775 1.2944323 1.2949633 1.2956024 1.2957214 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 1.2958913 1.2965987 1.2966103 1.2968904 1.2969397 1.2969907 1.2972371 1.2974908 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 1.2988048 1.2988789 1.2990069 1.2993267 1.2993536 1.2995483 1.2996534 1.2998976 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 1.2999052 1.3003827 1.3004342 1.3005524 1.3007313 1.3007905