Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>      t25      t54      t35      t38      t30       t7       t3      t86 
#> 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 
#>      t92      t91      t73      t36      t11      t17      t88      t81 
#> 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 
#>      t49      t55      t99      t70      t61      t46      t43      t58 
#> 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 
#>      t93      t27      t10      t40      t39      t14      t69      t82 
#> 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 
#>      t89       t6      t19      t12      t60      t65      t15      t94 
#> 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 
#>     t100      t96      t52      t37      t57      t45      t66      t64 
#> 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 
#>      t22       t2      t44      t42      t98      t59      t74      t16 
#> 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 
#>       t1      t71      t68      t47      t32      t18      t85       t5 
#> 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 
#>      t20      t63      t26      t87      t75       t9      t95      t41 
#> 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 
#>      t21      t33      t97      t51      t77      t90      t84      t79 
#> 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 
#>      t31      t72      t83      t78      t48      t80      t34      t23 
#> 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 
#>       t8      t76      t67      t28      t13       t4      t62      t56 
#> 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 4.439202 
#>      t50      t24      t53      t29    Node2    Node3    Node4    Node5 
#> 4.439202 4.439202 4.439202 4.439202 2.927133 3.921745 3.978941 3.998803 
#>    Node6    Node7    Node8    Node9   Node10   Node11   Node12   Node13 
#> 4.043820 4.142086 4.182598 4.230138 4.232453 4.253473 4.292001 4.293498 
#>   Node14   Node15   Node16   Node17   Node18   Node19   Node20   Node21 
#> 4.318661 4.330032 4.333670 4.334763 4.338237 4.339996 4.344198 4.347392 
#>   Node22   Node23   Node24   Node25   Node26   Node27   Node28   Node29 
#> 4.350233 4.353125 4.354337 4.361261 4.365824 4.366593 4.373070 4.381489 
#>   Node30   Node31   Node32   Node33   Node34   Node35   Node36   Node37 
#> 4.381817 4.385504 4.386344 4.394436 4.397761 4.397844 4.399912 4.401804 
#>   Node38   Node39   Node40   Node41   Node42   Node43   Node44   Node45 
#> 4.403317 4.407548 4.408929 4.411766 4.412747 4.414408 4.414600 4.415328 
#>   Node46   Node47   Node48   Node49   Node50   Node51   Node52   Node53 
#> 4.416198 4.416469 4.417068 4.417109 4.417179 4.417470 4.417596 4.418055 
#>   Node54   Node55   Node56   Node57   Node58   Node59   Node60   Node61 
#> 4.418894 4.420011 4.421258 4.422784 4.423927 4.424783 4.424981 4.425838 
#>   Node62   Node63   Node64   Node65   Node66   Node67   Node68   Node69 
#> 4.426226 4.429020 4.429332 4.429439 4.429892 4.429911 4.430036 4.430537 
#>   Node70   Node71   Node72   Node73   Node74   Node75   Node76   Node77 
#> 4.431399 4.431458 4.432093 4.432144 4.432233 4.432330 4.432543 4.433973 
#>   Node78   Node79   Node80   Node81   Node82   Node83   Node84   Node85 
#> 4.434524 4.434904 4.435314 4.435466 4.435546 4.435761 4.435765 4.435886 
#>   Node86   Node87   Node88   Node89   Node90   Node91   Node92   Node93 
#> 4.436078 4.436193 4.436283 4.437220 4.437277 4.437427 4.437545 4.437604 
#>   Node94   Node95   Node96   Node97   Node98   Node99 
#> 4.437873 4.438084 4.438170 4.438185 4.438598 4.439069