Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>       t39        t4       t48       t30        t7       t40       t85       t34 
#> 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 
#>       t94       t56       t65       t23       t97       t83       t29       t98 
#> 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 
#>       t35       t15       t82       t21        t2       t52       t43       t74 
#> 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 
#>       t44       t59       t80       t11       t33       t81       t69       t38 
#> 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 
#>       t49        t8       t14       t28       t17        t3       t67       t68 
#> 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 
#>       t31        t1       t72      t100       t57       t45       t24       t42 
#> 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 
#>       t26       t86       t99       t92       t19       t63       t12       t54 
#> 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 
#>       t96        t6       t25       t87       t75       t22       t84       t88 
#> 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 
#>       t55       t93       t78       t53       t89       t90       t62       t46 
#> 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 
#>       t76       t61       t13       t95       t71       t50       t79       t91 
#> 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 
#>       t37       t66       t20        t5       t51       t60       t10       t16 
#> 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 
#>       t64       t58       t77        t9       t32       t41       t27       t70 
#> 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 1.6734248 
#>       t36       t47       t18       t73     Node2     Node3     Node4     Node5 
#> 1.6734248 1.6734248 1.6734248 1.6734248 0.4252677 0.7688671 1.1839242 1.3821326 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 1.4561142 1.4610207 1.4685559 1.4740023 1.5485904 1.5557155 1.5656565 1.5672289 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 1.5740168 1.5789593 1.5851179 1.5875410 1.5968578 1.5977850 1.5988983 1.6126694 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 1.6147013 1.6162775 1.6193551 1.6199228 1.6246198 1.6250560 1.6296039 1.6297578 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 1.6300054 1.6312374 1.6321074 1.6357738 1.6360234 1.6363054 1.6372669 1.6383208 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 1.6416512 1.6417046 1.6433925 1.6435087 1.6436541 1.6476052 1.6478460 1.6482213 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 1.6500632 1.6505415 1.6510171 1.6518201 1.6543135 1.6545897 1.6554494 1.6555278 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 1.6555749 1.6560182 1.6567997 1.6571531 1.6588614 1.6589503 1.6593593 1.6593688 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 1.6601241 1.6606802 1.6613521 1.6617715 1.6617876 1.6617880 1.6620323 1.6630505 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 1.6631423 1.6633217 1.6639468 1.6647575 1.6650545 1.6655839 1.6664179 1.6669034 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 1.6674619 1.6674786 1.6675340 1.6676907 1.6677274 1.6680552 1.6684480 1.6690477 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 1.6692255 1.6694026 1.6698912 1.6702874 1.6703371 1.6707489 1.6714555 1.6716165 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 1.6720956 1.6727445 1.6728197 1.6730480 1.6730607 1.6731591