Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>      t58      t47      t52      t66      t91      t41      t59       t1 
#> 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 
#>      t68      t70       t2      t34      t32      t20      t17      t45 
#> 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 
#>      t10      t77      t78       t6      t21      t26       t4      t81 
#> 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 
#>      t51      t80      t89      t46      t75      t13      t44      t12 
#> 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 
#>      t48      t86      t60       t3      t31      t39      t16      t14 
#> 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 
#>      t50      t22      t36      t62      t56      t72      t54      t35 
#> 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 
#>      t99      t88      t69      t96      t92       t9       t5      t49 
#> 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 
#>      t82      t84      t90      t15      t65      t33      t85      t83 
#> 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 
#>      t67      t29      t28       t8      t37      t93      t19      t27 
#> 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 
#>      t64      t63      t43      t57      t42      t25      t74      t55 
#> 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 
#>      t98      t61      t23      t94      t97      t38      t79      t18 
#> 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 
#>      t71      t11      t40       t7      t76      t73      t87     t100 
#> 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 3.062288 
#>      t30      t95      t24      t53    Node2    Node3    Node4    Node5 
#> 3.062288 3.062288 3.062288 3.062288 2.386477 2.674288 2.713613 2.714913 
#>    Node6    Node7    Node8    Node9   Node10   Node11   Node12   Node13 
#> 2.795010 2.795316 2.822082 2.822092 2.825871 2.851158 2.883213 2.896241 
#>   Node14   Node15   Node16   Node17   Node18   Node19   Node20   Node21 
#> 2.901113 2.928199 2.941417 2.949021 2.959100 2.973642 2.985471 2.988575 
#>   Node22   Node23   Node24   Node25   Node26   Node27   Node28   Node29 
#> 2.991470 3.000389 3.001437 3.003926 3.007203 3.009196 3.009405 3.009983 
#>   Node30   Node31   Node32   Node33   Node34   Node35   Node36   Node37 
#> 3.009990 3.011926 3.012301 3.014222 3.016869 3.017760 3.018140 3.024086 
#>   Node38   Node39   Node40   Node41   Node42   Node43   Node44   Node45 
#> 3.025836 3.027476 3.029233 3.029371 3.029850 3.030107 3.032358 3.032750 
#>   Node46   Node47   Node48   Node49   Node50   Node51   Node52   Node53 
#> 3.034201 3.034795 3.035236 3.036079 3.036603 3.037457 3.038390 3.038623 
#>   Node54   Node55   Node56   Node57   Node58   Node59   Node60   Node61 
#> 3.039372 3.040325 3.040699 3.041115 3.043039 3.045325 3.045464 3.045653 
#>   Node62   Node63   Node64   Node65   Node66   Node67   Node68   Node69 
#> 3.045958 3.046327 3.046418 3.048085 3.048412 3.049458 3.049683 3.050926 
#>   Node70   Node71   Node72   Node73   Node74   Node75   Node76   Node77 
#> 3.051016 3.051022 3.051076 3.051974 3.052133 3.052192 3.054285 3.054614 
#>   Node78   Node79   Node80   Node81   Node82   Node83   Node84   Node85 
#> 3.054737 3.055013 3.055215 3.055793 3.056220 3.056282 3.057050 3.057288 
#>   Node86   Node87   Node88   Node89   Node90   Node91   Node92   Node93 
#> 3.057590 3.057811 3.058820 3.058963 3.060239 3.060323 3.060541 3.061013 
#>   Node94   Node95   Node96   Node97   Node98   Node99 
#> 3.061164 3.061186 3.061240 3.061298 3.061694 3.062252