Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>      t28      t60      t84      t47      t38      t50      t44       t8 
#> 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 
#>      t12      t25      t49      t94      t33     t100      t22      t10 
#> 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 
#>      t67      t56      t64      t45      t51      t31      t32      t61 
#> 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 
#>      t62      t39      t90      t26      t52      t14      t58      t83 
#> 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 
#>      t53      t43      t80      t82      t57       t4      t35      t69 
#> 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 
#>      t42      t86      t18      t68      t46      t55      t76      t21 
#> 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 
#>       t1      t93       t6      t96       t9      t74       t3      t81 
#> 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 
#>      t97      t89      t77      t95      t88      t99      t23      t27 
#> 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 
#>       t5      t98      t13      t29      t34      t30      t16      t59 
#> 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 
#>      t70       t2      t75      t15      t24      t54      t36      t41 
#> 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 
#>      t17      t71      t72      t63      t65      t85      t19      t37 
#> 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 
#>      t20      t79      t40       t7      t66      t92      t87      t73 
#> 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 2.524849 
#>      t78      t11      t48      t91    Node2    Node3    Node4    Node5 
#> 2.524849 2.524849 2.524849 2.524849 1.600015 1.849510 2.077204 2.166047 
#>    Node6    Node7    Node8    Node9   Node10   Node11   Node12   Node13 
#> 2.216173 2.247434 2.258656 2.300856 2.317047 2.336401 2.375109 2.386223 
#>   Node14   Node15   Node16   Node17   Node18   Node19   Node20   Node21 
#> 2.388927 2.399792 2.404574 2.413748 2.424493 2.446792 2.449853 2.450995 
#>   Node22   Node23   Node24   Node25   Node26   Node27   Node28   Node29 
#> 2.451827 2.466000 2.468222 2.469356 2.471918 2.475541 2.476676 2.477020 
#>   Node30   Node31   Node32   Node33   Node34   Node35   Node36   Node37 
#> 2.478602 2.481502 2.481622 2.483562 2.484735 2.486067 2.486348 2.486914 
#>   Node38   Node39   Node40   Node41   Node42   Node43   Node44   Node45 
#> 2.489070 2.489797 2.491706 2.491798 2.493172 2.493182 2.493767 2.496138 
#>   Node46   Node47   Node48   Node49   Node50   Node51   Node52   Node53 
#> 2.497770 2.499372 2.500536 2.500704 2.501300 2.502131 2.504989 2.505115 
#>   Node54   Node55   Node56   Node57   Node58   Node59   Node60   Node61 
#> 2.505664 2.506099 2.507023 2.507370 2.508864 2.508985 2.509281 2.509510 
#>   Node62   Node63   Node64   Node65   Node66   Node67   Node68   Node69 
#> 2.509996 2.513077 2.513169 2.513367 2.513855 2.513933 2.514045 2.514052 
#>   Node70   Node71   Node72   Node73   Node74   Node75   Node76   Node77 
#> 2.514359 2.514361 2.514364 2.514399 2.514461 2.514650 2.514855 2.514901 
#>   Node78   Node79   Node80   Node81   Node82   Node83   Node84   Node85 
#> 2.515237 2.517630 2.519172 2.519809 2.519858 2.520190 2.520590 2.520719 
#>   Node86   Node87   Node88   Node89   Node90   Node91   Node92   Node93 
#> 2.520895 2.522476 2.523273 2.523322 2.523474 2.523572 2.523677 2.524155 
#>   Node94   Node95   Node96   Node97   Node98   Node99 
#> 2.524256 2.524317 2.524371 2.524553 2.524715 2.524831