Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>       t70       t88       t89       t95       t33        t2       t26       t49 
#> 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 
#>       t35       t76       t13       t62       t39       t45       t78       t77 
#> 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 
#>       t55       t65        t1       t71       t81       t11       t84       t91 
#> 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 
#>       t61       t93        t8       t99       t58       t59       t34       t64 
#> 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 
#>       t54       t69       t46       t80       t31       t32       t21       t18 
#> 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 
#>       t66       t51       t37       t19       t79       t14       t97       t36 
#> 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 
#>       t43       t38       t85       t75       t25        t5       t90       t96 
#> 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 
#>       t44        t7       t83       t67       t30        t4       t50        t3 
#> 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 
#>       t28       t94       t24       t73       t56       t40       t12        t9 
#> 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 
#>       t57       t82       t72       t10       t17       t20       t22       t29 
#> 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 
#>       t60       t53       t15      t100        t6       t74       t27       t48 
#> 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 
#>       t52       t41       t92       t63       t16       t68       t23       t87 
#> 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 1.0656704 
#>       t47       t86       t98       t42     Node2     Node3     Node4     Node5 
#> 1.0656704 1.0656704 1.0656704 1.0656704 0.6675285 0.6772680 0.8174504 0.8692801 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 0.8906822 0.8921858 0.8930545 0.8951987 0.8985506 0.9181512 0.9199990 0.9235509 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 0.9247512 0.9416303 0.9488088 0.9792215 0.9801338 0.9809396 0.9818726 0.9846738 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 0.9900595 0.9929116 0.9936233 0.9958393 0.9967760 0.9983522 1.0077481 1.0116234 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 1.0119582 1.0144760 1.0232540 1.0242604 1.0260492 1.0275605 1.0294761 1.0301540 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 1.0304191 1.0309642 1.0316649 1.0321279 1.0343245 1.0350966 1.0357339 1.0363373 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 1.0363480 1.0400957 1.0419400 1.0447807 1.0469259 1.0469379 1.0472045 1.0474329 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 1.0483395 1.0484659 1.0493414 1.0505274 1.0506154 1.0512227 1.0512531 1.0513766 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 1.0525005 1.0528952 1.0538594 1.0538628 1.0561713 1.0565051 1.0567079 1.0568182 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 1.0568733 1.0574530 1.0575277 1.0587303 1.0593473 1.0595020 1.0604548 1.0607195 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 1.0618232 1.0619202 1.0620925 1.0623306 1.0623723 1.0623956 1.0624643 1.0627440 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 1.0632380 1.0634515 1.0638655 1.0639641 1.0640478 1.0640869 1.0642442 1.0645911 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 1.0649455 1.0651545 1.0652154 1.0653334 1.0653795 1.0655324