Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>       t17       t90       t41       t57       t46       t10       t66       t92 
#> 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 
#>       t19       t95       t44      t100       t89       t32        t2       t37 
#> 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 
#>       t64       t35       t11       t40       t29       t48       t93       t14 
#> 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 
#>       t82       t53       t50       t68       t84       t43       t27       t22 
#> 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 
#>       t85       t76       t28       t12       t45       t77       t39       t34 
#> 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 
#>       t18       t70       t42        t4       t96        t3       t33       t79 
#> 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 
#>       t71       t21       t91       t63       t26        t9       t25       t51 
#> 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 
#>       t69       t58       t86       t55        t8       t36       t87       t23 
#> 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 
#>       t75       t72       t81       t47       t52       t30       t65       t13 
#> 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 
#>        t5       t73       t60       t15       t99        t7       t78       t61 
#> 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 
#>       t98        t6       t31       t16       t24       t80       t94       t56 
#> 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 
#>       t88        t1       t20       t49       t38       t62       t54       t83 
#> 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 1.0607107 
#>       t67       t97       t59       t74     Node2     Node3     Node4     Node5 
#> 1.0607107 1.0607107 1.0607107 1.0607107 0.1325940 0.7005830 0.7044946 0.7086916 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 0.7143460 0.7450101 0.7562995 0.7624777 0.8031675 0.8885280 0.8930705 0.9097362 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 0.9401517 0.9436535 0.9500434 0.9526526 0.9574946 0.9715308 0.9760010 0.9838505 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 0.9859307 0.9958970 0.9974820 0.9997039 1.0044389 1.0075867 1.0151306 1.0161803 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 1.0192294 1.0193331 1.0210711 1.0228494 1.0231367 1.0241756 1.0242509 1.0244204 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 1.0245670 1.0249007 1.0267607 1.0283514 1.0288038 1.0308594 1.0330068 1.0333647 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 1.0334817 1.0337163 1.0368160 1.0378870 1.0380016 1.0387213 1.0387351 1.0390057 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 1.0392924 1.0393080 1.0396328 1.0412547 1.0417754 1.0423287 1.0439067 1.0440849 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 1.0454753 1.0457371 1.0457907 1.0463036 1.0476233 1.0477215 1.0483510 1.0484664 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 1.0494270 1.0503015 1.0507732 1.0510469 1.0515480 1.0516148 1.0530250 1.0531354 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 1.0533101 1.0540013 1.0540147 1.0541723 1.0545462 1.0548611 1.0554074 1.0558819 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 1.0565470 1.0585566 1.0587548 1.0590294 1.0594633 1.0598363 1.0598738 1.0600062 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 1.0602003 1.0602347 1.0602539 1.0604294 1.0605462 1.0606407