Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>       t87       t47       t49       t54       t81       t73       t83       t18 
#> 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 
#>       t98       t92       t57       t53       t80        t5       t36       t11 
#> 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 
#>       t51       t93       t89       t62       t39       t30       t37       t20 
#> 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 
#>       t99       t17       t15       t29       t67       t75       t63       t34 
#> 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 
#>       t74        t9       t79       t59       t70       t46       t77       t50 
#> 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 
#>       t35       t88       t12       t52       t38       t56        t8       t82 
#> 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 
#>        t7       t44       t65       t21       t25       t72        t2        t4 
#> 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 
#>       t40       t33      t100       t16       t13       t27       t94       t28 
#> 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 
#>       t58       t42       t64       t41       t97       t14       t26       t90 
#> 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 
#>       t19       t66       t69       t24       t22       t55       t86       t84 
#> 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 
#>       t60       t71       t76       t48       t45       t91       t85        t6 
#> 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 
#>       t95       t96       t43       t10       t23       t32        t3       t31 
#> 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 1.2538413 
#>       t61        t1       t68       t78     Node2     Node3     Node4     Node5 
#> 1.2538413 1.2538413 1.2538413 1.2538413 0.7527420 0.7573799 0.8660314 0.8698643 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 0.9037817 1.0061433 1.0096447 1.0286847 1.0303517 1.0393830 1.0765442 1.1401392 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 1.1401591 1.1488644 1.1495763 1.1510220 1.1654696 1.1720447 1.1785810 1.1802305 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 1.1824388 1.1849187 1.1873233 1.1873761 1.1928008 1.1932082 1.1960743 1.1970531 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 1.1982365 1.1991228 1.2016803 1.2034047 1.2095790 1.2108220 1.2149882 1.2150751 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 1.2152908 1.2166713 1.2183544 1.2205263 1.2231881 1.2254389 1.2254996 1.2268193 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 1.2280803 1.2288082 1.2291476 1.2301091 1.2317488 1.2318600 1.2321786 1.2329705 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 1.2345790 1.2354724 1.2357791 1.2360884 1.2365917 1.2370357 1.2375622 1.2386083 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 1.2392342 1.2395213 1.2407589 1.2408238 1.2417129 1.2417597 1.2424808 1.2425302 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 1.2436444 1.2436471 1.2441567 1.2445977 1.2455132 1.2462381 1.2464429 1.2469062 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 1.2469723 1.2474656 1.2476886 1.2477795 1.2482219 1.2483662 1.2489152 1.2500485 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 1.2500959 1.2502060 1.2503970 1.2509015 1.2511844 1.2514199 1.2515284 1.2519084 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 1.2520080 1.2521563 1.2527291 1.2530820 1.2535022 1.2538365