Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>       t42       t87       t37       t56       t66       t14       t74       t43 
#> 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 
#>       t18       t79       t38       t39        t7        t9       t47       t34 
#> 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 
#>       t20       t71       t40       t82       t98       t26        t5       t15 
#> 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 
#>       t97       t51       t86       t36        t2       t35       t55       t61 
#> 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 
#>       t78       t88        t8       t53       t27       t76        t3       t58 
#> 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 
#>       t21       t62       t49       t28       t25       t31       t17       t77 
#> 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 
#>      t100       t83       t33       t72       t52       t13       t89       t41 
#> 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 
#>       t12       t57       t69        t1       t63       t93       t54       t84 
#> 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 
#>       t85       t64       t50       t60       t91       t48       t90        t4 
#> 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 
#>       t24       t75       t80       t46       t59       t68       t11       t99 
#> 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 
#>       t94       t73        t6       t29       t45       t65       t30       t96 
#> 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 
#>       t67       t23       t95       t70       t32       t16       t19       t92 
#> 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 0.9673403 
#>       t22       t44       t10       t81     Node2     Node3     Node4     Node5 
#> 0.9673403 0.9673403 0.9673403 0.9673403 0.1157599 0.4996234 0.6187159 0.6780124 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 0.6916036 0.7049200 0.7054322 0.7312432 0.7953523 0.7985561 0.8047193 0.8204759 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 0.8540741 0.8570495 0.8593094 0.8840851 0.8883588 0.8918137 0.8993517 0.9004196 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 0.9009947 0.9017627 0.9044564 0.9046678 0.9073247 0.9075493 0.9096146 0.9099460 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 0.9165786 0.9202804 0.9215820 0.9217950 0.9276041 0.9294149 0.9322970 0.9337538 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 0.9388219 0.9395300 0.9401717 0.9402305 0.9414739 0.9428125 0.9429830 0.9438618 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 0.9441751 0.9454493 0.9463494 0.9470086 0.9478159 0.9483466 0.9490496 0.9491684 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 0.9493565 0.9504441 0.9515768 0.9516736 0.9525988 0.9526949 0.9537908 0.9538567 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 0.9544291 0.9544651 0.9545106 0.9545747 0.9546734 0.9559139 0.9568963 0.9575269 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 0.9576915 0.9582550 0.9584025 0.9594501 0.9596605 0.9599455 0.9601408 0.9605136 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 0.9607373 0.9610942 0.9611695 0.9626094 0.9629167 0.9631179 0.9635083 0.9636630 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 0.9637631 0.9638114 0.9638513 0.9641550 0.9641576 0.9648704 0.9650884 0.9655341 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 0.9657141 0.9657921 0.9660103 0.9660515 0.9663968 0.9665075