Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>        t29        t84        t34        t71        t89        t74       t100 
#> 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 
#>        t62        t80        t44        t12        t42        t56        t11 
#> 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 
#>        t24        t82        t51        t87         t6        t92        t26 
#> 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 
#>        t54        t65        t83        t93        t59        t60         t1 
#> 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 
#>        t61        t67        t35        t47        t16         t8        t99 
#> 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 
#>        t88        t58        t63        t96        t14        t90        t66 
#> 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 
#>        t52         t5        t21         t7        t50         t2        t18 
#> 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 
#>        t53        t46        t20        t77        t55        t91        t38 
#> 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 
#>        t68        t75        t79         t3        t57        t40        t10 
#> 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 
#>        t19         t9        t33        t43        t28        t98        t78 
#> 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 
#>        t48        t27        t69        t73        t41        t39        t72 
#> 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 
#>        t17        t15        t32        t23        t86        t36        t81 
#> 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 
#>        t85         t4        t76        t49        t25        t97        t64 
#> 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 
#>        t22        t94        t70        t13        t37        t30        t31 
#> 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 0.90500935 
#>        t45        t95      Node2      Node3      Node4      Node5      Node6 
#> 0.90500935 0.90500935 0.09322605 0.47856872 0.50674822 0.56200566 0.56840003 
#>      Node7      Node8      Node9     Node10     Node11     Node12     Node13 
#> 0.64852197 0.67069521 0.68512024 0.70347624 0.70909964 0.71406954 0.73122974 
#>     Node14     Node15     Node16     Node17     Node18     Node19     Node20 
#> 0.76742809 0.77233097 0.77243787 0.79035074 0.80550386 0.81118234 0.81992902 
#>     Node21     Node22     Node23     Node24     Node25     Node26     Node27 
#> 0.82695204 0.82932853 0.83361051 0.83433848 0.84266605 0.84348584 0.84482156 
#>     Node28     Node29     Node30     Node31     Node32     Node33     Node34 
#> 0.84727872 0.85019054 0.85215479 0.85677209 0.85746465 0.86076886 0.86727223 
#>     Node35     Node36     Node37     Node38     Node39     Node40     Node41 
#> 0.86849647 0.87141555 0.87213478 0.87269767 0.87338145 0.87434928 0.87516897 
#>     Node42     Node43     Node44     Node45     Node46     Node47     Node48 
#> 0.87520492 0.87529646 0.87573358 0.87626664 0.87896706 0.88311146 0.88432202 
#>     Node49     Node50     Node51     Node52     Node53     Node54     Node55 
#> 0.88437917 0.88445478 0.88496700 0.88542376 0.88636605 0.88712542 0.88922351 
#>     Node56     Node57     Node58     Node59     Node60     Node61     Node62 
#> 0.88952096 0.88979919 0.89001604 0.89009966 0.89036873 0.89057497 0.89117172 
#>     Node63     Node64     Node65     Node66     Node67     Node68     Node69 
#> 0.89177869 0.89192575 0.89222763 0.89226211 0.89230342 0.89319878 0.89473333 
#>     Node70     Node71     Node72     Node73     Node74     Node75     Node76 
#> 0.89490727 0.89495867 0.89537836 0.89604994 0.89645834 0.89673026 0.89697073 
#>     Node77     Node78     Node79     Node80     Node81     Node82     Node83 
#> 0.89725127 0.89740131 0.89759518 0.89971148 0.89983028 0.89987609 0.90079255 
#>     Node84     Node85     Node86     Node87     Node88     Node89     Node90 
#> 0.90111569 0.90128416 0.90131415 0.90155633 0.90225608 0.90227730 0.90228583 
#>     Node91     Node92     Node93     Node94     Node95     Node96     Node97 
#> 0.90232614 0.90261966 0.90307478 0.90326477 0.90336873 0.90361693 0.90364443 
#>     Node98     Node99 
#> 0.90444454 0.90492696