Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>         t3         t5        t14        t12        t74        t99        t35 
#> 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 
#>         t4        t49        t13        t76        t40        t28        t68 
#> 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 
#>        t18         t8        t55         t7        t62        t42        t81 
#> 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 
#>        t50        t51        t44        t97        t38        t36        t10 
#> 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 
#>        t95        t64        t78        t24        t58        t43        t46 
#> 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 
#>        t66        t15        t87        t41        t22        t52        t70 
#> 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 
#>        t83        t11        t85        t30        t32         t6        t69 
#> 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 
#>        t57        t27        t53        t77        t48         t1        t25 
#> 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 
#>        t54        t63        t80        t19        t20        t93        t16 
#> 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 
#>        t29        t94        t89        t39        t23        t31        t60 
#> 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 
#>        t75        t59        t33        t26        t37        t98        t56 
#> 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 
#>       t100        t96        t79        t67        t47        t71        t82 
#> 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 
#>        t72        t45        t90        t92        t88        t73        t21 
#> 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 
#>        t86        t61        t17        t34        t91         t9        t84 
#> 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 1.39588002 
#>        t65         t2      Node2      Node3      Node4      Node5      Node6 
#> 1.39588002 1.39588002 0.01284393 0.46671407 0.75143276 0.82447863 0.90949776 
#>      Node7      Node8      Node9     Node10     Node11     Node12     Node13 
#> 1.00463650 1.06145506 1.14377408 1.23652197 1.25589249 1.29158809 1.29367885 
#>     Node14     Node15     Node16     Node17     Node18     Node19     Node20 
#> 1.29554538 1.30091187 1.30918739 1.31144044 1.31178842 1.31531269 1.32949449 
#>     Node21     Node22     Node23     Node24     Node25     Node26     Node27 
#> 1.33654708 1.33665792 1.33857638 1.34333240 1.34652872 1.34929113 1.35009451 
#>     Node28     Node29     Node30     Node31     Node32     Node33     Node34 
#> 1.35114381 1.35377024 1.35397769 1.35461878 1.35602368 1.35719898 1.35851900 
#>     Node35     Node36     Node37     Node38     Node39     Node40     Node41 
#> 1.35899118 1.35990949 1.36175037 1.36255264 1.36614453 1.36633741 1.36721621 
#>     Node42     Node43     Node44     Node45     Node46     Node47     Node48 
#> 1.36821068 1.37072042 1.37123530 1.37227744 1.37460478 1.37514294 1.37598105 
#>     Node49     Node50     Node51     Node52     Node53     Node54     Node55 
#> 1.37599729 1.37830711 1.37848458 1.38027376 1.38049968 1.38158794 1.38316160 
#>     Node56     Node57     Node58     Node59     Node60     Node61     Node62 
#> 1.38328353 1.38397818 1.38455598 1.38456180 1.38550589 1.38566612 1.38688127 
#>     Node63     Node64     Node65     Node66     Node67     Node68     Node69 
#> 1.38765183 1.38778805 1.38779901 1.38814270 1.38827755 1.38865288 1.38874403 
#>     Node70     Node71     Node72     Node73     Node74     Node75     Node76 
#> 1.38881983 1.38886645 1.38890331 1.38898388 1.38937696 1.38975619 1.38993740 
#>     Node77     Node78     Node79     Node80     Node81     Node82     Node83 
#> 1.39069090 1.39095108 1.39100633 1.39167169 1.39169147 1.39224351 1.39236730 
#>     Node84     Node85     Node86     Node87     Node88     Node89     Node90 
#> 1.39249434 1.39283524 1.39290885 1.39325492 1.39343659 1.39346145 1.39353265 
#>     Node91     Node92     Node93     Node94     Node95     Node96     Node97 
#> 1.39364954 1.39370693 1.39431120 1.39435687 1.39503865 1.39510699 1.39557096 
#>     Node98     Node99 
#> 1.39572188 1.39583622