Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>      t96       t4      t32      t39      t22      t53      t62      t34 
#> 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 
#>      t19      t76      t61      t11      t35      t57      t49      t58 
#> 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 
#>      t63      t10      t71      t14      t44      t82      t38      t72 
#> 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 
#>      t80      t93      t45       t8      t47      t40       t6      t55 
#> 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 
#>      t77      t59      t81      t60      t64      t41      t69      t28 
#> 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 
#>      t78      t87       t5      t50      t21      t85      t48      t84 
#> 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 
#>      t88      t74      t15      t66      t83      t75       t3      t65 
#> 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 
#>      t31      t27      t18      t97      t23      t99      t29      t43 
#> 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 
#>      t54      t46      t25      t68      t37      t30       t1      t20 
#> 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 
#>      t33      t94      t98      t42      t92      t16      t91      t95 
#> 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 
#>      t70      t36      t79      t56      t89      t86      t90      t67 
#> 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 
#>      t51      t12      t73       t7       t9      t26       t2      t52 
#> 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 1.546727 
#>      t24     t100      t13      t17    Node2    Node3    Node4    Node5 
#> 1.546727 1.546727 1.546727 1.546727 1.074692 1.112435 1.159058 1.161110 
#>    Node6    Node7    Node8    Node9   Node10   Node11   Node12   Node13 
#> 1.182358 1.248892 1.364918 1.378946 1.387186 1.419134 1.420279 1.445829 
#>   Node14   Node15   Node16   Node17   Node18   Node19   Node20   Node21 
#> 1.456418 1.456942 1.462211 1.465465 1.471253 1.476763 1.482428 1.482667 
#>   Node22   Node23   Node24   Node25   Node26   Node27   Node28   Node29 
#> 1.483717 1.489589 1.490838 1.491474 1.494877 1.500144 1.501970 1.503950 
#>   Node30   Node31   Node32   Node33   Node34   Node35   Node36   Node37 
#> 1.503954 1.504506 1.506858 1.508473 1.510792 1.511539 1.512183 1.512231 
#>   Node38   Node39   Node40   Node41   Node42   Node43   Node44   Node45 
#> 1.513507 1.514626 1.516921 1.518002 1.518154 1.519323 1.520173 1.520439 
#>   Node46   Node47   Node48   Node49   Node50   Node51   Node52   Node53 
#> 1.520548 1.521344 1.522952 1.522971 1.523353 1.524195 1.525620 1.526410 
#>   Node54   Node55   Node56   Node57   Node58   Node59   Node60   Node61 
#> 1.526529 1.526883 1.526938 1.528346 1.528430 1.529355 1.529430 1.529464 
#>   Node62   Node63   Node64   Node65   Node66   Node67   Node68   Node69 
#> 1.531645 1.532120 1.532485 1.532872 1.533836 1.533937 1.534228 1.534803 
#>   Node70   Node71   Node72   Node73   Node74   Node75   Node76   Node77 
#> 1.536864 1.536930 1.537025 1.537302 1.537533 1.538040 1.539439 1.539444 
#>   Node78   Node79   Node80   Node81   Node82   Node83   Node84   Node85 
#> 1.539950 1.540366 1.541258 1.541451 1.541582 1.541770 1.541931 1.542040 
#>   Node86   Node87   Node88   Node89   Node90   Node91   Node92   Node93 
#> 1.542697 1.542726 1.543126 1.543129 1.543287 1.543558 1.543781 1.544526 
#>   Node94   Node95   Node96   Node97   Node98   Node99 
#> 1.544821 1.544869 1.545331 1.545872 1.545883 1.546655