Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>       t43       t26       t15       t21       t73       t49       t91       t98 
#> 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 
#>        t2       t48       t56       t72       t18       t24       t90       t75 
#> 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 
#>       t44       t50       t61       t65       t93        t3       t57       t82 
#> 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 
#>       t64       t92       t51       t36       t28       t53       t83       t16 
#> 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 
#>       t34        t8       t95       t63       t42        t4       t86       t85 
#> 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 
#>        t5       t67       t99       t12       t23       t77       t17       t32 
#> 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 
#>       t66       t69       t55        t6       t87       t11       t88       t58 
#> 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 
#>       t29       t74       t13       t31       t41       t71       t62       t84 
#> 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 
#>       t19       t35       t59       t94       t47       t40       t45        t1 
#> 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 
#>       t54      t100       t39       t52       t37        t7       t96       t76 
#> 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 
#>       t81       t10       t68       t20       t30       t38       t89       t14 
#> 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 
#>        t9       t46       t97       t80       t25       t70       t33       t78 
#> 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 0.8415545 
#>       t60       t79       t27       t22     Node2     Node3     Node4     Node5 
#> 0.8415545 0.8415545 0.8415545 0.8415545 0.2869791 0.3986644 0.5165693 0.5243719 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 0.5384396 0.5600678 0.5688339 0.5779480 0.6124867 0.6276888 0.6833118 0.7198864 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 0.7265372 0.7313219 0.7351272 0.7582543 0.7614955 0.7686461 0.7722633 0.7755482 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 0.7823088 0.7914485 0.7929690 0.7945614 0.7964991 0.7965689 0.7978936 0.7993008 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 0.8066064 0.8083266 0.8091365 0.8093236 0.8108197 0.8123112 0.8127128 0.8145529 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 0.8159668 0.8166977 0.8173107 0.8178872 0.8180697 0.8190715 0.8197958 0.8203049 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 0.8207901 0.8212761 0.8225872 0.8237605 0.8245406 0.8246247 0.8258239 0.8262214 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 0.8264473 0.8265561 0.8265983 0.8272604 0.8274396 0.8288633 0.8297496 0.8298483 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 0.8300816 0.8308195 0.8310353 0.8311899 0.8313535 0.8313672 0.8319039 0.8325997 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 0.8329226 0.8329809 0.8335062 0.8336497 0.8340448 0.8343096 0.8343122 0.8348098 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 0.8349950 0.8355668 0.8356861 0.8359961 0.8360065 0.8363084 0.8368406 0.8369872 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 0.8372923 0.8377574 0.8383970 0.8388047 0.8392596 0.8394766 0.8398175 0.8400702 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 0.8402413 0.8404250 0.8405678 0.8411797 0.8412113 0.8415261