Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>       t93       t55       t58       t49       t97       t74       t48       t44 
#> 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 
#>       t21       t41       t50       t88       t12        t5       t99       t46 
#> 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 
#>       t54       t29       t78       t79       t56       t87       t32       t11 
#> 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 
#>       t37       t51       t45       t66       t53       t31       t20       t89 
#> 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 
#>       t23       t69       t59       t28       t17       t65      t100       t80 
#> 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 
#>       t25       t81       t94       t52       t14       t64        t6       t76 
#> 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 
#>       t82       t62       t30       t35       t60       t57       t47       t72 
#> 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 
#>       t77        t2       t70        t3       t67       t86        t7       t95 
#> 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 
#>       t43       t63       t90        t1       t98       t91       t96       t18 
#> 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 
#>       t83        t4       t38       t73       t33       t16       t42       t19 
#> 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 
#>       t10       t40       t39       t34        t8       t61       t84       t36 
#> 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 
#>       t68       t26       t13       t24       t22       t27       t75       t85 
#> 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 1.1480686 
#>        t9       t92       t71       t15     Node2     Node3     Node4     Node5 
#> 1.1480686 1.1480686 1.1480686 1.1480686 0.2036243 0.5331039 0.7631346 0.7977339 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 0.8930281 0.9450536 0.9496320 0.9629284 0.9661557 0.9758053 0.9947434 1.0157826 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 1.0286788 1.0367991 1.0448090 1.0556796 1.0558417 1.0635907 1.0657018 1.0675856 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 1.0692781 1.0733403 1.0741181 1.0753805 1.0829570 1.0876926 1.0925832 1.0996419 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 1.0997859 1.1018587 1.1069433 1.1097058 1.1138455 1.1146717 1.1166966 1.1177944 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 1.1184038 1.1184924 1.1192864 1.1203680 1.1203858 1.1242612 1.1248996 1.1260100 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 1.1260195 1.1261728 1.1263997 1.1282759 1.1286190 1.1287984 1.1292213 1.1292473 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 1.1305465 1.1318255 1.1330257 1.1339752 1.1345199 1.1346127 1.1353265 1.1356570 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 1.1358740 1.1361557 1.1377264 1.1382688 1.1383740 1.1385155 1.1390797 1.1393516 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 1.1408363 1.1415029 1.1419320 1.1421757 1.1426433 1.1428176 1.1435111 1.1436667 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 1.1441866 1.1443197 1.1444744 1.1445119 1.1445245 1.1458185 1.1459355 1.1459817 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 1.1462057 1.1464727 1.1466653 1.1469022 1.1470455 1.1472462 1.1473992 1.1474552 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 1.1474657 1.1475498 1.1476357 1.1478266 1.1478451 1.1478620