Calculate the sum of features for each flow

Sums the phylogenetic flow features for each flow in a pfc, returning a vector of values.

Usage

pf_flow_sum(x, ...)

Arguments

x

A pfc object

...

Other arguments to pass to or from other methods

Value

A vector equal to the length of x with the sum of each flow's features

Examples

pf_flow_sum(rpfc(100))
#>        t7       t87       t89       t23       t48       t35       t99       t53 
#> 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 
#>       t84       t36       t40       t79       t86       t60       t10       t38 
#> 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 
#>       t82       t97       t67       t85       t98       t44       t17       t55 
#> 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 
#>       t20       t16       t63       t33       t69       t59       t43       t29 
#> 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 
#>       t21        t8       t22        t9       t91       t19        t5       t46 
#> 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 
#>       t47       t50       t70       t15       t39       t58       t31       t26 
#> 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 
#>       t81       t42       t83       t64       t61       t75       t80       t56 
#> 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 
#>       t93       t57       t34       t73       t32       t78       t92       t49 
#> 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 
#>       t68       t96       t76       t18        t6        t2       t65        t1 
#> 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 
#>       t41       t25       t27       t54       t12       t88       t51       t77 
#> 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 
#>       t62       t24       t71       t66       t13       t11       t74       t52 
#> 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 
#>       t14       t94       t45       t37      t100        t3       t28       t72 
#> 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 0.6413932 
#>       t90       t95       t30        t4     Node2     Node3     Node4     Node5 
#> 0.6413932 0.6413932 0.6413932 0.6413932 0.1178857 0.1577854 0.3084475 0.3292001 
#>     Node6     Node7     Node8     Node9    Node10    Node11    Node12    Node13 
#> 0.3756970 0.4028752 0.4436663 0.4712427 0.4782531 0.4830687 0.4891443 0.4906786 
#>    Node14    Node15    Node16    Node17    Node18    Node19    Node20    Node21 
#> 0.5148156 0.5185775 0.5296025 0.5409986 0.5429491 0.5448857 0.5450705 0.5546837 
#>    Node22    Node23    Node24    Node25    Node26    Node27    Node28    Node29 
#> 0.5639732 0.5767201 0.5785741 0.5850435 0.5882910 0.5921700 0.5928329 0.5934151 
#>    Node30    Node31    Node32    Node33    Node34    Node35    Node36    Node37 
#> 0.5948321 0.5949641 0.5963728 0.5976659 0.5980654 0.6002706 0.6050333 0.6073354 
#>    Node38    Node39    Node40    Node41    Node42    Node43    Node44    Node45 
#> 0.6073970 0.6083626 0.6093604 0.6094774 0.6122316 0.6125686 0.6145755 0.6147111 
#>    Node46    Node47    Node48    Node49    Node50    Node51    Node52    Node53 
#> 0.6184279 0.6188376 0.6203555 0.6203903 0.6210996 0.6223796 0.6229844 0.6230661 
#>    Node54    Node55    Node56    Node57    Node58    Node59    Node60    Node61 
#> 0.6234352 0.6237730 0.6243416 0.6248574 0.6249174 0.6249948 0.6258931 0.6262070 
#>    Node62    Node63    Node64    Node65    Node66    Node67    Node68    Node69 
#> 0.6266587 0.6267670 0.6274339 0.6276127 0.6277369 0.6280852 0.6282299 0.6295780 
#>    Node70    Node71    Node72    Node73    Node74    Node75    Node76    Node77 
#> 0.6303358 0.6309633 0.6325142 0.6327441 0.6340011 0.6341419 0.6344354 0.6347692 
#>    Node78    Node79    Node80    Node81    Node82    Node83    Node84    Node85 
#> 0.6349984 0.6353727 0.6366001 0.6368262 0.6368614 0.6368870 0.6371116 0.6372397 
#>    Node86    Node87    Node88    Node89    Node90    Node91    Node92    Node93 
#> 0.6374019 0.6375201 0.6383835 0.6384402 0.6385621 0.6389371 0.6391800 0.6394831 
#>    Node94    Node95    Node96    Node97    Node98    Node99 
#> 0.6395743 0.6397179 0.6401824 0.6405159 0.6410532 0.6412418