Replace feature with edge index in a pfc

This is similar to base R's col() function, it returns a pfc with features that are just the index of the edge. Useful for complex indexing procedures

Usage

pf_indexes(x)

Arguments

x

A pfc object

Value

A new pfc with the same structure as x, but with the features of each edge equal to the edge index (the column number in the sparse matrix representation)

Examples

pf_indexes(rpfc(100))
#> <pfc<e:198>[198]>
#> First  10 phylogenetic flows: 
#> [1] ◎──    2──→ Node3 ──    4──→ Node21 ──    6──→ t28 
#> 
#> [2] ◎──    2──→ Node3 ──    4──→ Node21 ──    5──→ Node47 ──    8──→ Node91 ──   10──→ t58 
#> 
#> [3] ◎──    2──→ Node3 ──    4──→ Node21 ──    5──→ Node47 ──    8──→ Node91 ──    9──→ t24 
#> 
#> [4] ◎──    2──→ Node3 ──    4──→ Node21 ──    5──→ Node47 ──    7──→ t41 
#> 
#> [5] ◎──    2──→ Node3 ──    3──→ Node30 ──   12──→ Node35 ──   14──→ Node87 ──   16──→ t82 
#> 
#> [6] ◎──    2──→ Node3 ──    3──→ Node30 ──   12──→ Node35 ──   14──→ Node87 ──   15──→ t10 
#> 
#> [7] ◎──    2──→ Node3 ──    3──→ Node30 ──   12──→ Node35 ──   13──→ Node51 ──   18──→ Node86 ──   20──→ t70 
#> 
#> [8] ◎──    2──→ Node3 ──    3──→ Node30 ──   12──→ Node35 ──   13──→ Node51 ──   18──→ Node86 ──   19──→ t36 
#> 
#> [9] ◎──    2──→ Node3 ──    3──→ Node30 ──   12──→ Node35 ──   13──→ Node51 ──   17──→ t77 
#> 
#> [10] ◎──    2──→ Node3 ──    3──→ Node30 ──   11──→ Node41 ──   22──→ t99