Calculate a kronecker product when a pfc is the multiplicand or the multiplier.

Calculate a kronecker product when a pfc is the multiplicand or the multiplier.

Usage

pf_kronecker(x, y, ...)

Arguments

x

The kronecker multiplicand.

y

The kronecker multiplier.

...

Other arguments passed to or from other methods.

Value

A pfc object

Examples

pf_kronecker(rpfc(20), rpfc(20))
#> <pfc<e:1444>[1444]>
#> First  10 phylogenetic flows: 
#> [1] ◎──5.666──→ Node6:Node2 ──1.575──→ Node6:Node8 ──0.637──→ Node6:t16 ──0.332──→ Node10:Node2 ──0.092──→ Node10:Node8 ──0.037──→ Node10:t16 ──0.108──→ t12:Node2 ──0.030──→ t12:Node8 ──0.012──→ t12:t16 
#> 
#> [2] ◎──5.66572──→ Node6:Node2 ──1.57500──→ Node6:Node8 ──0.63352──→ Node6:Node18 ──0.00315──→ Node6:t9 ──0.33208──→ Node10:Node2 ──0.09231──→ Node10:Node8 ──0.03713──→ Node10:Node18 ──0.00018──→ Node10:t9 ──0.10759──→ t12:Node2 ──0.02991──→ t12:Node8 ──0.01203──→ t12:Node18 ──0.00006──→ t12:t9 
#> 
#> [3] ◎──5.66572──→ Node6:Node2 ──1.57500──→ Node6:Node8 ──0.63352──→ Node6:Node18 ──0.00315──→ Node6:t4 ──0.33208──→ Node10:Node2 ──0.09231──→ Node10:Node8 ──0.03713──→ Node10:Node18 ──0.00018──→ Node10:t4 ──0.10759──→ t12:Node2 ──0.02991──→ t12:Node8 ──0.01203──→ t12:Node18 ──0.00006──→ t12:t4 
#> 
#> [4] ◎──5.66572──→ Node6:Node2 ──0.57008──→ Node6:Node3 ──0.01426──→ Node6:Node4 ──0.57867──→ Node6:Node6 ──1.04866──→ Node6:t1 ──0.33208──→ Node10:Node2 ──0.03341──→ Node10:Node3 ──0.00084──→ Node10:Node4 ──0.03392──→ Node10:Node6 ──0.06146──→ Node10:t1 ──0.10759──→ t12:Node2 ──0.01083──→ t12:Node3 ──0.00027──→ t12:Node4 ──0.01099──→ t12:Node6 ──0.01991──→ t12:t1 
#> 
#> [5] ◎──5.66572──→ Node6:Node2 ──0.57008──→ Node6:Node3 ──0.01426──→ Node6:Node4 ──0.57867──→ Node6:Node6 ──1.04866──→ Node6:t15 ──0.33208──→ Node10:Node2 ──0.03341──→ Node10:Node3 ──0.00084──→ Node10:Node4 ──0.03392──→ Node10:Node6 ──0.06146──→ Node10:t15 ──0.10759──→ t12:Node2 ──0.01083──→ t12:Node3 ──0.00027──→ t12:Node4 ──0.01099──→ t12:Node6 ──0.01991──→ t12:t15 
#> 
#> [6] ◎──5.66572──→ Node6:Node2 ──0.57008──→ Node6:Node3 ──0.01426──→ Node6:Node4 ──0.85658──→ Node6:Node7 ──0.44099──→ Node6:Node11 ──0.21676──→ Node6:Node15 ──0.04874──→ Node6:Node17 ──0.06426──→ Node6:t19 ──0.33208──→ Node10:Node2 ──0.03341──→ Node10:Node3 ──0.00084──→ Node10:Node4 ──0.05021──→ Node10:Node7 ──0.02585──→ Node10:Node11 ──0.01270──→ Node10:Node15 ──0.00286──→ Node10:Node17 ──0.00377──→ Node10:t19 ──0.10759──→ t12:Node2 ──0.01083──→ t12:Node3 ──0.00027──→ t12:Node4 ──0.01627──→ t12:Node7 ──0.00837──→ t12:Node11 ──0.00412──→ t12:Node15 ──0.00093──→ t12:Node17 ──0.00122──→ t12:t19 
#> 
#> [7] ◎──5.66572──→ Node6:Node2 ──0.57008──→ Node6:Node3 ──0.01426──→ Node6:Node4 ──0.85658──→ Node6:Node7 ──0.44099──→ Node6:Node11 ──0.21676──→ Node6:Node15 ──0.04874──→ Node6:Node17 ──0.06426──→ Node6:t7 ──0.33208──→ Node10:Node2 ──0.03341──→ Node10:Node3 ──0.00084──→ Node10:Node4 ──0.05021──→ Node10:Node7 ──0.02585──→ Node10:Node11 ──0.01270──→ Node10:Node15 ──0.00286──→ Node10:Node17 ──0.00377──→ Node10:t7 ──0.10759──→ t12:Node2 ──0.01083──→ t12:Node3 ──0.00027──→ t12:Node4 ──0.01627──→ t12:Node7 ──0.00837──→ t12:Node11 ──0.00412──→ t12:Node15 ──0.00093──→ t12:Node17 ──0.00122──→ t12:t7 
#> 
#> [8] ◎──5.66572──→ Node6:Node2 ──0.57008──→ Node6:Node3 ──0.01426──→ Node6:Node4 ──0.85658──→ Node6:Node7 ──0.44099──→ Node6:Node11 ──0.21676──→ Node6:Node15 ──0.11301──→ Node6:t14 ──0.33208──→ Node10:Node2 ──0.03341──→ Node10:Node3 ──0.00084──→ Node10:Node4 ──0.05021──→ Node10:Node7 ──0.02585──→ Node10:Node11 ──0.01270──→ Node10:Node15 ──0.00662──→ Node10:t14 ──0.10759──→ t12:Node2 ──0.01083──→ t12:Node3 ──0.00027──→ t12:Node4 ──0.01627──→ t12:Node7 ──0.00837──→ t12:Node11 ──0.00412──→ t12:Node15 ──0.00215──→ t12:t14 
#> 
#> [9] ◎──5.66572──→ Node6:Node2 ──0.57008──→ Node6:Node3 ──0.01426──→ Node6:Node4 ──0.85658──→ Node6:Node7 ──0.44099──→ Node6:Node11 ──0.03185──→ Node6:Node12 ──0.29791──→ Node6:t3 ──0.33208──→ Node10:Node2 ──0.03341──→ Node10:Node3 ──0.00084──→ Node10:Node4 ──0.05021──→ Node10:Node7 ──0.02585──→ Node10:Node11 ──0.00187──→ Node10:Node12 ──0.01746──→ Node10:t3 ──0.10759──→ t12:Node2 ──0.01083──→ t12:Node3 ──0.00027──→ t12:Node4 ──0.01627──→ t12:Node7 ──0.00837──→ t12:Node11 ──0.00060──→ t12:Node12 ──0.00566──→ t12:t3 
#> 
#> [10] ◎──5.66572──→ Node6:Node2 ──0.57008──→ Node6:Node3 ──0.01426──→ Node6:Node4 ──0.85658──→ Node6:Node7 ──0.44099──→ Node6:Node11 ──0.03185──→ Node6:Node12 ──0.29791──→ Node6:t6 ──0.33208──→ Node10:Node2 ──0.03341──→ Node10:Node3 ──0.00084──→ Node10:Node4 ──0.05021──→ Node10:Node7 ──0.02585──→ Node10:Node11 ──0.00187──→ Node10:Node12 ──0.01746──→ Node10:t6 ──0.10759──→ t12:Node2 ──0.01083──→ t12:Node3 ──0.00027──→ t12:Node4 ──0.01627──→ t12:Node7 ──0.00837──→ t12:Node11 ──0.00060──→ t12:Node12 ──0.00566──→ t12:t6