Calculate a kronecker product when a pfc is the multiplicand or the multiplier.

Calculate a kronecker product when a pfc is the multiplicand or the multiplier.

Usage

pf_kronecker(x, y, ...)

Arguments

x

The kronecker multiplicand.

y

The kronecker multiplier.

...

Other arguments passed to or from other methods.

Value

A pfc object

Examples

pf_kronecker(rpfc(20), rpfc(20))
#> <pfc<e:1444>[1444]>
#> First  10 phylogenetic flows: 
#> [1] ◎──1.3780──→ Node5:Node6 ──0.5163──→ Node5:Node11 ──0.0892──→ Node5:t10 ──0.0680──→ Node8:Node6 ──0.0255──→ Node8:Node11 ──0.0044──→ Node8:t10 ──0.0406──→ t4:Node6 ──0.0152──→ t4:Node11 ──0.0026──→ t4:t10 
#> 
#> [2] ◎──1.37801──→ Node5:Node6 ──0.51635──→ Node5:Node11 ──0.03277──→ Node5:Node14 ──0.05647──→ Node5:t8 ──0.06796──→ Node8:Node6 ──0.02547──→ Node8:Node11 ──0.00162──→ Node8:Node14 ──0.00279──→ Node8:t8 ──0.04060──→ t4:Node6 ──0.01521──→ t4:Node11 ──0.00097──→ t4:Node14 ──0.00166──→ t4:t8 
#> 
#> [3] ◎──1.37801──→ Node5:Node6 ──0.51635──→ Node5:Node11 ──0.03277──→ Node5:Node14 ──0.05647──→ Node5:t3 ──0.06796──→ Node8:Node6 ──0.02547──→ Node8:Node11 ──0.00162──→ Node8:Node14 ──0.00279──→ Node8:t3 ──0.04060──→ t4:Node6 ──0.01521──→ t4:Node11 ──0.00097──→ t4:Node14 ──0.00166──→ t4:t3 
#> 
#> [4] ◎──1.3780──→ Node5:Node6 ──0.2838──→ Node5:Node9 ──0.3218──→ Node5:t17 ──0.0680──→ Node8:Node6 ──0.0140──→ Node8:Node9 ──0.0159──→ Node8:t17 ──0.0406──→ t4:Node6 ──0.0084──→ t4:Node9 ──0.0095──→ t4:t17 
#> 
#> [5] ◎──1.3780──→ Node5:Node6 ──0.2838──→ Node5:Node9 ──0.3218──→ Node5:t12 ──0.0680──→ Node8:Node6 ──0.0140──→ Node8:Node9 ──0.0159──→ Node8:t12 ──0.0406──→ t4:Node6 ──0.0084──→ t4:Node9 ──0.0095──→ t4:t12 
#> 
#> [6] ◎──0.77376──→ Node5:Node2 ──0.01690──→ Node5:Node3 ──0.43745──→ Node5:Node5 ──0.75548──→ Node5:t6 ──0.03816──→ Node8:Node2 ──0.00083──→ Node8:Node3 ──0.02158──→ Node8:Node5 ──0.03726──→ Node8:t6 ──0.02280──→ t4:Node2 ──0.00050──→ t4:Node3 ──0.01289──→ t4:Node5 ──0.02226──→ t4:t6 
#> 
#> [7] ◎──0.77376──→ Node5:Node2 ──0.01690──→ Node5:Node3 ──0.43745──→ Node5:Node5 ──0.34411──→ Node5:Node7 ──0.08069──→ Node5:Node8 ──0.33067──→ Node5:t4 ──0.03816──→ Node8:Node2 ──0.00083──→ Node8:Node3 ──0.02158──→ Node8:Node5 ──0.01697──→ Node8:Node7 ──0.00398──→ Node8:Node8 ──0.01631──→ Node8:t4 ──0.02280──→ t4:Node2 ──0.00050──→ t4:Node3 ──0.01289──→ t4:Node5 ──0.01014──→ t4:Node7 ──0.00238──→ t4:Node8 ──0.00974──→ t4:t4 
#> 
#> [8] ◎──0.77376──→ Node5:Node2 ──0.01690──→ Node5:Node3 ──0.43745──→ Node5:Node5 ──0.34411──→ Node5:Node7 ──0.08069──→ Node5:Node8 ──0.33067──→ Node5:t1 ──0.03816──→ Node8:Node2 ──0.00083──→ Node8:Node3 ──0.02158──→ Node8:Node5 ──0.01697──→ Node8:Node7 ──0.00398──→ Node8:Node8 ──0.01631──→ Node8:t1 ──0.02280──→ t4:Node2 ──0.00050──→ t4:Node3 ──0.01289──→ t4:Node5 ──0.01014──→ t4:Node7 ──0.00238──→ t4:Node8 ──0.00974──→ t4:t1 
#> 
#> [9] ◎──0.77376──→ Node5:Node2 ──0.01690──→ Node5:Node3 ──0.43745──→ Node5:Node5 ──0.34411──→ Node5:Node7 ──0.41137──→ Node5:t11 ──0.03816──→ Node8:Node2 ──0.00083──→ Node8:Node3 ──0.02158──→ Node8:Node5 ──0.01697──→ Node8:Node7 ──0.02029──→ Node8:t11 ──0.02280──→ t4:Node2 ──0.00050──→ t4:Node3 ──0.01289──→ t4:Node5 ──0.01014──→ t4:Node7 ──0.01212──→ t4:t11 
#> 
#> [10] ◎──0.77376──→ Node5:Node2 ──0.01690──→ Node5:Node3 ──0.28684──→ Node5:Node4 ──0.82530──→ Node5:Node12 ──0.03257──→ Node5:Node16 ──0.04823──→ Node5:t18 ──0.03816──→ Node8:Node2 ──0.00083──→ Node8:Node3 ──0.01415──→ Node8:Node4 ──0.04070──→ Node8:Node12 ──0.00161──→ Node8:Node16 ──0.00238──→ Node8:t18 ──0.02280──→ t4:Node2 ──0.00050──→ t4:Node3 ──0.00845──→ t4:Node4 ──0.02432──→ t4:Node12 ──0.00096──→ t4:Node16 ──0.00142──→ t4:t18