Calculate a kronecker product when a pfc is the multiplicand or the multiplier.

Calculate a kronecker product when a pfc is the multiplicand or the multiplier.

Usage

pf_kronecker(x, y, ...)

Arguments

x

The kronecker multiplicand.

y

The kronecker multiplier.

...

Other arguments passed to or from other methods.

Value

A pfc object

Examples

pf_kronecker(rpfc(20), rpfc(20))
#> <pfc<e:1444>[1444]>
#> First  10 phylogenetic flows: 
#> [1] ◎──7.9e-01──→ Node2:Node2 ──2.5e-01──→ Node2:Node4 ──9.7e-04──→ Node2:Node5 ──1.5e-01──→ Node2:Node15 ──2.7e-02──→ Node2:t8 ──1.8e-02──→ Node3:Node2 ──5.7e-03──→ Node3:Node4 ──2.2e-05──→ Node3:Node5 ──3.4e-03──→ Node3:Node15 ──6.0e-04──→ Node3:t8 ──4.1e-01──→ Node7:Node2 ──1.3e-01──→ Node7:Node4 ──5.0e-04──→ Node7:Node5 ──7.8e-02──→ Node7:Node15 ──1.4e-02──→ Node7:t8 ──7.8e-02──→ Node9:Node2 ──2.5e-02──→ Node9:Node4 ──9.5e-05──→ Node9:Node5 ──1.5e-02──→ Node9:Node15 ──2.6e-03──→ Node9:t8 ──9.2e-02──→ t17:Node2 ──2.9e-02──→ t17:Node4 ──1.1e-04──→ t17:Node5 ──1.8e-02──→ t17:Node15 ──3.1e-03──→ t17:t8 
#> 
#> [2] ◎──7.9e-01──→ Node2:Node2 ──2.5e-01──→ Node2:Node4 ──9.7e-04──→ Node2:Node5 ──1.5e-01──→ Node2:Node15 ──2.7e-02──→ Node2:t15 ──1.8e-02──→ Node3:Node2 ──5.7e-03──→ Node3:Node4 ──2.2e-05──→ Node3:Node5 ──3.4e-03──→ Node3:Node15 ──6.0e-04──→ Node3:t15 ──4.1e-01──→ Node7:Node2 ──1.3e-01──→ Node7:Node4 ──5.0e-04──→ Node7:Node5 ──7.8e-02──→ Node7:Node15 ──1.4e-02──→ Node7:t15 ──7.8e-02──→ Node9:Node2 ──2.5e-02──→ Node9:Node4 ──9.5e-05──→ Node9:Node5 ──1.5e-02──→ Node9:Node15 ──2.6e-03──→ Node9:t15 ──9.2e-02──→ t17:Node2 ──2.9e-02──→ t17:Node4 ──1.1e-04──→ t17:Node5 ──1.8e-02──→ t17:Node15 ──3.1e-03──→ t17:t15 
#> 
#> [3] ◎──7.9e-01──→ Node2:Node2 ──2.5e-01──→ Node2:Node4 ──9.7e-04──→ Node2:Node5 ──1.3e-01──→ Node2:Node12 ──4.5e-02──→ Node2:t2 ──1.8e-02──→ Node3:Node2 ──5.7e-03──→ Node3:Node4 ──2.2e-05──→ Node3:Node5 ──3.0e-03──→ Node3:Node12 ──1.0e-03──→ Node3:t2 ──4.1e-01──→ Node7:Node2 ──1.3e-01──→ Node7:Node4 ──5.0e-04──→ Node7:Node5 ──6.9e-02──→ Node7:Node12 ──2.3e-02──→ Node7:t2 ──7.8e-02──→ Node9:Node2 ──2.5e-02──→ Node9:Node4 ──9.5e-05──→ Node9:Node5 ──1.3e-02──→ Node9:Node12 ──4.4e-03──→ Node9:t2 ──9.2e-02──→ t17:Node2 ──2.9e-02──→ t17:Node4 ──1.1e-04──→ t17:Node5 ──1.5e-02──→ t17:Node12 ──5.2e-03──→ t17:t2 
#> 
#> [4] ◎──7.9e-01──→ Node2:Node2 ──2.5e-01──→ Node2:Node4 ──9.7e-04──→ Node2:Node5 ──1.3e-01──→ Node2:Node12 ──4.5e-02──→ Node2:t11 ──1.8e-02──→ Node3:Node2 ──5.7e-03──→ Node3:Node4 ──2.2e-05──→ Node3:Node5 ──3.0e-03──→ Node3:Node12 ──1.0e-03──→ Node3:t11 ──4.1e-01──→ Node7:Node2 ──1.3e-01──→ Node7:Node4 ──5.0e-04──→ Node7:Node5 ──6.9e-02──→ Node7:Node12 ──2.3e-02──→ Node7:t11 ──7.8e-02──→ Node9:Node2 ──2.5e-02──→ Node9:Node4 ──9.5e-05──→ Node9:Node5 ──1.3e-02──→ Node9:Node12 ──4.4e-03──→ Node9:t11 ──9.2e-02──→ t17:Node2 ──2.9e-02──→ t17:Node4 ──1.1e-04──→ t17:Node5 ──1.5e-02──→ t17:Node12 ──5.2e-03──→ t17:t11 
#> 
#> [5] ◎──7.9e-01──→ Node2:Node2 ──2.5e-01──→ Node2:Node4 ──1.2e-01──→ Node2:Node10 ──1.1e-02──→ Node2:Node13 ──1.2e-03──→ Node2:Node14 ──4.3e-02──→ Node2:t3 ──1.8e-02──→ Node3:Node2 ──5.7e-03──→ Node3:Node4 ──2.8e-03──→ Node3:Node10 ──2.5e-04──→ Node3:Node13 ──2.7e-05──→ Node3:Node14 ──9.6e-04──→ Node3:t3 ──4.1e-01──→ Node7:Node2 ──1.3e-01──→ Node7:Node4 ──6.4e-02──→ Node7:Node10 ──5.8e-03──→ Node7:Node13 ──6.1e-04──→ Node7:Node14 ──2.2e-02──→ Node7:t3 ──7.8e-02──→ Node9:Node2 ──2.5e-02──→ Node9:Node4 ──1.2e-02──→ Node9:Node10 ──1.1e-03──→ Node9:Node13 ──1.2e-04──→ Node9:Node14 ──4.2e-03──→ Node9:t3 ──9.2e-02──→ t17:Node2 ──2.9e-02──→ t17:Node4 ──1.4e-02──→ t17:Node10 ──1.3e-03──→ t17:Node13 ──1.4e-04──→ t17:Node14 ──4.9e-03──→ t17:t3 
#> 
#> [6] ◎──7.9e-01──→ Node2:Node2 ──2.5e-01──→ Node2:Node4 ──1.2e-01──→ Node2:Node10 ──1.1e-02──→ Node2:Node13 ──1.2e-03──→ Node2:Node14 ──4.3e-02──→ Node2:t19 ──1.8e-02──→ Node3:Node2 ──5.7e-03──→ Node3:Node4 ──2.8e-03──→ Node3:Node10 ──2.5e-04──→ Node3:Node13 ──2.7e-05──→ Node3:Node14 ──9.6e-04──→ Node3:t19 ──4.1e-01──→ Node7:Node2 ──1.3e-01──→ Node7:Node4 ──6.4e-02──→ Node7:Node10 ──5.8e-03──→ Node7:Node13 ──6.1e-04──→ Node7:Node14 ──2.2e-02──→ Node7:t19 ──7.8e-02──→ Node9:Node2 ──2.5e-02──→ Node9:Node4 ──1.2e-02──→ Node9:Node10 ──1.1e-03──→ Node9:Node13 ──1.2e-04──→ Node9:Node14 ──4.2e-03──→ Node9:t19 ──9.2e-02──→ t17:Node2 ──2.9e-02──→ t17:Node4 ──1.4e-02──→ t17:Node10 ──1.3e-03──→ t17:Node13 ──1.4e-04──→ t17:Node14 ──4.9e-03──→ t17:t19 
#> 
#> [7] ◎──0.79409──→ Node2:Node2 ──0.25168──→ Node2:Node4 ──0.12465──→ Node2:Node10 ──0.01124──→ Node2:Node13 ──0.04400──→ Node2:t13 ──0.01785──→ Node3:Node2 ──0.00566──→ Node3:Node4 ──0.00280──→ Node3:Node10 ──0.00025──→ Node3:Node13 ──0.00099──→ Node3:t13 ──0.40879──→ Node7:Node2 ──0.12956──→ Node7:Node4 ──0.06417──→ Node7:Node10 ──0.00578──→ Node7:Node13 ──0.02265──→ Node7:t13 ──0.07761──→ Node9:Node2 ──0.02460──→ Node9:Node4 ──0.01218──→ Node9:Node10 ──0.00110──→ Node9:Node13 ──0.00430──→ Node9:t13 ──0.09162──→ t17:Node2 ──0.02904──→ t17:Node4 ──0.01438──→ t17:Node10 ──0.00130──→ t17:Node13 ──0.00508──→ t17:t13 
#> 
#> [8] ◎──0.7941──→ Node2:Node2 ──0.2517──→ Node2:Node4 ──0.1246──→ Node2:Node10 ──0.0552──→ Node2:t18 ──0.0179──→ Node3:Node2 ──0.0057──→ Node3:Node4 ──0.0028──→ Node3:Node10 ──0.0012──→ Node3:t18 ──0.4088──→ Node7:Node2 ──0.1296──→ Node7:Node4 ──0.0642──→ Node7:Node10 ──0.0284──→ Node7:t18 ──0.0776──→ Node9:Node2 ──0.0246──→ Node9:Node4 ──0.0122──→ Node9:Node10 ──0.0054──→ Node9:t18 ──0.0916──→ t17:Node2 ──0.0290──→ t17:Node4 ──0.0144──→ t17:Node10 ──0.0064──→ t17:t18 
#> 
#> [9] ◎──0.7941──→ Node2:Node2 ──0.4316──→ Node2:t9 ──0.0179──→ Node3:Node2 ──0.0097──→ Node3:t9 ──0.4088──→ Node7:Node2 ──0.2222──→ Node7:t9 ──0.0776──→ Node9:Node2 ──0.0422──→ Node9:t9 ──0.0916──→ t17:Node2 ──0.0498──→ t17:t9 
#> 
#> [10] ◎──0.99736──→ Node2:Node3 ──0.05468──→ Node2:Node6 ──0.00599──→ Node2:Node7 ──0.08785──→ Node2:Node8 ──0.07977──→ Node2:t5 ──0.02243──→ Node3:Node3 ──0.00123──→ Node3:Node6 ──0.00013──→ Node3:Node7 ──0.00198──→ Node3:Node8 ──0.00179──→ Node3:t5 ──0.51342──→ Node7:Node3 ──0.02815──→ Node7:Node6 ──0.00308──→ Node7:Node7 ──0.04523──→ Node7:Node8 ──0.04106──→ Node7:t5 ──0.09748──→ Node9:Node3 ──0.00534──→ Node9:Node6 ──0.00059──→ Node9:Node7 ──0.00859──→ Node9:Node8 ──0.00780──→ Node9:t5 ──0.11507──→ t17:Node3 ──0.00631──→ t17:Node6 ──0.00069──→ t17:Node7 ──0.01014──→ t17:Node8 ──0.00920──→ t17:t5