Extract pfp object from pfc, the paths of each flow from root to terminal node.

Extract pfp object from pfc, the paths of each flow from root to terminal node.

Usage

pf_path(x)

Arguments

x

A pfc object

Value

A pfp object

Examples

pf_path(rpfc(100))
#> <pfp[198]>
#>   [1] 2->4->6->8                                               
#>   [2] 2->4->6->7                                               
#>   [3] 2->4->5->10                                              
#>   [4] 2->4->5->9->12->14                                       
#>   [5] 2->4->5->9->12->13                                       
#>   [6] 2->4->5->9->11                                           
#>   [7] 2->3->16->18->20->22->24->26                             
#>   [8] 2->3->16->18->20->22->24->25                             
#>   [9] 2->3->16->18->20->22->23->28                             
#>  [10] 2->3->16->18->20->22->23->27->30                         
#>  [11] 2->3->16->18->20->22->23->27->29->32                     
#>  [12] 2->3->16->18->20->22->23->27->29->31                     
#>  [13] 2->3->16->18->20->21->34->36->38                         
#>  [14] 2->3->16->18->20->21->34->36->37                         
#>  [15] 2->3->16->18->20->21->34->35                             
#>  [16] 2->3->16->18->20->21->33->40->42->44                     
#>  [17] 2->3->16->18->20->21->33->40->42->43                     
#>  [18] 2->3->16->18->20->21->33->40->41->46->48->50             
#>  [19] 2->3->16->18->20->21->33->40->41->46->48->49             
#>  [20] 2->3->16->18->20->21->33->40->41->46->47                 
#>  [21] 2->3->16->18->20->21->33->40->41->45->52                 
#>  [22] 2->3->16->18->20->21->33->40->41->45->51                 
#>  [23] 2->3->16->18->20->21->33->39->54->56                     
#>  [24] 2->3->16->18->20->21->33->39->54->55                     
#>  [25] 2->3->16->18->20->21->33->39->53->58                     
#>  [26] 2->3->16->18->20->21->33->39->53->57->60                 
#>  [27] 2->3->16->18->20->21->33->39->53->57->59                 
#>  [28] 2->3->16->18->19->62->64                                 
#>  [29] 2->3->16->18->19->62->63                                 
#>  [30] 2->3->16->18->19->61->66->68->70->72                     
#>  [31] 2->3->16->18->19->61->66->68->70->71                     
#>  [32] 2->3->16->18->19->61->66->68->69                         
#>  [33] 2->3->16->18->19->61->66->67                             
#>  [34] 2->3->16->18->19->61->65->74                             
#>  [35] 2->3->16->18->19->61->65->73                             
#>  [36] 2->3->16->17->76                                         
#>  [37] 2->3->16->17->75                                         
#>  [38] 2->3->15->78->80                                         
#>  [39] 2->3->15->78->79                                         
#>  [40] 2->3->15->77                                             
#>  [41] 1->82->84->86->88->90->92                                
#>  [42] 1->82->84->86->88->90->91->94->96->98                    
#>  [43] 1->82->84->86->88->90->91->94->96->97->100               
#>  [44] 1->82->84->86->88->90->91->94->96->97->99                
#>  [45] 1->82->84->86->88->90->91->94->95                        
#>  [46] 1->82->84->86->88->90->91->93->102                       
#>  [47] 1->82->84->86->88->90->91->93->101                       
#>  [48] 1->82->84->86->88->89->104->106                          
#>  [49] 1->82->84->86->88->89->104->105                          
#>  [50] 1->82->84->86->88->89->103->108                          
#>  [51] 1->82->84->86->88->89->103->107                          
#>  [52] 1->82->84->86->87->110->112->114->116                    
#>  [53] 1->82->84->86->87->110->112->114->115                    
#>  [54] 1->82->84->86->87->110->112->113                         
#>  [55] 1->82->84->86->87->110->111->118->120                    
#>  [56] 1->82->84->86->87->110->111->118->119->122->124          
#>  [57] 1->82->84->86->87->110->111->118->119->122->123          
#>  [58] 1->82->84->86->87->110->111->118->119->121->126          
#>  [59] 1->82->84->86->87->110->111->118->119->121->125          
#>  [60] 1->82->84->86->87->110->111->117->128->130->132->134->136
#>  [61] 1->82->84->86->87->110->111->117->128->130->132->134->135
#>  [62] 1->82->84->86->87->110->111->117->128->130->132->133     
#>  [63] 1->82->84->86->87->110->111->117->128->130->131          
#>  [64] 1->82->84->86->87->110->111->117->128->129               
#>  [65] 1->82->84->86->87->110->111->117->127->138->140->142     
#>  [66] 1->82->84->86->87->110->111->117->127->138->140->141     
#>  [67] 1->82->84->86->87->110->111->117->127->138->139          
#>  [68] 1->82->84->86->87->110->111->117->127->137->144          
#>  [69] 1->82->84->86->87->110->111->117->127->137->143          
#>  [70] 1->82->84->86->87->109->146                              
#>  [71] 1->82->84->86->87->109->145                              
#>  [72] 1->82->84->85->148                                       
#>  [73] 1->82->84->85->147                                       
#>  [74] 1->82->83->150->152->154                                 
#>  [75] 1->82->83->150->152->153                                 
#>  [76] 1->82->83->150->151                                      
#>  [77] 1->82->83->149->156                                      
#>  [78] 1->82->83->149->155                                      
#>  [79] 1->81->158->160->162->164                                
#>  [80] 1->81->158->160->162->163->166->168->170                 
#>  [81] 1->81->158->160->162->163->166->168->169                 
#>  [82] 1->81->158->160->162->163->166->167                      
#>  [83] 1->81->158->160->162->163->165                           
#>  [84] 1->81->158->160->161->172->174->176->178                 
#>  [85] 1->81->158->160->161->172->174->176->177                 
#>  [86] 1->81->158->160->161->172->174->175                      
#>  [87] 1->81->158->160->161->172->173                           
#>  [88] 1->81->158->160->161->171                                
#>  [89] 1->81->158->159->180                                     
#>  [90] 1->81->158->159->179                                     
#>  [91] 1->81->157->182->184                                     
#>  [92] 1->81->157->182->183->186                                
#>  [93] 1->81->157->182->183->185                                
#>  [94] 1->81->157->181->188                                     
#>  [95] 1->81->157->181->187->190->192                           
#>  [96] 1->81->157->181->187->190->191                           
#>  [97] 1->81->157->181->187->189->194                           
#>  [98] 1->81->157->181->187->189->193->196                      
#>  [99] 1->81->157->181->187->189->193->195->198                 
#> [100] 1->81->157->181->187->189->193->195->197                 
#> [101] 1                                                        
#> [102] 2                                                        
#> [103] 1->82                                                    
#> [104] 1->82->84                                                
#> [105] 2->3                                                     
#> [106] 1->82->84->86                                            
#> [107] 2->3->16                                                 
#> [108] 2->3->16->18                                             
#> [109] 1->81                                                    
#> [110] 2->3->16->18->20                                         
#> [111] 1->82->84->85                                            
#> [112] 1->82->84->86->87                                        
#> [113] 2->4                                                     
#> [114] 2->3->16->18->20->21                                     
#> [115] 1->82->84->86->88                                        
#> [116] 2->4->5                                                  
#> [117] 1->81->158                                               
#> [118] 2->4->5->9                                               
#> [119] 1->82->84->86->87->110                                   
#> [120] 2->3->16->18->19                                         
#> [121] 2->3->16->18->20->21->33                                 
#> [122] 1->81->158->160                                          
#> [123] 1->82->84->86->87->110->111                              
#> [124] 1->82->84->86->88->90                                    
#> [125] 1->81->157                                               
#> [126] 1->82->84->86->87->110->111->117                         
#> [127] 2->3->15                                                 
#> [128] 1->82->84->86->88->90->91                                
#> [129] 1->82->84->86->87->110->111->118                         
#> [130] 1->82->84->86->88->89                                    
#> [131] 1->81->158->160->161                                     
#> [132] 2->3->16->18->19->61                                     
#> [133] 2->3->16->18->20->21->33->39                             
#> [134] 2->3->16->18->19->61->66                                 
#> [135] 1->81->157->181                                          
#> [136] 1->82->84->86->87->110->111->118->119                    
#> [137] 1->81->157->182                                          
#> [138] 2->3->16->17                                             
#> [139] 2->3->16->18->20->21->34                                 
#> [140] 1->82->84->86->87->110->111->118->119->122               
#> [141] 1->82->84->86->88->89->103                               
#> [142] 2->3->16->18->20->22                                     
#> [143] 1->82->84->86->87->110->111->117->127                    
#> [144] 2->3->16->18->20->21->33->40                             
#> [145] 1->81->157->181->187                                     
#> [146] 1->82->84->86->87->110->111->117->128                    
#> [147] 2->3->16->18->20->22->23                                 
#> [148] 2->3->16->18->20->22->23->27                             
#> [149] 1->81->158->160->162                                     
#> [150] 1->82->84->86->87->110->111->117->128->130               
#> [151] 2->3->15->78                                             
#> [152] 2->3->16->18->19->61->66->68                             
#> [153] 1->82->83                                                
#> [154] 2->3->16->18->20->21->33->40->42                         
#> [155] 1->82->83->150                                           
#> [156] 2->4->6                                                  
#> [157] 1->82->84->86->88->90->91->94                            
#> [158] 1->82->84->86->88->89->104                               
#> [159] 1->82->84->86->87->110->111->117->128->130->132          
#> [160] 2->3->16->18->20->21->33->40->41                         
#> [161] 1->81->157->181->187->190                                
#> [162] 1->81->157->182->183                                     
#> [163] 1->82->84->86->87->110->111->117->128->130->132->134     
#> [164] 1->81->157->181->187->189                                
#> [165] 2->3->16->18->20->21->34->36                             
#> [166] 1->81->158->160->161->172                                
#> [167] 2->3->16->18->19->61->66->68->70                         
#> [168] 1->82->84->86->88->90->91->94->96                        
#> [169] 2->3->16->18->20->21->33->40->41->46                     
#> [170] 1->81->158->160->161->172->174                           
#> [171] 1->82->84->86->87->110->112                              
#> [172] 2->4->5->9->12                                           
#> [173] 1->82->84->86->87->109                                   
#> [174] 1->82->84->86->88->90->91->93                            
#> [175] 2->3->16->18->20->22->24                                 
#> [176] 1->81->158->160->161->172->174->176                      
#> [177] 1->82->84->86->87->110->111->118->119->121               
#> [178] 2->3->16->18->19->62                                     
#> [179] 1->82->84->86->87->110->111->117->127->138               
#> [180] 1->81->158->160->162->163                                
#> [181] 2->3->16->18->20->21->33->40->41->45                     
#> [182] 2->3->16->18->20->21->33->39->53                         
#> [183] 2->3->16->18->20->22->23->27->29                         
#> [184] 1->82->84->86->87->110->111->117->127->138->140          
#> [185] 1->81->158->160->162->163->166                           
#> [186] 1->81->158->159                                          
#> [187] 2->3->16->18->20->21->33->39->54                         
#> [188] 1->82->84->86->87->110->111->117->127->137               
#> [189] 2->3->16->18->19->61->65                                 
#> [190] 1->81->158->160->162->163->166->168                      
#> [191] 1->82->84->86->88->90->91->94->96->97                    
#> [192] 1->82->83->150->152                                      
#> [193] 2->3->16->18->20->21->33->40->41->46->48                 
#> [194] 1->82->84->86->87->110->112->114                         
#> [195] 1->81->157->181->187->189->193                           
#> [196] 1->81->157->181->187->189->193->195                      
#> [197] 2->3->16->18->20->21->33->39->53->57                     
#> [198] 1->82->83->149