Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──1.246──→ Node2:Node2 ──2.164──→ Node2:t18 ──0.013──→ Node3:Node2 ──0.023──→ Node3:t18 ──0.125──→ Node5:Node2 ──0.216──→ Node5:t18 ──0.024──→ Node7:Node2 ──0.042──→ Node7:t18 ──0.064──→ t1:Node2 ──0.111──→ t1:t18 
#> 
#> [2] ◎──1.2458──→ Node2:Node2 ──1.5070──→ Node2:Node6 ──0.4581──→ Node2:Node15 ──0.1992──→ Node2:t15 ──0.0131──→ Node3:Node2 ──0.0158──→ Node3:Node6 ──0.0048──→ Node3:Node15 ──0.0021──→ Node3:t15 ──0.1245──→ Node5:Node2 ──0.1506──→ Node5:Node6 ──0.0458──→ Node5:Node15 ──0.0199──→ Node5:t15 ──0.0241──→ Node7:Node2 ──0.0292──→ Node7:Node6 ──0.0089──→ Node7:Node15 ──0.0039──→ Node7:t15 ──0.0640──→ t4:Node2 ──0.0774──→ t4:Node6 ──0.0235──→ t4:Node15 ──0.0102──→ t4:t15 
#> 
#> [3] ◎──1.24584──→ Node2:Node2 ──1.50702──→ Node2:Node6 ──0.45814──→ Node2:Node15 ──0.19918──→ Node2:t9 ──0.01309──→ Node3:Node2 ──0.01584──→ Node3:Node6 ──0.00481──→ Node3:Node15 ──0.00209──→ Node3:t9 ──0.12450──→ Node5:Node2 ──0.15061──→ Node5:Node6 ──0.04578──→ Node5:Node15 ──0.01991──→ Node5:t9 ──0.08553──→ Node18:Node2 ──0.10346──→ Node18:Node6 ──0.03145──→ Node18:Node15 ──0.01367──→ Node18:t9 ──0.00262──→ t19:Node2 ──0.00317──→ t19:Node6 ──0.00096──→ t19:Node15 ──0.00042──→ t19:t9 
#> 
#> [4] ◎──1.2e+00──→ Node2:Node2 ──1.5e+00──→ Node2:Node6 ──2.5e-02──→ Node2:Node7 ──4.4e-01──→ Node2:Node16 ──1.9e-01──→ Node2:t8 ──1.3e-02──→ Node3:Node2 ──1.6e-02──→ Node3:Node6 ──2.6e-04──→ Node3:Node7 ──4.6e-03──→ Node3:Node16 ──2.0e-03──→ Node3:t8 ──1.2e-01──→ Node5:Node2 ──1.5e-01──→ Node5:Node6 ──2.5e-03──→ Node5:Node7 ──4.4e-02──→ Node5:Node16 ──1.9e-02──→ Node5:t8 ──8.6e-02──→ Node18:Node2 ──1.0e-01──→ Node18:Node6 ──1.7e-03──→ Node18:Node7 ──3.0e-02──→ Node18:Node16 ──1.3e-02──→ Node18:t8 ──2.6e-03──→ t6:Node2 ──3.2e-03──→ t6:Node6 ──5.2e-05──→ t6:Node7 ──9.2e-04──→ t6:Node16 ──4.1e-04──→ t6:t8 
#> 
#> [5] ◎──1.2e+00──→ Node2:Node2 ──1.5e+00──→ Node2:Node6 ──2.5e-02──→ Node2:Node7 ──4.4e-01──→ Node2:Node16 ──1.9e-01──→ Node2:t14 ──1.3e-02──→ Node3:Node2 ──1.6e-02──→ Node3:Node6 ──2.6e-04──→ Node3:Node7 ──4.6e-03──→ Node3:Node16 ──2.0e-03──→ Node3:t14 ──4.2e-03──→ Node4:Node2 ──5.1e-03──→ Node4:Node6 ──8.3e-05──→ Node4:Node7 ──1.5e-03──→ Node4:Node16 ──6.6e-04──→ Node4:t14 ──1.6e-01──→ Node8:Node2 ──2.0e-01──→ Node8:Node6 ──3.2e-03──→ Node8:Node7 ──5.8e-02──→ Node8:Node16 ──2.6e-02──→ Node8:t14 ──2.3e-02──→ Node13:Node2 ──2.8e-02──→ Node13:Node6 ──4.6e-04──→ Node13:Node7 ──8.2e-03──→ Node13:Node16 ──3.6e-03──→ Node13:t14 ──2.1e-02──→ t14:Node2 ──2.6e-02──→ t14:Node6 ──4.2e-04──→ t14:Node7 ──7.4e-03──→ t14:Node16 ──3.3e-03──→ t14:t14 
#> 
#> [6] ◎──1.2e+00──→ Node2:Node2 ──1.5e+00──→ Node2:Node6 ──2.5e-02──→ Node2:Node7 ──1.9e-01──→ Node2:Node11 ──4.4e-01──→ Node2:t13 ──1.3e-02──→ Node3:Node2 ──1.6e-02──→ Node3:Node6 ──2.6e-04──→ Node3:Node7 ──2.0e-03──→ Node3:Node11 ──4.6e-03──→ Node3:t13 ──4.2e-03──→ Node4:Node2 ──5.1e-03──→ Node4:Node6 ──8.3e-05──→ Node4:Node7 ──6.5e-04──→ Node4:Node11 ──1.5e-03──→ Node4:t13 ──1.6e-01──→ Node8:Node2 ──2.0e-01──→ Node8:Node6 ──3.2e-03──→ Node8:Node7 ──2.5e-02──→ Node8:Node11 ──5.8e-02──→ Node8:t13 ──2.3e-02──→ Node13:Node2 ──2.8e-02──→ Node13:Node6 ──4.6e-04──→ Node13:Node7 ──3.6e-03──→ Node13:Node11 ──8.3e-03──→ Node13:t13 ──2.1e-02──→ t15:Node2 ──2.6e-02──→ t15:Node6 ──4.2e-04──→ t15:Node7 ──3.3e-03──→ t15:Node11 ──7.5e-03──→ t15:t13 
#> 
#> [7] ◎──1.2e+00──→ Node2:Node2 ──1.5e+00──→ Node2:Node6 ──2.5e-02──→ Node2:Node7 ──1.9e-01──→ Node2:Node11 ──4.4e-01──→ Node2:t4 ──1.3e-02──→ Node3:Node2 ──1.6e-02──→ Node3:Node6 ──2.6e-04──→ Node3:Node7 ──2.0e-03──→ Node3:Node11 ──4.6e-03──→ Node3:t4 ──4.2e-03──→ Node4:Node2 ──5.1e-03──→ Node4:Node6 ──8.3e-05──→ Node4:Node7 ──6.5e-04──→ Node4:Node11 ──1.5e-03──→ Node4:t4 ──1.6e-01──→ Node8:Node2 ──2.0e-01──→ Node8:Node6 ──3.2e-03──→ Node8:Node7 ──2.5e-02──→ Node8:Node11 ──5.8e-02──→ Node8:t4 ──7.7e-04──→ Node9:Node2 ──9.3e-04──→ Node9:Node6 ──1.5e-05──→ Node9:Node7 ──1.2e-04──→ Node9:Node11 ──2.7e-04──→ Node9:t4 ──3.3e-03──→ Node10:Node2 ──4.0e-03──→ Node10:Node6 ──6.6e-05──→ Node10:Node7 ──5.2e-04──→ Node10:Node11 ──1.2e-03──→ Node10:t4 ──4.0e-02──→ t18:Node2 ──4.9e-02──→ t18:Node6 ──7.9e-04──→ t18:Node7 ──6.3e-03──→ t18:Node11 ──1.4e-02──→ t18:t4 
#> 
#> [8] ◎──2.5e+00──→ Node2:Node3 ──1.6e-01──→ Node2:Node4 ──4.9e-01──→ Node2:Node13 ──2.5e-01──→ Node2:t5 ──2.6e-02──→ Node3:Node3 ──1.7e-03──→ Node3:Node4 ──5.2e-03──→ Node3:Node13 ──2.6e-03──→ Node3:t5 ──8.5e-03──→ Node4:Node3 ──5.4e-04──→ Node4:Node4 ──1.7e-03──→ Node4:Node13 ──8.5e-04──→ Node4:t5 ──3.3e-01──→ Node8:Node3 ──2.1e-02──→ Node8:Node4 ──6.5e-02──→ Node8:Node13 ──3.3e-02──→ Node8:t5 ──1.5e-03──→ Node9:Node3 ──9.9e-05──→ Node9:Node4 ──3.1e-04──→ Node9:Node13 ──1.5e-04──→ Node9:t5 ──6.7e-03──→ Node10:Node3 ──4.3e-04──→ Node10:Node4 ──1.3e-03──→ Node10:Node13 ──6.7e-04──→ Node10:t5 ──6.2e-02──→ Node15:Node3 ──4.0e-03──→ Node15:Node4 ──1.2e-02──→ Node15:Node13 ──6.2e-03──→ Node15:t5 ──1.9e-02──→ t13:Node3 ──1.2e-03──→ t13:Node4 ──3.7e-03──→ t13:Node13 ──1.9e-03──→ t13:t5 
#> 
#> [9] ◎──2.5e+00──→ Node2:Node3 ──1.6e-01──→ Node2:Node4 ──4.9e-01──→ Node2:Node13 ──2.5e-01──→ Node2:Node19 ──3.7e-03──→ Node2:t11 ──2.6e-02──→ Node3:Node3 ──1.7e-03──→ Node3:Node4 ──5.2e-03──→ Node3:Node13 ──2.6e-03──→ Node3:Node19 ──3.9e-05──→ Node3:t11 ──8.5e-03──→ Node4:Node3 ──5.4e-04──→ Node4:Node4 ──1.7e-03──→ Node4:Node13 ──8.3e-04──→ Node4:Node19 ──1.3e-05──→ Node4:t11 ──3.3e-01──→ Node8:Node3 ──2.1e-02──→ Node8:Node4 ──6.5e-02──→ Node8:Node13 ──3.2e-02──→ Node8:Node19 ──4.9e-04──→ Node8:t11 ──1.5e-03──→ Node9:Node3 ──9.9e-05──→ Node9:Node4 ──3.1e-04──→ Node9:Node13 ──1.5e-04──→ Node9:Node19 ──2.3e-06──→ Node9:t11 ──6.7e-03──→ Node10:Node3 ──4.3e-04──→ Node10:Node4 ──1.3e-03──→ Node10:Node13 ──6.6e-04──→ Node10:Node19 ──1.0e-05──→ Node10:t11 ──6.2e-02──→ Node15:Node3 ──4.0e-03──→ Node15:Node4 ──1.2e-02──→ Node15:Node13 ──6.1e-03──→ Node15:Node19 ──9.3e-05──→ Node15:t11 ──1.9e-02──→ t7:Node3 ──1.2e-03──→ t7:Node4 ──3.7e-03──→ t7:Node13 ──1.9e-03──→ t7:Node19 ──2.8e-05──→ t7:t11 
#> 
#> [10] ◎──2.5e+00──→ Node2:Node3 ──1.6e-01──→ Node2:Node4 ──4.9e-01──→ Node2:Node13 ──2.5e-01──→ Node2:Node19 ──3.7e-03──→ Node2:t2 ──2.6e-02──→ Node3:Node3 ──1.7e-03──→ Node3:Node4 ──5.2e-03──→ Node3:Node13 ──2.6e-03──→ Node3:Node19 ──3.9e-05──→ Node3:t2 ──8.5e-03──→ Node4:Node3 ──5.4e-04──→ Node4:Node4 ──1.7e-03──→ Node4:Node13 ──8.3e-04──→ Node4:Node19 ──1.3e-05──→ Node4:t2 ──3.3e-01──→ Node8:Node3 ──2.1e-02──→ Node8:Node4 ──6.5e-02──→ Node8:Node13 ──3.2e-02──→ Node8:Node19 ──4.9e-04──→ Node8:t2 ──1.5e-03──→ Node9:Node3 ──9.9e-05──→ Node9:Node4 ──3.1e-04──→ Node9:Node13 ──1.5e-04──→ Node9:Node19 ──2.3e-06──→ Node9:t2 ──8.8e-02──→ t11:Node3 ──5.6e-03──→ t11:Node4 ──1.7e-02──→ t11:Node13 ──8.6e-03──→ t11:Node19 ──1.3e-04──→ t11:t2