Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──0.1574──→ Node4:Node2 ──0.4221──→ Node4:Node4 ──0.4353──→ Node4:Node9 ──0.2476──→ Node4:t19 ──0.0076──→ Node9:Node2 ──0.0203──→ Node9:Node4 ──0.0209──→ Node9:Node9 ──0.0119──→ Node9:t19 ──0.0082──→ t2:Node2 ──0.0219──→ t2:Node4 ──0.0226──→ t2:Node9 ──0.0129──→ t2:t19 
#> 
#> [2] ◎──0.1574──→ Node4:Node2 ──0.4221──→ Node4:Node4 ──0.4353──→ Node4:Node9 ──0.2476──→ Node4:t18 ──0.0076──→ Node9:Node2 ──0.0203──→ Node9:Node4 ──0.0209──→ Node9:Node9 ──0.0119──→ Node9:t18 ──0.0053──→ Node15:Node2 ──0.0143──→ Node15:Node4 ──0.0147──→ Node15:Node9 ──0.0084──→ Node15:t18 ──0.0029──→ t13:Node2 ──0.0077──→ t13:Node4 ──0.0079──→ t13:Node9 ──0.0045──→ t13:t18 
#> 
#> [3] ◎──0.1574──→ Node4:Node2 ──0.4221──→ Node4:Node4 ──0.1881──→ Node4:Node5 ──0.1177──→ Node4:Node6 ──0.2276──→ Node4:Node12 ──0.1495──→ Node4:t2 ──0.0076──→ Node9:Node2 ──0.0203──→ Node9:Node4 ──0.0090──→ Node9:Node5 ──0.0057──→ Node9:Node6 ──0.0109──→ Node9:Node12 ──0.0072──→ Node9:t2 ──0.0053──→ Node15:Node2 ──0.0143──→ Node15:Node4 ──0.0064──→ Node15:Node5 ──0.0040──→ Node15:Node6 ──0.0077──→ Node15:Node12 ──0.0051──→ Node15:t2 ──0.0029──→ t15:Node2 ──0.0077──→ t15:Node4 ──0.0034──→ t15:Node5 ──0.0021──→ t15:Node6 ──0.0041──→ t15:Node12 ──0.0027──→ t15:t2 
#> 
#> [4] ◎──0.1574──→ Node4:Node2 ──0.4221──→ Node4:Node4 ──0.1881──→ Node4:Node5 ──0.1177──→ Node4:Node6 ──0.2276──→ Node4:Node12 ──0.1495──→ Node4:t13 ──0.0141──→ Node16:Node2 ──0.0378──→ Node16:Node4 ──0.0168──→ Node16:Node5 ──0.0105──→ Node16:Node6 ──0.0204──→ Node16:Node12 ──0.0134──→ Node16:t13 ──0.0017──→ t8:Node2 ──0.0044──→ t8:Node4 ──0.0020──→ t8:Node5 ──0.0012──→ t8:Node6 ──0.0024──→ t8:Node12 ──0.0016──→ t8:t13 
#> 
#> [5] ◎──1.6e-01──→ Node4:Node2 ──4.2e-01──→ Node4:Node4 ──1.9e-01──→ Node4:Node5 ──1.2e-01──→ Node4:Node6 ──3.8e-01──→ Node4:t7 ──1.4e-02──→ Node16:Node2 ──3.8e-02──→ Node16:Node4 ──1.7e-02──→ Node16:Node5 ──1.1e-02──→ Node16:Node6 ──3.4e-02──→ Node16:t7 ──7.2e-05──→ Node17:Node2 ──1.9e-04──→ Node17:Node4 ──8.6e-05──→ Node17:Node5 ──5.4e-05──→ Node17:Node6 ──1.7e-04──→ Node17:t7 ──1.6e-03──→ t20:Node2 ──4.2e-03──→ t20:Node4 ──1.9e-03──→ t20:Node5 ──1.2e-03──→ t20:Node6 ──3.8e-03──→ t20:t7 
#> 
#> [6] ◎──1.6e-01──→ Node4:Node2 ──4.2e-01──→ Node4:Node4 ──1.9e-01──→ Node4:Node5 ──1.7e-01──→ Node4:Node7 ──1.4e-01──→ Node4:Node11 ──1.9e-01──→ Node4:t10 ──1.4e-02──→ Node16:Node2 ──3.8e-02──→ Node16:Node4 ──1.7e-02──→ Node16:Node5 ──1.5e-02──→ Node16:Node7 ──1.3e-02──→ Node16:Node11 ──1.7e-02──→ Node16:t10 ──7.2e-05──→ Node17:Node2 ──1.9e-04──→ Node17:Node4 ──8.6e-05──→ Node17:Node5 ──7.7e-05──→ Node17:Node7 ──6.5e-05──→ Node17:Node11 ──8.5e-05──→ Node17:t10 ──1.6e-03──→ t17:Node2 ──4.2e-03──→ t17:Node4 ──1.9e-03──→ t17:Node5 ──1.7e-03──→ t17:Node7 ──1.4e-03──→ t17:Node11 ──1.9e-03──→ t17:t10 
#> 
#> [7] ◎──0.14970──→ Node2:Node2 ──0.40153──→ Node2:Node4 ──0.17892──→ Node2:Node5 ──0.15961──→ Node2:Node7 ──0.13436──→ Node2:Node11 ──0.17676──→ Node2:t6 ──0.00799──→ Node5:Node2 ──0.02144──→ Node5:Node4 ──0.00955──→ Node5:Node5 ──0.00852──→ Node5:Node7 ──0.00717──→ Node5:Node11 ──0.00944──→ Node5:t6 ──0.00346──→ Node6:Node2 ──0.00928──→ Node6:Node4 ──0.00413──→ Node6:Node5 ──0.00369──→ Node6:Node7 ──0.00310──→ Node6:Node11 ──0.00408──→ Node6:t6 ──0.00224──→ Node8:Node2 ──0.00600──→ Node8:Node4 ──0.00268──→ Node8:Node5 ──0.00239──→ Node8:Node7 ──0.00201──→ Node8:Node11 ──0.00264──→ Node8:t6 ──0.00229──→ Node10:Node2 ──0.00615──→ Node10:Node4 ──0.00274──→ Node10:Node5 ──0.00244──→ Node10:Node7 ──0.00206──→ Node10:Node11 ──0.00271──→ Node10:t6 ──0.00596──→ Node18:Node2 ──0.01597──→ Node18:Node4 ──0.00712──→ Node18:Node5 ──0.00635──→ Node18:Node7 ──0.00534──→ Node18:Node11 ──0.00703──→ Node18:t6 ──0.00048──→ Node19:Node2 ──0.00128──→ Node19:Node4 ──0.00057──→ Node19:Node5 ──0.00051──→ Node19:Node7 ──0.00043──→ Node19:Node11 ──0.00056──→ Node19:t6 ──0.00098──→ t19:Node2 ──0.00262──→ t19:Node4 ──0.00117──→ t19:Node5 ──0.00104──→ t19:Node7 ──0.00088──→ t19:Node11 ──0.00115──→ t19:t6 
#> 
#> [8] ◎──0.14970──→ Node2:Node2 ──0.40153──→ Node2:Node4 ──0.17892──→ Node2:Node5 ──0.15961──→ Node2:Node7 ──0.31111──→ Node2:t9 ──0.00799──→ Node5:Node2 ──0.02144──→ Node5:Node4 ──0.00955──→ Node5:Node5 ──0.00852──→ Node5:Node7 ──0.01661──→ Node5:t9 ──0.00346──→ Node6:Node2 ──0.00928──→ Node6:Node4 ──0.00413──→ Node6:Node5 ──0.00369──→ Node6:Node7 ──0.00719──→ Node6:t9 ──0.00224──→ Node8:Node2 ──0.00600──→ Node8:Node4 ──0.00268──→ Node8:Node5 ──0.00239──→ Node8:Node7 ──0.00465──→ Node8:t9 ──0.00229──→ Node10:Node2 ──0.00615──→ Node10:Node4 ──0.00274──→ Node10:Node5 ──0.00244──→ Node10:Node7 ──0.00476──→ Node10:t9 ──0.00596──→ Node18:Node2 ──0.01597──→ Node18:Node4 ──0.00712──→ Node18:Node5 ──0.00635──→ Node18:Node7 ──0.01238──→ Node18:t9 ──0.00048──→ Node19:Node2 ──0.00128──→ Node19:Node4 ──0.00057──→ Node19:Node5 ──0.00051──→ Node19:Node7 ──0.00099──→ Node19:t9 ──0.00098──→ t7:Node2 ──0.00262──→ t7:Node4 ──0.00117──→ t7:Node5 ──0.00104──→ t7:Node7 ──0.00203──→ t7:t9 
#> 
#> [9] ◎──0.14970──→ Node2:Node2 ──0.03409──→ Node2:Node3 ──0.99256──→ Node2:Node16 ──0.02452──→ Node2:t11 ──0.00799──→ Node5:Node2 ──0.00182──→ Node5:Node3 ──0.05299──→ Node5:Node16 ──0.00131──→ Node5:t11 ──0.00346──→ Node6:Node2 ──0.00079──→ Node6:Node3 ──0.02293──→ Node6:Node16 ──0.00057──→ Node6:t11 ──0.00224──→ Node8:Node2 ──0.00051──→ Node8:Node3 ──0.01484──→ Node8:Node16 ──0.00037──→ Node8:t11 ──0.00229──→ Node10:Node2 ──0.00052──→ Node10:Node3 ──0.01519──→ Node10:Node16 ──0.00038──→ Node10:t11 ──0.00596──→ Node18:Node2 ──0.00136──→ Node18:Node3 ──0.03949──→ Node18:Node16 ──0.00098──→ Node18:t11 ──0.00145──→ t9:Node2 ──0.00033──→ t9:Node3 ──0.00964──→ t9:Node16 ──0.00024──→ t9:t11 
#> 
#> [10] ◎──1.5e-01──→ Node2:Node2 ──3.4e-02──→ Node2:Node3 ──9.9e-01──→ Node2:Node16 ──2.5e-02──→ Node2:t16 ──8.0e-03──→ Node5:Node2 ──1.8e-03──→ Node5:Node3 ──5.3e-02──→ Node5:Node16 ──1.3e-03──→ Node5:t16 ──3.5e-03──→ Node6:Node2 ──7.9e-04──→ Node6:Node3 ──2.3e-02──→ Node6:Node16 ──5.7e-04──→ Node6:t16 ──2.2e-03──→ Node8:Node2 ──5.1e-04──→ Node8:Node3 ──1.5e-02──→ Node8:Node16 ──3.7e-04──→ Node8:t16 ──2.3e-03──→ Node10:Node2 ──5.2e-04──→ Node10:Node3 ──1.5e-02──→ Node10:Node16 ──3.8e-04──→ Node10:t16 ──2.8e-03──→ Node11:Node2 ──6.3e-04──→ Node11:Node3 ──1.8e-02──→ Node11:Node16 ──4.6e-04──→ Node11:t16 ──9.5e-04──→ Node12:Node2 ──2.2e-04──→ Node12:Node3 ──6.3e-03──→ Node12:Node16 ──1.6e-04──→ Node12:t16 ──6.7e-05──→ Node13:Node2 ──1.5e-05──→ Node13:Node3 ──4.4e-04──→ Node13:Node16 ──1.1e-05──→ Node13:t16 ──3.6e-03──→ t1:Node2 ──8.2e-04──→ t1:Node3 ──2.4e-02──→ t1:Node16 ──5.9e-04──→ t1:t16