Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──0.2015──→ Node2:Node2 ──0.0571──→ Node2:Node3 ──0.1335──→ Node2:Node4 ──0.1220──→ Node2:Node10 ──0.0239──→ Node2:t9 ──0.5249──→ Node3:Node2 ──0.1488──→ Node3:Node3 ──0.3477──→ Node3:Node4 ──0.3178──→ Node3:Node10 ──0.0624──→ Node3:t9 ──0.3886──→ Node14:Node2 ──0.1102──→ Node14:Node3 ──0.2574──→ Node14:Node4 ──0.2353──→ Node14:Node10 ──0.0462──→ Node14:t9 ──0.0271──→ t12:Node2 ──0.0077──→ t12:Node3 ──0.0180──→ t12:Node4 ──0.0164──→ t12:Node10 ──0.0032──→ t12:t9 
#> 
#> [2] ◎──0.20152──→ Node2:Node2 ──0.05712──→ Node2:Node3 ──0.13349──→ Node2:Node4 ──0.12200──→ Node2:Node10 ──0.02395──→ Node2:t19 ──0.52490──→ Node3:Node2 ──0.14878──→ Node3:Node3 ──0.34770──→ Node3:Node4 ──0.31777──→ Node3:Node10 ──0.06238──→ Node3:t19 ──0.38863──→ Node14:Node2 ──0.11015──→ Node14:Node3 ──0.25743──→ Node14:Node4 ──0.23527──→ Node14:Node10 ──0.04618──→ Node14:t19 ──0.00807──→ Node15:Node2 ──0.00229──→ Node15:Node3 ──0.00535──→ Node15:Node4 ──0.00489──→ Node15:Node10 ──0.00096──→ Node15:t19 ──0.01907──→ t5:Node2 ──0.00540──→ t5:Node3 ──0.01263──→ t5:Node4 ──0.01154──→ t5:Node10 ──0.00227──→ t5:t19 
#> 
#> [3] ◎──0.20152──→ Node2:Node2 ──0.05712──→ Node2:Node3 ──0.13349──→ Node2:Node4 ──0.06020──→ Node2:Node5 ──0.00678──→ Node2:Node6 ──0.07897──→ Node2:t14 ──0.52490──→ Node3:Node2 ──0.14878──→ Node3:Node3 ──0.34770──→ Node3:Node4 ──0.15679──→ Node3:Node5 ──0.01767──→ Node3:Node6 ──0.20568──→ Node3:t14 ──0.38863──→ Node14:Node2 ──0.11015──→ Node14:Node3 ──0.25743──→ Node14:Node4 ──0.11609──→ Node14:Node5 ──0.01308──→ Node14:Node6 ──0.15229──→ Node14:t14 ──0.00807──→ Node15:Node2 ──0.00229──→ Node15:Node3 ──0.00535──→ Node15:Node4 ──0.00241──→ Node15:Node5 ──0.00027──→ Node15:Node6 ──0.00316──→ Node15:t14 ──0.01907──→ t20:Node2 ──0.00540──→ t20:Node3 ──0.01263──→ t20:Node4 ──0.00570──→ t20:Node5 ──0.00064──→ t20:Node6 ──0.00747──→ t20:t14 
#> 
#> [4] ◎──0.2015──→ Node2:Node2 ──0.0571──→ Node2:Node3 ──0.1335──→ Node2:Node4 ──0.0602──→ Node2:Node5 ──0.0068──→ Node2:Node6 ──0.0584──→ Node2:Node11 ──0.0205──→ Node2:t13 ──0.5249──→ Node3:Node2 ──0.1488──→ Node3:Node3 ──0.3477──→ Node3:Node4 ──0.1568──→ Node3:Node5 ──0.0177──→ Node3:Node6 ──0.1522──→ Node3:Node11 ──0.0535──→ Node3:t13 ──0.2588──→ Node6:Node2 ──0.0733──→ Node6:Node3 ──0.1714──→ Node6:Node4 ──0.0773──→ Node6:Node5 ──0.0087──→ Node6:Node6 ──0.0750──→ Node6:Node11 ──0.0264──→ Node6:t13 ──0.1570──→ t1:Node2 ──0.0445──→ t1:Node3 ──0.1040──→ t1:Node4 ──0.0469──→ t1:Node5 ──0.0053──→ t1:Node6 ──0.0455──→ t1:Node11 ──0.0160──→ t1:t13 
#> 
#> [5] ◎──0.2015──→ Node2:Node2 ──0.0571──→ Node2:Node3 ──0.1335──→ Node2:Node4 ──0.0602──→ Node2:Node5 ──0.0068──→ Node2:Node6 ──0.0584──→ Node2:Node11 ──0.0081──→ Node2:Node16 ──0.0124──→ Node2:t12 ──0.5249──→ Node3:Node2 ──0.1488──→ Node3:Node3 ──0.3477──→ Node3:Node4 ──0.1568──→ Node3:Node5 ──0.0177──→ Node3:Node6 ──0.1522──→ Node3:Node11 ──0.0212──→ Node3:Node16 ──0.0323──→ Node3:t12 ──0.2588──→ Node6:Node2 ──0.0733──→ Node6:Node3 ──0.1714──→ Node6:Node4 ──0.0773──→ Node6:Node5 ──0.0087──→ Node6:Node6 ──0.0750──→ Node6:Node11 ──0.0105──→ Node6:Node16 ──0.0159──→ Node6:t12 ──0.0967──→ Node11:Node2 ──0.0274──→ Node11:Node3 ──0.0641──→ Node11:Node4 ──0.0289──→ Node11:Node5 ──0.0033──→ Node11:Node6 ──0.0280──→ Node11:Node11 ──0.0039──→ Node11:Node16 ──0.0059──→ Node11:t12 ──0.0603──→ t15:Node2 ──0.0171──→ t15:Node3 ──0.0399──→ t15:Node4 ──0.0180──→ t15:Node5 ──0.0020──→ t15:Node6 ──0.0175──→ t15:Node11 ──0.0024──→ t15:Node16 ──0.0037──→ t15:t12 
#> 
#> [6] ◎──0.20152──→ Node2:Node2 ──0.05712──→ Node2:Node3 ──0.13349──→ Node2:Node4 ──0.06020──→ Node2:Node5 ──0.00678──→ Node2:Node6 ──0.05844──→ Node2:Node11 ──0.00814──→ Node2:Node16 ──0.01238──→ Node2:t11 ──0.52490──→ Node3:Node2 ──0.14878──→ Node3:Node3 ──0.34770──→ Node3:Node4 ──0.15679──→ Node3:Node5 ──0.01767──→ Node3:Node6 ──0.15222──→ Node3:Node11 ──0.02120──→ Node3:Node16 ──0.03226──→ Node3:t11 ──0.25876──→ Node6:Node2 ──0.07334──→ Node6:Node3 ──0.17140──→ Node6:Node4 ──0.07729──→ Node6:Node5 ──0.00871──→ Node6:Node6 ──0.07504──→ Node6:Node11 ──0.01045──→ Node6:Node16 ──0.01590──→ Node6:t11 ──0.09672──→ Node11:Node2 ──0.02741──→ Node11:Node3 ──0.06407──→ Node11:Node4 ──0.02889──→ Node11:Node5 ──0.00326──→ Node11:Node6 ──0.02805──→ Node11:Node11 ──0.00391──→ Node11:Node16 ──0.00594──→ Node11:t11 ──0.05281──→ Node18:Node2 ──0.01497──→ Node18:Node3 ──0.03498──→ Node18:Node4 ──0.01578──→ Node18:Node5 ──0.00178──→ Node18:Node6 ──0.01532──→ Node18:Node11 ──0.00213──→ Node18:Node16 ──0.00325──→ Node18:t11 ──0.00748──→ t11:Node2 ──0.00212──→ t11:Node3 ──0.00496──→ t11:Node4 ──0.00223──→ t11:Node5 ──0.00025──→ t11:Node6 ──0.00217──→ t11:Node11 ──0.00030──→ t11:Node16 ──0.00046──→ t11:t11 
#> 
#> [7] ◎──0.20152──→ Node2:Node2 ──0.05712──→ Node2:Node3 ──0.13349──→ Node2:Node4 ──0.06020──→ Node2:Node5 ──0.06821──→ Node2:Node13 ──0.01754──→ Node2:t3 ──0.52490──→ Node3:Node2 ──0.14878──→ Node3:Node3 ──0.34770──→ Node3:Node4 ──0.15679──→ Node3:Node5 ──0.17767──→ Node3:Node13 ──0.04568──→ Node3:t3 ──0.25876──→ Node6:Node2 ──0.07334──→ Node6:Node3 ──0.17140──→ Node6:Node4 ──0.07729──→ Node6:Node5 ──0.08759──→ Node6:Node13 ──0.02252──→ Node6:t3 ──0.09672──→ Node11:Node2 ──0.02741──→ Node11:Node3 ──0.06407──→ Node11:Node4 ──0.02889──→ Node11:Node5 ──0.03274──→ Node11:Node13 ──0.00842──→ Node11:t3 ──0.05281──→ Node18:Node2 ──0.01497──→ Node18:Node3 ──0.03498──→ Node18:Node4 ──0.01578──→ Node18:Node5 ──0.01788──→ Node18:Node13 ──0.00460──→ Node18:t3 ──0.00748──→ t7:Node2 ──0.00212──→ t7:Node3 ──0.00496──→ t7:Node4 ──0.00223──→ t7:Node5 ──0.00253──→ t7:Node13 ──0.00065──→ t7:t3 
#> 
#> [8] ◎──0.2015──→ Node2:Node2 ──0.0571──→ Node2:Node3 ──0.1335──→ Node2:Node4 ──0.0602──→ Node2:Node5 ──0.0682──→ Node2:Node13 ──0.0046──→ Node2:Node15 ──0.0043──→ Node2:Node17 ──0.0085──→ Node2:t7 ──0.5291──→ Node4:Node2 ──0.1500──→ Node4:Node3 ──0.3505──→ Node4:Node4 ──0.1580──→ Node4:Node5 ──0.1791──→ Node4:Node13 ──0.0122──→ Node4:Node15 ──0.0114──→ Node4:Node17 ──0.0224──→ Node4:t7 ──0.3370──→ Node8:Node2 ──0.0955──→ Node8:Node3 ──0.2232──→ Node8:Node4 ──0.1007──→ Node8:Node5 ──0.1141──→ Node8:Node13 ──0.0078──→ Node8:Node15 ──0.0073──→ Node8:Node17 ──0.0143──→ Node8:t7 ──0.0746──→ t16:Node2 ──0.0211──→ t16:Node3 ──0.0494──→ t16:Node4 ──0.0223──→ t16:Node5 ──0.0252──→ t16:Node13 ──0.0017──→ t16:Node15 ──0.0016──→ t16:Node17 ──0.0032──→ t16:t7 
#> 
#> [9] ◎──0.2015──→ Node2:Node2 ──0.0571──→ Node2:Node3 ──0.1335──→ Node2:Node4 ──0.0602──→ Node2:Node5 ──0.0682──→ Node2:Node13 ──0.0046──→ Node2:Node15 ──0.0043──→ Node2:Node17 ──0.0085──→ Node2:t18 ──0.5291──→ Node4:Node2 ──0.1500──→ Node4:Node3 ──0.3505──→ Node4:Node4 ──0.1580──→ Node4:Node5 ──0.1791──→ Node4:Node13 ──0.0122──→ Node4:Node15 ──0.0114──→ Node4:Node17 ──0.0224──→ Node4:t18 ──0.3370──→ Node8:Node2 ──0.0955──→ Node8:Node3 ──0.2232──→ Node8:Node4 ──0.1007──→ Node8:Node5 ──0.1141──→ Node8:Node13 ──0.0078──→ Node8:Node15 ──0.0073──→ Node8:Node17 ──0.0143──→ Node8:t18 ──0.0746──→ t13:Node2 ──0.0211──→ t13:Node3 ──0.0494──→ t13:Node4 ──0.0223──→ t13:Node5 ──0.0252──→ t13:Node13 ──0.0017──→ t13:Node15 ──0.0016──→ t13:Node17 ──0.0032──→ t13:t18 
#> 
#> [10] ◎──2.0e-01──→ Node2:Node2 ──5.7e-02──→ Node2:Node3 ──1.3e-01──→ Node2:Node4 ──6.0e-02──→ Node2:Node5 ──6.8e-02──→ Node2:Node13 ──4.6e-03──→ Node2:Node15 ──1.3e-02──→ Node2:t2 ──5.3e-01──→ Node4:Node2 ──1.5e-01──→ Node4:Node3 ──3.5e-01──→ Node4:Node4 ──1.6e-01──→ Node4:Node5 ──1.8e-01──→ Node4:Node13 ──1.2e-02──→ Node4:Node15 ──3.4e-02──→ Node4:t2 ──3.1e-01──→ Node7:Node2 ──8.8e-02──→ Node7:Node3 ──2.0e-01──→ Node7:Node4 ──9.2e-02──→ Node7:Node5 ──1.0e-01──→ Node7:Node13 ──7.1e-03──→ Node7:Node15 ──2.0e-02──→ Node7:t2 ──9.9e-02──→ Node19:Node2 ──2.8e-02──→ Node19:Node3 ──6.6e-02──→ Node19:Node4 ──3.0e-02──→ Node19:Node5 ──3.4e-02──→ Node19:Node13 ──2.3e-03──→ Node19:Node15 ──6.4e-03──→ Node19:t2 ──2.7e-03──→ t17:Node2 ──7.5e-04──→ t17:Node3 ──1.8e-03──→ t17:Node4 ──8.0e-04──→ t17:Node5 ──9.0e-04──→ t17:Node13 ──6.1e-05──→ t17:Node15 ──1.7e-04──→ t17:t2