Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──1.1679──→ Node3:Node4 ──0.0331──→ Node3:Node5 ──0.0908──→ Node3:Node12 ──0.0059──→ Node3:Node13 ──0.0075──→ Node3:Node15 ──0.0089──→ Node3:Node16 ──0.0130──→ Node3:t7 ──1.9857──→ Node5:Node4 ──0.0563──→ Node5:Node5 ──0.1544──→ Node5:Node12 ──0.0100──→ Node5:Node13 ──0.0128──→ Node5:Node15 ──0.0151──→ Node5:Node16 ──0.0221──→ Node5:t7 ──0.4559──→ t19:Node4 ──0.0129──→ t19:Node5 ──0.0355──→ t19:Node12 ──0.0023──→ t19:Node13 ──0.0029──→ t19:Node15 ──0.0035──→ t19:Node16 ──0.0051──→ t19:t7 
#> 
#> [2] ◎──1.2e+00──→ Node3:Node4 ──3.3e-02──→ Node3:Node5 ──9.1e-02──→ Node3:Node12 ──5.9e-03──→ Node3:Node13 ──7.5e-03──→ Node3:Node15 ──8.9e-03──→ Node3:Node16 ──1.3e-02──→ Node3:Node19 ──1.3e-04──→ Node3:t18 ──2.0e+00──→ Node5:Node4 ──5.6e-02──→ Node5:Node5 ──1.5e-01──→ Node5:Node12 ──1.0e-02──→ Node5:Node13 ──1.3e-02──→ Node5:Node15 ──1.5e-02──→ Node5:Node16 ──2.2e-02──→ Node5:Node19 ──2.1e-04──→ Node5:t18 ──9.2e-02──→ Node7:Node4 ──2.6e-03──→ Node7:Node5 ──7.2e-03──→ Node7:Node12 ──4.7e-04──→ Node7:Node13 ──5.9e-04──→ Node7:Node15 ──7.0e-04──→ Node7:Node16 ──1.0e-03──→ Node7:Node19 ──9.9e-06──→ Node7:t18 ──3.6e-01──→ t18:Node4 ──1.0e-02──→ t18:Node5 ──2.8e-02──→ t18:Node12 ──1.8e-03──→ t18:Node13 ──2.3e-03──→ t18:Node15 ──2.8e-03──→ t18:Node16 ──4.0e-03──→ t18:Node19 ──3.9e-05──→ t18:t18 
#> 
#> [3] ◎──1.2e+00──→ Node3:Node4 ──3.3e-02──→ Node3:Node5 ──9.1e-02──→ Node3:Node12 ──5.9e-03──→ Node3:Node13 ──7.5e-03──→ Node3:Node15 ──8.9e-03──→ Node3:Node16 ──1.3e-02──→ Node3:Node19 ──1.3e-04──→ Node3:t19 ──2.0e+00──→ Node5:Node4 ──5.6e-02──→ Node5:Node5 ──1.5e-01──→ Node5:Node12 ──1.0e-02──→ Node5:Node13 ──1.3e-02──→ Node5:Node15 ──1.5e-02──→ Node5:Node16 ──2.2e-02──→ Node5:Node19 ──2.1e-04──→ Node5:t19 ──9.2e-02──→ Node7:Node4 ──2.6e-03──→ Node7:Node5 ──7.2e-03──→ Node7:Node12 ──4.7e-04──→ Node7:Node13 ──5.9e-04──→ Node7:Node15 ──7.0e-04──→ Node7:Node16 ──1.0e-03──→ Node7:Node19 ──9.9e-06──→ Node7:t19 ──2.3e-01──→ Node11:Node4 ──6.5e-03──→ Node11:Node5 ──1.8e-02──→ Node11:Node12 ──1.2e-03──→ Node11:Node13 ──1.5e-03──→ Node11:Node15 ──1.7e-03──→ Node11:Node16 ──2.5e-03──→ Node11:Node19 ──2.5e-05──→ Node11:t19 ──5.2e-02──→ Node13:Node4 ──1.5e-03──→ Node13:Node5 ──4.1e-03──→ Node13:Node12 ──2.6e-04──→ Node13:Node13 ──3.4e-04──→ Node13:Node15 ──4.0e-04──→ Node13:Node16 ──5.8e-04──→ Node13:Node19 ──5.6e-06──→ Node13:t19 ──6.0e-02──→ Node17:Node4 ──1.7e-03──→ Node17:Node5 ──4.7e-03──→ Node17:Node12 ──3.0e-04──→ Node17:Node13 ──3.9e-04──→ Node17:Node15 ──4.5e-04──→ Node17:Node16 ──6.6e-04──→ Node17:Node19 ──6.5e-06──→ Node17:t19 ──2.2e-02──→ t20:Node4 ──6.3e-04──→ t20:Node5 ──1.7e-03──→ t20:Node12 ──1.1e-04──→ t20:Node13 ──1.4e-04──→ t20:Node15 ──1.7e-04──→ t20:Node16 ──2.4e-04──→ t20:Node19 ──2.4e-06──→ t20:t19 
#> 
#> [4] ◎──1.16793──→ Node3:Node4 ──0.03312──→ Node3:Node5 ──0.09081──→ Node3:Node12 ──0.00589──→ Node3:Node13 ──0.00753──→ Node3:Node15 ──0.02184──→ Node3:t12 ──1.98572──→ Node5:Node4 ──0.05632──→ Node5:Node5 ──0.15440──→ Node5:Node12 ──0.01002──→ Node5:Node13 ──0.01280──→ Node5:Node15 ──0.03712──→ Node5:t12 ──0.09224──→ Node7:Node4 ──0.00262──→ Node7:Node5 ──0.00717──→ Node7:Node12 ──0.00047──→ Node7:Node13 ──0.00059──→ Node7:Node15 ──0.00172──→ Node7:t12 ──0.22930──→ Node11:Node4 ──0.00650──→ Node11:Node5 ──0.01783──→ Node11:Node12 ──0.00116──→ Node11:Node13 ──0.00148──→ Node11:Node15 ──0.00429──→ Node11:t12 ──0.05227──→ Node13:Node4 ──0.00148──→ Node13:Node5 ──0.00406──→ Node13:Node12 ──0.00026──→ Node13:Node13 ──0.00034──→ Node13:Node15 ──0.00098──→ Node13:t12 ──0.05998──→ Node17:Node4 ──0.00170──→ Node17:Node5 ──0.00466──→ Node17:Node12 ──0.00030──→ Node17:Node13 ──0.00039──→ Node17:Node15 ──0.00112──→ Node17:t12 ──0.02214──→ t3:Node4 ──0.00063──→ t3:Node5 ──0.00172──→ t3:Node12 ──0.00011──→ t3:Node13 ──0.00014──→ t3:Node15 ──0.00041──→ t3:t12 
#> 
#> [5] ◎──1.16793──→ Node3:Node4 ──0.03312──→ Node3:Node5 ──0.09081──→ Node3:Node12 ──0.00589──→ Node3:Node13 ──0.02936──→ Node3:t1 ──1.98572──→ Node5:Node4 ──0.05632──→ Node5:Node5 ──0.15440──→ Node5:Node12 ──0.01002──→ Node5:Node13 ──0.04992──→ Node5:t1 ──0.09224──→ Node7:Node4 ──0.00262──→ Node7:Node5 ──0.00717──→ Node7:Node12 ──0.00047──→ Node7:Node13 ──0.00232──→ Node7:t1 ──0.22930──→ Node11:Node4 ──0.00650──→ Node11:Node5 ──0.01783──→ Node11:Node12 ──0.00116──→ Node11:Node13 ──0.00576──→ Node11:t1 ──0.05227──→ Node13:Node4 ──0.00148──→ Node13:Node5 ──0.00406──→ Node13:Node12 ──0.00026──→ Node13:Node13 ──0.00131──→ Node13:t1 ──0.08212──→ t12:Node4 ──0.00233──→ t12:Node5 ──0.00639──→ t12:Node12 ──0.00041──→ t12:Node13 ──0.00206──→ t12:t1 
#> 
#> [6] ◎──1.1679──→ Node3:Node4 ──0.0331──→ Node3:Node5 ──0.0908──→ Node3:Node12 ──0.0353──→ Node3:t17 ──1.9857──→ Node5:Node4 ──0.0563──→ Node5:Node5 ──0.1544──→ Node5:Node12 ──0.0599──→ Node5:t17 ──0.0922──→ Node7:Node4 ──0.0026──→ Node7:Node5 ──0.0072──→ Node7:Node12 ──0.0028──→ Node7:t17 ──0.2293──→ Node11:Node4 ──0.0065──→ Node11:Node5 ──0.0178──→ Node11:Node12 ──0.0069──→ Node11:t17 ──0.1344──→ t6:Node4 ──0.0038──→ t6:Node5 ──0.0104──→ t6:Node12 ──0.0041──→ t6:t17 
#> 
#> [7] ◎──1.168──→ Node3:Node4 ──0.033──→ Node3:Node5 ──0.126──→ Node3:t3 ──2.442──→ t11:Node4 ──0.069──→ t11:Node5 ──0.264──→ t11:t3 
#> 
#> [8] ◎──0.684──→ Node2:Node4 ──0.093──→ Node2:t14 ──2.272──→ Node4:Node4 ──0.310──→ Node4:t14 ──0.653──→ t14:Node4 ──0.089──→ t14:t14 
#> 
#> [9] ◎──0.55692──→ Node2:Node2 ──0.02061──→ Node2:Node3 ──0.16747──→ Node2:Node10 ──0.03276──→ Node2:t15 ──1.84880──→ Node4:Node2 ──0.06842──→ Node4:Node3 ──0.55594──→ Node4:Node10 ──0.10876──→ Node4:t15 ──0.25902──→ Node8:Node2 ──0.00959──→ Node8:Node3 ──0.07789──→ Node8:Node10 ──0.01524──→ Node8:t15 ──0.17157──→ Node12:Node2 ──0.00635──→ Node12:Node3 ──0.05159──→ Node12:Node10 ──0.01009──→ Node12:t15 ──0.08125──→ Node16:Node2 ──0.00301──→ Node16:Node3 ──0.02443──→ Node16:Node10 ──0.00478──→ Node16:t15 ──0.01939──→ t15:Node2 ──0.00072──→ t15:Node3 ──0.00583──→ t15:Node10 ──0.00114──→ t15:t15 
#> 
#> [10] ◎──0.55692──→ Node2:Node2 ──0.02061──→ Node2:Node3 ──0.16747──→ Node2:Node10 ──0.03276──→ Node2:t11 ──1.84880──→ Node4:Node2 ──0.06842──→ Node4:Node3 ──0.55594──→ Node4:Node10 ──0.10876──→ Node4:t11 ──0.25902──→ Node8:Node2 ──0.00959──→ Node8:Node3 ──0.07789──→ Node8:Node10 ──0.01524──→ Node8:t11 ──0.17157──→ Node12:Node2 ──0.00635──→ Node12:Node3 ──0.05159──→ Node12:Node10 ──0.01009──→ Node12:t11 ──0.08125──→ Node16:Node2 ──0.00301──→ Node16:Node3 ──0.02443──→ Node16:Node10 ──0.00478──→ Node16:t11 ──0.01939──→ t8:Node2 ──0.00072──→ t8:Node3 ──0.00583──→ t8:Node10 ──0.00114──→ t8:t11