Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──3.7e-02──→ Node2:Node2 ──4.4e-01──→ Node2:Node10 ──4.8e-02──→ Node2:t14 ──4.7e-02──→ Node9:Node2 ──5.7e-01──→ Node9:Node10 ──6.1e-02──→ Node9:t14 ──3.1e-03──→ Node12:Node2 ──3.7e-02──→ Node12:Node10 ──4.0e-03──→ Node12:t14 ──2.9e-03──→ Node19:Node2 ──3.5e-02──→ Node19:Node10 ──3.7e-03──→ Node19:t14 ──5.5e-05──→ t16:Node2 ──6.6e-04──→ t16:Node10 ──7.1e-05──→ t16:t14 
#> 
#> [2] ◎──3.7e-02──→ Node2:Node2 ──4.4e-01──→ Node2:Node10 ──4.8e-02──→ Node2:t9 ──4.7e-02──→ Node9:Node2 ──5.7e-01──→ Node9:Node10 ──6.1e-02──→ Node9:t9 ──3.1e-03──→ Node12:Node2 ──3.7e-02──→ Node12:Node10 ──4.0e-03──→ Node12:t9 ──2.9e-03──→ Node19:Node2 ──3.5e-02──→ Node19:Node10 ──3.7e-03──→ Node19:t9 ──5.5e-05──→ t14:Node2 ──6.6e-04──→ t14:Node10 ──7.1e-05──→ t14:t9 
#> 
#> [3] ◎──0.0371──→ Node2:Node2 ──0.0577──→ Node2:Node3 ──0.4318──→ Node2:t16 ──0.0474──→ Node9:Node2 ──0.0738──→ Node9:Node3 ──0.5523──→ Node9:t16 ──0.0031──→ Node12:Node2 ──0.0048──→ Node12:Node3 ──0.0358──→ Node12:t16 ──0.0030──→ t12:Node2 ──0.0046──→ t12:Node3 ──0.0344──→ t12:t16 
#> 
#> [4] ◎──0.0371──→ Node2:Node2 ──0.0577──→ Node2:Node3 ──0.4318──→ Node2:t1 ──0.0474──→ Node9:Node2 ──0.0738──→ Node9:Node3 ──0.5523──→ Node9:t1 ──0.0060──→ t6:Node2 ──0.0094──→ t6:Node3 ──0.0703──→ t6:t1 
#> 
#> [5] ◎──0.240──→ Node2:Node4 ──0.138──→ Node2:Node5 ──0.148──→ Node2:t11 ──0.218──→ Node4:Node4 ──0.125──→ Node4:Node5 ──0.135──→ Node4:t11 ──0.128──→ t1:Node4 ──0.074──→ t1:Node5 ──0.079──→ t1:t11 
#> 
#> [6] ◎──0.2402──→ Node2:Node4 ──0.1379──→ Node2:Node5 ──0.0518──→ Node2:Node6 ──0.0741──→ Node2:Node16 ──0.0226──→ Node2:t10 ──0.2181──→ Node4:Node4 ──0.1252──→ Node4:Node5 ──0.0470──→ Node4:Node6 ──0.0673──→ Node4:Node16 ──0.0205──→ Node4:t10 ──0.0511──→ Node7:Node4 ──0.0293──→ Node7:Node5 ──0.0110──→ Node7:Node6 ──0.0158──→ Node7:Node16 ──0.0048──→ Node7:t10 ──0.0771──→ t19:Node4 ──0.0443──→ t19:Node5 ──0.0166──→ t19:Node6 ──0.0238──→ t19:Node16 ──0.0072──→ t19:t10 
#> 
#> [7] ◎──0.24018──→ Node2:Node4 ──0.13791──→ Node2:Node5 ──0.05177──→ Node2:Node6 ──0.07411──→ Node2:Node16 ──0.02258──→ Node2:t6 ──0.21811──→ Node4:Node4 ──0.12524──→ Node4:Node5 ──0.04702──→ Node4:Node6 ──0.06730──→ Node4:Node16 ──0.02050──→ Node4:t6 ──0.05107──→ Node7:Node4 ──0.02932──→ Node7:Node5 ──0.01101──→ Node7:Node6 ──0.01576──→ Node7:Node16 ──0.00480──→ Node7:t6 ──0.00486──→ Node8:Node4 ──0.00279──→ Node8:Node5 ──0.00105──→ Node8:Node6 ──0.00150──→ Node8:Node16 ──0.00046──→ Node8:t6 ──0.07225──→ t13:Node4 ──0.04149──→ t13:Node5 ──0.01557──→ t13:Node6 ──0.02229──→ t13:Node16 ──0.00679──→ t13:t6 
#> 
#> [8] ◎──0.24018──→ Node2:Node4 ──0.13791──→ Node2:Node5 ──0.05177──→ Node2:Node6 ──0.01394──→ Node2:Node8 ──0.08275──→ Node2:t13 ──0.21811──→ Node4:Node4 ──0.12524──→ Node4:Node5 ──0.04702──→ Node4:Node6 ──0.01266──→ Node4:Node8 ──0.07514──→ Node4:t13 ──0.05107──→ Node7:Node4 ──0.02932──→ Node7:Node5 ──0.01101──→ Node7:Node6 ──0.00296──→ Node7:Node8 ──0.01759──→ Node7:t13 ──0.00486──→ Node8:Node4 ──0.00279──→ Node8:Node5 ──0.00105──→ Node8:Node6 ──0.00028──→ Node8:Node8 ──0.00167──→ Node8:t13 ──0.03956──→ Node11:Node4 ──0.02272──→ Node11:Node5 ──0.00853──→ Node11:Node6 ──0.00230──→ Node11:Node8 ──0.01363──→ Node11:t13 ──0.02835──→ Node17:Node4 ──0.01628──→ Node17:Node5 ──0.00611──→ Node17:Node6 ──0.00165──→ Node17:Node8 ──0.00977──→ Node17:t13 ──0.00434──→ t15:Node4 ──0.00249──→ t15:Node5 ──0.00094──→ t15:Node6 ──0.00025──→ t15:Node8 ──0.00150──→ t15:t13 
#> 
#> [9] ◎──0.24018──→ Node2:Node4 ──0.13791──→ Node2:Node5 ──0.05177──→ Node2:Node6 ──0.01394──→ Node2:Node8 ──0.08275──→ Node2:t4 ──0.21811──→ Node4:Node4 ──0.12524──→ Node4:Node5 ──0.04702──→ Node4:Node6 ──0.01266──→ Node4:Node8 ──0.07514──→ Node4:t4 ──0.05107──→ Node7:Node4 ──0.02932──→ Node7:Node5 ──0.01101──→ Node7:Node6 ──0.00296──→ Node7:Node8 ──0.01759──→ Node7:t4 ──0.00486──→ Node8:Node4 ──0.00279──→ Node8:Node5 ──0.00105──→ Node8:Node6 ──0.00028──→ Node8:Node8 ──0.00167──→ Node8:t4 ──0.03956──→ Node11:Node4 ──0.02272──→ Node11:Node5 ──0.00853──→ Node11:Node6 ──0.00230──→ Node11:Node8 ──0.01363──→ Node11:t4 ──0.02835──→ Node17:Node4 ──0.01628──→ Node17:Node5 ──0.00611──→ Node17:Node6 ──0.00165──→ Node17:Node8 ──0.00977──→ Node17:t4 ──0.00434──→ t3:Node4 ──0.00249──→ t3:Node5 ──0.00094──→ t3:Node6 ──0.00025──→ t3:Node8 ──0.00150──→ t3:t4 
#> 
#> [10] ◎──0.24018──→ Node2:Node4 ──0.19747──→ Node2:Node7 ──0.03404──→ Node2:Node9 ──0.01073──→ Node2:Node11 ──0.04413──→ Node2:t3 ──0.21811──→ Node4:Node4 ──0.17932──→ Node4:Node7 ──0.03091──→ Node4:Node9 ──0.00974──→ Node4:Node11 ──0.04008──→ Node4:t3 ──0.05107──→ Node7:Node4 ──0.04198──→ Node7:Node7 ──0.00724──→ Node7:Node9 ──0.00228──→ Node7:Node11 ──0.00938──→ Node7:t3 ──0.00486──→ Node8:Node4 ──0.00399──→ Node8:Node7 ──0.00069──→ Node8:Node9 ──0.00022──→ Node8:Node11 ──0.00089──→ Node8:t3 ──0.03956──→ Node11:Node4 ──0.03253──→ Node11:Node7 ──0.00561──→ Node11:Node9 ──0.00177──→ Node11:Node11 ──0.00727──→ Node11:t3 ──0.03269──→ t5:Node4 ──0.02688──→ t5:Node7 ──0.00463──→ t5:Node9 ──0.00146──→ t5:Node11 ──0.00601──→ t5:t3