Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──5.369──→ t1:Node4 ──1.241──→ t1:Node8 ──0.294──→ t1:Node18 ──0.034──→ t1:t5 
#> 
#> [2] ◎──2.62163──→ Node2:Node4 ──0.60579──→ Node2:Node8 ──0.14343──→ Node2:Node18 ──0.01640──→ Node2:t1 ──1.41287──→ Node3:Node4 ──0.32648──→ Node3:Node8 ──0.07730──→ Node3:Node18 ──0.00884──→ Node3:t1 ──0.76523──→ Node5:Node4 ──0.17682──→ Node5:Node8 ──0.04187──→ Node5:Node18 ──0.00479──→ Node5:t1 ──0.22846──→ Node6:Node4 ──0.05279──→ Node6:Node8 ──0.01250──→ Node6:Node18 ──0.00143──→ Node6:t1 ──0.19451──→ Node10:Node4 ──0.04495──→ Node10:Node8 ──0.01064──→ Node10:Node18 ──0.00122──→ Node10:t1 ──0.08264──→ Node14:Node4 ──0.01910──→ Node14:Node8 ──0.00452──→ Node14:Node18 ──0.00052──→ Node14:t1 ──0.02516──→ Node18:Node4 ──0.00581──→ Node18:Node8 ──0.00138──→ Node18:Node18 ──0.00016──→ Node18:t1 ──0.03809──→ t18:Node4 ──0.00880──→ t18:Node8 ──0.00208──→ t18:Node18 ──0.00024──→ t18:t1 
#> 
#> [3] ◎──2.62163──→ Node2:Node4 ──0.60579──→ Node2:Node8 ──0.03219──→ Node2:Node9 ──0.05305──→ Node2:Node12 ──0.07460──→ Node2:t2 ──1.41287──→ Node3:Node4 ──0.32648──→ Node3:Node8 ──0.01735──→ Node3:Node9 ──0.02859──→ Node3:Node12 ──0.04020──→ Node3:t2 ──0.76523──→ Node5:Node4 ──0.17682──→ Node5:Node8 ──0.00940──→ Node5:Node9 ──0.01549──→ Node5:Node12 ──0.02177──→ Node5:t2 ──0.22846──→ Node6:Node4 ──0.05279──→ Node6:Node8 ──0.00280──→ Node6:Node9 ──0.00462──→ Node6:Node12 ──0.00650──→ Node6:t2 ──0.19451──→ Node10:Node4 ──0.04495──→ Node10:Node8 ──0.00239──→ Node10:Node9 ──0.00394──→ Node10:Node12 ──0.00553──→ Node10:t2 ──0.08264──→ Node14:Node4 ──0.01910──→ Node14:Node8 ──0.00101──→ Node14:Node9 ──0.00167──→ Node14:Node12 ──0.00235──→ Node14:t2 ──0.02516──→ Node18:Node4 ──0.00581──→ Node18:Node8 ──0.00031──→ Node18:Node9 ──0.00051──→ Node18:Node12 ──0.00072──→ Node18:t2 ──0.03809──→ t7:Node4 ──0.00880──→ t7:Node8 ──0.00047──→ t7:Node9 ──0.00077──→ t7:Node12 ──0.00108──→ t7:t2 
#> 
#> [4] ◎──2.62163──→ Node2:Node4 ──0.60579──→ Node2:Node8 ──0.03219──→ Node2:Node9 ──0.05305──→ Node2:Node12 ──0.07460──→ Node2:t14 ──1.41287──→ Node3:Node4 ──0.32648──→ Node3:Node8 ──0.01735──→ Node3:Node9 ──0.02859──→ Node3:Node12 ──0.04020──→ Node3:t14 ──0.76523──→ Node5:Node4 ──0.17682──→ Node5:Node8 ──0.00940──→ Node5:Node9 ──0.01549──→ Node5:Node12 ──0.02177──→ Node5:t14 ──0.22846──→ Node6:Node4 ──0.05279──→ Node6:Node8 ──0.00280──→ Node6:Node9 ──0.00462──→ Node6:Node12 ──0.00650──→ Node6:t14 ──0.19451──→ Node10:Node4 ──0.04495──→ Node10:Node8 ──0.00239──→ Node10:Node9 ──0.00394──→ Node10:Node12 ──0.00553──→ Node10:t14 ──0.08264──→ Node14:Node4 ──0.01910──→ Node14:Node8 ──0.00101──→ Node14:Node9 ──0.00167──→ Node14:Node12 ──0.00235──→ Node14:t14 ──0.02568──→ Node19:Node4 ──0.00593──→ Node19:Node8 ──0.00032──→ Node19:Node9 ──0.00052──→ Node19:Node12 ──0.00073──→ Node19:t14 ──0.03757──→ t3:Node4 ──0.00868──→ t3:Node8 ──0.00046──→ t3:Node9 ──0.00076──→ t3:Node12 ──0.00107──→ t3:t14 
#> 
#> [5] ◎──2.62163──→ Node2:Node4 ──0.60579──→ Node2:Node8 ──0.03219──→ Node2:Node9 ──0.03047──→ Node2:Node11 ──0.05240──→ Node2:Node16 ──0.04478──→ Node2:t20 ──1.41287──→ Node3:Node4 ──0.32648──→ Node3:Node8 ──0.01735──→ Node3:Node9 ──0.01642──→ Node3:Node11 ──0.02824──→ Node3:Node16 ──0.02414──→ Node3:t20 ──0.76523──→ Node5:Node4 ──0.17682──→ Node5:Node8 ──0.00940──→ Node5:Node9 ──0.00889──→ Node5:Node11 ──0.01530──→ Node5:Node16 ──0.01307──→ Node5:t20 ──0.22846──→ Node6:Node4 ──0.05279──→ Node6:Node8 ──0.00280──→ Node6:Node9 ──0.00265──→ Node6:Node11 ──0.00457──→ Node6:Node16 ──0.00390──→ Node6:t20 ──0.19451──→ Node10:Node4 ──0.04495──→ Node10:Node8 ──0.00239──→ Node10:Node9 ──0.00226──→ Node10:Node11 ──0.00389──→ Node10:Node16 ──0.00332──→ Node10:t20 ──0.08264──→ Node14:Node4 ──0.01910──→ Node14:Node8 ──0.00101──→ Node14:Node9 ──0.00096──→ Node14:Node11 ──0.00165──→ Node14:Node16 ──0.00141──→ Node14:t20 ──0.02568──→ Node19:Node4 ──0.00593──→ Node19:Node8 ──0.00032──→ Node19:Node9 ──0.00030──→ Node19:Node11 ──0.00051──→ Node19:Node16 ──0.00044──→ Node19:t20 ──0.03757──→ t4:Node4 ──0.00868──→ t4:Node8 ──0.00046──→ t4:Node9 ──0.00044──→ t4:Node11 ──0.00075──→ t4:Node16 ──0.00064──→ t4:t20 
#> 
#> [6] ◎──2.62163──→ Node2:Node4 ──0.60579──→ Node2:Node8 ──0.03219──→ Node2:Node9 ──0.03047──→ Node2:Node11 ──0.05240──→ Node2:Node16 ──0.04478──→ Node2:t16 ──1.41287──→ Node3:Node4 ──0.32648──→ Node3:Node8 ──0.01735──→ Node3:Node9 ──0.01642──→ Node3:Node11 ──0.02824──→ Node3:Node16 ──0.02414──→ Node3:t16 ──0.76523──→ Node5:Node4 ──0.17682──→ Node5:Node8 ──0.00940──→ Node5:Node9 ──0.00889──→ Node5:Node11 ──0.01530──→ Node5:Node16 ──0.01307──→ Node5:t16 ──0.22846──→ Node6:Node4 ──0.05279──→ Node6:Node8 ──0.00280──→ Node6:Node9 ──0.00265──→ Node6:Node11 ──0.00457──→ Node6:Node16 ──0.00390──→ Node6:t16 ──0.19451──→ Node10:Node4 ──0.04495──→ Node10:Node8 ──0.00239──→ Node10:Node9 ──0.00226──→ Node10:Node11 ──0.00389──→ Node10:Node16 ──0.00332──→ Node10:t16 ──0.04701──→ Node13:Node4 ──0.01086──→ Node13:Node8 ──0.00058──→ Node13:Node9 ──0.00055──→ Node13:Node11 ──0.00094──→ Node13:Node16 ──0.00080──→ Node13:t16 ──0.09889──→ t13:Node4 ──0.02285──→ t13:Node8 ──0.00121──→ t13:Node9 ──0.00115──→ t13:Node11 ──0.00198──→ t13:Node16 ──0.00169──→ t13:t16 
#> 
#> [7] ◎──2.62163──→ Node2:Node4 ──0.60579──→ Node2:Node8 ──0.03219──→ Node2:Node9 ──0.03047──→ Node2:Node11 ──0.09719──→ Node2:t18 ──1.41287──→ Node3:Node4 ──0.32648──→ Node3:Node8 ──0.01735──→ Node3:Node9 ──0.01642──→ Node3:Node11 ──0.05238──→ Node3:t18 ──0.76523──→ Node5:Node4 ──0.17682──→ Node5:Node8 ──0.00940──→ Node5:Node9 ──0.00889──→ Node5:Node11 ──0.02837──→ Node5:t18 ──0.22846──→ Node6:Node4 ──0.05279──→ Node6:Node8 ──0.00280──→ Node6:Node9 ──0.00265──→ Node6:Node11 ──0.00847──→ Node6:t18 ──0.19451──→ Node10:Node4 ──0.04495──→ Node10:Node8 ──0.00239──→ Node10:Node9 ──0.00226──→ Node10:Node11 ──0.00721──→ Node10:t18 ──0.04701──→ Node13:Node4 ──0.01086──→ Node13:Node8 ──0.00058──→ Node13:Node9 ──0.00055──→ Node13:Node11 ──0.00174──→ Node13:t18 ──0.04762──→ Node16:Node4 ──0.01100──→ Node16:Node8 ──0.00058──→ Node16:Node9 ──0.00055──→ Node16:Node11 ──0.00177──→ Node16:t18 ──0.05126──→ t20:Node4 ──0.01185──→ t20:Node8 ──0.00063──→ t20:Node9 ──0.00060──→ t20:Node11 ──0.00190──→ t20:t18 
#> 
#> [8] ◎──2.62163──→ Node2:Node4 ──0.19542──→ Node2:Node6 ──0.50659──→ Node2:Node14 ──0.06362──→ Node2:t15 ──1.41287──→ Node3:Node4 ──0.10532──→ Node3:Node6 ──0.27301──→ Node3:Node14 ──0.03429──→ Node3:t15 ──0.76523──→ Node5:Node4 ──0.05704──→ Node5:Node6 ──0.14787──→ Node5:Node14 ──0.01857──→ Node5:t15 ──0.22846──→ Node6:Node4 ──0.01703──→ Node6:Node6 ──0.04415──→ Node6:Node14 ──0.00554──→ Node6:t15 ──0.19451──→ Node10:Node4 ──0.01450──→ Node10:Node6 ──0.03759──→ Node10:Node14 ──0.00472──→ Node10:t15 ──0.04701──→ Node13:Node4 ──0.00350──→ Node13:Node6 ──0.00908──→ Node13:Node14 ──0.00114──→ Node13:t15 ──0.04762──→ Node16:Node4 ──0.00355──→ Node16:Node6 ──0.00920──→ Node16:Node14 ──0.00116──→ Node16:t15 ──0.01172──→ Node17:Node4 ──0.00087──→ Node17:Node6 ──0.00227──→ Node17:Node14 ──0.00028──→ Node17:t15 ──0.03954──→ t11:Node4 ──0.00295──→ t11:Node6 ──0.00764──→ t11:Node14 ──0.00096──→ t11:t15 
#> 
#> [9] ◎──2.6e+00──→ Node2:Node4 ──2.0e-01──→ Node2:Node6 ──5.1e-01──→ Node2:Node14 ──1.9e-03──→ Node2:Node15 ──2.9e-02──→ Node2:Node17 ──3.2e-02──→ Node2:t10 ──1.4e+00──→ Node3:Node4 ──1.1e-01──→ Node3:Node6 ──2.7e-01──→ Node3:Node14 ──1.0e-03──→ Node3:Node15 ──1.6e-02──→ Node3:Node17 ──1.7e-02──→ Node3:t10 ──7.7e-01──→ Node5:Node4 ──5.7e-02──→ Node5:Node6 ──1.5e-01──→ Node5:Node14 ──5.6e-04──→ Node5:Node15 ──8.6e-03──→ Node5:Node17 ──9.4e-03──→ Node5:t10 ──2.3e-01──→ Node6:Node4 ──1.7e-02──→ Node6:Node6 ──4.4e-02──→ Node6:Node14 ──1.7e-04──→ Node6:Node15 ──2.6e-03──→ Node6:Node17 ──2.8e-03──→ Node6:t10 ──1.9e-01──→ Node10:Node4 ──1.4e-02──→ Node10:Node6 ──3.8e-02──→ Node10:Node14 ──1.4e-04──→ Node10:Node15 ──2.2e-03──→ Node10:Node17 ──2.4e-03──→ Node10:t10 ──4.7e-02──→ Node13:Node4 ──3.5e-03──→ Node13:Node6 ──9.1e-03──→ Node13:Node14 ──3.5e-05──→ Node13:Node15 ──5.3e-04──→ Node13:Node17 ──5.8e-04──→ Node13:t10 ──4.8e-02──→ Node16:Node4 ──3.6e-03──→ Node16:Node6 ──9.2e-03──→ Node16:Node14 ──3.5e-05──→ Node16:Node15 ──5.3e-04──→ Node16:Node17 ──5.9e-04──→ Node16:t10 ──1.2e-02──→ Node17:Node4 ──8.7e-04──→ Node17:Node6 ──2.3e-03──→ Node17:Node14 ──8.6e-06──→ Node17:Node15 ──1.3e-04──→ Node17:Node17 ──1.4e-04──→ Node17:t10 ──4.0e-02──→ t10:Node4 ──2.9e-03──→ t10:Node6 ──7.6e-03──→ t10:Node14 ──2.9e-05──→ t10:Node15 ──4.4e-04──→ t10:Node17 ──4.9e-04──→ t10:t10 
#> 
#> [10] ◎──2.62163──→ Node2:Node4 ──0.19542──→ Node2:Node6 ──0.50659──→ Node2:Node14 ──0.00193──→ Node2:Node15 ──0.02937──→ Node2:Node17 ──0.03233──→ Node2:t11 ──1.41287──→ Node3:Node4 ──0.10532──→ Node3:Node6 ──0.27301──→ Node3:Node14 ──0.00104──→ Node3:Node15 ──0.01583──→ Node3:Node17 ──0.01742──→ Node3:t11 ──0.76523──→ Node5:Node4 ──0.05704──→ Node5:Node6 ──0.14787──→ Node5:Node14 ──0.00056──→ Node5:Node15 ──0.00857──→ Node5:Node17 ──0.00944──→ Node5:t11 ──0.22846──→ Node6:Node4 ──0.01703──→ Node6:Node6 ──0.04415──→ Node6:Node14 ──0.00017──→ Node6:Node15 ──0.00256──→ Node6:Node17 ──0.00282──→ Node6:t11 ──0.16270──→ Node9:Node4 ──0.01213──→ Node9:Node6 ──0.03144──→ Node9:Node14 ──0.00012──→ Node9:Node15 ──0.00182──→ Node9:Node17 ──0.00201──→ Node9:t11 ──0.17771──→ t16:Node4 ──0.01325──→ t16:Node6 ──0.03434──→ t16:Node14 ──0.00013──→ t16:Node15 ──0.00199──→ t16:Node17 ──0.00219──→ t16:t11