Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──0.895──→ Node3:Node2 ──0.405──→ Node3:Node15 ──0.038──→ Node3:t2 ──0.366──→ t8:Node2 ──0.166──→ t8:Node15 ──0.016──→ t8:t2 
#> 
#> [2] ◎──0.8952──→ Node3:Node2 ──0.4048──→ Node3:Node15 ──0.0380──→ Node3:t3 ──0.0282──→ Node4:Node2 ──0.0128──→ Node4:Node15 ──0.0012──→ Node4:t3 ──0.0769──→ Node5:Node2 ──0.0348──→ Node5:Node15 ──0.0033──→ Node5:t3 ──0.1173──→ Node10:Node2 ──0.0530──→ Node10:Node15 ──0.0050──→ Node10:t3 ──0.0984──→ Node15:Node2 ──0.0445──→ Node15:Node15 ──0.0042──→ Node15:t3 ──0.0453──→ t14:Node2 ──0.0205──→ t14:Node15 ──0.0019──→ t14:t3 
#> 
#> [3] ◎──9.0e-01──→ Node3:Node2 ──1.1e-01──→ Node3:Node5 ──2.2e-01──→ Node3:Node9 ──1.0e-01──→ Node3:Node19 ──6.2e-03──→ Node3:t20 ──2.8e-02──→ Node4:Node2 ──3.5e-03──→ Node4:Node5 ──7.0e-03──→ Node4:Node9 ──3.3e-03──→ Node4:Node19 ──1.9e-04──→ Node4:t20 ──7.7e-02──→ Node5:Node2 ──9.6e-03──→ Node5:Node5 ──1.9e-02──→ Node5:Node9 ──8.9e-03──→ Node5:Node19 ──5.3e-04──→ Node5:t20 ──1.2e-01──→ Node10:Node2 ──1.5e-02──→ Node10:Node5 ──2.9e-02──→ Node10:Node9 ──1.4e-02──→ Node10:Node19 ──8.1e-04──→ Node10:t20 ──9.8e-02──→ Node15:Node2 ──1.2e-02──→ Node15:Node5 ──2.4e-02──→ Node15:Node9 ──1.1e-02──→ Node15:Node19 ──6.8e-04──→ Node15:t20 ──3.6e-03──→ Node16:Node2 ──4.5e-04──→ Node16:Node5 ──8.9e-04──→ Node16:Node9 ──4.2e-04──→ Node16:Node19 ──2.5e-05──→ Node16:t20 ──4.2e-02──→ t20:Node2 ──5.2e-03──→ t20:Node5 ──1.0e-02──→ t20:Node9 ──4.8e-03──→ t20:Node19 ──2.9e-04──→ t20:t20 
#> 
#> [4] ◎──9.0e-01──→ Node3:Node2 ──1.1e-01──→ Node3:Node5 ──2.2e-01──→ Node3:Node9 ──1.0e-01──→ Node3:Node19 ──6.2e-03──→ Node3:t19 ──2.8e-02──→ Node4:Node2 ──3.5e-03──→ Node4:Node5 ──7.0e-03──→ Node4:Node9 ──3.3e-03──→ Node4:Node19 ──1.9e-04──→ Node4:t19 ──7.7e-02──→ Node5:Node2 ──9.6e-03──→ Node5:Node5 ──1.9e-02──→ Node5:Node9 ──8.9e-03──→ Node5:Node19 ──5.3e-04──→ Node5:t19 ──1.2e-01──→ Node10:Node2 ──1.5e-02──→ Node10:Node5 ──2.9e-02──→ Node10:Node9 ──1.4e-02──→ Node10:Node19 ──8.1e-04──→ Node10:t19 ──9.8e-02──→ Node15:Node2 ──1.2e-02──→ Node15:Node5 ──2.4e-02──→ Node15:Node9 ──1.1e-02──→ Node15:Node19 ──6.8e-04──→ Node15:t19 ──3.6e-03──→ Node16:Node2 ──4.5e-04──→ Node16:Node5 ──8.9e-04──→ Node16:Node9 ──4.2e-04──→ Node16:Node19 ──2.5e-05──→ Node16:t19 ──4.2e-02──→ t2:Node2 ──5.2e-03──→ t2:Node5 ──1.0e-02──→ t2:Node9 ──4.8e-03──→ t2:Node19 ──2.9e-04──→ t2:t19 
#> 
#> [5] ◎──0.8952──→ Node3:Node2 ──0.1123──→ Node3:Node5 ──0.2207──→ Node3:Node9 ──0.1097──→ Node3:t5 ──0.0282──→ Node4:Node2 ──0.0035──→ Node4:Node5 ──0.0070──→ Node4:Node9 ──0.0035──→ Node4:t5 ──0.0769──→ Node5:Node2 ──0.0096──→ Node5:Node5 ──0.0190──→ Node5:Node9 ──0.0094──→ Node5:t5 ──0.1173──→ Node10:Node2 ──0.0147──→ Node10:Node5 ──0.0289──→ Node10:Node9 ──0.0144──→ Node10:t5 ──0.1437──→ t9:Node2 ──0.0180──→ t9:Node5 ──0.0354──→ t9:Node9 ──0.0176──→ t9:t5 
#> 
#> [6] ◎──0.8952──→ Node3:Node2 ──0.1123──→ Node3:Node5 ──0.1176──→ Node3:Node7 ──0.0384──→ Node3:Node8 ──0.1238──→ Node3:Node13 ──0.0507──→ Node3:t15 ──0.0282──→ Node4:Node2 ──0.0035──→ Node4:Node5 ──0.0037──→ Node4:Node7 ──0.0012──→ Node4:Node8 ──0.0039──→ Node4:Node13 ──0.0016──→ Node4:t15 ──0.0769──→ Node5:Node2 ──0.0096──→ Node5:Node5 ──0.0101──→ Node5:Node7 ──0.0033──→ Node5:Node8 ──0.0106──→ Node5:Node13 ──0.0044──→ Node5:t15 ──0.0438──→ Node7:Node2 ──0.0055──→ Node7:Node5 ──0.0058──→ Node7:Node7 ──0.0019──→ Node7:Node8 ──0.0061──→ Node7:Node13 ──0.0025──→ Node7:t15 ──0.2171──→ t6:Node2 ──0.0272──→ t6:Node5 ──0.0285──→ t6:Node7 ──0.0093──→ t6:Node8 ──0.0300──→ t6:Node13 ──0.0123──→ t6:t15 
#> 
#> [7] ◎──0.8952──→ Node3:Node2 ──0.1123──→ Node3:Node5 ──0.1176──→ Node3:Node7 ──0.0384──→ Node3:Node8 ──0.1238──→ Node3:Node13 ──0.0507──→ Node3:t4 ──0.0282──→ Node4:Node2 ──0.0035──→ Node4:Node5 ──0.0037──→ Node4:Node7 ──0.0012──→ Node4:Node8 ──0.0039──→ Node4:Node13 ──0.0016──→ Node4:t4 ──0.0769──→ Node5:Node2 ──0.0096──→ Node5:Node5 ──0.0101──→ Node5:Node7 ──0.0033──→ Node5:Node8 ──0.0106──→ Node5:Node13 ──0.0044──→ Node5:t4 ──0.0438──→ Node7:Node2 ──0.0055──→ Node7:Node5 ──0.0058──→ Node7:Node7 ──0.0019──→ Node7:Node8 ──0.0061──→ Node7:Node13 ──0.0025──→ Node7:t4 ──0.2171──→ t4:Node2 ──0.0272──→ t4:Node5 ──0.0285──→ t4:Node7 ──0.0093──→ t4:Node8 ──0.0300──→ t4:Node13 ──0.0123──→ t4:t4 
#> 
#> [8] ◎──0.89524──→ Node3:Node2 ──0.11233──→ Node3:Node5 ──0.11756──→ Node3:Node7 ──0.03838──→ Node3:Node8 ──0.08743──→ Node3:Node10 ──0.06659──→ Node3:Node17 ──0.00778──→ Node3:Node18 ──0.01274──→ Node3:t7 ──0.02823──→ Node4:Node2 ──0.00354──→ Node4:Node5 ──0.00371──→ Node4:Node7 ──0.00121──→ Node4:Node8 ──0.00276──→ Node4:Node10 ──0.00210──→ Node4:Node17 ──0.00025──→ Node4:Node18 ──0.00040──→ Node4:t7 ──0.10963──→ Node6:Node2 ──0.01376──→ Node6:Node5 ──0.01440──→ Node6:Node7 ──0.00470──→ Node6:Node8 ──0.01071──→ Node6:Node10 ──0.00815──→ Node6:Node17 ──0.00095──→ Node6:Node18 ──0.00156──→ Node6:t7 ──0.21344──→ Node19:Node2 ──0.02678──→ Node19:Node5 ──0.02803──→ Node19:Node7 ──0.00915──→ Node19:Node8 ──0.02084──→ Node19:Node10 ──0.01588──→ Node19:Node17 ──0.00185──→ Node19:Node18 ──0.00304──→ Node19:t7 ──0.01478──→ t7:Node2 ──0.00185──→ t7:Node5 ──0.00194──→ t7:Node7 ──0.00063──→ t7:Node8 ──0.00144──→ t7:Node10 ──0.00110──→ t7:Node17 ──0.00013──→ t7:Node18 ──0.00021──→ t7:t7 
#> 
#> [9] ◎──0.89524──→ Node3:Node2 ──0.11233──→ Node3:Node5 ──0.11756──→ Node3:Node7 ──0.03838──→ Node3:Node8 ──0.08743──→ Node3:Node10 ──0.06659──→ Node3:Node17 ──0.00778──→ Node3:Node18 ──0.01274──→ Node3:t12 ──0.02823──→ Node4:Node2 ──0.00354──→ Node4:Node5 ──0.00371──→ Node4:Node7 ──0.00121──→ Node4:Node8 ──0.00276──→ Node4:Node10 ──0.00210──→ Node4:Node17 ──0.00025──→ Node4:Node18 ──0.00040──→ Node4:t12 ──0.10963──→ Node6:Node2 ──0.01376──→ Node6:Node5 ──0.01440──→ Node6:Node7 ──0.00470──→ Node6:Node8 ──0.01071──→ Node6:Node10 ──0.00815──→ Node6:Node17 ──0.00095──→ Node6:Node18 ──0.00156──→ Node6:t12 ──0.21344──→ Node19:Node2 ──0.02678──→ Node19:Node5 ──0.02803──→ Node19:Node7 ──0.00915──→ Node19:Node8 ──0.02084──→ Node19:Node10 ──0.01588──→ Node19:Node17 ──0.00185──→ Node19:Node18 ──0.00304──→ Node19:t12 ──0.01478──→ t17:Node2 ──0.00185──→ t17:Node5 ──0.00194──→ t17:Node7 ──0.00063──→ t17:Node8 ──0.00144──→ t17:Node10 ──0.00110──→ t17:Node17 ──0.00013──→ t17:Node18 ──0.00021──→ t17:t12 
#> 
#> [10] ◎──0.89524──→ Node3:Node2 ──0.11233──→ Node3:Node5 ──0.11756──→ Node3:Node7 ──0.03838──→ Node3:Node8 ──0.08743──→ Node3:Node10 ──0.06659──→ Node3:Node17 ──0.02051──→ Node3:t17 ──0.02823──→ Node4:Node2 ──0.00354──→ Node4:Node5 ──0.00371──→ Node4:Node7 ──0.00121──→ Node4:Node8 ──0.00276──→ Node4:Node10 ──0.00210──→ Node4:Node17 ──0.00065──→ Node4:t17 ──0.10963──→ Node6:Node2 ──0.01376──→ Node6:Node5 ──0.01440──→ Node6:Node7 ──0.00470──→ Node6:Node8 ──0.01071──→ Node6:Node10 ──0.00815──→ Node6:Node17 ──0.00251──→ Node6:t17 ──0.04964──→ Node8:Node2 ──0.00623──→ Node8:Node5 ──0.00652──→ Node8:Node7 ──0.00213──→ Node8:Node8 ──0.00485──→ Node8:Node10 ──0.00369──→ Node8:Node17 ──0.00114──→ Node8:t17 ──0.02231──→ Node9:Node2 ──0.00280──→ Node9:Node5 ──0.00293──→ Node9:Node7 ──0.00096──→ Node9:Node8 ──0.00218──→ Node9:Node10 ──0.00166──→ Node9:Node17 ──0.00051──→ Node9:t17 ──0.09447──→ Node14:Node2 ──0.01185──→ Node14:Node5 ──0.01241──→ Node14:Node7 ──0.00405──→ Node14:Node8 ──0.00923──→ Node14:Node10 ──0.00703──→ Node14:Node17 ──0.00216──→ Node14:t17 ──0.06180──→ t10:Node2 ──0.00775──→ t10:Node5 ──0.00812──→ t10:Node7 ──0.00265──→ t10:Node8 ──0.00604──→ t10:Node10 ──0.00460──→ t10:Node17 ──0.00142──→ t10:t17