Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──8.9e+00──→ Node3:Node2 ──2.5e+00──→ Node3:Node4 ──5.5e-01──→ Node3:Node6 ──4.9e-01──→ Node3:Node18 ──7.4e-03──→ Node3:t11 ──5.3e-01──→ Node5:Node2 ──1.5e-01──→ Node5:Node4 ──3.3e-02──→ Node5:Node6 ──2.9e-02──→ Node5:Node18 ──4.4e-04──→ Node5:t11 ──6.5e-01──→ Node14:Node2 ──1.8e-01──→ Node14:Node4 ──4.0e-02──→ Node14:Node6 ──3.6e-02──→ Node14:Node18 ──5.4e-04──→ Node14:t11 ──6.9e-02──→ t17:Node2 ──1.9e-02──→ t17:Node4 ──4.3e-03──→ t17:Node6 ──3.8e-03──→ t17:Node18 ──5.7e-05──→ t17:t11 
#> 
#> [2] ◎──8.9e+00──→ Node3:Node2 ──2.5e+00──→ Node3:Node4 ──5.5e-01──→ Node3:Node6 ──4.9e-01──→ Node3:Node18 ──7.4e-03──→ Node3:t4 ──5.3e-01──→ Node5:Node2 ──1.5e-01──→ Node5:Node4 ──3.3e-02──→ Node5:Node6 ──2.9e-02──→ Node5:Node18 ──4.4e-04──→ Node5:t4 ──6.5e-01──→ Node14:Node2 ──1.8e-01──→ Node14:Node4 ──4.0e-02──→ Node14:Node6 ──3.6e-02──→ Node14:Node18 ──5.4e-04──→ Node14:t4 ──6.9e-02──→ t10:Node2 ──1.9e-02──→ t10:Node4 ──4.3e-03──→ t10:Node6 ──3.8e-03──→ t10:Node18 ──5.7e-05──→ t10:t4 
#> 
#> [3] ◎──8.9e+00──→ Node3:Node2 ──2.5e+00──→ Node3:Node4 ──5.5e-01──→ Node3:Node6 ──5.2e-02──→ Node3:Node8 ──4.4e-01──→ Node3:t9 ──5.3e-01──→ Node5:Node2 ──1.5e-01──→ Node5:Node4 ──3.3e-02──→ Node5:Node6 ──3.1e-03──→ Node5:Node8 ──2.6e-02──→ Node5:t9 ──4.6e-01──→ Node8:Node2 ──1.3e-01──→ Node8:Node4 ──2.8e-02──→ Node8:Node6 ──2.7e-03──→ Node8:Node8 ──2.3e-02──→ Node8:t9 ──1.7e-02──→ Node9:Node2 ──4.7e-03──→ Node9:Node4 ──1.0e-03──→ Node9:Node6 ──9.7e-05──→ Node9:Node8 ──8.3e-04──→ Node9:t9 ──2.5e-01──→ t1:Node2 ──7.0e-02──→ t1:Node4 ──1.5e-02──→ t1:Node6 ──1.4e-03──→ t1:Node8 ──1.2e-02──→ t1:t9 
#> 
#> [4] ◎──8.9e+00──→ Node3:Node2 ──2.5e+00──→ Node3:Node4 ──5.5e-01──→ Node3:Node6 ──5.2e-02──→ Node3:Node8 ──1.3e-01──→ Node3:Node11 ──3.1e-01──→ Node3:t14 ──5.3e-01──→ Node5:Node2 ──1.5e-01──→ Node5:Node4 ──3.3e-02──→ Node5:Node6 ──3.1e-03──→ Node5:Node8 ──7.8e-03──→ Node5:Node11 ──1.9e-02──→ Node5:t14 ──4.6e-01──→ Node8:Node2 ──1.3e-01──→ Node8:Node4 ──2.8e-02──→ Node8:Node6 ──2.7e-03──→ Node8:Node8 ──6.7e-03──→ Node8:Node11 ──1.6e-02──→ Node8:t14 ──1.7e-02──→ Node9:Node2 ──4.7e-03──→ Node9:Node4 ──1.0e-03──→ Node9:Node6 ──9.7e-05──→ Node9:Node8 ──2.5e-04──→ Node9:Node11 ──5.8e-04──→ Node9:t14 ──2.7e-02──→ Node10:Node2 ──7.7e-03──→ Node10:Node4 ──1.7e-03──→ Node10:Node6 ──1.6e-04──→ Node10:Node8 ──4.0e-04──→ Node10:Node11 ──9.5e-04──→ Node10:t14 ──2.2e-01──→ t6:Node2 ──6.2e-02──→ t6:Node4 ──1.4e-02──→ t6:Node6 ──1.3e-03──→ t6:Node8 ──3.3e-03──→ t6:Node11 ──7.8e-03──→ t6:t14 
#> 
#> [5] ◎──8.9e+00──→ Node3:Node2 ──2.5e+00──→ Node3:Node4 ──5.5e-01──→ Node3:Node6 ──5.2e-02──→ Node3:Node8 ──1.3e-01──→ Node3:Node11 ──3.1e-01──→ Node3:t10 ──5.3e-01──→ Node5:Node2 ──1.5e-01──→ Node5:Node4 ──3.3e-02──→ Node5:Node6 ──3.1e-03──→ Node5:Node8 ──7.8e-03──→ Node5:Node11 ──1.9e-02──→ Node5:t10 ──4.6e-01──→ Node8:Node2 ──1.3e-01──→ Node8:Node4 ──2.8e-02──→ Node8:Node6 ──2.7e-03──→ Node8:Node8 ──6.7e-03──→ Node8:Node11 ──1.6e-02──→ Node8:t10 ──1.7e-02──→ Node9:Node2 ──4.7e-03──→ Node9:Node4 ──1.0e-03──→ Node9:Node6 ──9.7e-05──→ Node9:Node8 ──2.5e-04──→ Node9:Node11 ──5.8e-04──→ Node9:t10 ──2.7e-02──→ Node10:Node2 ──7.7e-03──→ Node10:Node4 ──1.7e-03──→ Node10:Node6 ──1.6e-04──→ Node10:Node8 ──4.0e-04──→ Node10:Node11 ──9.5e-04──→ Node10:t10 ──4.7e-02──→ Node12:Node2 ──1.3e-02──→ Node12:Node4 ──2.9e-03──→ Node12:Node6 ──2.7e-04──→ Node12:Node8 ──6.9e-04──→ Node12:Node11 ──1.6e-03──→ Node12:t10 ──3.8e-02──→ Node13:Node2 ──1.1e-02──→ Node13:Node4 ──2.3e-03──→ Node13:Node6 ──2.2e-04──→ Node13:Node8 ──5.6e-04──→ Node13:Node11 ──1.3e-03──→ Node13:t10 ──1.4e-01──→ t14:Node2 ──3.9e-02──→ t14:Node4 ──8.4e-03──→ t14:Node6 ──8.0e-04──→ t14:Node8 ──2.0e-03──→ t14:Node11 ──4.8e-03──→ t14:t10 
#> 
#> [6] ◎──8.9e+00──→ Node3:Node2 ──2.5e+00──→ Node3:Node4 ──6.4e-01──→ Node3:Node10 ──2.0e-01──→ Node3:Node14 ──2.1e-01──→ Node3:t1 ──5.3e-01──→ Node5:Node2 ──1.5e-01──→ Node5:Node4 ──3.8e-02──→ Node5:Node10 ──1.2e-02──→ Node5:Node14 ──1.2e-02──→ Node5:t1 ──4.6e-01──→ Node8:Node2 ──1.3e-01──→ Node8:Node4 ──3.3e-02──→ Node8:Node10 ──1.0e-02──→ Node8:Node14 ──1.1e-02──→ Node8:t1 ──1.7e-02──→ Node9:Node2 ──4.7e-03──→ Node9:Node4 ──1.2e-03──→ Node9:Node10 ──3.7e-04──→ Node9:Node14 ──3.9e-04──→ Node9:t1 ──2.7e-02──→ Node10:Node2 ──7.7e-03──→ Node10:Node4 ──2.0e-03──→ Node10:Node10 ──6.0e-04──→ Node10:Node14 ──6.4e-04──→ Node10:t1 ──4.7e-02──→ Node12:Node2 ──1.3e-02──→ Node12:Node4 ──3.4e-03──→ Node12:Node10 ──1.0e-03──→ Node12:Node14 ──1.1e-03──→ Node12:t1 ──3.8e-02──→ Node13:Node2 ──1.1e-02──→ Node13:Node4 ──2.7e-03──→ Node13:Node10 ──8.4e-04──→ Node13:Node14 ──8.8e-04──→ Node13:t1 ──1.3e-01──→ Node19:Node2 ──3.8e-02──→ Node19:Node4 ──9.6e-03──→ Node19:Node10 ──2.9e-03──→ Node19:Node14 ──3.1e-03──→ Node19:t1 ──3.8e-03──→ t12:Node2 ──1.1e-03──→ t12:Node4 ──2.7e-04──→ t12:Node10 ──8.4e-05──→ t12:Node14 ──8.9e-05──→ t12:t1 
#> 
#> [7] ◎──8.9e+00──→ Node3:Node2 ──2.5e+00──→ Node3:Node4 ──6.4e-01──→ Node3:Node10 ──2.0e-01──→ Node3:Node14 ──1.3e-01──→ Node3:Node16 ──7.5e-02──→ Node3:t17 ──5.3e-01──→ Node5:Node2 ──1.5e-01──→ Node5:Node4 ──3.8e-02──→ Node5:Node10 ──1.2e-02──→ Node5:Node14 ──8.0e-03──→ Node5:Node16 ──4.5e-03──→ Node5:t17 ──4.6e-01──→ Node8:Node2 ──1.3e-01──→ Node8:Node4 ──3.3e-02──→ Node8:Node10 ──1.0e-02──→ Node8:Node14 ──6.8e-03──→ Node8:Node16 ──3.8e-03──→ Node8:t17 ──1.7e-02──→ Node9:Node2 ──4.7e-03──→ Node9:Node4 ──1.2e-03──→ Node9:Node10 ──3.7e-04──→ Node9:Node14 ──2.5e-04──→ Node9:Node16 ──1.4e-04──→ Node9:t17 ──2.7e-02──→ Node10:Node2 ──7.7e-03──→ Node10:Node4 ──2.0e-03──→ Node10:Node10 ──6.0e-04──→ Node10:Node14 ──4.1e-04──→ Node10:Node16 ──2.3e-04──→ Node10:t17 ──4.7e-02──→ Node12:Node2 ──1.3e-02──→ Node12:Node4 ──3.4e-03──→ Node12:Node10 ──1.0e-03──→ Node12:Node14 ──7.0e-04──→ Node12:Node16 ──3.9e-04──→ Node12:t17 ──3.8e-02──→ Node13:Node2 ──1.1e-02──→ Node13:Node4 ──2.7e-03──→ Node13:Node10 ──8.4e-04──→ Node13:Node14 ──5.7e-04──→ Node13:Node16 ──3.2e-04──→ Node13:t17 ──1.3e-01──→ Node19:Node2 ──3.8e-02──→ Node19:Node4 ──9.6e-03──→ Node19:Node10 ──2.9e-03──→ Node19:Node14 ──2.0e-03──→ Node19:Node16 ──1.1e-03──→ Node19:t17 ──3.8e-03──→ t9:Node2 ──1.1e-03──→ t9:Node4 ──2.7e-04──→ t9:Node10 ──8.4e-05──→ t9:Node14 ──5.7e-05──→ t9:Node16 ──3.2e-05──→ t9:t17 
#> 
#> [8] ◎──8.91153──→ Node3:Node2 ──2.51089──→ Node3:Node4 ──0.63918──→ Node3:Node10 ──0.19641──→ Node3:Node14 ──0.13323──→ Node3:Node16 ──0.07454──→ Node3:t6 ──0.53232──→ Node5:Node2 ──0.14998──→ Node5:Node4 ──0.03818──→ Node5:Node10 ──0.01173──→ Node5:Node14 ──0.00796──→ Node5:Node16 ──0.00445──→ Node5:t6 ──0.45652──→ Node8:Node2 ──0.12863──→ Node8:Node4 ──0.03274──→ Node8:Node10 ──0.01006──→ Node8:Node14 ──0.00683──→ Node8:Node16 ──0.00382──→ Node8:t6 ──0.01669──→ Node9:Node2 ──0.00470──→ Node9:Node4 ──0.00120──→ Node9:Node10 ──0.00037──→ Node9:Node14 ──0.00025──→ Node9:Node16 ──0.00014──→ Node9:t6 ──0.02727──→ Node10:Node2 ──0.00768──→ Node10:Node4 ──0.00196──→ Node10:Node10 ──0.00060──→ Node10:Node14 ──0.00041──→ Node10:Node16 ──0.00023──→ Node10:t6 ──0.04680──→ Node12:Node2 ──0.01319──→ Node12:Node4 ──0.00336──→ Node12:Node10 ──0.00103──→ Node12:Node14 ──0.00070──→ Node12:Node16 ──0.00039──→ Node12:t6 ──0.17502──→ t2:Node2 ──0.04931──→ t2:Node4 ──0.01255──→ t2:Node10 ──0.00386──→ t2:Node14 ──0.00262──→ t2:Node16 ──0.00146──→ t2:t6 
#> 
#> [9] ◎──8.9115──→ Node3:Node2 ──2.5109──→ Node3:Node4 ──0.6392──→ Node3:Node10 ──0.2967──→ Node3:Node15 ──0.1075──→ Node3:t5 ──0.5323──→ Node5:Node2 ──0.1500──→ Node5:Node4 ──0.0382──→ Node5:Node10 ──0.0177──→ Node5:Node15 ──0.0064──→ Node5:t5 ──0.4565──→ Node8:Node2 ──0.1286──→ Node8:Node4 ──0.0327──→ Node8:Node10 ──0.0152──→ Node8:Node15 ──0.0055──→ Node8:t5 ──0.2658──→ t4:Node2 ──0.0749──→ t4:Node4 ──0.0191──→ t4:Node10 ──0.0088──→ t4:Node15 ──0.0032──→ t4:t5 
#> 
#> [10] ◎──8.912──→ Node3:Node2 ──2.511──→ Node3:Node4 ──0.639──→ Node3:Node10 ──0.297──→ Node3:Node15 ──0.108──→ Node3:t3 ──1.255──→ t13:Node2 ──0.353──→ t13:Node4 ──0.090──→ t13:Node10 ──0.042──→ t13:Node15 ──0.015──→ t13:t3