Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──1.01981──→ t18:Node2 ──0.09844──→ t18:Node3 ──0.43911──→ t18:Node4 ──0.27657──→ t18:Node18 ──0.02811──→ t18:Node19 ──0.00093──→ t18:t17 
#> 
#> [2] ◎──0.31311──→ Node2:Node2 ──0.03022──→ Node2:Node3 ──0.13482──→ Node2:Node4 ──0.08491──→ Node2:Node18 ──0.00863──→ Node2:Node19 ──0.00028──→ Node2:t9 ──0.40076──→ Node4:Node2 ──0.03869──→ Node4:Node3 ──0.17256──→ Node4:Node4 ──0.10868──→ Node4:Node18 ──0.01105──→ Node4:Node19 ──0.00036──→ Node4:t9 ──0.30593──→ t17:Node2 ──0.02953──→ t17:Node3 ──0.13173──→ t17:Node4 ──0.08297──→ t17:Node18 ──0.00843──→ t17:Node19 ──0.00028──→ t17:t9 
#> 
#> [3] ◎──0.31311──→ Node2:Node2 ──0.03022──→ Node2:Node3 ──0.13482──→ Node2:Node4 ──0.08491──→ Node2:Node18 ──0.00892──→ Node2:t12 ──0.40076──→ Node4:Node2 ──0.03869──→ Node4:Node3 ──0.17256──→ Node4:Node4 ──0.10868──→ Node4:Node18 ──0.01141──→ Node4:t12 ──0.07606──→ Node5:Node2 ──0.00734──→ Node5:Node3 ──0.03275──→ Node5:Node4 ──0.02063──→ Node5:Node18 ──0.00217──→ Node5:t12 ──0.10016──→ Node8:Node2 ──0.00967──→ Node8:Node3 ──0.04313──→ Node8:Node4 ──0.02716──→ Node8:Node18 ──0.00285──→ Node8:t12 ──0.02369──→ Node9:Node2 ──0.00229──→ Node9:Node3 ──0.01020──→ Node9:Node4 ──0.00642──→ Node9:Node18 ──0.00067──→ Node9:t12 ──0.10602──→ t7:Node2 ──0.01023──→ t7:Node3 ──0.04565──→ t7:Node4 ──0.02875──→ t7:Node18 ──0.00302──→ t7:t12 
#> 
#> [4] ◎──0.31311──→ Node2:Node2 ──0.03022──→ Node2:Node3 ──0.13482──→ Node2:Node4 ──0.04728──→ Node2:Node7 ──0.03653──→ Node2:Node17 ──0.01002──→ Node2:t16 ──0.40076──→ Node4:Node2 ──0.03869──→ Node4:Node3 ──0.17256──→ Node4:Node4 ──0.06052──→ Node4:Node7 ──0.04675──→ Node4:Node17 ──0.01282──→ Node4:t16 ──0.07606──→ Node5:Node2 ──0.00734──→ Node5:Node3 ──0.03275──→ Node5:Node4 ──0.01149──→ Node5:Node7 ──0.00887──→ Node5:Node17 ──0.00243──→ Node5:t16 ──0.10016──→ Node8:Node2 ──0.00967──→ Node8:Node3 ──0.04313──→ Node8:Node4 ──0.01513──→ Node8:Node7 ──0.01168──→ Node8:Node17 ──0.00321──→ Node8:t16 ──0.02369──→ Node9:Node2 ──0.00229──→ Node9:Node3 ──0.01020──→ Node9:Node4 ──0.00358──→ Node9:Node7 ──0.00276──→ Node9:Node17 ──0.00076──→ Node9:t16 ──0.10602──→ t19:Node2 ──0.01023──→ t19:Node3 ──0.04565──→ t19:Node4 ──0.01601──→ t19:Node7 ──0.01237──→ t19:Node17 ──0.00339──→ t19:t16 
#> 
#> [5] ◎──0.3131──→ Node2:Node2 ──0.0302──→ Node2:Node3 ──0.1348──→ Node2:Node4 ──0.0473──→ Node2:Node7 ──0.0365──→ Node2:Node17 ──0.0100──→ Node2:t11 ──0.4008──→ Node4:Node2 ──0.0387──→ Node4:Node3 ──0.1726──→ Node4:Node4 ──0.0605──→ Node4:Node7 ──0.0468──→ Node4:Node17 ──0.0128──→ Node4:t11 ──0.0761──→ Node5:Node2 ──0.0073──→ Node5:Node3 ──0.0327──→ Node5:Node4 ──0.0115──→ Node5:Node7 ──0.0089──→ Node5:Node17 ──0.0024──→ Node5:t11 ──0.1002──→ Node8:Node2 ──0.0097──→ Node8:Node3 ──0.0431──→ Node8:Node4 ──0.0151──→ Node8:Node7 ──0.0117──→ Node8:Node17 ──0.0032──→ Node8:t11 ──0.1297──→ t20:Node2 ──0.0125──→ t20:Node3 ──0.0559──→ t20:Node4 ──0.0196──→ t20:Node7 ──0.0151──→ t20:Node17 ──0.0042──→ t20:t11 
#> 
#> [6] ◎──0.31311──→ Node2:Node2 ──0.03022──→ Node2:Node3 ──0.13482──→ Node2:Node4 ──0.04728──→ Node2:Node7 ──0.00877──→ Node2:Node8 ──0.03777──→ Node2:t3 ──0.40076──→ Node4:Node2 ──0.03869──→ Node4:Node3 ──0.17256──→ Node4:Node4 ──0.06052──→ Node4:Node7 ──0.01123──→ Node4:Node8 ──0.04835──→ Node4:t3 ──0.07606──→ Node5:Node2 ──0.00734──→ Node5:Node3 ──0.03275──→ Node5:Node4 ──0.01149──→ Node5:Node7 ──0.00213──→ Node5:Node8 ──0.00918──→ Node5:t3 ──0.08243──→ Node7:Node2 ──0.00796──→ Node7:Node3 ──0.03549──→ Node7:Node4 ──0.01245──→ Node7:Node7 ──0.00231──→ Node7:Node8 ──0.00994──→ Node7:t3 ──0.05906──→ Node10:Node2 ──0.00570──→ Node10:Node3 ──0.02543──→ Node10:Node4 ──0.00892──→ Node10:Node7 ──0.00165──→ Node10:Node8 ──0.00712──→ Node10:t3 ──0.03046──→ Node13:Node2 ──0.00294──→ Node13:Node3 ──0.01311──→ Node13:Node4 ──0.00460──→ Node13:Node7 ──0.00085──→ Node13:Node8 ──0.00367──→ Node13:t3 ──0.00790──→ Node14:Node2 ──0.00076──→ Node14:Node3 ──0.00340──→ Node14:Node4 ──0.00119──→ Node14:Node7 ──0.00022──→ Node14:Node8 ──0.00095──→ Node14:t3 ──0.05003──→ t5:Node2 ──0.00483──→ t5:Node3 ──0.02154──→ t5:Node4 ──0.00755──→ t5:Node7 ──0.00140──→ t5:Node8 ──0.00604──→ t5:t3 
#> 
#> [7] ◎──0.31311──→ Node2:Node2 ──0.03022──→ Node2:Node3 ──0.13482──→ Node2:Node4 ──0.04728──→ Node2:Node7 ──0.00877──→ Node2:Node8 ──0.00862──→ Node2:Node10 ──0.02916──→ Node2:t10 ──0.40076──→ Node4:Node2 ──0.03869──→ Node4:Node3 ──0.17256──→ Node4:Node4 ──0.06052──→ Node4:Node7 ──0.01123──→ Node4:Node8 ──0.01103──→ Node4:Node10 ──0.03732──→ Node4:t10 ──0.07606──→ Node5:Node2 ──0.00734──→ Node5:Node3 ──0.03275──→ Node5:Node4 ──0.01149──→ Node5:Node7 ──0.00213──→ Node5:Node8 ──0.00209──→ Node5:Node10 ──0.00708──→ Node5:t10 ──0.08243──→ Node7:Node2 ──0.00796──→ Node7:Node3 ──0.03549──→ Node7:Node4 ──0.01245──→ Node7:Node7 ──0.00231──→ Node7:Node8 ──0.00227──→ Node7:Node10 ──0.00768──→ Node7:t10 ──0.05906──→ Node10:Node2 ──0.00570──→ Node10:Node3 ──0.02543──→ Node10:Node4 ──0.00892──→ Node10:Node7 ──0.00165──→ Node10:Node8 ──0.00163──→ Node10:Node10 ──0.00550──→ Node10:t10 ──0.03046──→ Node13:Node2 ──0.00294──→ Node13:Node3 ──0.01311──→ Node13:Node4 ──0.00460──→ Node13:Node7 ──0.00085──→ Node13:Node8 ──0.00084──→ Node13:Node10 ──0.00284──→ Node13:t10 ──0.00790──→ Node14:Node2 ──0.00076──→ Node14:Node3 ──0.00340──→ Node14:Node4 ──0.00119──→ Node14:Node7 ──0.00022──→ Node14:Node8 ──0.00022──→ Node14:Node10 ──0.00074──→ Node14:t10 ──0.03133──→ Node17:Node2 ──0.00302──→ Node17:Node3 ──0.01349──→ Node17:Node4 ──0.00473──→ Node17:Node7 ──0.00088──→ Node17:Node8 ──0.00086──→ Node17:Node10 ──0.00292──→ Node17:t10 ──0.01870──→ t3:Node2 ──0.00181──→ t3:Node3 ──0.00805──→ t3:Node4 ──0.00282──→ t3:Node7 ──0.00052──→ t3:Node8 ──0.00051──→ t3:Node10 ──0.00174──→ t3:t10 
#> 
#> [8] ◎──3.1e-01──→ Node2:Node2 ──3.0e-02──→ Node2:Node3 ──1.3e-01──→ Node2:Node4 ──4.7e-02──→ Node2:Node7 ──8.8e-03──→ Node2:Node8 ──8.6e-03──→ Node2:Node10 ──2.3e-03──→ Node2:Node11 ──2.7e-02──→ Node2:t5 ──4.0e-01──→ Node4:Node2 ──3.9e-02──→ Node4:Node3 ──1.7e-01──→ Node4:Node4 ──6.1e-02──→ Node4:Node7 ──1.1e-02──→ Node4:Node8 ──1.1e-02──→ Node4:Node10 ──2.9e-03──→ Node4:Node11 ──3.4e-02──→ Node4:t5 ──7.6e-02──→ Node5:Node2 ──7.3e-03──→ Node5:Node3 ──3.3e-02──→ Node5:Node4 ──1.1e-02──→ Node5:Node7 ──2.1e-03──→ Node5:Node8 ──2.1e-03──→ Node5:Node10 ──5.6e-04──→ Node5:Node11 ──6.5e-03──→ Node5:t5 ──8.2e-02──→ Node7:Node2 ──8.0e-03──→ Node7:Node3 ──3.5e-02──→ Node7:Node4 ──1.2e-02──→ Node7:Node7 ──2.3e-03──→ Node7:Node8 ──2.3e-03──→ Node7:Node10 ──6.1e-04──→ Node7:Node11 ──7.1e-03──→ Node7:t5 ──5.9e-02──→ Node10:Node2 ──5.7e-03──→ Node10:Node3 ──2.5e-02──→ Node10:Node4 ──8.9e-03──→ Node10:Node7 ──1.7e-03──→ Node10:Node8 ──1.6e-03──→ Node10:Node10 ──4.3e-04──→ Node10:Node11 ──5.1e-03──→ Node10:t5 ──3.0e-02──→ Node13:Node2 ──2.9e-03──→ Node13:Node3 ──1.3e-02──→ Node13:Node4 ──4.6e-03──→ Node13:Node7 ──8.5e-04──→ Node13:Node8 ──8.4e-04──→ Node13:Node10 ──2.2e-04──→ Node13:Node11 ──2.6e-03──→ Node13:t5 ──7.9e-03──→ Node14:Node2 ──7.6e-04──→ Node14:Node3 ──3.4e-03──→ Node14:Node4 ──1.2e-03──→ Node14:Node7 ──2.2e-04──→ Node14:Node8 ──2.2e-04──→ Node14:Node10 ──5.8e-05──→ Node14:Node11 ──6.8e-04──→ Node14:t5 ──3.1e-02──→ Node17:Node2 ──3.0e-03──→ Node17:Node3 ──1.3e-02──→ Node17:Node4 ──4.7e-03──→ Node17:Node7 ──8.8e-04──→ Node17:Node8 ──8.6e-04──→ Node17:Node10 ──2.3e-04──→ Node17:Node11 ──2.7e-03──→ Node17:t5 ──1.9e-02──→ t1:Node2 ──1.8e-03──→ t1:Node3 ──8.1e-03──→ t1:Node4 ──2.8e-03──→ t1:Node7 ──5.2e-04──→ t1:Node8 ──5.1e-04──→ t1:Node10 ──1.4e-04──→ t1:Node11 ──1.6e-03──→ t1:t5 
#> 
#> [9] ◎──0.31311──→ Node2:Node2 ──0.03022──→ Node2:Node3 ──0.13482──→ Node2:Node4 ──0.04728──→ Node2:Node7 ──0.00877──→ Node2:Node8 ──0.00862──→ Node2:Node10 ──0.00230──→ Node2:Node11 ──0.01044──→ Node2:Node14 ──0.01642──→ Node2:t2 ──0.40076──→ Node4:Node2 ──0.03869──→ Node4:Node3 ──0.17256──→ Node4:Node4 ──0.06052──→ Node4:Node7 ──0.01123──→ Node4:Node8 ──0.01103──→ Node4:Node10 ──0.00294──→ Node4:Node11 ──0.01336──→ Node4:Node14 ──0.02101──→ Node4:t2 ──0.07606──→ Node5:Node2 ──0.00734──→ Node5:Node3 ──0.03275──→ Node5:Node4 ──0.01149──→ Node5:Node7 ──0.00213──→ Node5:Node8 ──0.00209──→ Node5:Node10 ──0.00056──→ Node5:Node11 ──0.00254──→ Node5:Node14 ──0.00399──→ Node5:t2 ──0.08243──→ Node7:Node2 ──0.00796──→ Node7:Node3 ──0.03549──→ Node7:Node4 ──0.01245──→ Node7:Node7 ──0.00231──→ Node7:Node8 ──0.00227──→ Node7:Node10 ──0.00061──→ Node7:Node11 ──0.00275──→ Node7:Node14 ──0.00432──→ Node7:t2 ──0.05906──→ Node10:Node2 ──0.00570──→ Node10:Node3 ──0.02543──→ Node10:Node4 ──0.00892──→ Node10:Node7 ──0.00165──→ Node10:Node8 ──0.00163──→ Node10:Node10 ──0.00043──→ Node10:Node11 ──0.00197──→ Node10:Node14 ──0.00310──→ Node10:t2 ──0.03046──→ Node13:Node2 ──0.00294──→ Node13:Node3 ──0.01311──→ Node13:Node4 ──0.00460──→ Node13:Node7 ──0.00085──→ Node13:Node8 ──0.00084──→ Node13:Node10 ──0.00022──→ Node13:Node11 ──0.00102──→ Node13:Node14 ──0.00160──→ Node13:t2 ──0.03651──→ Node16:Node2 ──0.00352──→ Node16:Node3 ──0.01572──→ Node16:Node4 ──0.00551──→ Node16:Node7 ──0.00102──→ Node16:Node8 ──0.00101──→ Node16:Node10 ──0.00027──→ Node16:Node11 ──0.00122──→ Node16:Node14 ──0.00191──→ Node16:t2 ──0.02142──→ t4:Node2 ──0.00207──→ t4:Node3 ──0.00922──→ t4:Node4 ──0.00323──→ t4:Node7 ──0.00060──→ t4:Node8 ──0.00059──→ t4:Node10 ──0.00016──→ t4:Node11 ──0.00071──→ t4:Node14 ──0.00112──→ t4:t2 
#> 
#> [10] ◎──3.1e-01──→ Node2:Node2 ──3.0e-02──→ Node2:Node3 ──1.3e-01──→ Node2:Node4 ──4.7e-02──→ Node2:Node7 ──8.8e-03──→ Node2:Node8 ──8.6e-03──→ Node2:Node10 ──2.3e-03──→ Node2:Node11 ──1.0e-02──→ Node2:Node14 ──1.6e-02──→ Node2:t20 ──4.0e-01──→ Node4:Node2 ──3.9e-02──→ Node4:Node3 ──1.7e-01──→ Node4:Node4 ──6.1e-02──→ Node4:Node7 ──1.1e-02──→ Node4:Node8 ──1.1e-02──→ Node4:Node10 ──2.9e-03──→ Node4:Node11 ──1.3e-02──→ Node4:Node14 ──2.1e-02──→ Node4:t20 ──7.6e-02──→ Node5:Node2 ──7.3e-03──→ Node5:Node3 ──3.3e-02──→ Node5:Node4 ──1.1e-02──→ Node5:Node7 ──2.1e-03──→ Node5:Node8 ──2.1e-03──→ Node5:Node10 ──5.6e-04──→ Node5:Node11 ──2.5e-03──→ Node5:Node14 ──4.0e-03──→ Node5:t20 ──8.2e-02──→ Node7:Node2 ──8.0e-03──→ Node7:Node3 ──3.5e-02──→ Node7:Node4 ──1.2e-02──→ Node7:Node7 ──2.3e-03──→ Node7:Node8 ──2.3e-03──→ Node7:Node10 ──6.1e-04──→ Node7:Node11 ──2.7e-03──→ Node7:Node14 ──4.3e-03──→ Node7:t20 ──5.9e-02──→ Node10:Node2 ──5.7e-03──→ Node10:Node3 ──2.5e-02──→ Node10:Node4 ──8.9e-03──→ Node10:Node7 ──1.7e-03──→ Node10:Node8 ──1.6e-03──→ Node10:Node10 ──4.3e-04──→ Node10:Node11 ──2.0e-03──→ Node10:Node14 ──3.1e-03──→ Node10:t20 ──3.0e-02──→ Node13:Node2 ──2.9e-03──→ Node13:Node3 ──1.3e-02──→ Node13:Node4 ──4.6e-03──→ Node13:Node7 ──8.5e-04──→ Node13:Node8 ──8.4e-04──→ Node13:Node10 ──2.2e-04──→ Node13:Node11 ──1.0e-03──→ Node13:Node14 ──1.6e-03──→ Node13:t20 ──3.7e-02──→ Node16:Node2 ──3.5e-03──→ Node16:Node3 ──1.6e-02──→ Node16:Node4 ──5.5e-03──→ Node16:Node7 ──1.0e-03──→ Node16:Node8 ──1.0e-03──→ Node16:Node10 ──2.7e-04──→ Node16:Node11 ──1.2e-03──→ Node16:Node14 ──1.9e-03──→ Node16:t20 ──1.9e-02──→ Node19:Node2 ──1.8e-03──→ Node19:Node3 ──8.0e-03──→ Node19:Node4 ──2.8e-03──→ Node19:Node7 ──5.2e-04──→ Node19:Node8 ──5.1e-04──→ Node19:Node10 ──1.4e-04──→ Node19:Node11 ──6.2e-04──→ Node19:Node14 ──9.7e-04──→ Node19:t20 ──2.9e-03──→ t2:Node2 ──2.8e-04──→ t2:Node3 ──1.2e-03──→ t2:Node4 ──4.3e-04──→ t2:Node7 ──8.0e-05──→ t2:Node8 ──7.9e-05──→ t2:Node10 ──2.1e-05──→ t2:Node11 ──9.6e-05──→ t2:Node14 ──1.5e-04──→ t2:t20