Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──0.379──→ Node3:Node2 ──0.307──→ Node3:Node6 ──0.233──→ Node3:t13 ──0.223──→ Node7:Node2 ──0.180──→ Node7:Node6 ──0.137──→ Node7:t13 ──0.041──→ t19:Node2 ──0.033──→ t19:Node6 ──0.025──→ t19:t13 
#> 
#> [2] ◎──0.3790──→ Node3:Node2 ──0.3067──→ Node3:Node6 ──0.1690──→ Node3:Node12 ──0.0396──→ Node3:Node16 ──0.0248──→ Node3:t18 ──0.2229──→ Node7:Node2 ──0.1804──→ Node7:Node6 ──0.0994──→ Node7:Node12 ──0.0233──→ Node7:Node16 ──0.0146──→ Node7:t18 ──0.0163──→ Node9:Node2 ──0.0132──→ Node9:Node6 ──0.0073──→ Node9:Node12 ──0.0017──→ Node9:Node16 ──0.0011──→ Node9:t18 ──0.0246──→ t15:Node2 ──0.0199──→ t15:Node6 ──0.0110──→ t15:Node12 ──0.0026──→ t15:Node16 ──0.0016──→ t15:t18 
#> 
#> [3] ◎──0.3790──→ Node3:Node2 ──0.3067──→ Node3:Node6 ──0.1690──→ Node3:Node12 ──0.0396──→ Node3:Node16 ──0.0248──→ Node3:t19 ──0.2229──→ Node7:Node2 ──0.1804──→ Node7:Node6 ──0.0994──→ Node7:Node12 ──0.0233──→ Node7:Node16 ──0.0146──→ Node7:t19 ──0.0163──→ Node9:Node2 ──0.0132──→ Node9:Node6 ──0.0073──→ Node9:Node12 ──0.0017──→ Node9:Node16 ──0.0011──→ Node9:t19 ──0.0246──→ t13:Node2 ──0.0199──→ t13:Node6 ──0.0110──→ t13:Node12 ──0.0026──→ t13:Node16 ──0.0016──→ t13:t19 
#> 
#> [4] ◎──3.8e-01──→ Node3:Node2 ──3.1e-01──→ Node3:Node6 ──1.7e-01──→ Node3:Node12 ──6.2e-02──→ Node3:Node19 ──2.0e-03──→ Node3:t7 ──2.6e-03──→ Node4:Node2 ──2.1e-03──→ Node4:Node6 ──1.2e-03──→ Node4:Node12 ──4.3e-04──→ Node4:Node19 ──1.4e-05──→ Node4:t7 ──8.5e-02──→ Node5:Node2 ──6.9e-02──→ Node5:Node6 ──3.8e-02──→ Node5:Node12 ──1.4e-02──→ Node5:Node19 ──4.4e-04──→ Node5:t7 ──7.8e-02──→ Node6:Node2 ──6.3e-02──→ Node6:Node6 ──3.5e-02──→ Node6:Node12 ──1.3e-02──→ Node6:Node19 ──4.1e-04──→ Node6:t7 ──9.8e-02──→ t6:Node2 ──8.0e-02──→ t6:Node6 ──4.4e-02──→ t6:Node12 ──1.6e-02──→ t6:Node19 ──5.1e-04──→ t6:t7 
#> 
#> [5] ◎──3.8e-01──→ Node3:Node2 ──3.1e-01──→ Node3:Node6 ──1.7e-01──→ Node3:Node12 ──6.2e-02──→ Node3:Node19 ──2.0e-03──→ Node3:t15 ──2.6e-03──→ Node4:Node2 ──2.1e-03──→ Node4:Node6 ──1.2e-03──→ Node4:Node12 ──4.3e-04──→ Node4:Node19 ──1.4e-05──→ Node4:t15 ──8.5e-02──→ Node5:Node2 ──6.9e-02──→ Node5:Node6 ──3.8e-02──→ Node5:Node12 ──1.4e-02──→ Node5:Node19 ──4.4e-04──→ Node5:t15 ──7.8e-02──→ Node6:Node2 ──6.3e-02──→ Node6:Node6 ──3.5e-02──→ Node6:Node12 ──1.3e-02──→ Node6:Node19 ──4.1e-04──→ Node6:t15 ──7.8e-02──→ Node13:Node2 ──6.3e-02──→ Node13:Node6 ──3.5e-02──→ Node13:Node12 ──1.3e-02──→ Node13:Node19 ──4.1e-04──→ Node13:t15 ──2.1e-02──→ t5:Node2 ──1.7e-02──→ t5:Node6 ──9.2e-03──→ t5:Node12 ──3.4e-03──→ t5:Node19 ──1.1e-04──→ t5:t15 
#> 
#> [6] ◎──0.3790──→ Node3:Node2 ──0.1797──→ Node3:Node3 ──0.1902──→ Node3:Node8 ──0.1700──→ Node3:t16 ──0.0026──→ Node4:Node2 ──0.0012──→ Node4:Node3 ──0.0013──→ Node4:Node8 ──0.0012──→ Node4:t16 ──0.0847──→ Node5:Node2 ──0.0402──→ Node5:Node3 ──0.0425──→ Node5:Node8 ──0.0380──→ Node5:t16 ──0.0782──→ Node6:Node2 ──0.0371──→ Node6:Node3 ──0.0393──→ Node6:Node8 ──0.0351──→ Node6:t16 ──0.0777──→ Node13:Node2 ──0.0369──→ Node13:Node3 ──0.0390──→ Node13:Node8 ──0.0349──→ Node13:t16 ──0.0205──→ t20:Node2 ──0.0097──→ t20:Node3 ──0.0103──→ t20:Node8 ──0.0092──→ t20:t16 
#> 
#> [7] ◎──3.8e-01──→ Node3:Node2 ──1.8e-01──→ Node3:Node3 ──1.9e-01──→ Node3:Node8 ──5.1e-03──→ Node3:Node9 ──2.2e-02──→ Node3:Node10 ──1.3e-01──→ Node3:Node17 ──8.1e-03──→ Node3:t12 ──2.6e-03──→ Node4:Node2 ──1.2e-03──→ Node4:Node3 ──1.3e-03──→ Node4:Node8 ──3.5e-05──→ Node4:Node9 ──1.5e-04──→ Node4:Node10 ──9.2e-04──→ Node4:Node17 ──5.6e-05──→ Node4:t12 ──8.5e-02──→ Node5:Node2 ──4.0e-02──→ Node5:Node3 ──4.3e-02──→ Node5:Node8 ──1.1e-03──→ Node5:Node9 ──5.0e-03──→ Node5:Node10 ──3.0e-02──→ Node5:Node17 ──1.8e-03──→ Node5:t12 ──1.8e-01──→ t16:Node2 ──8.4e-02──→ t16:Node3 ──8.9e-02──→ t16:Node8 ──2.4e-03──→ t16:Node9 ──1.0e-02──→ t16:Node10 ──6.3e-02──→ t16:Node17 ──3.8e-03──→ t16:t12 
#> 
#> [8] ◎──3.8e-01──→ Node3:Node2 ──1.8e-01──→ Node3:Node3 ──1.9e-01──→ Node3:Node8 ──5.1e-03──→ Node3:Node9 ──2.2e-02──→ Node3:Node10 ──1.3e-01──→ Node3:Node17 ──8.1e-03──→ Node3:t6 ──2.6e-03──→ Node4:Node2 ──1.2e-03──→ Node4:Node3 ──1.3e-03──→ Node4:Node8 ──3.5e-05──→ Node4:Node9 ──1.5e-04──→ Node4:Node10 ──9.2e-04──→ Node4:Node17 ──5.6e-05──→ Node4:t6 ──2.6e-01──→ t10:Node2 ──1.2e-01──→ t10:Node3 ──1.3e-01──→ t10:Node8 ──3.5e-03──→ t10:Node9 ──1.5e-02──→ t10:Node10 ──9.3e-02──→ t10:Node17 ──5.6e-03──→ t10:t6 
#> 
#> [9] ◎──3.1e-01──→ Node2:Node2 ──1.5e-01──→ Node2:Node3 ──1.5e-01──→ Node2:Node8 ──4.1e-03──→ Node2:Node9 ──1.8e-02──→ Node2:Node10 ──3.3e-02──→ Node2:Node11 ──8.2e-02──→ Node2:t8 ──3.1e-01──→ Node11:Node2 ──1.5e-01──→ Node11:Node3 ──1.6e-01──→ Node11:Node8 ──4.2e-03──→ Node11:Node9 ──1.9e-02──→ Node11:Node10 ──3.4e-02──→ Node11:Node11 ──8.4e-02──→ Node11:t8 ──1.1e-03──→ Node12:Node2 ──5.0e-04──→ Node12:Node3 ──5.3e-04──→ Node12:Node8 ──1.4e-05──→ Node12:Node9 ──6.3e-05──→ Node12:Node10 ──1.2e-04──→ Node12:Node11 ──2.8e-04──→ Node12:t8 ──2.1e-02──→ t18:Node2 ──1.0e-02──→ t18:Node3 ──1.1e-02──→ t18:Node8 ──2.8e-04──→ t18:Node9 ──1.3e-03──→ t18:Node10 ──2.3e-03──→ t18:Node11 ──5.7e-03──→ t18:t8 
#> 
#> [10] ◎──3.1e-01──→ Node2:Node2 ──1.5e-01──→ Node2:Node3 ──1.5e-01──→ Node2:Node8 ──4.1e-03──→ Node2:Node9 ──1.8e-02──→ Node2:Node10 ──3.3e-02──→ Node2:Node11 ──3.3e-02──→ Node2:Node13 ──4.9e-02──→ Node2:t20 ──3.1e-01──→ Node11:Node2 ──1.5e-01──→ Node11:Node3 ──1.6e-01──→ Node11:Node8 ──4.2e-03──→ Node11:Node9 ──1.9e-02──→ Node11:Node10 ──3.4e-02──→ Node11:Node11 ──3.4e-02──→ Node11:Node13 ──5.0e-02──→ Node11:t20 ──1.1e-03──→ Node12:Node2 ──5.0e-04──→ Node12:Node3 ──5.3e-04──→ Node12:Node8 ──1.4e-05──→ Node12:Node9 ──6.3e-05──→ Node12:Node10 ──1.2e-04──→ Node12:Node11 ──1.1e-04──→ Node12:Node13 ──1.7e-04──→ Node12:t20 ──3.1e-03──→ Node14:Node2 ──1.5e-03──→ Node14:Node3 ──1.6e-03──→ Node14:Node8 ──4.2e-05──→ Node14:Node9 ──1.8e-04──→ Node14:Node10 ──3.4e-04──→ Node14:Node11 ──3.4e-04──→ Node14:Node13 ──5.0e-04──→ Node14:t20 ──4.5e-03──→ Node16:Node2 ──2.2e-03──→ Node16:Node3 ──2.3e-03──→ Node16:Node8 ──6.1e-05──→ Node16:Node9 ──2.7e-04──→ Node16:Node10 ──4.9e-04──→ Node16:Node11 ──4.9e-04──→ Node16:Node13 ──7.2e-04──→ Node16:t20 ──1.4e-02──→ t1:Node2 ──6.5e-03──→ t1:Node3 ──6.8e-03──→ t1:Node8 ──1.8e-04──→ t1:Node9 ──8.1e-04──→ t1:Node10 ──1.5e-03──→ t1:Node11 ──1.5e-03──→ t1:Node13 ──2.2e-03──→ t1:t20