Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎── 1.24──→ t8:Node2 ── 0.30──→ t8:Node3 ── 0.41──→ t8:Node13 ── 0.15──→ t8:t17 
#> 
#> [2] ◎──0.8324──→ Node2:Node2 ──0.2042──→ Node2:Node3 ──0.2732──→ Node2:Node13 ──0.0986──→ Node2:t4 ──0.0447──→ Node3:Node2 ──0.0110──→ Node3:Node3 ──0.0147──→ Node3:Node13 ──0.0053──→ Node3:t4 ──0.0457──→ Node4:Node2 ──0.0112──→ Node4:Node3 ──0.0150──→ Node4:Node13 ──0.0054──→ Node4:t4 ──0.1639──→ Node6:Node2 ──0.0402──→ Node6:Node3 ──0.0538──→ Node6:Node13 ──0.0194──→ Node6:t4 ──0.1507──→ t7:Node2 ──0.0370──→ t7:Node3 ──0.0495──→ t7:Node13 ──0.0178──→ t7:t4 
#> 
#> [3] ◎──0.8324──→ Node2:Node2 ──0.2042──→ Node2:Node3 ──0.2702──→ Node2:Node11 ──0.0276──→ Node2:Node14 ──0.0740──→ Node2:t20 ──0.0447──→ Node3:Node2 ──0.0110──→ Node3:Node3 ──0.0145──→ Node3:Node11 ──0.0015──→ Node3:Node14 ──0.0040──→ Node3:t20 ──0.0457──→ Node4:Node2 ──0.0112──→ Node4:Node3 ──0.0148──→ Node4:Node11 ──0.0015──→ Node4:Node14 ──0.0041──→ Node4:t20 ──0.1639──→ Node6:Node2 ──0.0402──→ Node6:Node3 ──0.0532──→ Node6:Node11 ──0.0054──→ Node6:Node14 ──0.0146──→ Node6:t20 ──0.1199──→ Node11:Node2 ──0.0294──→ Node11:Node3 ──0.0389──→ Node11:Node11 ──0.0040──→ Node11:Node14 ──0.0107──→ Node11:t20 ──0.0308──→ t14:Node2 ──0.0076──→ t14:Node3 ──0.0100──→ t14:Node11 ──0.0010──→ t14:Node14 ──0.0027──→ t14:t20 
#> 
#> [4] ◎──8.3e-01──→ Node2:Node2 ──2.0e-01──→ Node2:Node3 ──2.7e-01──→ Node2:Node11 ──2.8e-02──→ Node2:Node14 ──7.4e-02──→ Node2:t11 ──4.5e-02──→ Node3:Node2 ──1.1e-02──→ Node3:Node3 ──1.5e-02──→ Node3:Node11 ──1.5e-03──→ Node3:Node14 ──4.0e-03──→ Node3:t11 ──4.6e-02──→ Node4:Node2 ──1.1e-02──→ Node4:Node3 ──1.5e-02──→ Node4:Node11 ──1.5e-03──→ Node4:Node14 ──4.1e-03──→ Node4:t11 ──1.6e-01──→ Node6:Node2 ──4.0e-02──→ Node6:Node3 ──5.3e-02──→ Node6:Node11 ──5.4e-03──→ Node6:Node14 ──1.5e-02──→ Node6:t11 ──1.2e-01──→ Node11:Node2 ──2.9e-02──→ Node11:Node3 ──3.9e-02──→ Node11:Node11 ──4.0e-03──→ Node11:Node14 ──1.1e-02──→ Node11:t11 ──1.1e-03──→ Node12:Node2 ──2.8e-04──→ Node12:Node3 ──3.7e-04──→ Node12:Node11 ──3.8e-05──→ Node12:Node14 ──1.0e-04──→ Node12:t11 ──3.0e-02──→ t4:Node2 ──7.3e-03──→ t4:Node3 ──9.6e-03──→ t4:Node11 ──9.8e-04──→ t4:Node14 ──2.6e-03──→ t4:t11 
#> 
#> [5] ◎──0.83244──→ Node2:Node2 ──0.20415──→ Node2:Node3 ──0.27017──→ Node2:Node11 ──0.10159──→ Node2:t1 ──0.04473──→ Node3:Node2 ──0.01097──→ Node3:Node3 ──0.01452──→ Node3:Node11 ──0.00546──→ Node3:t1 ──0.04569──→ Node4:Node2 ──0.01120──→ Node4:Node3 ──0.01483──→ Node4:Node11 ──0.00558──→ Node4:t1 ──0.16391──→ Node6:Node2 ──0.04020──→ Node6:Node3 ──0.05320──→ Node6:Node11 ──0.02000──→ Node6:t1 ──0.11991──→ Node11:Node2 ──0.02941──→ Node11:Node3 ──0.03892──→ Node11:Node11 ──0.01463──→ Node11:t1 ──0.00114──→ Node12:Node2 ──0.00028──→ Node12:Node3 ──0.00037──→ Node12:Node11 ──0.00014──→ Node12:t1 ──0.00640──→ Node14:Node2 ──0.00157──→ Node14:Node3 ──0.00208──→ Node14:Node11 ──0.00078──→ Node14:t1 ──0.02328──→ t2:Node2 ──0.00571──→ t2:Node3 ──0.00756──→ t2:Node11 ──0.00284──→ t2:t1 
#> 
#> [6] ◎──8.3e-01──→ Node2:Node2 ──4.6e-01──→ Node2:Node10 ──1.1e-01──→ Node2:Node19 ──6.5e-03──→ Node2:t6 ──4.5e-02──→ Node3:Node2 ──2.4e-02──→ Node3:Node10 ──6.1e-03──→ Node3:Node19 ──3.5e-04──→ Node3:t6 ──4.6e-02──→ Node4:Node2 ──2.5e-02──→ Node4:Node10 ──6.3e-03──→ Node4:Node19 ──3.6e-04──→ Node4:t6 ──1.6e-01──→ Node6:Node2 ──9.0e-02──→ Node6:Node10 ──2.2e-02──→ Node6:Node19 ──1.3e-03──→ Node6:t6 ──1.2e-01──→ Node11:Node2 ──6.6e-02──→ Node11:Node10 ──1.6e-02──→ Node11:Node19 ──9.4e-04──→ Node11:t6 ──1.1e-03──→ Node12:Node2 ──6.3e-04──→ Node12:Node10 ──1.6e-04──→ Node12:Node19 ──8.9e-06──→ Node12:t6 ──6.4e-03──→ Node14:Node2 ──3.5e-03──→ Node14:Node10 ──8.8e-04──→ Node14:Node19 ──5.0e-05──→ Node14:t6 ──2.3e-02──→ t17:Node2 ──1.3e-02──→ t17:Node10 ──3.2e-03──→ t17:Node19 ──1.8e-04──→ t17:t6 
#> 
#> [7] ◎──0.83244──→ Node2:Node2 ──0.45524──→ Node2:Node10 ──0.11418──→ Node2:Node19 ──0.00649──→ Node2:t9 ──0.04473──→ Node3:Node2 ──0.02446──→ Node3:Node10 ──0.00614──→ Node3:Node19 ──0.00035──→ Node3:t9 ──0.04569──→ Node4:Node2 ──0.02498──→ Node4:Node10 ──0.00627──→ Node4:Node19 ──0.00036──→ Node4:t9 ──0.14782──→ Node5:Node2 ──0.08084──→ Node5:Node10 ──0.02028──→ Node5:Node19 ──0.00115──→ Node5:t9 ──0.13172──→ Node10:Node2 ──0.07204──→ Node10:Node10 ──0.01807──→ Node10:Node19 ──0.00103──→ Node10:t9 ──0.03510──→ t5:Node2 ──0.01920──→ t5:Node10 ──0.00481──→ t5:Node19 ──0.00027──→ t5:t9 
#> 
#> [8] ◎──0.8324──→ Node2:Node2 ──0.4552──→ Node2:Node10 ──0.1207──→ Node2:t14 ──0.0447──→ Node3:Node2 ──0.0245──→ Node3:Node10 ──0.0065──→ Node3:t14 ──0.0457──→ Node4:Node2 ──0.0250──→ Node4:Node10 ──0.0066──→ Node4:t14 ──0.1478──→ Node5:Node2 ──0.0808──→ Node5:Node10 ──0.0214──→ Node5:t14 ──0.1317──→ Node10:Node2 ──0.0720──→ Node10:Node10 ──0.0191──→ Node10:t14 ──0.0195──→ Node16:Node2 ──0.0107──→ Node16:Node10 ──0.0028──→ Node16:t14 ──0.0156──→ t13:Node2 ──0.0085──→ t13:Node10 ──0.0023──→ t13:t14 
#> 
#> [9] ◎──1.08828──→ Node2:Node4 ──0.05563──→ Node2:Node5 ──0.19634──→ Node2:Node16 ──0.06811──→ Node2:t19 ──0.05848──→ Node3:Node4 ──0.00299──→ Node3:Node5 ──0.01055──→ Node3:Node16 ──0.00366──→ Node3:t19 ──0.05973──→ Node4:Node4 ──0.00305──→ Node4:Node5 ──0.01078──→ Node4:Node16 ──0.00374──→ Node4:t19 ──0.19326──→ Node5:Node4 ──0.00988──→ Node5:Node5 ──0.03487──→ Node5:Node16 ──0.01210──→ Node5:t19 ──0.17221──→ Node10:Node4 ──0.00880──→ Node10:Node5 ──0.03107──→ Node10:Node16 ──0.01078──→ Node10:t19 ──0.02553──→ Node16:Node4 ──0.00130──→ Node16:Node5 ──0.00461──→ Node16:Node16 ──0.00160──→ Node16:t19 ──0.01652──→ Node18:Node4 ──0.00084──→ Node18:Node5 ──0.00298──→ Node18:Node16 ──0.00103──→ Node18:t19 ──0.00384──→ t12:Node4 ──0.00020──→ t12:Node5 ──0.00069──→ t12:Node16 ──0.00024──→ t12:t19 
#> 
#> [10] ◎──1.08828──→ Node2:Node4 ──0.05563──→ Node2:Node5 ──0.19634──→ Node2:Node16 ──0.06811──→ Node2:t2 ──0.05848──→ Node3:Node4 ──0.00299──→ Node3:Node5 ──0.01055──→ Node3:Node16 ──0.00366──→ Node3:t2 ──0.05973──→ Node4:Node4 ──0.00305──→ Node4:Node5 ──0.01078──→ Node4:Node16 ──0.00374──→ Node4:t2 ──0.19326──→ Node5:Node4 ──0.00988──→ Node5:Node5 ──0.03487──→ Node5:Node16 ──0.01210──→ Node5:t2 ──0.17221──→ Node10:Node4 ──0.00880──→ Node10:Node5 ──0.03107──→ Node10:Node16 ──0.01078──→ Node10:t2 ──0.02553──→ Node16:Node4 ──0.00130──→ Node16:Node5 ──0.00461──→ Node16:Node16 ──0.00160──→ Node16:t2 ──0.01652──→ Node18:Node4 ──0.00084──→ Node18:Node5 ──0.00298──→ Node18:Node16 ──0.00103──→ Node18:t2 ──0.00384──→ t1:Node4 ──0.00020──→ t1:Node5 ──0.00069──→ t1:Node16 ──0.00024──→ t1:t2