Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──0.60226──→ Node4:Node8 ──0.11580──→ Node4:Node9 ──0.10569──→ Node4:Node17 ──0.02297──→ Node4:t6 ──0.00491──→ Node5:Node8 ──0.00094──→ Node5:Node9 ──0.00086──→ Node5:Node17 ──0.00019──→ Node5:t6 ──0.19457──→ t3:Node8 ──0.03741──→ t3:Node9 ──0.03414──→ t3:Node17 ──0.00742──→ t3:t6 
#> 
#> [2] ◎──0.60226──→ Node4:Node8 ──0.11580──→ Node4:Node9 ──0.10569──→ Node4:Node17 ──0.02297──→ Node4:t5 ──0.00491──→ Node5:Node8 ──0.00094──→ Node5:Node9 ──0.00086──→ Node5:Node17 ──0.00019──→ Node5:t5 ──0.19457──→ t2:Node8 ──0.03741──→ t2:Node9 ──0.03414──→ t2:Node17 ──0.00742──→ t2:t5 
#> 
#> [3] ◎──0.602──→ Node4:Node8 ──0.116──→ Node4:Node9 ──0.129──→ Node4:t18 ──0.199──→ t11:Node8 ──0.038──→ t11:Node9 ──0.043──→ t11:t18 
#> 
#> [4] ◎──0.423──→ Node2:Node8 ──0.105──→ Node2:Node12 ──0.067──→ Node2:t11 ──0.379──→ t1:Node8 ──0.094──→ t1:Node12 ──0.060──→ t1:t11 
#> 
#> [5] ◎──0.42251──→ Node2:Node8 ──0.10455──→ Node2:Node12 ──0.03939──→ Node2:Node16 ──0.02755──→ Node2:t13 ──0.17214──→ Node3:Node8 ──0.04260──→ Node3:Node12 ──0.01605──→ Node3:Node16 ──0.01122──→ Node3:t13 ──0.05077──→ Node6:Node8 ──0.01256──→ Node6:Node12 ──0.00473──→ Node6:Node16 ──0.00331──→ Node6:t13 ──0.00703──→ Node7:Node8 ──0.00174──→ Node7:Node12 ──0.00066──→ Node7:Node16 ──0.00046──→ Node7:t13 ──0.04579──→ Node9:Node8 ──0.01133──→ Node9:Node12 ──0.00427──→ Node9:Node16 ──0.00299──→ Node9:t13 ──0.08394──→ Node13:Node8 ──0.02077──→ Node13:Node12 ──0.00783──→ Node13:Node16 ──0.00547──→ Node13:t13 ──0.01676──→ Node19:Node8 ──0.00415──→ Node19:Node12 ──0.00156──→ Node19:Node16 ──0.00109──→ Node19:t13 ──0.00279──→ t8:Node8 ──0.00069──→ t8:Node12 ──0.00026──→ t8:Node16 ──0.00018──→ t8:t13 
#> 
#> [6] ◎──0.42251──→ Node2:Node8 ──0.10455──→ Node2:Node12 ──0.03939──→ Node2:Node16 ──0.02755──→ Node2:t4 ──0.17214──→ Node3:Node8 ──0.04260──→ Node3:Node12 ──0.01605──→ Node3:Node16 ──0.01122──→ Node3:t4 ──0.05077──→ Node6:Node8 ──0.01256──→ Node6:Node12 ──0.00473──→ Node6:Node16 ──0.00331──→ Node6:t4 ──0.00703──→ Node7:Node8 ──0.00174──→ Node7:Node12 ──0.00066──→ Node7:Node16 ──0.00046──→ Node7:t4 ──0.04579──→ Node9:Node8 ──0.01133──→ Node9:Node12 ──0.00427──→ Node9:Node16 ──0.00299──→ Node9:t4 ──0.08394──→ Node13:Node8 ──0.02077──→ Node13:Node12 ──0.00783──→ Node13:Node16 ──0.00547──→ Node13:t4 ──0.01676──→ Node19:Node8 ──0.00415──→ Node19:Node12 ──0.00156──→ Node19:Node16 ──0.00109──→ Node19:t4 ──0.00279──→ t19:Node8 ──0.00069──→ t19:Node12 ──0.00026──→ t19:Node16 ──0.00018──→ t19:t4 
#> 
#> [7] ◎──6.0e-02──→ Node2:Node2 ──1.2e-01──→ Node2:Node4 ──2.4e-01──→ Node2:Node7 ──1.7e-01──→ Node2:Node18 ──5.4e-03──→ Node2:t14 ──2.4e-02──→ Node3:Node2 ──4.8e-02──→ Node3:Node4 ──9.9e-02──→ Node3:Node7 ──6.8e-02──→ Node3:Node18 ──2.2e-03──→ Node3:t14 ──7.2e-03──→ Node6:Node2 ──1.4e-02──→ Node6:Node4 ──2.9e-02──→ Node6:Node7 ──2.0e-02──→ Node6:Node18 ──6.5e-04──→ Node6:t14 ──9.9e-04──→ Node7:Node2 ──2.0e-03──→ Node7:Node4 ──4.0e-03──→ Node7:Node7 ──2.8e-03──→ Node7:Node18 ──9.0e-05──→ Node7:t14 ──6.5e-03──→ Node9:Node2 ──1.3e-02──→ Node9:Node4 ──2.6e-02──→ Node9:Node7 ──1.8e-02──→ Node9:Node18 ──5.9e-04──→ Node9:t14 ──1.2e-02──→ Node13:Node2 ──2.3e-02──→ Node13:Node4 ──4.8e-02──→ Node13:Node7 ──3.3e-02──→ Node13:Node18 ──1.1e-03──→ Node13:t14 ──2.3e-03──→ Node18:Node2 ──4.6e-03──→ Node18:Node4 ──9.5e-03──→ Node18:Node7 ──6.6e-03──→ Node18:Node18 ──2.1e-04──→ Node18:t14 ──4.2e-04──→ t9:Node2 ──8.4e-04──→ t9:Node4 ──1.7e-03──→ t9:Node7 ──1.2e-03──→ t9:Node18 ──3.9e-05──→ t9:t14 
#> 
#> [8] ◎──6.0e-02──→ Node2:Node2 ──1.2e-01──→ Node2:Node4 ──2.4e-01──→ Node2:Node7 ──1.7e-01──→ Node2:Node18 ──5.4e-03──→ Node2:t15 ──2.4e-02──→ Node3:Node2 ──4.8e-02──→ Node3:Node4 ──9.9e-02──→ Node3:Node7 ──6.8e-02──→ Node3:Node18 ──2.2e-03──→ Node3:t15 ──7.2e-03──→ Node6:Node2 ──1.4e-02──→ Node6:Node4 ──2.9e-02──→ Node6:Node7 ──2.0e-02──→ Node6:Node18 ──6.5e-04──→ Node6:t15 ──9.9e-04──→ Node7:Node2 ──2.0e-03──→ Node7:Node4 ──4.0e-03──→ Node7:Node7 ──2.8e-03──→ Node7:Node18 ──9.0e-05──→ Node7:t15 ──6.5e-03──→ Node9:Node2 ──1.3e-02──→ Node9:Node4 ──2.6e-02──→ Node9:Node7 ──1.8e-02──→ Node9:Node18 ──5.9e-04──→ Node9:t15 ──1.2e-02──→ Node13:Node2 ──2.3e-02──→ Node13:Node4 ──4.8e-02──→ Node13:Node7 ──3.3e-02──→ Node13:Node18 ──1.1e-03──→ Node13:t15 ──2.3e-03──→ Node18:Node2 ──4.6e-03──→ Node18:Node4 ──9.5e-03──→ Node18:Node7 ──6.6e-03──→ Node18:Node18 ──2.1e-04──→ Node18:t15 ──4.2e-04──→ t15:Node2 ──8.4e-04──→ t15:Node4 ──1.7e-03──→ t15:Node7 ──1.2e-03──→ t15:Node18 ──3.9e-05──→ t15:t15 
#> 
#> [9] ◎──6.0e-02──→ Node2:Node2 ──1.2e-01──→ Node2:Node4 ──2.4e-01──→ Node2:Node7 ──1.3e-01──→ Node2:Node14 ──3.8e-02──→ Node2:Node19 ──3.9e-03──→ Node2:t17 ──2.4e-02──→ Node3:Node2 ──4.8e-02──→ Node3:Node4 ──9.9e-02──→ Node3:Node7 ──5.3e-02──→ Node3:Node14 ──1.6e-02──→ Node3:Node19 ──1.6e-03──→ Node3:t17 ──7.2e-03──→ Node6:Node2 ──1.4e-02──→ Node6:Node4 ──2.9e-02──→ Node6:Node7 ──1.6e-02──→ Node6:Node14 ──4.6e-03──→ Node6:Node19 ──4.7e-04──→ Node6:t17 ──9.9e-04──→ Node7:Node2 ──2.0e-03──→ Node7:Node4 ──4.0e-03──→ Node7:Node7 ──2.2e-03──→ Node7:Node14 ──6.4e-04──→ Node7:Node19 ──6.5e-05──→ Node7:t17 ──6.5e-03──→ Node9:Node2 ──1.3e-02──→ Node9:Node4 ──2.6e-02──→ Node9:Node7 ──1.4e-02──→ Node9:Node14 ──4.2e-03──→ Node9:Node19 ──4.2e-04──→ Node9:t17 ──1.5e-02──→ t13:Node2 ──2.9e-02──→ t13:Node4 ──6.0e-02──→ t13:Node7 ──3.2e-02──→ t13:Node14 ──9.4e-03──→ t13:Node19 ──9.6e-04──→ t13:t17 
#> 
#> [10] ◎──6.0e-02──→ Node2:Node2 ──1.2e-01──→ Node2:Node4 ──2.4e-01──→ Node2:Node7 ──1.3e-01──→ Node2:Node14 ──3.8e-02──→ Node2:Node19 ──3.9e-03──→ Node2:t3 ──2.4e-02──→ Node3:Node2 ──4.8e-02──→ Node3:Node4 ──9.9e-02──→ Node3:Node7 ──5.3e-02──→ Node3:Node14 ──1.6e-02──→ Node3:Node19 ──1.6e-03──→ Node3:t3 ──7.2e-03──→ Node6:Node2 ──1.4e-02──→ Node6:Node4 ──2.9e-02──→ Node6:Node7 ──1.6e-02──→ Node6:Node14 ──4.6e-03──→ Node6:Node19 ──4.7e-04──→ Node6:t3 ──9.9e-04──→ Node7:Node2 ──2.0e-03──→ Node7:Node4 ──4.0e-03──→ Node7:Node7 ──2.2e-03──→ Node7:Node14 ──6.4e-04──→ Node7:Node19 ──6.5e-05──→ Node7:t3 ──1.4e-02──→ Node10:Node2 ──2.8e-02──→ Node10:Node4 ──5.8e-02──→ Node10:Node7 ──3.1e-02──→ Node10:Node14 ──9.1e-03──→ Node10:Node19 ──9.2e-04──→ Node10:t3 ──7.0e-03──→ t4:Node2 ──1.4e-02──→ t4:Node4 ──2.8e-02──→ t4:Node7 ──1.5e-02──→ t4:Node14 ──4.5e-03──→ t4:Node19 ──4.6e-04──→ t4:t3