Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──0.6791──→ Node2:Node4 ──0.1449──→ Node2:Node6 ──0.0649──→ Node2:Node12 ──0.0449──→ Node2:t17 ──0.0891──→ Node3:Node4 ──0.0190──→ Node3:Node6 ──0.0085──→ Node3:Node12 ──0.0059──→ Node3:t17 ──0.1929──→ Node4:Node4 ──0.0412──→ Node4:Node6 ──0.0184──→ Node4:Node12 ──0.0128──→ Node4:t17 ──0.3164──→ Node7:Node4 ──0.0675──→ Node7:Node6 ──0.0302──→ Node7:Node12 ──0.0209──→ Node7:t17 ──0.2216──→ t18:Node4 ──0.0473──→ t18:Node6 ──0.0212──→ t18:Node12 ──0.0146──→ t18:t17 
#> 
#> [2] ◎──6.8e-01──→ Node2:Node4 ──1.4e-01──→ Node2:Node6 ──6.5e-02──→ Node2:Node12 ──9.8e-03──→ Node2:Node15 ──3.4e-02──→ Node2:Node19 ──9.2e-04──→ Node2:t2 ──8.9e-02──→ Node3:Node4 ──1.9e-02──→ Node3:Node6 ──8.5e-03──→ Node3:Node12 ──1.3e-03──→ Node3:Node15 ──4.5e-03──→ Node3:Node19 ──1.2e-04──→ Node3:t2 ──1.9e-01──→ Node4:Node4 ──4.1e-02──→ Node4:Node6 ──1.8e-02──→ Node4:Node12 ──2.8e-03──→ Node4:Node15 ──9.7e-03──→ Node4:Node19 ──2.6e-04──→ Node4:t2 ──3.2e-01──→ Node7:Node4 ──6.8e-02──→ Node7:Node6 ──3.0e-02──→ Node7:Node12 ──4.6e-03──→ Node7:Node15 ──1.6e-02──→ Node7:Node19 ──4.3e-04──→ Node7:t2 ──2.0e-01──→ Node15:Node4 ──4.2e-02──→ Node15:Node6 ──1.9e-02──→ Node15:Node12 ──2.9e-03──→ Node15:Node15 ──1.0e-02──→ Node15:Node19 ──2.7e-04──→ Node15:t2 ──7.2e-03──→ Node16:Node4 ──1.5e-03──→ Node16:Node6 ──6.9e-04──→ Node16:Node12 ──1.0e-04──→ Node16:Node15 ──3.6e-04──→ Node16:Node19 ──9.8e-06──→ Node16:t2 ──4.1e-04──→ Node17:Node4 ──8.9e-05──→ Node17:Node6 ──4.0e-05──→ Node17:Node12 ──6.0e-06──→ Node17:Node15 ──2.1e-05──→ Node17:Node19 ──5.6e-07──→ Node17:t2 ──1.5e-02──→ t14:Node4 ──3.2e-03──→ t14:Node6 ──1.4e-03──→ t14:Node12 ──2.2e-04──→ t14:Node15 ──7.6e-04──→ t14:Node19 ──2.0e-05──→ t14:t2 
#> 
#> [3] ◎──6.8e-01──→ Node2:Node4 ──1.4e-01──→ Node2:Node6 ──6.5e-02──→ Node2:Node12 ──9.8e-03──→ Node2:Node15 ──3.4e-02──→ Node2:Node19 ──9.2e-04──→ Node2:t7 ──8.9e-02──→ Node3:Node4 ──1.9e-02──→ Node3:Node6 ──8.5e-03──→ Node3:Node12 ──1.3e-03──→ Node3:Node15 ──4.5e-03──→ Node3:Node19 ──1.2e-04──→ Node3:t7 ──1.9e-01──→ Node4:Node4 ──4.1e-02──→ Node4:Node6 ──1.8e-02──→ Node4:Node12 ──2.8e-03──→ Node4:Node15 ──9.7e-03──→ Node4:Node19 ──2.6e-04──→ Node4:t7 ──3.2e-01──→ Node7:Node4 ──6.8e-02──→ Node7:Node6 ──3.0e-02──→ Node7:Node12 ──4.6e-03──→ Node7:Node15 ──1.6e-02──→ Node7:Node19 ──4.3e-04──→ Node7:t7 ──2.0e-01──→ Node15:Node4 ──4.2e-02──→ Node15:Node6 ──1.9e-02──→ Node15:Node12 ──2.9e-03──→ Node15:Node15 ──1.0e-02──→ Node15:Node19 ──2.7e-04──→ Node15:t7 ──7.2e-03──→ Node16:Node4 ──1.5e-03──→ Node16:Node6 ──6.9e-04──→ Node16:Node12 ──1.0e-04──→ Node16:Node15 ──3.6e-04──→ Node16:Node19 ──9.8e-06──→ Node16:t7 ──4.1e-04──→ Node17:Node4 ──8.9e-05──→ Node17:Node6 ──4.0e-05──→ Node17:Node12 ──6.0e-06──→ Node17:Node15 ──2.1e-05──→ Node17:Node19 ──5.6e-07──→ Node17:t7 ──1.5e-02──→ t12:Node4 ──3.2e-03──→ t12:Node6 ──1.4e-03──→ t12:Node12 ──2.2e-04──→ t12:Node15 ──7.6e-04──→ t12:Node19 ──2.0e-05──→ t12:t7 
#> 
#> [4] ◎──6.8e-01──→ Node2:Node4 ──1.4e-01──→ Node2:Node6 ──6.5e-02──→ Node2:Node12 ──9.8e-03──→ Node2:Node15 ──3.5e-02──→ Node2:t8 ──8.9e-02──→ Node3:Node4 ──1.9e-02──→ Node3:Node6 ──8.5e-03──→ Node3:Node12 ──1.3e-03──→ Node3:Node15 ──4.6e-03──→ Node3:t8 ──1.9e-01──→ Node4:Node4 ──4.1e-02──→ Node4:Node6 ──1.8e-02──→ Node4:Node12 ──2.8e-03──→ Node4:Node15 ──1.0e-02──→ Node4:t8 ──3.2e-01──→ Node7:Node4 ──6.8e-02──→ Node7:Node6 ──3.0e-02──→ Node7:Node12 ──4.6e-03──→ Node7:Node15 ──1.6e-02──→ Node7:t8 ──2.0e-01──→ Node15:Node4 ──4.2e-02──→ Node15:Node6 ──1.9e-02──→ Node15:Node12 ──2.9e-03──→ Node15:Node15 ──1.0e-02──→ Node15:t8 ──7.2e-03──→ Node16:Node4 ──1.5e-03──→ Node16:Node6 ──6.9e-04──→ Node16:Node12 ──1.0e-04──→ Node16:Node15 ──3.7e-04──→ Node16:t8 ──1.2e-03──→ Node18:Node4 ──2.5e-04──→ Node18:Node6 ──1.1e-04──→ Node18:Node12 ──1.7e-05──→ Node18:Node15 ──6.1e-05──→ Node18:t8 ──1.4e-02──→ t16:Node4 ──3.1e-03──→ t16:Node6 ──1.4e-03──→ t16:Node12 ──2.1e-04──→ t16:Node15 ──7.4e-04──→ t16:t8 
#> 
#> [5] ◎──6.8e-01──→ Node2:Node4 ──1.4e-01──→ Node2:Node6 ──2.8e-02──→ Node2:Node7 ──8.2e-02──→ Node2:t11 ──8.9e-02──→ Node3:Node4 ──1.9e-02──→ Node3:Node6 ──3.7e-03──→ Node3:Node7 ──1.1e-02──→ Node3:t11 ──1.9e-01──→ Node4:Node4 ──4.1e-02──→ Node4:Node6 ──7.9e-03──→ Node4:Node7 ──2.3e-02──→ Node4:t11 ──3.2e-01──→ Node7:Node4 ──6.8e-02──→ Node7:Node6 ──1.3e-02──→ Node7:Node7 ──3.8e-02──→ Node7:t11 ──2.0e-01──→ Node15:Node4 ──4.2e-02──→ Node15:Node6 ──8.2e-03──→ Node15:Node7 ──2.4e-02──→ Node15:t11 ──7.2e-03──→ Node16:Node4 ──1.5e-03──→ Node16:Node6 ──3.0e-04──→ Node16:Node7 ──8.7e-04──→ Node16:t11 ──1.2e-03──→ Node18:Node4 ──2.5e-04──→ Node18:Node6 ──4.8e-05──→ Node18:Node7 ──1.4e-04──→ Node18:t11 ──1.4e-02──→ t1:Node4 ──3.1e-03──→ t1:Node6 ──5.9e-04──→ t1:Node7 ──1.7e-03──→ t1:t11 
#> 
#> [6] ◎──0.67909──→ Node2:Node4 ──0.14493──→ Node2:Node6 ──0.02792──→ Node2:Node7 ──0.02398──→ Node2:Node9 ──0.00903──→ Node2:Node10 ──0.01077──→ Node2:Node14 ──0.03804──→ Node2:t19 ──0.08905──→ Node3:Node4 ──0.01901──→ Node3:Node6 ──0.00366──→ Node3:Node7 ──0.00314──→ Node3:Node9 ──0.00118──→ Node3:Node10 ──0.00141──→ Node3:Node14 ──0.00499──→ Node3:t19 ──0.19295──→ Node4:Node4 ──0.04118──→ Node4:Node6 ──0.00793──→ Node4:Node7 ──0.00681──→ Node4:Node9 ──0.00256──→ Node4:Node10 ──0.00306──→ Node4:Node14 ──0.01081──→ Node4:t19 ──0.31635──→ Node7:Node4 ──0.06752──→ Node7:Node6 ──0.01301──→ Node7:Node7 ──0.01117──→ Node7:Node9 ──0.00420──→ Node7:Node10 ──0.00502──→ Node7:Node14 ──0.01772──→ Node7:t19 ──0.19887──→ Node15:Node4 ──0.04244──→ Node15:Node6 ──0.00818──→ Node15:Node7 ──0.00702──→ Node15:Node9 ──0.00264──→ Node15:Node10 ──0.00316──→ Node15:Node14 ──0.01114──→ Node15:t19 ──0.02273──→ t6:Node4 ──0.00485──→ t6:Node6 ──0.00093──→ t6:Node7 ──0.00080──→ t6:Node9 ──0.00030──→ t6:Node10 ──0.00036──→ t6:Node14 ──0.00127──→ t6:t19 
#> 
#> [7] ◎──0.67909──→ Node2:Node4 ──0.14493──→ Node2:Node6 ──0.02792──→ Node2:Node7 ──0.02398──→ Node2:Node9 ──0.00903──→ Node2:Node10 ──0.01077──→ Node2:Node14 ──0.02819──→ Node2:Node17 ──0.00723──→ Node2:Node18 ──0.00262──→ Node2:t16 ──0.08905──→ Node3:Node4 ──0.01901──→ Node3:Node6 ──0.00366──→ Node3:Node7 ──0.00314──→ Node3:Node9 ──0.00118──→ Node3:Node10 ──0.00141──→ Node3:Node14 ──0.00370──→ Node3:Node17 ──0.00095──→ Node3:Node18 ──0.00034──→ Node3:t16 ──0.19295──→ Node4:Node4 ──0.04118──→ Node4:Node6 ──0.00793──→ Node4:Node7 ──0.00681──→ Node4:Node9 ──0.00256──→ Node4:Node10 ──0.00306──→ Node4:Node14 ──0.00801──→ Node4:Node17 ──0.00205──→ Node4:Node18 ──0.00075──→ Node4:t16 ──0.15325──→ Node5:Node4 ──0.03271──→ Node5:Node6 ──0.00630──→ Node5:Node7 ──0.00541──→ Node5:Node9 ──0.00204──→ Node5:Node10 ──0.00243──→ Node5:Node14 ──0.00636──→ Node5:Node17 ──0.00163──→ Node5:Node18 ──0.00059──→ Node5:t16 ──0.38471──→ t15:Node4 ──0.08210──→ t15:Node6 ──0.01582──→ t15:Node7 ──0.01359──→ t15:Node9 ──0.00511──→ t15:Node10 ──0.00610──→ t15:Node14 ──0.01597──→ t15:Node17 ──0.00410──→ t15:Node18 ──0.00149──→ t15:t16 
#> 
#> [8] ◎──0.67909──→ Node2:Node4 ──0.14493──→ Node2:Node6 ──0.02792──→ Node2:Node7 ──0.02398──→ Node2:Node9 ──0.00903──→ Node2:Node10 ──0.01077──→ Node2:Node14 ──0.02819──→ Node2:Node17 ──0.00723──→ Node2:Node18 ──0.00262──→ Node2:t18 ──0.08905──→ Node3:Node4 ──0.01901──→ Node3:Node6 ──0.00366──→ Node3:Node7 ──0.00314──→ Node3:Node9 ──0.00118──→ Node3:Node10 ──0.00141──→ Node3:Node14 ──0.00370──→ Node3:Node17 ──0.00095──→ Node3:Node18 ──0.00034──→ Node3:t18 ──0.19295──→ Node4:Node4 ──0.04118──→ Node4:Node6 ──0.00793──→ Node4:Node7 ──0.00681──→ Node4:Node9 ──0.00256──→ Node4:Node10 ──0.00306──→ Node4:Node14 ──0.00801──→ Node4:Node17 ──0.00205──→ Node4:Node18 ──0.00075──→ Node4:t18 ──0.15325──→ Node5:Node4 ──0.03271──→ Node5:Node6 ──0.00630──→ Node5:Node7 ──0.00541──→ Node5:Node9 ──0.00204──→ Node5:Node10 ──0.00243──→ Node5:Node14 ──0.00636──→ Node5:Node17 ──0.00163──→ Node5:Node18 ──0.00059──→ Node5:t18 ──0.32704──→ Node13:Node4 ──0.06980──→ Node13:Node6 ──0.01345──→ Node13:Node7 ──0.01155──→ Node13:Node9 ──0.00435──→ Node13:Node10 ──0.00519──→ Node13:Node14 ──0.01357──→ Node13:Node17 ──0.00348──→ Node13:Node18 ──0.00126──→ Node13:t18 ──0.05766──→ t4:Node4 ──0.01231──→ t4:Node6 ──0.00237──→ t4:Node7 ──0.00204──→ t4:Node9 ──0.00077──→ t4:Node10 ──0.00091──→ t4:Node14 ──0.00239──→ t4:Node17 ──0.00061──→ t4:Node18 ──0.00022──→ t4:t18 
#> 
#> [9] ◎──0.67909──→ Node2:Node4 ──0.14493──→ Node2:Node6 ──0.02792──→ Node2:Node7 ──0.02398──→ Node2:Node9 ──0.00903──→ Node2:Node10 ──0.01077──→ Node2:Node14 ──0.02819──→ Node2:Node17 ──0.00985──→ Node2:t15 ──0.08905──→ Node3:Node4 ──0.01901──→ Node3:Node6 ──0.00366──→ Node3:Node7 ──0.00314──→ Node3:Node9 ──0.00118──→ Node3:Node10 ──0.00141──→ Node3:Node14 ──0.00370──→ Node3:Node17 ──0.00129──→ Node3:t15 ──0.19295──→ Node4:Node4 ──0.04118──→ Node4:Node6 ──0.00793──→ Node4:Node7 ──0.00681──→ Node4:Node9 ──0.00256──→ Node4:Node10 ──0.00306──→ Node4:Node14 ──0.00801──→ Node4:Node17 ──0.00280──→ Node4:t15 ──0.15325──→ Node5:Node4 ──0.03271──→ Node5:Node6 ──0.00630──→ Node5:Node7 ──0.00541──→ Node5:Node9 ──0.00204──→ Node5:Node10 ──0.00243──→ Node5:Node14 ──0.00636──→ Node5:Node17 ──0.00222──→ Node5:t15 ──0.32704──→ Node13:Node4 ──0.06980──→ Node13:Node6 ──0.01345──→ Node13:Node7 ──0.01155──→ Node13:Node9 ──0.00435──→ Node13:Node10 ──0.00519──→ Node13:Node14 ──0.01357──→ Node13:Node17 ──0.00475──→ Node13:t15 ──0.05766──→ t8:Node4 ──0.01231──→ t8:Node6 ──0.00237──→ t8:Node7 ──0.00204──→ t8:Node9 ──0.00077──→ t8:Node10 ──0.00091──→ t8:Node14 ──0.00239──→ t8:Node17 ──0.00084──→ t8:t15 
#> 
#> [10] ◎──6.8e-01──→ Node2:Node4 ──1.4e-01──→ Node2:Node6 ──2.8e-02──→ Node2:Node7 ──2.4e-02──→ Node2:Node9 ──9.0e-03──→ Node2:Node10 ──4.9e-02──→ Node2:t3 ──8.9e-02──→ Node3:Node4 ──1.9e-02──→ Node3:Node6 ──3.7e-03──→ Node3:Node7 ──3.1e-03──→ Node3:Node9 ──1.2e-03──→ Node3:Node10 ──6.4e-03──→ Node3:t3 ──4.9e-01──→ Node6:Node4 ──1.0e-01──→ Node6:Node6 ──2.0e-02──→ Node6:Node7 ──1.7e-02──→ Node6:Node9 ──6.5e-03──→ Node6:Node10 ──3.5e-02──→ Node6:t3 ──1.6e-01──→ Node9:Node4 ──3.4e-02──→ Node9:Node6 ──6.5e-03──→ Node9:Node7 ──5.6e-03──→ Node9:Node9 ──2.1e-03──→ Node9:Node10 ──1.1e-02──→ Node9:t3 ──4.5e-03──→ Node11:Node4 ──9.5e-04──→ Node11:Node6 ──1.8e-04──→ Node11:Node7 ──1.6e-04──→ Node11:Node9 ──5.9e-05──→ Node11:Node10 ──3.2e-04──→ Node11:t3 ──7.6e-02──→ t7:Node4 ──1.6e-02──→ t7:Node6 ──3.1e-03──→ t7:Node7 ──2.7e-03──→ t7:Node9 ──1.0e-03──→ t7:Node10 ──5.5e-03──→ t7:t3