Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──0.331──→ Node2:Node2 ──0.046──→ Node2:Node3 ──0.038──→ Node2:Node7 ──0.041──→ Node2:t9 ──0.947──→ Node3:Node2 ──0.130──→ Node3:Node3 ──0.109──→ Node3:Node7 ──0.116──→ Node3:t9 ──0.642──→ Node6:Node2 ──0.088──→ Node6:Node3 ──0.074──→ Node6:Node7 ──0.079──→ Node6:t9 ──0.211──→ Node7:Node2 ──0.029──→ Node7:Node3 ──0.024──→ Node7:Node7 ──0.026──→ Node7:t9 ──0.663──→ t18:Node2 ──0.091──→ t18:Node3 ──0.076──→ t18:Node7 ──0.081──→ t18:t9 
#> 
#> [2] ◎──0.331──→ Node2:Node2 ──0.046──→ Node2:Node3 ──0.038──→ Node2:Node7 ──0.041──→ Node2:t16 ──0.947──→ Node3:Node2 ──0.130──→ Node3:Node3 ──0.109──→ Node3:Node7 ──0.116──→ Node3:t16 ──0.642──→ Node6:Node2 ──0.088──→ Node6:Node3 ──0.074──→ Node6:Node7 ──0.079──→ Node6:t16 ──0.211──→ Node7:Node2 ──0.029──→ Node7:Node3 ──0.024──→ Node7:Node7 ──0.026──→ Node7:t16 ──0.427──→ Node11:Node2 ──0.059──→ Node11:Node3 ──0.049──→ Node11:Node7 ──0.052──→ Node11:t16 ──0.236──→ t1:Node2 ──0.033──→ t1:Node3 ──0.027──→ t1:Node7 ──0.029──→ t1:t16 
#> 
#> [3] ◎──0.3306──→ Node2:Node2 ──0.0455──→ Node2:Node3 ──0.0717──→ Node2:Node14 ──0.0068──→ Node2:t20 ──0.9470──→ Node3:Node2 ──0.1304──→ Node3:Node3 ──0.2055──→ Node3:Node14 ──0.0196──→ Node3:t20 ──0.6422──→ Node6:Node2 ──0.0884──→ Node6:Node3 ──0.1394──→ Node6:Node14 ──0.0133──→ Node6:t20 ──0.2111──→ Node7:Node2 ──0.0291──→ Node7:Node3 ──0.0458──→ Node7:Node14 ──0.0044──→ Node7:t20 ──0.4269──→ Node11:Node2 ──0.0588──→ Node11:Node3 ──0.0927──→ Node11:Node14 ──0.0088──→ Node11:t20 ──0.2365──→ t15:Node2 ──0.0326──→ t15:Node3 ──0.0513──→ t15:Node14 ──0.0049──→ t15:t20 
#> 
#> [4] ◎──0.3306──→ Node2:Node2 ──0.0455──→ Node2:Node3 ──0.0717──→ Node2:Node14 ──0.0068──→ Node2:t5 ──0.9470──→ Node3:Node2 ──0.1304──→ Node3:Node3 ──0.2055──→ Node3:Node14 ──0.0196──→ Node3:t5 ──0.6422──→ Node6:Node2 ──0.0884──→ Node6:Node3 ──0.1394──→ Node6:Node14 ──0.0133──→ Node6:t5 ──0.6088──→ Node9:Node2 ──0.0838──→ Node9:Node3 ──0.1321──→ Node9:Node14 ──0.0126──→ Node9:t5 ──0.2657──→ t2:Node2 ──0.0366──→ t2:Node3 ──0.0577──→ t2:Node14 ──0.0055──→ t2:t5 
#> 
#> [5] ◎──0.3306──→ Node2:Node2 ──0.0694──→ Node2:Node4 ──0.0072──→ Node2:Node5 ──0.0474──→ Node2:t6 ──0.9470──→ Node3:Node2 ──0.1988──→ Node3:Node4 ──0.0207──→ Node3:Node5 ──0.1359──→ Node3:t6 ──0.6422──→ Node6:Node2 ──0.1348──→ Node6:Node4 ──0.0140──→ Node6:Node5 ──0.0922──→ Node6:t6 ──0.6088──→ Node9:Node2 ──0.1278──→ Node9:Node4 ──0.0133──→ Node9:Node5 ──0.0874──→ Node9:t6 ──0.2657──→ t16:Node2 ──0.0558──→ t16:Node4 ──0.0058──→ t16:Node5 ──0.0381──→ t16:t6 
#> 
#> [6] ◎──0.3306──→ Node2:Node2 ──0.0694──→ Node2:Node4 ──0.0072──→ Node2:Node5 ──0.0436──→ Node2:Node18 ──0.0039──→ Node2:t10 ──0.9470──→ Node3:Node2 ──0.1988──→ Node3:Node4 ──0.0207──→ Node3:Node5 ──0.1248──→ Node3:Node18 ──0.0112──→ Node3:t10 ──1.2900──→ Node13:Node2 ──0.2709──→ Node13:Node4 ──0.0282──→ Node13:Node5 ──0.1699──→ Node13:Node18 ──0.0152──→ Node13:t10 ──0.2266──→ t8:Node2 ──0.0476──→ t8:Node4 ──0.0050──→ t8:Node5 ──0.0299──→ t8:Node18 ──0.0027──→ t8:t10 
#> 
#> [7] ◎──0.33057──→ Node2:Node2 ──0.06941──→ Node2:Node4 ──0.00723──→ Node2:Node5 ──0.04355──→ Node2:Node18 ──0.00390──→ Node2:t19 ──0.94702──→ Node3:Node2 ──0.19884──→ Node3:Node4 ──0.02070──→ Node3:Node5 ──0.12476──→ Node3:Node18 ──0.01116──→ Node3:t19 ──1.29002──→ Node13:Node2 ──0.27086──→ Node13:Node4 ──0.02820──→ Node13:Node5 ──0.16995──→ Node13:Node18 ──0.01521──→ Node13:t19 ──0.20602──→ Node19:Node2 ──0.04326──→ Node19:Node4 ──0.00450──→ Node19:Node5 ──0.02714──→ Node19:Node18 ──0.00243──→ Node19:t19 ──0.02060──→ t10:Node2 ──0.00433──→ t10:Node4 ──0.00045──→ t10:Node5 ──0.00271──→ t10:Node18 ──0.00024──→ t10:t19 
#> 
#> [8] ◎──0.3306──→ Node2:Node2 ──0.0694──→ Node2:Node4 ──0.0547──→ Node2:t8 ──0.9470──→ Node3:Node2 ──0.1988──→ Node3:Node4 ──0.1566──→ Node3:t8 ──1.2900──→ Node13:Node2 ──0.2709──→ Node13:Node4 ──0.2134──→ Node13:t8 ──0.2060──→ Node19:Node2 ──0.0433──→ Node19:Node4 ──0.0341──→ Node19:t8 ──0.0206──→ t4:Node2 ──0.0043──→ t4:Node4 ──0.0034──→ t4:t8 
#> 
#> [9] ◎──0.409──→ Node2:Node6 ──0.018──→ Node2:Node9 ──0.017──→ Node2:Node11 ──0.011──→ Node2:t3 ──1.963──→ Node5:Node6 ──0.087──→ Node5:Node9 ──0.081──→ Node5:Node11 ──0.052──→ Node5:t3 ──1.084──→ t6:Node6 ──0.048──→ t6:Node9 ──0.045──→ t6:Node11 ──0.029──→ t6:t3 
#> 
#> [10] ◎──0.40881──→ Node2:Node6 ──0.01802──→ Node2:Node9 ──0.01697──→ Node2:Node11 ──0.01086──→ Node2:t13 ──1.96305──→ Node5:Node6 ──0.08651──→ Node5:Node9 ──0.08147──→ Node5:Node11 ──0.05216──→ Node5:t13 ──0.59822──→ Node8:Node6 ──0.02636──→ Node8:Node9 ──0.02483──→ Node8:Node11 ──0.01590──→ Node8:t13 ──0.45430──→ Node18:Node6 ──0.02002──→ Node18:Node9 ──0.01885──→ Node18:Node11 ──0.01207──→ Node18:t13 ──0.03116──→ t5:Node6 ──0.00137──→ t5:Node9 ──0.00129──→ t5:Node11 ──0.00083──→ t5:t13