Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──5.068──→ t8:Node2 ──0.828──→ t8:Node6 ──0.526──→ t8:Node17 ──0.071──→ t8:t5 
#> 
#> [2] ◎──1.6042──→ Node2:Node2 ──0.2620──→ Node2:Node6 ──0.1665──→ Node2:Node17 ──0.0226──→ Node2:t14 ──2.4395──→ Node4:Node2 ──0.3984──→ Node4:Node6 ──0.2533──→ Node4:Node17 ──0.0344──→ Node4:t14 ──0.7552──→ Node10:Node2 ──0.1233──→ Node10:Node6 ──0.0784──→ Node10:Node17 ──0.0107──→ Node10:t14 ──0.2687──→ t19:Node2 ──0.0439──→ t19:Node6 ──0.0279──→ t19:Node17 ──0.0038──→ t19:t14 
#> 
#> [3] ◎──1.604──→ Node2:Node2 ──0.262──→ Node2:Node6 ──0.189──→ Node2:t16 ──2.439──→ Node4:Node2 ──0.398──→ Node4:Node6 ──0.288──→ Node4:t16 ──0.755──→ Node10:Node2 ──0.123──→ Node10:Node6 ──0.089──→ Node10:t16 ──0.269──→ t13:Node2 ──0.044──→ t13:Node6 ──0.032──→ t13:t16 
#> 
#> [4] ◎──1.60423──→ Node2:Node2 ──0.28808──→ Node2:Node7 ──0.15564──→ Node2:Node18 ──0.00746──→ Node2:t11 ──2.43947──→ Node4:Node2 ──0.43807──→ Node4:Node7 ──0.23668──→ Node4:Node18 ──0.01135──→ Node4:t11 ──0.83445──→ Node14:Node2 ──0.14985──→ Node14:Node7 ──0.08096──→ Node14:Node18 ──0.00388──→ Node14:t11 ──0.18941──→ t14:Node2 ──0.03401──→ t14:Node7 ──0.01838──→ t14:Node18 ──0.00088──→ t14:t11 
#> 
#> [5] ◎──1.60423──→ Node2:Node2 ──0.28808──→ Node2:Node7 ──0.15564──→ Node2:Node18 ──0.00746──→ Node2:t2 ──2.43947──→ Node4:Node2 ──0.43807──→ Node4:Node7 ──0.23668──→ Node4:Node18 ──0.01135──→ Node4:t2 ──0.83445──→ Node14:Node2 ──0.14985──→ Node14:Node7 ──0.08096──→ Node14:Node18 ──0.00388──→ Node14:t2 ──0.18941──→ t12:Node2 ──0.03401──→ t12:Node7 ──0.01838──→ t12:Node18 ──0.00088──→ t12:t2 
#> 
#> [6] ◎──1.6042──→ Node2:Node2 ──0.2881──→ Node2:Node7 ──0.1631──→ Node2:t10 ──0.8739──→ Node3:Node2 ──0.1569──→ Node3:Node7 ──0.0888──→ Node3:t10 ──1.9361──→ Node5:Node2 ──0.3477──→ Node5:Node7 ──0.1968──→ Node5:t10 ──0.2969──→ Node8:Node2 ──0.0533──→ Node8:Node7 ──0.0302──→ Node8:t10 ──0.0941──→ Node11:Node2 ──0.0169──→ Node11:Node7 ──0.0096──→ Node11:t10 ──0.0656──→ Node13:Node2 ──0.0118──→ Node13:Node7 ──0.0067──→ Node13:t10 ──0.1967──→ t5:Node2 ──0.0353──→ t5:Node7 ──0.0200──→ t5:t10 
#> 
#> [7] ◎──1.6323──→ Node2:Node3 ──0.0989──→ Node2:Node4 ──0.1052──→ Node2:Node5 ──0.2191──→ Node2:t7 ──0.8891──→ Node3:Node3 ──0.0539──→ Node3:Node4 ──0.0573──→ Node3:Node5 ──0.1194──→ Node3:t7 ──1.9699──→ Node5:Node3 ──0.1193──→ Node5:Node4 ──0.1269──→ Node5:Node5 ──0.2644──→ Node5:t7 ──0.3021──→ Node8:Node3 ──0.0183──→ Node8:Node4 ──0.0195──→ Node8:Node5 ──0.0406──→ Node8:t7 ──0.0958──→ Node11:Node3 ──0.0058──→ Node11:Node4 ──0.0062──→ Node11:Node5 ──0.0129──→ Node11:t7 ──0.0668──→ Node13:Node3 ──0.0040──→ Node13:Node4 ──0.0043──→ Node13:Node5 ──0.0090──→ Node13:t7 ──0.2001──→ t4:Node3 ──0.0121──→ t4:Node4 ──0.0129──→ t4:Node5 ──0.0269──→ t4:t7 
#> 
#> [8] ◎──1.63227──→ Node2:Node3 ──0.09886──→ Node2:Node4 ──0.10515──→ Node2:Node5 ──0.14372──→ Node2:Node10 ──0.03757──→ Node2:Node14 ──0.03783──→ Node2:t4 ──0.88915──→ Node3:Node3 ──0.05385──→ Node3:Node4 ──0.05728──→ Node3:Node5 ──0.07829──→ Node3:Node10 ──0.02047──→ Node3:Node14 ──0.02061──→ Node3:t4 ──1.96994──→ Node5:Node3 ──0.11932──→ Node5:Node4 ──0.12691──→ Node5:Node5 ──0.17345──→ Node5:Node10 ──0.04534──→ Node5:Node14 ──0.04565──→ Node5:t4 ──0.30208──→ Node8:Node3 ──0.01830──→ Node8:Node4 ──0.01946──→ Node8:Node5 ──0.02660──→ Node8:Node10 ──0.00695──→ Node8:Node14 ──0.00700──→ Node8:t4 ──0.09578──→ Node11:Node3 ──0.00580──→ Node11:Node4 ──0.00617──→ Node11:Node5 ──0.00843──→ Node11:Node10 ──0.00220──→ Node11:Node14 ──0.00222──→ Node11:t4 ──0.02396──→ Node12:Node3 ──0.00145──→ Node12:Node4 ──0.00154──→ Node12:Node5 ──0.00211──→ Node12:Node10 ──0.00055──→ Node12:Node14 ──0.00056──→ Node12:t4 ──0.24297──→ t16:Node3 ──0.01472──→ t16:Node4 ──0.01565──→ t16:Node5 ──0.02139──→ t16:Node10 ──0.00559──→ t16:Node14 ──0.00563──→ t16:t4 
#> 
#> [9] ◎──1.63227──→ Node2:Node3 ──0.09886──→ Node2:Node4 ──0.10515──→ Node2:Node5 ──0.14372──→ Node2:Node10 ──0.03757──→ Node2:Node14 ──0.03783──→ Node2:t8 ──0.88915──→ Node3:Node3 ──0.05385──→ Node3:Node4 ──0.05728──→ Node3:Node5 ──0.07829──→ Node3:Node10 ──0.02047──→ Node3:Node14 ──0.02061──→ Node3:t8 ──1.96994──→ Node5:Node3 ──0.11932──→ Node5:Node4 ──0.12691──→ Node5:Node5 ──0.17345──→ Node5:Node10 ──0.04534──→ Node5:Node14 ──0.04565──→ Node5:t8 ──0.30208──→ Node8:Node3 ──0.01830──→ Node8:Node4 ──0.01946──→ Node8:Node5 ──0.02660──→ Node8:Node10 ──0.00695──→ Node8:Node14 ──0.00700──→ Node8:t8 ──0.09578──→ Node11:Node3 ──0.00580──→ Node11:Node4 ──0.00617──→ Node11:Node5 ──0.00843──→ Node11:Node10 ──0.00220──→ Node11:Node14 ──0.00222──→ Node11:t8 ──0.02396──→ Node12:Node3 ──0.00145──→ Node12:Node4 ──0.00154──→ Node12:Node5 ──0.00211──→ Node12:Node10 ──0.00055──→ Node12:Node14 ──0.00056──→ Node12:t8 ──0.24297──→ t6:Node3 ──0.01472──→ t6:Node4 ──0.01565──→ t6:Node5 ──0.02139──→ t6:Node10 ──0.00559──→ t6:Node14 ──0.00563──→ t6:t8 
#> 
#> [10] ◎──1.6e+00──→ Node2:Node3 ──9.9e-02──→ Node2:Node4 ──1.1e-01──→ Node2:Node5 ──1.4e-01──→ Node2:Node10 ──7.4e-02──→ Node2:Node19 ──1.8e-03──→ Node2:t9 ──8.9e-01──→ Node3:Node3 ──5.4e-02──→ Node3:Node4 ──5.7e-02──→ Node3:Node5 ──7.8e-02──→ Node3:Node10 ──4.0e-02──→ Node3:Node19 ──9.7e-04──→ Node3:t9 ──2.0e+00──→ Node5:Node3 ──1.2e-01──→ Node5:Node4 ──1.3e-01──→ Node5:Node5 ──1.7e-01──→ Node5:Node10 ──8.9e-02──→ Node5:Node19 ──2.1e-03──→ Node5:t9 ──3.0e-01──→ Node8:Node3 ──1.8e-02──→ Node8:Node4 ──1.9e-02──→ Node8:Node5 ──2.7e-02──→ Node8:Node10 ──1.4e-02──→ Node8:Node19 ──3.3e-04──→ Node8:t9 ──3.3e-01──→ Node17:Node3 ──2.0e-02──→ Node17:Node4 ──2.1e-02──→ Node17:Node5 ──2.9e-02──→ Node17:Node10 ──1.5e-02──→ Node17:Node19 ──3.6e-04──→ Node17:t9 ──3.5e-02──→ t15:Node3 ──2.1e-03──→ t15:Node4 ──2.3e-03──→ t15:Node5 ──3.1e-03──→ t15:Node10 ──1.6e-03──→ t15:Node19 ──3.8e-05──→ t15:t9