Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──3.443──→ Node5:Node2 ──2.046──→ Node5:t2 ──0.325──→ Node12:Node2 ──0.193──→ Node12:t2 ──0.083──→ Node16:Node2 ──0.049──→ Node16:t2 ──0.023──→ t3:Node2 ──0.013──→ t3:t2 
#> 
#> [2] ◎──3.443──→ Node5:Node2 ──2.046──→ Node5:t16 ──0.325──→ Node12:Node2 ──0.193──→ Node12:t16 ──0.083──→ Node16:Node2 ──0.049──→ Node16:t16 ──0.023──→ t18:Node2 ──0.013──→ t18:t16 
#> 
#> [3] ◎──4.4710──→ Node5:Node3 ──0.2502──→ Node5:Node4 ──0.3114──→ Node5:Node7 ──0.4566──→ Node5:t15 ──0.4221──→ Node12:Node3 ──0.0236──→ Node12:Node4 ──0.0294──→ Node12:Node7 ──0.0431──→ Node12:t15 ──0.1373──→ t4:Node3 ──0.0077──→ t4:Node4 ──0.0096──→ t4:Node7 ──0.0140──→ t4:t15 
#> 
#> [4] ◎──4.471──→ Node5:Node3 ──0.250──→ Node5:Node4 ──0.311──→ Node5:Node7 ──0.111──→ Node5:Node10 ──0.345──→ Node5:t3 ──0.559──→ t7:Node3 ──0.031──→ t7:Node4 ──0.039──→ t7:Node7 ──0.014──→ t7:Node10 ──0.043──→ t7:t3 
#> 
#> [5] ◎──3.4341──→ Node2:Node3 ──0.1922──→ Node2:Node4 ──0.2392──→ Node2:Node7 ──0.0855──→ Node2:Node10 ──0.1404──→ Node2:Node12 ──0.1248──→ Node2:t17 ──0.4151──→ Node3:Node3 ──0.0232──→ Node3:Node4 ──0.0289──→ Node3:Node7 ──0.0103──→ Node3:Node10 ──0.0170──→ Node3:Node12 ──0.0151──→ Node3:t17 ──0.4592──→ Node4:Node3 ──0.0257──→ Node4:Node4 ──0.0320──→ Node4:Node7 ──0.0114──→ Node4:Node10 ──0.0188──→ Node4:Node12 ──0.0167──→ Node4:t17 ──0.2626──→ Node7:Node3 ──0.0147──→ Node7:Node4 ──0.0183──→ Node7:Node7 ──0.0065──→ Node7:Node10 ──0.0107──→ Node7:Node12 ──0.0095──→ Node7:t17 ──0.4594──→ t9:Node3 ──0.0257──→ t9:Node4 ──0.0320──→ t9:Node7 ──0.0114──→ t9:Node10 ──0.0188──→ t9:Node12 ──0.0167──→ t9:t17 
#> 
#> [6] ◎──3.43414──→ Node2:Node3 ──0.19222──→ Node2:Node4 ──0.23917──→ Node2:Node7 ──0.08550──→ Node2:Node10 ──0.14044──→ Node2:Node12 ──0.03532──→ Node2:Node14 ──0.08944──→ Node2:t18 ──0.41507──→ Node3:Node3 ──0.02323──→ Node3:Node4 ──0.02891──→ Node3:Node7 ──0.01033──→ Node3:Node10 ──0.01697──→ Node3:Node12 ──0.00427──→ Node3:Node14 ──0.01081──→ Node3:t18 ──0.45922──→ Node4:Node3 ──0.02570──→ Node4:Node4 ──0.03198──→ Node4:Node7 ──0.01143──→ Node4:Node10 ──0.01878──→ Node4:Node12 ──0.00472──→ Node4:Node14 ──0.01196──→ Node4:t18 ──0.26259──→ Node7:Node3 ──0.01470──→ Node7:Node4 ──0.01829──→ Node7:Node7 ──0.00654──→ Node7:Node10 ──0.01074──→ Node7:Node12 ──0.00270──→ Node7:Node14 ──0.00684──→ Node7:t18 ──0.29625──→ Node11:Node3 ──0.01658──→ Node11:Node4 ──0.02063──→ Node11:Node7 ──0.00738──→ Node11:Node10 ──0.01212──→ Node11:Node12 ──0.00305──→ Node11:Node14 ──0.00772──→ Node11:t18 ──0.15271──→ Node19:Node3 ──0.00855──→ Node19:Node4 ──0.01064──→ Node19:Node7 ──0.00380──→ Node19:Node10 ──0.00625──→ Node19:Node12 ──0.00157──→ Node19:Node14 ──0.00398──→ Node19:t18 ──0.01043──→ t1:Node3 ──0.00058──→ t1:Node4 ──0.00073──→ t1:Node7 ──0.00026──→ t1:Node10 ──0.00043──→ t1:Node12 ──0.00011──→ t1:Node14 ──0.00027──→ t1:t18 
#> 
#> [7] ◎──3.4e+00──→ Node2:Node3 ──1.9e-01──→ Node2:Node4 ──2.4e-01──→ Node2:Node7 ──8.6e-02──→ Node2:Node10 ──1.4e-01──→ Node2:Node12 ──3.5e-02──→ Node2:Node14 ──3.0e-02──→ Node2:Node17 ──5.9e-02──→ Node2:t9 ──4.2e-01──→ Node3:Node3 ──2.3e-02──→ Node3:Node4 ──2.9e-02──→ Node3:Node7 ──1.0e-02──→ Node3:Node10 ──1.7e-02──→ Node3:Node12 ──4.3e-03──→ Node3:Node14 ──3.6e-03──→ Node3:Node17 ──7.2e-03──→ Node3:t9 ──4.6e-01──→ Node4:Node3 ──2.6e-02──→ Node4:Node4 ──3.2e-02──→ Node4:Node7 ──1.1e-02──→ Node4:Node10 ──1.9e-02──→ Node4:Node12 ──4.7e-03──→ Node4:Node14 ──4.0e-03──→ Node4:Node17 ──7.9e-03──→ Node4:t9 ──2.6e-01──→ Node7:Node3 ──1.5e-02──→ Node7:Node4 ──1.8e-02──→ Node7:Node7 ──6.5e-03──→ Node7:Node10 ──1.1e-02──→ Node7:Node12 ──2.7e-03──→ Node7:Node14 ──2.3e-03──→ Node7:Node17 ──4.5e-03──→ Node7:t9 ──3.0e-01──→ Node11:Node3 ──1.7e-02──→ Node11:Node4 ──2.1e-02──→ Node11:Node7 ──7.4e-03──→ Node11:Node10 ──1.2e-02──→ Node11:Node12 ──3.0e-03──→ Node11:Node14 ──2.6e-03──→ Node11:Node17 ──5.1e-03──→ Node11:t9 ──1.5e-01──→ Node19:Node3 ──8.5e-03──→ Node19:Node4 ──1.1e-02──→ Node19:Node7 ──3.8e-03──→ Node19:Node10 ──6.2e-03──→ Node19:Node12 ──1.6e-03──→ Node19:Node14 ──1.3e-03──→ Node19:Node17 ──2.6e-03──→ Node19:t9 ──1.0e-02──→ t17:Node3 ──5.8e-04──→ t17:Node4 ──7.3e-04──→ t17:Node7 ──2.6e-04──→ t17:Node10 ──4.3e-04──→ t17:Node12 ──1.1e-04──→ t17:Node14 ──9.2e-05──→ t17:Node17 ──1.8e-04──→ t17:t9 
#> 
#> [8] ◎──3.4341──→ Node2:Node3 ──0.1922──→ Node2:Node4 ──0.2392──→ Node2:Node7 ──0.0855──→ Node2:Node10 ──0.1404──→ Node2:Node12 ──0.0353──→ Node2:Node14 ──0.0302──→ Node2:Node17 ──0.0593──→ Node2:t19 ──0.4151──→ Node3:Node3 ──0.0232──→ Node3:Node4 ──0.0289──→ Node3:Node7 ──0.0103──→ Node3:Node10 ──0.0170──→ Node3:Node12 ──0.0043──→ Node3:Node14 ──0.0036──→ Node3:Node17 ──0.0072──→ Node3:t19 ──0.4592──→ Node4:Node3 ──0.0257──→ Node4:Node4 ──0.0320──→ Node4:Node7 ──0.0114──→ Node4:Node10 ──0.0188──→ Node4:Node12 ──0.0047──→ Node4:Node14 ──0.0040──→ Node4:Node17 ──0.0079──→ Node4:t19 ──0.2626──→ Node7:Node3 ──0.0147──→ Node7:Node4 ──0.0183──→ Node7:Node7 ──0.0065──→ Node7:Node10 ──0.0107──→ Node7:Node12 ──0.0027──→ Node7:Node14 ──0.0023──→ Node7:Node17 ──0.0045──→ Node7:t19 ──0.2962──→ Node11:Node3 ──0.0166──→ Node11:Node4 ──0.0206──→ Node11:Node7 ──0.0074──→ Node11:Node10 ──0.0121──→ Node11:Node12 ──0.0030──→ Node11:Node14 ──0.0026──→ Node11:Node17 ──0.0051──→ Node11:t19 ──0.1631──→ t15:Node3 ──0.0091──→ t15:Node4 ──0.0114──→ t15:Node7 ──0.0041──→ t15:Node10 ──0.0067──→ t15:Node12 ──0.0017──→ t15:Node14 ──0.0014──→ t15:Node17 ──0.0028──→ t15:t19 
#> 
#> [9] ◎──3.4341──→ Node2:Node3 ──0.1922──→ Node2:Node4 ──0.2202──→ Node2:Node6 ──0.3697──→ Node2:t13 ──0.4151──→ Node3:Node3 ──0.0232──→ Node3:Node4 ──0.0266──→ Node3:Node6 ──0.0447──→ Node3:t13 ──0.4592──→ Node4:Node3 ──0.0257──→ Node4:Node4 ──0.0294──→ Node4:Node6 ──0.0494──→ Node4:t13 ──0.1777──→ Node6:Node3 ──0.0099──→ Node6:Node4 ──0.0114──→ Node6:Node6 ──0.0191──→ Node6:t13 ──0.5443──→ t8:Node3 ──0.0305──→ t8:Node4 ──0.0349──→ t8:Node6 ──0.0586──→ t8:t13 
#> 
#> [10] ◎──3.4341──→ Node2:Node3 ──0.1922──→ Node2:Node4 ──0.2202──→ Node2:Node6 ──0.2739──→ Node2:Node13 ──0.0958──→ Node2:t7 ──0.4151──→ Node3:Node3 ──0.0232──→ Node3:Node4 ──0.0266──→ Node3:Node6 ──0.0331──→ Node3:Node13 ──0.0116──→ Node3:t7 ──0.4592──→ Node4:Node3 ──0.0257──→ Node4:Node4 ──0.0294──→ Node4:Node6 ──0.0366──→ Node4:Node13 ──0.0128──→ Node4:t7 ──0.1777──→ Node6:Node3 ──0.0099──→ Node6:Node4 ──0.0114──→ Node6:Node6 ──0.0142──→ Node6:Node13 ──0.0050──→ Node6:t7 ──0.5443──→ t2:Node3 ──0.0305──→ t2:Node4 ──0.0349──→ t2:Node6 ──0.0434──→ t2:Node13 ──0.0152──→ t2:t7