Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──5.0e-02──→ Node4:Node2 ──2.4e-01──→ Node4:Node3 ──5.7e-01──→ Node4:Node12 ──1.7e-03──→ Node4:Node13 ──4.6e-02──→ Node4:t4 ──1.1e-03──→ Node6:Node2 ──5.3e-03──→ Node6:Node3 ──1.2e-02──→ Node6:Node12 ──3.8e-05──→ Node6:Node13 ──1.0e-03──→ Node6:t4 ──8.1e-03──→ t19:Node2 ──3.9e-02──→ t19:Node3 ──9.1e-02──→ t19:Node12 ──2.8e-04──→ t19:Node13 ──7.3e-03──→ t19:t4 
#> 
#> [2] ◎──5.0e-02──→ Node4:Node2 ──2.4e-01──→ Node4:Node3 ──5.7e-01──→ Node4:Node12 ──1.7e-03──→ Node4:Node13 ──4.6e-02──→ Node4:t14 ──1.1e-03──→ Node6:Node2 ──5.3e-03──→ Node6:Node3 ──1.2e-02──→ Node6:Node12 ──3.8e-05──→ Node6:Node13 ──1.0e-03──→ Node6:t14 ──6.8e-04──→ Node7:Node2 ──3.3e-03──→ Node7:Node3 ──7.7e-03──→ Node7:Node12 ──2.4e-05──→ Node7:Node13 ──6.2e-04──→ Node7:t14 ──1.8e-03──→ Node8:Node2 ──8.9e-03──→ Node8:Node3 ──2.1e-02──→ Node8:Node12 ──6.4e-05──→ Node8:Node13 ──1.7e-03──→ Node8:t14 ──5.6e-03──→ t5:Node2 ──2.7e-02──→ t5:Node3 ──6.3e-02──→ t5:Node12 ──1.9e-04──→ t5:Node13 ──5.1e-03──→ t5:t14 
#> 
#> [3] ◎──0.05045──→ Node4:Node2 ──0.24299──→ Node4:Node3 ──0.56817──→ Node4:Node12 ──0.04742──→ Node4:t5 ──0.00110──→ Node6:Node2 ──0.00531──→ Node6:Node3 ──0.01241──→ Node6:Node12 ──0.00104──→ Node6:t5 ──0.00068──→ Node7:Node2 ──0.00329──→ Node7:Node3 ──0.00770──→ Node7:Node12 ──0.00064──→ Node7:t5 ──0.00184──→ Node8:Node2 ──0.00885──→ Node8:Node3 ──0.02070──→ Node8:Node12 ──0.00173──→ Node8:t5 ──0.00206──→ Node12:Node2 ──0.00991──→ Node12:Node3 ──0.02318──→ Node12:Node12 ──0.00193──→ Node12:t5 ──0.00130──→ Node15:Node2 ──0.00627──→ Node15:Node3 ──0.01467──→ Node15:Node12 ──0.00122──→ Node15:t5 ──0.00223──→ t20:Node2 ──0.01073──→ t20:Node3 ──0.02510──→ t20:Node12 ──0.00209──→ t20:t5 
#> 
#> [4] ◎──0.05045──→ Node4:Node2 ──0.24299──→ Node4:Node3 ──0.30915──→ Node4:Node4 ──0.12751──→ Node4:Node8 ──0.17894──→ Node4:t16 ──0.00110──→ Node6:Node2 ──0.00531──→ Node6:Node3 ──0.00675──→ Node6:Node4 ──0.00279──→ Node6:Node8 ──0.00391──→ Node6:t16 ──0.00068──→ Node7:Node2 ──0.00329──→ Node7:Node3 ──0.00419──→ Node7:Node4 ──0.00173──→ Node7:Node8 ──0.00242──→ Node7:t16 ──0.00184──→ Node8:Node2 ──0.00885──→ Node8:Node3 ──0.01126──→ Node8:Node4 ──0.00464──→ Node8:Node8 ──0.00652──→ Node8:t16 ──0.00206──→ Node12:Node2 ──0.00991──→ Node12:Node3 ──0.01261──→ Node12:Node4 ──0.00520──→ Node12:Node8 ──0.00730──→ Node12:t16 ──0.00130──→ Node15:Node2 ──0.00627──→ Node15:Node3 ──0.00798──→ Node15:Node4 ──0.00329──→ Node15:Node8 ──0.00462──→ Node15:t16 ──0.00103──→ Node16:Node2 ──0.00494──→ Node16:Node3 ──0.00628──→ Node16:Node4 ──0.00259──→ Node16:Node8 ──0.00364──→ Node16:t16 ──0.00120──→ t18:Node2 ──0.00579──→ t18:Node3 ──0.00737──→ t18:Node4 ──0.00304──→ t18:Node8 ──0.00427──→ t18:t16 
#> 
#> [5] ◎──0.05045──→ Node4:Node2 ──0.24299──→ Node4:Node3 ──0.30915──→ Node4:Node4 ──0.12751──→ Node4:Node8 ──0.13589──→ Node4:Node14 ──0.04305──→ Node4:t1 ──0.00110──→ Node6:Node2 ──0.00531──→ Node6:Node3 ──0.00675──→ Node6:Node4 ──0.00279──→ Node6:Node8 ──0.00297──→ Node6:Node14 ──0.00094──→ Node6:t1 ──0.00068──→ Node7:Node2 ──0.00329──→ Node7:Node3 ──0.00419──→ Node7:Node4 ──0.00173──→ Node7:Node8 ──0.00184──→ Node7:Node14 ──0.00058──→ Node7:t1 ──0.00184──→ Node8:Node2 ──0.00885──→ Node8:Node3 ──0.01126──→ Node8:Node4 ──0.00464──→ Node8:Node8 ──0.00495──→ Node8:Node14 ──0.00157──→ Node8:t1 ──0.00206──→ Node12:Node2 ──0.00991──→ Node12:Node3 ──0.01261──→ Node12:Node4 ──0.00520──→ Node12:Node8 ──0.00554──→ Node12:Node14 ──0.00176──→ Node12:t1 ──0.00130──→ Node15:Node2 ──0.00627──→ Node15:Node3 ──0.00798──→ Node15:Node4 ──0.00329──→ Node15:Node8 ──0.00351──→ Node15:Node14 ──0.00111──→ Node15:t1 ──0.00103──→ Node16:Node2 ──0.00494──→ Node16:Node3 ──0.00628──→ Node16:Node4 ──0.00259──→ Node16:Node8 ──0.00276──→ Node16:Node14 ──0.00088──→ Node16:t1 ──0.00120──→ t4:Node2 ──0.00579──→ t4:Node3 ──0.00737──→ t4:Node4 ──0.00304──→ t4:Node8 ──0.00324──→ t4:Node14 ──0.00103──→ t4:t1 
#> 
#> [6] ◎──0.05045──→ Node4:Node2 ──0.24299──→ Node4:Node3 ──0.30915──→ Node4:Node4 ──0.12751──→ Node4:Node8 ──0.13589──→ Node4:Node14 ──0.00773──→ Node4:Node15 ──0.03532──→ Node4:t8 ──0.00110──→ Node6:Node2 ──0.00531──→ Node6:Node3 ──0.00675──→ Node6:Node4 ──0.00279──→ Node6:Node8 ──0.00297──→ Node6:Node14 ──0.00017──→ Node6:Node15 ──0.00077──→ Node6:t8 ──0.00068──→ Node7:Node2 ──0.00329──→ Node7:Node3 ──0.00419──→ Node7:Node4 ──0.00173──→ Node7:Node8 ──0.00184──→ Node7:Node14 ──0.00010──→ Node7:Node15 ──0.00048──→ Node7:t8 ──0.00184──→ Node8:Node2 ──0.00885──→ Node8:Node3 ──0.01126──→ Node8:Node4 ──0.00464──→ Node8:Node8 ──0.00495──→ Node8:Node14 ──0.00028──→ Node8:Node15 ──0.00129──→ Node8:t8 ──0.00206──→ Node12:Node2 ──0.00991──→ Node12:Node3 ──0.01261──→ Node12:Node4 ──0.00520──→ Node12:Node8 ──0.00554──→ Node12:Node14 ──0.00032──→ Node12:Node15 ──0.00144──→ Node12:t8 ──0.00078──→ Node13:Node2 ──0.00374──→ Node13:Node3 ──0.00476──→ Node13:Node4 ──0.00196──→ Node13:Node8 ──0.00209──→ Node13:Node14 ──0.00012──→ Node13:Node15 ──0.00054──→ Node13:t8 ──0.00275──→ t11:Node2 ──0.01327──→ t11:Node3 ──0.01688──→ t11:Node4 ──0.00696──→ t11:Node8 ──0.00742──→ t11:Node14 ──0.00042──→ t11:Node15 ──0.00193──→ t11:t8 
#> 
#> [7] ◎──5.0e-02──→ Node4:Node2 ──2.4e-01──→ Node4:Node3 ──3.1e-01──→ Node4:Node4 ──1.3e-01──→ Node4:Node8 ──1.4e-01──→ Node4:Node14 ──7.7e-03──→ Node4:Node15 ──3.0e-02──→ Node4:Node19 ──4.8e-03──→ Node4:t20 ──1.1e-03──→ Node6:Node2 ──5.3e-03──→ Node6:Node3 ──6.8e-03──→ Node6:Node4 ──2.8e-03──→ Node6:Node8 ──3.0e-03──→ Node6:Node14 ──1.7e-04──→ Node6:Node15 ──6.7e-04──→ Node6:Node19 ──1.1e-04──→ Node6:t20 ──6.8e-04──→ Node7:Node2 ──3.3e-03──→ Node7:Node3 ──4.2e-03──→ Node7:Node4 ──1.7e-03──→ Node7:Node8 ──1.8e-03──→ Node7:Node14 ──1.0e-04──→ Node7:Node15 ──4.1e-04──→ Node7:Node19 ──6.5e-05──→ Node7:t20 ──1.8e-03──→ Node8:Node2 ──8.9e-03──→ Node8:Node3 ──1.1e-02──→ Node8:Node4 ──4.6e-03──→ Node8:Node8 ──4.9e-03──→ Node8:Node14 ──2.8e-04──→ Node8:Node15 ──1.1e-03──→ Node8:Node19 ──1.8e-04──→ Node8:t20 ──2.1e-03──→ Node12:Node2 ──9.9e-03──→ Node12:Node3 ──1.3e-02──→ Node12:Node4 ──5.2e-03──→ Node12:Node8 ──5.5e-03──→ Node12:Node14 ──3.2e-04──→ Node12:Node15 ──1.2e-03──→ Node12:Node19 ──2.0e-04──→ Node12:t20 ──7.8e-04──→ Node13:Node2 ──3.7e-03──→ Node13:Node3 ──4.8e-03──→ Node13:Node4 ──2.0e-03──→ Node13:Node8 ──2.1e-03──→ Node13:Node14 ──1.2e-04──→ Node13:Node15 ──4.7e-04──→ Node13:Node19 ──7.4e-05──→ Node13:t20 ──1.6e-03──→ Node17:Node2 ──7.9e-03──→ Node17:Node3 ──1.0e-02──→ Node17:Node4 ──4.1e-03──→ Node17:Node8 ──4.4e-03──→ Node17:Node14 ──2.5e-04──→ Node17:Node15 ──9.9e-04──→ Node17:Node19 ──1.6e-04──→ Node17:t20 ──1.1e-03──→ t2:Node2 ──5.4e-03──→ t2:Node3 ──6.9e-03──→ t2:Node4 ──2.8e-03──→ t2:Node8 ──3.0e-03──→ t2:Node14 ──1.7e-04──→ t2:Node15 ──6.8e-04──→ t2:Node19 ──1.1e-04──→ t2:t20 
#> 
#> [8] ◎──5.0e-02──→ Node4:Node2 ──2.4e-01──→ Node4:Node3 ──3.1e-01──→ Node4:Node4 ──1.3e-01──→ Node4:Node8 ──1.4e-01──→ Node4:Node14 ──7.7e-03──→ Node4:Node15 ──3.0e-02──→ Node4:Node19 ──4.8e-03──→ Node4:t11 ──1.1e-03──→ Node6:Node2 ──5.3e-03──→ Node6:Node3 ──6.8e-03──→ Node6:Node4 ──2.8e-03──→ Node6:Node8 ──3.0e-03──→ Node6:Node14 ──1.7e-04──→ Node6:Node15 ──6.7e-04──→ Node6:Node19 ──1.1e-04──→ Node6:t11 ──6.8e-04──→ Node7:Node2 ──3.3e-03──→ Node7:Node3 ──4.2e-03──→ Node7:Node4 ──1.7e-03──→ Node7:Node8 ──1.8e-03──→ Node7:Node14 ──1.0e-04──→ Node7:Node15 ──4.1e-04──→ Node7:Node19 ──6.5e-05──→ Node7:t11 ──1.8e-03──→ Node8:Node2 ──8.9e-03──→ Node8:Node3 ──1.1e-02──→ Node8:Node4 ──4.6e-03──→ Node8:Node8 ──4.9e-03──→ Node8:Node14 ──2.8e-04──→ Node8:Node15 ──1.1e-03──→ Node8:Node19 ──1.8e-04──→ Node8:t11 ──2.1e-03──→ Node12:Node2 ──9.9e-03──→ Node12:Node3 ──1.3e-02──→ Node12:Node4 ──5.2e-03──→ Node12:Node8 ──5.5e-03──→ Node12:Node14 ──3.2e-04──→ Node12:Node15 ──1.2e-03──→ Node12:Node19 ──2.0e-04──→ Node12:t11 ──7.8e-04──→ Node13:Node2 ──3.7e-03──→ Node13:Node3 ──4.8e-03──→ Node13:Node4 ──2.0e-03──→ Node13:Node8 ──2.1e-03──→ Node13:Node14 ──1.2e-04──→ Node13:Node15 ──4.7e-04──→ Node13:Node19 ──7.4e-05──→ Node13:t11 ──1.6e-03──→ Node17:Node2 ──7.9e-03──→ Node17:Node3 ──1.0e-02──→ Node17:Node4 ──4.1e-03──→ Node17:Node8 ──4.4e-03──→ Node17:Node14 ──2.5e-04──→ Node17:Node15 ──9.9e-04──→ Node17:Node19 ──1.6e-04──→ Node17:t11 ──1.1e-03──→ t14:Node2 ──5.4e-03──→ t14:Node3 ──6.9e-03──→ t14:Node4 ──2.8e-03──→ t14:Node8 ──3.0e-03──→ t14:Node14 ──1.7e-04──→ t14:Node15 ──6.8e-04──→ t14:Node19 ──1.1e-04──→ t14:t11 
#> 
#> [9] ◎──0.05045──→ Node4:Node2 ──0.24299──→ Node4:Node3 ──0.30915──→ Node4:Node4 ──0.02436──→ Node4:Node5 ──0.10882──→ Node4:Node9 ──0.15744──→ Node4:Node17 ──0.01582──→ Node4:t9 ──0.00110──→ Node6:Node2 ──0.00531──→ Node6:Node3 ──0.00675──→ Node6:Node4 ──0.00053──→ Node6:Node5 ──0.00238──→ Node6:Node9 ──0.00344──→ Node6:Node17 ──0.00035──→ Node6:t9 ──0.00068──→ Node7:Node2 ──0.00329──→ Node7:Node3 ──0.00419──→ Node7:Node4 ──0.00033──→ Node7:Node5 ──0.00147──→ Node7:Node9 ──0.00213──→ Node7:Node17 ──0.00021──→ Node7:t9 ──0.00187──→ Node9:Node2 ──0.00902──→ Node9:Node3 ──0.01148──→ Node9:Node4 ──0.00090──→ Node9:Node5 ──0.00404──→ Node9:Node9 ──0.00585──→ Node9:Node17 ──0.00059──→ Node9:t9 ──0.00119──→ Node10:Node2 ──0.00573──→ Node10:Node3 ──0.00729──→ Node10:Node4 ──0.00057──→ Node10:Node5 ──0.00257──→ Node10:Node9 ──0.00372──→ Node10:Node17 ──0.00037──→ Node10:t9 ──0.00436──→ t3:Node2 ──0.02101──→ t3:Node3 ──0.02673──→ t3:Node4 ──0.00211──→ t3:Node5 ──0.00941──→ t3:Node9 ──0.01361──→ t3:Node17 ──0.00137──→ t3:t9 
#> 
#> [10] ◎──0.05045──→ Node4:Node2 ──0.24299──→ Node4:Node3 ──0.30915──→ Node4:Node4 ──0.02436──→ Node4:Node5 ──0.10882──→ Node4:Node9 ──0.15744──→ Node4:Node17 ──0.01582──→ Node4:t7 ──0.00110──→ Node6:Node2 ──0.00531──→ Node6:Node3 ──0.00675──→ Node6:Node4 ──0.00053──→ Node6:Node5 ──0.00238──→ Node6:Node9 ──0.00344──→ Node6:Node17 ──0.00035──→ Node6:t7 ──0.00068──→ Node7:Node2 ──0.00329──→ Node7:Node3 ──0.00419──→ Node7:Node4 ──0.00033──→ Node7:Node5 ──0.00147──→ Node7:Node9 ──0.00213──→ Node7:Node17 ──0.00021──→ Node7:t7 ──0.00187──→ Node9:Node2 ──0.00902──→ Node9:Node3 ──0.01148──→ Node9:Node4 ──0.00090──→ Node9:Node5 ──0.00404──→ Node9:Node9 ──0.00585──→ Node9:Node17 ──0.00059──→ Node9:t7 ──0.00119──→ Node10:Node2 ──0.00573──→ Node10:Node3 ──0.00729──→ Node10:Node4 ──0.00057──→ Node10:Node5 ──0.00257──→ Node10:Node9 ──0.00372──→ Node10:Node17 ──0.00037──→ Node10:t7 ──0.00436──→ t7:Node2 ──0.02101──→ t7:Node3 ──0.02673──→ t7:Node4 ──0.00211──→ t7:Node5 ──0.00941──→ t7:Node9 ──0.01361──→ t7:Node17 ──0.00137──→ t7:t7