Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Calculate a rowwise kronecker product when the multiplicand or multiplier is a pfc.

Usage

pf_row_kron(x, y, ...)

Arguments

x

The rowwise kronecker multiplicand

y

The rowwise kronecker multiplier

...

Other arguments passed to or from other methods.

Value

A pfc object.

Examples

pf_row_kron(rpfc(20), rpfc(20))
#> <pfc<e:1444>[38]>
#> First  10 phylogenetic flows: 
#> [1] ◎──9.4155──→ Node3:Node14 ──0.0177──→ Node3:Node16 ──0.1390──→ Node3:t3 ──0.9881──→ Node6:Node14 ──0.0019──→ Node6:Node16 ──0.0146──→ Node6:t3 ──0.7258──→ t9:Node14 ──0.0014──→ t9:Node16 ──0.0107──→ t9:t3 
#> 
#> [2] ◎──9.41551──→ Node3:Node14 ──0.01774──→ Node3:Node16 ──0.13903──→ Node3:t6 ──0.98806──→ Node6:Node14 ──0.00186──→ Node6:Node16 ──0.01459──→ Node6:t6 ──0.36524──→ Node11:Node14 ──0.00069──→ Node11:Node16 ──0.00539──→ Node11:t6 ──0.14119──→ Node12:Node14 ──0.00027──→ Node12:Node16 ──0.00208──→ Node12:t6 ──0.21940──→ t13:Node14 ──0.00041──→ t13:Node16 ──0.00324──→ t13:t6 
#> 
#> [3] ◎──9.4155──→ Node3:Node14 ──0.1568──→ Node3:t2 ──0.9881──→ Node6:Node14 ──0.0165──→ Node6:t2 ──0.3652──→ Node11:Node14 ──0.0061──→ Node11:t2 ──0.1412──→ Node12:Node14 ──0.0024──→ Node12:t2 ──0.2194──→ t18:Node14 ──0.0037──→ t18:t2 
#> 
#> [4] ◎──6.5146──→ Node3:Node2 ──0.3352──→ Node3:Node3 ──2.6290──→ Node3:Node19 ──0.0936──→ Node3:t15 ──0.6836──→ Node6:Node2 ──0.0352──→ Node6:Node3 ──0.2759──→ Node6:Node19 ──0.0098──→ Node6:t15 ──0.2527──→ Node11:Node2 ──0.0130──→ Node11:Node3 ──0.1020──→ Node11:Node19 ──0.0036──→ Node11:t15 ──0.2495──→ t20:Node2 ──0.0128──→ t20:Node3 ──0.1007──→ t20:Node19 ──0.0036──→ t20:t15 
#> 
#> [5] ◎──6.5146──→ Node3:Node2 ──0.3352──→ Node3:Node3 ──2.6290──→ Node3:Node19 ──0.0936──→ Node3:t14 ──1.1099──→ Node16:Node2 ──0.0571──→ Node16:Node3 ──0.4479──→ Node16:Node19 ──0.0159──→ Node16:t14 ──0.0759──→ t8:Node2 ──0.0039──→ t8:Node3 ──0.0306──→ t8:Node19 ──0.0011──→ t8:t14 
#> 
#> [6] ◎──6.5146──→ Node3:Node2 ──0.3352──→ Node3:Node3 ──0.5098──→ Node3:Node4 ──0.7596──→ Node3:Node5 ──0.9800──→ Node3:Node7 ──0.3200──→ Node3:Node15 ──0.1531──→ Node3:t10 ──1.1099──→ Node16:Node2 ──0.0571──→ Node16:Node3 ──0.0869──→ Node16:Node4 ──0.1294──→ Node16:Node5 ──0.1670──→ Node16:Node7 ──0.0545──→ Node16:Node15 ──0.0261──→ Node16:t10 ──0.0759──→ t2:Node2 ──0.0039──→ t2:Node3 ──0.0059──→ t2:Node4 ──0.0089──→ t2:Node5 ──0.0114──→ t2:Node7 ──0.0037──→ t2:Node15 ──0.0018──→ t2:t10 
#> 
#> [7] ◎──4.85843──→ Node2:Node2 ──0.24997──→ Node2:Node3 ──0.38022──→ Node2:Node4 ──0.56647──→ Node2:Node5 ──0.73086──→ Node2:Node7 ──0.23867──→ Node2:Node15 ──0.11416──→ Node2:t4 ──2.21073──→ Node5:Node2 ──0.11374──→ Node5:Node3 ──0.17301──→ Node5:Node4 ──0.25776──→ Node5:Node5 ──0.33256──→ Node5:Node7 ──0.10860──→ Node5:Node15 ──0.05195──→ Node5:t4 ──0.62482──→ Node19:Node2 ──0.03215──→ Node19:Node3 ──0.04890──→ Node19:Node4 ──0.07285──→ Node19:Node5 ──0.09399──→ Node19:Node7 ──0.03069──→ Node19:Node15 ──0.01468──→ Node19:t4 ──0.00645──→ t1:Node2 ──0.00033──→ t1:Node3 ──0.00051──→ t1:Node4 ──0.00075──→ t1:Node5 ──0.00097──→ t1:Node7 ──0.00032──→ t1:Node15 ──0.00015──→ t1:t4 
#> 
#> [8] ◎──4.85843──→ Node2:Node2 ──0.24997──→ Node2:Node3 ──0.38022──→ Node2:Node4 ──0.56647──→ Node2:Node5 ──0.73086──→ Node2:Node7 ──0.35284──→ Node2:t12 ──2.21073──→ Node5:Node2 ──0.11374──→ Node5:Node3 ──0.17301──→ Node5:Node4 ──0.25776──→ Node5:Node5 ──0.33256──→ Node5:Node7 ──0.16055──→ Node5:t12 ──0.62482──→ Node19:Node2 ──0.03215──→ Node19:Node3 ──0.04890──→ Node19:Node4 ──0.07285──→ Node19:Node5 ──0.09399──→ Node19:Node7 ──0.04538──→ Node19:t12 ──0.00645──→ t14:Node2 ──0.00033──→ t14:Node3 ──0.00051──→ t14:Node4 ──0.00075──→ t14:Node5 ──0.00097──→ t14:Node7 ──0.00047──→ t14:t12 
#> 
#> [9] ◎──4.8584──→ Node2:Node2 ──0.2500──→ Node2:Node3 ──0.3802──→ Node2:Node4 ──0.5665──→ Node2:Node5 ──0.9489──→ Node2:Node13 ──0.0550──→ Node2:Node18 ──0.0798──→ Node2:t9 ──2.2107──→ Node5:Node2 ──0.1137──→ Node5:Node3 ──0.1730──→ Node5:Node4 ──0.2578──→ Node5:Node5 ──0.4318──→ Node5:Node13 ──0.0250──→ Node5:Node18 ──0.0363──→ Node5:t9 ──0.1429──→ Node7:Node2 ──0.0074──→ Node7:Node3 ──0.0112──→ Node7:Node4 ──0.0167──→ Node7:Node5 ──0.0279──→ Node7:Node13 ──0.0016──→ Node7:Node18 ──0.0023──→ Node7:t9 ──0.4884──→ t7:Node2 ──0.0251──→ t7:Node3 ──0.0382──→ t7:Node4 ──0.0569──→ t7:Node5 ──0.0954──→ t7:Node13 ──0.0055──→ t7:Node18 ──0.0080──→ t7:t9 
#> 
#> [10] ◎──4.9e+00──→ Node2:Node2 ──2.5e-01──→ Node2:Node3 ──3.8e-01──→ Node2:Node4 ──5.7e-01──→ Node2:Node5 ──9.5e-01──→ Node2:Node13 ──5.5e-02──→ Node2:Node18 ──8.0e-02──→ Node2:t13 ──2.2e+00──→ Node5:Node2 ──1.1e-01──→ Node5:Node3 ──1.7e-01──→ Node5:Node4 ──2.6e-01──→ Node5:Node5 ──4.3e-01──→ Node5:Node13 ──2.5e-02──→ Node5:Node18 ──3.6e-02──→ Node5:t13 ──1.4e-01──→ Node7:Node2 ──7.4e-03──→ Node7:Node3 ──1.1e-02──→ Node7:Node4 ──1.7e-02──→ Node7:Node5 ──2.8e-02──→ Node7:Node13 ──1.6e-03──→ Node7:Node18 ──2.3e-03──→ Node7:t13 ──7.5e-03──→ Node8:Node2 ──3.9e-04──→ Node8:Node3 ──5.9e-04──→ Node8:Node4 ──8.7e-04──→ Node8:Node5 ──1.5e-03──→ Node8:Node13 ──8.5e-05──→ Node8:Node18 ──1.2e-04──→ Node8:t13 ──4.3e-01──→ Node17:Node2 ──2.2e-02──→ Node17:Node3 ──3.4e-02──→ Node17:Node4 ──5.0e-02──→ Node17:Node5 ──8.5e-02──→ Node17:Node13 ──4.9e-03──→ Node17:Node18 ──7.1e-03──→ Node17:t13 ──4.8e-02──→ t19:Node2 ──2.5e-03──→ t19:Node3 ──3.7e-03──→ t19:Node4 ──5.6e-03──→ t19:Node5 ──9.3e-03──→ t19:Node13 ──5.4e-04──→ t19:Node18 ──7.9e-04──→ t19:t13