Scale the phylogenetic flow features to a constant sum.

Can be used to standardise branch lengths to reasonable values

Usage

pf_scale_flow_sum(x, scale_to = 1)

Arguments

x

a pfc object

scale_to

The value to scale the sums to. Default is 1.

Value

a scaled pfc object

Examples

pf_scale_flow_sum(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] --0.6608->Node3--0.3073->Node33--0.0318->t90                                                                                                                                                                             
#>   [2] --0.6608->Node3--0.3073->Node33--0.0318->t15                                                                                                                                                                             
#>   [3] --0.6608->Node3--0.0490->Node4--0.1305->Node7--0.0780->Node13--0.0316->Node22--0.0500->t9                                                                                                                                
#>   [4] --0.6608->Node3--0.0490->Node4--0.1305->Node7--0.0780->Node13--0.0316->Node22--0.0500->t95                                                                                                                               
#>   [5] -0.660844>Node3-0.049039>Node4-0.130535>Node7-0.077970>Node13-0.013273>Node15-0.035986>Node32-0.027117>Node79-0.003011>Node87-0.000538>Node90-0.001687>t34                                                               
#>   [6] -0.660844>Node3-0.049039>Node4-0.130535>Node7-0.077970>Node13-0.013273>Node15-0.035986>Node32-0.027117>Node79-0.003011>Node87-0.000538>Node90-0.001687>t69                                                               
#>   [7] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.07797->Node13-0.01327->Node15-0.03599->Node32-0.02712->Node79-0.00301->Node87-0.00223->t35                                                                               
#>   [8] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.07797->Node13-0.01327->Node15-0.03599->Node32-0.02712->Node79-0.00524->t83                                                                                               
#>   [9] --0.6608->Node3--0.0490->Node4--0.1305->Node7--0.0780->Node13--0.0133->Node15--0.0360->Node32--0.0324->t75                                                                                                               
#>  [10] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.07797->Node13-0.01327->Node15-0.04110->Node37-0.01472->Node57-0.00705->Node76-0.00548->t27                                                                               
#>  [11] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.07797->Node13-0.01327->Node15-0.04110->Node37-0.01472->Node57-0.00705->Node76-0.00548->t24                                                                               
#>  [12] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.07797->Node13-0.01327->Node15-0.04110->Node37-0.01472->Node57-0.00413->Node66-0.00839->t41                                                                               
#>  [13] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.07797->Node13-0.01327->Node15-0.04110->Node37-0.01472->Node57-0.00413->Node66-0.00149->Node70-0.00690->t38                                                               
#>  [14] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.07797->Node13-0.01327->Node15-0.04110->Node37-0.01472->Node57-0.00413->Node66-0.00149->Node70-0.00690->t53                                                               
#>  [15] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.07797->Node13-0.01327->Node15-0.04110->Node37-0.00153->Node38-0.02571->t19                                                                                               
#>  [16] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.07797->Node13-0.01327->Node15-0.04110->Node37-0.00153->Node38-0.00979->Node52-0.01593->t36                                                                               
#>  [17] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.07797->Node13-0.01327->Node15-0.04110->Node37-0.00153->Node38-0.00979->Node52-0.00990->Node74-0.00602->t60                                                               
#>  [18] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.07797->Node13-0.01327->Node15-0.04110->Node37-0.00153->Node38-0.00979->Node52-0.00990->Node74-0.00602->t5                                                                
#>  [19] --0.6608->Node3--0.0490->Node4--0.1305->Node7--0.0100->Node8--0.0937->Node18--0.0188->Node29--0.0371->t54                                                                                                                
#>  [20] --0.6608->Node3--0.0490->Node4--0.1305->Node7--0.0100->Node8--0.0937->Node18--0.0188->Node29--0.0371->t80                                                                                                                
#>  [21] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.01004->Node8-0.09368->Node18-0.00303->Node19-0.03147->Node43-0.00317->Node47-0.01024->Node67-0.00494->Node85-0.00300->t96                                                
#>  [22] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.01004->Node8-0.09368->Node18-0.00303->Node19-0.03147->Node43-0.00317->Node47-0.01024->Node67-0.00494->Node85-0.00300->t79                                                
#>  [23] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.01004->Node8-0.09368->Node18-0.00303->Node19-0.03147->Node43-0.00317->Node47-0.01024->Node67-0.00795->t59                                                                
#>  [24] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.01004->Node8-0.09368->Node18-0.00303->Node19-0.03147->Node43-0.00317->Node47-0.01819->t18                                                                                
#>  [25] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.01004->Node8-0.09368->Node18-0.00303->Node19-0.03147->Node43-0.02136->t37                                                                                                
#>  [26] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.01004->Node8-0.09368->Node18-0.00303->Node19-0.00406->Node24-0.04877->t7                                                                                                 
#>  [27] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.01004->Node8-0.09368->Node18-0.00303->Node19-0.00406->Node24-0.02139->Node36-0.02738->t40                                                                                
#>  [28] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.01004->Node8-0.09368->Node18-0.00303->Node19-0.00406->Node24-0.02139->Node36-0.01257->Node54-0.01481->t1                                                                 
#>  [29] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.01004->Node8-0.09368->Node18-0.00303->Node19-0.00406->Node24-0.02139->Node36-0.01257->Node54-0.01481->t66                                                                
#>  [30] --0.6608->Node3--0.0490->Node4--0.1305->Node7--0.0100->Node8--0.0694->Node14--0.0505->Node34--0.0297->t61                                                                                                                
#>  [31] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.01004->Node8-0.06939->Node14-0.05046->Node34-0.02057->Node63-0.00912->t10                                                                                                
#>  [32] -0.660844>Node3-0.049039>Node4-0.130535>Node7-0.010036>Node8-0.069394>Node14-0.050464>Node34-0.020570>Node63-0.000418>Node65-0.003631>Node80-0.001346>Node83-0.003723>t23                                                
#>  [33] -0.660844>Node3-0.049039>Node4-0.130535>Node7-0.010036>Node8-0.069394>Node14-0.050464>Node34-0.020570>Node63-0.000418>Node65-0.003631>Node80-0.001346>Node83-0.001692>Node89-0.002031>t87                                
#>  [34] -0.660844>Node3-0.049039>Node4-0.130535>Node7-0.010036>Node8-0.069394>Node14-0.050464>Node34-0.020570>Node63-0.000418>Node65-0.003631>Node80-0.001346>Node83-0.001692>Node89-0.002031>t55                                
#>  [35] -0.660844>Node3-0.049039>Node4-0.130535>Node7-0.010036>Node8-0.069394>Node14-0.050464>Node34-0.020570>Node63-0.000418>Node65-0.003631>Node80-0.005069>t45                                                                
#>  [36] -0.660844>Node3-0.049039>Node4-0.130535>Node7-0.010036>Node8-0.069394>Node14-0.050464>Node34-0.020570>Node63-0.000418>Node65-0.008700>t21                                                                                
#>  [37] -0.660844>Node3-0.049039>Node4-0.130535>Node7-0.010036>Node8-0.069394>Node14-0.045745>Node30-0.009943>Node39-0.024421>Node98-0.000042>t85                                                                                
#>  [38] -0.660844>Node3-0.049039>Node4-0.130535>Node7-0.010036>Node8-0.069394>Node14-0.045745>Node30-0.009943>Node39-0.024421>Node98-0.000042>t67                                                                                
#>  [39] -0.66084->Node3-0.04904->Node4-0.13054->Node7-0.01004->Node8-0.06939->Node14-0.04575->Node30-0.00994->Node39-0.00747->Node48-0.01700->t84                                                                                
#>  [40] -0.660844>Node3-0.049039>Node4-0.130535>Node7-0.010036>Node8-0.069394>Node14-0.045745>Node30-0.009943>Node39-0.007468>Node48-0.000229>Node49-0.016767>t48                                                                
#>  [41] -0.660844>Node3-0.049039>Node4-0.130535>Node7-0.010036>Node8-0.069394>Node14-0.045745>Node30-0.009943>Node39-0.007468>Node48-0.000229>Node49-0.007104>Node62-0.002917>Node71-0.006746>t76                                
#>  [42] -0.660844>Node3-0.049039>Node4-0.130535>Node7-0.010036>Node8-0.069394>Node14-0.045745>Node30-0.009943>Node39-0.007468>Node48-0.000229>Node49-0.007104>Node62-0.002917>Node71-0.000860>Node75-0.005886>t31                
#>  [43] -0.660844>Node3-0.049039>Node4-0.130535>Node7-0.010036>Node8-0.069394>Node14-0.045745>Node30-0.009943>Node39-0.007468>Node48-0.000229>Node49-0.007104>Node62-0.002917>Node71-0.000860>Node75-0.003019>Node86-0.002866>t86
#>  [44] -0.660844>Node3-0.049039>Node4-0.130535>Node7-0.010036>Node8-0.069394>Node14-0.045745>Node30-0.009943>Node39-0.007468>Node48-0.000229>Node49-0.007104>Node62-0.002917>Node71-0.000860>Node75-0.003019>Node86-0.002866>t81
#>  [45] -0.660844>Node3-0.049039>Node4-0.130535>Node7-0.010036>Node8-0.069394>Node14-0.045745>Node30-0.009943>Node39-0.007468>Node48-0.000229>Node49-0.007104>Node62-0.009662>t26                                                
#>  [46] --0.6608->Node3--0.0490->Node4--0.1305->Node7--0.0100->Node8--0.0694->Node14--0.0457->Node30--0.0344->t28                                                                                                                
#>  [47] --0.661-->Node3--0.049-->Node4--0.176-->Node11--0.114-->t43                                                                                                                                                              
#>  [48] --0.661-->Node3--0.049-->Node4--0.176-->Node11--0.114-->t99                                                                                                                                                              
#>  [49] -0.56546->Node2-0.19987->Node6-0.11451->Node9-0.05591->Node16-0.04779->Node50-0.00548->Node60-0.00567->Node78-0.00531->t89                                                                                               
#>  [50] -0.56546->Node2-0.19987->Node6-0.11451->Node9-0.05591->Node16-0.04779->Node50-0.00548->Node60-0.00567->Node78-0.00531->t16                                                                                               
#>  [51] -0.565457>Node2-0.199867>Node6-0.114515>Node9-0.055908>Node16-0.047791>Node50-0.005485>Node60-0.010222>Node95-0.000756>t39                                                                                               
#>  [52] -0.565457>Node2-0.199867>Node6-0.114515>Node9-0.055908>Node16-0.047791>Node50-0.005485>Node60-0.010222>Node95-0.000756>t58                                                                                               
#>  [53] -0.565457>Node2-0.199867>Node6-0.114515>Node9-0.055908>Node16-0.047791>Node50-0.000367>Node51-0.002590>Node56-0.012098>Node93-0.001409>t46                                                                               
#>  [54] -0.565457>Node2-0.199867>Node6-0.114515>Node9-0.055908>Node16-0.047791>Node50-0.000367>Node51-0.002590>Node56-0.012098>Node93-0.001409>t77                                                                               
#>  [55] -0.565457>Node2-0.199867>Node6-0.114515>Node9-0.055908>Node16-0.047791>Node50-0.000367>Node51-0.002590>Node56-0.009029>Node81-0.004478>t14                                                                               
#>  [56] -0.565457>Node2-0.199867>Node6-0.114515>Node9-0.055908>Node16-0.047791>Node50-0.000367>Node51-0.002590>Node56-0.009029>Node81-0.004478>t3                                                                                
#>  [57] -0.565457>Node2-0.199867>Node6-0.114515>Node9-0.055908>Node16-0.047791>Node50-0.000367>Node51-0.013057>Node84-0.003039>t25                                                                                               
#>  [58] -5.65e-01>Node2-2.00e-01>Node6-1.15e-01>Node9-5.59e-02>Node16-4.78e-02>Node50-3.67e-04>Node51-1.31e-02>Node84-2.77e-03>Node97-2.45e-04>Node99-2.19e-05>t82                                                               
#>  [59] -5.65e-01>Node2-2.00e-01>Node6-1.15e-01>Node9-5.59e-02>Node16-4.78e-02>Node50-3.67e-04>Node51-1.31e-02>Node84-2.77e-03>Node97-2.45e-04>Node99-2.19e-05>t57                                                               
#>  [60] -0.565457>Node2-0.199867>Node6-0.114515>Node9-0.055908>Node16-0.047791>Node50-0.000367>Node51-0.013057>Node84-0.002772>Node97-0.000267>t33                                                                               
#>  [61] --0.5655->Node2--0.1999->Node6--0.1145->Node9--0.0559->Node16--0.0147->Node23--0.0309->Node46--0.0187->t22                                                                                                               
#>  [62] --0.5655->Node2--0.1999->Node6--0.1145->Node9--0.0559->Node16--0.0147->Node23--0.0309->Node46--0.0187->t20                                                                                                               
#>  [63] --0.5655->Node2--0.1999->Node6--0.1145->Node9--0.0559->Node16--0.0147->Node23--0.0304->Node45--0.0192->t52                                                                                                               
#>  [64] --0.5655->Node2--0.1999->Node6--0.1145->Node9--0.0559->Node16--0.0147->Node23--0.0304->Node45--0.0192->t72                                                                                                               
#>  [65] -0.565457>Node2-0.199867>Node6-0.114515>Node9-0.000706>Node10-0.085866>Node31-0.033589>t17                                                                                                                               
#>  [66] -0.565457>Node2-0.199867>Node6-0.114515>Node9-0.000706>Node10-0.085866>Node31-0.025762>Node68-0.005657>Node88-0.002170>t74                                                                                               
#>  [67] -0.565457>Node2-0.199867>Node6-0.114515>Node9-0.000706>Node10-0.085866>Node31-0.025762>Node68-0.005657>Node88-0.000899>Node94-0.001271>t56                                                                               
#>  [68] -0.565457>Node2-0.199867>Node6-0.114515>Node9-0.000706>Node10-0.085866>Node31-0.025762>Node68-0.005657>Node88-0.000899>Node94-0.001271>t91                                                                               
#>  [69] -0.565457>Node2-0.199867>Node6-0.114515>Node9-0.000706>Node10-0.085866>Node31-0.025762>Node68-0.007827>t11                                                                                                               
#>  [70] -0.565457>Node2-0.199867>Node6-0.114515>Node9-0.000706>Node10-0.105403>Node55-0.014052>t12                                                                                                                               
#>  [71] -0.565457>Node2-0.199867>Node6-0.114515>Node9-0.000706>Node10-0.105403>Node55-0.014052>t50                                                                                                                               
#>  [72] -0.56546->Node2-0.19987->Node6-0.17870->Node17-0.00439->Node20-0.03165->Node44-0.01349->Node73-0.00645->t93                                                                                                              
#>  [73] -0.56546->Node2-0.19987->Node6-0.17870->Node17-0.00439->Node20-0.03165->Node44-0.01349->Node73-0.00645->t32                                                                                                              
#>  [74] -0.565457>Node2-0.199867>Node6-0.178698>Node17-0.004390>Node20-0.031647>Node44-0.008624>Node59-0.000411>Node61-0.010907>t51                                                                                              
#>  [75] -0.565457>Node2-0.199867>Node6-0.178698>Node17-0.004390>Node20-0.031647>Node44-0.008624>Node59-0.000411>Node61-0.010907>t94                                                                                              
#>  [76] -0.56546->Node2-0.19987->Node6-0.17870->Node17-0.00439->Node20-0.03165->Node44-0.00862->Node59-0.01132->t71                                                                                                              
#>  [77] -0.56546->Node2-0.19987->Node6-0.17870->Node17-0.00439->Node20-0.03995->Node58-0.00452->Node69-0.00173->Node77-0.00539->t98                                                                                              
#>  [78] -0.56546->Node2-0.19987->Node6-0.17870->Node17-0.00439->Node20-0.03995->Node58-0.00452->Node69-0.00173->Node77-0.00539->t78                                                                                              
#>  [79] -0.56546->Node2-0.19987->Node6-0.17870->Node17-0.00439->Node20-0.03995->Node58-0.00452->Node69-0.00712->t49                                                                                                              
#>  [80] -0.56546->Node2-0.19987->Node6-0.17870->Node17-0.00439->Node20-0.03995->Node58-0.01164->t2                                                                                                                               
#>  [81] -0.56546->Node2-0.19987->Node6-0.17870->Node17-0.00846->Node25-0.00889->Node28-0.03862->t29                                                                                                                              
#>  [82] -0.56546->Node2-0.19987->Node6-0.17870->Node17-0.00846->Node25-0.00889->Node28-0.03862->t100                                                                                                                             
#>  [83] -0.56546->Node2-0.19987->Node6-0.17870->Node17-0.00846->Node25-0.00293->Node26-0.04296->Node92-0.00162->t42                                                                                                              
#>  [84] -0.56546->Node2-0.19987->Node6-0.17870->Node17-0.00846->Node25-0.00293->Node26-0.04296->Node92-0.00162->t30                                                                                                              
#>  [85] -0.56546->Node2-0.19987->Node6-0.17870->Node17-0.00846->Node25-0.00293->Node26-0.04459->t68                                                                                                                              
#>  [86] -0.56546->Node2-0.18951->Node5-0.19468->Node21-0.02687->Node42-0.01692->Node72-0.00656->t4                                                                                                                               
#>  [87] -0.56546->Node2-0.18951->Node5-0.19468->Node21-0.02687->Node42-0.01692->Node72-0.00494->Node91-0.00163->t73                                                                                                              
#>  [88] -0.56546->Node2-0.18951->Node5-0.19468->Node21-0.02687->Node42-0.01692->Node72-0.00494->Node91-0.00163->t62                                                                                                              
#>  [89] --0.5655->Node2--0.1895->Node5--0.1947->Node21--0.0269->Node42--0.0235->t97                                                                                                                                              
#>  [90] --0.5655->Node2--0.1895->Node5--0.1947->Node21--0.0348->Node53--0.0156->t63                                                                                                                                              
#>  [91] --0.5655->Node2--0.1895->Node5--0.1947->Node21--0.0348->Node53--0.0156->t13                                                                                                                                              
#>  [92] -0.56546->Node2-0.18951->Node5-0.14295->Node12-0.07342->Node35-0.00442->Node40-0.02424->t92                                                                                                                              
#>  [93] -0.56546->Node2-0.18951->Node5-0.14295->Node12-0.07342->Node35-0.00442->Node40-0.01515->Node64-0.00472->Node82-0.00437->t8                                                                                               
#>  [94] -0.56546->Node2-0.18951->Node5-0.14295->Node12-0.07342->Node35-0.00442->Node40-0.01515->Node64-0.00472->Node82-0.00437->t64                                                                                              
#>  [95] -0.56546->Node2-0.18951->Node5-0.14295->Node12-0.07342->Node35-0.00442->Node40-0.01515->Node64-0.00910->t65                                                                                                              
#>  [96] --0.5655->Node2--0.1895->Node5--0.1429->Node12--0.0734->Node35--0.0287->t88                                                                                                                                              
#>  [97] -0.565457>Node2-0.189511>Node5-0.142945>Node12-0.060339>Node27-0.041044>Node96-0.000704>t70                                                                                                                              
#>  [98] -0.565457>Node2-0.189511>Node5-0.142945>Node12-0.060339>Node27-0.041044>Node96-0.000704>t44                                                                                                                              
#>  [99] --0.5655->Node2--0.1895->Node5--0.1429->Node12--0.0603->Node27--0.0179->Node41--0.0239->t47                                                                                                                              
#> [100] --0.5655->Node2--0.1895->Node5--0.1429->Node12--0.0603->Node27--0.0179->Node41--0.0239->t6                                                                                                                               
#> [101] ----1---->Node2                                                                                                                                                                                                          
#> [102] ----1---->Node3                                                                                                                                                                                                          
#> [103] --0.9309->Node3--0.0691->Node4                                                                                                                                                                                           
#> [104] --0.749-->Node2--0.251-->Node5                                                                                                                                                                                           
#> [105] --0.739-->Node2--0.261-->Node6                                                                                                                                                                                           
#> [106] --0.7863->Node3--0.0584->Node4--0.1553->Node7                                                                                                                                                                            
#> [107] --0.7770->Node3--0.0577->Node4--0.1535->Node7--0.0118->Node8                                                                                                                                                             
#> [108] --0.643-->Node2--0.227-->Node6--0.130-->Node9                                                                                                                                                                            
#> [109] -0.642167>Node2-0.226981>Node6-0.130050>Node9-0.000802>Node10                                                                                                                                                            
#> [110] --0.7461->Node3--0.0554->Node4--0.1985->Node11                                                                                                                                                                           
#> [111] --0.630-->Node2--0.211-->Node5--0.159-->Node12                                                                                                                                                                           
#> [112] --0.7196->Node3--0.0534->Node4--0.1421->Node7--0.0849->Node13                                                                                                                                                            
#> [113] --0.7184->Node3--0.0533->Node4--0.1419->Node7--0.0109->Node8--0.0754->Node14                                                                                                                                             
#> [114] --0.7093->Node3--0.0526->Node4--0.1401->Node7--0.0837->Node13--0.0142->Node15                                                                                                                                            
#> [115] --0.6043->Node2--0.2136->Node6--0.1224->Node9--0.0597->Node16                                                                                                                                                            
#> [116] --0.599-->Node2--0.212-->Node6--0.189-->Node17                                                                                                                                                                           
#> [117] --0.6999->Node3--0.0519->Node4--0.1383->Node7--0.0106->Node8--0.0992->Node18                                                                                                                                             
#> [118] --0.6977->Node3--0.0518->Node4--0.1378->Node7--0.0106->Node8--0.0989->Node18--0.0032->Node19                                                                                                                             
#> [119] -0.59621->Node2-0.21074->Node6-0.18842->Node17-0.00463->Node20                                                                                                                                                           
#> [120] --0.595-->Node2--0.200-->Node5--0.205-->Node21                                                                                                                                                                           
#> [121] --0.6956->Node3--0.0516->Node4--0.1374->Node7--0.0821->Node13--0.0333->Node22                                                                                                                                            
#> [122] --0.5950->Node2--0.2103->Node6--0.1205->Node9--0.0588->Node16--0.0154->Node23                                                                                                                                            
#> [123] -0.69473->Node3-0.05155->Node4-0.13723->Node7-0.01055->Node8-0.09849->Node18-0.00319->Node19-0.00427->Node24                                                                                                             
#> [124] -0.59366->Node2-0.20984->Node6-0.18761->Node17-0.00889->Node25                                                                                                                                                           
#> [125] -0.59184->Node2-0.20919->Node6-0.18704->Node17-0.00886->Node25-0.00307->Node26                                                                                                                                           
#> [126] --0.590-->Node2--0.198-->Node5--0.149-->Node12--0.063-->Node27                                                                                                                                                           
#> [127] -0.58817->Node2-0.20790->Node6-0.18588->Node17-0.00880->Node25-0.00925->Node28                                                                                                                                           
#> [128] --0.6863->Node3--0.0509->Node4--0.1356->Node7--0.0104->Node8--0.0973->Node18--0.0195->Node29                                                                                                                             
#> [129] --0.6844->Node3--0.0508->Node4--0.1352->Node7--0.0104->Node8--0.0719->Node14--0.0474->Node30                                                                                                                             
#> [130] -0.585110>Node2-0.206813>Node6-0.118495>Node9-0.000731>Node10-0.088851>Node31                                                                                                                                            
#> [131] --0.6829->Node3--0.0507->Node4--0.1349->Node7--0.0806->Node13--0.0137->Node15--0.0372->Node32                                                                                                                            
#> [132] --0.683-->Node3--0.317-->Node33                                                                                                                                                                                          
#> [133] --0.6811->Node3--0.0505->Node4--0.1345->Node7--0.0103->Node8--0.0715->Node14--0.0520->Node34                                                                                                                             
#> [134] --0.5821->Node2--0.1951->Node5--0.1472->Node12--0.0756->Node35                                                                                                                                                           
#> [135] -0.67945->Node3-0.05042->Node4-0.13421->Node7-0.01032->Node8-0.09632->Node18-0.00312->Node19-0.00417->Node24-0.02199->Node36                                                                                             
#> [136] --0.6794->Node3--0.0504->Node4--0.1342->Node7--0.0802->Node13--0.0136->Node15--0.0422->Node37                                                                                                                            
#> [137] -0.67829->Node3-0.05033->Node4-0.13398->Node7-0.08003->Node13-0.01362->Node15-0.04218->Node37-0.00157->Node38                                                                                                            
#> [138] --0.6774->Node3--0.0503->Node4--0.1338->Node7--0.0103->Node8--0.0711->Node14--0.0469->Node30--0.0102->Node39                                                                                                             
#> [139] -0.57951->Node2-0.19422->Node5-0.14650->Node12-0.07524->Node35-0.00453->Node40                                                                                                                                           
#> [140] --0.5793->Node2--0.1941->Node5--0.1464->Node12--0.0618->Node27--0.0183->Node41                                                                                                                                           
#> [141] --0.5791->Node2--0.1941->Node5--0.1994->Node21--0.0275->Node42                                                                                                                                                           
#> [142] --0.6753->Node3--0.0501->Node4--0.1334->Node7--0.0103->Node8--0.0957->Node18--0.0031->Node19--0.0322->Node43                                                                                                             
#> [143] -0.57696->Node2-0.20393->Node6-0.18233->Node17-0.00448->Node20-0.03229->Node44                                                                                                                                           
#> [144] --0.577-->Node2--0.204-->Node6--0.117-->Node9--0.057-->Node16--0.015-->Node23--0.031-->Node45                                                                                                                            
#> [145] --0.5762->Node2--0.2037->Node6--0.1167->Node9--0.0570->Node16--0.0150->Node23--0.0315->Node46                                                                                                                            
#> [146] -0.67309->Node3-0.04995->Node4-0.13295->Node7-0.01022->Node8-0.09542->Node18-0.00309->Node19-0.03205->Node43-0.00323->Node47                                                                                             
#> [147] --0.6723->Node3--0.0499->Node4--0.1328->Node7--0.0102->Node8--0.0706->Node14--0.0465->Node30--0.0101->Node39--0.0076->Node48                                                                                             
#> [148] -0.672113>Node3-0.049876>Node4-0.132761>Node7-0.010207>Node8-0.070577>Node14-0.046525>Node30-0.010113>Node39-0.007595>Node48-0.000233>Node49                                                                             
#> [149] --0.5749->Node2--0.2032->Node6--0.1164->Node9--0.0568->Node16--0.0486->Node50                                                                                                                                            
#> [150] -0.574707>Node2-0.203137>Node6-0.116388>Node9-0.056822>Node16-0.048572>Node50-0.000373>Node51                                                                                                                            
#> [151] -0.67154->Node3-0.04983->Node4-0.13265->Node7-0.07923->Node13-0.01349->Node15-0.04176->Node37-0.00155->Node38-0.00995->Node52                                                                                            
#> [152] --0.5744->Node2--0.1925->Node5--0.1978->Node21--0.0353->Node53                                                                                                                                                           
#> [153] -0.67078->Node3-0.04978->Node4-0.13250->Node7-0.01019->Node8-0.09509->Node18-0.00308->Node19-0.00412->Node24-0.02171->Node36-0.01276->Node54                                                                             
#> [154] -0.573516>Node2-0.202715>Node6-0.116147>Node9-0.000717>Node10-0.106905>Node55                                                                                                                                            
#> [155] -0.573199>Node2-0.202603>Node6-0.116083>Node9-0.056673>Node16-0.048445>Node50-0.000372>Node51-0.002625>Node56                                                                                                            
#> [156] --0.6692->Node3--0.0497->Node4--0.1322->Node7--0.0790->Node13--0.0134->Node15--0.0416->Node37--0.0149->Node57                                                                                                            
#> [157] -0.57212->Node2-0.20222->Node6-0.18080->Node17-0.00444->Node20-0.04042->Node58                                                                                                                                           
#> [158] -0.57193->Node2-0.20215->Node6-0.18074->Node17-0.00444->Node20-0.03201->Node44-0.00872->Node59                                                                                                                           
#> [159] -0.57173->Node2-0.20209->Node6-0.11579->Node9-0.05653->Node16-0.04832->Node50-0.00555->Node60                                                                                                                            
#> [160] -0.571692>Node2-0.202071>Node6-0.180668>Node17-0.004438>Node20-0.031996>Node44-0.008719>Node59-0.000415>Node61                                                                                                           
#> [161] -0.667292>Node3-0.049518>Node4-0.131809>Node7-0.010134>Node8-0.070071>Node14-0.046192>Node30-0.010040>Node39-0.007540>Node48-0.000232>Node49-0.007173>Node62                                                             
#> [162] --0.6669->Node3--0.0495->Node4--0.1317->Node7--0.0101->Node8--0.0700->Node14--0.0509->Node34--0.0208->Node63                                                                                                             
#> [163] -0.57065->Node2-0.19125->Node5-0.14426->Node12-0.07409->Node35-0.00446->Node40-0.01529->Node64                                                                                                                           
#> [164] -0.666643>Node3-0.049470>Node4-0.131681>Node7-0.010124>Node8-0.070002>Node14-0.050907>Node34-0.020750>Node63-0.000422>Node65                                                                                             
#> [165] -0.66644->Node3-0.04945->Node4-0.13164->Node7-0.07863->Node13-0.01339->Node15-0.04145->Node37-0.01484->Node57-0.00417->Node66                                                                                            
#> [166] -0.66614->Node3-0.04943->Node4-0.13158->Node7-0.01012->Node8-0.09444->Node18-0.00305->Node19-0.03172->Node43-0.00320->Node47-0.01032->Node67                                                                             
#> [167] -0.569917>Node2-0.201443>Node6-0.115418>Node9-0.000712>Node10-0.086544>Node31-0.025965>Node68                                                                                                                            
#> [168] -0.56951->Node2-0.20130->Node6-0.17998->Node17-0.00442->Node20-0.04023->Node58-0.00455->Node69                                                                                                                           
#> [169] -0.66544->Node3-0.04938->Node4-0.13144->Node7-0.07851->Node13-0.01337->Node15-0.04138->Node37-0.01482->Node57-0.00416->Node66-0.00151->Node70                                                                            
#> [170] -0.665332>Node3-0.049372>Node4-0.131422>Node7-0.010104>Node8-0.069865>Node14-0.046056>Node30-0.010011>Node39-0.007518>Node48-0.000231>Node49-0.007152>Node62-0.002936>Node71                                             
#> [171] --0.569-->Node2--0.191-->Node5--0.196-->Node21--0.027-->Node42--0.017-->Node72                                                                                                                                           
#> [172] -0.56913->Node2-0.20117->Node6-0.17986->Node17-0.00442->Node20-0.03185->Node44-0.01358->Node73                                                                                                                           
#> [173] -0.66485->Node3-0.04934->Node4-0.13133->Node7-0.07844->Node13-0.01335->Node15-0.04135->Node37-0.00154->Node38-0.00985->Node52-0.00996->Node74                                                                            
#> [174] -0.664756>Node3-0.049330>Node4-0.131308>Node7-0.010095>Node8-0.069804>Node14-0.046016>Node30-0.010002>Node39-0.007512>Node48-0.000231>Node49-0.007146>Node62-0.002934>Node71-0.000865>Node75                             
#> [175] -0.66448->Node3-0.04931->Node4-0.13125->Node7-0.07840->Node13-0.01335->Node15-0.04132->Node37-0.01480->Node57-0.00709->Node76                                                                                            
#> [176] -0.56852->Node2-0.20095->Node6-0.17967->Node17-0.00441->Node20-0.04016->Node58-0.00454->Node69-0.00174->Node77                                                                                                           
#> [177] -0.56847->Node2-0.20093->Node6-0.11513->Node9-0.05621->Node16-0.04805->Node50-0.00551->Node60-0.00570->Node78                                                                                                            
#> [178] --0.6643->Node3--0.0493->Node4--0.1312->Node7--0.0784->Node13--0.0133->Node15--0.0362->Node32--0.0273->Node79                                                                                                            
#> [179] -0.66421->Node3-0.04929->Node4-0.13120->Node7-0.01009->Node8-0.06975->Node14-0.05072->Node34-0.02067->Node63-0.00042->Node65-0.00365->Node80                                                                             
#> [180] -0.568000>Node2-0.200766>Node6-0.115030>Node9-0.056159>Node16-0.048005>Node50-0.000368>Node51-0.002601>Node56-0.009069>Node81                                                                                            
#> [181] -0.56794->Node2-0.19034->Node5-0.14357->Node12-0.07374->Node35-0.00444->Node40-0.01521->Node64-0.00474->Node82                                                                                                           
#> [182] -0.66331->Node3-0.04922->Node4-0.13102->Node7-0.01007->Node8-0.06965->Node14-0.05065->Node34-0.02065->Node63-0.00042->Node65-0.00364->Node80-0.00135->Node83                                                             
#> [183] -0.567180>Node2-0.200476>Node6-0.114864>Node9-0.056078>Node16-0.047936>Node50-0.000368>Node51-0.013097>Node84                                                                                                            
#> [184] -0.66284->Node3-0.04919->Node4-0.13093->Node7-0.01007->Node8-0.09397->Node18-0.00304->Node19-0.03157->Node43-0.00318->Node47-0.01027->Node67-0.00496->Node85                                                             
#> [185] -0.662744>Node3-0.049180>Node4-0.130910>Node7-0.010065>Node8-0.069593>Node14-0.045877>Node30-0.009972>Node39-0.007489>Node48-0.000230>Node49-0.007125>Node62-0.002925>Node71-0.000863>Node75-0.003028>Node86             
#> [186] -0.66232->Node3-0.04915->Node4-0.13083->Node7-0.07814->Node13-0.01330->Node15-0.03607->Node32-0.02718->Node79-0.00302->Node87                                                                                            
#> [187] -0.566686>Node2-0.200301>Node6-0.114764>Node9-0.000708>Node10-0.086053>Node31-0.025818>Node68-0.005669>Node88                                                                                                            
#> [188] -0.662189>Node3-0.049139>Node4-0.130801>Node7-0.010056>Node8-0.069535>Node14-0.050567>Node34-0.020612>Node63-0.000419>Node65-0.003638>Node80-0.001348>Node83-0.001695>Node89                                             
#> [189] -0.661961>Node3-0.049122>Node4-0.130756>Node7-0.078102>Node13-0.013296>Node15-0.036047>Node32-0.027163>Node79-0.003016>Node87-0.000539>Node90                                                                            
#> [190] -0.56638->Node2-0.18982->Node5-0.19500->Node21-0.02692->Node42-0.01695->Node72-0.00495->Node91                                                                                                                           
#> [191] -0.56638->Node2-0.20019->Node6-0.17899->Node17-0.00848->Node25-0.00293->Node26-0.04303->Node92                                                                                                                           
#> [192] -0.566255>Node2-0.200149>Node6-0.114677>Node9-0.055987>Node16-0.047858>Node50-0.000367>Node51-0.002593>Node56-0.012115>Node93                                                                                            
#> [193] -0.566176>Node2-0.200121>Node6-0.114661>Node9-0.000707>Node10-0.085976>Node31-0.025795>Node68-0.005664>Node88-0.000900>Node94                                                                                            
#> [194] -0.56588->Node2-0.20002->Node6-0.11460->Node9-0.05595->Node16-0.04783->Node50-0.00549->Node60-0.01023->Node95                                                                                                            
#> [195] --0.5659->Node2--0.1896->Node5--0.1430->Node12--0.0604->Node27--0.0411->Node96                                                                                                                                           
#> [196] -0.565608>Node2-0.199920>Node6-0.114546>Node9-0.055923>Node16-0.047803>Node50-0.000367>Node51-0.013061>Node84-0.002773>Node97                                                                                            
#> [197] -0.66087->Node3-0.04904->Node4-0.13054->Node7-0.01004->Node8-0.06940->Node14-0.04575->Node30-0.00994->Node39-0.02442->Node98                                                                                             
#> [198] -0.565469>Node2-0.199871>Node6-0.114518>Node9-0.055909>Node16-0.047792>Node50-0.000367>Node51-0.013058>Node84-0.002772>Node97-0.000245>Node99