Scale the phylogenetic flow features to a constant sum.

Can be used to standardise branch lengths to reasonable values

Usage

pf_scale_flow_sum(x, scale_to = 1)

Arguments

x

a pfc object

scale_to

The value to scale the sums to. Default is 1.

Value

a scaled pfc object

Examples

pf_scale_flow_sum(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] -0.90132->Node10-0.08871->Node47-0.00413->Node62-0.00584->t68                                                                                                                           
#>   [2] -0.90132->Node10-0.08871->Node47-0.00413->Node62-0.00180->Node72-0.00404->t84                                                                                                           
#>   [3] -0.90132->Node10-0.08871->Node47-0.00413->Node62-0.00180->Node72-0.00404->t90                                                                                                           
#>   [4] -0.901316>Node10-0.088705>Node47-0.009578>Node96-0.000401>t82                                                                                                                           
#>   [5] -0.901316>Node10-0.088705>Node47-0.009578>Node96-0.000401>t12                                                                                                                           
#>   [6] -0.90132->Node10-0.09414->Node70-0.00454->t27                                                                                                                                           
#>   [7] -0.90132->Node10-0.09414->Node70-0.00454->t9                                                                                                                                            
#>   [8] --0.3316->Node2--0.4548->Node4--0.0985->Node7--0.0675->Node14--0.0234->Node24--0.0243->t97                                                                                              
#>   [9] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06748->Node14-0.02336->Node24-0.00649->Node31-0.01641->Node82-0.00135->t59                                                              
#>  [10] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06748->Node14-0.02336->Node24-0.00649->Node31-0.01641->Node82-0.00135->t58                                                              
#>  [11] -0.331599>Node2-0.454767>Node4-0.098539>Node7-0.067482>Node14-0.023363>Node24-0.006490>Node31-0.012755>Node67-0.003057>Node79-0.001604>Node97-0.000346>t23                              
#>  [12] -0.331599>Node2-0.454767>Node4-0.098539>Node7-0.067482>Node14-0.023363>Node24-0.006490>Node31-0.012755>Node67-0.003057>Node79-0.001604>Node97-0.000346>t41                              
#>  [13] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06748->Node14-0.02336->Node24-0.00649->Node31-0.01275->Node67-0.00306->Node79-0.00195->t54                                              
#>  [14] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06748->Node14-0.02336->Node24-0.00649->Node31-0.01275->Node67-0.00501->t76                                                              
#>  [15] --0.3316->Node2--0.4548->Node4--0.0985->Node7--0.0675->Node14--0.0476->t24                                                                                                              
#>  [16] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.00853->Node17-0.02818->Node38-0.01305->t52                                                                              
#>  [17] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.00853->Node17-0.02818->Node38-0.00250->Node45-0.00697->Node73-0.00358->t71                                              
#>  [18] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.00853->Node17-0.02818->Node38-0.00250->Node45-0.00697->Node73-0.00358->t26                                              
#>  [19] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.00853->Node17-0.02818->Node38-0.00250->Node45-0.01055->t19                                                              
#>  [20] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.00853->Node17-0.00849->Node20-0.01862->Node36-0.00433->Node48-0.00978->t34                                              
#>  [21] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.00853->Node17-0.00849->Node20-0.01862->Node36-0.00433->Node48-0.00978->t28                                              
#>  [22] -0.331599>Node2-0.454767>Node4-0.098539>Node7-0.065334>Node13-0.008533>Node17-0.008493>Node20-0.018618>Node36-0.008896>Node65-0.000253>Node68-0.004969>t3                               
#>  [23] -0.331599>Node2-0.454767>Node4-0.098539>Node7-0.065334>Node13-0.008533>Node17-0.008493>Node20-0.018618>Node36-0.008896>Node65-0.000253>Node68-0.000606>Node71-0.004363>t36              
#>  [24] -0.331599>Node2-0.454767>Node4-0.098539>Node7-0.065334>Node13-0.008533>Node17-0.008493>Node20-0.018618>Node36-0.008896>Node65-0.000253>Node68-0.000606>Node71-0.004363>t2               
#>  [25] -0.331599>Node2-0.454767>Node4-0.098539>Node7-0.065334>Node13-0.008533>Node17-0.008493>Node20-0.018618>Node36-0.008896>Node65-0.004700>Node93-0.000395>Node99-0.000128>t100             
#>  [26] -0.331599>Node2-0.454767>Node4-0.098539>Node7-0.065334>Node13-0.008533>Node17-0.008493>Node20-0.018618>Node36-0.008896>Node65-0.004700>Node93-0.000395>Node99-0.000128>t88              
#>  [27] -0.331599>Node2-0.454767>Node4-0.098539>Node7-0.065334>Node13-0.008533>Node17-0.008493>Node20-0.018618>Node36-0.008896>Node65-0.004700>Node93-0.000523>t14                              
#>  [28] -0.331599>Node2-0.454767>Node4-0.098539>Node7-0.065334>Node13-0.008533>Node17-0.008493>Node20-0.008452>Node23-0.017752>Node59-0.000762>Node63-0.005769>t6                               
#>  [29] -0.331599>Node2-0.454767>Node4-0.098539>Node7-0.065334>Node13-0.008533>Node17-0.008493>Node20-0.008452>Node23-0.017752>Node59-0.000762>Node63-0.002344>Node74-0.003425>t55              
#>  [30] -0.331599>Node2-0.454767>Node4-0.098539>Node7-0.065334>Node13-0.008533>Node17-0.008493>Node20-0.008452>Node23-0.017752>Node59-0.000762>Node63-0.002344>Node74-0.003425>t94              
#>  [31] -0.331599>Node2-0.454767>Node4-0.098539>Node7-0.065334>Node13-0.008533>Node17-0.008493>Node20-0.008452>Node23-0.017752>Node59-0.005754>Node88-0.000777>t70                              
#>  [32] -0.331599>Node2-0.454767>Node4-0.098539>Node7-0.065334>Node13-0.008533>Node17-0.008493>Node20-0.008452>Node23-0.017752>Node59-0.005754>Node88-0.000777>t33                              
#>  [33] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.00853->Node17-0.00849->Node20-0.00845->Node23-0.02428->t31                                                              
#>  [34] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.01426->Node19-0.00768->Node21-0.01806->Node49-0.00144->Node52-0.00240->Node61-0.00593->t56                              
#>  [35] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.01426->Node19-0.00768->Node21-0.01806->Node49-0.00144->Node52-0.00240->Node61-0.00593->t32                              
#>  [36] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.01426->Node19-0.00768->Node21-0.01806->Node49-0.00144->Node52-0.00833->t74                                              
#>  [37] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.01426->Node19-0.00768->Node21-0.01806->Node49-0.00977->t47                                                              
#>  [38] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.01426->Node19-0.00768->Node21-0.02783->t11                                                                              
#>  [39] --0.3316->Node2--0.4548->Node4--0.0985->Node7--0.0653->Node13--0.0143->Node19--0.0216->Node37--0.0139->t91                                                                              
#>  [40] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.01426->Node19-0.02161->Node37-0.00335->Node46-0.00502->Node64-0.00552->t67                                              
#>  [41] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.01426->Node19-0.02161->Node37-0.00335->Node46-0.00502->Node64-0.00268->Node75-0.00161->Node86-0.00123->t83              
#>  [42] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.01426->Node19-0.02161->Node37-0.00335->Node46-0.00502->Node64-0.00268->Node75-0.00161->Node86-0.00123->t53              
#>  [43] -0.331599>Node2-0.454767>Node4-0.098539>Node7-0.065334>Node13-0.014260>Node19-0.021611>Node37-0.003348>Node46-0.005021>Node64-0.002676>Node75-0.002356>Node94-0.000491>t85              
#>  [44] -0.331599>Node2-0.454767>Node4-0.098539>Node7-0.065334>Node13-0.014260>Node19-0.021611>Node37-0.003348>Node46-0.005021>Node64-0.002676>Node75-0.002356>Node94-0.000491>t42              
#>  [45] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.01426->Node19-0.02161->Node37-0.00335->Node46-0.00400->Node58-0.00179->Node69-0.00475->t8                               
#>  [46] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.01426->Node19-0.02161->Node37-0.00335->Node46-0.00400->Node58-0.00179->Node69-0.00475->t40                              
#>  [47] -0.33160->Node2-0.45477->Node4-0.09854->Node7-0.06533->Node13-0.01426->Node19-0.02161->Node37-0.00335->Node46-0.00400->Node58-0.00654->t16                                              
#>  [48] -0.331599>Node2-0.454767>Node4-0.169159>Node15-0.036321>Node53-0.000644>Node55-0.007511>t51                                                                                             
#>  [49] -0.331599>Node2-0.454767>Node4-0.169159>Node15-0.036321>Node53-0.000644>Node55-0.006789>Node89-0.000722>t48                                                                             
#>  [50] -0.331599>Node2-0.454767>Node4-0.169159>Node15-0.036321>Node53-0.000644>Node55-0.006789>Node89-0.000722>t89                                                                             
#>  [51] -0.33160->Node2-0.45477->Node4-0.16916->Node15-0.03632->Node53-0.00815->t87                                                                                                             
#>  [52] -0.331599>Node2-0.454767>Node4-0.169159>Node15-0.029836>Node33-0.008173>Node60-0.006062>Node95-0.000405>t69                                                                             
#>  [53] -0.331599>Node2-0.454767>Node4-0.169159>Node15-0.029836>Node33-0.008173>Node60-0.006062>Node95-0.000405>t39                                                                             
#>  [54] -0.33160->Node2-0.45477->Node4-0.16916->Node15-0.02984->Node33-0.00817->Node60-0.00647->t66                                                                                             
#>  [55] -0.33160->Node2-0.45477->Node4-0.16916->Node15-0.02984->Node33-0.00687->Node54-0.00777->t62                                                                                             
#>  [56] -0.33160->Node2-0.45477->Node4-0.16916->Node15-0.02984->Node33-0.00687->Node54-0.00777->t75                                                                                             
#>  [57] --0.3316->Node2--0.3541->Node3--0.2594->Node12--0.0334->Node26--0.0215->t95                                                                                                             
#>  [58] --0.3316->Node2--0.3541->Node3--0.2594->Node12--0.0334->Node26--0.0215->t15                                                                                                             
#>  [59] --0.3316->Node2--0.3541->Node3--0.2594->Node12--0.0404->Node35--0.0146->t78                                                                                                             
#>  [60] -0.33160->Node2-0.35409->Node3-0.25941->Node12-0.04036->Node35-0.00560->Node50-0.00772->Node85-0.00124->t29                                                                             
#>  [61] -0.33160->Node2-0.35409->Node3-0.25941->Node12-0.04036->Node35-0.00560->Node50-0.00772->Node85-0.00124->t86                                                                             
#>  [62] -0.33160->Node2-0.35409->Node3-0.25941->Node12-0.04036->Node35-0.00560->Node50-0.00231->Node57-0.00664->t57                                                                             
#>  [63] -0.33160->Node2-0.35409->Node3-0.25941->Node12-0.04036->Node35-0.00560->Node50-0.00231->Node57-0.00664->t98                                                                             
#>  [64] --0.3316->Node2--0.3541->Node3--0.1849->Node5--0.0142->Node6--0.0719->Node16--0.0433->t10                                                                                               
#>  [65] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.07189->Node16-0.02101->Node25-0.00405->Node29-0.00354->Node32-0.00400->Node43-0.00889->Node80-0.00183->t81               
#>  [66] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.07189->Node16-0.02101->Node25-0.00405->Node29-0.00354->Node32-0.00400->Node43-0.00889->Node80-0.00183->t38               
#>  [67] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.07189->Node16-0.02101->Node25-0.00405->Node29-0.00354->Node32-0.00400->Node43-0.00938->Node83-0.00134->t80               
#>  [68] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.07189->Node16-0.02101->Node25-0.00405->Node29-0.00354->Node32-0.00400->Node43-0.00938->Node83-0.00134->t46               
#>  [69] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.07189->Node16-0.02101->Node25-0.00405->Node29-0.00354->Node32-0.01472->t65                                               
#>  [70] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.07189->Node16-0.02101->Node25-0.00405->Node29-0.01570->Node77-0.00129->Node84-0.00126->t79                               
#>  [71] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.07189->Node16-0.02101->Node25-0.00405->Node29-0.01570->Node77-0.00129->Node84-0.00126->t18                               
#>  [72] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.07189->Node16-0.02101->Node25-0.00405->Node29-0.01570->Node77-0.00255->t30                                               
#>  [73] --0.3316->Node2--0.3541->Node3--0.1849->Node5--0.0142->Node6--0.0719->Node16--0.0210->Node25--0.0223->t61                                                                               
#>  [74] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.00655->Node8-0.06919->Node18-0.01391->Node22-0.01285->Node40-0.00205->Node44-0.01065->t44                                
#>  [75] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.00655->Node8-0.06919->Node18-0.01391->Node22-0.01285->Node40-0.00205->Node44-0.01065->t17                                
#>  [76] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.00655->Node8-0.06919->Node18-0.01391->Node22-0.01285->Node40-0.01270->t99                                                
#>  [77] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.00655->Node8-0.06919->Node18-0.01391->Node22-0.01685->Node51-0.00870->t77                                                
#>  [78] -0.331599>Node2-0.354085>Node3-0.184921>Node5-0.014195>Node6-0.006547>Node8-0.069189>Node18-0.013911>Node22-0.016854>Node51-0.006422>Node78-0.000507>Node81-0.001770>t1                 
#>  [79] -0.331599>Node2-0.354085>Node3-0.184921>Node5-0.014195>Node6-0.006547>Node8-0.069189>Node18-0.013911>Node22-0.016854>Node51-0.006422>Node78-0.000507>Node81-0.001770>t7                 
#>  [80] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.00655->Node8-0.06919->Node18-0.01391->Node22-0.01685->Node51-0.00642->Node78-0.00124->Node87-0.00103->t20                
#>  [81] -0.331599>Node2-0.354085>Node3-0.184921>Node5-0.014195>Node6-0.006547>Node8-0.069189>Node18-0.013911>Node22-0.016854>Node51-0.006422>Node78-0.001242>Node87-0.000829>Node98-0.000206>t25
#>  [82] -0.331599>Node2-0.354085>Node3-0.184921>Node5-0.014195>Node6-0.006547>Node8-0.069189>Node18-0.013911>Node22-0.016854>Node51-0.006422>Node78-0.001242>Node87-0.000829>Node98-0.000206>t72
#>  [83] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.00655->Node8-0.06919->Node18-0.03946->t63                                                                                
#>  [84] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.00655->Node8-0.00412->Node9-0.00801->Node11-0.08349->Node39-0.01303->t64                                                 
#>  [85] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.00655->Node8-0.00412->Node9-0.00801->Node11-0.08349->Node39-0.01303->t73                                                 
#>  [86] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.00655->Node8-0.00412->Node9-0.00801->Node11-0.07857->Node30-0.01795->t93                                                 
#>  [87] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.00655->Node8-0.00412->Node9-0.00801->Node11-0.07857->Node30-0.01795->t50                                                 
#>  [88] -0.331599>Node2-0.354085>Node3-0.184921>Node5-0.014195>Node6-0.006547>Node8-0.004120>Node9-0.086245>Node28-0.006535>Node42-0.011097>Node91-0.000066>Node92-0.000589>t5                  
#>  [89] -0.331599>Node2-0.354085>Node3-0.184921>Node5-0.014195>Node6-0.006547>Node8-0.004120>Node9-0.086245>Node28-0.006535>Node42-0.011097>Node91-0.000066>Node92-0.000589>t96                 
#>  [90] -0.331599>Node2-0.354085>Node3-0.184921>Node5-0.014195>Node6-0.006547>Node8-0.004120>Node9-0.086245>Node28-0.006535>Node42-0.011097>Node91-0.000656>t4                                  
#>  [91] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.00655->Node8-0.00412->Node9-0.08624->Node28-0.00654->Node42-0.00663->Node66-0.00512->t21                                 
#>  [92] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.00655->Node8-0.00412->Node9-0.08624->Node28-0.00654->Node42-0.00663->Node66-0.00512->t60                                 
#>  [93] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.00655->Node8-0.00412->Node9-0.08624->Node28-0.00585->Node41-0.01244->t92                                                 
#>  [94] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.00655->Node8-0.00412->Node9-0.08624->Node28-0.00585->Node41-0.00564->Node56-0.00403->Node76-0.00276->t43                 
#>  [95] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.00655->Node8-0.00412->Node9-0.08624->Node28-0.00585->Node41-0.00564->Node56-0.00403->Node76-0.00276->t13                 
#>  [96] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.01420->Node6-0.00655->Node8-0.00412->Node9-0.08624->Node28-0.00585->Node41-0.00564->Node56-0.00679->t45                                 
#>  [97] -0.331599>Node2-0.354085>Node3-0.184921>Node5-0.108905>Node27-0.005883>Node34-0.013895>Node90-0.000711>t37                                                                              
#>  [98] -0.331599>Node2-0.354085>Node3-0.184921>Node5-0.108905>Node27-0.005883>Node34-0.013895>Node90-0.000711>t35                                                                              
#>  [99] -0.33160->Node2-0.35409->Node3-0.18492->Node5-0.10890->Node27-0.00588->Node34-0.01461->t22                                                                                              
#> [100] --0.3316->Node2--0.3541->Node3--0.1849->Node5--0.1089->Node27--0.0205->t49                                                                                                              
#> [101] ----1---->Node2                                                                                                                                                                         
#> [102] --0.484-->Node2--0.516-->Node3                                                                                                                                                          
#> [103] --0.422-->Node2--0.578-->Node4                                                                                                                                                          
#> [104] --0.381-->Node2--0.407-->Node3--0.212-->Node5                                                                                                                                           
#> [105] --0.375-->Node2--0.400-->Node3--0.209-->Node5--0.016-->Node6                                                                                                                            
#> [106] --0.375-->Node2--0.514-->Node4--0.111-->Node7                                                                                                                                           
#> [107] -0.37202->Node2-0.39725->Node3-0.20746->Node5-0.01593->Node6-0.00734->Node8                                                                                                             
#> [108] -0.37031->Node2-0.39542->Node3-0.20651->Node5-0.01585->Node6-0.00731->Node8-0.00460->Node9                                                                                              
#> [109] ----1---->Node10                                                                                                                                                                        
#> [110] -0.36702->Node2-0.39191->Node3-0.20468->Node5-0.01571->Node6-0.00725->Node8-0.00456->Node9-0.00887->Node11                                                                              
#> [111] --0.351-->Node2--0.375-->Node3--0.274-->Node12                                                                                                                                          
#> [112] --0.3490->Node2--0.4786->Node4--0.1037->Node7--0.0688->Node13                                                                                                                           
#> [113] --0.3482->Node2--0.4775->Node4--0.1035->Node7--0.0709->Node14                                                                                                                           
#> [114] --0.347-->Node2--0.476-->Node4--0.177-->Node15                                                                                                                                          
#> [115] --0.3466->Node2--0.3701->Node3--0.1933->Node5--0.0148->Node6--0.0751->Node16                                                                                                            
#> [116] --0.3459->Node2--0.4743->Node4--0.1028->Node7--0.0681->Node13--0.0089->Node17                                                                                                           
#> [117] -0.34522->Node2-0.36863->Node3-0.19252->Node5-0.01478->Node6-0.00682->Node8-0.07203->Node18                                                                                             
#> [118] --0.3438->Node2--0.4715->Node4--0.1022->Node7--0.0677->Node13--0.0148->Node19                                                                                                           
#> [119] -0.34282->Node2-0.47016->Node4-0.10187->Node7-0.06754->Node13-0.00882->Node17-0.00878->Node20                                                                                           
#> [120] -0.34109->Node2-0.46778->Node4-0.10136->Node7-0.06720->Node13-0.01467->Node19-0.00789->Node21                                                                                           
#> [121] -0.34029->Node2-0.36337->Node3-0.18977->Node5-0.01457->Node6-0.00672->Node8-0.07100->Node18-0.01428->Node22                                                                             
#> [122] -0.33985->Node2-0.46608->Node4-0.10099->Node7-0.06696->Node13-0.00875->Node17-0.00870->Node20-0.00866->Node23                                                                           
#> [123] --0.3398->Node2--0.4661->Node4--0.1010->Node7--0.0692->Node14--0.0239->Node24                                                                                                           
#> [124] --0.3392->Node2--0.3622->Node3--0.1891->Node5--0.0145->Node6--0.0735->Node16--0.0215->Node25                                                                                            
#> [125] --0.3389->Node2--0.3619->Node3--0.2651->Node12--0.0341->Node26                                                                                                                          
#> [126] --0.339-->Node2--0.361-->Node3--0.189-->Node5--0.111-->Node27                                                                                                                           
#> [127] -0.33778->Node2-0.36068->Node3-0.18837->Node5-0.01446->Node6-0.00667->Node8-0.00420->Node9-0.08785->Node28                                                                              
#> [128] -0.33777->Node2-0.36067->Node3-0.18836->Node5-0.01446->Node6-0.07322->Node16-0.02140->Node25-0.00412->Node29                                                                            
#> [129] -0.33766->Node2-0.36056->Node3-0.18830->Node5-0.01445->Node6-0.00667->Node8-0.00420->Node9-0.00816->Node11-0.08001->Node30                                                              
#> [130] -0.33759->Node2-0.46299->Node4-0.10032->Node7-0.06870->Node14-0.02379->Node24-0.00661->Node31                                                                                           
#> [131] -0.33655->Node2-0.35937->Node3-0.18768->Node5-0.01441->Node6-0.07296->Node16-0.02132->Node25-0.00411->Node29-0.00359->Node32                                                            
#> [132] --0.3365->Node2--0.4615->Node4--0.1717->Node15--0.0303->Node33                                                                                                                          
#> [133] -0.33651->Node2-0.35933->Node3-0.18766->Node5-0.11052->Node27-0.00597->Node34                                                                                                           
#> [134] --0.336-->Node2--0.359-->Node3--0.263-->Node12--0.041-->Node35                                                                                                                          
#> [135] -0.33635->Node2-0.46128->Node4-0.09995->Node7-0.06627->Node13-0.00865->Node17-0.00862->Node20-0.01888->Node36                                                                           
#> [136] --0.3363->Node2--0.4612->Node4--0.0999->Node7--0.0663->Node13--0.0145->Node19--0.0219->Node37                                                                                           
#> [137] -0.33598->Node2-0.46078->Node4-0.09984->Node7-0.06620->Node13-0.00865->Node17-0.02855->Node38                                                                                           
#> [138] -0.33598->Node2-0.35876->Node3-0.18736->Node5-0.01438->Node6-0.00663->Node8-0.00417->Node9-0.00812->Node11-0.08459->Node39                                                              
#> [139] -0.33586->Node2-0.35864->Node3-0.18730->Node5-0.01438->Node6-0.00663->Node8-0.07008->Node18-0.01409->Node22-0.01302->Node40                                                             
#> [140] -0.33577->Node2-0.35854->Node3-0.18725->Node5-0.01437->Node6-0.00663->Node8-0.00417->Node9-0.08733->Node28-0.00592->Node41                                                              
#> [141] -0.33554->Node2-0.35830->Node3-0.18712->Node5-0.01436->Node6-0.00662->Node8-0.00417->Node9-0.08727->Node28-0.00661->Node42                                                              
#> [142] -0.33519->Node2-0.35792->Node3-0.18692->Node5-0.01435->Node6-0.07267->Node16-0.02124->Node25-0.00409->Node29-0.00358->Node32-0.00404->Node43                                            
#> [143] -0.33517->Node2-0.35790->Node3-0.18691->Node5-0.01435->Node6-0.00662->Node8-0.06993->Node18-0.01406->Node22-0.01299->Node40-0.00207->Node44                                             
#> [144] -0.33514->Node2-0.45962->Node4-0.09959->Node7-0.06603->Node13-0.00862->Node17-0.02848->Node38-0.00252->Node45                                                                           
#> [145] -0.33513->Node2-0.45961->Node4-0.09959->Node7-0.06603->Node13-0.01441->Node19-0.02184->Node37-0.00338->Node46                                                                           
#> [146] --0.9104->Node10--0.0896->Node47                                                                                                                                                        
#> [147] -0.33488->Node2-0.45926->Node4-0.09951->Node7-0.06598->Node13-0.00862->Node17-0.00858->Node20-0.01880->Node36-0.00438->Node48                                                           
#> [148] -0.33487->Node2-0.45925->Node4-0.09951->Node7-0.06598->Node13-0.01440->Node19-0.00775->Node21-0.01823->Node49                                                                           
#> [149] -0.33459->Node2-0.35728->Node3-0.26175->Node12-0.04072->Node35-0.00565->Node50                                                                                                          
#> [150] --0.3345->Node2--0.3572->Node3--0.1865->Node5--0.0143->Node6--0.0066->Node8--0.0698->Node18--0.0140->Node22--0.0170->Node51                                                             
#> [151] -0.33438->Node2-0.45859->Node4-0.09937->Node7-0.06588->Node13-0.01438->Node19-0.00774->Node21-0.01821->Node49-0.00145->Node52                                                           
#> [152] --0.3343->Node2--0.4585->Node4--0.1705->Node15--0.0366->Node53                                                                                                                          
#> [153] -0.33419->Node2-0.45833->Node4-0.17048->Node15-0.03007->Node33-0.00693->Node54                                                                                                          
#> [154] -0.334108>Node2-0.458208>Node4-0.170439>Node15-0.036595>Node53-0.000649>Node55                                                                                                          
#> [155] -0.33387->Node2-0.35651->Node3-0.18619->Node5-0.01429->Node6-0.00659->Node8-0.00415->Node9-0.08683->Node28-0.00589->Node41-0.00568->Node56                                              
#> [156] -0.33382->Node2-0.35645->Node3-0.26114->Node12-0.04063->Node35-0.00564->Node50-0.00232->Node57                                                                                          
#> [157] -0.33378->Node2-0.45776->Node4-0.09919->Node7-0.06576->Node13-0.01435->Node19-0.02175->Node37-0.00337->Node46-0.00403->Node58                                                           
#> [158] -0.33378->Node2-0.45776->Node4-0.09919->Node7-0.06576->Node13-0.00859->Node17-0.00855->Node20-0.00851->Node23-0.01787->Node59                                                           
#> [159] -0.33376->Node2-0.45773->Node4-0.17026->Node15-0.03003->Node33-0.00823->Node60                                                                                                          
#> [160] -0.33358->Node2-0.45748->Node4-0.09913->Node7-0.06572->Node13-0.01435->Node19-0.00772->Node21-0.01816->Node49-0.00145->Node52-0.00241->Node61                                           
#> [161] -0.90661->Node10-0.08923->Node47-0.00416->Node62                                                                                                                                        
#> [162] -0.333523>Node2-0.457405>Node4-0.099111>Node7-0.065713>Node13-0.008582>Node17-0.008543>Node20-0.008502>Node23-0.017855>Node59-0.000767>Node63                                           
#> [163] -0.33344->Node2-0.45729->Node4-0.09909->Node7-0.06570->Node13-0.01434->Node19-0.02173->Node37-0.00337->Node46-0.00505->Node64                                                           
#> [164] -0.33334->Node2-0.45715->Node4-0.09906->Node7-0.06568->Node13-0.00858->Node17-0.00854->Node20-0.01872->Node36-0.00894->Node65                                                           
#> [165] -0.33331->Node2-0.35591->Node3-0.18587->Node5-0.01427->Node6-0.00658->Node8-0.00414->Node9-0.08669->Node28-0.00657->Node42-0.00666->Node66                                              
#> [166] -0.33327->Node2-0.45705->Node4-0.09903->Node7-0.06782->Node14-0.02348->Node24-0.00652->Node31-0.01282->Node67                                                                           
#> [167] -0.333255>Node2-0.457038>Node4-0.099031>Node7-0.065660>Node13-0.008575>Node17-0.008536>Node20-0.018711>Node36-0.008940>Node65-0.000255>Node68                                           
#> [168] -0.33318->Node2-0.45694->Node4-0.09901->Node7-0.06565->Node13-0.01433->Node19-0.02171->Node37-0.00336->Node46-0.00402->Node58-0.00180->Node69                                           
#> [169] --0.9054->Node10--0.0946->Node70                                                                                                                                                        
#> [170] -0.333052>Node2-0.456759>Node4-0.098971>Node7-0.065620>Node13-0.008570>Node17-0.008531>Node20-0.018699>Node36-0.008935>Node65-0.000255>Node68-0.000609>Node71                           
#> [171] -0.90497->Node10-0.08906->Node47-0.00415->Node62-0.00181->Node72                                                                                                                        
#> [172] -0.33279->Node2-0.45640->Node4-0.09889->Node7-0.06557->Node13-0.00856->Node17-0.02828->Node38-0.00250->Node45-0.00700->Node73                                                           
#> [173] -0.332739>Node2-0.456330>Node4-0.098878>Node7-0.065558>Node13-0.008562>Node17-0.008523>Node20-0.008482>Node23-0.017813>Node59-0.000765>Node63-0.002352>Node74                           
#> [174] -0.33255->Node2-0.45606->Node4-0.09882->Node7-0.06552->Node13-0.01430->Node19-0.02167->Node37-0.00336->Node46-0.00503->Node64-0.00268->Node75                                           
#> [175] -0.33252->Node2-0.35507->Node3-0.18543->Node5-0.01423->Node6-0.00656->Node8-0.00413->Node9-0.08648->Node28-0.00587->Node41-0.00566->Node56-0.00404->Node76                              
#> [176] -0.33245->Node2-0.35499->Node3-0.18539->Node5-0.01423->Node6-0.07207->Node16-0.02106->Node25-0.00406->Node29-0.01575->Node77                                                            
#> [177] -0.33236->Node2-0.35489->Node3-0.18534->Node5-0.01423->Node6-0.00656->Node8-0.06935->Node18-0.01394->Node22-0.01689->Node51-0.00644->Node78                                             
#> [178] -0.33225->Node2-0.45566->Node4-0.09873->Node7-0.06761->Node14-0.02341->Node24-0.00650->Node31-0.01278->Node67-0.00306->Node79                                                           
#> [179] -0.33221->Node2-0.35474->Node3-0.18526->Node5-0.01422->Node6-0.07202->Node16-0.02105->Node25-0.00405->Node29-0.00355->Node32-0.00401->Node43-0.00890->Node80                            
#> [180] -0.332187>Node2-0.354713>Node3-0.185249>Node5-0.014220>Node6-0.006558>Node8-0.069311>Node18-0.013936>Node22-0.016884>Node51-0.006433>Node78-0.000508>Node81                             
#> [181] --0.3320->Node2--0.4554->Node4--0.0987->Node7--0.0676->Node14--0.0234->Node24--0.0065->Node31--0.0164->Node82                                                                           
#> [182] -0.33204->Node2-0.35456->Node3-0.18517->Node5-0.01421->Node6-0.07198->Node16-0.02104->Node25-0.00405->Node29-0.00355->Node32-0.00400->Node43-0.00939->Node83                            
#> [183] -0.33202->Node2-0.35453->Node3-0.18515->Node5-0.01421->Node6-0.07198->Node16-0.02104->Node25-0.00405->Node29-0.01572->Node77-0.00129->Node84                                            
#> [184] -0.33201->Node2-0.35452->Node3-0.25973->Node12-0.04041->Node35-0.00561->Node50-0.00772->Node85                                                                                          
#> [185] -0.33201->Node2-0.45533->Node4-0.09866->Node7-0.06541->Node13-0.01428->Node19-0.02164->Node37-0.00335->Node46-0.00503->Node64-0.00268->Node75-0.00161->Node86                           
#> [186] -0.33194->Node2-0.35445->Node3-0.18511->Node5-0.01421->Node6-0.00655->Node8-0.06926->Node18-0.01393->Node22-0.01687->Node51-0.00643->Node78-0.00124->Node87                             
#> [187] -0.33186->Node2-0.45512->Node4-0.09862->Node7-0.06538->Node13-0.00854->Node17-0.00850->Node20-0.00846->Node23-0.01777->Node59-0.00576->Node88                                           
#> [188] -0.331838>Node2-0.455095>Node4-0.169281>Node15-0.036347>Node53-0.000644>Node55-0.006794>Node89                                                                                          
#> [189] -0.33183->Node2-0.35434->Node3-0.18505->Node5-0.10898->Node27-0.00589->Node34-0.01391->Node90                                                                                           
#> [190] -0.33182->Node2-0.35432->Node3-0.18504->Node5-0.01420->Node6-0.00655->Node8-0.00412->Node9-0.08630->Node28-0.00654->Node42-0.01110->Node91                                              
#> [191] -3.32e-01>Node2-3.54e-01>Node3-1.85e-01>Node5-1.42e-02>Node6-6.55e-03>Node8-4.12e-03>Node9-8.63e-02>Node28-6.54e-03>Node42-1.11e-02>Node91-6.61e-05>Node92                              
#> [192] -0.33177->Node2-0.45500->Node4-0.09859->Node7-0.06537->Node13-0.00854->Node17-0.00850->Node20-0.01863->Node36-0.00890->Node65-0.00470->Node93                                           
#> [193] -0.33176->Node2-0.45499->Node4-0.09859->Node7-0.06537->Node13-0.01427->Node19-0.02162->Node37-0.00335->Node46-0.00502->Node64-0.00268->Node75-0.00236->Node94                           
#> [194] -0.33173->Node2-0.45495->Node4-0.16923->Node15-0.02985->Node33-0.00818->Node60-0.00606->Node95                                                                                          
#> [195] -0.90168->Node10-0.08874->Node47-0.00958->Node96                                                                                                                                        
#> [196] -0.33171->Node2-0.45492->Node4-0.09857->Node7-0.06751->Node14-0.02337->Node24-0.00649->Node31-0.01276->Node67-0.00306->Node79-0.00160->Node97                                           
#> [197] -0.331667>Node2-0.354158>Node3-0.184959>Node5-0.014198>Node6-0.006548>Node8-0.069203>Node18-0.013914>Node22-0.016857>Node51-0.006423>Node78-0.001243>Node87-0.000829>Node98             
#> [198] -0.331641>Node2-0.454825>Node4-0.098552>Node7-0.065342>Node13-0.008534>Node17-0.008495>Node20-0.018620>Node36-0.008897>Node65-0.004700>Node93-0.000395>Node99