Scale the phylogenetic flow features to a constant sum.

Can be used to standardise branch lengths to reasonable values

Usage

pf_scale_flow_sum(x, scale_to = 1)

Arguments

x

a pfc object

scale_to

The value to scale the sums to. Default is 1.

Value

a scaled pfc object

Examples

pf_scale_flow_sum(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] --0.5210->Node2--0.3188->Node5--0.1257->Node25--0.0154->Node43--0.0191->t3                                                                                                                                                                               
#>   [2] --0.5210->Node2--0.3188->Node5--0.1257->Node25--0.0154->Node43--0.0191->t47                                                                                                                                                                              
#>   [3] --0.5210->Node2--0.3188->Node5--0.1257->Node25--0.0345->t94                                                                                                                                                                                              
#>   [4] -0.52099->Node2-0.31879->Node5-0.14832->Node58-0.00106->Node61-0.01084->t84                                                                                                                                                                              
#>   [5] -0.52099->Node2-0.31879->Node5-0.14832->Node58-0.00106->Node61-0.00720->Node84-0.00365->t62                                                                                                                                                              
#>   [6] -0.52099->Node2-0.31879->Node5-0.14832->Node58-0.00106->Node61-0.00720->Node84-0.00365->t30                                                                                                                                                              
#>   [7] -0.52099->Node2-0.31879->Node5-0.14832->Node58-0.00798->Node82-0.00392->t32                                                                                                                                                                              
#>   [8] -0.52099->Node2-0.31879->Node5-0.14832->Node58-0.00798->Node82-0.00392->t79                                                                                                                                                                              
#>   [9] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.00802->Node8-0.02216->Node16-0.04749->t60                                                                                                                                                                
#>  [10] -5.21e-01>Node2-1.69e-01>Node3-2.33e-01>Node7-8.02e-03>Node8-2.22e-02>Node16-8.59e-05>Node17-2.22e-02>Node35-2.92e-03>Node39-2.23e-02>t23                                                                                                                
#>  [11] -5.21e-01>Node2-1.69e-01>Node3-2.33e-01>Node7-8.02e-03>Node8-2.22e-02>Node16-8.59e-05>Node17-2.22e-02>Node35-2.92e-03>Node39-9.17e-03>Node54-1.32e-02>t50                                                                                                
#>  [12] -5.21e-01>Node2-1.69e-01>Node3-2.33e-01>Node7-8.02e-03>Node8-2.22e-02>Node16-8.59e-05>Node17-2.22e-02>Node35-2.92e-03>Node39-9.17e-03>Node54-1.11e-02>Node88-2.04e-03>t22                                                                                
#>  [13] -5.21e-01>Node2-1.69e-01>Node3-2.33e-01>Node7-8.02e-03>Node8-2.22e-02>Node16-8.59e-05>Node17-2.22e-02>Node35-2.92e-03>Node39-9.17e-03>Node54-1.11e-02>Node88-2.04e-03>t78                                                                                
#>  [14] -5.21e-01>Node2-1.69e-01>Node3-2.33e-01>Node7-8.02e-03>Node8-2.22e-02>Node16-8.59e-05>Node17-2.22e-02>Node35-9.63e-03>Node47-1.56e-02>t28                                                                                                                
#>  [15] -5.21e-01>Node2-1.69e-01>Node3-2.33e-01>Node7-8.02e-03>Node8-2.22e-02>Node16-8.59e-05>Node17-2.22e-02>Node35-9.63e-03>Node47-1.56e-02>t17                                                                                                                
#>  [16] -5.21e-01>Node2-1.69e-01>Node3-2.33e-01>Node7-8.02e-03>Node8-2.22e-02>Node16-8.59e-05>Node17-2.42e-02>Node37-9.27e-03>Node52-1.40e-02>t77                                                                                                                
#>  [17] -5.21e-01>Node2-1.69e-01>Node3-2.33e-01>Node7-8.02e-03>Node8-2.22e-02>Node16-8.59e-05>Node17-2.42e-02>Node37-9.27e-03>Node52-1.40e-02>t76                                                                                                                
#>  [18] -5.21e-01>Node2-1.69e-01>Node3-2.33e-01>Node7-8.02e-03>Node8-2.22e-02>Node16-8.59e-05>Node17-2.42e-02>Node37-2.32e-02>t88                                                                                                                                
#>  [19] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.029957>Node57-0.012483>t42                                                                                                
#>  [20] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.029957>Node57-0.012483>t27                                                                                                
#>  [21] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.017320>Node42-0.003547>Node46-0.014688>Node92-0.001073>t6                                                 
#>  [22] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.017320>Node42-0.003547>Node46-0.014688>Node92-0.001073>t85                                                
#>  [23] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.017320>Node42-0.003547>Node46-0.003193>Node56-0.000974>Node59-0.011595>t83                                
#>  [24] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.017320>Node42-0.003547>Node46-0.003193>Node56-0.000974>Node59-0.001135>Node62-0.005424>Node80-0.005036>t31
#>  [25] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.017320>Node42-0.003547>Node46-0.003193>Node56-0.000974>Node59-0.001135>Node62-0.005424>Node80-0.005036>t43
#>  [26] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.017320>Node42-0.003547>Node46-0.003193>Node56-0.000974>Node59-0.001135>Node62-0.010460>t40                
#>  [27] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.017320>Node42-0.003547>Node46-0.003193>Node56-0.003852>Node64-0.001338>Node69-0.007378>t38                
#>  [28] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.017320>Node42-0.003547>Node46-0.003193>Node56-0.003852>Node64-0.001338>Node69-0.007378>t97                
#>  [29] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.017320>Node42-0.003547>Node46-0.003193>Node56-0.003852>Node64-0.008716>t64                                
#>  [30] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.017320>Node42-0.019308>t49                                                                                
#>  [31] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.002603>Node26-0.026455>Node68-0.007571>t99                                                                
#>  [32] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.002603>Node26-0.026455>Node68-0.000303>Node70-0.001162>Node73-0.006106>t89                                
#>  [33] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.002603>Node26-0.026455>Node68-0.000303>Node70-0.001162>Node73-0.005939>Node98-0.000167>t68                
#>  [34] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.002603>Node26-0.026455>Node68-0.000303>Node70-0.001162>Node73-0.005939>Node98-0.000167>t11                
#>  [35] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.002603>Node26-0.026455>Node68-0.000303>Node70-0.006945>Node97-0.000323>t51                                
#>  [36] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.002603>Node26-0.026455>Node68-0.000303>Node70-0.006945>Node97-0.000323>t45                                
#>  [37] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.006380>Node20-0.005812>Node23-0.002603>Node26-0.034025>t44                                                                                
#>  [38] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.045120>Node83-0.003701>t10                                                                                                                
#>  [39] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.000308>Node15-0.045120>Node83-0.003701>t4                                                                                                                 
#>  [40] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.016768>Node29-0.021947>Node63-0.007914>Node87-0.001553>Node94-0.000946>t18                                                                                
#>  [41] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.008525>Node14-0.016768>Node29-0.021947>Node63-0.007914>Node87-0.001553>Node94-0.000946>t87                                                                                
#>  [42] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.00802->Node8-0.01200->Node11-0.00853->Node14-0.01677->Node29-0.02195->Node63-0.00791->Node87-0.00250->t65                                                                                                
#>  [43] -5.21e-01>Node2-1.69e-01>Node3-2.33e-01>Node7-8.02e-03>Node8-1.20e-02>Node11-8.53e-03>Node14-1.68e-02>Node29-2.19e-02>Node63-1.03e-02>Node99-6.76e-05>t5                                                                                                 
#>  [44] -5.21e-01>Node2-1.69e-01>Node3-2.33e-01>Node7-8.02e-03>Node8-1.20e-02>Node11-8.53e-03>Node14-1.68e-02>Node29-2.19e-02>Node63-1.03e-02>Node99-6.76e-05>t36                                                                                                
#>  [45] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.00802->Node8-0.01200->Node11-0.00853->Node14-0.01677->Node29-0.03236->t73                                                                                                                                
#>  [46] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.00802->Node8-0.01200->Node11-0.00274->Node12-0.02779->Node32-0.00531->Node40-0.00883->Node55-0.01299->t25                                                                                                
#>  [47] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.00802->Node8-0.01200->Node11-0.00274->Node12-0.02779->Node32-0.00531->Node40-0.00883->Node55-0.01023->Node86-0.00276->t19                                                                                
#>  [48] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.00802->Node8-0.01200->Node11-0.00274->Node12-0.02779->Node32-0.00531->Node40-0.00883->Node55-0.01023->Node86-0.00276->t41                                                                                
#>  [49] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.00802->Node8-0.01200->Node11-0.00274->Node12-0.02779->Node32-0.00531->Node40-0.02182->t26                                                                                                                
#>  [50] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.00802->Node8-0.01200->Node11-0.00274->Node12-0.02779->Node32-0.02713->t58                                                                                                                                
#>  [51] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.002736>Node12-0.002336>Node13-0.015468>Node22-0.004292>Node28-0.018089>Node51-0.001029>Node53-0.013217>Node96-0.000486>t95                                                
#>  [52] -0.520991>Node2-0.168587>Node3-0.232740>Node7-0.008025>Node8-0.012003>Node11-0.002736>Node12-0.002336>Node13-0.015468>Node22-0.004292>Node28-0.018089>Node51-0.001029>Node53-0.013217>Node96-0.000486>t57                                                
#>  [53] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.00802->Node8-0.01200->Node11-0.00274->Node12-0.00234->Node13-0.01547->Node22-0.00429->Node28-0.01809->Node51-0.00103->Node53-0.01370->t66                                                                
#>  [54] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.00802->Node8-0.01200->Node11-0.00274->Node12-0.00234->Node13-0.01547->Node22-0.00429->Node28-0.01809->Node51-0.00882->Node76-0.00591->t91                                                                
#>  [55] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.00802->Node8-0.01200->Node11-0.00274->Node12-0.00234->Node13-0.01547->Node22-0.00429->Node28-0.01809->Node51-0.00882->Node76-0.00591->t63                                                                
#>  [56] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.00802->Node8-0.01200->Node11-0.00274->Node12-0.00234->Node13-0.01547->Node22-0.00429->Node28-0.03282->t86                                                                                                
#>  [57] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.00802->Node8-0.01200->Node11-0.00274->Node12-0.00234->Node13-0.01547->Node22-0.03711->t54                                                                                                                
#>  [58] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.00802->Node8-0.01200->Node11-0.00274->Node12-0.00234->Node13-0.01803->Node24-0.01980->Node50-0.01475->t37                                                                                                
#>  [59] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.00802->Node8-0.01200->Node11-0.00274->Node12-0.00234->Node13-0.01803->Node24-0.01980->Node50-0.01475->t70                                                                                                
#>  [60] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.00802->Node8-0.01200->Node11-0.00274->Node12-0.00234->Node13-0.01803->Node24-0.03455->t2                                                                                                                 
#>  [61] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.04578->Node30-0.00254->Node31-0.00804->Node41-0.01523->Node74-0.00609->t72                                                                                                                               
#>  [62] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.04578->Node30-0.00254->Node31-0.00804->Node41-0.01523->Node74-0.00311->Node85-0.00298->t80                                                                                                               
#>  [63] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.04578->Node30-0.00254->Node31-0.00804->Node41-0.01523->Node74-0.00311->Node85-0.00298->t90                                                                                                               
#>  [64] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.04578->Node30-0.00254->Node31-0.00804->Node41-0.02132->t46                                                                                                                                               
#>  [65] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.04578->Node30-0.00254->Node31-0.02936->t7                                                                                                                                                                
#>  [66] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.04578->Node30-0.03077->Node91-0.00113->t55                                                                                                                                                               
#>  [67] -0.52099->Node2-0.16859->Node3-0.23274->Node7-0.04578->Node30-0.03077->Node91-0.00113->t13                                                                                                                                                               
#>  [68] -0.52099->Node2-0.16859->Node3-0.14098->Node4-0.10138->Node9-0.00254->Node10-0.05433->Node60-0.01119->t33                                                                                                                                                
#>  [69] -0.52099->Node2-0.16859->Node3-0.14098->Node4-0.10138->Node9-0.00254->Node10-0.05433->Node60-0.00676->Node81-0.00443->t98                                                                                                                                
#>  [70] -0.52099->Node2-0.16859->Node3-0.14098->Node4-0.10138->Node9-0.00254->Node10-0.05433->Node60-0.00676->Node81-0.00443->t16                                                                                                                                
#>  [71] -0.52099->Node2-0.16859->Node3-0.14098->Node4-0.10138->Node9-0.00254->Node10-0.04723->Node44-0.00343->Node49-0.00654->Node66-0.00833->t81                                                                                                                
#>  [72] -0.52099->Node2-0.16859->Node3-0.14098->Node4-0.10138->Node9-0.00254->Node10-0.04723->Node44-0.00343->Node49-0.00654->Node66-0.00833->t74                                                                                                                
#>  [73] -0.520991>Node2-0.168587>Node3-0.140979>Node4-0.101383>Node9-0.002539>Node10-0.047226>Node44-0.003426>Node49-0.014293>Node95-0.000575>t21                                                                                                                
#>  [74] -0.520991>Node2-0.168587>Node3-0.140979>Node4-0.101383>Node9-0.002539>Node10-0.047226>Node44-0.003426>Node49-0.014293>Node95-0.000575>t56                                                                                                                
#>  [75] -0.52099->Node2-0.16859->Node3-0.14098->Node4-0.10138->Node9-0.00254->Node10-0.04723->Node44-0.01829->t14                                                                                                                                                
#>  [76] --0.5210->Node2--0.1686->Node3--0.1410->Node4--0.1014->Node9--0.0424->Node34--0.0257->t92                                                                                                                                                                
#>  [77] --0.5210->Node2--0.1686->Node3--0.1410->Node4--0.1014->Node9--0.0424->Node34--0.0257->t12                                                                                                                                                                
#>  [78] --0.5210->Node2--0.1686->Node3--0.1410->Node4--0.1233->Node19--0.0461->t59                                                                                                                                                                               
#>  [79] -0.52099->Node2-0.16859->Node3-0.14098->Node4-0.12332->Node19-0.03743->Node65-0.00738->Node89-0.00130->t15                                                                                                                                               
#>  [80] -0.52099->Node2-0.16859->Node3-0.14098->Node4-0.12332->Node19-0.03743->Node65-0.00738->Node89-0.00130->t8                                                                                                                                                
#>  [81] -0.52099->Node2-0.16859->Node3-0.14098->Node4-0.12332->Node19-0.03743->Node65-0.00869->t35                                                                                                                                                               
#>  [82] -0.87918->Node6-0.07420->Node18-0.00642->Node21-0.04020->t61                                                                                                                                                                                             
#>  [83] -0.87918->Node6-0.07420->Node18-0.00642->Node21-0.04020->t75                                                                                                                                                                                             
#>  [84] -0.87918->Node6-0.07420->Node18-0.04041->Node72-0.00622->t34                                                                                                                                                                                             
#>  [85] -0.87918->Node6-0.07420->Node18-0.04041->Node72-0.00622->t20                                                                                                                                                                                             
#>  [86] -0.87918->Node6-0.08756->Node27-0.00671->Node33-0.02656->t82                                                                                                                                                                                             
#>  [87] -0.87918->Node6-0.08756->Node27-0.00671->Node33-0.02137->Node79-0.00519->t96                                                                                                                                                                             
#>  [88] -0.87918->Node6-0.08756->Node27-0.00671->Node33-0.02137->Node79-0.00519->t100                                                                                                                                                                            
#>  [89] -0.87918->Node6-0.08756->Node27-0.00804->Node36-0.00228->Node38-0.01489->Node67-0.00709->Node93-0.00097->t24                                                                                                                                             
#>  [90] -0.87918->Node6-0.08756->Node27-0.00804->Node36-0.00228->Node38-0.01489->Node67-0.00709->Node93-0.00097->t1                                                                                                                                              
#>  [91] -0.87918->Node6-0.08756->Node27-0.00804->Node36-0.00228->Node38-0.01489->Node67-0.00806->t93                                                                                                                                                             
#>  [92] -0.879177>Node6-0.087558>Node27-0.008041>Node36-0.002282>Node38-0.006647>Node45-0.000781>Node48-0.009509>Node75-0.006006>t9                                                                                                                              
#>  [93] -0.879177>Node6-0.087558>Node27-0.008041>Node36-0.002282>Node38-0.006647>Node45-0.000781>Node48-0.009509>Node75-0.006006>t48                                                                                                                             
#>  [94] -0.879177>Node6-0.087558>Node27-0.008041>Node36-0.002282>Node38-0.006647>Node45-0.000781>Node48-0.015514>t71                                                                                                                                             
#>  [95] -0.87918->Node6-0.08756->Node27-0.00804->Node36-0.00228->Node38-0.00665->Node45-0.00993->Node71-0.00636->t69                                                                                                                                             
#>  [96] -0.87918->Node6-0.08756->Node27-0.00804->Node36-0.00228->Node38-0.00665->Node45-0.00993->Node71-0.00063->Node77-0.00444->Node90-0.00130->t52                                                                                                             
#>  [97] -0.87918->Node6-0.08756->Node27-0.00804->Node36-0.00228->Node38-0.00665->Node45-0.00993->Node71-0.00063->Node77-0.00444->Node90-0.00130->t67                                                                                                             
#>  [98] -0.87918->Node6-0.08756->Node27-0.00804->Node36-0.00228->Node38-0.00665->Node45-0.00993->Node71-0.00063->Node77-0.00573->t53                                                                                                                             
#>  [99] -0.87918->Node6-0.08756->Node27-0.00804->Node36-0.01977->Node78-0.00545->t29                                                                                                                                                                             
#> [100] -0.87918->Node6-0.08756->Node27-0.00804->Node36-0.01977->Node78-0.00545->t39                                                                                                                                                                             
#> [101] ----1---->Node2                                                                                                                                                                                                                                          
#> [102] --0.756-->Node2--0.244-->Node3                                                                                                                                                                                                                           
#> [103] --0.627-->Node2--0.203-->Node3--0.170-->Node4                                                                                                                                                                                                            
#> [104] ---0.62-->Node2---0.38-->Node5                                                                                                                                                                                                                           
#> [105] ----1---->Node6                                                                                                                                                                                                                                          
#> [106] --0.565-->Node2--0.183-->Node3--0.252-->Node7                                                                                                                                                                                                            
#> [107] -0.56000->Node2-0.18121->Node3-0.25017->Node7-0.00863->Node8                                                                                                                                                                                             
#> [108] --0.559-->Node2--0.181-->Node3--0.151-->Node4--0.109-->Node9                                                                                                                                                                                             
#> [109] -0.55752->Node2-0.18041->Node3-0.15086->Node4-0.10849->Node9-0.00272->Node10                                                                                                                                                                             
#> [110] -0.55287->Node2-0.17890->Node3-0.24698->Node7-0.00852->Node8-0.01274->Node11                                                                                                                                                                             
#> [111] -0.55127->Node2-0.17838->Node3-0.24626->Node7-0.00849->Node8-0.01270->Node11-0.00289->Node12                                                                                                                                                             
#> [112] -0.54991->Node2-0.17794->Node3-0.24566->Node7-0.00847->Node8-0.01267->Node11-0.00289->Node12-0.00247->Node13                                                                                                                                             
#> [113] -0.54791->Node2-0.17730->Node3-0.24476->Node7-0.00844->Node8-0.01262->Node11-0.00897->Node14                                                                                                                                                             
#> [114] -0.547732>Node2-0.177240>Node3-0.244685>Node7-0.008437>Node8-0.012619>Node11-0.008963>Node14-0.000324>Node15                                                                                                                                             
#> [115] -0.54697->Node2-0.17699->Node3-0.24434->Node7-0.00843->Node8-0.02327->Node16                                                                                                                                                                             
#> [116] -5.47e-01>Node2-1.77e-01>Node3-2.44e-01>Node7-8.42e-03>Node8-2.33e-02>Node16-9.02e-05>Node17                                                                                                                                                             
#> [117] --0.9222->Node6--0.0778->Node18                                                                                                                                                                                                                          
#> [118] --0.546-->Node2--0.177-->Node3--0.148-->Node4--0.129-->Node19                                                                                                                                                                                            
#> [119] -0.544082>Node2-0.176059>Node3-0.243055>Node7-0.008381>Node8-0.012535>Node11-0.008903>Node14-0.000322>Node15-0.006663>Node20                                                                                                                             
#> [120] -0.91600->Node6-0.07731->Node18-0.00669->Node21                                                                                                                                                                                                          
#> [121] -0.54107->Node2-0.17509->Node3-0.24171->Node7-0.00833->Node8-0.01247->Node11-0.00284->Node12-0.00243->Node13-0.01606->Node22                                                                                                                             
#> [122] -0.54080->Node2-0.17500->Node3-0.24159->Node7-0.00833->Node8-0.01246->Node11-0.00885->Node14-0.00032->Node15-0.00662->Node20-0.00603->Node23                                                                                                             
#> [123] -0.53964->Node2-0.17462->Node3-0.24107->Node7-0.00831->Node8-0.01243->Node11-0.00283->Node12-0.00242->Node13-0.01868->Node24                                                                                                                             
#> [124] ---0.54-->Node2---0.33-->Node5---0.13-->Node25                                                                                                                                                                                                           
#> [125] -0.539343>Node2-0.174525>Node3-0.240938>Node7-0.008308>Node8-0.012426>Node11-0.008826>Node14-0.000319>Node15-0.006605>Node20-0.006017>Node23-0.002695>Node26                                                                                             
#> [126] --0.9094->Node6--0.0906->Node27                                                                                                                                                                                                                          
#> [127] -0.53867->Node2-0.17431->Node3-0.24064->Node7-0.00830->Node8-0.01241->Node11-0.00283->Node12-0.00242->Node13-0.01599->Node22-0.00444->Node28                                                                                                             
#> [128] -0.53841->Node2-0.17423->Node3-0.24052->Node7-0.00829->Node8-0.01240->Node11-0.00881->Node14-0.01733->Node29                                                                                                                                             
#> [129] --0.5382->Node2--0.1741->Node3--0.2404->Node7--0.0473->Node30                                                                                                                                                                                            
#> [130] -0.53675->Node2-0.17369->Node3-0.23978->Node7-0.04717->Node30-0.00261->Node31                                                                                                                                                                            
#> [131] -0.53552->Node2-0.17329->Node3-0.23923->Node7-0.00825->Node8-0.01234->Node11-0.00281->Node12-0.02856->Node32                                                                                                                                             
#> [132] -0.90316->Node6-0.08995->Node27-0.00689->Node33                                                                                                                                                                                                          
#> [133] --0.5347->Node2--0.1730->Node3--0.1447->Node4--0.1041->Node9--0.0435->Node34                                                                                                                                                                             
#> [134] -5.34e-01>Node2-1.73e-01>Node3-2.39e-01>Node7-8.23e-03>Node8-2.27e-02>Node16-8.82e-05>Node17-2.27e-02>Node35                                                                                                                                             
#> [135] -0.90193->Node6-0.08982->Node27-0.00825->Node36                                                                                                                                                                                                          
#> [136] -0.533384>Node2-0.172597>Node3-0.238276>Node7-0.008216>Node8-0.022689>Node16-0.000088>Node17-0.024750>Node37                                                                                                                                             
#> [137] -0.89982->Node6-0.08961->Node27-0.00823->Node36-0.00234->Node38                                                                                                                                                                                          
#> [138] -5.33e-01>Node2-1.72e-01>Node3-2.38e-01>Node7-8.21e-03>Node8-2.27e-02>Node16-8.79e-05>Node17-2.27e-02>Node35-2.98e-03>Node39                                                                                                                             
#> [139] -0.53261->Node2-0.17235->Node3-0.23793->Node7-0.00820->Node8-0.01227->Node11-0.00280->Node12-0.02841->Node32-0.00543->Node40                                                                                                                             
#> [140] -0.53234->Node2-0.17226->Node3-0.23781->Node7-0.04678->Node30-0.00259->Node31-0.00822->Node41                                                                                                                                                            
#> [141] -0.531249>Node2-0.171906>Node3-0.237322>Node7-0.008183>Node8-0.012240>Node11-0.008693>Node14-0.000314>Node15-0.006506>Node20-0.005926>Node23-0.017661>Node42                                                                                             
#> [142] --0.5312->Node2--0.3250->Node5--0.1282->Node25--0.0157->Node43                                                                                                                                                                                           
#> [143] -0.53070->Node2-0.17173->Node3-0.14361->Node4-0.10327->Node9-0.00259->Node10-0.04811->Node44                                                                                                                                                             
#> [144] -0.89374->Node6-0.08901->Node27-0.00817->Node36-0.00232->Node38-0.00676->Node45                                                                                                                                                                          
#> [145] -0.529334>Node2-0.171287>Node3-0.236467>Node7-0.008153>Node8-0.012196>Node11-0.008662>Node14-0.000313>Node15-0.006482>Node20-0.005905>Node23-0.017598>Node42-0.003604>Node46                                                                             
#> [146] -5.29e-01>Node2-1.71e-01>Node3-2.36e-01>Node7-8.15e-03>Node8-2.25e-02>Node16-8.73e-05>Node17-2.25e-02>Node35-9.78e-03>Node47                                                                                                                             
#> [147] -0.893031>Node6-0.088938>Node27-0.008168>Node36-0.002318>Node38-0.006752>Node45-0.000793>Node48                                                                                                                                                          
#> [148] -0.52885->Node2-0.17113->Node3-0.14311->Node4-0.10291->Node9-0.00258->Node10-0.04794->Node44-0.00348->Node49                                                                                                                                             
#> [149] -0.52879->Node2-0.17111->Node3-0.23622->Node7-0.00815->Node8-0.01218->Node11-0.00278->Node12-0.00237->Node13-0.01830->Node24-0.02010->Node50                                                                                                             
#> [150] -0.52878->Node2-0.17111->Node3-0.23622->Node7-0.00814->Node8-0.01218->Node11-0.00278->Node12-0.00237->Node13-0.01570->Node22-0.00436->Node28-0.01836->Node51                                                                                             
#> [151] -5.28e-01>Node2-1.71e-01>Node3-2.36e-01>Node7-8.14e-03>Node8-2.25e-02>Node16-8.72e-05>Node17-2.45e-02>Node37-9.40e-03>Node52                                                                                                                             
#> [152] -0.52823->Node2-0.17093->Node3-0.23597->Node7-0.00814->Node8-0.01217->Node11-0.00277->Node12-0.00237->Node13-0.01568->Node22-0.00435->Node28-0.01834->Node51-0.00104->Node53                                                                             
#> [153] -5.28e-01>Node2-1.71e-01>Node3-2.36e-01>Node7-8.13e-03>Node8-2.25e-02>Node16-8.71e-05>Node17-2.24e-02>Node35-2.95e-03>Node39-9.30e-03>Node54                                                                                                             
#> [154] -0.52785->Node2-0.17081->Node3-0.23580->Node7-0.00813->Node8-0.01216->Node11-0.00277->Node12-0.02815->Node32-0.00538->Node40-0.00895->Node55                                                                                                             
#> [155] -0.527623>Node2-0.170733>Node3-0.235702>Node7-0.008127>Node8-0.012156>Node11-0.008634>Node14-0.000312>Node15-0.006461>Node20-0.005886>Node23-0.017541>Node42-0.003592>Node46-0.003233>Node56                                                             
#> [156] -0.527577>Node2-0.170718>Node3-0.235682>Node7-0.008126>Node8-0.012155>Node11-0.008633>Node14-0.000312>Node15-0.006461>Node20-0.030335>Node57                                                                                                             
#> [157] --0.527-->Node2--0.323-->Node5--0.150-->Node58                                                                                                                                                                                                           
#> [158] -0.527103>Node2-0.170565>Node3-0.235470>Node7-0.008119>Node8-0.012144>Node11-0.008625>Node14-0.000312>Node15-0.006455>Node20-0.005880>Node23-0.017524>Node42-0.003588>Node46-0.003230>Node56-0.000985>Node59                                             
#> [159] -0.52689->Node2-0.17050->Node3-0.14257->Node4-0.10253->Node9-0.00257->Node10-0.05494->Node60                                                                                                                                                             
#> [160] -0.52670->Node2-0.32228->Node5-0.14994->Node58-0.00107->Node61                                                                                                                                                                                           
#> [161] -0.526499>Node2-0.170369>Node3-0.235200>Node7-0.008110>Node8-0.012130>Node11-0.008615>Node14-0.000311>Node15-0.006448>Node20-0.005873>Node23-0.017504>Node42-0.003584>Node46-0.003226>Node56-0.000984>Node59-0.001146>Node62                             
#> [162] -0.52647->Node2-0.17036->Node3-0.23519->Node7-0.00811->Node8-0.01213->Node11-0.00861->Node14-0.01694->Node29-0.02218->Node63                                                                                                                             
#> [163] -0.525572>Node2-0.170070>Node3-0.234786>Node7-0.008095>Node8-0.012109>Node11-0.008600>Node14-0.000311>Node15-0.006436>Node20-0.005863>Node23-0.017473>Node42-0.003578>Node46-0.003221>Node56-0.003886>Node64                                             
#> [164] --0.5256->Node2--0.1701->Node3--0.1422->Node4--0.1244->Node19--0.0378->Node65                                                                                                                                                                            
#> [165] -0.52537->Node2-0.17000->Node3-0.14216->Node4-0.10223->Node9-0.00256->Node10-0.04762->Node44-0.00345->Node49-0.00660->Node66                                                                                                                             
#> [166] -0.88632->Node6-0.08827->Node27-0.00811->Node36-0.00230->Node38-0.01501->Node67                                                                                                                                                                          
#> [167] -0.52497->Node2-0.16987->Node3-0.23452->Node7-0.00809->Node8-0.01209->Node11-0.00859->Node14-0.00031->Node15-0.00643->Node20-0.00586->Node23-0.00262->Node26-0.02666->Node68                                                                             
#> [168] -0.52486->Node2-0.16984->Node3-0.23447->Node7-0.00808->Node8-0.01209->Node11-0.00859->Node14-0.00031->Node15-0.00643->Node20-0.00586->Node23-0.01745->Node42-0.00357->Node46-0.00322->Node56-0.00388->Node64-0.00135->Node69                             
#> [169] -0.524806>Node2-0.169821>Node3-0.234444>Node7-0.008084>Node8-0.012091>Node11-0.008588>Node14-0.000310>Node15-0.006427>Node20-0.005854>Node23-0.002622>Node26-0.026648>Node68-0.000305>Node70                                                             
#> [170] -0.88481->Node6-0.08812->Node27-0.00809->Node36-0.00230->Node38-0.00669->Node45-0.01000->Node71                                                                                                                                                          
#> [171] --0.8847->Node6--0.0747->Node18--0.0407->Node72                                                                                                                                                                                                          
#> [172] -0.524192>Node2-0.169623>Node3-0.234170>Node7-0.008074>Node8-0.012077>Node11-0.008578>Node14-0.000310>Node15-0.006419>Node20-0.005848>Node23-0.002619>Node26-0.026617>Node68-0.000304>Node70-0.001169>Node73                                             
#> [173] -0.52418->Node2-0.16962->Node3-0.23417->Node7-0.04606->Node30-0.00255->Node31-0.00809->Node41-0.01533->Node74                                                                                                                                            
#> [174] -0.884489>Node6-0.088087>Node27-0.008090>Node36-0.002296>Node38-0.006687>Node45-0.000786>Node48-0.009566>Node75                                                                                                                                          
#> [175] -0.52409->Node2-0.16959->Node3-0.23412->Node7-0.00807->Node8-0.01207->Node11-0.00275->Node12-0.00235->Node13-0.01556->Node22-0.00432->Node28-0.01820->Node51-0.00888->Node76                                                                             
#> [176] -0.884246>Node6-0.088063>Node27-0.008087>Node36-0.002295>Node38-0.006685>Node45-0.009990>Node71-0.000634>Node77                                                                                                                                          
#> [177] -0.88400->Node6-0.08804->Node27-0.00809->Node36-0.01988->Node78                                                                                                                                                                                          
#> [178] -0.88377->Node6-0.08802->Node27-0.00674->Node33-0.02148->Node79                                                                                                                                                                                          
#> [179] -0.523629>Node2-0.169441>Node3-0.233918>Node7-0.008065>Node8-0.012064>Node11-0.008568>Node14-0.000310>Node15-0.006412>Node20-0.005841>Node23-0.017408>Node42-0.003565>Node46-0.003209>Node56-0.000979>Node59-0.001140>Node62-0.005451>Node80             
#> [180] -0.52331->Node2-0.16934->Node3-0.14161->Node4-0.10183->Node9-0.00255->Node10-0.05457->Node60-0.00679->Node81                                                                                                                                             
#> [181] -0.52304->Node2-0.32004->Node5-0.14890->Node58-0.00801->Node82                                                                                                                                                                                           
#> [182] -0.522927>Node2-0.169213>Node3-0.233604>Node7-0.008055>Node8-0.012048>Node11-0.008557>Node14-0.000309>Node15-0.045287>Node83                                                                                                                             
#> [183] -0.52290->Node2-0.31995->Node5-0.14886->Node58-0.00107->Node61-0.00722->Node84                                                                                                                                                                           
#> [184] -0.52255->Node2-0.16909->Node3-0.23344->Node7-0.04592->Node30-0.00254->Node31-0.00807->Node41-0.01528->Node74-0.00312->Node85                                                                                                                            
#> [185] -0.52243->Node2-0.16905->Node3-0.23338->Node7-0.00805->Node8-0.01204->Node11-0.00274->Node12-0.02787->Node32-0.00532->Node40-0.00886->Node55-0.01026->Node86                                                                                             
#> [186] -0.52230->Node2-0.16901->Node3-0.23332->Node7-0.00804->Node8-0.01203->Node11-0.00855->Node14-0.01681->Node29-0.02200->Node63-0.00793->Node87                                                                                                             
#> [187] -5.22e-01>Node2-1.69e-01>Node3-2.33e-01>Node7-8.04e-03>Node8-2.22e-02>Node16-8.61e-05>Node17-2.22e-02>Node35-2.92e-03>Node39-9.19e-03>Node54-1.12e-02>Node88                                                                                             
#> [188] -0.52167->Node2-0.16881->Node3-0.14116->Node4-0.12348->Node19-0.03748->Node65-0.00739->Node89                                                                                                                                                            
#> [189] -0.880319>Node6-0.087672>Node27-0.008052>Node36-0.002285>Node38-0.006656>Node45-0.009945>Node71-0.000631>Node77-0.004441>Node90                                                                                                                          
#> [190] --0.5216->Node2--0.1688->Node3--0.2330->Node7--0.0458->Node30--0.0308->Node91                                                                                                                                                                            
#> [191] -0.521551>Node2-0.168768>Node3-0.232990>Node7-0.008033>Node8-0.012016>Node11-0.008534>Node14-0.000308>Node15-0.006387>Node20-0.005818>Node23-0.017339>Node42-0.003551>Node46-0.014704>Node92                                                             
#> [192] -0.88003->Node6-0.08764->Node27-0.00805->Node36-0.00228->Node38-0.01490->Node67-0.00709->Node93                                                                                                                                                          
#> [193] -0.52148->Node2-0.16875->Node3-0.23296->Node7-0.00803->Node8-0.01201->Node11-0.00853->Node14-0.01678->Node29-0.02197->Node63-0.00792->Node87-0.00155->Node94                                                                                             
#> [194] -0.52129->Node2-0.16868->Node3-0.14106->Node4-0.10144->Node9-0.00254->Node10-0.04725->Node44-0.00343->Node49-0.01430->Node95                                                                                                                             
#> [195] -0.52124->Node2-0.16867->Node3-0.23285->Node7-0.00803->Node8-0.01201->Node11-0.00274->Node12-0.00234->Node13-0.01548->Node22-0.00429->Node28-0.01810->Node51-0.00103->Node53-0.01322->Node96                                                             
#> [196] -0.521160>Node2-0.168642>Node3-0.232815>Node7-0.008027>Node8-0.012007>Node11-0.008528>Node14-0.000308>Node15-0.006382>Node20-0.005814>Node23-0.002604>Node26-0.026463>Node68-0.000303>Node70-0.006947>Node97                                             
#> [197] -0.521079>Node2-0.168615>Node3-0.232779>Node7-0.008026>Node8-0.012005>Node11-0.008527>Node14-0.000308>Node15-0.006381>Node20-0.005813>Node23-0.002603>Node26-0.026459>Node68-0.000303>Node70-0.001162>Node73-0.005940>Node98                             
#> [198] -0.52103->Node2-0.16860->Node3-0.23276->Node7-0.00803->Node8-0.01200->Node11-0.00853->Node14-0.01677->Node29-0.02195->Node63-0.01035->Node99