Scale the phylogenetic flow features to a constant sum.

Can be used to standardise branch lengths to reasonable values

Usage

pf_scale_flow_sum(x, scale_to = 1)

Arguments

x

a pfc object

scale_to

The value to scale the sums to. Default is 1.

Value

a scaled pfc object

Examples

pf_scale_flow_sum(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] -0.96231->Node11-0.00603->Node15-0.00471->Node17-0.02097->Node47-0.00456->Node80-0.00142->t20                                                                                             
#>   [2] -0.96231->Node11-0.00603->Node15-0.00471->Node17-0.02097->Node47-0.00456->Node80-0.00142->t89                                                                                             
#>   [3] -0.96231->Node11-0.00603->Node15-0.00471->Node17-0.02097->Node47-0.00598->t48                                                                                                             
#>   [4] -0.96231->Node11-0.00603->Node15-0.00471->Node17-0.01961->Node42-0.00734->t38                                                                                                             
#>   [5] -0.96231->Node11-0.00603->Node15-0.00471->Node17-0.01961->Node42-0.00236->Node53-0.00178->Node63-0.00320->t90                                                                             
#>   [6] -0.96231->Node11-0.00603->Node15-0.00471->Node17-0.01961->Node42-0.00236->Node53-0.00178->Node63-0.00267->Node90-0.00053->t98                                                             
#>   [7] -0.96231->Node11-0.00603->Node15-0.00471->Node17-0.01961->Node42-0.00236->Node53-0.00178->Node63-0.00267->Node90-0.00053->t5                                                              
#>   [8] -0.96231->Node11-0.00603->Node15-0.00471->Node17-0.01961->Node42-0.00236->Node53-0.00498->t52                                                                                             
#>   [9] -0.96231->Node11-0.00603->Node15-0.01358->Node25-0.01808->t14                                                                                                                             
#>  [10] -0.96231->Node11-0.00603->Node15-0.01358->Node25-0.00367->Node28-0.01016->Node55-0.00425->t94                                                                                             
#>  [11] -0.96231->Node11-0.00603->Node15-0.01358->Node25-0.00367->Node28-0.01016->Node55-0.00425->t77                                                                                             
#>  [12] -0.96231->Node11-0.00603->Node15-0.01358->Node25-0.00367->Node28-0.01323->Node82-0.00118->t24                                                                                             
#>  [13] -0.96231->Node11-0.00603->Node15-0.01358->Node25-0.00367->Node28-0.01323->Node82-0.00118->t23                                                                                             
#>  [14] -0.962314>Node11-0.010897>Node18-0.000594>Node19-0.026195>t62                                                                                                                             
#>  [15] -0.962314>Node11-0.010897>Node18-0.000594>Node19-0.019754>Node45-0.003239>Node62-0.003202>t10                                                                                             
#>  [16] -0.962314>Node11-0.010897>Node18-0.000594>Node19-0.019754>Node45-0.003239>Node62-0.003202>t50                                                                                             
#>  [17] -0.962314>Node11-0.010897>Node18-0.000594>Node19-0.019754>Node45-0.005995>Node94-0.000446>t83                                                                                             
#>  [18] -0.962314>Node11-0.010897>Node18-0.000594>Node19-0.019754>Node45-0.005995>Node94-0.000446>t58                                                                                             
#>  [19] -0.96231->Node11-0.01090->Node18-0.01907->Node41-0.00592->Node77-0.00180->t76                                                                                                             
#>  [20] -0.96231->Node11-0.01090->Node18-0.01907->Node41-0.00592->Node77-0.00180->t69                                                                                                             
#>  [21] -0.96231->Node11-0.01090->Node18-0.01907->Node41-0.00772->t61                                                                                                                             
#>  [22] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.00539->Node10-0.02724->Node30-0.00885->Node54-0.00443->t88                                                                 
#>  [23] -0.818111>Node2-0.001217>Node3-0.014590>Node4-0.120171>Node9-0.005391>Node10-0.027244>Node30-0.008850>Node54-0.002642>Node78-0.001162>Node89-0.000623>t18                                 
#>  [24] -0.818111>Node2-0.001217>Node3-0.014590>Node4-0.120171>Node9-0.005391>Node10-0.027244>Node30-0.008850>Node54-0.002642>Node78-0.001162>Node89-0.000623>t30                                 
#>  [25] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.00539->Node10-0.02724->Node30-0.00885->Node54-0.00264->Node78-0.00178->t34                                                 
#>  [26] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.00539->Node10-0.02724->Node30-0.00735->Node48-0.00226->Node58-0.00141->Node73-0.00226->t27                                 
#>  [27] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.00539->Node10-0.02724->Node30-0.00735->Node48-0.00226->Node58-0.00141->Node73-0.00226->t39                                 
#>  [28] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.00539->Node10-0.02724->Node30-0.00735->Node48-0.00226->Node58-0.00367->t56                                                 
#>  [29] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.00539->Node10-0.02724->Node30-0.00735->Node48-0.00276->Node64-0.00316->t13                                                 
#>  [30] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.00539->Node10-0.02724->Node30-0.00735->Node48-0.00276->Node64-0.00316->t66                                                 
#>  [31] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.00539->Node10-0.01567->Node22-0.02485->t12                                                                                 
#>  [32] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.00539->Node10-0.01567->Node22-0.01285->Node33-0.01200->t45                                                                 
#>  [33] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.00539->Node10-0.01567->Node22-0.01285->Node33-0.00668->Node51-0.00532->t17                                                 
#>  [34] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.00539->Node10-0.01567->Node22-0.01285->Node33-0.00668->Node51-0.00367->Node79-0.00165->t91                                 
#>  [35] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.00539->Node10-0.01567->Node22-0.01285->Node33-0.00668->Node51-0.00367->Node79-0.00165->t36                                 
#>  [36] -0.818111>Node2-0.001217>Node3-0.014590>Node4-0.120171>Node9-0.010730>Node13-0.003701>Node16-0.019213>Node31-0.002998>Node38-0.000212>Node39-0.009058>t65                                 
#>  [37] -0.818111>Node2-0.001217>Node3-0.014590>Node4-0.120171>Node9-0.010730>Node13-0.003701>Node16-0.019213>Node31-0.002998>Node38-0.000212>Node39-0.000290>Node40-0.008768>t11                 
#>  [38] -0.818111>Node2-0.001217>Node3-0.014590>Node4-0.120171>Node9-0.010730>Node13-0.003701>Node16-0.019213>Node31-0.002998>Node38-0.000212>Node39-0.000290>Node40-0.005882>Node65-0.002886>t19 
#>  [39] -0.818111>Node2-0.001217>Node3-0.014590>Node4-0.120171>Node9-0.010730>Node13-0.003701>Node16-0.019213>Node31-0.002998>Node38-0.000212>Node39-0.000290>Node40-0.005882>Node65-0.002886>t31 
#>  [40] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.01073->Node13-0.00370->Node16-0.01921->Node31-0.00300->Node38-0.00927->t95                                                 
#>  [41] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.01073->Node13-0.00370->Node16-0.01921->Node31-0.01227->t22                                                                 
#>  [42] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.01073->Node13-0.00370->Node16-0.03148->t73                                                                                 
#>  [43] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.01073->Node13-0.01759->Node26-0.00396->Node29-0.01158->Node74-0.00204->t85                                                 
#>  [44] -0.818111>Node2-0.001217>Node3-0.014590>Node4-0.120171>Node9-0.010730>Node13-0.017592>Node26-0.003963>Node29-0.011585>Node74-0.001666>Node96-0.000375>t75                                 
#>  [45] -0.818111>Node2-0.001217>Node3-0.014590>Node4-0.120171>Node9-0.010730>Node13-0.017592>Node26-0.003963>Node29-0.011585>Node74-0.001666>Node96-0.000375>t41                                 
#>  [46] -0.818111>Node2-0.001217>Node3-0.014590>Node4-0.120171>Node9-0.010730>Node13-0.017592>Node26-0.003963>Node29-0.001781>Node34-0.011442>Node95-0.000404>t63                                 
#>  [47] -0.818111>Node2-0.001217>Node3-0.014590>Node4-0.120171>Node9-0.010730>Node13-0.017592>Node26-0.003963>Node29-0.001781>Node34-0.011442>Node95-0.000404>t55                                 
#>  [48] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.01073->Node13-0.01759->Node26-0.00396->Node29-0.00178->Node34-0.01086->Node85-0.00099->t60                                 
#>  [49] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.01073->Node13-0.01759->Node26-0.00396->Node29-0.00178->Node34-0.01086->Node85-0.00099->t9                                  
#>  [50] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.12017->Node9-0.01073->Node13-0.01759->Node26-0.01759->t82                                                                                 
#>  [51] -0.81811->Node2-0.00122->Node3-0.01459->Node4-0.16608->t51                                                                                                                                
#>  [52] -0.818111>Node2-0.001217>Node3-0.131937>Node8-0.047988>Node86-0.000271>Node93-0.000477>t79                                                                                                
#>  [53] -0.818111>Node2-0.001217>Node3-0.131937>Node8-0.047988>Node86-0.000271>Node93-0.000477>t21                                                                                                
#>  [54] -0.818111>Node2-0.001217>Node3-0.131937>Node8-0.047988>Node86-0.000748>t8                                                                                                                 
#>  [55] -0.81811->Node2-0.00122->Node3-0.13194->Node8-0.01344->Node12-0.01671->Node23-0.01468->Node56-0.00390->t37                                                                                
#>  [56] -0.818111>Node2-0.001217>Node3-0.131937>Node8-0.013438>Node12-0.016712>Node23-0.014684>Node56-0.000616>Node61-0.003286>t97                                                                
#>  [57] -0.818111>Node2-0.001217>Node3-0.131937>Node8-0.013438>Node12-0.016712>Node23-0.014684>Node56-0.000616>Node61-0.001409>Node76-0.001877>t53                                                
#>  [58] -0.818111>Node2-0.001217>Node3-0.131937>Node8-0.013438>Node12-0.016712>Node23-0.014684>Node56-0.000616>Node61-0.001409>Node76-0.001877>t3                                                 
#>  [59] -0.81811->Node2-0.00122->Node3-0.13194->Node8-0.01344->Node12-0.01671->Node23-0.01125->Node43-0.00733->t6                                                                                 
#>  [60] -0.81811->Node2-0.00122->Node3-0.13194->Node8-0.01344->Node12-0.01671->Node23-0.01125->Node43-0.00381->Node59-0.00228->Node81-0.00124->t16                                                
#>  [61] -0.81811->Node2-0.00122->Node3-0.13194->Node8-0.01344->Node12-0.01671->Node23-0.01125->Node43-0.00381->Node59-0.00228->Node81-0.00124->t71                                                
#>  [62] -0.818111>Node2-0.001217>Node3-0.131937>Node8-0.013438>Node12-0.016712>Node23-0.011252>Node43-0.003814>Node59-0.003296>Node97-0.000224>t86                                                
#>  [63] -0.818111>Node2-0.001217>Node3-0.131937>Node8-0.013438>Node12-0.016712>Node23-0.011252>Node43-0.003814>Node59-0.003296>Node97-0.000224>t68                                                
#>  [64] -0.81811->Node2-0.00122->Node3-0.13194->Node8-0.01344->Node12-0.00922->Node20-0.02608->t25                                                                                                
#>  [65] -0.81811->Node2-0.00122->Node3-0.13194->Node8-0.01344->Node12-0.00922->Node20-0.01402->Node32-0.01206->t80                                                                                
#>  [66] -0.81811->Node2-0.00122->Node3-0.13194->Node8-0.01344->Node12-0.00922->Node20-0.01402->Node32-0.00664->Node50-0.00542->t4                                                                 
#>  [67] -0.818111>Node2-0.001217>Node3-0.131937>Node8-0.013438>Node12-0.009220>Node20-0.014022>Node32-0.006635>Node50-0.002566>Node66-0.002201>Node87-0.000549>Node98-0.000106>t33                
#>  [68] -8.18e-01>Node2-1.22e-03>Node3-1.32e-01>Node8-1.34e-02>Node12-9.22e-03>Node20-1.40e-02>Node32-6.64e-03>Node50-2.57e-03>Node66-2.20e-03>Node87-5.49e-04>Node98-7.69e-05>Node99-2.90e-05>t43
#>  [69] -8.18e-01>Node2-1.22e-03>Node3-1.32e-01>Node8-1.34e-02>Node12-9.22e-03>Node20-1.40e-02>Node32-6.64e-03>Node50-2.57e-03>Node66-2.20e-03>Node87-5.49e-04>Node98-7.69e-05>Node99-2.90e-05>t59
#>  [70] -0.818111>Node2-0.001217>Node3-0.131937>Node8-0.013438>Node12-0.009220>Node20-0.014022>Node32-0.006635>Node50-0.002566>Node66-0.002201>Node87-0.000655>t57                                
#>  [71] -0.81811->Node2-0.00122->Node3-0.13194->Node8-0.01344->Node12-0.00922->Node20-0.01402->Node32-0.00664->Node50-0.00257->Node66-0.00286->t64                                                
#>  [72] -0.81811->Node2-0.04336->Node5-0.08565->Node7-0.04653->Node46-0.00298->Node60-0.00135->Node75-0.00203->t7                                                                                 
#>  [73] -0.81811->Node2-0.04336->Node5-0.08565->Node7-0.04653->Node46-0.00298->Node60-0.00135->Node75-0.00203->t2                                                                                 
#>  [74] -0.818111>Node2-0.043355>Node5-0.085650>Node7-0.046528>Node46-0.002981>Node60-0.000758>Node69-0.002618>t92                                                                                
#>  [75] -0.818111>Node2-0.043355>Node5-0.085650>Node7-0.046528>Node46-0.002981>Node60-0.000758>Node69-0.001975>Node88-0.000643>t26                                                                
#>  [76] -0.818111>Node2-0.043355>Node5-0.085650>Node7-0.046528>Node46-0.002981>Node60-0.000758>Node69-0.001975>Node88-0.000643>t32                                                                
#>  [77] -0.81811->Node2-0.04336->Node5-0.08565->Node7-0.04653->Node46-0.00636->t1                                                                                                                 
#>  [78] --0.8181->Node2--0.0434->Node5--0.0856->Node7--0.0529->t40                                                                                                                                
#>  [79] --0.8181->Node2--0.0434->Node5--0.0410->Node6--0.0638->Node14--0.0154->Node24--0.0183->t28                                                                                                
#>  [80] -0.818111>Node2-0.043355>Node5-0.041005>Node6-0.063850>Node14-0.015352>Node24-0.008831>Node36-0.004317>Node52-0.004676>Node91-0.000504>t15                                                
#>  [81] -0.818111>Node2-0.043355>Node5-0.041005>Node6-0.063850>Node14-0.015352>Node24-0.008831>Node36-0.004317>Node52-0.004676>Node91-0.000504>t81                                                
#>  [82] -0.81811->Node2-0.04336->Node5-0.04101->Node6-0.06385->Node14-0.01535->Node24-0.00883->Node36-0.00432->Node52-0.00518->t46                                                                
#>  [83] -0.81811->Node2-0.04336->Node5-0.04101->Node6-0.06385->Node14-0.01535->Node24-0.00883->Node36-0.00689->Node70-0.00148->Node83-0.00113->t84                                                
#>  [84] -0.81811->Node2-0.04336->Node5-0.04101->Node6-0.06385->Node14-0.01535->Node24-0.00883->Node36-0.00689->Node70-0.00148->Node83-0.00113->t29                                                
#>  [85] -8.18e-01>Node2-4.34e-02>Node5-4.10e-02>Node6-6.38e-02>Node14-1.54e-02>Node24-8.83e-03>Node36-6.89e-03>Node70-5.70e-06>Node71-2.61e-03>t100                                               
#>  [86] -8.18e-01>Node2-4.34e-02>Node5-4.10e-02>Node6-6.38e-02>Node14-1.54e-02>Node24-8.83e-03>Node36-6.89e-03>Node70-5.70e-06>Node71-2.61e-03>t72                                                
#>  [87] -0.81811->Node2-0.04336->Node5-0.04101->Node6-0.06385->Node14-0.00879->Node21-0.00922->Node27-0.01284->Node67-0.00282->t70                                                                
#>  [88] -0.81811->Node2-0.04336->Node5-0.04101->Node6-0.06385->Node14-0.00879->Node21-0.00922->Node27-0.01284->Node67-0.00282->t44                                                                
#>  [89] -0.81811->Node2-0.04336->Node5-0.04101->Node6-0.06385->Node14-0.00879->Node21-0.00922->Node27-0.00552->Node35-0.01015->t96                                                                
#>  [90] -0.81811->Node2-0.04336->Node5-0.04101->Node6-0.06385->Node14-0.00879->Node21-0.00922->Node27-0.00552->Node35-0.01015->t47                                                                
#>  [91] -0.818111>Node2-0.043355>Node5-0.041005>Node6-0.063850>Node14-0.008785>Node21-0.018441>Node44-0.000873>Node49-0.005580>t67                                                                
#>  [92] -0.818111>Node2-0.043355>Node5-0.041005>Node6-0.063850>Node14-0.008785>Node21-0.018441>Node44-0.000873>Node49-0.005580>t74                                                                
#>  [93] -0.81811->Node2-0.04336->Node5-0.04101->Node6-0.06385->Node14-0.00879->Node21-0.01844->Node44-0.00645->t35                                                                                
#>  [94] -0.818111>Node2-0.043355>Node5-0.041005>Node6-0.088178>Node37-0.005501>Node57-0.003361>Node92-0.000489>t42                                                                                
#>  [95] -0.818111>Node2-0.043355>Node5-0.041005>Node6-0.088178>Node37-0.005501>Node57-0.003361>Node92-0.000489>t99                                                                                
#>  [96] -0.81811->Node2-0.04336->Node5-0.04101->Node6-0.08818->Node37-0.00550->Node57-0.00105->Node68-0.00280->t87                                                                                
#>  [97] -0.81811->Node2-0.04336->Node5-0.04101->Node6-0.08818->Node37-0.00550->Node57-0.00105->Node68-0.00171->Node84-0.00108->t54                                                                
#>  [98] -0.81811->Node2-0.04336->Node5-0.04101->Node6-0.08818->Node37-0.00550->Node57-0.00105->Node68-0.00171->Node84-0.00108->t93                                                                
#>  [99] -0.81811->Node2-0.04336->Node5-0.04101->Node6-0.08818->Node37-0.00707->Node72-0.00228->t49                                                                                                
#> [100] -0.81811->Node2-0.04336->Node5-0.04101->Node6-0.08818->Node37-0.00707->Node72-0.00228->t78                                                                                                
#> [101] ----1---->Node2                                                                                                                                                                           
#> [102] -0.99851->Node2-0.00149->Node3                                                                                                                                                            
#> [103] -0.98105->Node2-0.00146->Node3-0.01750->Node4                                                                                                                                             
#> [104] --0.9497->Node2--0.0503->Node5                                                                                                                                                            
#> [105] --0.9065->Node2--0.0480->Node5--0.0454->Node6                                                                                                                                             
#> [106] --0.8638->Node2--0.0458->Node5--0.0904->Node7                                                                                                                                             
#> [107] -0.86002->Node2-0.00128->Node3-0.13870->Node8                                                                                                                                             
#> [108] -0.85748->Node2-0.00128->Node3-0.01529->Node4-0.12595->Node9                                                                                                                              
#> [109] -0.85266->Node2-0.00127->Node3-0.01521->Node4-0.12525->Node9-0.00562->Node10                                                                                                              
#> [110] ----1---->Node11                                                                                                                                                                          
#> [111] -0.84804->Node2-0.00126->Node3-0.13676->Node8-0.01393->Node12                                                                                                                             
#> [112] -0.84794->Node2-0.00126->Node3-0.01512->Node4-0.12455->Node9-0.01112->Node13                                                                                                              
#> [113] --0.8466->Node2--0.0449->Node5--0.0424->Node6--0.0661->Node14                                                                                                                             
#> [114] -0.99378->Node11-0.00622->Node15                                                                                                                                                          
#> [115] -0.84470->Node2-0.00126->Node3-0.01506->Node4-0.12408->Node9-0.01108->Node13-0.00382->Node16                                                                                              
#> [116] -0.98897->Node11-0.00619->Node15-0.00484->Node17                                                                                                                                          
#> [117] --0.9888->Node11--0.0112->Node18                                                                                                                                                          
#> [118] -0.98820->Node11-0.01119->Node18-0.00061->Node19                                                                                                                                          
#> [119] -0.84002->Node2-0.00125->Node3-0.13547->Node8-0.01380->Node12-0.00947->Node20                                                                                                             
#> [120] -0.83900->Node2-0.04446->Node5-0.04205->Node6-0.06548->Node14-0.00901->Node21                                                                                                             
#> [121] -0.83896->Node2-0.00125->Node3-0.01496->Node4-0.12323->Node9-0.00553->Node10-0.01607->Node22                                                                                              
#> [122] -0.83360->Node2-0.00124->Node3-0.13444->Node8-0.01369->Node12-0.01703->Node23                                                                                                             
#> [123] --0.8334->Node2--0.0442->Node5--0.0418->Node6--0.0650->Node14--0.0156->Node24                                                                                                             
#> [124] -0.98004->Node11-0.00614->Node15-0.01383->Node25                                                                                                                                          
#> [125] -0.83276->Node2-0.00124->Node3-0.01485->Node4-0.12232->Node9-0.01092->Node13-0.01791->Node26                                                                                              
#> [126] -0.83113->Node2-0.04405->Node5-0.04166->Node6-0.06487->Node14-0.00892->Node21-0.00937->Node27                                                                                             
#> [127] -0.97638->Node11-0.00611->Node15-0.01377->Node25-0.00373->Node28                                                                                                                          
#> [128] -0.82941->Node2-0.00123->Node3-0.01479->Node4-0.12183->Node9-0.01088->Node13-0.01784->Node26-0.00402->Node29                                                                              
#> [129] -0.82912->Node2-0.00123->Node3-0.01479->Node4-0.12179->Node9-0.00546->Node10-0.02761->Node30                                                                                              
#> [130] -0.82827->Node2-0.00123->Node3-0.01477->Node4-0.12166->Node9-0.01086->Node13-0.00375->Node16-0.01945->Node31                                                                              
#> [131] -0.82809->Node2-0.00123->Node3-0.13355->Node8-0.01360->Node12-0.00933->Node20-0.01419->Node32                                                                                             
#> [132] -0.82805->Node2-0.00123->Node3-0.01477->Node4-0.12163->Node9-0.00546->Node10-0.01586->Node22-0.01301->Node33                                                                              
#> [133] -0.82792->Node2-0.00123->Node3-0.01476->Node4-0.12161->Node9-0.01086->Node13-0.01780->Node26-0.00401->Node29-0.00180->Node34                                                              
#> [134] -0.82650->Node2-0.04380->Node5-0.04143->Node6-0.06450->Node14-0.00888->Node21-0.00932->Node27-0.00558->Node35                                                                             
#> [135] -0.82595->Node2-0.04377->Node5-0.04140->Node6-0.06446->Node14-0.01550->Node24-0.00892->Node36                                                                                             
#> [136] --0.8258->Node2--0.0438->Node5--0.0414->Node6--0.0890->Node37                                                                                                                             
#> [137] -0.82577->Node2-0.00123->Node3-0.01473->Node4-0.12130->Node9-0.01083->Node13-0.00374->Node16-0.01939->Node31-0.00303->Node38                                                              
#> [138] -0.825589>Node2-0.001228>Node3-0.014723>Node4-0.121269>Node9-0.010828>Node13-0.003735>Node16-0.019388>Node31-0.003025>Node38-0.000214>Node39                                              
#> [139] -0.825347>Node2-0.001228>Node3-0.014719>Node4-0.121234>Node9-0.010825>Node13-0.003734>Node16-0.019383>Node31-0.003024>Node38-0.000214>Node39-0.000293>Node40                              
#> [140] --0.9698->Node11--0.0110->Node18--0.0192->Node41                                                                                                                                          
#> [141] -0.96943->Node11-0.00607->Node15-0.00475->Node17-0.01975->Node42                                                                                                                          
#> [142] -0.82415->Node2-0.00123->Node3-0.13291->Node8-0.01354->Node12-0.01684->Node23-0.01134->Node43                                                                                             
#> [143] -0.82342->Node2-0.04364->Node5-0.04127->Node6-0.06426->Node14-0.00884->Node21-0.01856->Node44                                                                                             
#> [144] -0.968553>Node11-0.010967>Node18-0.000598>Node19-0.019882>Node45                                                                                                                          
#> [145] --0.8233->Node2--0.0436->Node5--0.0862->Node7--0.0468->Node46                                                                                                                             
#> [146] -0.96811->Node11-0.00606->Node15-0.00474->Node17-0.02109->Node47                                                                                                                          
#> [147] -0.82298->Node2-0.00122->Node3-0.01468->Node4-0.12089->Node9-0.00542->Node10-0.02741->Node30-0.00740->Node48                                                                              
#> [148] -0.822701>Node2-0.043598>Node5-0.041236>Node6-0.064208>Node14-0.008834>Node21-0.018545>Node44-0.000878>Node49                                                                             
#> [149] -0.82257->Node2-0.00122->Node3-0.13266->Node8-0.01351->Node12-0.00927->Node20-0.01410->Node32-0.00667->Node50                                                                             
#> [150] -0.82249->Node2-0.00122->Node3-0.01467->Node4-0.12081->Node9-0.00542->Node10-0.01575->Node22-0.01292->Node33-0.00671->Node51                                                              
#> [151] -0.82237->Node2-0.04358->Node5-0.04122->Node6-0.06418->Node14-0.01543->Node24-0.00888->Node36-0.00434->Node52                                                                             
#> [152] -0.96713->Node11-0.00606->Node15-0.00473->Node17-0.01971->Node42-0.00237->Node53                                                                                                          
#> [153] -0.82175->Node2-0.00122->Node3-0.01465->Node4-0.12071->Node9-0.00542->Node10-0.02737->Node30-0.00889->Node54                                                                              
#> [154] -0.96642->Node11-0.00605->Node15-0.01363->Node25-0.00369->Node28-0.01020->Node55                                                                                                          
#> [155] -0.82132->Node2-0.00122->Node3-0.13245->Node8-0.01349->Node12-0.01678->Node23-0.01474->Node56                                                                                             
#> [156] -0.82127->Node2-0.04352->Node5-0.04116->Node6-0.08852->Node37-0.00552->Node57                                                                                                             
#> [157] -0.82112->Node2-0.00122->Node3-0.01464->Node4-0.12061->Node9-0.00541->Node10-0.02734->Node30-0.00738->Node48-0.00226->Node58                                                              
#> [158] -0.82100->Node2-0.00122->Node3-0.13240->Node8-0.01349->Node12-0.01677->Node23-0.01129->Node43-0.00383->Node59                                                                             
#> [159] -0.82088->Node2-0.04350->Node5-0.08594->Node7-0.04669->Node46-0.00299->Node60                                                                                                             
#> [160] -0.820808>Node2-0.001221>Node3-0.132372>Node8-0.013482>Node12-0.016767>Node23-0.014733>Node56-0.000618>Node61                                                                             
#> [161] -0.965406>Node11-0.010932>Node18-0.000596>Node19-0.019818>Node45-0.003249>Node62                                                                                                          
#> [162] -0.96540->Node11-0.00604->Node15-0.00473->Node17-0.01967->Node42-0.00237->Node53-0.00179->Node63                                                                                          
#> [163] -0.82070->Node2-0.00122->Node3-0.01464->Node4-0.12055->Node9-0.00541->Node10-0.02733->Node30-0.00738->Node48-0.00277->Node64                                                              
#> [164] -0.820478>Node2-0.001220>Node3-0.014632>Node4-0.120519>Node9-0.010761>Node13-0.003712>Node16-0.019268>Node31-0.003007>Node38-0.000213>Node39-0.000291>Node40-0.005899>Node65              
#> [165] -0.82045->Node2-0.00122->Node3-0.13231->Node8-0.01348->Node12-0.00925->Node20-0.01406->Node32-0.00665->Node50-0.00257->Node66                                                             
#> [166] -0.82043->Node2-0.04348->Node5-0.04112->Node6-0.06403->Node14-0.00881->Node21-0.00925->Node27-0.01288->Node67                                                                             
#> [167] -0.82041->Node2-0.04348->Node5-0.04112->Node6-0.08843->Node37-0.00552->Node57-0.00106->Node68                                                                                             
#> [168] -0.82026->Node2-0.04347->Node5-0.08587->Node7-0.04665->Node46-0.00299->Node60-0.00076->Node69                                                                                             
#> [169] -0.82025->Node2-0.04347->Node5-0.04111->Node6-0.06402->Node14-0.01539->Node24-0.00885->Node36-0.00690->Node70                                                                             
#> [170] -8.20e-01>Node2-4.35e-02>Node5-4.11e-02>Node6-6.40e-02>Node14-1.54e-02>Node24-8.85e-03>Node36-6.90e-03>Node70-5.71e-06>Node71                                                             
#> [171] -0.81998->Node2-0.04345->Node5-0.04110->Node6-0.08838->Node37-0.00708->Node72                                                                                                             
#> [172] -0.81996->Node2-0.00122->Node3-0.01462->Node4-0.12044->Node9-0.00540->Node10-0.02731->Node30-0.00737->Node48-0.00226->Node58-0.00141->Node73                                              
#> [173] -0.81978->Node2-0.00122->Node3-0.01462->Node4-0.12042->Node9-0.01075->Node13-0.01763->Node26-0.00397->Node29-0.01161->Node74                                                              
#> [174] -0.81977->Node2-0.04344->Node5-0.08582->Node7-0.04662->Node46-0.00299->Node60-0.00135->Node75                                                                                             
#> [175] -0.819649>Node2-0.001219>Node3-0.132185>Node8-0.013463>Node12-0.016743>Node23-0.014712>Node56-0.000617>Node61-0.001411>Node76                                                             
#> [176] -0.96405->Node11-0.01092->Node18-0.01911->Node41-0.00593->Node77                                                                                                                          
#> [177] -0.81957->Node2-0.00122->Node3-0.01462->Node4-0.12039->Node9-0.00540->Node10-0.02729->Node30-0.00887->Node54-0.00265->Node78                                                              
#> [178] -0.81946->Node2-0.00122->Node3-0.01461->Node4-0.12037->Node9-0.00540->Node10-0.01569->Node22-0.01288->Node33-0.00669->Node51-0.00367->Node79                                              
#> [179] -0.96368->Node11-0.00603->Node15-0.00472->Node17-0.02099->Node47-0.00457->Node80                                                                                                          
#> [180] -0.81913->Node2-0.00122->Node3-0.13210->Node8-0.01345->Node12-0.01673->Node23-0.01127->Node43-0.00382->Node59-0.00228->Node81                                                             
#> [181] -0.96345->Node11-0.00603->Node15-0.01359->Node25-0.00368->Node28-0.01325->Node82                                                                                                          
#> [182] -0.81904->Node2-0.04340->Node5-0.04105->Node6-0.06392->Node14-0.01537->Node24-0.00884->Node36-0.00689->Node70-0.00148->Node83                                                             
#> [183] -0.81900->Node2-0.04340->Node5-0.04105->Node6-0.08827->Node37-0.00551->Node57-0.00105->Node68-0.00171->Node84                                                                             
#> [184] -0.81892->Node2-0.00122->Node3-0.01460->Node4-0.12029->Node9-0.01074->Node13-0.01761->Node26-0.00397->Node29-0.00178->Node34-0.01087->Node85                                              
#> [185] -0.81872->Node2-0.00122->Node3-0.13204->Node8-0.04802->Node86                                                                                                                             
#> [186] -0.81865->Node2-0.00122->Node3-0.13202->Node8-0.01345->Node12-0.00923->Node20-0.01403->Node32-0.00664->Node50-0.00257->Node66-0.00220->Node87                                             
#> [187] -0.818637>Node2-0.043383>Node5-0.085705>Node7-0.046558>Node46-0.002983>Node60-0.000758>Node69-0.001976>Node88                                                                             
#> [188] -0.81862->Node2-0.00122->Node3-0.01460->Node4-0.12025->Node9-0.00539->Node10-0.02726->Node30-0.00886->Node54-0.00264->Node78-0.00116->Node89                                              
#> [189] -0.96282->Node11-0.00603->Node15-0.00471->Node17-0.01962->Node42-0.00236->Node53-0.00178->Node63-0.00267->Node90                                                                          
#> [190] -0.81852->Node2-0.04338->Node5-0.04103->Node6-0.06388->Node14-0.01536->Node24-0.00884->Node36-0.00432->Node52-0.00468->Node91                                                             
#> [191] -0.81851->Node2-0.04338->Node5-0.04103->Node6-0.08822->Node37-0.00550->Node57-0.00336->Node92                                                                                             
#> [192] -0.818501>Node2-0.001217>Node3-0.132000>Node8-0.048010>Node86-0.000271>Node93                                                                                                             
#> [193] -0.962744>Node11-0.010901>Node18-0.000594>Node19-0.019763>Node45-0.005998>Node94                                                                                                          
#> [194] -0.81844->Node2-0.00122->Node3-0.01460->Node4-0.12022->Node9-0.01073->Node13-0.01760->Node26-0.00396->Node29-0.00178->Node34-0.01145->Node95                                              
#> [195] -0.81842->Node2-0.00122->Node3-0.01460->Node4-0.12022->Node9-0.01073->Node13-0.01760->Node26-0.00396->Node29-0.01159->Node74-0.00167->Node96                                              
#> [196] -0.81829->Node2-0.00122->Node3-0.13197->Node8-0.01344->Node12-0.01672->Node23-0.01125->Node43-0.00382->Node59-0.00330->Node97                                                             
#> [197] -0.818197>Node2-0.001217>Node3-0.131951>Node8-0.013439>Node12-0.009221>Node20-0.014023>Node32-0.006636>Node50-0.002566>Node66-0.002201>Node87-0.000549>Node98                             
#> [198] -8.18e-01>Node2-1.22e-03>Node3-1.32e-01>Node8-1.34e-02>Node12-9.22e-03>Node20-1.40e-02>Node32-6.64e-03>Node50-2.57e-03>Node66-2.20e-03>Node87-5.49e-04>Node98-7.69e-05>Node99