Scale the phylogenetic flow features to a constant sum.

Can be used to standardise branch lengths to reasonable values

Usage

pf_scale_flow_sum(x, scale_to = 1)

Arguments

x

a pfc object

scale_to

The value to scale the sums to. Default is 1.

Value

a scaled pfc object

Examples

pf_scale_flow_sum(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] -4.19e-01>Node2-2.12e-02>Node3-2.95e-01>Node10-2.25e-01>Node39-1.64e-02>Node48-7.15e-05>Node49-2.32e-02>t82                                                                                                              
#>   [2] -4.19e-01>Node2-2.12e-02>Node3-2.95e-01>Node10-2.25e-01>Node39-1.64e-02>Node48-7.15e-05>Node49-2.32e-02>t92                                                                                                              
#>   [3] -0.41920->Node2-0.02121->Node3-0.29482->Node10-0.22510->Node39-0.01639->Node48-0.00469->Node54-0.00995->Node69-0.00508->Node87-0.00180->Node92-0.00175->t19                                                              
#>   [4] -0.41920->Node2-0.02121->Node3-0.29482->Node10-0.22510->Node39-0.01639->Node48-0.00469->Node54-0.00995->Node69-0.00508->Node87-0.00180->Node92-0.00175->t11                                                              
#>   [5] -0.41920->Node2-0.02121->Node3-0.29482->Node10-0.22510->Node39-0.01639->Node48-0.00469->Node54-0.00995->Node69-0.00508->Node87-0.00355->t3                                                                               
#>   [6] -0.41920->Node2-0.02121->Node3-0.29482->Node10-0.22510->Node39-0.01639->Node48-0.00469->Node54-0.00995->Node69-0.00863->t10                                                                                              
#>   [7] -0.419205>Node2-0.021210>Node3-0.294822>Node10-0.225101>Node39-0.016387>Node48-0.004693>Node54-0.000673>Node55-0.017910>t62                                                                                              
#>   [8] -0.419205>Node2-0.021210>Node3-0.294822>Node10-0.225101>Node39-0.016387>Node48-0.004693>Node54-0.000673>Node55-0.017910>t51                                                                                              
#>   [9] --0.4192->Node2--0.0212->Node3--0.2948->Node10--0.2251->Node39--0.0397->t89                                                                                                                                              
#>  [10] -0.41920->Node2-0.02121->Node3-0.29482->Node10-0.23211->Node42-0.00522->Node44-0.02744->t73                                                                                                                              
#>  [11] -0.41920->Node2-0.02121->Node3-0.29482->Node10-0.23211->Node42-0.00522->Node44-0.00546->Node50-0.00563->Node58-0.01634->t83                                                                                              
#>  [12] -0.41920->Node2-0.02121->Node3-0.29482->Node10-0.23211->Node42-0.00522->Node44-0.00546->Node50-0.00563->Node58-0.01634->t70                                                                                              
#>  [13] -0.41920->Node2-0.02121->Node3-0.29482->Node10-0.23211->Node42-0.00522->Node44-0.00546->Node50-0.01410->Node71-0.00787->t30                                                                                              
#>  [14] -0.41920->Node2-0.02121->Node3-0.29482->Node10-0.23211->Node42-0.00522->Node44-0.00546->Node50-0.01410->Node71-0.00787->t14                                                                                              
#>  [15] -0.419205>Node2-0.021210>Node3-0.294822>Node10-0.232111>Node42-0.024387>Node70-0.000508>Node72-0.007757>t54                                                                                                              
#>  [16] -0.419205>Node2-0.021210>Node3-0.294822>Node10-0.232111>Node42-0.024387>Node70-0.000508>Node72-0.006547>Node94-0.001210>t36                                                                                              
#>  [17] -0.419205>Node2-0.021210>Node3-0.294822>Node10-0.232111>Node42-0.024387>Node70-0.000508>Node72-0.006547>Node94-0.001210>t78                                                                                              
#>  [18] -0.41920->Node2-0.02121->Node3-0.29482->Node10-0.23211->Node42-0.02439->Node70-0.00827->t41                                                                                                                              
#>  [19] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.13350->Node9-0.22184->Node14-0.07325->Node27-0.00821->Node36-0.02717->Node59-0.01136->Node82-0.00262->Node89-0.00206->t45                                                
#>  [20] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.13350->Node9-0.22184->Node14-0.07325->Node27-0.00821->Node36-0.02717->Node59-0.01136->Node82-0.00262->Node89-0.00206->t48                                                
#>  [21] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.13350->Node9-0.22184->Node14-0.07325->Node27-0.00821->Node36-0.02717->Node59-0.01136->Node82-0.00468->t16                                                                
#>  [22] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.221843>Node14-0.073245>Node27-0.008214>Node36-0.027165>Node59-0.000707>Node60-0.015327>t43                                                                
#>  [23] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.221843>Node14-0.073245>Node27-0.008214>Node36-0.027165>Node59-0.000707>Node60-0.009192>Node77-0.006135>t71                                                
#>  [24] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.221843>Node14-0.073245>Node27-0.008214>Node36-0.027165>Node59-0.000707>Node60-0.009192>Node77-0.006135>t77                                                
#>  [25] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.13350->Node9-0.22184->Node14-0.07325->Node27-0.00821->Node36-0.04320->t85                                                                                                
#>  [26] --0.4192->Node2--0.0212->Node3--0.0796->Node5--0.1335->Node9--0.2218->Node14--0.0732->Node27--0.0514->t9                                                                                                                 
#>  [27] --0.4192->Node2--0.0212->Node3--0.0796->Node5--0.1335->Node9--0.2218->Node14--0.0643->Node24--0.0170->Node35--0.0434->t72                                                                                                
#>  [28] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.13350->Node9-0.22184->Node14-0.06433->Node24-0.01697->Node35-0.01233->Node43-0.00986->Node51-0.02117->t75                                                                
#>  [29] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.13350->Node9-0.22184->Node14-0.06433->Node24-0.01697->Node35-0.01233->Node43-0.00986->Node51-0.00124->Node52-0.01993->t47                                                
#>  [30] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.221843>Node14-0.064329>Node24-0.016971>Node35-0.012332>Node43-0.009857>Node51-0.001237>Node52-0.000525>Node53-0.018169>Node93-0.001239>t84                
#>  [31] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.221843>Node14-0.064329>Node24-0.016971>Node35-0.012332>Node43-0.009857>Node51-0.001237>Node52-0.000525>Node53-0.018169>Node93-0.001239>t46                
#>  [32] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.221843>Node14-0.064329>Node24-0.016971>Node35-0.012332>Node43-0.009857>Node51-0.001237>Node52-0.000525>Node53-0.019407>t86                                
#>  [33] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.13350->Node9-0.22184->Node14-0.06433->Node24-0.01697->Node35-0.01233->Node43-0.02482->Node76-0.00621->t49                                                                
#>  [34] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.13350->Node9-0.22184->Node14-0.06433->Node24-0.01697->Node35-0.01233->Node43-0.02482->Node76-0.00621->t68                                                                
#>  [35] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.13350->Node9-0.22184->Node14-0.06433->Node24-0.00657->Node26-0.00259->Node29-0.05117->t100                                                                               
#>  [36] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.221843>Node14-0.064329>Node24-0.006571>Node26-0.002586>Node29-0.000631>Node30-0.038309>Node63-0.012232>t57                                                
#>  [37] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.221843>Node14-0.064329>Node24-0.006571>Node26-0.002586>Node29-0.000631>Node30-0.038309>Node63-0.010437>Node91-0.001794>t22                                
#>  [38] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.221843>Node14-0.064329>Node24-0.006571>Node26-0.002586>Node29-0.000631>Node30-0.038309>Node63-0.010437>Node91-0.001794>t18                                
#>  [39] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.221843>Node14-0.064329>Node24-0.006571>Node26-0.002586>Node29-0.000631>Node30-0.006403>Node34-0.034040>Node66-0.001275>Node68-0.002050>Node75-0.006774>t53
#>  [40] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.221843>Node14-0.064329>Node24-0.006571>Node26-0.002586>Node29-0.000631>Node30-0.006403>Node34-0.034040>Node66-0.001275>Node68-0.002050>Node75-0.006774>t23
#>  [41] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.221843>Node14-0.064329>Node24-0.006571>Node26-0.002586>Node29-0.000631>Node30-0.006403>Node34-0.034040>Node66-0.001275>Node68-0.008823>t1                 
#>  [42] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.221843>Node14-0.064329>Node24-0.006571>Node26-0.002586>Node29-0.000631>Node30-0.006403>Node34-0.034040>Node66-0.010098>t13                                
#>  [43] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.221843>Node14-0.064329>Node24-0.006571>Node26-0.002586>Node29-0.000631>Node30-0.006403>Node34-0.044138>t38                                                
#>  [44] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.13350->Node9-0.22184->Node14-0.06433->Node24-0.00657->Node26-0.00692->Node32-0.04684->t61                                                                                
#>  [45] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.221843>Node14-0.064329>Node24-0.006571>Node26-0.006916>Node32-0.044794>Node90-0.000945>Node96-0.001104>t39                                                
#>  [46] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.221843>Node14-0.064329>Node24-0.006571>Node26-0.006916>Node32-0.044794>Node90-0.000945>Node96-0.001104>t65                                                
#>  [47] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.13350->Node9-0.22184->Node14-0.06433->Node24-0.00657->Node26-0.00692->Node32-0.04479->Node90-0.00205->t17                                                                
#>  [48] --0.4192->Node2--0.0212->Node3--0.0796->Node5--0.1335->Node9--0.1393->Node13--0.1249->Node20--0.0227->Node25--0.0335->Node47--0.0134->Node62--0.0127->t6                                                                 
#>  [49] --0.4192->Node2--0.0212->Node3--0.0796->Node5--0.1335->Node9--0.1393->Node13--0.1249->Node20--0.0227->Node25--0.0335->Node47--0.0134->Node62--0.0127->t21                                                                
#>  [50] --0.4192->Node2--0.0212->Node3--0.0796->Node5--0.1335->Node9--0.1393->Node13--0.1249->Node20--0.0227->Node25--0.0335->Node47--0.0145->Node64--0.0116->t4                                                                 
#>  [51] --0.4192->Node2--0.0212->Node3--0.0796->Node5--0.1335->Node9--0.1393->Node13--0.1249->Node20--0.0227->Node25--0.0335->Node47--0.0145->Node64--0.0116->t94                                                                
#>  [52] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.13350->Node9-0.13928->Node13-0.12489->Node20-0.02273->Node25-0.00834->Node28-0.03366->Node56-0.01759->t88                                                                
#>  [53] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.139285>Node13-0.124892>Node20-0.022735>Node25-0.008339>Node28-0.033660>Node56-0.000357>Node57-0.011650>Node78-0.005583>t90                                
#>  [54] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.139285>Node13-0.124892>Node20-0.022735>Node25-0.008339>Node28-0.033660>Node56-0.000357>Node57-0.011650>Node78-0.005583>t52                                
#>  [55] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.133505>Node9-0.139285>Node13-0.124892>Node20-0.022735>Node25-0.008339>Node28-0.033660>Node56-0.000357>Node57-0.017233>t42                                                
#>  [56] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.13350->Node9-0.13928->Node13-0.12489->Node20-0.02273->Node25-0.00834->Node28-0.05049->Node98-0.00076->t24                                                                
#>  [57] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.13350->Node9-0.13928->Node13-0.12489->Node20-0.02273->Node25-0.00834->Node28-0.05049->Node98-0.00076->t26                                                                
#>  [58] --0.4192->Node2--0.0212->Node3--0.0796->Node5--0.1335->Node9--0.1393->Node13--0.1249->Node20--0.0474->Node41--0.0349->t93                                                                                                
#>  [59] --0.4192->Node2--0.0212->Node3--0.0796->Node5--0.1335->Node9--0.1393->Node13--0.1249->Node20--0.0474->Node41--0.0349->t96                                                                                                
#>  [60] --0.4192->Node2--0.0212->Node3--0.0796->Node5--0.1335->Node9--0.1393->Node13--0.1280->Node21--0.0429->Node40--0.0363->t63                                                                                                
#>  [61] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.13350->Node9-0.13928->Node13-0.12800->Node21-0.04290->Node40-0.03526->Node97-0.00106->t31                                                                                
#>  [62] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.13350->Node9-0.13928->Node13-0.12800->Node21-0.04290->Node40-0.03526->Node97-0.00106->t25                                                                                
#>  [63] --0.4192->Node2--0.0212->Node3--0.0796->Node5--0.1335->Node9--0.1393->Node13--0.1280->Node21--0.0792->t33                                                                                                                
#>  [64] --0.4192->Node2--0.0212->Node3--0.0796->Node5--0.1289->Node8--0.1080->Node11--0.2431->t95                                                                                                                                
#>  [65] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.12891->Node8-0.10800->Node11-0.14092->Node16-0.00706->Node18-0.08811->Node74-0.00701->t37                                                                                
#>  [66] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.12891->Node8-0.10800->Node11-0.14092->Node16-0.00706->Node18-0.08811->Node74-0.00465->Node88-0.00236->t7                                                                 
#>  [67] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.12891->Node8-0.10800->Node11-0.14092->Node16-0.00706->Node18-0.08811->Node74-0.00465->Node88-0.00236->t81                                                                
#>  [68] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.12891->Node8-0.10800->Node11-0.14092->Node16-0.00706->Node18-0.05287->Node38-0.01552->Node45-0.02672->t66                                                                
#>  [69] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.12891->Node8-0.10800->Node11-0.14092->Node16-0.00706->Node18-0.05287->Node38-0.01552->Node45-0.02672->t44                                                                
#>  [70] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.128911>Node8-0.108001>Node11-0.140918>Node16-0.007057>Node18-0.052873>Node38-0.041769>Node99-0.000478>t97                                                                
#>  [71] -0.419205>Node2-0.021210>Node3-0.079579>Node5-0.128911>Node8-0.108001>Node11-0.140918>Node16-0.007057>Node18-0.052873>Node38-0.041769>Node99-0.000478>t60                                                                
#>  [72] --0.4192->Node2--0.0212->Node3--0.0796->Node5--0.1289->Node8--0.1080->Node11--0.1409->Node16--0.1022->t28                                                                                                                
#>  [73] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.12891->Node8-0.24866->Node15-0.09494->Node73-0.00352->Node83-0.00397->t58                                                                                                
#>  [74] -0.41920->Node2-0.02121->Node3-0.07958->Node5-0.12891->Node8-0.24866->Node15-0.09494->Node73-0.00352->Node83-0.00397->t79                                                                                                
#>  [75] --0.4192->Node2--0.0212->Node3--0.0796->Node5--0.1289->Node8--0.2487->Node15--0.0949->Node73--0.0075->t29                                                                                                                
#>  [76] --0.4192->Node2--0.0212->Node3--0.0796->Node5--0.1289->Node8--0.2487->Node15--0.1024->t74                                                                                                                                
#>  [77] -0.41920->Node2-0.08556->Node4-0.39842->Node17-0.02066->Node23-0.03062->Node33-0.00324->Node37-0.03724->Node80-0.00125->Node84-0.00381->t40                                                                              
#>  [78] -0.41920->Node2-0.08556->Node4-0.39842->Node17-0.02066->Node23-0.03062->Node33-0.00324->Node37-0.03724->Node80-0.00125->Node84-0.00381->t8                                                                               
#>  [79] -0.419205>Node2-0.085561>Node4-0.398415>Node17-0.020664>Node23-0.030620>Node33-0.003244>Node37-0.037236>Node80-0.000235>Node81-0.004821>t67                                                                              
#>  [80] -0.419205>Node2-0.085561>Node4-0.398415>Node17-0.020664>Node23-0.030620>Node33-0.003244>Node37-0.037236>Node80-0.000235>Node81-0.004821>t87                                                                              
#>  [81] -0.41920->Node2-0.08556->Node4-0.39842->Node17-0.02066->Node23-0.03062->Node33-0.00324->Node37-0.04229->t15                                                                                                              
#>  [82] -0.41920->Node2-0.08556->Node4-0.39842->Node17-0.02066->Node23-0.03062->Node33-0.04007->Node79-0.00547->t76                                                                                                              
#>  [83] -0.41920->Node2-0.08556->Node4-0.39842->Node17-0.02066->Node23-0.03062->Node33-0.04007->Node79-0.00547->t20                                                                                                              
#>  [84] --0.4192->Node2--0.0856->Node4--0.3984->Node17--0.0207->Node23--0.0762->t69                                                                                                                                              
#>  [85] --0.4192->Node2--0.0856->Node4--0.3984->Node17--0.0184->Node22--0.0785->t80                                                                                                                                              
#>  [86] --0.4192->Node2--0.0856->Node4--0.3984->Node17--0.0184->Node22--0.0524->Node46--0.0261->t56                                                                                                                              
#>  [87] --0.4192->Node2--0.0856->Node4--0.3984->Node17--0.0184->Node22--0.0524->Node46--0.0261->t91                                                                                                                              
#>  [88] --0.4192->Node2--0.0856->Node4--0.0173->Node6--0.4278->Node31--0.0502->t50                                                                                                                                               
#>  [89] -0.41920->Node2-0.08556->Node4-0.01725->Node6-0.42778->Node31-0.04908->Node95-0.00112->t34                                                                                                                               
#>  [90] -0.41920->Node2-0.08556->Node4-0.01725->Node6-0.42778->Node31-0.04908->Node95-0.00112->t55                                                                                                                               
#>  [91] --0.4192->Node2--0.0856->Node4--0.0173->Node6--0.0714->Node7--0.4066->t64                                                                                                                                                
#>  [92] -0.41920->Node2-0.08556->Node4-0.01725->Node6-0.07139->Node7-0.40293->Node85-0.00366->t99                                                                                                                                
#>  [93] -0.41920->Node2-0.08556->Node4-0.01725->Node6-0.07139->Node7-0.40293->Node85-0.00366->t2                                                                                                                                 
#>  [94] --0.761-->Node12--0.226-->Node61--0.013-->t35                                                                                                                                                                            
#>  [95] -0.76095->Node12-0.22608->Node61-0.00217->Node65-0.00114->Node67-0.00966->t59                                                                                                                                            
#>  [96] -0.76095->Node12-0.22608->Node61-0.00217->Node65-0.00114->Node67-0.00611->Node86-0.00356->t27                                                                                                                            
#>  [97] -0.76095->Node12-0.22608->Node61-0.00217->Node65-0.00114->Node67-0.00611->Node86-0.00356->t12                                                                                                                            
#>  [98] -0.76095->Node12-0.22608->Node61-0.00217->Node65-0.01080->t98                                                                                                                                                            
#>  [99] --0.7609->Node12--0.1514->Node19--0.0876->t5                                                                                                                                                                             
#> [100] --0.7609->Node12--0.1514->Node19--0.0876->t32                                                                                                                                                                            
#> [101] ----1---->Node2                                                                                                                                                                                                          
#> [102] --0.9518->Node2--0.0482->Node3                                                                                                                                                                                           
#> [103] ---0.83-->Node2---0.17-->Node4                                                                                                                                                                                           
#> [104] --0.8062->Node2--0.0408->Node3--0.1530->Node5                                                                                                                                                                            
#> [105] --0.8030->Node2--0.1639->Node4--0.0331->Node6                                                                                                                                                                            
#> [106] --0.7064->Node2--0.1442->Node4--0.0291->Node6--0.1203->Node7                                                                                                                                                             
#> [107] --0.6460->Node2--0.0327->Node3--0.1226->Node5--0.1987->Node8                                                                                                                                                             
#> [108] --0.6415->Node2--0.0325->Node3--0.1218->Node5--0.2043->Node9                                                                                                                                                             
#> [109] --0.5702->Node2--0.0288->Node3--0.4010->Node10                                                                                                                                                                           
#> [110] --0.554-->Node2--0.028-->Node3--0.105-->Node5--0.170-->Node8--0.143-->Node11                                                                                                                                             
#> [111] ----1---->Node12                                                                                                                                                                                                         
#> [112] --0.5288->Node2--0.0268->Node3--0.1004->Node5--0.1684->Node9--0.1757->Node13                                                                                                                                             
#> [113] --0.4789->Node2--0.0242->Node3--0.0909->Node5--0.1525->Node9--0.2534->Node14                                                                                                                                             
#> [114] --0.4670->Node2--0.0236->Node3--0.0887->Node5--0.1436->Node8--0.2770->Node15                                                                                                                                             
#> [115] --0.4669->Node2--0.0236->Node3--0.0886->Node5--0.1436->Node8--0.1203->Node11--0.1570->Node16                                                                                                                             
#> [116] --0.4641->Node2--0.0947->Node4--0.4411->Node17                                                                                                                                                                           
#> [117] --0.4633->Node2--0.0234->Node3--0.0879->Node5--0.1425->Node8--0.1194->Node11--0.1557->Node16--0.0078->Node18                                                                                                             
#> [118] --0.834-->Node12--0.166-->Node19                                                                                                                                                                                         
#> [119] --0.4568->Node2--0.0231->Node3--0.0867->Node5--0.1455->Node9--0.1518->Node13--0.1361->Node20                                                                                                                             
#> [120] --0.4553->Node2--0.0230->Node3--0.0864->Node5--0.1450->Node9--0.1513->Node13--0.1390->Node21                                                                                                                             
#> [121] --0.4549->Node2--0.0928->Node4--0.4323->Node17--0.0199->Node22                                                                                                                                                           
#> [122] --0.4538->Node2--0.0926->Node4--0.4313->Node17--0.0224->Node23                                                                                                                                                           
#> [123] --0.4461->Node2--0.0226->Node3--0.0847->Node5--0.1421->Node9--0.2361->Node14--0.0685->Node24                                                                                                                             
#> [124] --0.4458->Node2--0.0226->Node3--0.0846->Node5--0.1420->Node9--0.1481->Node13--0.1328->Node20--0.0242->Node25                                                                                                             
#> [125] -0.44302->Node2-0.02241->Node3-0.08410->Node5-0.14109->Node9-0.23445->Node14-0.06798->Node24-0.00694->Node26                                                                                                             
#> [126] --0.4419->Node2--0.0224->Node3--0.0839->Node5--0.1407->Node9--0.2339->Node14--0.0772->Node27                                                                                                                             
#> [127] -0.44185->Node2-0.02236->Node3-0.08388->Node5-0.14072->Node9-0.14681->Node13-0.13164->Node20-0.02396->Node25-0.00879->Node28                                                                                             
#> [128] -0.44181->Node2-0.02235->Node3-0.08387->Node5-0.14071->Node9-0.23381->Node14-0.06780->Node24-0.00693->Node26-0.00273->Node29                                                                                             
#> [129] -0.441520>Node2-0.022339>Node3-0.083815>Node5-0.140612>Node9-0.233652>Node14-0.067754>Node24-0.006921>Node26-0.002724>Node29-0.000664>Node30                                                                             
#> [130] --0.4414->Node2--0.0901->Node4--0.0182->Node6--0.4504->Node31                                                                                                                                                            
#> [131] -0.43981->Node2-0.02225->Node3-0.08349->Node5-0.14007->Node9-0.23275->Node14-0.06749->Node24-0.00689->Node26-0.00726->Node32                                                                                             
#> [132] --0.4392->Node2--0.0896->Node4--0.4174->Node17--0.0216->Node23--0.0321->Node33                                                                                                                                           
#> [133] -0.43856->Node2-0.02219->Node3-0.08325->Node5-0.13967->Node9-0.23209->Node14-0.06730->Node24-0.00687->Node26-0.00271->Node29-0.00066->Node30-0.00670->Node34                                                             
#> [134] --0.4382->Node2--0.0222->Node3--0.0832->Node5--0.1396->Node9--0.2319->Node14--0.0672->Node24--0.0177->Node35                                                                                                             
#> [135] -0.43813->Node2-0.02217->Node3-0.08317->Node5-0.13953->Node9-0.23186->Node14-0.07655->Node27-0.00858->Node36                                                                                                             
#> [136] -0.43772->Node2-0.08934->Node4-0.41601->Node17-0.02158->Node23-0.03197->Node33-0.00339->Node37                                                                                                                           
#> [137] -0.43770->Node2-0.02215->Node3-0.08309->Node5-0.13460->Node8-0.11277->Node11-0.14713->Node16-0.00737->Node18-0.05521->Node38                                                                                             
#> [138] --0.4365->Node2--0.0221->Node3--0.3070->Node10--0.2344->Node39                                                                                                                                                           
#> [139] --0.4350->Node2--0.0220->Node3--0.0826->Node5--0.1385->Node9--0.1445->Node13--0.1328->Node21--0.0445->Node40                                                                                                             
#> [140] --0.4344->Node2--0.0220->Node3--0.0825->Node5--0.1383->Node9--0.1443->Node13--0.1294->Node20--0.0491->Node41                                                                                                             
#> [141] --0.4334->Node2--0.0219->Node3--0.3048->Node10--0.2399->Node42                                                                                                                                                           
#> [142] --0.4326->Node2--0.0219->Node3--0.0821->Node5--0.1378->Node9--0.2289->Node14--0.0664->Node24--0.0175->Node35--0.0127->Node43                                                                                             
#> [143] -0.43103->Node2-0.02181->Node3-0.30314->Node10-0.23866->Node42-0.00536->Node44                                                                                                                                           
#> [144] -0.43071->Node2-0.02179->Node3-0.08176->Node5-0.13245->Node8-0.11097->Node11-0.14479->Node16-0.00725->Node18-0.05433->Node38-0.01595->Node45                                                                             
#> [145] --0.4304->Node2--0.0879->Node4--0.4091->Node17--0.0188->Node22--0.0538->Node46                                                                                                                                           
#> [146] --0.4304->Node2--0.0218->Node3--0.0817->Node5--0.1371->Node9--0.1430->Node13--0.1282->Node20--0.0233->Node25--0.0344->Node47                                                                                             
#> [147] --0.4292->Node2--0.0217->Node3--0.3018->Node10--0.2305->Node39--0.0168->Node48                                                                                                                                           
#> [148] -4.29e-01>Node2-2.17e-02>Node3-3.02e-01>Node10-2.30e-01>Node39-1.68e-02>Node48-7.32e-05>Node49                                                                                                                           
#> [149] -0.42862->Node2-0.02169->Node3-0.30145->Node10-0.23733->Node42-0.00533->Node44-0.00559->Node50                                                                                                                           
#> [150] --0.4283->Node2--0.0217->Node3--0.0813->Node5--0.1364->Node9--0.2266->Node14--0.0657->Node24--0.0173->Node35--0.0126->Node43--0.0101->Node51                                                                             
#> [151] -0.42773->Node2-0.02164->Node3-0.08120->Node5-0.13622->Node9-0.22636->Node14-0.06564->Node24-0.01732->Node35-0.01258->Node43-0.01006->Node51-0.00126->Node52                                                             
#> [152] -0.427502>Node2-0.021629>Node3-0.081154>Node5-0.136147>Node9-0.226234>Node14-0.065602>Node24-0.017307>Node35-0.012576>Node43-0.010052>Node51-0.001261>Node52-0.000535>Node53                                             
#> [153] -0.42714->Node2-0.02161->Node3-0.30040->Node10-0.22936->Node39-0.01670->Node48-0.00478->Node54                                                                                                                           
#> [154] -0.426850>Node2-0.021596>Node3-0.300199>Node10-0.229206>Node39-0.016686>Node48-0.004778>Node54-0.000685>Node55                                                                                                           
#> [155] -0.42671->Node2-0.02159->Node3-0.08100->Node5-0.13590->Node9-0.14178->Node13-0.12713->Node20-0.02314->Node25-0.00849->Node28-0.03426->Node56                                                                             
#> [156] -0.426556>Node2-0.021582>Node3-0.080974>Node5-0.135846>Node9-0.141727>Node13-0.127082>Node20-0.023134>Node25-0.008486>Node28-0.034251>Node56-0.000363>Node57                                                             
#> [157] -0.42617->Node2-0.02156->Node3-0.29972->Node10-0.23597->Node42-0.00530->Node44-0.00555->Node50-0.00573->Node58                                                                                                           
#> [158] -0.42604->Node2-0.02156->Node3-0.08088->Node5-0.13568->Node9-0.22546->Node14-0.07444->Node27-0.00835->Node36-0.02761->Node59                                                                                             
#> [159] -0.425730>Node2-0.021540>Node3-0.080817>Node5-0.135583>Node9-0.225296>Node14-0.074386>Node27-0.008342>Node36-0.027588>Node59-0.000718>Node60                                                                             
#> [160] --0.771-->Node12--0.229-->Node61                                                                                                                                                                                         
#> [161] --0.4246->Node2--0.0215->Node3--0.0806->Node5--0.1352->Node9--0.1411->Node13--0.1265->Node20--0.0230->Node25--0.0339->Node47--0.0136->Node62                                                                             
#> [162] -0.424396>Node2-0.021472>Node3-0.080564>Node5-0.135158>Node9-0.224590>Node14-0.065126>Node24-0.006652>Node26-0.002618>Node29-0.000639>Node30-0.038784>Node63                                                             
#> [163] --0.4241->Node2--0.0215->Node3--0.0805->Node5--0.1351->Node9--0.1409->Node13--0.1264->Node20--0.0230->Node25--0.0339->Node47--0.0146->Node64                                                                             
#> [164] --0.7693->Node12--0.2285->Node61--0.0022->Node65                                                                                                                                                                         
#> [165] -0.423481>Node2-0.021426>Node3-0.080391>Node5-0.134867>Node9-0.224106>Node14-0.064986>Node24-0.006638>Node26-0.002613>Node29-0.000637>Node30-0.006468>Node34-0.034387>Node66                                             
#> [166] -0.76837->Node12-0.22828->Node61-0.00219->Node65-0.00115->Node67                                                                                                                                                         
#> [167] -0.422936>Node2-0.021398>Node3-0.080287>Node5-0.134693>Node9-0.223818>Node14-0.064902>Node24-0.006629>Node26-0.002609>Node29-0.000636>Node30-0.006460>Node34-0.034343>Node66-0.001286>Node68                             
#> [168] -0.42285->Node2-0.02139->Node3-0.29739->Node10-0.22706->Node39-0.01653->Node48-0.00473->Node54-0.01004->Node69                                                                                                           
#> [169] --0.4227->Node2--0.0214->Node3--0.2973->Node10--0.2340->Node42--0.0246->Node70                                                                                                                                           
#> [170] -0.42253->Node2-0.02138->Node3-0.29716->Node10-0.23395->Node42-0.00526->Node44-0.00551->Node50-0.01422->Node71                                                                                                           
#> [171] -0.422482>Node2-0.021375>Node3-0.297127>Node10-0.233926>Node42-0.024578>Node70-0.000512>Node72                                                                                                                           
#> [172] --0.4224->Node2--0.0214->Node3--0.0802->Node5--0.1299->Node8--0.2505->Node15--0.0957->Node73                                                                                                                             
#> [173] -0.42216->Node2-0.02136->Node3-0.08014->Node5-0.12982->Node8-0.10876->Node11-0.14191->Node16-0.00711->Node18-0.08873->Node74                                                                                             
#> [174] -0.422064>Node2-0.021354>Node3-0.080121>Node5-0.134415>Node9-0.223356>Node14-0.064768>Node24-0.006616>Node26-0.002604>Node29-0.000635>Node30-0.006447>Node34-0.034272>Node66-0.001283>Node68-0.002064>Node75             
#> [175] --0.4218->Node2--0.0213->Node3--0.0801->Node5--0.1343->Node9--0.2232->Node14--0.0647->Node24--0.0171->Node35--0.0124->Node43--0.0250->Node76                                                                             
#> [176] -0.421793>Node2-0.021341>Node3-0.080070>Node5-0.134329>Node9-0.223213>Node14-0.073698>Node27-0.008265>Node36-0.027333>Node59-0.000712>Node60-0.009248>Node77                                                             
#> [177] -0.421558>Node2-0.021329>Node3-0.080026>Node5-0.134254>Node9-0.140067>Node13-0.125593>Node20-0.022863>Node25-0.008386>Node28-0.033849>Node56-0.000359>Node57-0.011716>Node78                                             
#> [178] --0.4215->Node2--0.0860->Node4--0.4006->Node17--0.0208->Node23--0.0308->Node33--0.0403->Node79                                                                                                                           
#> [179] -0.42133->Node2-0.08600->Node4-0.40044->Node17-0.02077->Node23-0.03078->Node33-0.00326->Node37-0.03742->Node80                                                                                                           
#> [180] -0.421235>Node2-0.085975>Node4-0.400345>Node17-0.020764>Node23-0.030769>Node33-0.003260>Node37-0.037416>Node80-0.000236>Node81                                                                                           
#> [181] -0.42117->Node2-0.02131->Node3-0.07995->Node5-0.13413->Node9-0.22289->Node14-0.07359->Node27-0.00825->Node36-0.02729->Node59-0.01141->Node82                                                                             
#> [182] -0.42088->Node2-0.02129->Node3-0.07990->Node5-0.12942->Node8-0.24965->Node15-0.09532->Node73-0.00354->Node83                                                                                                             
#> [183] -0.42081->Node2-0.08589->Node4-0.39994->Node17-0.02074->Node23-0.03074->Node33-0.00326->Node37-0.03738->Node80-0.00125->Node84                                                                                           
#> [184] --0.4207->Node2--0.0859->Node4--0.0173->Node6--0.0717->Node7--0.4044->Node85                                                                                                                                             
#> [185] -0.76366->Node12-0.22688->Node61-0.00218->Node65-0.00114->Node67-0.00613->Node86                                                                                                                                         
#> [186] -0.42070->Node2-0.02129->Node3-0.29587->Node10-0.22590->Node39-0.01645->Node48-0.00471->Node54-0.00999->Node69-0.00510->Node87                                                                                           
#> [187] -0.42020->Node2-0.02126->Node3-0.07977->Node5-0.12922->Node8-0.10826->Node11-0.14125->Node16-0.00707->Node18-0.08832->Node74-0.00466->Node88                                                                             
#> [188] -0.42007->Node2-0.02125->Node3-0.07974->Node5-0.13378->Node9-0.22230->Node14-0.07340->Node27-0.00823->Node36-0.02722->Node59-0.01138->Node82-0.00263->Node89                                                             
#> [189] -0.42007->Node2-0.02125->Node3-0.07974->Node5-0.13378->Node9-0.22230->Node14-0.06446->Node24-0.00658->Node26-0.00693->Node32-0.04489->Node90                                                                             
#> [190] -0.419958>Node2-0.021248>Node3-0.079722>Node5-0.133745>Node9-0.222242>Node14-0.064445>Node24-0.006583>Node26-0.002591>Node29-0.000632>Node30-0.038378>Node63-0.010456>Node91                                             
#> [191] -0.41994->Node2-0.02125->Node3-0.29534->Node10-0.22549->Node39-0.01642->Node48-0.00470->Node54-0.00997->Node69-0.00509->Node87-0.00181->Node92                                                                           
#> [192] -0.419725>Node2-0.021236>Node3-0.079677>Node5-0.133671>Node9-0.222118>Node14-0.064409>Node24-0.016992>Node35-0.012347>Node43-0.009869>Node51-0.001238>Node52-0.000526>Node53-0.018191>Node93                             
#> [193] -0.419713>Node2-0.021235>Node3-0.295179>Node10-0.232393>Node42-0.024416>Node70-0.000509>Node72-0.006555>Node94                                                                                                           
#> [194] --0.4197->Node2--0.0857->Node4--0.0173->Node6--0.4283->Node31--0.0491->Node95                                                                                                                                            
#> [195] -0.419668>Node2-0.021233>Node3-0.079667>Node5-0.133652>Node9-0.222088>Node14-0.064400>Node24-0.006578>Node26-0.006923>Node32-0.044843>Node90-0.000946>Node96                                                             
#> [196] --0.4196->Node2--0.0212->Node3--0.0797->Node5--0.1336->Node9--0.1394->Node13--0.1281->Node21--0.0429->Node40--0.0353->Node97                                                                                             
#> [197] -0.41952->Node2-0.02123->Node3-0.07964->Node5-0.13361->Node9-0.13939->Node13-0.12499->Node20-0.02275->Node25-0.00835->Node28-0.05053->Node98                                                                             
#> [198] -0.41941->Node2-0.02122->Node3-0.07962->Node5-0.12897->Node8-0.10805->Node11-0.14099->Node16-0.00706->Node18-0.05290->Node38-0.04179->Node99