Scale the phylogenetic flow features to a constant sum.

Can be used to standardise branch lengths to reasonable values

Usage

pf_scale_flow_sum(x, scale_to = 1)

Arguments

x

a pfc object

scale_to

The value to scale the sums to. Default is 1.

Value

a scaled pfc object

Examples

pf_scale_flow_sum(rpfc(100))
#> <pfc<e:198>[198]>
#>   [1] -0.91483->Node6-0.05478->Node19-0.02776->Node71-0.00264->t79                                                                                                                              
#>   [2] -0.914828>Node6-0.054776>Node19-0.027757>Node71-0.002007>Node90-0.000632>t49                                                                                                              
#>   [3] -0.914828>Node6-0.054776>Node19-0.027757>Node71-0.002007>Node90-0.000632>t14                                                                                                              
#>   [4] --0.9148->Node6--0.0548->Node19--0.0194->Node36--0.0110->t51                                                                                                                              
#>   [5] --0.9148->Node6--0.0548->Node19--0.0194->Node36--0.0110->t56                                                                                                                              
#>   [6] -9.15e-01>Node6-2.54e-02>Node8-7.13e-03>Node9-9.69e-03>Node11-2.95e-03>Node12-3.39e-02>Node50-3.27e-03>Node69-2.57e-03>Node94-4.25e-05>Node95-2.62e-04>t37                                
#>   [7] -9.15e-01>Node6-2.54e-02>Node8-7.13e-03>Node9-9.69e-03>Node11-2.95e-03>Node12-3.39e-02>Node50-3.27e-03>Node69-2.57e-03>Node94-4.25e-05>Node95-2.62e-04>t84                                
#>   [8] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.009689>Node11-0.002946>Node12-0.033867>Node50-0.003269>Node69-0.002571>Node94-0.000304>t46                                                
#>   [9] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.00969->Node11-0.00295->Node12-0.03387->Node50-0.00327->Node69-0.00288->t15                                                                
#>  [10] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.00969->Node11-0.00295->Node12-0.03387->Node50-0.00614->t43                                                                                
#>  [11] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.00969->Node11-0.00295->Node12-0.01256->Node20-0.00428->Node22-0.01471->Node41-0.00846->t28                                                
#>  [12] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.00969->Node11-0.00295->Node12-0.01256->Node20-0.00428->Node22-0.01471->Node41-0.00846->t77                                                
#>  [13] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.00969->Node11-0.00295->Node12-0.01256->Node20-0.00428->Node22-0.01455->Node40-0.00862->t10                                                
#>  [14] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.00969->Node11-0.00295->Node12-0.01256->Node20-0.00428->Node22-0.01455->Node40-0.00360->Node56-0.00501->t73                                
#>  [15] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.00969->Node11-0.00295->Node12-0.01256->Node20-0.00428->Node22-0.01455->Node40-0.00360->Node56-0.00293->Node76-0.00208->t78                
#>  [16] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.00969->Node11-0.00295->Node12-0.01256->Node20-0.00428->Node22-0.01455->Node40-0.00360->Node56-0.00293->Node76-0.00208->t8                 
#>  [17] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.009689>Node11-0.002946>Node12-0.012560>Node20-0.010747>Node25-0.009263>Node45-0.000983>Node47-0.002918>Node63-0.003540>t35                
#>  [18] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.009689>Node11-0.002946>Node12-0.012560>Node20-0.010747>Node25-0.009263>Node45-0.000983>Node47-0.002918>Node63-0.000965>Node72-0.002575>t88
#>  [19] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.009689>Node11-0.002946>Node12-0.012560>Node20-0.010747>Node25-0.009263>Node45-0.000983>Node47-0.002918>Node63-0.000965>Node72-0.002575>t53
#>  [20] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.009689>Node11-0.002946>Node12-0.012560>Node20-0.010747>Node25-0.009263>Node45-0.000983>Node47-0.006458>t90                                
#>  [21] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.009689>Node11-0.002946>Node12-0.012560>Node20-0.010747>Node25-0.009263>Node45-0.006594>Node87-0.000847>t30                                
#>  [22] -9.15e-01>Node6-2.54e-02>Node8-7.13e-03>Node9-9.69e-03>Node11-2.95e-03>Node12-1.26e-02>Node20-1.07e-02>Node25-9.26e-03>Node45-6.59e-03>Node87-8.32e-04>Node98-1.52e-05>t83                
#>  [23] -9.15e-01>Node6-2.54e-02>Node8-7.13e-03>Node9-9.69e-03>Node11-2.95e-03>Node12-1.26e-02>Node20-1.07e-02>Node25-9.26e-03>Node45-6.59e-03>Node87-8.32e-04>Node98-1.52e-05>t54                
#>  [24] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.00969->Node11-0.00295->Node12-0.01256->Node20-0.01075->Node25-0.01670->t76                                                                
#>  [25] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.00969->Node11-0.03207->Node37-0.00551->Node52-0.00537->t64                                                                                
#>  [26] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.00969->Node11-0.03207->Node37-0.00551->Node52-0.00537->t23                                                                                
#>  [27] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.00969->Node11-0.03207->Node37-0.01088->t69                                                                                                
#>  [28] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.023624>Node28-0.001809>Node31-0.011392>Node82-0.001405>t100                                               
#>  [29] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.023624>Node28-0.001809>Node31-0.011392>Node82-0.001405>t11                                                
#>  [30] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.023624>Node28-0.001809>Node31-0.007358>Node51-0.000242>Node54-0.005198>t55                                
#>  [31] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.023624>Node28-0.001809>Node31-0.007358>Node51-0.000242>Node54-0.005198>t29                                
#>  [32] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.023624>Node28-0.001809>Node31-0.007358>Node51-0.001938>Node64-0.003501>t72                                
#>  [33] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.023624>Node28-0.001809>Node31-0.007358>Node51-0.001938>Node64-0.003501>t4                                 
#>  [34] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.023624>Node28-0.001041>Node29-0.013565>t97                                                                
#>  [35] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.023624>Node28-0.001041>Node29-0.002028>Node34-0.005380>Node49-0.006157>t25                                
#>  [36] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.023624>Node28-0.001041>Node29-0.002028>Node34-0.005380>Node49-0.006157>t57                                
#>  [37] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.023624>Node28-0.001041>Node29-0.002028>Node34-0.010033>Node80-0.001505>t87                                
#>  [38] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.023624>Node28-0.001041>Node29-0.002028>Node34-0.010033>Node80-0.001505>t38                                
#>  [39] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.004331>Node17-0.006689>Node21-0.025548>Node79-0.001662>t86                                                
#>  [40] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.004331>Node17-0.006689>Node21-0.025548>Node79-0.000943>Node89-0.000719>t75                                
#>  [41] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.004331>Node17-0.006689>Node21-0.025548>Node79-0.000943>Node89-0.000719>t44                                
#>  [42] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.004331>Node17-0.006689>Node21-0.023845>Node66-0.003365>t19                                                
#>  [43] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.004331>Node17-0.006689>Node21-0.023845>Node66-0.002547>Node88-0.000818>t89                                
#>  [44] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.004331>Node17-0.006689>Node21-0.023845>Node66-0.002547>Node88-0.000818>t12                                
#>  [45] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.004331>Node17-0.020479>Node30-0.002740>Node39-0.006743>Node60-0.000442>Node65-0.003319>Node96-0.000176>t61
#>  [46] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.004331>Node17-0.020479>Node30-0.002740>Node39-0.006743>Node60-0.000442>Node65-0.003319>Node96-0.000176>t70
#>  [47] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.004331>Node17-0.020479>Node30-0.002740>Node39-0.006743>Node60-0.000442>Node65-0.003495>t47                
#>  [48] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.004331>Node17-0.020479>Node30-0.002740>Node39-0.006743>Node60-0.003937>t32                                
#>  [49] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.004331>Node17-0.020479>Node30-0.002740>Node39-0.006562>Node59-0.001677>Node74-0.002441>t3                 
#>  [50] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.004331>Node17-0.020479>Node30-0.002740>Node39-0.006562>Node59-0.001677>Node74-0.002441>t82                
#>  [51] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.004331>Node17-0.020479>Node30-0.002740>Node39-0.006562>Node59-0.004118>t93                                
#>  [52] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.004331>Node17-0.020479>Node30-0.001972>Node35-0.011448>t20                                                
#>  [53] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.004331>Node17-0.020479>Node30-0.001972>Node35-0.003985>Node43-0.007464>t42                                
#>  [54] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.000584>Node14-0.004331>Node17-0.020479>Node30-0.001972>Node35-0.003985>Node43-0.007464>t52                                
#>  [55] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.01383->Node13-0.01887->Node23-0.01634->Node62-0.00360->t66                                                                                
#>  [56] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.01383->Node13-0.01887->Node23-0.01634->Node62-0.00360->t80                                                                                
#>  [57] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.01383->Node13-0.01887->Node23-0.00835->Node33-0.00348->Node42-0.00169->Node48-0.00643->t40                                                
#>  [58] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.01383->Node13-0.01887->Node23-0.00835->Node33-0.00348->Node42-0.00169->Node48-0.00260->Node61-0.00383->t24                                
#>  [59] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.018871>Node23-0.008346>Node33-0.003475>Node42-0.001688>Node48-0.002603>Node61-0.000516>Node67-0.002432>Node86-0.000882>t71
#>  [60] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.018871>Node23-0.008346>Node33-0.003475>Node42-0.001688>Node48-0.002603>Node61-0.000516>Node67-0.002432>Node86-0.000882>t7 
#>  [61] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.018871>Node23-0.008346>Node33-0.003475>Node42-0.001688>Node48-0.002603>Node61-0.000516>Node67-0.000230>Node68-0.003084>t67
#>  [62] -0.914828>Node6-0.025394>Node8-0.007131>Node9-0.013832>Node13-0.018871>Node23-0.008346>Node33-0.003475>Node42-0.001688>Node48-0.002603>Node61-0.000516>Node67-0.000230>Node68-0.003084>t36
#>  [63] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.01383->Node13-0.01887->Node23-0.00835->Node33-0.00348->Node42-0.00812->t33                                                                
#>  [64] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.01383->Node13-0.01887->Node23-0.00835->Node33-0.00648->Node55-0.00382->Node83-0.00130->t48                                                
#>  [65] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.01383->Node13-0.01887->Node23-0.00835->Node33-0.00648->Node55-0.00382->Node83-0.00130->t13                                                
#>  [66] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.01383->Node13-0.01887->Node23-0.00835->Node33-0.00648->Node55-0.00228->Node70-0.00118->Node78-0.00166->t41                                
#>  [67] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.01383->Node13-0.01887->Node23-0.00835->Node33-0.00648->Node55-0.00228->Node70-0.00118->Node78-0.00166->t45                                
#>  [68] -0.91483->Node6-0.02539->Node8-0.00713->Node9-0.01383->Node13-0.01887->Node23-0.00835->Node33-0.00648->Node55-0.00228->Node70-0.00284->t5                                                 
#>  [69] -0.914828>Node6-0.025394>Node8-0.011336>Node10-0.011789>Node16-0.034846>Node77-0.001451>Node92-0.000356>t65                                                                               
#>  [70] -0.914828>Node6-0.025394>Node8-0.011336>Node10-0.011789>Node16-0.034846>Node77-0.001451>Node92-0.000356>t27                                                                               
#>  [71] -0.91483->Node6-0.02539->Node8-0.01134->Node10-0.01179->Node16-0.03485->Node77-0.00181->t1                                                                                                
#>  [72] -0.91483->Node6-0.02539->Node8-0.01134->Node10-0.01179->Node16-0.00420->Node18-0.02034->Node32-0.00716->Node57-0.00495->t92                                                               
#>  [73] -0.91483->Node6-0.02539->Node8-0.01134->Node10-0.01179->Node16-0.00420->Node18-0.02034->Node32-0.00716->Node57-0.00495->t99                                                               
#>  [74] --0.9148->Node6--0.0254->Node8--0.0113->Node10--0.0118->Node16--0.0042->Node18--0.0203->Node32--0.0121->t59                                                                               
#>  [75] --0.9148->Node6--0.0254->Node8--0.0113->Node10--0.0118->Node16--0.0042->Node18--0.0128->Node24--0.0197->t85                                                                               
#>  [76] -0.914828>Node6-0.025394>Node8-0.011336>Node10-0.011789>Node16-0.004199>Node18-0.012762>Node24-0.017169>Node73-0.000266>Node75-0.002256>t91                                               
#>  [77] -0.914828>Node6-0.025394>Node8-0.011336>Node10-0.011789>Node16-0.004199>Node18-0.012762>Node24-0.017169>Node73-0.000266>Node75-0.002256>t21                                               
#>  [78] -9.15e-01>Node6-2.54e-02>Node8-1.13e-02>Node10-1.18e-02>Node16-4.20e-03>Node18-1.28e-02>Node24-1.72e-02>Node73-2.52e-03>Node99-4.94e-06>t50                                               
#>  [79] -9.15e-01>Node6-2.54e-02>Node8-1.13e-02>Node10-1.18e-02>Node16-4.20e-03>Node18-1.28e-02>Node24-1.72e-02>Node73-2.52e-03>Node99-4.94e-06>t94                                               
#>  [80] -0.91483->Node6-0.02539->Node8-0.01134->Node10-0.04315->Node53-0.00529->t16                                                                                                               
#>  [81] -0.91483->Node6-0.02539->Node8-0.01134->Node10-0.04315->Node53-0.00529->t31                                                                                                               
#>  [82] --0.792-->Node2--0.022-->Node3--0.023-->Node4--0.147-->Node26--0.016-->t62                                                                                                                
#>  [83] -0.79247->Node2-0.02200->Node3-0.02298->Node4-0.14657->Node26-0.01474->Node84-0.00125->t26                                                                                                
#>  [84] -0.79247->Node2-0.02200->Node3-0.02298->Node4-0.14657->Node26-0.01474->Node84-0.00125->t18                                                                                                
#>  [85] -0.79247->Node2-0.02200->Node3-0.02298->Node4-0.06570->Node5-0.09209->Node58-0.00477->t34                                                                                                 
#>  [86] -0.79247->Node2-0.02200->Node3-0.02298->Node4-0.06570->Node5-0.09209->Node58-0.00477->t58                                                                                                 
#>  [87] --0.7925->Node2--0.0220->Node3--0.0230->Node4--0.0657->Node5--0.0364->Node7--0.0495->Node38--0.0109->t9                                                                                   
#>  [88] --0.7925->Node2--0.0220->Node3--0.0230->Node4--0.0657->Node5--0.0364->Node7--0.0495->Node38--0.0109->t68                                                                                  
#>  [89] --0.7925->Node2--0.0220->Node3--0.0230->Node4--0.0657->Node5--0.0364->Node7--0.0604->t22                                                                                                  
#>  [90] -0.79247->Node2-0.02200->Node3-0.17077->Node27-0.00823->Node46-0.00654->t2                                                                                                                
#>  [91] -0.79247->Node2-0.02200->Node3-0.17077->Node27-0.00823->Node46-0.00654->t96                                                                                                               
#>  [92] -0.792471>Node2-0.021995>Node3-0.170768>Node27-0.013355>Node81-0.000501>Node85-0.000910>t81                                                                                               
#>  [93] -0.792471>Node2-0.021995>Node3-0.170768>Node27-0.013355>Node81-0.000501>Node85-0.000910>t95                                                                                               
#>  [94] -0.79247->Node2-0.02200->Node3-0.17077->Node27-0.01335->Node81-0.00141->t60                                                                                                               
#>  [95] -0.792471>Node2-0.170593>Node15-0.036469>Node91-0.000468>t63                                                                                                                              
#>  [96] -0.792471>Node2-0.170593>Node15-0.036469>Node91-0.000117>Node93-0.000246>Node97-0.000105>t6                                                                                               
#>  [97] -0.792471>Node2-0.170593>Node15-0.036469>Node91-0.000117>Node93-0.000246>Node97-0.000105>t39                                                                                              
#>  [98] -0.792471>Node2-0.170593>Node15-0.036469>Node91-0.000117>Node93-0.000351>t17                                                                                                              
#>  [99] -0.79247->Node2-0.17059->Node15-0.02948->Node44-0.00746->t74                                                                                                                              
#> [100] -0.79247->Node2-0.17059->Node15-0.02948->Node44-0.00746->t98                                                                                                                              
#> [101] ----1---->Node2                                                                                                                                                                           
#> [102] --0.973-->Node2--0.027-->Node3                                                                                                                                                            
#> [103] --0.9463->Node2--0.0263->Node3--0.0274->Node4                                                                                                                                             
#> [104] --0.8775->Node2--0.0244->Node3--0.0254->Node4--0.0727->Node5                                                                                                                              
#> [105] ----1---->Node6                                                                                                                                                                           
#> [106] --0.8434->Node2--0.0234->Node3--0.0245->Node4--0.0699->Node5--0.0388->Node7                                                                                                               
#> [107] --0.973-->Node6--0.027-->Node8                                                                                                                                                            
#> [108] -0.96567->Node6-0.02681->Node8-0.00753->Node9                                                                                                                                             
#> [109] --0.9614->Node6--0.0267->Node8--0.0119->Node10                                                                                                                                            
#> [110] -0.95589->Node6-0.02653->Node8-0.00745->Node9-0.01012->Node11                                                                                                                             
#> [111] -0.95296->Node6-0.02645->Node8-0.00743->Node9-0.01009->Node11-0.00307->Node12                                                                                                             
#> [112] -0.95177->Node6-0.02642->Node8-0.00742->Node9-0.01439->Node13                                                                                                                             
#> [113] -0.951192>Node6-0.026404>Node8-0.007414>Node9-0.014382>Node13-0.000608>Node14                                                                                                             
#> [114] --0.823-->Node2--0.177-->Node15                                                                                                                                                           
#> [115] --0.9496->Node6--0.0264->Node8--0.0118->Node10--0.0122->Node16                                                                                                                            
#> [116] -0.946928>Node6-0.026285>Node8-0.007381>Node9-0.014318>Node13-0.000605>Node14-0.004483>Node17                                                                                             
#> [117] -0.94551->Node6-0.02625->Node8-0.01172->Node10-0.01218->Node16-0.00434->Node18                                                                                                            
#> [118] --0.9435->Node6--0.0565->Node19                                                                                                                                                           
#> [119] -0.94065->Node6-0.02611->Node8-0.00733->Node9-0.00996->Node11-0.00303->Node12-0.01291->Node20                                                                                             
#> [120] -0.940417>Node6-0.026104>Node8-0.007330>Node9-0.014219>Node13-0.000601>Node14-0.004452>Node17-0.006876>Node21                                                                             
#> [121] -0.93653->Node6-0.02600->Node8-0.00730->Node9-0.00992->Node11-0.00302->Node12-0.01286->Node20-0.00439->Node22                                                                             
#> [122] -0.93344->Node6-0.02591->Node8-0.00728->Node9-0.01411->Node13-0.01926->Node23                                                                                                             
#> [123] -0.93320->Node6-0.02590->Node8-0.01156->Node10-0.01203->Node16-0.00428->Node18-0.01302->Node24                                                                                            
#> [124] -0.93037->Node6-0.02583->Node8-0.00725->Node9-0.00985->Node11-0.00300->Node12-0.01277->Node20-0.01093->Node25                                                                             
#> [125] --0.8053->Node2--0.0224->Node3--0.0234->Node4--0.1490->Node26                                                                                                                             
#> [126] --0.8043->Node2--0.0223->Node3--0.1733->Node27                                                                                                                                            
#> [127] -0.928389>Node6-0.025771>Node8-0.007236>Node9-0.014037>Node13-0.000593>Node14-0.023974>Node28                                                                                             
#> [128] -0.927409>Node6-0.025743>Node8-0.007229>Node9-0.014022>Node13-0.000592>Node14-0.023949>Node28-0.001055>Node29                                                                             
#> [129] -0.927272>Node6-0.025740>Node8-0.007228>Node9-0.014020>Node13-0.000592>Node14-0.004390>Node17-0.020757>Node30                                                                             
#> [130] -0.926687>Node6-0.025723>Node8-0.007223>Node9-0.014012>Node13-0.000592>Node14-0.023930>Node28-0.001833>Node31                                                                             
#> [131] -0.92604->Node6-0.02571->Node8-0.01147->Node10-0.01193->Node16-0.00425->Node18-0.02059->Node32                                                                                            
#> [132] -0.92556->Node6-0.02569->Node8-0.00721->Node9-0.01399->Node13-0.01909->Node23-0.00844->Node33                                                                                             
#> [133] -0.925506>Node6-0.025691>Node8-0.007214>Node9-0.013994>Node13-0.000591>Node14-0.023900>Node28-0.001053>Node29-0.002051>Node34                                                             
#> [134] -0.925423>Node6-0.025688>Node8-0.007213>Node9-0.013992>Node13-0.000591>Node14-0.004381>Node17-0.020716>Node30-0.001995>Node35                                                             
#> [135] --0.9250->Node6--0.0554->Node19--0.0197->Node36                                                                                                                                           
#> [136] -0.92489->Node6-0.02567->Node8-0.00721->Node9-0.00980->Node11-0.03243->Node37                                                                                                             
#> [137] --0.8012->Node2--0.0222->Node3--0.0232->Node4--0.0664->Node5--0.0368->Node7--0.0501->Node38                                                                                               
#> [138] -0.924704>Node6-0.025668>Node8-0.007208>Node9-0.013982>Node13-0.000591>Node14-0.004378>Node17-0.020700>Node30-0.002770>Node39                                                             
#> [139] -0.92278->Node6-0.02561->Node8-0.00719->Node9-0.00977->Node11-0.00297->Node12-0.01267->Node20-0.00432->Node22-0.01468->Node40                                                             
#> [140] -0.92263->Node6-0.02561->Node8-0.00719->Node9-0.00977->Node11-0.00297->Node12-0.01267->Node20-0.00432->Node22-0.01484->Node41                                                             
#> [141] -0.92232->Node6-0.02560->Node8-0.00719->Node9-0.01395->Node13-0.01903->Node23-0.00841->Node33-0.00350->Node42                                                                             
#> [142] -0.921708>Node6-0.025585>Node8-0.007184>Node9-0.013936>Node13-0.000589>Node14-0.004364>Node17-0.020633>Node30-0.001987>Node35-0.004015>Node43                                             
#> [143] --0.7984->Node2--0.1719->Node15--0.0297->Node44                                                                                                                                           
#> [144] -0.92169->Node6-0.02558->Node8-0.00718->Node9-0.00976->Node11-0.00297->Node12-0.01265->Node20-0.01083->Node25-0.00933->Node45                                                             
#> [145] -0.79768->Node2-0.02214->Node3-0.17189->Node27-0.00828->Node46                                                                                                                            
#> [146] -0.92077->Node6-0.02556->Node8-0.00718->Node9-0.00975->Node11-0.00297->Node12-0.01264->Node20-0.01082->Node25-0.00932->Node45-0.00099->Node47                                             
#> [147] -0.92075->Node6-0.02556->Node8-0.00718->Node9-0.01392->Node13-0.01899->Node23-0.00840->Node33-0.00350->Node42-0.00170->Node48                                                             
#> [148] -0.920496>Node6-0.025552>Node8-0.007175>Node9-0.013918>Node13-0.000588>Node14-0.023770>Node28-0.001047>Node29-0.002040>Node34-0.005414>Node49                                             
#> [149] -0.92048->Node6-0.02555->Node8-0.00717->Node9-0.00975->Node11-0.00296->Node12-0.03408->Node50                                                                                             
#> [150] -0.919831>Node6-0.025533>Node8-0.007170>Node9-0.013908>Node13-0.000588>Node14-0.023753>Node28-0.001819>Node31-0.007398>Node51                                                             
#> [151] -0.91977->Node6-0.02553->Node8-0.00717->Node9-0.00974->Node11-0.03225->Node37-0.00554->Node52                                                                                             
#> [152] --0.9197->Node6--0.0255->Node8--0.0114->Node10--0.0434->Node53                                                                                                                            
#> [153] -0.919608>Node6-0.025527>Node8-0.007168>Node9-0.013905>Node13-0.000587>Node14-0.023747>Node28-0.001819>Node31-0.007396>Node51-0.000243>Node54                                             
#> [154] -0.91953->Node6-0.02552->Node8-0.00717->Node9-0.01390->Node13-0.01897->Node23-0.00839->Node33-0.00651->Node55                                                                             
#> [155] -0.91944->Node6-0.02552->Node8-0.00717->Node9-0.00974->Node11-0.00296->Node12-0.01262->Node20-0.00431->Node22-0.01463->Node40-0.00362->Node56                                             
#> [156] -0.91938->Node6-0.02552->Node8-0.01139->Node10-0.01185->Node16-0.00422->Node18-0.02044->Node32-0.00719->Node57                                                                            
#> [157] --0.7963->Node2--0.0221->Node3--0.0231->Node4--0.0660->Node5--0.0925->Node58                                                                                                              
#> [158] -0.918611>Node6-0.025499>Node8-0.007160>Node9-0.013889>Node13-0.000587>Node14-0.004349>Node17-0.020563>Node30-0.002752>Node39-0.006589>Node59                                             
#> [159] -0.918444>Node6-0.025495>Node8-0.007159>Node9-0.013887>Node13-0.000587>Node14-0.004348>Node17-0.020560>Node30-0.002751>Node39-0.006769>Node60                                             
#> [160] -0.91835->Node6-0.02549->Node8-0.00716->Node9-0.01389->Node13-0.01894->Node23-0.00838->Node33-0.00349->Node42-0.00169->Node48-0.00261->Node61                                             
#> [161] -0.91813->Node6-0.02549->Node8-0.00716->Node9-0.01388->Node13-0.01894->Node23-0.01640->Node62                                                                                             
#> [162] -0.918079>Node6-0.025484>Node8-0.007156>Node9-0.009723>Node11-0.002957>Node12-0.012605>Node20-0.010785>Node25-0.009296>Node45-0.000987>Node47-0.002928>Node63                             
#> [163] -0.918043>Node6-0.025483>Node8-0.007156>Node9-0.013881>Node13-0.000586>Node14-0.023707>Node28-0.001816>Node31-0.007384>Node51-0.001944>Node64                                             
#> [164] -0.918037>Node6-0.025483>Node8-0.007156>Node9-0.013881>Node13-0.000586>Node14-0.004346>Node17-0.020551>Node30-0.002750>Node39-0.006766>Node60-0.000444>Node65                             
#> [165] -0.917917>Node6-0.025480>Node8-0.007155>Node9-0.013879>Node13-0.000586>Node14-0.004346>Node17-0.006712>Node21-0.023926>Node66                                                             
#> [166] -0.917870>Node6-0.025479>Node8-0.007154>Node9-0.013878>Node13-0.018934>Node23-0.008374>Node33-0.003487>Node42-0.001694>Node48-0.002611>Node61-0.000518>Node67                             
#> [167] -0.917658>Node6-0.025473>Node8-0.007153>Node9-0.013875>Node13-0.018930>Node23-0.008372>Node33-0.003486>Node42-0.001693>Node48-0.002611>Node61-0.000518>Node67-0.000231>Node68             
#> [168] -0.91747->Node6-0.02547->Node8-0.00715->Node9-0.00972->Node11-0.00295->Node12-0.03396->Node50-0.00328->Node69                                                                             
#> [169] -0.91743->Node6-0.02547->Node8-0.00715->Node9-0.01387->Node13-0.01893->Node23-0.00837->Node33-0.00650->Node55-0.00228->Node70                                                             
#> [170] --0.9172->Node6--0.0549->Node19--0.0278->Node71                                                                                                                                           
#> [171] -0.917190>Node6-0.025460>Node8-0.007149>Node9-0.009714>Node11-0.002954>Node12-0.012592>Node20-0.010775>Node25-0.009287>Node45-0.000986>Node47-0.002925>Node63-0.000968>Node72             
#> [172] -0.91714->Node6-0.02546->Node8-0.01136->Node10-0.01182->Node16-0.00421->Node18-0.01279->Node24-0.01721->Node73                                                                            
#> [173] -0.917067>Node6-0.025456>Node8-0.007148>Node9-0.013866>Node13-0.000586>Node14-0.004342>Node17-0.020529>Node30-0.002747>Node39-0.006578>Node59-0.001681>Node74                             
#> [174] -0.916897>Node6-0.025452>Node8-0.011361>Node10-0.011816>Node16-0.004209>Node18-0.012791>Node24-0.017208>Node73-0.000267>Node75                                                            
#> [175] -0.91674->Node6-0.02545->Node8-0.00715->Node9-0.00971->Node11-0.00295->Node12-0.01259->Node20-0.00429->Node22-0.01458->Node40-0.00361->Node56-0.00294->Node76                             
#> [176] --0.9165->Node6--0.0254->Node8--0.0114->Node10--0.0118->Node16--0.0349->Node77                                                                                                            
#> [177] -0.91635->Node6-0.02544->Node8-0.00714->Node9-0.01386->Node13-0.01890->Node23-0.00836->Node33-0.00649->Node55-0.00228->Node70-0.00118->Node78                                             
#> [178] -0.916351>Node6-0.025436>Node8-0.007143>Node9-0.013855>Node13-0.000585>Node14-0.004338>Node17-0.006700>Node21-0.025590>Node79                                                             
#> [179] -0.916207>Node6-0.025432>Node8-0.007142>Node9-0.013853>Node13-0.000585>Node14-0.023659>Node28-0.001042>Node29-0.002031>Node34-0.010048>Node80                                             
#> [180] --0.7936->Node2--0.0220->Node3--0.1710->Node27--0.0134->Node81                                                                                                                            
#> [181] -0.916115>Node6-0.025430>Node8-0.007141>Node9-0.013852>Node13-0.000585>Node14-0.023657>Node28-0.001812>Node31-0.011408>Node82                                                             
#> [182] -0.91602->Node6-0.02543->Node8-0.00714->Node9-0.01385->Node13-0.01890->Node23-0.00836->Node33-0.00649->Node55-0.00382->Node83                                                             
#> [183] --0.7935->Node2--0.0220->Node3--0.0230->Node4--0.1468->Node26--0.0148->Node84                                                                                                             
#> [184] -0.793192>Node2-0.022015>Node3-0.170924>Node27-0.013367>Node81-0.000502>Node85                                                                                                            
#> [185] -0.915636>Node6-0.025417>Node8-0.007137>Node9-0.013844>Node13-0.018888>Node23-0.008354>Node33-0.003478>Node42-0.001690>Node48-0.002605>Node61-0.000517>Node67-0.002434>Node86             
#> [186] -0.91560->Node6-0.02542->Node8-0.00714->Node9-0.00970->Node11-0.00295->Node12-0.01257->Node20-0.01076->Node25-0.00927->Node45-0.00660->Node87                                             
#> [187] -0.915577>Node6-0.025415>Node8-0.007137>Node9-0.013844>Node13-0.000585>Node14-0.004335>Node17-0.006695>Node21-0.023865>Node66-0.002549>Node88                                             
#> [188] -0.915486>Node6-0.025412>Node8-0.007136>Node9-0.013842>Node13-0.000585>Node14-0.004334>Node17-0.006694>Node21-0.025566>Node79-0.000944>Node89                                             
#> [189] -0.91541->Node6-0.05481->Node19-0.02777->Node71-0.00201->Node90                                                                                                                           
#> [190] --0.7928->Node2--0.1707->Node15--0.0365->Node91                                                                                                                                           
#> [191] -0.91515->Node6-0.02540->Node8-0.01134->Node10-0.01179->Node16-0.03486->Node77-0.00145->Node92                                                                                            
#> [192] -0.792749>Node2-0.170653>Node15-0.036481>Node91-0.000117>Node93                                                                                                                           
#> [193] -0.91511->Node6-0.02540->Node8-0.00713->Node9-0.00969->Node11-0.00295->Node12-0.03388->Node50-0.00327->Node69-0.00257->Node94                                                             
#> [194] -9.15e-01>Node6-2.54e-02>Node8-7.13e-03>Node9-9.69e-03>Node11-2.95e-03>Node12-3.39e-02>Node50-3.27e-03>Node69-2.57e-03>Node94-4.25e-05>Node95                                             
#> [195] -0.914989>Node6-0.025399>Node8-0.007132>Node9-0.013835>Node13-0.000585>Node14-0.004332>Node17-0.020482>Node30-0.002741>Node39-0.006744>Node60-0.000442>Node65-0.003319>Node96             
#> [196] -0.792554>Node2-0.170611>Node15-0.036473>Node91-0.000117>Node93-0.000246>Node97                                                                                                           
#> [197] -0.914842>Node6-0.025395>Node8-0.007131>Node9-0.009689>Node11-0.002946>Node12-0.012560>Node20-0.010747>Node25-0.009263>Node45-0.006594>Node87-0.000832>Node98                             
#> [198] -0.91483->Node6-0.02539->Node8-0.01134->Node10-0.01179->Node16-0.00420->Node18-0.01276->Node24-0.01717->Node73-0.00252->Node99