Scale the phylogenetic flow features to a constant sum.

Can be used to standardise branch lengths to reasonable values

Usage

pf_scale_flow_sum(x, scale_to = 1)

Arguments

x

a pfc object

scale_to

The value to scale the sums to. Default is 1.

Value

a scaled pfc object

Examples

pf_scale_flow_sum(rpfc(100))
#> <pfc<e:198>[198]>
#> First  10 phylogenetic flows: 
#> [1] ◎──0.588──→ Node4 ──0.086──→ Node6 ──0.256──→ Node23 ──0.057──→ Node64 ──0.013──→ t70 
#> 
#> [2] ◎──0.5884──→ Node4 ──0.0855──→ Node6 ──0.2563──→ Node23 ──0.0571──→ Node64 ──0.0018──→ Node67 ──0.0110──→ t1 
#> 
#> [3] ◎──0.5884──→ Node4 ──0.0855──→ Node6 ──0.2563──→ Node23 ──0.0571──→ Node64 ──0.0018──→ Node67 ──0.0110──→ t63 
#> 
#> [4] ◎──0.588──→ Node4 ──0.086──→ Node6 ──0.256──→ Node23 ──0.070──→ t51 
#> 
#> [5] ◎──0.588──→ Node4 ──0.086──→ Node6 ──0.131──→ Node8 ──0.195──→ t49 
#> 
#> [6] ◎──0.588──→ Node4 ──0.086──→ Node6 ──0.131──→ Node8 ──0.173──→ Node46 ──0.023──→ t95 
#> 
#> [7] ◎──0.58838──→ Node4 ──0.08554──→ Node6 ──0.13102──→ Node8 ──0.17254──→ Node46 ──0.00046──→ Node47 ──0.00849──→ Node60 ──0.00032──→ Node62 ──0.00505──→ Node72 ──0.00757──→ Node96 ──0.00062──→ t68 
#> 
#> [8] ◎──0.58838──→ Node4 ──0.08554──→ Node6 ──0.13102──→ Node8 ──0.17254──→ Node46 ──0.00046──→ Node47 ──0.00849──→ Node60 ──0.00032──→ Node62 ──0.00505──→ Node72 ──0.00757──→ Node96 ──0.00062──→ t5 
#> 
#> [9] ◎──0.58838──→ Node4 ──0.08554──→ Node6 ──0.13102──→ Node8 ──0.17254──→ Node46 ──0.00046──→ Node47 ──0.00849──→ Node60 ──0.00032──→ Node62 ──0.00505──→ Node72 ──0.00730──→ Node95 ──0.00090──→ t96 
#> 
#> [10] ◎──0.58838──→ Node4 ──0.08554──→ Node6 ──0.13102──→ Node8 ──0.17254──→ Node46 ──0.00046──→ Node47 ──0.00849──→ Node60 ──0.00032──→ Node62 ──0.00505──→ Node72 ──0.00730──→ Node95 ──0.00090──→ t97