Calculate a 'second order' pfc, which represents summed branch length from each node to the terminal node along each flow. When used in modelling traits, a second order pfc implies that rates of evolution of the trait are themselves evolving according to Brownian motion.

Calculate a 'second order' pfc, which represents summed branch length from each node to the terminal node along each flow. When used in modelling traits, a second order pfc implies that rates of evolution of the trait are themselves evolving according to Brownian motion.

Usage

pf_second_order(x, ...)

Arguments

x

A pfc object

...

Additional arguments passed to or other methods.

Value

A pfc object

Examples

pf_second_order(rpfc(100))
#> <pfc<e:198>[198]>
#> First  10 phylogenetic flows: 
#> [1] ◎──0.812──→ Node2 ──0.630──→ Node28 ──0.040──→ Node43 ──0.024──→ t51 
#> 
#> [2] ◎──0.812──→ Node2 ──0.630──→ Node28 ──0.040──→ Node43 ──0.024──→ Node53 ──0.018──→ t22 
#> 
#> [3] ◎──0.812──→ Node2 ──0.630──→ Node28 ──0.040──→ Node43 ──0.024──→ Node53 ──0.018──→ t56 
#> 
#> [4] ◎──0.8118──→ Node2 ──0.6299──→ Node28 ──0.0404──→ Node69 ──0.0095──→ t93 
#> 
#> [5] ◎──0.8118──→ Node2 ──0.6299──→ Node28 ──0.0404──→ Node69 ──0.0095──→ t62 
#> 
#> [6] ◎──0.812──→ Node2 ──0.630──→ Node4 ──0.413──→ Node6 ──0.351──→ Node8 ──0.297──→ Node12 ──0.113──→ Node20 ──0.083──→ Node60 ──0.014──→ t40 
#> 
#> [7] ◎──0.812──→ Node2 ──0.630──→ Node4 ──0.413──→ Node6 ──0.351──→ Node8 ──0.297──→ Node12 ──0.113──→ Node20 ──0.083──→ Node60 ──0.014──→ t4 
#> 
#> [8] ◎──0.812──→ Node2 ──0.630──→ Node4 ──0.413──→ Node6 ──0.351──→ Node8 ──0.297──→ Node12 ──0.113──→ Node20 ──0.083──→ Node21 ──0.082──→ Node34 ──0.034──→ Node44 ──0.024──→ t97 
#> 
#> [9] ◎──0.8118──→ Node2 ──0.6299──→ Node4 ──0.4125──→ Node6 ──0.3512──→ Node8 ──0.2969──→ Node12 ──0.1133──→ Node20 ──0.0833──→ Node21 ──0.0822──→ Node34 ──0.0335──→ Node44 ──0.0238──→ Node75 ──0.0071──→ t50 
#> 
#> [10] ◎──0.8118──→ Node2 ──0.6299──→ Node4 ──0.4125──→ Node6 ──0.3512──→ Node8 ──0.2969──→ Node12 ──0.1133──→ Node20 ──0.0833──→ Node21 ──0.0822──→ Node34 ──0.0335──→ Node44 ──0.0238──→ Node75 ──0.0071──→ t42