Calculate a 'second order' pfc, which represents summed branch length from each node to the terminal node along each flow. When used in modelling traits, a second order pfc implies that rates of evolution of the trait are themselves evolving according to Brownian motion.

Calculate a 'second order' pfc, which represents summed branch length from each node to the terminal node along each flow. When used in modelling traits, a second order pfc implies that rates of evolution of the trait are themselves evolving according to Brownian motion.

Usage

pf_second_order(x, ...)

Arguments

x

A pfc object

...

Additional arguments passed to or other methods.

Value

A pfc object

Examples

pf_second_order(rpfc(100))
#> <pfc<e:198>[198]>
#> First  10 phylogenetic flows: 
#> [1] ◎── 6.45──→ Node9 ── 0.22──→ t97 
#> 
#> [2] ◎──6.450──→ Node9 ──0.219──→ Node21 ──0.086──→ Node37 ──0.035──→ t58 
#> 
#> [3] ◎──6.450──→ Node9 ──0.219──→ Node21 ──0.086──→ Node37 ──0.035──→ t81 
#> 
#> [4] ◎──6.4503──→ Node9 ──0.2195──→ Node21 ──0.0859──→ Node81 ──0.0049──→ t14 
#> 
#> [5] ◎──6.4503──→ Node9 ──0.2195──→ Node21 ──0.0859──→ Node81 ──0.0049──→ t42 
#> 
#> [6] ◎──6.450──→ Node2 ──1.710──→ Node3 ──0.649──→ Node5 ──0.502──→ Node6 ──0.338──→ Node34 ──0.045──→ Node41 ──0.033──→ t98 
#> 
#> [7] ◎──6.4503──→ Node2 ──1.7104──→ Node3 ──0.6495──→ Node5 ──0.5022──→ Node6 ──0.3383──→ Node34 ──0.0453──→ Node41 ──0.0328──→ Node90 ──0.0023──→ t94 
#> 
#> [8] ◎──6.4503──→ Node2 ──1.7104──→ Node3 ──0.6495──→ Node5 ──0.5022──→ Node6 ──0.3383──→ Node34 ──0.0453──→ Node41 ──0.0328──→ Node90 ──0.0023──→ t28 
#> 
#> [9] ◎──6.4503──→ Node2 ──1.7104──→ Node3 ──0.6495──→ Node5 ──0.5022──→ Node6 ──0.3383──→ Node34 ──0.0453──→ Node70 ──0.0085──→ Node99 ──0.0002──→ t61 
#> 
#> [10] ◎──6.4503──→ Node2 ──1.7104──→ Node3 ──0.6495──→ Node5 ──0.5022──→ Node6 ──0.3383──→ Node34 ──0.0453──→ Node70 ──0.0085──→ Node99 ──0.0002──→ t75